Exponential Linear Unit (ELU) Activation Function

Exponential Linear Unit (ELU) Activation Function

Usage

ELU(x, alpha = 1)

Arguments

x

A numeric vector. All elements must be finite and non-missing.

alpha

A numeric scalar. All elements must be finite and non-missing.

Value

A numeric vector where the ReLU function has been applied to each element of x.

See also

Author

Matthew Fidler

Examples

ELU(c(-1, 0, 1, 2), 2)
#> [1] -1.264241  0.000000  1.000000  2.000000

# Can also be used in rxode2:
x <- rxode2({
   r=SELU(time)
})
#>  
#>  
#> C model:
#> // Define 1 LHS values
#> #define _LHS_0_ 0 // r
#> #define _getRxSolve_ _rxb250bf28dfa8aab06587778de36541940__getRxSolve__b90S
#> #define _evalUdf _rxb250bf28dfa8aab06587778de36541941__evalUdf_UHjl
#> #define _solveData _rxb250bf28dfa8aab06587778de36541942__solveData_sA2J
#> #define _assign_ptr _rxb250bf28dfa8aab06587778de36541943__assign_ptr_OCP3
#> #define _rxRmModelLib _rxb250bf28dfa8aab06587778de36541944__rxRmModelLib_OyHg
#> #define _rxGetModelLib _rxb250bf28dfa8aab06587778de36541945__rxGetModelLib_jCyj
#> #define _old_c _rxb250bf28dfa8aab06587778de36541946__old_c_3Whg
#> #define _ptrid _rxb250bf28dfa8aab06587778de36541947__ptrid_1DId
#> #define _rxIsCurrentC _rxb250bf28dfa8aab06587778de36541948__rxIsCurrentC_TQhw
#> #define _sumPS _rxb250bf28dfa8aab06587778de36541949__sumPS_BDgU
#> #define _prodPS _rxb250bf28dfa8aab06587778de365419410__prodPS_de5l
#> #define _prodType _rxb250bf28dfa8aab06587778de365419411__prodType_lzyO
#> #define _sumType _rxb250bf28dfa8aab06587778de365419412__sumType_SJAW
#> #define _update_par_ptr _rxb250bf28dfa8aab06587778de365419413__update_par_ptr_UmCi
#> #define _getParCov _rxb250bf28dfa8aab06587778de365419414__getParCov_4LpG
#> #define _rxode2_rxAssignPtr _rxb250bf28dfa8aab06587778de365419415__rxode2_rxAssignPtr_ZKaB
#> #define _rxQr _rxb250bf28dfa8aab06587778de365419416__rxQr_wXXH
#> #define _compareFactorVal _rxb250bf28dfa8aab06587778de365419417__compareFactorVal_RN1S
#> #define _sum _rxb250bf28dfa8aab06587778de365419418__sum_O1NY
#> #define _udf _rxb250bf28dfa8aab06587778de365419419__udf_rNWb
#> #define _sign _rxb250bf28dfa8aab06587778de365419420__sign_GHTv
#> #define _prod _rxb250bf28dfa8aab06587778de365419421__prod_ODf3
#> #define _max _rxb250bf28dfa8aab06587778de365419422__max_z9Wq
#> #define _min _rxb250bf28dfa8aab06587778de365419423__min_iEXp
#> #define _transit4P _rxb250bf28dfa8aab06587778de365419424__transit4P_1Tno
#> #define _transit3P _rxb250bf28dfa8aab06587778de365419425__transit3P_PDPf
#> #define _assignFuns0 _rxb250bf28dfa8aab06587778de365419426__assignFuns0_VeVp
#> #define _assignFuns _rxb250bf28dfa8aab06587778de365419427__assignFuns_hs0t
#> #define _rxord _rxb250bf28dfa8aab06587778de365419428__rxord_zVMm
#> #define __assignFuns2 _rxb250bf28dfa8aab06587778de365419429___assignFuns2_adrZ
#> #define _llikCauchyDscale _rxb250bf28dfa8aab06587778de365419430__llikCauchyDscale_FurF
#> #define _llikCauchyDlocation _rxb250bf28dfa8aab06587778de365419431__llikCauchyDlocation_Ol7C
#> #define _llikCauchy _rxb250bf28dfa8aab06587778de365419432__llikCauchy_Afrv
#> #define _llikGammaDrate _rxb250bf28dfa8aab06587778de365419433__llikGammaDrate_YZgF
#> #define _llikGammaDshape _rxb250bf28dfa8aab06587778de365419434__llikGammaDshape_2u4z
#> #define _llikGamma _rxb250bf28dfa8aab06587778de365419435__llikGamma_bAu2
#> #define _llikWeibullDscale _rxb250bf28dfa8aab06587778de365419436__llikWeibullDscale_aEuo
#> #define _llikWeibullDshape _rxb250bf28dfa8aab06587778de365419437__llikWeibullDshape_hDjD
#> #define _llikWeibull _rxb250bf28dfa8aab06587778de365419438__llikWeibull_SKZF
#> #define _llikUnifDbeta _rxb250bf28dfa8aab06587778de365419439__llikUnifDbeta_Suyo
#> #define _llikUnifDalpha _rxb250bf28dfa8aab06587778de365419440__llikUnifDalpha_talz
#> #define _llikUnif _rxb250bf28dfa8aab06587778de365419441__llikUnif_iDIi
#> #define _llikGeomDp _rxb250bf28dfa8aab06587778de365419442__llikGeomDp_rE0w
#> #define _llikGeom _rxb250bf28dfa8aab06587778de365419443__llikGeom_RkfP
#> #define _llikFDdf2 _rxb250bf28dfa8aab06587778de365419444__llikFDdf2_Ljf5
#> #define _llikFDdf1 _rxb250bf28dfa8aab06587778de365419445__llikFDdf1_r4q6
#> #define _llikF _rxb250bf28dfa8aab06587778de365419446__llikF_UFgA
#> #define _llikExpDrate _rxb250bf28dfa8aab06587778de365419447__llikExpDrate_xJMt
#> #define _llikExp _rxb250bf28dfa8aab06587778de365419448__llikExp_p6ul
#> #define _llikChisqDdf _rxb250bf28dfa8aab06587778de365419449__llikChisqDdf_NzLw
#> #define _llikChisq _rxb250bf28dfa8aab06587778de365419450__llikChisq_D19q
#> #define _llikTDsd _rxb250bf28dfa8aab06587778de365419451__llikTDsd_DD2w
#> #define _llikTDmean _rxb250bf28dfa8aab06587778de365419452__llikTDmean_FmHs
#> #define _llikTDdf _rxb250bf28dfa8aab06587778de365419453__llikTDdf_Iyal
#> #define _llikT _rxb250bf28dfa8aab06587778de365419454__llikT_plP8
#> #define _llikBetaDshape2 _rxb250bf28dfa8aab06587778de365419455__llikBetaDshape2_5ebe
#> #define _llikBetaDshape1 _rxb250bf28dfa8aab06587778de365419456__llikBetaDshape1_Kyis
#> #define _llikBeta _rxb250bf28dfa8aab06587778de365419457__llikBeta_yDfO
#> #define _llikNbinomMuDmu _rxb250bf28dfa8aab06587778de365419458__llikNbinomMuDmu_2cbP
#> #define _llikNbinomMu _rxb250bf28dfa8aab06587778de365419459__llikNbinomMu_rWy1
#> #define _llikNbinomDprob _rxb250bf28dfa8aab06587778de365419460__llikNbinomDprob_fNug
#> #define _llikNbinom _rxb250bf28dfa8aab06587778de365419461__llikNbinom_fkXE
#> #define _llikBinomDprob _rxb250bf28dfa8aab06587778de365419462__llikBinomDprob_VPjd
#> #define _llikBinom _rxb250bf28dfa8aab06587778de365419463__llikBinom_OxI3
#> #define _llikPoisDlambda _rxb250bf28dfa8aab06587778de365419464__llikPoisDlambda_ZD7w
#> #define _llikPois _rxb250bf28dfa8aab06587778de365419465__llikPois_nkXX
#> #define _llikNormDsd _rxb250bf28dfa8aab06587778de365419466__llikNormDsd_5OAg
#> #define _llikNormDmean _rxb250bf28dfa8aab06587778de365419467__llikNormDmean_fwCy
#> #define _llikNorm _rxb250bf28dfa8aab06587778de365419468__llikNorm_Neo4
#> #define simeps _rxb250bf28dfa8aab06587778de365419469_simeps_WlQY
#> #define simeta _rxb250bf28dfa8aab06587778de365419470_simeta_C1yB
#> #define expit _rxb250bf28dfa8aab06587778de365419471_expit_mspb
#> #define logit _rxb250bf28dfa8aab06587778de365419472_logit_u285
#> #define gammapDer _rxb250bf28dfa8aab06587778de365419473_gammapDer_hP47
#> #define lowergamma _rxb250bf28dfa8aab06587778de365419474_lowergamma_wDfC
#> #define uppergamma _rxb250bf28dfa8aab06587778de365419475_uppergamma_Yp7G
#> #define gammaqInva _rxb250bf28dfa8aab06587778de365419476_gammaqInva_H0qP
#> #define gammaqInv _rxb250bf28dfa8aab06587778de365419477_gammaqInv_kiVy
#> #define gammapInva _rxb250bf28dfa8aab06587778de365419478_gammapInva_MAm1
#> #define gammapInv _rxb250bf28dfa8aab06587778de365419479_gammapInv_mA1m
#> #define gammaq _rxb250bf28dfa8aab06587778de365419480_gammaq_pxNA
#> #define gammap _rxb250bf28dfa8aab06587778de365419481_gammap_bn2j
#> #define phi _rxb250bf28dfa8aab06587778de365419482_phi_D07g
#> #define d2ELUa _rxb250bf28dfa8aab06587778de365419483_d2ELUa_dKOh
#> #define dELUa _rxb250bf28dfa8aab06587778de365419484_dELUa_hvSa
#> #define d2aELU _rxb250bf28dfa8aab06587778de365419485_d2aELU_AZta
#> #define d2ELU _rxb250bf28dfa8aab06587778de365419486_d2ELU_pamO
#> #define dELU _rxb250bf28dfa8aab06587778de365419487_dELU_qjJ6
#> #define ELU _rxb250bf28dfa8aab06587778de365419488_ELU_6nN3
#> #define dPReLUa1 _rxb250bf28dfa8aab06587778de365419489_dPReLUa1_zb9X
#> #define dPReLUa _rxb250bf28dfa8aab06587778de365419490_dPReLUa_7x2A
#> #define dPReLU _rxb250bf28dfa8aab06587778de365419491_dPReLU_wyis
#> #define PReLU _rxb250bf28dfa8aab06587778de365419492_PReLU_y2Rd
#> #define dSwish _rxb250bf28dfa8aab06587778de365419493_dSwish_zJrX
#> #define Swish _rxb250bf28dfa8aab06587778de365419494_Swish_sVrr
#> #define dlReLU _rxb250bf28dfa8aab06587778de365419495_dlReLU_L0Cw
#> #define lReLU _rxb250bf28dfa8aab06587778de365419496_lReLU_Hsx0
#> #define dSELU _rxb250bf28dfa8aab06587778de365419497_dSELU_YDej
#> #define SELU _rxb250bf28dfa8aab06587778de365419498_SELU_sNtq
#> #define d4softplus _rxb250bf28dfa8aab06587778de365419499_d4softplus_j2Kk
#> #define d3softplus _rxb250bf28dfa8aab06587778de3654194100_d3softplus_uCTN
#> #define d2softplus _rxb250bf28dfa8aab06587778de3654194101_d2softplus_OD0M
#> #define dsoftplus _rxb250bf28dfa8aab06587778de3654194102_dsoftplus_Wq8N
#> #define softplus _rxb250bf28dfa8aab06587778de3654194103_softplus_GU15
#> #define d4GELU _rxb250bf28dfa8aab06587778de3654194104_d4GELU_zw77
#> #define d3GELU _rxb250bf28dfa8aab06587778de3654194105_d3GELU_MTmU
#> #define d2GELU _rxb250bf28dfa8aab06587778de3654194106_d2GELU_6whV
#> #define dGELU _rxb250bf28dfa8aab06587778de3654194107_dGELU_WGiy
#> #define GELU _rxb250bf28dfa8aab06587778de3654194108_GELU_en4s
#> #define dReLU _rxb250bf28dfa8aab06587778de3654194109_dReLU_piMW
#> #define ReLU _rxb250bf28dfa8aab06587778de3654194110_ReLU_T6VR
#> #define riweibull _rxb250bf28dfa8aab06587778de3654194111_riweibull_Ye9I
#> #define riunif _rxb250bf28dfa8aab06587778de3654194112_riunif_tadG
#> #define rit_ _rxb250bf28dfa8aab06587778de3654194113_rit__JLl0
#> #define ripois _rxb250bf28dfa8aab06587778de3654194114_ripois_3GaI
#> #define ribeta _rxb250bf28dfa8aab06587778de3654194115_ribeta_5llO
#> #define rigamma _rxb250bf28dfa8aab06587778de3654194116_rigamma_3kWH
#> #define rigeom _rxb250bf28dfa8aab06587778de3654194117_rigeom_GXCd
#> #define rif _rxb250bf28dfa8aab06587778de3654194118_rif_kGsq
#> #define riexp _rxb250bf28dfa8aab06587778de3654194119_riexp_2NBv
#> #define richisq _rxb250bf28dfa8aab06587778de3654194120_richisq_uBr0
#> #define ricauchy _rxb250bf28dfa8aab06587778de3654194121_ricauchy_sO4e
#> #define rinbinomMu _rxb250bf28dfa8aab06587778de3654194122_rinbinomMu_upVb
#> #define rinbinom _rxb250bf28dfa8aab06587778de3654194123_rinbinom_TsfO
#> #define ribinom _rxb250bf28dfa8aab06587778de3654194124_ribinom_YFC4
#> #define rinorm _rxb250bf28dfa8aab06587778de3654194125_rinorm_Jz7i
#> #define rxweibull _rxb250bf28dfa8aab06587778de3654194126_rxweibull_E8vt
#> #define rxunif _rxb250bf28dfa8aab06587778de3654194127_rxunif_CFcR
#> #define rxt_ _rxb250bf28dfa8aab06587778de3654194128_rxt__J6Ge
#> #define rxpois _rxb250bf28dfa8aab06587778de3654194129_rxpois_v1Yq
#> #define rxbeta _rxb250bf28dfa8aab06587778de3654194130_rxbeta_4CUG
#> #define rxgamma _rxb250bf28dfa8aab06587778de3654194131_rxgamma_PRB4
#> #define rxgeom _rxb250bf28dfa8aab06587778de3654194132_rxgeom_1jyX
#> #define rxf _rxb250bf28dfa8aab06587778de3654194133_rxf_A3Sc
#> #define rxexp _rxb250bf28dfa8aab06587778de3654194134_rxexp_akWU
#> #define rxchisq _rxb250bf28dfa8aab06587778de3654194135_rxchisq_0xSy
#> #define rxcauchy _rxb250bf28dfa8aab06587778de3654194136_rxcauchy_XMV4
#> #define rxnbinomMu _rxb250bf28dfa8aab06587778de3654194137_rxnbinomMu_yezX
#> #define rxnbinom _rxb250bf28dfa8aab06587778de3654194138_rxnbinom_lYmd
#> #define rxbinom _rxb250bf28dfa8aab06587778de3654194139_rxbinom_NMqS
#> #define rxnorm _rxb250bf28dfa8aab06587778de3654194140_rxnorm_Bxi1
#> #define linCmtB _rxb250bf28dfa8aab06587778de3654194141_linCmtB_cpdh
#> #define linCmtA _rxb250bf28dfa8aab06587778de3654194142_linCmtA_rHA6
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){}
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){}
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  rx_371c5187714088175e3e83183aec6649__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *rx_371c5187714088175e3e83183aec6649__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP rx_371c5187714088175e3e83183aec6649__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void rx_371c5187714088175e3e83183aec6649__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double r = NA_REAL;
#> 
#>   (void)t;
#>   (void)r;
#> 
#>   r = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   r=SELU(t);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void rx_371c5187714088175e3e83183aec6649__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void rx_371c5187714088175e3e83183aec6649__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void rx_371c5187714088175e3e83183aec6649__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double r = NA_REAL;
#> 
#>   (void)t;
#>   (void)r;
#> 
#>   r = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   r=SELU(t);
#> 
#>   _lhs[_LHS_0_]=r;
#> }
#> // Functional based bioavailability
#> double rx_371c5187714088175e3e83183aec6649__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double rx_371c5187714088175e3e83183aec6649__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double rx_371c5187714088175e3e83183aec6649__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double rx_371c5187714088175e3e83183aec6649__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void rx_371c5187714088175e3e83183aec6649__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void rx_371c5187714088175e3e83183aec6649__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void rx_371c5187714088175e3e83183aec6649__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP rx_371c5187714088175e3e83183aec6649__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_371c5187714088175e3e83183aec6649__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAAv(zAAAv7#aT)cmT,MAg)(Zgt~kaBaifxmBHOOiZ.vi3Rzvixv;oGtSIz>Z2pvZD*JjsV/U+</vKy]SWYiMgh1y)kisK=`BIO;z%%_HAiAmB(#kQ*~EjxI<aGxfFVJZue~*#^~>=F2YB!^v_TSrLu6c{3~tG,rUlQq#X<^8XARvbclc94e7s_UfZer`2v:CqfZe#]^h=i{ID22)\")\"`ZPT/#IKoF]0w0\"@x!@H1^6VT2zmv0L:ymcO|0bx)1[Wjc`,IfCKkE&``Z#[xMn|XGJf!`$`+u.DD.[P[O!jR#:!+]*~/jv!1Py#m<1\"3V<T=b!@f{*\?,m!OH+%%$a(qNZ4)^D!2[\?5>MLYx@&MozCaq2a*_zlTA*wgwAgAHTPC5g%%L/}Fx#/SJ\"XzjHFWG+`~JySY!)ziG5wOu,dq!y}\?nHpqn`/;>^[5,%%eZoEjl%%XD+AeZ}1R!2uhj>x!Xs#PdRk`J,2fmT^y,orH}=Bl}n&h0&^|{hf%%lq3RTQ!kvp[GdG].7XCI5L/>{B/E&MrL9t4bHyy]Hh+x3R9o:x39pi:X(;rcZy3O\"pm$srvGVtMdq2A\?qFb%%`F1Km[V@t8{uIKLuOwoOc2)[Bg@<Da/_Oa&<Lr`\"\?}+z@B`K9_/<8;5k%%{No;&k2ac}i)G|\"6I)$pJf_|[azi{FRK!szG8k^;oLHr$cW9rZW<z$6>P9~kn;}r5<sb+W(U0*OOd1c=kyHB~j10Cdv{&2Nb^YL6f[OQ^DJ)(LA*$k&GN5`M\"/:W7FWd:!aLw(X{qO#F>hyuSDzB.t74h$TK}vL]&j;n:CTD92nI0~D2:tX#$yGiQV+Znt0wl^DlOqD#Lsj)XEJ]Ou>D*j($6JCoP9tL/i=#i:3k~yx}vFf\?eFmh2qgx!5WG%%nHwC@5:0\"8O{0D;GW@i6]}zegRA9\"))r2fw)vv7`J>S]KFMBCpTb><c,x~YANY<b^D|Fjg7G7#[oi}Du`pY]LR<Xy$tdONV|&>F}tQ!ntt|cY!gJ[j9cO*ERfzNVkjcntC+hx1#$4+J");
#>     char buf[1060];
#> #define __doBufN__ 1060
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void rx_371c5187714088175e3e83183aec6649__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   rx_371c5187714088175e3e83183aec6649__dydt(neq, *t, A, DADT);
#> }
#> extern int rx_371c5187714088175e3e83183aec6649__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   rx_371c5187714088175e3e83183aec6649__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void rx_371c5187714088175e3e83183aec6649__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   rx_371c5187714088175e3e83183aec6649__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void rx_371c5187714088175e3e83183aec6649__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_rx_371c5187714088175e3e83183aec6649_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("rx_371c5187714088175e3e83183aec6649_","rx_371c5187714088175e3e83183aec6649__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("rx_371c5187714088175e3e83183aec6649_","rx_371c5187714088175e3e83183aec6649__assignFuns", (DL_FUNC) rx_371c5187714088175e3e83183aec6649__assignFuns);
#>   R_RegisterCCallable("rx_371c5187714088175e3e83183aec6649_","rx_371c5187714088175e3e83183aec6649__inis",(DL_FUNC) rx_371c5187714088175e3e83183aec6649__inis);
#>   R_RegisterCCallable("rx_371c5187714088175e3e83183aec6649_","rx_371c5187714088175e3e83183aec6649__dydt",(DL_FUNC) rx_371c5187714088175e3e83183aec6649__dydt);
#>   R_RegisterCCallable("rx_371c5187714088175e3e83183aec6649_","rx_371c5187714088175e3e83183aec6649__calc_lhs",(DL_FUNC) rx_371c5187714088175e3e83183aec6649__calc_lhs);
#>   R_RegisterCCallable("rx_371c5187714088175e3e83183aec6649_","rx_371c5187714088175e3e83183aec6649__calc_jac",(DL_FUNC) rx_371c5187714088175e3e83183aec6649__calc_jac);
#>   R_RegisterCCallable("rx_371c5187714088175e3e83183aec6649_","rx_371c5187714088175e3e83183aec6649__dydt_lsoda", (DL_FUNC) rx_371c5187714088175e3e83183aec6649__dydt_lsoda);
#>   R_RegisterCCallable("rx_371c5187714088175e3e83183aec6649_","rx_371c5187714088175e3e83183aec6649__calc_jac_lsoda", (DL_FUNC) rx_371c5187714088175e3e83183aec6649__calc_jac_lsoda);
#>   R_RegisterCCallable("rx_371c5187714088175e3e83183aec6649_","rx_371c5187714088175e3e83183aec6649__ode_solver_solvedata", (DL_FUNC) rx_371c5187714088175e3e83183aec6649__ode_solver_solvedata);
#>   R_RegisterCCallable("rx_371c5187714088175e3e83183aec6649_","rx_371c5187714088175e3e83183aec6649__ode_solver_get_solvedata", (DL_FUNC) rx_371c5187714088175e3e83183aec6649__ode_solver_get_solvedata);
#>   R_RegisterCCallable("rx_371c5187714088175e3e83183aec6649_","rx_371c5187714088175e3e83183aec6649__F", (DL_FUNC) rx_371c5187714088175e3e83183aec6649__F);
#>   R_RegisterCCallable("rx_371c5187714088175e3e83183aec6649_","rx_371c5187714088175e3e83183aec6649__Lag", (DL_FUNC) rx_371c5187714088175e3e83183aec6649__Lag);
#>   R_RegisterCCallable("rx_371c5187714088175e3e83183aec6649_","rx_371c5187714088175e3e83183aec6649__Rate", (DL_FUNC) rx_371c5187714088175e3e83183aec6649__Rate);
#>   R_RegisterCCallable("rx_371c5187714088175e3e83183aec6649_","rx_371c5187714088175e3e83183aec6649__Dur", (DL_FUNC) rx_371c5187714088175e3e83183aec6649__Dur);
#>   R_RegisterCCallable("rx_371c5187714088175e3e83183aec6649_","rx_371c5187714088175e3e83183aec6649__mtime", (DL_FUNC) rx_371c5187714088175e3e83183aec6649__mtime);
#>   R_RegisterCCallable("rx_371c5187714088175e3e83183aec6649_","rx_371c5187714088175e3e83183aec6649__ME", (DL_FUNC) rx_371c5187714088175e3e83183aec6649__ME);
#>   R_RegisterCCallable("rx_371c5187714088175e3e83183aec6649_","rx_371c5187714088175e3e83183aec6649__IndF", (DL_FUNC) rx_371c5187714088175e3e83183aec6649__IndF);
#>   R_RegisterCCallable("rx_371c5187714088175e3e83183aec6649_","rx_371c5187714088175e3e83183aec6649__dydt_liblsoda", (DL_FUNC) rx_371c5187714088175e3e83183aec6649__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_rx_371c5187714088175e3e83183aec6649_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_rx_371c5187714088175e3e83183aec6649_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"rx_371c5187714088175e3e83183aec6649__model_vars", (DL_FUNC) &rx_371c5187714088175e3e83183aec6649__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_rx_371c5187714088175e3e83183aec6649_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_371c5187714088175e3e83183aec6649__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("rx_371c5187714088175e3e83183aec6649__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’

e <- et(c(-1, 0, 1, 2))

rxSolve(x, e)
#> Warning: 
#> with negative times, compartments initialize at first negative observed time
#> with positive times, compartments initialize at time zero
#> use 'rxSetIni0(FALSE)' to initialize at first observed time
#> this warning is displayed once per session
#> ── Solved rxode2 object ──
#> ── Parameters (value$params): ──
#> # A tibble: 1 × 0
#> ── Initial Conditions (value$inits): ──
#> named numeric(0)
#> ── First part of data (object): ──
#> # A tibble: 4 × 2
#>    time     r
#>   <dbl> <dbl>
#> 1    -1 -1.11
#> 2     0  0   
#> 3     1  1.05
#> 4     2  2.10