GELU activation function

GELU activation function

Usage

GELU(x)

Arguments

x

numeric vector

Value

numeric vector

See also

Examples

GELU(c(-2, -1, 0, 1, 2))
#> [1] -0.04550026 -0.15865525  0.00000000  0.84134475  1.95449974

# you can use rxode2 as well
r <- rxode2({
  r = GELU(time)
})
#>  
#>  
#> C model:
#> // Define 1 LHS values
#> #define _LHS_0_ 0 // r
#> #define _getRxSolve_ _rx8fa2bca14fb8444085b8182567ef0093143__getRxSolve__UHjl
#> #define _evalUdf _rx8fa2bca14fb8444085b8182567ef0093144__evalUdf_sA2J
#> #define _solveData _rx8fa2bca14fb8444085b8182567ef0093145__solveData_OCP3
#> #define _assign_ptr _rx8fa2bca14fb8444085b8182567ef0093146__assign_ptr_OyHg
#> #define _rxRmModelLib _rx8fa2bca14fb8444085b8182567ef0093147__rxRmModelLib_jCyj
#> #define _rxGetModelLib _rx8fa2bca14fb8444085b8182567ef0093148__rxGetModelLib_3Whg
#> #define _old_c _rx8fa2bca14fb8444085b8182567ef0093149__old_c_1DId
#> #define _ptrid _rx8fa2bca14fb8444085b8182567ef0093150__ptrid_TQhw
#> #define _rxIsCurrentC _rx8fa2bca14fb8444085b8182567ef0093151__rxIsCurrentC_BDgU
#> #define _sumPS _rx8fa2bca14fb8444085b8182567ef0093152__sumPS_de5l
#> #define _prodPS _rx8fa2bca14fb8444085b8182567ef0093153__prodPS_lzyO
#> #define _prodType _rx8fa2bca14fb8444085b8182567ef0093154__prodType_SJAW
#> #define _sumType _rx8fa2bca14fb8444085b8182567ef0093155__sumType_UmCi
#> #define _update_par_ptr _rx8fa2bca14fb8444085b8182567ef0093156__update_par_ptr_4LpG
#> #define _getParCov _rx8fa2bca14fb8444085b8182567ef0093157__getParCov_ZKaB
#> #define _rxode2_rxAssignPtr _rx8fa2bca14fb8444085b8182567ef0093158__rxode2_rxAssignPtr_wXXH
#> #define _rxQr _rx8fa2bca14fb8444085b8182567ef0093159__rxQr_RN1S
#> #define _compareFactorVal _rx8fa2bca14fb8444085b8182567ef0093160__compareFactorVal_O1NY
#> #define _sum _rx8fa2bca14fb8444085b8182567ef0093161__sum_rNWb
#> #define _udf _rx8fa2bca14fb8444085b8182567ef0093162__udf_GHTv
#> #define _sign _rx8fa2bca14fb8444085b8182567ef0093163__sign_ODf3
#> #define _prod _rx8fa2bca14fb8444085b8182567ef0093164__prod_z9Wq
#> #define _max _rx8fa2bca14fb8444085b8182567ef0093165__max_iEXp
#> #define _min _rx8fa2bca14fb8444085b8182567ef0093166__min_1Tno
#> #define _transit4P _rx8fa2bca14fb8444085b8182567ef0093167__transit4P_PDPf
#> #define _transit3P _rx8fa2bca14fb8444085b8182567ef0093168__transit3P_VeVp
#> #define _assignFuns0 _rx8fa2bca14fb8444085b8182567ef0093169__assignFuns0_hs0t
#> #define _assignFuns _rx8fa2bca14fb8444085b8182567ef0093170__assignFuns_zVMm
#> #define _rxord _rx8fa2bca14fb8444085b8182567ef0093171__rxord_adrZ
#> #define __assignFuns2 _rx8fa2bca14fb8444085b8182567ef0093172___assignFuns2_FurF
#> #define _llikCauchyDscale _rx8fa2bca14fb8444085b8182567ef0093173__llikCauchyDscale_Ol7C
#> #define _llikCauchyDlocation _rx8fa2bca14fb8444085b8182567ef0093174__llikCauchyDlocation_Afrv
#> #define _llikCauchy _rx8fa2bca14fb8444085b8182567ef0093175__llikCauchy_YZgF
#> #define _llikGammaDrate _rx8fa2bca14fb8444085b8182567ef0093176__llikGammaDrate_2u4z
#> #define _llikGammaDshape _rx8fa2bca14fb8444085b8182567ef0093177__llikGammaDshape_bAu2
#> #define _llikGamma _rx8fa2bca14fb8444085b8182567ef0093178__llikGamma_aEuo
#> #define _llikWeibullDscale _rx8fa2bca14fb8444085b8182567ef0093179__llikWeibullDscale_hDjD
#> #define _llikWeibullDshape _rx8fa2bca14fb8444085b8182567ef0093180__llikWeibullDshape_SKZF
#> #define _llikWeibull _rx8fa2bca14fb8444085b8182567ef0093181__llikWeibull_Suyo
#> #define _llikUnifDbeta _rx8fa2bca14fb8444085b8182567ef0093182__llikUnifDbeta_talz
#> #define _llikUnifDalpha _rx8fa2bca14fb8444085b8182567ef0093183__llikUnifDalpha_iDIi
#> #define _llikUnif _rx8fa2bca14fb8444085b8182567ef0093184__llikUnif_rE0w
#> #define _llikGeomDp _rx8fa2bca14fb8444085b8182567ef0093185__llikGeomDp_RkfP
#> #define _llikGeom _rx8fa2bca14fb8444085b8182567ef0093186__llikGeom_Ljf5
#> #define _llikFDdf2 _rx8fa2bca14fb8444085b8182567ef0093187__llikFDdf2_r4q6
#> #define _llikFDdf1 _rx8fa2bca14fb8444085b8182567ef0093188__llikFDdf1_UFgA
#> #define _llikF _rx8fa2bca14fb8444085b8182567ef0093189__llikF_xJMt
#> #define _llikExpDrate _rx8fa2bca14fb8444085b8182567ef0093190__llikExpDrate_p6ul
#> #define _llikExp _rx8fa2bca14fb8444085b8182567ef0093191__llikExp_NzLw
#> #define _llikChisqDdf _rx8fa2bca14fb8444085b8182567ef0093192__llikChisqDdf_D19q
#> #define _llikChisq _rx8fa2bca14fb8444085b8182567ef0093193__llikChisq_DD2w
#> #define _llikTDsd _rx8fa2bca14fb8444085b8182567ef0093194__llikTDsd_FmHs
#> #define _llikTDmean _rx8fa2bca14fb8444085b8182567ef0093195__llikTDmean_Iyal
#> #define _llikTDdf _rx8fa2bca14fb8444085b8182567ef0093196__llikTDdf_plP8
#> #define _llikT _rx8fa2bca14fb8444085b8182567ef0093197__llikT_5ebe
#> #define _llikBetaDshape2 _rx8fa2bca14fb8444085b8182567ef0093198__llikBetaDshape2_Kyis
#> #define _llikBetaDshape1 _rx8fa2bca14fb8444085b8182567ef0093199__llikBetaDshape1_yDfO
#> #define _llikBeta _rx8fa2bca14fb8444085b8182567ef0093200__llikBeta_2cbP
#> #define _llikNbinomMuDmu _rx8fa2bca14fb8444085b8182567ef0093201__llikNbinomMuDmu_rWy1
#> #define _llikNbinomMu _rx8fa2bca14fb8444085b8182567ef0093202__llikNbinomMu_fNug
#> #define _llikNbinomDprob _rx8fa2bca14fb8444085b8182567ef0093203__llikNbinomDprob_fkXE
#> #define _llikNbinom _rx8fa2bca14fb8444085b8182567ef0093204__llikNbinom_VPjd
#> #define _llikBinomDprob _rx8fa2bca14fb8444085b8182567ef0093205__llikBinomDprob_OxI3
#> #define _llikBinom _rx8fa2bca14fb8444085b8182567ef0093206__llikBinom_ZD7w
#> #define _llikPoisDlambda _rx8fa2bca14fb8444085b8182567ef0093207__llikPoisDlambda_nkXX
#> #define _llikPois _rx8fa2bca14fb8444085b8182567ef0093208__llikPois_5OAg
#> #define _llikNormDsd _rx8fa2bca14fb8444085b8182567ef0093209__llikNormDsd_fwCy
#> #define _llikNormDmean _rx8fa2bca14fb8444085b8182567ef0093210__llikNormDmean_Neo4
#> #define _llikNorm _rx8fa2bca14fb8444085b8182567ef0093211__llikNorm_WlQY
#> #define simeps _rx8fa2bca14fb8444085b8182567ef0093212_simeps_C1yB
#> #define simeta _rx8fa2bca14fb8444085b8182567ef0093213_simeta_mspb
#> #define expit _rx8fa2bca14fb8444085b8182567ef0093214_expit_u285
#> #define logit _rx8fa2bca14fb8444085b8182567ef0093215_logit_hP47
#> #define gammapDer _rx8fa2bca14fb8444085b8182567ef0093216_gammapDer_wDfC
#> #define lowergamma _rx8fa2bca14fb8444085b8182567ef0093217_lowergamma_Yp7G
#> #define uppergamma _rx8fa2bca14fb8444085b8182567ef0093218_uppergamma_H0qP
#> #define gammaqInva _rx8fa2bca14fb8444085b8182567ef0093219_gammaqInva_kiVy
#> #define gammaqInv _rx8fa2bca14fb8444085b8182567ef0093220_gammaqInv_MAm1
#> #define gammapInva _rx8fa2bca14fb8444085b8182567ef0093221_gammapInva_mA1m
#> #define gammapInv _rx8fa2bca14fb8444085b8182567ef0093222_gammapInv_pxNA
#> #define gammaq _rx8fa2bca14fb8444085b8182567ef0093223_gammaq_bn2j
#> #define gammap _rx8fa2bca14fb8444085b8182567ef0093224_gammap_D07g
#> #define phi _rx8fa2bca14fb8444085b8182567ef0093225_phi_dKOh
#> #define d2ELUa _rx8fa2bca14fb8444085b8182567ef0093226_d2ELUa_hvSa
#> #define dELUa _rx8fa2bca14fb8444085b8182567ef0093227_dELUa_AZta
#> #define d2aELU _rx8fa2bca14fb8444085b8182567ef0093228_d2aELU_pamO
#> #define d2ELU _rx8fa2bca14fb8444085b8182567ef0093229_d2ELU_qjJ6
#> #define dELU _rx8fa2bca14fb8444085b8182567ef0093230_dELU_6nN3
#> #define ELU _rx8fa2bca14fb8444085b8182567ef0093231_ELU_zb9X
#> #define dPReLUa1 _rx8fa2bca14fb8444085b8182567ef0093232_dPReLUa1_7x2A
#> #define dPReLUa _rx8fa2bca14fb8444085b8182567ef0093233_dPReLUa_wyis
#> #define dPReLU _rx8fa2bca14fb8444085b8182567ef0093234_dPReLU_y2Rd
#> #define PReLU _rx8fa2bca14fb8444085b8182567ef0093235_PReLU_zJrX
#> #define dSwish _rx8fa2bca14fb8444085b8182567ef0093236_dSwish_sVrr
#> #define Swish _rx8fa2bca14fb8444085b8182567ef0093237_Swish_L0Cw
#> #define dlReLU _rx8fa2bca14fb8444085b8182567ef0093238_dlReLU_Hsx0
#> #define lReLU _rx8fa2bca14fb8444085b8182567ef0093239_lReLU_YDej
#> #define dSELU _rx8fa2bca14fb8444085b8182567ef0093240_dSELU_sNtq
#> #define SELU _rx8fa2bca14fb8444085b8182567ef0093241_SELU_j2Kk
#> #define d4softplus _rx8fa2bca14fb8444085b8182567ef0093242_d4softplus_uCTN
#> #define d3softplus _rx8fa2bca14fb8444085b8182567ef0093243_d3softplus_OD0M
#> #define d2softplus _rx8fa2bca14fb8444085b8182567ef0093244_d2softplus_Wq8N
#> #define dsoftplus _rx8fa2bca14fb8444085b8182567ef0093245_dsoftplus_GU15
#> #define softplus _rx8fa2bca14fb8444085b8182567ef0093246_softplus_zw77
#> #define d4GELU _rx8fa2bca14fb8444085b8182567ef0093247_d4GELU_MTmU
#> #define d3GELU _rx8fa2bca14fb8444085b8182567ef0093248_d3GELU_6whV
#> #define d2GELU _rx8fa2bca14fb8444085b8182567ef0093249_d2GELU_WGiy
#> #define dGELU _rx8fa2bca14fb8444085b8182567ef0093250_dGELU_en4s
#> #define GELU _rx8fa2bca14fb8444085b8182567ef0093251_GELU_piMW
#> #define dReLU _rx8fa2bca14fb8444085b8182567ef0093252_dReLU_T6VR
#> #define ReLU _rx8fa2bca14fb8444085b8182567ef0093253_ReLU_Ye9I
#> #define riweibull _rx8fa2bca14fb8444085b8182567ef0093254_riweibull_tadG
#> #define riunif _rx8fa2bca14fb8444085b8182567ef0093255_riunif_JLl0
#> #define rit_ _rx8fa2bca14fb8444085b8182567ef0093256_rit__3GaI
#> #define ripois _rx8fa2bca14fb8444085b8182567ef0093257_ripois_5llO
#> #define ribeta _rx8fa2bca14fb8444085b8182567ef0093258_ribeta_3kWH
#> #define rigamma _rx8fa2bca14fb8444085b8182567ef0093259_rigamma_GXCd
#> #define rigeom _rx8fa2bca14fb8444085b8182567ef0093260_rigeom_kGsq
#> #define rif _rx8fa2bca14fb8444085b8182567ef0093261_rif_2NBv
#> #define riexp _rx8fa2bca14fb8444085b8182567ef0093262_riexp_uBr0
#> #define richisq _rx8fa2bca14fb8444085b8182567ef0093263_richisq_sO4e
#> #define ricauchy _rx8fa2bca14fb8444085b8182567ef0093264_ricauchy_upVb
#> #define rinbinomMu _rx8fa2bca14fb8444085b8182567ef0093265_rinbinomMu_TsfO
#> #define rinbinom _rx8fa2bca14fb8444085b8182567ef0093266_rinbinom_YFC4
#> #define ribinom _rx8fa2bca14fb8444085b8182567ef0093267_ribinom_Jz7i
#> #define rinorm _rx8fa2bca14fb8444085b8182567ef0093268_rinorm_E8vt
#> #define rxweibull _rx8fa2bca14fb8444085b8182567ef0093269_rxweibull_CFcR
#> #define rxunif _rx8fa2bca14fb8444085b8182567ef0093270_rxunif_J6Ge
#> #define rxt_ _rx8fa2bca14fb8444085b8182567ef0093271_rxt__v1Yq
#> #define rxpois _rx8fa2bca14fb8444085b8182567ef0093272_rxpois_4CUG
#> #define rxbeta _rx8fa2bca14fb8444085b8182567ef0093273_rxbeta_PRB4
#> #define rxgamma _rx8fa2bca14fb8444085b8182567ef0093274_rxgamma_1jyX
#> #define rxgeom _rx8fa2bca14fb8444085b8182567ef0093275_rxgeom_A3Sc
#> #define rxf _rx8fa2bca14fb8444085b8182567ef0093276_rxf_akWU
#> #define rxexp _rx8fa2bca14fb8444085b8182567ef0093277_rxexp_0xSy
#> #define rxchisq _rx8fa2bca14fb8444085b8182567ef0093278_rxchisq_XMV4
#> #define rxcauchy _rx8fa2bca14fb8444085b8182567ef0093279_rxcauchy_yezX
#> #define rxnbinomMu _rx8fa2bca14fb8444085b8182567ef0093280_rxnbinomMu_lYmd
#> #define rxnbinom _rx8fa2bca14fb8444085b8182567ef0093281_rxnbinom_NMqS
#> #define rxbinom _rx8fa2bca14fb8444085b8182567ef0093282_rxbinom_Bxi1
#> #define rxnorm _rx8fa2bca14fb8444085b8182567ef0093283_rxnorm_cpdh
#> #define linCmtB _rx8fa2bca14fb8444085b8182567ef0093284_linCmtB_rHA6
#> #define linCmtA _rx8fa2bca14fb8444085b8182567ef0093285_linCmtA_aqvf
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){}
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){}
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  rx_c457ff7accb95e63435d1139ab816b59__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *rx_c457ff7accb95e63435d1139ab816b59__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP rx_c457ff7accb95e63435d1139ab816b59__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void rx_c457ff7accb95e63435d1139ab816b59__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double r = NA_REAL;
#> 
#>   (void)t;
#>   (void)r;
#> 
#>   r = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   r=GELU(t);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void rx_c457ff7accb95e63435d1139ab816b59__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void rx_c457ff7accb95e63435d1139ab816b59__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void rx_c457ff7accb95e63435d1139ab816b59__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double r = NA_REAL;
#> 
#>   (void)t;
#>   (void)r;
#> 
#>   r = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   r=GELU(t);
#> 
#>   _lhs[_LHS_0_]=r;
#> }
#> // Functional based bioavailability
#> double rx_c457ff7accb95e63435d1139ab816b59__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double rx_c457ff7accb95e63435d1139ab816b59__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double rx_c457ff7accb95e63435d1139ab816b59__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double rx_c457ff7accb95e63435d1139ab816b59__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void rx_c457ff7accb95e63435d1139ab816b59__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void rx_c457ff7accb95e63435d1139ab816b59__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void rx_c457ff7accb95e63435d1139ab816b59__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP rx_c457ff7accb95e63435d1139ab816b59__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_c457ff7accb95e63435d1139ab816b59__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAA+>zAAAv7#aT)cmC:MAg)<+h\";*Tu1)=J=qaErnbmwo8#_@@I[MJV\"8uYi85^0+tTd6\?r1{zuBS^mj\?:Lk;BH3TAP;`G:<hPVcQ$MXL5FBAFt}#nt!\?x6C2|G8n\"yT#V2J/7`g*oKmr]W|s;&}*\?55@(iL{puY_k%%XWl205h2rYGcps3<k}gHV1z\?}~{p;6o6z\?x9]e^z,Q2LC/>,*5S0,3R_|6\?/j:Bhm<*Vpap0R.`,GE>*}zsy`8<kojU/\"Mh);nZG[TjH=vs&mmQW,X(V!l`&XWfjNh0+&!%%^#htVWIf.}5O!Q>GEbEePhmc):66k_tTu@_v65lpwgF{m;cXDyKIqJFF<7;<IqxjKO5:\"t<ISxm~E1;XGmSuW|LFIBmWU;vnChx.RRRbXkr._G{8T(kuqB1JMRYd)Gh/GI$[^00]exRZI+eJeQ:x{VghG!YJFnLLLpKqVeON^ve`2S<^N!`\"OZ_R~;;)9]GBDl1[Xb8>=U(*D6`(04Uj/GZ}8QS/!6&qjz`oTdhH&R<.eB1|RLMM`2l|[WG,{w]WG1F}\"a<MqX/5e$yR_0$zX;0}z|XuL1w@WLHby`o8Y[A8~I\"(J,N^p}eP{|t1p<BZquH|^bz!2qkH8Frx|R|:nZ16\?8~Hd%%{XUhMH+k|WF9#{+N}46`5w<q@wKo$r~=N_>d00%%)8]zN$bc\?|AV%%<y3!\?E_@JS|z8Sib`2Fy//n~qZ8gv^D[]M64geiGPkVO=$PL$#y3:{!CK|RTz3WFksSxoPDM01cSF1B&6~z8I)MlR|x(T\?J{xbvmvoS9*B/:44AfZI>(~W9}l*zz|$EIURFAd_aQwJbRet|q/xh]36CIh,/P(AmjCP=>o^r{P]&njY\?!CIWt`9^3k7,X)0ZGIIB<nNVsQy@xF)j8<xZ%%!_i.1ytq[B}lY|;;Co=.|71\"v&wLXQBdzMoj4fd[EFN(I|`d)Z.SFSn;0J}{7\?^,0ic}Km#\?c]=w|VqkXq:cg)nM76lJ7}ba@xYdQ2lvGFiSNrG7pwjfZ[,f7]9r=PC");
#>     char buf[1061];
#> #define __doBufN__ 1061
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void rx_c457ff7accb95e63435d1139ab816b59__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   rx_c457ff7accb95e63435d1139ab816b59__dydt(neq, *t, A, DADT);
#> }
#> extern int rx_c457ff7accb95e63435d1139ab816b59__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   rx_c457ff7accb95e63435d1139ab816b59__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void rx_c457ff7accb95e63435d1139ab816b59__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   rx_c457ff7accb95e63435d1139ab816b59__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void rx_c457ff7accb95e63435d1139ab816b59__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_rx_c457ff7accb95e63435d1139ab816b59_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("rx_c457ff7accb95e63435d1139ab816b59_","rx_c457ff7accb95e63435d1139ab816b59__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("rx_c457ff7accb95e63435d1139ab816b59_","rx_c457ff7accb95e63435d1139ab816b59__assignFuns", (DL_FUNC) rx_c457ff7accb95e63435d1139ab816b59__assignFuns);
#>   R_RegisterCCallable("rx_c457ff7accb95e63435d1139ab816b59_","rx_c457ff7accb95e63435d1139ab816b59__inis",(DL_FUNC) rx_c457ff7accb95e63435d1139ab816b59__inis);
#>   R_RegisterCCallable("rx_c457ff7accb95e63435d1139ab816b59_","rx_c457ff7accb95e63435d1139ab816b59__dydt",(DL_FUNC) rx_c457ff7accb95e63435d1139ab816b59__dydt);
#>   R_RegisterCCallable("rx_c457ff7accb95e63435d1139ab816b59_","rx_c457ff7accb95e63435d1139ab816b59__calc_lhs",(DL_FUNC) rx_c457ff7accb95e63435d1139ab816b59__calc_lhs);
#>   R_RegisterCCallable("rx_c457ff7accb95e63435d1139ab816b59_","rx_c457ff7accb95e63435d1139ab816b59__calc_jac",(DL_FUNC) rx_c457ff7accb95e63435d1139ab816b59__calc_jac);
#>   R_RegisterCCallable("rx_c457ff7accb95e63435d1139ab816b59_","rx_c457ff7accb95e63435d1139ab816b59__dydt_lsoda", (DL_FUNC) rx_c457ff7accb95e63435d1139ab816b59__dydt_lsoda);
#>   R_RegisterCCallable("rx_c457ff7accb95e63435d1139ab816b59_","rx_c457ff7accb95e63435d1139ab816b59__calc_jac_lsoda", (DL_FUNC) rx_c457ff7accb95e63435d1139ab816b59__calc_jac_lsoda);
#>   R_RegisterCCallable("rx_c457ff7accb95e63435d1139ab816b59_","rx_c457ff7accb95e63435d1139ab816b59__ode_solver_solvedata", (DL_FUNC) rx_c457ff7accb95e63435d1139ab816b59__ode_solver_solvedata);
#>   R_RegisterCCallable("rx_c457ff7accb95e63435d1139ab816b59_","rx_c457ff7accb95e63435d1139ab816b59__ode_solver_get_solvedata", (DL_FUNC) rx_c457ff7accb95e63435d1139ab816b59__ode_solver_get_solvedata);
#>   R_RegisterCCallable("rx_c457ff7accb95e63435d1139ab816b59_","rx_c457ff7accb95e63435d1139ab816b59__F", (DL_FUNC) rx_c457ff7accb95e63435d1139ab816b59__F);
#>   R_RegisterCCallable("rx_c457ff7accb95e63435d1139ab816b59_","rx_c457ff7accb95e63435d1139ab816b59__Lag", (DL_FUNC) rx_c457ff7accb95e63435d1139ab816b59__Lag);
#>   R_RegisterCCallable("rx_c457ff7accb95e63435d1139ab816b59_","rx_c457ff7accb95e63435d1139ab816b59__Rate", (DL_FUNC) rx_c457ff7accb95e63435d1139ab816b59__Rate);
#>   R_RegisterCCallable("rx_c457ff7accb95e63435d1139ab816b59_","rx_c457ff7accb95e63435d1139ab816b59__Dur", (DL_FUNC) rx_c457ff7accb95e63435d1139ab816b59__Dur);
#>   R_RegisterCCallable("rx_c457ff7accb95e63435d1139ab816b59_","rx_c457ff7accb95e63435d1139ab816b59__mtime", (DL_FUNC) rx_c457ff7accb95e63435d1139ab816b59__mtime);
#>   R_RegisterCCallable("rx_c457ff7accb95e63435d1139ab816b59_","rx_c457ff7accb95e63435d1139ab816b59__ME", (DL_FUNC) rx_c457ff7accb95e63435d1139ab816b59__ME);
#>   R_RegisterCCallable("rx_c457ff7accb95e63435d1139ab816b59_","rx_c457ff7accb95e63435d1139ab816b59__IndF", (DL_FUNC) rx_c457ff7accb95e63435d1139ab816b59__IndF);
#>   R_RegisterCCallable("rx_c457ff7accb95e63435d1139ab816b59_","rx_c457ff7accb95e63435d1139ab816b59__dydt_liblsoda", (DL_FUNC) rx_c457ff7accb95e63435d1139ab816b59__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_rx_c457ff7accb95e63435d1139ab816b59_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_rx_c457ff7accb95e63435d1139ab816b59_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"rx_c457ff7accb95e63435d1139ab816b59__model_vars", (DL_FUNC) &rx_c457ff7accb95e63435d1139ab816b59__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_rx_c457ff7accb95e63435d1139ab816b59_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_c457ff7accb95e63435d1139ab816b59__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("rx_c457ff7accb95e63435d1139ab816b59__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’
et <- et(c(-2, -1, 0, 1, 2))
rxSolve(r, et)
#> ── Solved rxode2 object ──
#> ── Parameters (value$params): ──
#> # A tibble: 1 × 0
#> ── Initial Conditions (value$inits): ──
#> named numeric(0)
#> ── First part of data (object): ──
#> # A tibble: 5 × 2
#>    time       r
#>   <dbl>   <dbl>
#> 1    -2 -0.0455
#> 2    -1 -0.159 
#> 3     0  0     
#> 4     1  0.841 
#> 5     2  1.95