Parametric ReLU Activation Function

Parametric ReLU Activation Function

Usage

PReLU(x, alpha = 1)

Arguments

x

A numeric vector. All elements must be finite and non-missing.

alpha

A numeric scalar. All elements must be finite and non-missing.

Value

A numeric vector where the ReLU function has been applied to each element of x.

See also

Author

Matthew Fidler

Examples

PReLU(c(-1, 0, 1, 2), 2)
#> [1] -2  0  1  2

# Can also be used in rxode2:
x <- rxode2({
   r=PReLU(time, 2)
})
#>  
#>  
#> C model:
#> // Define 1 LHS values
#> #define _LHS_0_ 0 // r
#> #define _getRxSolve_ _rxc9b7d4b5325bb8c123cb7d10c4d558b6286__getRxSolve__sA2J
#> #define _evalUdf _rxc9b7d4b5325bb8c123cb7d10c4d558b6287__evalUdf_OCP3
#> #define _solveData _rxc9b7d4b5325bb8c123cb7d10c4d558b6288__solveData_OyHg
#> #define _assign_ptr _rxc9b7d4b5325bb8c123cb7d10c4d558b6289__assign_ptr_jCyj
#> #define _rxRmModelLib _rxc9b7d4b5325bb8c123cb7d10c4d558b6290__rxRmModelLib_3Whg
#> #define _rxGetModelLib _rxc9b7d4b5325bb8c123cb7d10c4d558b6291__rxGetModelLib_1DId
#> #define _old_c _rxc9b7d4b5325bb8c123cb7d10c4d558b6292__old_c_TQhw
#> #define _ptrid _rxc9b7d4b5325bb8c123cb7d10c4d558b6293__ptrid_BDgU
#> #define _rxIsCurrentC _rxc9b7d4b5325bb8c123cb7d10c4d558b6294__rxIsCurrentC_de5l
#> #define _sumPS _rxc9b7d4b5325bb8c123cb7d10c4d558b6295__sumPS_lzyO
#> #define _prodPS _rxc9b7d4b5325bb8c123cb7d10c4d558b6296__prodPS_SJAW
#> #define _prodType _rxc9b7d4b5325bb8c123cb7d10c4d558b6297__prodType_UmCi
#> #define _sumType _rxc9b7d4b5325bb8c123cb7d10c4d558b6298__sumType_4LpG
#> #define _update_par_ptr _rxc9b7d4b5325bb8c123cb7d10c4d558b6299__update_par_ptr_ZKaB
#> #define _getParCov _rxc9b7d4b5325bb8c123cb7d10c4d558b6300__getParCov_wXXH
#> #define _rxode2_rxAssignPtr _rxc9b7d4b5325bb8c123cb7d10c4d558b6301__rxode2_rxAssignPtr_RN1S
#> #define _rxQr _rxc9b7d4b5325bb8c123cb7d10c4d558b6302__rxQr_O1NY
#> #define _compareFactorVal _rxc9b7d4b5325bb8c123cb7d10c4d558b6303__compareFactorVal_rNWb
#> #define _sum _rxc9b7d4b5325bb8c123cb7d10c4d558b6304__sum_GHTv
#> #define _udf _rxc9b7d4b5325bb8c123cb7d10c4d558b6305__udf_ODf3
#> #define _sign _rxc9b7d4b5325bb8c123cb7d10c4d558b6306__sign_z9Wq
#> #define _prod _rxc9b7d4b5325bb8c123cb7d10c4d558b6307__prod_iEXp
#> #define _max _rxc9b7d4b5325bb8c123cb7d10c4d558b6308__max_1Tno
#> #define _min _rxc9b7d4b5325bb8c123cb7d10c4d558b6309__min_PDPf
#> #define _transit4P _rxc9b7d4b5325bb8c123cb7d10c4d558b6310__transit4P_VeVp
#> #define _transit3P _rxc9b7d4b5325bb8c123cb7d10c4d558b6311__transit3P_hs0t
#> #define _assignFuns0 _rxc9b7d4b5325bb8c123cb7d10c4d558b6312__assignFuns0_zVMm
#> #define _assignFuns _rxc9b7d4b5325bb8c123cb7d10c4d558b6313__assignFuns_adrZ
#> #define _rxord _rxc9b7d4b5325bb8c123cb7d10c4d558b6314__rxord_FurF
#> #define __assignFuns2 _rxc9b7d4b5325bb8c123cb7d10c4d558b6315___assignFuns2_Ol7C
#> #define _llikCauchyDscale _rxc9b7d4b5325bb8c123cb7d10c4d558b6316__llikCauchyDscale_Afrv
#> #define _llikCauchyDlocation _rxc9b7d4b5325bb8c123cb7d10c4d558b6317__llikCauchyDlocation_YZgF
#> #define _llikCauchy _rxc9b7d4b5325bb8c123cb7d10c4d558b6318__llikCauchy_2u4z
#> #define _llikGammaDrate _rxc9b7d4b5325bb8c123cb7d10c4d558b6319__llikGammaDrate_bAu2
#> #define _llikGammaDshape _rxc9b7d4b5325bb8c123cb7d10c4d558b6320__llikGammaDshape_aEuo
#> #define _llikGamma _rxc9b7d4b5325bb8c123cb7d10c4d558b6321__llikGamma_hDjD
#> #define _llikWeibullDscale _rxc9b7d4b5325bb8c123cb7d10c4d558b6322__llikWeibullDscale_SKZF
#> #define _llikWeibullDshape _rxc9b7d4b5325bb8c123cb7d10c4d558b6323__llikWeibullDshape_Suyo
#> #define _llikWeibull _rxc9b7d4b5325bb8c123cb7d10c4d558b6324__llikWeibull_talz
#> #define _llikUnifDbeta _rxc9b7d4b5325bb8c123cb7d10c4d558b6325__llikUnifDbeta_iDIi
#> #define _llikUnifDalpha _rxc9b7d4b5325bb8c123cb7d10c4d558b6326__llikUnifDalpha_rE0w
#> #define _llikUnif _rxc9b7d4b5325bb8c123cb7d10c4d558b6327__llikUnif_RkfP
#> #define _llikGeomDp _rxc9b7d4b5325bb8c123cb7d10c4d558b6328__llikGeomDp_Ljf5
#> #define _llikGeom _rxc9b7d4b5325bb8c123cb7d10c4d558b6329__llikGeom_r4q6
#> #define _llikFDdf2 _rxc9b7d4b5325bb8c123cb7d10c4d558b6330__llikFDdf2_UFgA
#> #define _llikFDdf1 _rxc9b7d4b5325bb8c123cb7d10c4d558b6331__llikFDdf1_xJMt
#> #define _llikF _rxc9b7d4b5325bb8c123cb7d10c4d558b6332__llikF_p6ul
#> #define _llikExpDrate _rxc9b7d4b5325bb8c123cb7d10c4d558b6333__llikExpDrate_NzLw
#> #define _llikExp _rxc9b7d4b5325bb8c123cb7d10c4d558b6334__llikExp_D19q
#> #define _llikChisqDdf _rxc9b7d4b5325bb8c123cb7d10c4d558b6335__llikChisqDdf_DD2w
#> #define _llikChisq _rxc9b7d4b5325bb8c123cb7d10c4d558b6336__llikChisq_FmHs
#> #define _llikTDsd _rxc9b7d4b5325bb8c123cb7d10c4d558b6337__llikTDsd_Iyal
#> #define _llikTDmean _rxc9b7d4b5325bb8c123cb7d10c4d558b6338__llikTDmean_plP8
#> #define _llikTDdf _rxc9b7d4b5325bb8c123cb7d10c4d558b6339__llikTDdf_5ebe
#> #define _llikT _rxc9b7d4b5325bb8c123cb7d10c4d558b6340__llikT_Kyis
#> #define _llikBetaDshape2 _rxc9b7d4b5325bb8c123cb7d10c4d558b6341__llikBetaDshape2_yDfO
#> #define _llikBetaDshape1 _rxc9b7d4b5325bb8c123cb7d10c4d558b6342__llikBetaDshape1_2cbP
#> #define _llikBeta _rxc9b7d4b5325bb8c123cb7d10c4d558b6343__llikBeta_rWy1
#> #define _llikNbinomMuDmu _rxc9b7d4b5325bb8c123cb7d10c4d558b6344__llikNbinomMuDmu_fNug
#> #define _llikNbinomMu _rxc9b7d4b5325bb8c123cb7d10c4d558b6345__llikNbinomMu_fkXE
#> #define _llikNbinomDprob _rxc9b7d4b5325bb8c123cb7d10c4d558b6346__llikNbinomDprob_VPjd
#> #define _llikNbinom _rxc9b7d4b5325bb8c123cb7d10c4d558b6347__llikNbinom_OxI3
#> #define _llikBinomDprob _rxc9b7d4b5325bb8c123cb7d10c4d558b6348__llikBinomDprob_ZD7w
#> #define _llikBinom _rxc9b7d4b5325bb8c123cb7d10c4d558b6349__llikBinom_nkXX
#> #define _llikPoisDlambda _rxc9b7d4b5325bb8c123cb7d10c4d558b6350__llikPoisDlambda_5OAg
#> #define _llikPois _rxc9b7d4b5325bb8c123cb7d10c4d558b6351__llikPois_fwCy
#> #define _llikNormDsd _rxc9b7d4b5325bb8c123cb7d10c4d558b6352__llikNormDsd_Neo4
#> #define _llikNormDmean _rxc9b7d4b5325bb8c123cb7d10c4d558b6353__llikNormDmean_WlQY
#> #define _llikNorm _rxc9b7d4b5325bb8c123cb7d10c4d558b6354__llikNorm_C1yB
#> #define simeps _rxc9b7d4b5325bb8c123cb7d10c4d558b6355_simeps_mspb
#> #define simeta _rxc9b7d4b5325bb8c123cb7d10c4d558b6356_simeta_u285
#> #define expit _rxc9b7d4b5325bb8c123cb7d10c4d558b6357_expit_hP47
#> #define logit _rxc9b7d4b5325bb8c123cb7d10c4d558b6358_logit_wDfC
#> #define gammapDer _rxc9b7d4b5325bb8c123cb7d10c4d558b6359_gammapDer_Yp7G
#> #define lowergamma _rxc9b7d4b5325bb8c123cb7d10c4d558b6360_lowergamma_H0qP
#> #define uppergamma _rxc9b7d4b5325bb8c123cb7d10c4d558b6361_uppergamma_kiVy
#> #define gammaqInva _rxc9b7d4b5325bb8c123cb7d10c4d558b6362_gammaqInva_MAm1
#> #define gammaqInv _rxc9b7d4b5325bb8c123cb7d10c4d558b6363_gammaqInv_mA1m
#> #define gammapInva _rxc9b7d4b5325bb8c123cb7d10c4d558b6364_gammapInva_pxNA
#> #define gammapInv _rxc9b7d4b5325bb8c123cb7d10c4d558b6365_gammapInv_bn2j
#> #define gammaq _rxc9b7d4b5325bb8c123cb7d10c4d558b6366_gammaq_D07g
#> #define gammap _rxc9b7d4b5325bb8c123cb7d10c4d558b6367_gammap_dKOh
#> #define phi _rxc9b7d4b5325bb8c123cb7d10c4d558b6368_phi_hvSa
#> #define d2ELUa _rxc9b7d4b5325bb8c123cb7d10c4d558b6369_d2ELUa_AZta
#> #define dELUa _rxc9b7d4b5325bb8c123cb7d10c4d558b6370_dELUa_pamO
#> #define d2aELU _rxc9b7d4b5325bb8c123cb7d10c4d558b6371_d2aELU_qjJ6
#> #define d2ELU _rxc9b7d4b5325bb8c123cb7d10c4d558b6372_d2ELU_6nN3
#> #define dELU _rxc9b7d4b5325bb8c123cb7d10c4d558b6373_dELU_zb9X
#> #define ELU _rxc9b7d4b5325bb8c123cb7d10c4d558b6374_ELU_7x2A
#> #define dPReLUa1 _rxc9b7d4b5325bb8c123cb7d10c4d558b6375_dPReLUa1_wyis
#> #define dPReLUa _rxc9b7d4b5325bb8c123cb7d10c4d558b6376_dPReLUa_y2Rd
#> #define dPReLU _rxc9b7d4b5325bb8c123cb7d10c4d558b6377_dPReLU_zJrX
#> #define PReLU _rxc9b7d4b5325bb8c123cb7d10c4d558b6378_PReLU_sVrr
#> #define dSwish _rxc9b7d4b5325bb8c123cb7d10c4d558b6379_dSwish_L0Cw
#> #define Swish _rxc9b7d4b5325bb8c123cb7d10c4d558b6380_Swish_Hsx0
#> #define dlReLU _rxc9b7d4b5325bb8c123cb7d10c4d558b6381_dlReLU_YDej
#> #define lReLU _rxc9b7d4b5325bb8c123cb7d10c4d558b6382_lReLU_sNtq
#> #define dSELU _rxc9b7d4b5325bb8c123cb7d10c4d558b6383_dSELU_j2Kk
#> #define SELU _rxc9b7d4b5325bb8c123cb7d10c4d558b6384_SELU_uCTN
#> #define d4softplus _rxc9b7d4b5325bb8c123cb7d10c4d558b6385_d4softplus_OD0M
#> #define d3softplus _rxc9b7d4b5325bb8c123cb7d10c4d558b6386_d3softplus_Wq8N
#> #define d2softplus _rxc9b7d4b5325bb8c123cb7d10c4d558b6387_d2softplus_GU15
#> #define dsoftplus _rxc9b7d4b5325bb8c123cb7d10c4d558b6388_dsoftplus_zw77
#> #define softplus _rxc9b7d4b5325bb8c123cb7d10c4d558b6389_softplus_MTmU
#> #define d4GELU _rxc9b7d4b5325bb8c123cb7d10c4d558b6390_d4GELU_6whV
#> #define d3GELU _rxc9b7d4b5325bb8c123cb7d10c4d558b6391_d3GELU_WGiy
#> #define d2GELU _rxc9b7d4b5325bb8c123cb7d10c4d558b6392_d2GELU_en4s
#> #define dGELU _rxc9b7d4b5325bb8c123cb7d10c4d558b6393_dGELU_piMW
#> #define GELU _rxc9b7d4b5325bb8c123cb7d10c4d558b6394_GELU_T6VR
#> #define dReLU _rxc9b7d4b5325bb8c123cb7d10c4d558b6395_dReLU_Ye9I
#> #define ReLU _rxc9b7d4b5325bb8c123cb7d10c4d558b6396_ReLU_tadG
#> #define riweibull _rxc9b7d4b5325bb8c123cb7d10c4d558b6397_riweibull_JLl0
#> #define riunif _rxc9b7d4b5325bb8c123cb7d10c4d558b6398_riunif_3GaI
#> #define rit_ _rxc9b7d4b5325bb8c123cb7d10c4d558b6399_rit__5llO
#> #define ripois _rxc9b7d4b5325bb8c123cb7d10c4d558b6400_ripois_3kWH
#> #define ribeta _rxc9b7d4b5325bb8c123cb7d10c4d558b6401_ribeta_GXCd
#> #define rigamma _rxc9b7d4b5325bb8c123cb7d10c4d558b6402_rigamma_kGsq
#> #define rigeom _rxc9b7d4b5325bb8c123cb7d10c4d558b6403_rigeom_2NBv
#> #define rif _rxc9b7d4b5325bb8c123cb7d10c4d558b6404_rif_uBr0
#> #define riexp _rxc9b7d4b5325bb8c123cb7d10c4d558b6405_riexp_sO4e
#> #define richisq _rxc9b7d4b5325bb8c123cb7d10c4d558b6406_richisq_upVb
#> #define ricauchy _rxc9b7d4b5325bb8c123cb7d10c4d558b6407_ricauchy_TsfO
#> #define rinbinomMu _rxc9b7d4b5325bb8c123cb7d10c4d558b6408_rinbinomMu_YFC4
#> #define rinbinom _rxc9b7d4b5325bb8c123cb7d10c4d558b6409_rinbinom_Jz7i
#> #define ribinom _rxc9b7d4b5325bb8c123cb7d10c4d558b6410_ribinom_E8vt
#> #define rinorm _rxc9b7d4b5325bb8c123cb7d10c4d558b6411_rinorm_CFcR
#> #define rxweibull _rxc9b7d4b5325bb8c123cb7d10c4d558b6412_rxweibull_J6Ge
#> #define rxunif _rxc9b7d4b5325bb8c123cb7d10c4d558b6413_rxunif_v1Yq
#> #define rxt_ _rxc9b7d4b5325bb8c123cb7d10c4d558b6414_rxt__4CUG
#> #define rxpois _rxc9b7d4b5325bb8c123cb7d10c4d558b6415_rxpois_PRB4
#> #define rxbeta _rxc9b7d4b5325bb8c123cb7d10c4d558b6416_rxbeta_1jyX
#> #define rxgamma _rxc9b7d4b5325bb8c123cb7d10c4d558b6417_rxgamma_A3Sc
#> #define rxgeom _rxc9b7d4b5325bb8c123cb7d10c4d558b6418_rxgeom_akWU
#> #define rxf _rxc9b7d4b5325bb8c123cb7d10c4d558b6419_rxf_0xSy
#> #define rxexp _rxc9b7d4b5325bb8c123cb7d10c4d558b6420_rxexp_XMV4
#> #define rxchisq _rxc9b7d4b5325bb8c123cb7d10c4d558b6421_rxchisq_yezX
#> #define rxcauchy _rxc9b7d4b5325bb8c123cb7d10c4d558b6422_rxcauchy_lYmd
#> #define rxnbinomMu _rxc9b7d4b5325bb8c123cb7d10c4d558b6423_rxnbinomMu_NMqS
#> #define rxnbinom _rxc9b7d4b5325bb8c123cb7d10c4d558b6424_rxnbinom_Bxi1
#> #define rxbinom _rxc9b7d4b5325bb8c123cb7d10c4d558b6425_rxbinom_cpdh
#> #define rxnorm _rxc9b7d4b5325bb8c123cb7d10c4d558b6426_rxnorm_rHA6
#> #define linCmtB _rxc9b7d4b5325bb8c123cb7d10c4d558b6427_linCmtB_aqvf
#> #define linCmtA _rxc9b7d4b5325bb8c123cb7d10c4d558b6428_linCmtA_L0Qi
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){}
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){}
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  rx_cd0af269d6c39e55c39293e7debf7ee2__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *rx_cd0af269d6c39e55c39293e7debf7ee2__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP rx_cd0af269d6c39e55c39293e7debf7ee2__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void rx_cd0af269d6c39e55c39293e7debf7ee2__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double r = NA_REAL;
#> 
#>   (void)t;
#>   (void)r;
#> 
#>   r = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   r=PReLU(t,2);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void rx_cd0af269d6c39e55c39293e7debf7ee2__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void rx_cd0af269d6c39e55c39293e7debf7ee2__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void rx_cd0af269d6c39e55c39293e7debf7ee2__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double r = NA_REAL;
#> 
#>   (void)t;
#>   (void)r;
#> 
#>   r = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   r=PReLU(t,2);
#> 
#>   _lhs[_LHS_0_]=r;
#> }
#> // Functional based bioavailability
#> double rx_cd0af269d6c39e55c39293e7debf7ee2__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double rx_cd0af269d6c39e55c39293e7debf7ee2__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double rx_cd0af269d6c39e55c39293e7debf7ee2__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double rx_cd0af269d6c39e55c39293e7debf7ee2__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void rx_cd0af269d6c39e55c39293e7debf7ee2__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void rx_cd0af269d6c39e55c39293e7debf7ee2__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void rx_cd0af269d6c39e55c39293e7debf7ee2__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP rx_cd0af269d6c39e55c39293e7debf7ee2__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_cd0af269d6c39e55c39293e7debf7ee2__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAAN/zAAAv7#aT)im4.MAl*$+<WGlaB~1u%%RS@ZeShROGD%%Ab=8)iON{ZX01X`/OM/qbxgi\"yaB1DjNl]A\"cnu<vuhr$aN]B2zou#+CIY:C,h}5TT8sNg[Iq31B3KG!+NE=3i5~I27Q{Dj%%34)l{WAb96=}xEa]:{f_ss_<q(UI&{4142R#5>FyU)}=/Y8~XaKi$dXN,zEQHA/U)T!yVDK$X0+)`4R}Kv/{%%,i,7j^MF^^g`HDkq*{N17P&p41IL>xP2Kzvq:XRcphB!TP~:_Pq=(EUaNKzjju#NoKTUvO[y,E6qftf~a8qdLP_z)bw&JkeVPG~Q8\"J%%q`zbXoI0.rJ88/ztHxI)m0wAGn#gJq3ZOO_{DDCDAH\"~tq6>W\?sqCpn;ghx]WT\"*5z.%%TPp&x<1QCRdsH*M+t/c2c{+=Q`ZXN)c.,FvZ8.Q@RTCJxWh`QwYZ&Ifs);nM$KLA4,:5|[Ys@UJ;{bo7=SLatY(`6]^&P8=^Dt=Fd^*CTLwB%%p$zN}c58I_t{_6S0e3pd/8R=F<rUuEJ%%a)A:O%%.\"h>K!y7r~@gtOj3$ciXQ@Dl|k%%`O/o[Ev>b7+@Y*`rD:&MDy9T|8}4\"V|%%UQ@8eXrldj=gETq7<1bP3AZ56MD45WG*;H3|3vqUwQoFYU[|bl)R,8h9}I1iY_F3}+p2H`9N@LHP3V2o:QD/{[0j%%p!D^1U/leTow7}UnYxp`p^:N3wf++EXE/e*Nn|z2;jNMt<uYicm,Ko;,1x<AT2]t\"kOA3W{aU,PtG\"t]`cjL%%(HDlGLDzBaX%%mnU^FV::,7\?6y{/e$\?9Z\?\"H<FVn,WH\?%%_FN!BYOHpO63>J.s{sDERc!3EHVv;:(m*QyMl3\"QEf;obx:;/G|EL8)9K<&W2)IgZLS!\"HwH[(:E\"gS{0D;HD*YDg<Z$>vBsK5VA_&Qts6:E*3!;h=Meg`[|dkLQ#0nGq[JD3%%_;pvH]~(D%%UFlwj0!CSnHj.fWa1`CT0aPV96SU.6Sg6O>e7>/$P0is8]5H+pQ{a!+~I9S<s*RuB");
#>     char buf[1070];
#> #define __doBufN__ 1070
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void rx_cd0af269d6c39e55c39293e7debf7ee2__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   rx_cd0af269d6c39e55c39293e7debf7ee2__dydt(neq, *t, A, DADT);
#> }
#> extern int rx_cd0af269d6c39e55c39293e7debf7ee2__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   rx_cd0af269d6c39e55c39293e7debf7ee2__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void rx_cd0af269d6c39e55c39293e7debf7ee2__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   rx_cd0af269d6c39e55c39293e7debf7ee2__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void rx_cd0af269d6c39e55c39293e7debf7ee2__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_rx_cd0af269d6c39e55c39293e7debf7ee2_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("rx_cd0af269d6c39e55c39293e7debf7ee2_","rx_cd0af269d6c39e55c39293e7debf7ee2__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("rx_cd0af269d6c39e55c39293e7debf7ee2_","rx_cd0af269d6c39e55c39293e7debf7ee2__assignFuns", (DL_FUNC) rx_cd0af269d6c39e55c39293e7debf7ee2__assignFuns);
#>   R_RegisterCCallable("rx_cd0af269d6c39e55c39293e7debf7ee2_","rx_cd0af269d6c39e55c39293e7debf7ee2__inis",(DL_FUNC) rx_cd0af269d6c39e55c39293e7debf7ee2__inis);
#>   R_RegisterCCallable("rx_cd0af269d6c39e55c39293e7debf7ee2_","rx_cd0af269d6c39e55c39293e7debf7ee2__dydt",(DL_FUNC) rx_cd0af269d6c39e55c39293e7debf7ee2__dydt);
#>   R_RegisterCCallable("rx_cd0af269d6c39e55c39293e7debf7ee2_","rx_cd0af269d6c39e55c39293e7debf7ee2__calc_lhs",(DL_FUNC) rx_cd0af269d6c39e55c39293e7debf7ee2__calc_lhs);
#>   R_RegisterCCallable("rx_cd0af269d6c39e55c39293e7debf7ee2_","rx_cd0af269d6c39e55c39293e7debf7ee2__calc_jac",(DL_FUNC) rx_cd0af269d6c39e55c39293e7debf7ee2__calc_jac);
#>   R_RegisterCCallable("rx_cd0af269d6c39e55c39293e7debf7ee2_","rx_cd0af269d6c39e55c39293e7debf7ee2__dydt_lsoda", (DL_FUNC) rx_cd0af269d6c39e55c39293e7debf7ee2__dydt_lsoda);
#>   R_RegisterCCallable("rx_cd0af269d6c39e55c39293e7debf7ee2_","rx_cd0af269d6c39e55c39293e7debf7ee2__calc_jac_lsoda", (DL_FUNC) rx_cd0af269d6c39e55c39293e7debf7ee2__calc_jac_lsoda);
#>   R_RegisterCCallable("rx_cd0af269d6c39e55c39293e7debf7ee2_","rx_cd0af269d6c39e55c39293e7debf7ee2__ode_solver_solvedata", (DL_FUNC) rx_cd0af269d6c39e55c39293e7debf7ee2__ode_solver_solvedata);
#>   R_RegisterCCallable("rx_cd0af269d6c39e55c39293e7debf7ee2_","rx_cd0af269d6c39e55c39293e7debf7ee2__ode_solver_get_solvedata", (DL_FUNC) rx_cd0af269d6c39e55c39293e7debf7ee2__ode_solver_get_solvedata);
#>   R_RegisterCCallable("rx_cd0af269d6c39e55c39293e7debf7ee2_","rx_cd0af269d6c39e55c39293e7debf7ee2__F", (DL_FUNC) rx_cd0af269d6c39e55c39293e7debf7ee2__F);
#>   R_RegisterCCallable("rx_cd0af269d6c39e55c39293e7debf7ee2_","rx_cd0af269d6c39e55c39293e7debf7ee2__Lag", (DL_FUNC) rx_cd0af269d6c39e55c39293e7debf7ee2__Lag);
#>   R_RegisterCCallable("rx_cd0af269d6c39e55c39293e7debf7ee2_","rx_cd0af269d6c39e55c39293e7debf7ee2__Rate", (DL_FUNC) rx_cd0af269d6c39e55c39293e7debf7ee2__Rate);
#>   R_RegisterCCallable("rx_cd0af269d6c39e55c39293e7debf7ee2_","rx_cd0af269d6c39e55c39293e7debf7ee2__Dur", (DL_FUNC) rx_cd0af269d6c39e55c39293e7debf7ee2__Dur);
#>   R_RegisterCCallable("rx_cd0af269d6c39e55c39293e7debf7ee2_","rx_cd0af269d6c39e55c39293e7debf7ee2__mtime", (DL_FUNC) rx_cd0af269d6c39e55c39293e7debf7ee2__mtime);
#>   R_RegisterCCallable("rx_cd0af269d6c39e55c39293e7debf7ee2_","rx_cd0af269d6c39e55c39293e7debf7ee2__ME", (DL_FUNC) rx_cd0af269d6c39e55c39293e7debf7ee2__ME);
#>   R_RegisterCCallable("rx_cd0af269d6c39e55c39293e7debf7ee2_","rx_cd0af269d6c39e55c39293e7debf7ee2__IndF", (DL_FUNC) rx_cd0af269d6c39e55c39293e7debf7ee2__IndF);
#>   R_RegisterCCallable("rx_cd0af269d6c39e55c39293e7debf7ee2_","rx_cd0af269d6c39e55c39293e7debf7ee2__dydt_liblsoda", (DL_FUNC) rx_cd0af269d6c39e55c39293e7debf7ee2__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_rx_cd0af269d6c39e55c39293e7debf7ee2_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_rx_cd0af269d6c39e55c39293e7debf7ee2_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"rx_cd0af269d6c39e55c39293e7debf7ee2__model_vars", (DL_FUNC) &rx_cd0af269d6c39e55c39293e7debf7ee2__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_rx_cd0af269d6c39e55c39293e7debf7ee2_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_cd0af269d6c39e55c39293e7debf7ee2__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("rx_cd0af269d6c39e55c39293e7debf7ee2__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’

e <- et(c(-1, 0, 1, 2))

rxSolve(x, e)
#> ── Solved rxode2 object ──
#> ── Parameters (value$params): ──
#> # A tibble: 1 × 0
#> ── Initial Conditions (value$inits): ──
#> named numeric(0)
#> ── First part of data (object): ──
#> # A tibble: 4 × 2
#>    time     r
#>   <dbl> <dbl>
#> 1    -1    -2
#> 2     0     0
#> 3     1     1
#> 4     2     2