Rectified Linear Unit (ReLU) Activation Function

This function applies the Rectified Linear Unit (ReLU) activation function to the input numeric vector. The ReLU function is defined as the positive part of its argument: \(f(x) = max(0, x)\).

Usage

ReLU(x)

Arguments

x

A numeric vector. All elements must be finite and non-missing.

Value

A numeric vector where the ReLU function has been applied to each element of x.

See also

Author

Matthew Fidler

Examples

ReLU(c(-1, 0, 1, 2))
#> [1] 0 0 1 2

# Can also be used in rxode2:
x <- rxode2({
   r=ReLU(time)
})
#>  
#>  
#> C model:
#> // Define 1 LHS values
#> #define _LHS_0_ 0 // r
#> #define _getRxSolve_ _rx6023fafa5b63d8824469492bc57027e3429__getRxSolve__OCP3
#> #define _evalUdf _rx6023fafa5b63d8824469492bc57027e3430__evalUdf_OyHg
#> #define _solveData _rx6023fafa5b63d8824469492bc57027e3431__solveData_jCyj
#> #define _assign_ptr _rx6023fafa5b63d8824469492bc57027e3432__assign_ptr_3Whg
#> #define _rxRmModelLib _rx6023fafa5b63d8824469492bc57027e3433__rxRmModelLib_1DId
#> #define _rxGetModelLib _rx6023fafa5b63d8824469492bc57027e3434__rxGetModelLib_TQhw
#> #define _old_c _rx6023fafa5b63d8824469492bc57027e3435__old_c_BDgU
#> #define _ptrid _rx6023fafa5b63d8824469492bc57027e3436__ptrid_de5l
#> #define _rxIsCurrentC _rx6023fafa5b63d8824469492bc57027e3437__rxIsCurrentC_lzyO
#> #define _sumPS _rx6023fafa5b63d8824469492bc57027e3438__sumPS_SJAW
#> #define _prodPS _rx6023fafa5b63d8824469492bc57027e3439__prodPS_UmCi
#> #define _prodType _rx6023fafa5b63d8824469492bc57027e3440__prodType_4LpG
#> #define _sumType _rx6023fafa5b63d8824469492bc57027e3441__sumType_ZKaB
#> #define _update_par_ptr _rx6023fafa5b63d8824469492bc57027e3442__update_par_ptr_wXXH
#> #define _getParCov _rx6023fafa5b63d8824469492bc57027e3443__getParCov_RN1S
#> #define _rxode2_rxAssignPtr _rx6023fafa5b63d8824469492bc57027e3444__rxode2_rxAssignPtr_O1NY
#> #define _rxQr _rx6023fafa5b63d8824469492bc57027e3445__rxQr_rNWb
#> #define _compareFactorVal _rx6023fafa5b63d8824469492bc57027e3446__compareFactorVal_GHTv
#> #define _sum _rx6023fafa5b63d8824469492bc57027e3447__sum_ODf3
#> #define _udf _rx6023fafa5b63d8824469492bc57027e3448__udf_z9Wq
#> #define _sign _rx6023fafa5b63d8824469492bc57027e3449__sign_iEXp
#> #define _prod _rx6023fafa5b63d8824469492bc57027e3450__prod_1Tno
#> #define _max _rx6023fafa5b63d8824469492bc57027e3451__max_PDPf
#> #define _min _rx6023fafa5b63d8824469492bc57027e3452__min_VeVp
#> #define _transit4P _rx6023fafa5b63d8824469492bc57027e3453__transit4P_hs0t
#> #define _transit3P _rx6023fafa5b63d8824469492bc57027e3454__transit3P_zVMm
#> #define _assignFuns0 _rx6023fafa5b63d8824469492bc57027e3455__assignFuns0_adrZ
#> #define _assignFuns _rx6023fafa5b63d8824469492bc57027e3456__assignFuns_FurF
#> #define _rxord _rx6023fafa5b63d8824469492bc57027e3457__rxord_Ol7C
#> #define __assignFuns2 _rx6023fafa5b63d8824469492bc57027e3458___assignFuns2_Afrv
#> #define _llikCauchyDscale _rx6023fafa5b63d8824469492bc57027e3459__llikCauchyDscale_YZgF
#> #define _llikCauchyDlocation _rx6023fafa5b63d8824469492bc57027e3460__llikCauchyDlocation_2u4z
#> #define _llikCauchy _rx6023fafa5b63d8824469492bc57027e3461__llikCauchy_bAu2
#> #define _llikGammaDrate _rx6023fafa5b63d8824469492bc57027e3462__llikGammaDrate_aEuo
#> #define _llikGammaDshape _rx6023fafa5b63d8824469492bc57027e3463__llikGammaDshape_hDjD
#> #define _llikGamma _rx6023fafa5b63d8824469492bc57027e3464__llikGamma_SKZF
#> #define _llikWeibullDscale _rx6023fafa5b63d8824469492bc57027e3465__llikWeibullDscale_Suyo
#> #define _llikWeibullDshape _rx6023fafa5b63d8824469492bc57027e3466__llikWeibullDshape_talz
#> #define _llikWeibull _rx6023fafa5b63d8824469492bc57027e3467__llikWeibull_iDIi
#> #define _llikUnifDbeta _rx6023fafa5b63d8824469492bc57027e3468__llikUnifDbeta_rE0w
#> #define _llikUnifDalpha _rx6023fafa5b63d8824469492bc57027e3469__llikUnifDalpha_RkfP
#> #define _llikUnif _rx6023fafa5b63d8824469492bc57027e3470__llikUnif_Ljf5
#> #define _llikGeomDp _rx6023fafa5b63d8824469492bc57027e3471__llikGeomDp_r4q6
#> #define _llikGeom _rx6023fafa5b63d8824469492bc57027e3472__llikGeom_UFgA
#> #define _llikFDdf2 _rx6023fafa5b63d8824469492bc57027e3473__llikFDdf2_xJMt
#> #define _llikFDdf1 _rx6023fafa5b63d8824469492bc57027e3474__llikFDdf1_p6ul
#> #define _llikF _rx6023fafa5b63d8824469492bc57027e3475__llikF_NzLw
#> #define _llikExpDrate _rx6023fafa5b63d8824469492bc57027e3476__llikExpDrate_D19q
#> #define _llikExp _rx6023fafa5b63d8824469492bc57027e3477__llikExp_DD2w
#> #define _llikChisqDdf _rx6023fafa5b63d8824469492bc57027e3478__llikChisqDdf_FmHs
#> #define _llikChisq _rx6023fafa5b63d8824469492bc57027e3479__llikChisq_Iyal
#> #define _llikTDsd _rx6023fafa5b63d8824469492bc57027e3480__llikTDsd_plP8
#> #define _llikTDmean _rx6023fafa5b63d8824469492bc57027e3481__llikTDmean_5ebe
#> #define _llikTDdf _rx6023fafa5b63d8824469492bc57027e3482__llikTDdf_Kyis
#> #define _llikT _rx6023fafa5b63d8824469492bc57027e3483__llikT_yDfO
#> #define _llikBetaDshape2 _rx6023fafa5b63d8824469492bc57027e3484__llikBetaDshape2_2cbP
#> #define _llikBetaDshape1 _rx6023fafa5b63d8824469492bc57027e3485__llikBetaDshape1_rWy1
#> #define _llikBeta _rx6023fafa5b63d8824469492bc57027e3486__llikBeta_fNug
#> #define _llikNbinomMuDmu _rx6023fafa5b63d8824469492bc57027e3487__llikNbinomMuDmu_fkXE
#> #define _llikNbinomMu _rx6023fafa5b63d8824469492bc57027e3488__llikNbinomMu_VPjd
#> #define _llikNbinomDprob _rx6023fafa5b63d8824469492bc57027e3489__llikNbinomDprob_OxI3
#> #define _llikNbinom _rx6023fafa5b63d8824469492bc57027e3490__llikNbinom_ZD7w
#> #define _llikBinomDprob _rx6023fafa5b63d8824469492bc57027e3491__llikBinomDprob_nkXX
#> #define _llikBinom _rx6023fafa5b63d8824469492bc57027e3492__llikBinom_5OAg
#> #define _llikPoisDlambda _rx6023fafa5b63d8824469492bc57027e3493__llikPoisDlambda_fwCy
#> #define _llikPois _rx6023fafa5b63d8824469492bc57027e3494__llikPois_Neo4
#> #define _llikNormDsd _rx6023fafa5b63d8824469492bc57027e3495__llikNormDsd_WlQY
#> #define _llikNormDmean _rx6023fafa5b63d8824469492bc57027e3496__llikNormDmean_C1yB
#> #define _llikNorm _rx6023fafa5b63d8824469492bc57027e3497__llikNorm_mspb
#> #define simeps _rx6023fafa5b63d8824469492bc57027e3498_simeps_u285
#> #define simeta _rx6023fafa5b63d8824469492bc57027e3499_simeta_hP47
#> #define expit _rx6023fafa5b63d8824469492bc57027e3500_expit_wDfC
#> #define logit _rx6023fafa5b63d8824469492bc57027e3501_logit_Yp7G
#> #define gammapDer _rx6023fafa5b63d8824469492bc57027e3502_gammapDer_H0qP
#> #define lowergamma _rx6023fafa5b63d8824469492bc57027e3503_lowergamma_kiVy
#> #define uppergamma _rx6023fafa5b63d8824469492bc57027e3504_uppergamma_MAm1
#> #define gammaqInva _rx6023fafa5b63d8824469492bc57027e3505_gammaqInva_mA1m
#> #define gammaqInv _rx6023fafa5b63d8824469492bc57027e3506_gammaqInv_pxNA
#> #define gammapInva _rx6023fafa5b63d8824469492bc57027e3507_gammapInva_bn2j
#> #define gammapInv _rx6023fafa5b63d8824469492bc57027e3508_gammapInv_D07g
#> #define gammaq _rx6023fafa5b63d8824469492bc57027e3509_gammaq_dKOh
#> #define gammap _rx6023fafa5b63d8824469492bc57027e3510_gammap_hvSa
#> #define phi _rx6023fafa5b63d8824469492bc57027e3511_phi_AZta
#> #define d2ELUa _rx6023fafa5b63d8824469492bc57027e3512_d2ELUa_pamO
#> #define dELUa _rx6023fafa5b63d8824469492bc57027e3513_dELUa_qjJ6
#> #define d2aELU _rx6023fafa5b63d8824469492bc57027e3514_d2aELU_6nN3
#> #define d2ELU _rx6023fafa5b63d8824469492bc57027e3515_d2ELU_zb9X
#> #define dELU _rx6023fafa5b63d8824469492bc57027e3516_dELU_7x2A
#> #define ELU _rx6023fafa5b63d8824469492bc57027e3517_ELU_wyis
#> #define dPReLUa1 _rx6023fafa5b63d8824469492bc57027e3518_dPReLUa1_y2Rd
#> #define dPReLUa _rx6023fafa5b63d8824469492bc57027e3519_dPReLUa_zJrX
#> #define dPReLU _rx6023fafa5b63d8824469492bc57027e3520_dPReLU_sVrr
#> #define PReLU _rx6023fafa5b63d8824469492bc57027e3521_PReLU_L0Cw
#> #define dSwish _rx6023fafa5b63d8824469492bc57027e3522_dSwish_Hsx0
#> #define Swish _rx6023fafa5b63d8824469492bc57027e3523_Swish_YDej
#> #define dlReLU _rx6023fafa5b63d8824469492bc57027e3524_dlReLU_sNtq
#> #define lReLU _rx6023fafa5b63d8824469492bc57027e3525_lReLU_j2Kk
#> #define dSELU _rx6023fafa5b63d8824469492bc57027e3526_dSELU_uCTN
#> #define SELU _rx6023fafa5b63d8824469492bc57027e3527_SELU_OD0M
#> #define d4softplus _rx6023fafa5b63d8824469492bc57027e3528_d4softplus_Wq8N
#> #define d3softplus _rx6023fafa5b63d8824469492bc57027e3529_d3softplus_GU15
#> #define d2softplus _rx6023fafa5b63d8824469492bc57027e3530_d2softplus_zw77
#> #define dsoftplus _rx6023fafa5b63d8824469492bc57027e3531_dsoftplus_MTmU
#> #define softplus _rx6023fafa5b63d8824469492bc57027e3532_softplus_6whV
#> #define d4GELU _rx6023fafa5b63d8824469492bc57027e3533_d4GELU_WGiy
#> #define d3GELU _rx6023fafa5b63d8824469492bc57027e3534_d3GELU_en4s
#> #define d2GELU _rx6023fafa5b63d8824469492bc57027e3535_d2GELU_piMW
#> #define dGELU _rx6023fafa5b63d8824469492bc57027e3536_dGELU_T6VR
#> #define GELU _rx6023fafa5b63d8824469492bc57027e3537_GELU_Ye9I
#> #define dReLU _rx6023fafa5b63d8824469492bc57027e3538_dReLU_tadG
#> #define ReLU _rx6023fafa5b63d8824469492bc57027e3539_ReLU_JLl0
#> #define riweibull _rx6023fafa5b63d8824469492bc57027e3540_riweibull_3GaI
#> #define riunif _rx6023fafa5b63d8824469492bc57027e3541_riunif_5llO
#> #define rit_ _rx6023fafa5b63d8824469492bc57027e3542_rit__3kWH
#> #define ripois _rx6023fafa5b63d8824469492bc57027e3543_ripois_GXCd
#> #define ribeta _rx6023fafa5b63d8824469492bc57027e3544_ribeta_kGsq
#> #define rigamma _rx6023fafa5b63d8824469492bc57027e3545_rigamma_2NBv
#> #define rigeom _rx6023fafa5b63d8824469492bc57027e3546_rigeom_uBr0
#> #define rif _rx6023fafa5b63d8824469492bc57027e3547_rif_sO4e
#> #define riexp _rx6023fafa5b63d8824469492bc57027e3548_riexp_upVb
#> #define richisq _rx6023fafa5b63d8824469492bc57027e3549_richisq_TsfO
#> #define ricauchy _rx6023fafa5b63d8824469492bc57027e3550_ricauchy_YFC4
#> #define rinbinomMu _rx6023fafa5b63d8824469492bc57027e3551_rinbinomMu_Jz7i
#> #define rinbinom _rx6023fafa5b63d8824469492bc57027e3552_rinbinom_E8vt
#> #define ribinom _rx6023fafa5b63d8824469492bc57027e3553_ribinom_CFcR
#> #define rinorm _rx6023fafa5b63d8824469492bc57027e3554_rinorm_J6Ge
#> #define rxweibull _rx6023fafa5b63d8824469492bc57027e3555_rxweibull_v1Yq
#> #define rxunif _rx6023fafa5b63d8824469492bc57027e3556_rxunif_4CUG
#> #define rxt_ _rx6023fafa5b63d8824469492bc57027e3557_rxt__PRB4
#> #define rxpois _rx6023fafa5b63d8824469492bc57027e3558_rxpois_1jyX
#> #define rxbeta _rx6023fafa5b63d8824469492bc57027e3559_rxbeta_A3Sc
#> #define rxgamma _rx6023fafa5b63d8824469492bc57027e3560_rxgamma_akWU
#> #define rxgeom _rx6023fafa5b63d8824469492bc57027e3561_rxgeom_0xSy
#> #define rxf _rx6023fafa5b63d8824469492bc57027e3562_rxf_XMV4
#> #define rxexp _rx6023fafa5b63d8824469492bc57027e3563_rxexp_yezX
#> #define rxchisq _rx6023fafa5b63d8824469492bc57027e3564_rxchisq_lYmd
#> #define rxcauchy _rx6023fafa5b63d8824469492bc57027e3565_rxcauchy_NMqS
#> #define rxnbinomMu _rx6023fafa5b63d8824469492bc57027e3566_rxnbinomMu_Bxi1
#> #define rxnbinom _rx6023fafa5b63d8824469492bc57027e3567_rxnbinom_cpdh
#> #define rxbinom _rx6023fafa5b63d8824469492bc57027e3568_rxbinom_rHA6
#> #define rxnorm _rx6023fafa5b63d8824469492bc57027e3569_rxnorm_aqvf
#> #define linCmtB _rx6023fafa5b63d8824469492bc57027e3570_linCmtB_L0Qi
#> #define linCmtA _rx6023fafa5b63d8824469492bc57027e3571_linCmtA_PY92
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){}
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){}
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  rx_05683aab5201e3f155e0493fe7f54729__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *rx_05683aab5201e3f155e0493fe7f54729__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP rx_05683aab5201e3f155e0493fe7f54729__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void rx_05683aab5201e3f155e0493fe7f54729__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double r = NA_REAL;
#> 
#>   (void)t;
#>   (void)r;
#> 
#>   r = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   r=ReLU(t);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void rx_05683aab5201e3f155e0493fe7f54729__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void rx_05683aab5201e3f155e0493fe7f54729__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void rx_05683aab5201e3f155e0493fe7f54729__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double r = NA_REAL;
#> 
#>   (void)t;
#>   (void)r;
#> 
#>   r = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   r=ReLU(t);
#> 
#>   _lhs[_LHS_0_]=r;
#> }
#> // Functional based bioavailability
#> double rx_05683aab5201e3f155e0493fe7f54729__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double rx_05683aab5201e3f155e0493fe7f54729__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double rx_05683aab5201e3f155e0493fe7f54729__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double rx_05683aab5201e3f155e0493fe7f54729__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void rx_05683aab5201e3f155e0493fe7f54729__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void rx_05683aab5201e3f155e0493fe7f54729__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void rx_05683aab5201e3f155e0493fe7f54729__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP rx_05683aab5201e3f155e0493fe7f54729__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_05683aab5201e3f155e0493fe7f54729__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAAN/zAAAv7#aT)cm4.MAg).+=WjmaB~15wu1tE9*u/kYHjAknu5mseF3[ZK*/F:I76{GhVT|v6oI{QJ)0v6{/%%T~HYU4R:*\?B.FU=yeWXLZLS\"Kg+t(h63C2{G}>iG^ipS9hs;t(t}dr+_&b~.!7{kVp<;\"s\"vd=w`OjclcW%%mUMzEFQ,wHsT)\"ZX3UL8{7)da4ERrCE%%3sr~v|No6CG9O^*s;=XO\?{gnjgO=bvm*856sX7&c;[8|U~,T2(m@0aY*ZUWFX9(pB>W\"cl&2uQWf)N8r~LM#_Wi^R^0[0:\?W=qmGp\?s^!u!L$zIn<ZGV}8nd9f,9]u*2=&KY3=>L~x7Ft{$L)E5OHP0e:J,rZ>MF0lpjwxO<m~vy/VBlB!!xON[=_^,1xxz(T01P,WY|L+`kLjV{y%%YRY0SbCYP^iC~Lzn[XtC:;>5bkdg*pf{T{k5BE\?aS;j6M3EpzkT&Et.ag!`si7:#k0jfgC3ykdFz^.`,d>$hUp0=80;=:,uyUqOg3q:<7nCd2L/fTC/Y>Rb0;<arD6+T8gjyX2eklyX<Uhl6>_&85QdHt|8T(s)s{RNJt%%LnBJo:CKQ][2u*BckFu\"1po|R[WaRzb9Zv9aC07g+{fXR!TQ6:6%%y=m8z6kWNK/H]vfN2>Mi<@Fo&js~E%%#,)=,P@NHOnt[j|N6HXS(8u>ZJ6qg:1q!L_Y(dgB:;P||V#.5OirpLx,kc1c=*fXGYrcQzISh@lL$8z:%%@+2i3!dJ\?H*XXv7RHOBbzNGFm\"[1m0f\"G\"t]`cjL%%(HDlGLDzBaX%%mnU^FV:q,7\?5FmGdT\?QL~FS4BtkkLN)gqxz>\"0wl^DlOqD#Lsj)XEJ]Ou>D*j($6JCoP9tL/i=#i:3k~yx}vFf\?eFmh2qgx!5WG%%nHwC@5:0\"8O{0D;GW@i6]}zegRA9\"))r2fw)vv7`J>S]KFMBCpTb><c,x~YANY<b^D|Fjg7G7#[oi}Du`pY]LR<Xy$tdONV|&>F}tQ!ntt|cY!gJ[j9cO*ERfzNVkjcntC+hx!_Sk)f");
#>     char buf[1064];
#> #define __doBufN__ 1064
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void rx_05683aab5201e3f155e0493fe7f54729__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   rx_05683aab5201e3f155e0493fe7f54729__dydt(neq, *t, A, DADT);
#> }
#> extern int rx_05683aab5201e3f155e0493fe7f54729__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   rx_05683aab5201e3f155e0493fe7f54729__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void rx_05683aab5201e3f155e0493fe7f54729__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   rx_05683aab5201e3f155e0493fe7f54729__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void rx_05683aab5201e3f155e0493fe7f54729__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_rx_05683aab5201e3f155e0493fe7f54729_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("rx_05683aab5201e3f155e0493fe7f54729_","rx_05683aab5201e3f155e0493fe7f54729__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("rx_05683aab5201e3f155e0493fe7f54729_","rx_05683aab5201e3f155e0493fe7f54729__assignFuns", (DL_FUNC) rx_05683aab5201e3f155e0493fe7f54729__assignFuns);
#>   R_RegisterCCallable("rx_05683aab5201e3f155e0493fe7f54729_","rx_05683aab5201e3f155e0493fe7f54729__inis",(DL_FUNC) rx_05683aab5201e3f155e0493fe7f54729__inis);
#>   R_RegisterCCallable("rx_05683aab5201e3f155e0493fe7f54729_","rx_05683aab5201e3f155e0493fe7f54729__dydt",(DL_FUNC) rx_05683aab5201e3f155e0493fe7f54729__dydt);
#>   R_RegisterCCallable("rx_05683aab5201e3f155e0493fe7f54729_","rx_05683aab5201e3f155e0493fe7f54729__calc_lhs",(DL_FUNC) rx_05683aab5201e3f155e0493fe7f54729__calc_lhs);
#>   R_RegisterCCallable("rx_05683aab5201e3f155e0493fe7f54729_","rx_05683aab5201e3f155e0493fe7f54729__calc_jac",(DL_FUNC) rx_05683aab5201e3f155e0493fe7f54729__calc_jac);
#>   R_RegisterCCallable("rx_05683aab5201e3f155e0493fe7f54729_","rx_05683aab5201e3f155e0493fe7f54729__dydt_lsoda", (DL_FUNC) rx_05683aab5201e3f155e0493fe7f54729__dydt_lsoda);
#>   R_RegisterCCallable("rx_05683aab5201e3f155e0493fe7f54729_","rx_05683aab5201e3f155e0493fe7f54729__calc_jac_lsoda", (DL_FUNC) rx_05683aab5201e3f155e0493fe7f54729__calc_jac_lsoda);
#>   R_RegisterCCallable("rx_05683aab5201e3f155e0493fe7f54729_","rx_05683aab5201e3f155e0493fe7f54729__ode_solver_solvedata", (DL_FUNC) rx_05683aab5201e3f155e0493fe7f54729__ode_solver_solvedata);
#>   R_RegisterCCallable("rx_05683aab5201e3f155e0493fe7f54729_","rx_05683aab5201e3f155e0493fe7f54729__ode_solver_get_solvedata", (DL_FUNC) rx_05683aab5201e3f155e0493fe7f54729__ode_solver_get_solvedata);
#>   R_RegisterCCallable("rx_05683aab5201e3f155e0493fe7f54729_","rx_05683aab5201e3f155e0493fe7f54729__F", (DL_FUNC) rx_05683aab5201e3f155e0493fe7f54729__F);
#>   R_RegisterCCallable("rx_05683aab5201e3f155e0493fe7f54729_","rx_05683aab5201e3f155e0493fe7f54729__Lag", (DL_FUNC) rx_05683aab5201e3f155e0493fe7f54729__Lag);
#>   R_RegisterCCallable("rx_05683aab5201e3f155e0493fe7f54729_","rx_05683aab5201e3f155e0493fe7f54729__Rate", (DL_FUNC) rx_05683aab5201e3f155e0493fe7f54729__Rate);
#>   R_RegisterCCallable("rx_05683aab5201e3f155e0493fe7f54729_","rx_05683aab5201e3f155e0493fe7f54729__Dur", (DL_FUNC) rx_05683aab5201e3f155e0493fe7f54729__Dur);
#>   R_RegisterCCallable("rx_05683aab5201e3f155e0493fe7f54729_","rx_05683aab5201e3f155e0493fe7f54729__mtime", (DL_FUNC) rx_05683aab5201e3f155e0493fe7f54729__mtime);
#>   R_RegisterCCallable("rx_05683aab5201e3f155e0493fe7f54729_","rx_05683aab5201e3f155e0493fe7f54729__ME", (DL_FUNC) rx_05683aab5201e3f155e0493fe7f54729__ME);
#>   R_RegisterCCallable("rx_05683aab5201e3f155e0493fe7f54729_","rx_05683aab5201e3f155e0493fe7f54729__IndF", (DL_FUNC) rx_05683aab5201e3f155e0493fe7f54729__IndF);
#>   R_RegisterCCallable("rx_05683aab5201e3f155e0493fe7f54729_","rx_05683aab5201e3f155e0493fe7f54729__dydt_liblsoda", (DL_FUNC) rx_05683aab5201e3f155e0493fe7f54729__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_rx_05683aab5201e3f155e0493fe7f54729_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_rx_05683aab5201e3f155e0493fe7f54729_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"rx_05683aab5201e3f155e0493fe7f54729__model_vars", (DL_FUNC) &rx_05683aab5201e3f155e0493fe7f54729__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_rx_05683aab5201e3f155e0493fe7f54729_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_05683aab5201e3f155e0493fe7f54729__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("rx_05683aab5201e3f155e0493fe7f54729__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’

e <- et(c(-1, 0, 1, 2))

rxSolve(x, e)
#> ── Solved rxode2 object ──
#> ── Parameters (value$params): ──
#> # A tibble: 1 × 0
#> ── Initial Conditions (value$inits): ──
#> named numeric(0)
#> ── First part of data (object): ──
#> # A tibble: 4 × 2
#>    time     r
#>   <dbl> <dbl>
#> 1    -1     0
#> 2     0     0
#> 3     1     1
#> 4     2     2