Scaled Exponential Linear Unit (SELU) Activation Function

Scaled Exponential Linear Unit (SELU) Activation Function

Usage

SELU(x)

Arguments

x

A numeric vector. All elements must be finite and non-missing.

Value

A numeric vector where the ReLU function has been applied to each element of x.

See also

Author

Matthew Fidler

Examples

SELU(c(-1, 0, 1, 2))
#> [1] -1.111331  0.000000  1.050701  2.101402

# Can also be used in rxode2:
x <- rxode2({
   r=SELU(time)
})
#>  
#>  
#> C model:
#> // Define 1 LHS values
#> #define _LHS_0_ 0 // r
#> #define _getRxSolve_ _rxe2adab73c2168560d4493698814d3641572__getRxSolve__OyHg
#> #define _evalUdf _rxe2adab73c2168560d4493698814d3641573__evalUdf_jCyj
#> #define _solveData _rxe2adab73c2168560d4493698814d3641574__solveData_3Whg
#> #define _assign_ptr _rxe2adab73c2168560d4493698814d3641575__assign_ptr_1DId
#> #define _rxRmModelLib _rxe2adab73c2168560d4493698814d3641576__rxRmModelLib_TQhw
#> #define _rxGetModelLib _rxe2adab73c2168560d4493698814d3641577__rxGetModelLib_BDgU
#> #define _old_c _rxe2adab73c2168560d4493698814d3641578__old_c_de5l
#> #define _ptrid _rxe2adab73c2168560d4493698814d3641579__ptrid_lzyO
#> #define _rxIsCurrentC _rxe2adab73c2168560d4493698814d3641580__rxIsCurrentC_SJAW
#> #define _sumPS _rxe2adab73c2168560d4493698814d3641581__sumPS_UmCi
#> #define _prodPS _rxe2adab73c2168560d4493698814d3641582__prodPS_4LpG
#> #define _prodType _rxe2adab73c2168560d4493698814d3641583__prodType_ZKaB
#> #define _sumType _rxe2adab73c2168560d4493698814d3641584__sumType_wXXH
#> #define _update_par_ptr _rxe2adab73c2168560d4493698814d3641585__update_par_ptr_RN1S
#> #define _getParCov _rxe2adab73c2168560d4493698814d3641586__getParCov_O1NY
#> #define _rxode2_rxAssignPtr _rxe2adab73c2168560d4493698814d3641587__rxode2_rxAssignPtr_rNWb
#> #define _rxQr _rxe2adab73c2168560d4493698814d3641588__rxQr_GHTv
#> #define _compareFactorVal _rxe2adab73c2168560d4493698814d3641589__compareFactorVal_ODf3
#> #define _sum _rxe2adab73c2168560d4493698814d3641590__sum_z9Wq
#> #define _udf _rxe2adab73c2168560d4493698814d3641591__udf_iEXp
#> #define _sign _rxe2adab73c2168560d4493698814d3641592__sign_1Tno
#> #define _prod _rxe2adab73c2168560d4493698814d3641593__prod_PDPf
#> #define _max _rxe2adab73c2168560d4493698814d3641594__max_VeVp
#> #define _min _rxe2adab73c2168560d4493698814d3641595__min_hs0t
#> #define _transit4P _rxe2adab73c2168560d4493698814d3641596__transit4P_zVMm
#> #define _transit3P _rxe2adab73c2168560d4493698814d3641597__transit3P_adrZ
#> #define _assignFuns0 _rxe2adab73c2168560d4493698814d3641598__assignFuns0_FurF
#> #define _assignFuns _rxe2adab73c2168560d4493698814d3641599__assignFuns_Ol7C
#> #define _rxord _rxe2adab73c2168560d4493698814d3641600__rxord_Afrv
#> #define __assignFuns2 _rxe2adab73c2168560d4493698814d3641601___assignFuns2_YZgF
#> #define _llikCauchyDscale _rxe2adab73c2168560d4493698814d3641602__llikCauchyDscale_2u4z
#> #define _llikCauchyDlocation _rxe2adab73c2168560d4493698814d3641603__llikCauchyDlocation_bAu2
#> #define _llikCauchy _rxe2adab73c2168560d4493698814d3641604__llikCauchy_aEuo
#> #define _llikGammaDrate _rxe2adab73c2168560d4493698814d3641605__llikGammaDrate_hDjD
#> #define _llikGammaDshape _rxe2adab73c2168560d4493698814d3641606__llikGammaDshape_SKZF
#> #define _llikGamma _rxe2adab73c2168560d4493698814d3641607__llikGamma_Suyo
#> #define _llikWeibullDscale _rxe2adab73c2168560d4493698814d3641608__llikWeibullDscale_talz
#> #define _llikWeibullDshape _rxe2adab73c2168560d4493698814d3641609__llikWeibullDshape_iDIi
#> #define _llikWeibull _rxe2adab73c2168560d4493698814d3641610__llikWeibull_rE0w
#> #define _llikUnifDbeta _rxe2adab73c2168560d4493698814d3641611__llikUnifDbeta_RkfP
#> #define _llikUnifDalpha _rxe2adab73c2168560d4493698814d3641612__llikUnifDalpha_Ljf5
#> #define _llikUnif _rxe2adab73c2168560d4493698814d3641613__llikUnif_r4q6
#> #define _llikGeomDp _rxe2adab73c2168560d4493698814d3641614__llikGeomDp_UFgA
#> #define _llikGeom _rxe2adab73c2168560d4493698814d3641615__llikGeom_xJMt
#> #define _llikFDdf2 _rxe2adab73c2168560d4493698814d3641616__llikFDdf2_p6ul
#> #define _llikFDdf1 _rxe2adab73c2168560d4493698814d3641617__llikFDdf1_NzLw
#> #define _llikF _rxe2adab73c2168560d4493698814d3641618__llikF_D19q
#> #define _llikExpDrate _rxe2adab73c2168560d4493698814d3641619__llikExpDrate_DD2w
#> #define _llikExp _rxe2adab73c2168560d4493698814d3641620__llikExp_FmHs
#> #define _llikChisqDdf _rxe2adab73c2168560d4493698814d3641621__llikChisqDdf_Iyal
#> #define _llikChisq _rxe2adab73c2168560d4493698814d3641622__llikChisq_plP8
#> #define _llikTDsd _rxe2adab73c2168560d4493698814d3641623__llikTDsd_5ebe
#> #define _llikTDmean _rxe2adab73c2168560d4493698814d3641624__llikTDmean_Kyis
#> #define _llikTDdf _rxe2adab73c2168560d4493698814d3641625__llikTDdf_yDfO
#> #define _llikT _rxe2adab73c2168560d4493698814d3641626__llikT_2cbP
#> #define _llikBetaDshape2 _rxe2adab73c2168560d4493698814d3641627__llikBetaDshape2_rWy1
#> #define _llikBetaDshape1 _rxe2adab73c2168560d4493698814d3641628__llikBetaDshape1_fNug
#> #define _llikBeta _rxe2adab73c2168560d4493698814d3641629__llikBeta_fkXE
#> #define _llikNbinomMuDmu _rxe2adab73c2168560d4493698814d3641630__llikNbinomMuDmu_VPjd
#> #define _llikNbinomMu _rxe2adab73c2168560d4493698814d3641631__llikNbinomMu_OxI3
#> #define _llikNbinomDprob _rxe2adab73c2168560d4493698814d3641632__llikNbinomDprob_ZD7w
#> #define _llikNbinom _rxe2adab73c2168560d4493698814d3641633__llikNbinom_nkXX
#> #define _llikBinomDprob _rxe2adab73c2168560d4493698814d3641634__llikBinomDprob_5OAg
#> #define _llikBinom _rxe2adab73c2168560d4493698814d3641635__llikBinom_fwCy
#> #define _llikPoisDlambda _rxe2adab73c2168560d4493698814d3641636__llikPoisDlambda_Neo4
#> #define _llikPois _rxe2adab73c2168560d4493698814d3641637__llikPois_WlQY
#> #define _llikNormDsd _rxe2adab73c2168560d4493698814d3641638__llikNormDsd_C1yB
#> #define _llikNormDmean _rxe2adab73c2168560d4493698814d3641639__llikNormDmean_mspb
#> #define _llikNorm _rxe2adab73c2168560d4493698814d3641640__llikNorm_u285
#> #define simeps _rxe2adab73c2168560d4493698814d3641641_simeps_hP47
#> #define simeta _rxe2adab73c2168560d4493698814d3641642_simeta_wDfC
#> #define expit _rxe2adab73c2168560d4493698814d3641643_expit_Yp7G
#> #define logit _rxe2adab73c2168560d4493698814d3641644_logit_H0qP
#> #define gammapDer _rxe2adab73c2168560d4493698814d3641645_gammapDer_kiVy
#> #define lowergamma _rxe2adab73c2168560d4493698814d3641646_lowergamma_MAm1
#> #define uppergamma _rxe2adab73c2168560d4493698814d3641647_uppergamma_mA1m
#> #define gammaqInva _rxe2adab73c2168560d4493698814d3641648_gammaqInva_pxNA
#> #define gammaqInv _rxe2adab73c2168560d4493698814d3641649_gammaqInv_bn2j
#> #define gammapInva _rxe2adab73c2168560d4493698814d3641650_gammapInva_D07g
#> #define gammapInv _rxe2adab73c2168560d4493698814d3641651_gammapInv_dKOh
#> #define gammaq _rxe2adab73c2168560d4493698814d3641652_gammaq_hvSa
#> #define gammap _rxe2adab73c2168560d4493698814d3641653_gammap_AZta
#> #define phi _rxe2adab73c2168560d4493698814d3641654_phi_pamO
#> #define d2ELUa _rxe2adab73c2168560d4493698814d3641655_d2ELUa_qjJ6
#> #define dELUa _rxe2adab73c2168560d4493698814d3641656_dELUa_6nN3
#> #define d2aELU _rxe2adab73c2168560d4493698814d3641657_d2aELU_zb9X
#> #define d2ELU _rxe2adab73c2168560d4493698814d3641658_d2ELU_7x2A
#> #define dELU _rxe2adab73c2168560d4493698814d3641659_dELU_wyis
#> #define ELU _rxe2adab73c2168560d4493698814d3641660_ELU_y2Rd
#> #define dPReLUa1 _rxe2adab73c2168560d4493698814d3641661_dPReLUa1_zJrX
#> #define dPReLUa _rxe2adab73c2168560d4493698814d3641662_dPReLUa_sVrr
#> #define dPReLU _rxe2adab73c2168560d4493698814d3641663_dPReLU_L0Cw
#> #define PReLU _rxe2adab73c2168560d4493698814d3641664_PReLU_Hsx0
#> #define dSwish _rxe2adab73c2168560d4493698814d3641665_dSwish_YDej
#> #define Swish _rxe2adab73c2168560d4493698814d3641666_Swish_sNtq
#> #define dlReLU _rxe2adab73c2168560d4493698814d3641667_dlReLU_j2Kk
#> #define lReLU _rxe2adab73c2168560d4493698814d3641668_lReLU_uCTN
#> #define dSELU _rxe2adab73c2168560d4493698814d3641669_dSELU_OD0M
#> #define SELU _rxe2adab73c2168560d4493698814d3641670_SELU_Wq8N
#> #define d4softplus _rxe2adab73c2168560d4493698814d3641671_d4softplus_GU15
#> #define d3softplus _rxe2adab73c2168560d4493698814d3641672_d3softplus_zw77
#> #define d2softplus _rxe2adab73c2168560d4493698814d3641673_d2softplus_MTmU
#> #define dsoftplus _rxe2adab73c2168560d4493698814d3641674_dsoftplus_6whV
#> #define softplus _rxe2adab73c2168560d4493698814d3641675_softplus_WGiy
#> #define d4GELU _rxe2adab73c2168560d4493698814d3641676_d4GELU_en4s
#> #define d3GELU _rxe2adab73c2168560d4493698814d3641677_d3GELU_piMW
#> #define d2GELU _rxe2adab73c2168560d4493698814d3641678_d2GELU_T6VR
#> #define dGELU _rxe2adab73c2168560d4493698814d3641679_dGELU_Ye9I
#> #define GELU _rxe2adab73c2168560d4493698814d3641680_GELU_tadG
#> #define dReLU _rxe2adab73c2168560d4493698814d3641681_dReLU_JLl0
#> #define ReLU _rxe2adab73c2168560d4493698814d3641682_ReLU_3GaI
#> #define riweibull _rxe2adab73c2168560d4493698814d3641683_riweibull_5llO
#> #define riunif _rxe2adab73c2168560d4493698814d3641684_riunif_3kWH
#> #define rit_ _rxe2adab73c2168560d4493698814d3641685_rit__GXCd
#> #define ripois _rxe2adab73c2168560d4493698814d3641686_ripois_kGsq
#> #define ribeta _rxe2adab73c2168560d4493698814d3641687_ribeta_2NBv
#> #define rigamma _rxe2adab73c2168560d4493698814d3641688_rigamma_uBr0
#> #define rigeom _rxe2adab73c2168560d4493698814d3641689_rigeom_sO4e
#> #define rif _rxe2adab73c2168560d4493698814d3641690_rif_upVb
#> #define riexp _rxe2adab73c2168560d4493698814d3641691_riexp_TsfO
#> #define richisq _rxe2adab73c2168560d4493698814d3641692_richisq_YFC4
#> #define ricauchy _rxe2adab73c2168560d4493698814d3641693_ricauchy_Jz7i
#> #define rinbinomMu _rxe2adab73c2168560d4493698814d3641694_rinbinomMu_E8vt
#> #define rinbinom _rxe2adab73c2168560d4493698814d3641695_rinbinom_CFcR
#> #define ribinom _rxe2adab73c2168560d4493698814d3641696_ribinom_J6Ge
#> #define rinorm _rxe2adab73c2168560d4493698814d3641697_rinorm_v1Yq
#> #define rxweibull _rxe2adab73c2168560d4493698814d3641698_rxweibull_4CUG
#> #define rxunif _rxe2adab73c2168560d4493698814d3641699_rxunif_PRB4
#> #define rxt_ _rxe2adab73c2168560d4493698814d3641700_rxt__1jyX
#> #define rxpois _rxe2adab73c2168560d4493698814d3641701_rxpois_A3Sc
#> #define rxbeta _rxe2adab73c2168560d4493698814d3641702_rxbeta_akWU
#> #define rxgamma _rxe2adab73c2168560d4493698814d3641703_rxgamma_0xSy
#> #define rxgeom _rxe2adab73c2168560d4493698814d3641704_rxgeom_XMV4
#> #define rxf _rxe2adab73c2168560d4493698814d3641705_rxf_yezX
#> #define rxexp _rxe2adab73c2168560d4493698814d3641706_rxexp_lYmd
#> #define rxchisq _rxe2adab73c2168560d4493698814d3641707_rxchisq_NMqS
#> #define rxcauchy _rxe2adab73c2168560d4493698814d3641708_rxcauchy_Bxi1
#> #define rxnbinomMu _rxe2adab73c2168560d4493698814d3641709_rxnbinomMu_cpdh
#> #define rxnbinom _rxe2adab73c2168560d4493698814d3641710_rxnbinom_rHA6
#> #define rxbinom _rxe2adab73c2168560d4493698814d3641711_rxbinom_aqvf
#> #define rxnorm _rxe2adab73c2168560d4493698814d3641712_rxnorm_L0Qi
#> #define linCmtB _rxe2adab73c2168560d4493698814d3641713_linCmtB_PY92
#> #define linCmtA _rxe2adab73c2168560d4493698814d3641714_linCmtA_gVwq
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){}
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){}
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  rx_51bb622b72f018aaa685322695a476d3__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *rx_51bb622b72f018aaa685322695a476d3__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP rx_51bb622b72f018aaa685322695a476d3__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void rx_51bb622b72f018aaa685322695a476d3__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double r = NA_REAL;
#> 
#>   (void)t;
#>   (void)r;
#> 
#>   r = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   r=SELU(t);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void rx_51bb622b72f018aaa685322695a476d3__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void rx_51bb622b72f018aaa685322695a476d3__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void rx_51bb622b72f018aaa685322695a476d3__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double r = NA_REAL;
#> 
#>   (void)t;
#>   (void)r;
#> 
#>   r = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   r=SELU(t);
#> 
#>   _lhs[_LHS_0_]=r;
#> }
#> // Functional based bioavailability
#> double rx_51bb622b72f018aaa685322695a476d3__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double rx_51bb622b72f018aaa685322695a476d3__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double rx_51bb622b72f018aaa685322695a476d3__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double rx_51bb622b72f018aaa685322695a476d3__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void rx_51bb622b72f018aaa685322695a476d3__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void rx_51bb622b72f018aaa685322695a476d3__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void rx_51bb622b72f018aaa685322695a476d3__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP rx_51bb622b72f018aaa685322695a476d3__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_51bb622b72f018aaa685322695a476d3__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAAN/zAAAv7#aT)cm4.MAg).+h\";*7t1)=JDQOv^o^FG+[W\"n7(!ZXt<MNMDd#f%%MrG,xE4:[fiF)9Sh\?Tn!V@^V|D8~;j<Xt9_rfLQPL5F6FIA.#:4^8Qd^}7uYK)MREP[#5{<]Pd$x4;,(eN/v!imXp;;(B2L5$.`T77l3`wTrz1!!@4QZ_8nHMdk\"ve3NmIz.)].{(l8{.>/_e4M[T|/g64HtF{E_lnj%%,l^Fv8zG[{.JdLqT:&%%hp.P6ns`qOENY^`B7\"_vHzd&7^~eE.jQ}:TS>CQ&\"Q}1q,+w0`4mTmGp[q^!tRO=,ah<#X_Sxqu6f,q73&PfPnLF3n_tgcwWs=^4@/]Z)zPHJ,sZ>MF0b0i(KeGTp>YQaARtI$\"B@k>q|jwY{Z7.!M/RGG[aT3R+\"EWjZb=1*Z.)(l3Zn.i!n[lq~/Af$qNR`}_;#XyYP!`E0)\"jyP04ffzh1P*7M;=)p:gF!bHW~Whn`Lo4fkD#Z,Z=\?s$P\?#OrCW4_)vP|u~dY8fe%%a>vo}gt8:Pm9z7(s.nj\"(2I$~)[8V$ru%%v=Y]#zr{oe,QQg3fHH<c>gHtciZO5hFaY*j`K!cVt7puIAJ8G0x)0{2qLc&:/6l|\"N9BGo[g6E(8[O/HsLFZ{4E_]h]8XBa%%+Bvl[Fo0]yeIJ)`8`o2Lr]j@&`[4EMq[bo|dWWh)CBs\?qh2K(7OT!{42V0*oGvd<%%++Inh.C}yl!!W:Msoa\?1aJGQ5{/kmj~B}uEi@ytigShD1/rz\"klL<v4Om<7y+>LrVH>v_Qa)2B.t](e/R=Ky\")G{>R^TlB4uf.Ux.~C.]W5fgZE!23@M%%WaY$pvSHVw<G(S\?MvEpI)_B[!i=:EBU1e*ykR_fS]c#t+Ys*C!_9v$Qc&!*<cLD\?$DYC`clat8O{0D;GW@i6]}zegRA9\"))r2fw)vv7`J>S]KFMBCpTb><c,x~YANY<b^D|Fjg7G7#[oi}Du`pY]LR<Xy$tdONV|&>F}tQ!ntt|cY!gJ[j9cO*ERfzNVkjcntC+hxUy#r#o");
#>     char buf[1061];
#> #define __doBufN__ 1061
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void rx_51bb622b72f018aaa685322695a476d3__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   rx_51bb622b72f018aaa685322695a476d3__dydt(neq, *t, A, DADT);
#> }
#> extern int rx_51bb622b72f018aaa685322695a476d3__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   rx_51bb622b72f018aaa685322695a476d3__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void rx_51bb622b72f018aaa685322695a476d3__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   rx_51bb622b72f018aaa685322695a476d3__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void rx_51bb622b72f018aaa685322695a476d3__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_rx_51bb622b72f018aaa685322695a476d3_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("rx_51bb622b72f018aaa685322695a476d3_","rx_51bb622b72f018aaa685322695a476d3__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("rx_51bb622b72f018aaa685322695a476d3_","rx_51bb622b72f018aaa685322695a476d3__assignFuns", (DL_FUNC) rx_51bb622b72f018aaa685322695a476d3__assignFuns);
#>   R_RegisterCCallable("rx_51bb622b72f018aaa685322695a476d3_","rx_51bb622b72f018aaa685322695a476d3__inis",(DL_FUNC) rx_51bb622b72f018aaa685322695a476d3__inis);
#>   R_RegisterCCallable("rx_51bb622b72f018aaa685322695a476d3_","rx_51bb622b72f018aaa685322695a476d3__dydt",(DL_FUNC) rx_51bb622b72f018aaa685322695a476d3__dydt);
#>   R_RegisterCCallable("rx_51bb622b72f018aaa685322695a476d3_","rx_51bb622b72f018aaa685322695a476d3__calc_lhs",(DL_FUNC) rx_51bb622b72f018aaa685322695a476d3__calc_lhs);
#>   R_RegisterCCallable("rx_51bb622b72f018aaa685322695a476d3_","rx_51bb622b72f018aaa685322695a476d3__calc_jac",(DL_FUNC) rx_51bb622b72f018aaa685322695a476d3__calc_jac);
#>   R_RegisterCCallable("rx_51bb622b72f018aaa685322695a476d3_","rx_51bb622b72f018aaa685322695a476d3__dydt_lsoda", (DL_FUNC) rx_51bb622b72f018aaa685322695a476d3__dydt_lsoda);
#>   R_RegisterCCallable("rx_51bb622b72f018aaa685322695a476d3_","rx_51bb622b72f018aaa685322695a476d3__calc_jac_lsoda", (DL_FUNC) rx_51bb622b72f018aaa685322695a476d3__calc_jac_lsoda);
#>   R_RegisterCCallable("rx_51bb622b72f018aaa685322695a476d3_","rx_51bb622b72f018aaa685322695a476d3__ode_solver_solvedata", (DL_FUNC) rx_51bb622b72f018aaa685322695a476d3__ode_solver_solvedata);
#>   R_RegisterCCallable("rx_51bb622b72f018aaa685322695a476d3_","rx_51bb622b72f018aaa685322695a476d3__ode_solver_get_solvedata", (DL_FUNC) rx_51bb622b72f018aaa685322695a476d3__ode_solver_get_solvedata);
#>   R_RegisterCCallable("rx_51bb622b72f018aaa685322695a476d3_","rx_51bb622b72f018aaa685322695a476d3__F", (DL_FUNC) rx_51bb622b72f018aaa685322695a476d3__F);
#>   R_RegisterCCallable("rx_51bb622b72f018aaa685322695a476d3_","rx_51bb622b72f018aaa685322695a476d3__Lag", (DL_FUNC) rx_51bb622b72f018aaa685322695a476d3__Lag);
#>   R_RegisterCCallable("rx_51bb622b72f018aaa685322695a476d3_","rx_51bb622b72f018aaa685322695a476d3__Rate", (DL_FUNC) rx_51bb622b72f018aaa685322695a476d3__Rate);
#>   R_RegisterCCallable("rx_51bb622b72f018aaa685322695a476d3_","rx_51bb622b72f018aaa685322695a476d3__Dur", (DL_FUNC) rx_51bb622b72f018aaa685322695a476d3__Dur);
#>   R_RegisterCCallable("rx_51bb622b72f018aaa685322695a476d3_","rx_51bb622b72f018aaa685322695a476d3__mtime", (DL_FUNC) rx_51bb622b72f018aaa685322695a476d3__mtime);
#>   R_RegisterCCallable("rx_51bb622b72f018aaa685322695a476d3_","rx_51bb622b72f018aaa685322695a476d3__ME", (DL_FUNC) rx_51bb622b72f018aaa685322695a476d3__ME);
#>   R_RegisterCCallable("rx_51bb622b72f018aaa685322695a476d3_","rx_51bb622b72f018aaa685322695a476d3__IndF", (DL_FUNC) rx_51bb622b72f018aaa685322695a476d3__IndF);
#>   R_RegisterCCallable("rx_51bb622b72f018aaa685322695a476d3_","rx_51bb622b72f018aaa685322695a476d3__dydt_liblsoda", (DL_FUNC) rx_51bb622b72f018aaa685322695a476d3__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_rx_51bb622b72f018aaa685322695a476d3_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_rx_51bb622b72f018aaa685322695a476d3_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"rx_51bb622b72f018aaa685322695a476d3__model_vars", (DL_FUNC) &rx_51bb622b72f018aaa685322695a476d3__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_rx_51bb622b72f018aaa685322695a476d3_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_51bb622b72f018aaa685322695a476d3__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("rx_51bb622b72f018aaa685322695a476d3__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’

e <- et(c(-1, 0, 1, 2))

rxSolve(x, e)
#> ── Solved rxode2 object ──
#> ── Parameters (value$params): ──
#> # A tibble: 1 × 0
#> ── Initial Conditions (value$inits): ──
#> named numeric(0)
#> ── First part of data (object): ──
#> # A tibble: 4 × 2
#>    time     r
#>   <dbl> <dbl>
#> 1    -1 -1.11
#> 2     0  0   
#> 3     1  1.05
#> 4     2  2.10