Derivatives of the Exponential Linear Unit (ELU) Activation Function

Derivatives of the Exponential Linear Unit (ELU) Activation Function

Usage

dELU(x, alpha = 1)

d2ELU(x, alpha = 1)

d2aELU(x, alpha = 1)

dELUa(x, alpha = 1)

d2ELUa(x, alpha = 1)

Arguments

x

A numeric vector. All elements must be finite and non-missing.

alpha

A numeric scalar. All elements must be finite and non-missing.

Value

A numeric vector where the derivative(s) of the ELU function has been applied to each element of x.

See also

Author

Matthew L. Fidler

Examples

dELU(c(-1, 0, 1, 2), 2)
#> [1] 0.7357589 2.0000000 1.0000000 1.0000000
d2ELU(c(-1, 0, 1, 2), 2)
#> [1] 0.7357589 2.0000000 0.0000000 0.0000000
d2aELU(c(-1, 0, 1, 2), 2)
#> [1] 0.3678794 1.0000000 0.0000000 0.0000000
dELUa(c(-1, 0, 1, 2), 2)
#> [1] -0.6321206  0.0000000  0.0000000  0.0000000
d2ELUa(c(-1, 0, 1, 2), 2)
#> [1] 0.3678794 1.0000000 0.0000000 0.0000000

# Can also be used in rxode2:
r <- rxode2({
  r1=dELU(time, 2)
  r2=d2ELU(time, 2)
  r2a=d2aELU(time, 2)
  ra=dELUa(time, 2)
  r2a=d2ELUa(time, 2)
})
#>  
#>  
#> C model:
#> // Define 4 LHS values
#> #define _LHS_0_ 0 // r1
#> #define _LHS_1_ 1 // r2
#> #define _LHS_2_ 2 // r2a
#> #define _LHS_3_ 3 // ra
#> #define _getRxSolve_ _rx6c048fd240063677d85c588de1976dee858__getRxSolve__RRbP
#> #define _evalUdf _rx6c048fd240063677d85c588de1976dee859__evalUdf_xyu5
#> #define _solveData _rx6c048fd240063677d85c588de1976dee860__solveData_TLfd
#> #define _assign_ptr _rx6c048fd240063677d85c588de1976dee861__assign_ptr_rwZX
#> #define _rxRmModelLib _rx6c048fd240063677d85c588de1976dee862__rxRmModelLib_wtP3
#> #define _rxGetModelLib _rx6c048fd240063677d85c588de1976dee863__rxGetModelLib_mvNF
#> #define _old_c _rx6c048fd240063677d85c588de1976dee864__old_c_5OB8
#> #define _ptrid _rx6c048fd240063677d85c588de1976dee865__ptrid_FY9x
#> #define _rxIsCurrentC _rx6c048fd240063677d85c588de1976dee866__rxIsCurrentC_wzh8
#> #define _sumPS _rx6c048fd240063677d85c588de1976dee867__sumPS_0ggu
#> #define _prodPS _rx6c048fd240063677d85c588de1976dee868__prodPS_PQXP
#> #define _prodType _rx6c048fd240063677d85c588de1976dee869__prodType_6NCi
#> #define _sumType _rx6c048fd240063677d85c588de1976dee870__sumType_0utk
#> #define _update_par_ptr _rx6c048fd240063677d85c588de1976dee871__update_par_ptr_xceN
#> #define _getParCov _rx6c048fd240063677d85c588de1976dee872__getParCov_4g7I
#> #define _rxode2_rxAssignPtr _rx6c048fd240063677d85c588de1976dee873__rxode2_rxAssignPtr_dO4r
#> #define _rxQr _rx6c048fd240063677d85c588de1976dee874__rxQr_Nf1a
#> #define _compareFactorVal _rx6c048fd240063677d85c588de1976dee875__compareFactorVal_HO9p
#> #define _sum _rx6c048fd240063677d85c588de1976dee876__sum_8utB
#> #define _udf _rx6c048fd240063677d85c588de1976dee877__udf_XW9W
#> #define _sign _rx6c048fd240063677d85c588de1976dee878__sign_DHV9
#> #define _prod _rx6c048fd240063677d85c588de1976dee879__prod_vUck
#> #define _max _rx6c048fd240063677d85c588de1976dee880__max_cv41
#> #define _min _rx6c048fd240063677d85c588de1976dee881__min_XEdN
#> #define _transit4P _rx6c048fd240063677d85c588de1976dee882__transit4P_22xE
#> #define _transit3P _rx6c048fd240063677d85c588de1976dee883__transit3P_0D3E
#> #define _assignFuns0 _rx6c048fd240063677d85c588de1976dee884__assignFuns0_O0Ma
#> #define _assignFuns _rx6c048fd240063677d85c588de1976dee885__assignFuns_bGfw
#> #define _rxord _rx6c048fd240063677d85c588de1976dee886__rxord_0mo4
#> #define __assignFuns2 _rx6c048fd240063677d85c588de1976dee887___assignFuns2_ZX8H
#> #define _llikCauchyDscale _rx6c048fd240063677d85c588de1976dee888__llikCauchyDscale_LoAW
#> #define _llikCauchyDlocation _rx6c048fd240063677d85c588de1976dee889__llikCauchyDlocation_lHRf
#> #define _llikCauchy _rx6c048fd240063677d85c588de1976dee890__llikCauchy_2YvQ
#> #define _llikGammaDrate _rx6c048fd240063677d85c588de1976dee891__llikGammaDrate_wyfI
#> #define _llikGammaDshape _rx6c048fd240063677d85c588de1976dee892__llikGammaDshape_CJq3
#> #define _llikGamma _rx6c048fd240063677d85c588de1976dee893__llikGamma_BTKo
#> #define _llikWeibullDscale _rx6c048fd240063677d85c588de1976dee894__llikWeibullDscale_r1pO
#> #define _llikWeibullDshape _rx6c048fd240063677d85c588de1976dee895__llikWeibullDshape_f7Tv
#> #define _llikWeibull _rx6c048fd240063677d85c588de1976dee896__llikWeibull_Zukd
#> #define _llikUnifDbeta _rx6c048fd240063677d85c588de1976dee897__llikUnifDbeta_nvAB
#> #define _llikUnifDalpha _rx6c048fd240063677d85c588de1976dee898__llikUnifDalpha_6pnH
#> #define _llikUnif _rx6c048fd240063677d85c588de1976dee899__llikUnif_RcMN
#> #define _llikGeomDp _rx6c048fd240063677d85c588de1976dee900__llikGeomDp_uO4h
#> #define _llikGeom _rx6c048fd240063677d85c588de1976dee901__llikGeom_g52O
#> #define _llikFDdf2 _rx6c048fd240063677d85c588de1976dee902__llikFDdf2_0ZwL
#> #define _llikFDdf1 _rx6c048fd240063677d85c588de1976dee903__llikFDdf1_FX7N
#> #define _llikF _rx6c048fd240063677d85c588de1976dee904__llikF_Kwul
#> #define _llikExpDrate _rx6c048fd240063677d85c588de1976dee905__llikExpDrate_lzZD
#> #define _llikExp _rx6c048fd240063677d85c588de1976dee906__llikExp_8nyy
#> #define _llikChisqDdf _rx6c048fd240063677d85c588de1976dee907__llikChisqDdf_7cr8
#> #define _llikChisq _rx6c048fd240063677d85c588de1976dee908__llikChisq_M9RU
#> #define _llikTDsd _rx6c048fd240063677d85c588de1976dee909__llikTDsd_wuwl
#> #define _llikTDmean _rx6c048fd240063677d85c588de1976dee910__llikTDmean_51KE
#> #define _llikTDdf _rx6c048fd240063677d85c588de1976dee911__llikTDdf_nUV4
#> #define _llikT _rx6c048fd240063677d85c588de1976dee912__llikT_LkN0
#> #define _llikBetaDshape2 _rx6c048fd240063677d85c588de1976dee913__llikBetaDshape2_iY4L
#> #define _llikBetaDshape1 _rx6c048fd240063677d85c588de1976dee914__llikBetaDshape1_PJjO
#> #define _llikBeta _rx6c048fd240063677d85c588de1976dee915__llikBeta_ccne
#> #define _llikNbinomMuDmu _rx6c048fd240063677d85c588de1976dee916__llikNbinomMuDmu_gPf1
#> #define _llikNbinomMu _rx6c048fd240063677d85c588de1976dee917__llikNbinomMu_o4t4
#> #define _llikNbinomDprob _rx6c048fd240063677d85c588de1976dee918__llikNbinomDprob_BmCP
#> #define _llikNbinom _rx6c048fd240063677d85c588de1976dee919__llikNbinom_RbZ3
#> #define _llikBinomDprob _rx6c048fd240063677d85c588de1976dee920__llikBinomDprob_BINi
#> #define _llikBinom _rx6c048fd240063677d85c588de1976dee921__llikBinom_7Q1S
#> #define _llikPoisDlambda _rx6c048fd240063677d85c588de1976dee922__llikPoisDlambda_nTEs
#> #define _llikPois _rx6c048fd240063677d85c588de1976dee923__llikPois_7Gsk
#> #define _llikNormDsd _rx6c048fd240063677d85c588de1976dee924__llikNormDsd_6or7
#> #define _llikNormDmean _rx6c048fd240063677d85c588de1976dee925__llikNormDmean_8X3U
#> #define _llikNorm _rx6c048fd240063677d85c588de1976dee926__llikNorm_0q3o
#> #define simeps _rx6c048fd240063677d85c588de1976dee927_simeps_XU1Y
#> #define simeta _rx6c048fd240063677d85c588de1976dee928_simeta_iEoQ
#> #define expit _rx6c048fd240063677d85c588de1976dee929_expit_Z9CA
#> #define logit _rx6c048fd240063677d85c588de1976dee930_logit_p6Gn
#> #define gammapDer _rx6c048fd240063677d85c588de1976dee931_gammapDer_Ir2L
#> #define lowergamma _rx6c048fd240063677d85c588de1976dee932_lowergamma_J8NS
#> #define uppergamma _rx6c048fd240063677d85c588de1976dee933_uppergamma_2zuy
#> #define gammaqInva _rx6c048fd240063677d85c588de1976dee934_gammaqInva_A28M
#> #define gammaqInv _rx6c048fd240063677d85c588de1976dee935_gammaqInv_XwdU
#> #define gammapInva _rx6c048fd240063677d85c588de1976dee936_gammapInva_5wLM
#> #define gammapInv _rx6c048fd240063677d85c588de1976dee937_gammapInv_cjCa
#> #define gammaq _rx6c048fd240063677d85c588de1976dee938_gammaq_qz0o
#> #define gammap _rx6c048fd240063677d85c588de1976dee939_gammap_C7gG
#> #define phi _rx6c048fd240063677d85c588de1976dee940_phi_pdW1
#> #define d2ELUa _rx6c048fd240063677d85c588de1976dee941_d2ELUa_H2t8
#> #define dELUa _rx6c048fd240063677d85c588de1976dee942_dELUa_3Mht
#> #define d2aELU _rx6c048fd240063677d85c588de1976dee943_d2aELU_FOcY
#> #define d2ELU _rx6c048fd240063677d85c588de1976dee944_d2ELU_2TwE
#> #define dELU _rx6c048fd240063677d85c588de1976dee945_dELU_gu28
#> #define ELU _rx6c048fd240063677d85c588de1976dee946_ELU_6PyM
#> #define dPReLUa1 _rx6c048fd240063677d85c588de1976dee947_dPReLUa1_ih1w
#> #define dPReLUa _rx6c048fd240063677d85c588de1976dee948_dPReLUa_Xh2u
#> #define dPReLU _rx6c048fd240063677d85c588de1976dee949_dPReLU_2mWG
#> #define PReLU _rx6c048fd240063677d85c588de1976dee950_PReLU_VX5u
#> #define dSwish _rx6c048fd240063677d85c588de1976dee951_dSwish_Y3fc
#> #define Swish _rx6c048fd240063677d85c588de1976dee952_Swish_4Rdo
#> #define dlReLU _rx6c048fd240063677d85c588de1976dee953_dlReLU_Olsc
#> #define lReLU _rx6c048fd240063677d85c588de1976dee954_lReLU_Mvk3
#> #define dSELU _rx6c048fd240063677d85c588de1976dee955_dSELU_6on3
#> #define SELU _rx6c048fd240063677d85c588de1976dee956_SELU_uitX
#> #define d4softplus _rx6c048fd240063677d85c588de1976dee957_d4softplus_LpnW
#> #define d3softplus _rx6c048fd240063677d85c588de1976dee958_d3softplus_m3S1
#> #define d2softplus _rx6c048fd240063677d85c588de1976dee959_d2softplus_jh0T
#> #define dsoftplus _rx6c048fd240063677d85c588de1976dee960_dsoftplus_aH3N
#> #define softplus _rx6c048fd240063677d85c588de1976dee961_softplus_WFwR
#> #define d4GELU _rx6c048fd240063677d85c588de1976dee962_d4GELU_hAF1
#> #define d3GELU _rx6c048fd240063677d85c588de1976dee963_d3GELU_KP6q
#> #define d2GELU _rx6c048fd240063677d85c588de1976dee964_d2GELU_0ynV
#> #define dGELU _rx6c048fd240063677d85c588de1976dee965_dGELU_vTVp
#> #define GELU _rx6c048fd240063677d85c588de1976dee966_GELU_9WQK
#> #define dReLU _rx6c048fd240063677d85c588de1976dee967_dReLU_R2F6
#> #define ReLU _rx6c048fd240063677d85c588de1976dee968_ReLU_MeQ4
#> #define riweibull _rx6c048fd240063677d85c588de1976dee969_riweibull_gqr5
#> #define riunif _rx6c048fd240063677d85c588de1976dee970_riunif_0QyU
#> #define rit_ _rx6c048fd240063677d85c588de1976dee971_rit__zU7Y
#> #define ripois _rx6c048fd240063677d85c588de1976dee972_ripois_EimR
#> #define ribeta _rx6c048fd240063677d85c588de1976dee973_ribeta_TOtY
#> #define rigamma _rx6c048fd240063677d85c588de1976dee974_rigamma_gK5w
#> #define rigeom _rx6c048fd240063677d85c588de1976dee975_rigeom_f07q
#> #define rif _rx6c048fd240063677d85c588de1976dee976_rif_ROQf
#> #define riexp _rx6c048fd240063677d85c588de1976dee977_riexp_Bpun
#> #define richisq _rx6c048fd240063677d85c588de1976dee978_richisq_VFfw
#> #define ricauchy _rx6c048fd240063677d85c588de1976dee979_ricauchy_y3Nn
#> #define rinbinomMu _rx6c048fd240063677d85c588de1976dee980_rinbinomMu_XIO1
#> #define rinbinom _rx6c048fd240063677d85c588de1976dee981_rinbinom_q1eb
#> #define ribinom _rx6c048fd240063677d85c588de1976dee982_ribinom_Jzpb
#> #define rinorm _rx6c048fd240063677d85c588de1976dee983_rinorm_8rZn
#> #define rxweibull _rx6c048fd240063677d85c588de1976dee984_rxweibull_1Fkp
#> #define rxunif _rx6c048fd240063677d85c588de1976dee985_rxunif_WdUa
#> #define rxt_ _rx6c048fd240063677d85c588de1976dee986_rxt__6qOl
#> #define rxpois _rx6c048fd240063677d85c588de1976dee987_rxpois_wGpe
#> #define rxbeta _rx6c048fd240063677d85c588de1976dee988_rxbeta_nBkj
#> #define rxgamma _rx6c048fd240063677d85c588de1976dee989_rxgamma_kK3f
#> #define rxgeom _rx6c048fd240063677d85c588de1976dee990_rxgeom_2dIO
#> #define rxf _rx6c048fd240063677d85c588de1976dee991_rxf_vgii
#> #define rxexp _rx6c048fd240063677d85c588de1976dee992_rxexp_EPZs
#> #define rxchisq _rx6c048fd240063677d85c588de1976dee993_rxchisq_auLF
#> #define rxcauchy _rx6c048fd240063677d85c588de1976dee994_rxcauchy_iDRh
#> #define rxnbinomMu _rx6c048fd240063677d85c588de1976dee995_rxnbinomMu_fBjK
#> #define rxnbinom _rx6c048fd240063677d85c588de1976dee996_rxnbinom_LmMH
#> #define rxbinom _rx6c048fd240063677d85c588de1976dee997_rxbinom_G1ek
#> #define rxnorm _rx6c048fd240063677d85c588de1976dee998_rxnorm_d3W5
#> #define linCmtB _rx6c048fd240063677d85c588de1976dee999_linCmtB_G0hn
#> #define linCmtA _rx6c048fd240063677d85c588de1976dee1000_linCmtA_8bMg
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){}
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){}
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  rx_bc75a7f04291176b28de80cf062f65ac__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *rx_bc75a7f04291176b28de80cf062f65ac__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP rx_bc75a7f04291176b28de80cf062f65ac__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void rx_bc75a7f04291176b28de80cf062f65ac__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double r1 = NA_REAL;
#>   double r2 = NA_REAL;
#>   double r2a = NA_REAL;
#>   double ra = NA_REAL;
#> 
#>   (void)t;
#>   (void)r1;
#>   (void)r2;
#>   (void)r2a;
#>   (void)ra;
#> 
#>   r1 = _PL[0];
#>   r2 = _PL[1];
#>   r2a = _PL[2];
#>   ra = _PL[3];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   r1=dELU(t,2);
#>   r2=d2ELU(t,2);
#>   r2a=d2aELU(t,2);
#>   ra=dELUa(t,2);
#>   r2a=d2ELUa(t,2);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void rx_bc75a7f04291176b28de80cf062f65ac__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void rx_bc75a7f04291176b28de80cf062f65ac__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void rx_bc75a7f04291176b28de80cf062f65ac__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double r1 = NA_REAL;
#>   double r2 = NA_REAL;
#>   double r2a = NA_REAL;
#>   double ra = NA_REAL;
#> 
#>   (void)t;
#>   (void)r1;
#>   (void)r2;
#>   (void)r2a;
#>   (void)ra;
#> 
#>   r1 = _PL[0];
#>   r2 = _PL[1];
#>   r2a = _PL[2];
#>   ra = _PL[3];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   r1=dELU(t,2);
#>   r2=d2ELU(t,2);
#>   r2a=d2aELU(t,2);
#>   ra=dELUa(t,2);
#>   r2a=d2ELUa(t,2);
#> 
#>   _lhs[_LHS_0_]=r1;
#>   _lhs[_LHS_1_]=r2;
#>   _lhs[_LHS_2_]=r2a;
#>   _lhs[_LHS_3_]=ra;
#> }
#> // Functional based bioavailability
#> double rx_bc75a7f04291176b28de80cf062f65ac__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double rx_bc75a7f04291176b28de80cf062f65ac__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double rx_bc75a7f04291176b28de80cf062f65ac__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double rx_bc75a7f04291176b28de80cf062f65ac__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void rx_bc75a7f04291176b28de80cf062f65ac__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void rx_bc75a7f04291176b28de80cf062f65ac__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void rx_bc75a7f04291176b28de80cf062f65ac__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP rx_bc75a7f04291176b28de80cf062f65ac__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_bc75a7f04291176b28de80cf062f65ac__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAAk_1AAAv7#aT)FpmeNAy>9,=WdwYB~1u%%C8AeMm\?BjM{+tx+,,\?>zq}@ceE}^gHG/&PnBTOWUMM!M_2V_4>N~6Je~mr_#qRxm5I{UTX#Tx(QA=6|fj:9I1+tBwe@TI3\?#>6b12Y[>rE>J#oJott\"!g[s\"#,Q|z6%%]x$;EcL$VV]98/rD>&BIWf@H{TZst|=c@t_rb:MSq!H4G\?bXq7@fHCx{G%%INg6EDIKlN9TC{0q*ZbKEGQ8*&GK`~`\">L0H`T);KR^V!TL#_;WtBzVh}:cRJXRD*I5yJ48+tK)(/y:/P~u%%4<zLu.\?9(S>7Zha|N{/_fP{oOQw\"ga9Ua};\?^>Nq1]tZoL6D6~&GLG05j,g[8Z/=eh<WuM3gPzfTMPEy\"g<(^ED9gfP#|HthO2%%%%*&z6Q\"AOOv=nuSDP/vX#u)Sw^2Q+Mh.{LKznA+BIULszPyo!m+:~AVqqCAA>:*gC]F]Ya,4^5*WWh$#|q*:PsD*l&O\?)U5M6#k>d@jeBf!3I;DM[|d=_+\"8{W7pW8e}B&{$>x*f8dn5gRmM9Nmzkl|O^8klk,.P71/\?<N47:Bdo8icqO:aa/)SGi:jDl6k#C+t>NlT*UR8DZIgW\"h:P!5gt!YxK;Dx}%%y=9EmQl_(\"fcXe(H:x;@WB`Vd|=W6QU#=*#j7/gUo_P_sS`^Is;0r>k>;P083f(a,`cQi<=~7W>e0w0%%d60#$3h~#TmmDT6x[7P)kTs&k)aKrtd^%%dtRn{14P1;OuC*a:oj$c8)@gg{O.U(dJdJMS0uCMF0FN\"/YATt44^dnCL\"L0@5K\"[FgG:MGGDDKYS$aF_GzA5X&r!m^*](P$/syc>Li22BGhV&MsyBiXZLY\?tmozWFq{C=xeq^T\"xi|Q$b{P4JlISv~AH0EG<M]yZrJ\?qttc9v/7y{yE\?+UA}Dp68x`,et0S{0F;j6`;)IsU`*7\?7bLX:+[+27N[c^@)|2#$ku39K\"@*AMF_V^ov_KTP|H*1fXWVWv6atU%%:v8\?h1(5VoE|wvt#i5K&HnY`/U`8z}OFu~@<YeO{)q%%`[,x[^M8~kt\?oMge$LIa^7A.R*1");
#>     char buf[1105];
#> #define __doBufN__ 1105
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void rx_bc75a7f04291176b28de80cf062f65ac__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   rx_bc75a7f04291176b28de80cf062f65ac__dydt(neq, *t, A, DADT);
#> }
#> extern int rx_bc75a7f04291176b28de80cf062f65ac__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   rx_bc75a7f04291176b28de80cf062f65ac__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void rx_bc75a7f04291176b28de80cf062f65ac__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   rx_bc75a7f04291176b28de80cf062f65ac__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void rx_bc75a7f04291176b28de80cf062f65ac__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_rx_bc75a7f04291176b28de80cf062f65ac_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("rx_bc75a7f04291176b28de80cf062f65ac_","rx_bc75a7f04291176b28de80cf062f65ac__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("rx_bc75a7f04291176b28de80cf062f65ac_","rx_bc75a7f04291176b28de80cf062f65ac__assignFuns", (DL_FUNC) rx_bc75a7f04291176b28de80cf062f65ac__assignFuns);
#>   R_RegisterCCallable("rx_bc75a7f04291176b28de80cf062f65ac_","rx_bc75a7f04291176b28de80cf062f65ac__inis",(DL_FUNC) rx_bc75a7f04291176b28de80cf062f65ac__inis);
#>   R_RegisterCCallable("rx_bc75a7f04291176b28de80cf062f65ac_","rx_bc75a7f04291176b28de80cf062f65ac__dydt",(DL_FUNC) rx_bc75a7f04291176b28de80cf062f65ac__dydt);
#>   R_RegisterCCallable("rx_bc75a7f04291176b28de80cf062f65ac_","rx_bc75a7f04291176b28de80cf062f65ac__calc_lhs",(DL_FUNC) rx_bc75a7f04291176b28de80cf062f65ac__calc_lhs);
#>   R_RegisterCCallable("rx_bc75a7f04291176b28de80cf062f65ac_","rx_bc75a7f04291176b28de80cf062f65ac__calc_jac",(DL_FUNC) rx_bc75a7f04291176b28de80cf062f65ac__calc_jac);
#>   R_RegisterCCallable("rx_bc75a7f04291176b28de80cf062f65ac_","rx_bc75a7f04291176b28de80cf062f65ac__dydt_lsoda", (DL_FUNC) rx_bc75a7f04291176b28de80cf062f65ac__dydt_lsoda);
#>   R_RegisterCCallable("rx_bc75a7f04291176b28de80cf062f65ac_","rx_bc75a7f04291176b28de80cf062f65ac__calc_jac_lsoda", (DL_FUNC) rx_bc75a7f04291176b28de80cf062f65ac__calc_jac_lsoda);
#>   R_RegisterCCallable("rx_bc75a7f04291176b28de80cf062f65ac_","rx_bc75a7f04291176b28de80cf062f65ac__ode_solver_solvedata", (DL_FUNC) rx_bc75a7f04291176b28de80cf062f65ac__ode_solver_solvedata);
#>   R_RegisterCCallable("rx_bc75a7f04291176b28de80cf062f65ac_","rx_bc75a7f04291176b28de80cf062f65ac__ode_solver_get_solvedata", (DL_FUNC) rx_bc75a7f04291176b28de80cf062f65ac__ode_solver_get_solvedata);
#>   R_RegisterCCallable("rx_bc75a7f04291176b28de80cf062f65ac_","rx_bc75a7f04291176b28de80cf062f65ac__F", (DL_FUNC) rx_bc75a7f04291176b28de80cf062f65ac__F);
#>   R_RegisterCCallable("rx_bc75a7f04291176b28de80cf062f65ac_","rx_bc75a7f04291176b28de80cf062f65ac__Lag", (DL_FUNC) rx_bc75a7f04291176b28de80cf062f65ac__Lag);
#>   R_RegisterCCallable("rx_bc75a7f04291176b28de80cf062f65ac_","rx_bc75a7f04291176b28de80cf062f65ac__Rate", (DL_FUNC) rx_bc75a7f04291176b28de80cf062f65ac__Rate);
#>   R_RegisterCCallable("rx_bc75a7f04291176b28de80cf062f65ac_","rx_bc75a7f04291176b28de80cf062f65ac__Dur", (DL_FUNC) rx_bc75a7f04291176b28de80cf062f65ac__Dur);
#>   R_RegisterCCallable("rx_bc75a7f04291176b28de80cf062f65ac_","rx_bc75a7f04291176b28de80cf062f65ac__mtime", (DL_FUNC) rx_bc75a7f04291176b28de80cf062f65ac__mtime);
#>   R_RegisterCCallable("rx_bc75a7f04291176b28de80cf062f65ac_","rx_bc75a7f04291176b28de80cf062f65ac__ME", (DL_FUNC) rx_bc75a7f04291176b28de80cf062f65ac__ME);
#>   R_RegisterCCallable("rx_bc75a7f04291176b28de80cf062f65ac_","rx_bc75a7f04291176b28de80cf062f65ac__IndF", (DL_FUNC) rx_bc75a7f04291176b28de80cf062f65ac__IndF);
#>   R_RegisterCCallable("rx_bc75a7f04291176b28de80cf062f65ac_","rx_bc75a7f04291176b28de80cf062f65ac__dydt_liblsoda", (DL_FUNC) rx_bc75a7f04291176b28de80cf062f65ac__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_rx_bc75a7f04291176b28de80cf062f65ac_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_rx_bc75a7f04291176b28de80cf062f65ac_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"rx_bc75a7f04291176b28de80cf062f65ac__model_vars", (DL_FUNC) &rx_bc75a7f04291176b28de80cf062f65ac__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_rx_bc75a7f04291176b28de80cf062f65ac_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_bc75a7f04291176b28de80cf062f65ac__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("rx_bc75a7f04291176b28de80cf062f65ac__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’

e <- et(c(-1, 0, 1, 2))
rxSolve(r, e)
#> ── Solved rxode2 object ──
#> ── Parameters (value$params): ──
#> # A tibble: 1 × 0
#> ── Initial Conditions (value$inits): ──
#> named numeric(0)
#> ── First part of data (object): ──
#> # A tibble: 4 × 5
#>    time    r1    r2   r2a     ra
#>   <dbl> <dbl> <dbl> <dbl>  <dbl>
#> 1    -1 0.736 0.736 0.368 -0.632
#> 2     0 2     2     1      0    
#> 3     1 1     0     0      0    
#> 4     2 1     0     0      0