Derivatives of GELU

Derivatives of GELU

Usage

dGELU(x)

d2GELU(x)

d3GELU(x)

d4GELU(x)

Arguments

x

numeric vector

Value

numeric vector

See also

Examples

dGELU(c(-2, -1, 0, 1, 2))
#> [1] -0.08523180 -0.08331547  0.50000000  1.08331547  1.08523180
d2GELU(c(-2, -1, 0, 1, 2))
#> [1] -0.1079819  0.2419707  0.7978846  0.2419707 -0.1079819
d3GELU(c(-2, -1, 0, 1, 2))
#> [1]  0.0000000  0.7259122  0.0000000 -0.7259122  0.0000000
d4GELU(c(-2, -1, 0, 1, 2))
#> [1]  0.4319277  0.4839414 -1.5957691  0.4839414  0.4319277
# you can use rxode2 as well
r <- rxode2({
   r1 <- dGELU(time)
   r2 <- d2GELU(time)
   r3 <- d3GELU(time)
   r4 <- d4GELU(time)
})
#>  
#>  
#> C model:
#> // Define 4 LHS values
#> #define _LHS_0_ 0 // r1
#> #define _LHS_1_ 1 // r2
#> #define _LHS_2_ 2 // r3
#> #define _LHS_3_ 3 // r4
#> #define _getRxSolve_ _rx0e10af87d090fb5e386a8254e0eb885d1001__getRxSolve__xyu5
#> #define _evalUdf _rx0e10af87d090fb5e386a8254e0eb885d1002__evalUdf_TLfd
#> #define _solveData _rx0e10af87d090fb5e386a8254e0eb885d1003__solveData_rwZX
#> #define _assign_ptr _rx0e10af87d090fb5e386a8254e0eb885d1004__assign_ptr_wtP3
#> #define _rxRmModelLib _rx0e10af87d090fb5e386a8254e0eb885d1005__rxRmModelLib_mvNF
#> #define _rxGetModelLib _rx0e10af87d090fb5e386a8254e0eb885d1006__rxGetModelLib_5OB8
#> #define _old_c _rx0e10af87d090fb5e386a8254e0eb885d1007__old_c_FY9x
#> #define _ptrid _rx0e10af87d090fb5e386a8254e0eb885d1008__ptrid_wzh8
#> #define _rxIsCurrentC _rx0e10af87d090fb5e386a8254e0eb885d1009__rxIsCurrentC_0ggu
#> #define _sumPS _rx0e10af87d090fb5e386a8254e0eb885d1010__sumPS_PQXP
#> #define _prodPS _rx0e10af87d090fb5e386a8254e0eb885d1011__prodPS_6NCi
#> #define _prodType _rx0e10af87d090fb5e386a8254e0eb885d1012__prodType_0utk
#> #define _sumType _rx0e10af87d090fb5e386a8254e0eb885d1013__sumType_xceN
#> #define _update_par_ptr _rx0e10af87d090fb5e386a8254e0eb885d1014__update_par_ptr_4g7I
#> #define _getParCov _rx0e10af87d090fb5e386a8254e0eb885d1015__getParCov_dO4r
#> #define _rxode2_rxAssignPtr _rx0e10af87d090fb5e386a8254e0eb885d1016__rxode2_rxAssignPtr_Nf1a
#> #define _rxQr _rx0e10af87d090fb5e386a8254e0eb885d1017__rxQr_HO9p
#> #define _compareFactorVal _rx0e10af87d090fb5e386a8254e0eb885d1018__compareFactorVal_8utB
#> #define _sum _rx0e10af87d090fb5e386a8254e0eb885d1019__sum_XW9W
#> #define _udf _rx0e10af87d090fb5e386a8254e0eb885d1020__udf_DHV9
#> #define _sign _rx0e10af87d090fb5e386a8254e0eb885d1021__sign_vUck
#> #define _prod _rx0e10af87d090fb5e386a8254e0eb885d1022__prod_cv41
#> #define _max _rx0e10af87d090fb5e386a8254e0eb885d1023__max_XEdN
#> #define _min _rx0e10af87d090fb5e386a8254e0eb885d1024__min_22xE
#> #define _transit4P _rx0e10af87d090fb5e386a8254e0eb885d1025__transit4P_0D3E
#> #define _transit3P _rx0e10af87d090fb5e386a8254e0eb885d1026__transit3P_O0Ma
#> #define _assignFuns0 _rx0e10af87d090fb5e386a8254e0eb885d1027__assignFuns0_bGfw
#> #define _assignFuns _rx0e10af87d090fb5e386a8254e0eb885d1028__assignFuns_0mo4
#> #define _rxord _rx0e10af87d090fb5e386a8254e0eb885d1029__rxord_ZX8H
#> #define __assignFuns2 _rx0e10af87d090fb5e386a8254e0eb885d1030___assignFuns2_LoAW
#> #define _llikCauchyDscale _rx0e10af87d090fb5e386a8254e0eb885d1031__llikCauchyDscale_lHRf
#> #define _llikCauchyDlocation _rx0e10af87d090fb5e386a8254e0eb885d1032__llikCauchyDlocation_2YvQ
#> #define _llikCauchy _rx0e10af87d090fb5e386a8254e0eb885d1033__llikCauchy_wyfI
#> #define _llikGammaDrate _rx0e10af87d090fb5e386a8254e0eb885d1034__llikGammaDrate_CJq3
#> #define _llikGammaDshape _rx0e10af87d090fb5e386a8254e0eb885d1035__llikGammaDshape_BTKo
#> #define _llikGamma _rx0e10af87d090fb5e386a8254e0eb885d1036__llikGamma_r1pO
#> #define _llikWeibullDscale _rx0e10af87d090fb5e386a8254e0eb885d1037__llikWeibullDscale_f7Tv
#> #define _llikWeibullDshape _rx0e10af87d090fb5e386a8254e0eb885d1038__llikWeibullDshape_Zukd
#> #define _llikWeibull _rx0e10af87d090fb5e386a8254e0eb885d1039__llikWeibull_nvAB
#> #define _llikUnifDbeta _rx0e10af87d090fb5e386a8254e0eb885d1040__llikUnifDbeta_6pnH
#> #define _llikUnifDalpha _rx0e10af87d090fb5e386a8254e0eb885d1041__llikUnifDalpha_RcMN
#> #define _llikUnif _rx0e10af87d090fb5e386a8254e0eb885d1042__llikUnif_uO4h
#> #define _llikGeomDp _rx0e10af87d090fb5e386a8254e0eb885d1043__llikGeomDp_g52O
#> #define _llikGeom _rx0e10af87d090fb5e386a8254e0eb885d1044__llikGeom_0ZwL
#> #define _llikFDdf2 _rx0e10af87d090fb5e386a8254e0eb885d1045__llikFDdf2_FX7N
#> #define _llikFDdf1 _rx0e10af87d090fb5e386a8254e0eb885d1046__llikFDdf1_Kwul
#> #define _llikF _rx0e10af87d090fb5e386a8254e0eb885d1047__llikF_lzZD
#> #define _llikExpDrate _rx0e10af87d090fb5e386a8254e0eb885d1048__llikExpDrate_8nyy
#> #define _llikExp _rx0e10af87d090fb5e386a8254e0eb885d1049__llikExp_7cr8
#> #define _llikChisqDdf _rx0e10af87d090fb5e386a8254e0eb885d1050__llikChisqDdf_M9RU
#> #define _llikChisq _rx0e10af87d090fb5e386a8254e0eb885d1051__llikChisq_wuwl
#> #define _llikTDsd _rx0e10af87d090fb5e386a8254e0eb885d1052__llikTDsd_51KE
#> #define _llikTDmean _rx0e10af87d090fb5e386a8254e0eb885d1053__llikTDmean_nUV4
#> #define _llikTDdf _rx0e10af87d090fb5e386a8254e0eb885d1054__llikTDdf_LkN0
#> #define _llikT _rx0e10af87d090fb5e386a8254e0eb885d1055__llikT_iY4L
#> #define _llikBetaDshape2 _rx0e10af87d090fb5e386a8254e0eb885d1056__llikBetaDshape2_PJjO
#> #define _llikBetaDshape1 _rx0e10af87d090fb5e386a8254e0eb885d1057__llikBetaDshape1_ccne
#> #define _llikBeta _rx0e10af87d090fb5e386a8254e0eb885d1058__llikBeta_gPf1
#> #define _llikNbinomMuDmu _rx0e10af87d090fb5e386a8254e0eb885d1059__llikNbinomMuDmu_o4t4
#> #define _llikNbinomMu _rx0e10af87d090fb5e386a8254e0eb885d1060__llikNbinomMu_BmCP
#> #define _llikNbinomDprob _rx0e10af87d090fb5e386a8254e0eb885d1061__llikNbinomDprob_RbZ3
#> #define _llikNbinom _rx0e10af87d090fb5e386a8254e0eb885d1062__llikNbinom_BINi
#> #define _llikBinomDprob _rx0e10af87d090fb5e386a8254e0eb885d1063__llikBinomDprob_7Q1S
#> #define _llikBinom _rx0e10af87d090fb5e386a8254e0eb885d1064__llikBinom_nTEs
#> #define _llikPoisDlambda _rx0e10af87d090fb5e386a8254e0eb885d1065__llikPoisDlambda_7Gsk
#> #define _llikPois _rx0e10af87d090fb5e386a8254e0eb885d1066__llikPois_6or7
#> #define _llikNormDsd _rx0e10af87d090fb5e386a8254e0eb885d1067__llikNormDsd_8X3U
#> #define _llikNormDmean _rx0e10af87d090fb5e386a8254e0eb885d1068__llikNormDmean_0q3o
#> #define _llikNorm _rx0e10af87d090fb5e386a8254e0eb885d1069__llikNorm_XU1Y
#> #define simeps _rx0e10af87d090fb5e386a8254e0eb885d1070_simeps_iEoQ
#> #define simeta _rx0e10af87d090fb5e386a8254e0eb885d1071_simeta_Z9CA
#> #define expit _rx0e10af87d090fb5e386a8254e0eb885d1072_expit_p6Gn
#> #define logit _rx0e10af87d090fb5e386a8254e0eb885d1073_logit_Ir2L
#> #define gammapDer _rx0e10af87d090fb5e386a8254e0eb885d1074_gammapDer_J8NS
#> #define lowergamma _rx0e10af87d090fb5e386a8254e0eb885d1075_lowergamma_2zuy
#> #define uppergamma _rx0e10af87d090fb5e386a8254e0eb885d1076_uppergamma_A28M
#> #define gammaqInva _rx0e10af87d090fb5e386a8254e0eb885d1077_gammaqInva_XwdU
#> #define gammaqInv _rx0e10af87d090fb5e386a8254e0eb885d1078_gammaqInv_5wLM
#> #define gammapInva _rx0e10af87d090fb5e386a8254e0eb885d1079_gammapInva_cjCa
#> #define gammapInv _rx0e10af87d090fb5e386a8254e0eb885d1080_gammapInv_qz0o
#> #define gammaq _rx0e10af87d090fb5e386a8254e0eb885d1081_gammaq_C7gG
#> #define gammap _rx0e10af87d090fb5e386a8254e0eb885d1082_gammap_pdW1
#> #define phi _rx0e10af87d090fb5e386a8254e0eb885d1083_phi_H2t8
#> #define d2ELUa _rx0e10af87d090fb5e386a8254e0eb885d1084_d2ELUa_3Mht
#> #define dELUa _rx0e10af87d090fb5e386a8254e0eb885d1085_dELUa_FOcY
#> #define d2aELU _rx0e10af87d090fb5e386a8254e0eb885d1086_d2aELU_2TwE
#> #define d2ELU _rx0e10af87d090fb5e386a8254e0eb885d1087_d2ELU_gu28
#> #define dELU _rx0e10af87d090fb5e386a8254e0eb885d1088_dELU_6PyM
#> #define ELU _rx0e10af87d090fb5e386a8254e0eb885d1089_ELU_ih1w
#> #define dPReLUa1 _rx0e10af87d090fb5e386a8254e0eb885d1090_dPReLUa1_Xh2u
#> #define dPReLUa _rx0e10af87d090fb5e386a8254e0eb885d1091_dPReLUa_2mWG
#> #define dPReLU _rx0e10af87d090fb5e386a8254e0eb885d1092_dPReLU_VX5u
#> #define PReLU _rx0e10af87d090fb5e386a8254e0eb885d1093_PReLU_Y3fc
#> #define dSwish _rx0e10af87d090fb5e386a8254e0eb885d1094_dSwish_4Rdo
#> #define Swish _rx0e10af87d090fb5e386a8254e0eb885d1095_Swish_Olsc
#> #define dlReLU _rx0e10af87d090fb5e386a8254e0eb885d1096_dlReLU_Mvk3
#> #define lReLU _rx0e10af87d090fb5e386a8254e0eb885d1097_lReLU_6on3
#> #define dSELU _rx0e10af87d090fb5e386a8254e0eb885d1098_dSELU_uitX
#> #define SELU _rx0e10af87d090fb5e386a8254e0eb885d1099_SELU_LpnW
#> #define d4softplus _rx0e10af87d090fb5e386a8254e0eb885d1100_d4softplus_m3S1
#> #define d3softplus _rx0e10af87d090fb5e386a8254e0eb885d1101_d3softplus_jh0T
#> #define d2softplus _rx0e10af87d090fb5e386a8254e0eb885d1102_d2softplus_aH3N
#> #define dsoftplus _rx0e10af87d090fb5e386a8254e0eb885d1103_dsoftplus_WFwR
#> #define softplus _rx0e10af87d090fb5e386a8254e0eb885d1104_softplus_hAF1
#> #define d4GELU _rx0e10af87d090fb5e386a8254e0eb885d1105_d4GELU_KP6q
#> #define d3GELU _rx0e10af87d090fb5e386a8254e0eb885d1106_d3GELU_0ynV
#> #define d2GELU _rx0e10af87d090fb5e386a8254e0eb885d1107_d2GELU_vTVp
#> #define dGELU _rx0e10af87d090fb5e386a8254e0eb885d1108_dGELU_9WQK
#> #define GELU _rx0e10af87d090fb5e386a8254e0eb885d1109_GELU_R2F6
#> #define dReLU _rx0e10af87d090fb5e386a8254e0eb885d1110_dReLU_MeQ4
#> #define ReLU _rx0e10af87d090fb5e386a8254e0eb885d1111_ReLU_gqr5
#> #define riweibull _rx0e10af87d090fb5e386a8254e0eb885d1112_riweibull_0QyU
#> #define riunif _rx0e10af87d090fb5e386a8254e0eb885d1113_riunif_zU7Y
#> #define rit_ _rx0e10af87d090fb5e386a8254e0eb885d1114_rit__EimR
#> #define ripois _rx0e10af87d090fb5e386a8254e0eb885d1115_ripois_TOtY
#> #define ribeta _rx0e10af87d090fb5e386a8254e0eb885d1116_ribeta_gK5w
#> #define rigamma _rx0e10af87d090fb5e386a8254e0eb885d1117_rigamma_f07q
#> #define rigeom _rx0e10af87d090fb5e386a8254e0eb885d1118_rigeom_ROQf
#> #define rif _rx0e10af87d090fb5e386a8254e0eb885d1119_rif_Bpun
#> #define riexp _rx0e10af87d090fb5e386a8254e0eb885d1120_riexp_VFfw
#> #define richisq _rx0e10af87d090fb5e386a8254e0eb885d1121_richisq_y3Nn
#> #define ricauchy _rx0e10af87d090fb5e386a8254e0eb885d1122_ricauchy_XIO1
#> #define rinbinomMu _rx0e10af87d090fb5e386a8254e0eb885d1123_rinbinomMu_q1eb
#> #define rinbinom _rx0e10af87d090fb5e386a8254e0eb885d1124_rinbinom_Jzpb
#> #define ribinom _rx0e10af87d090fb5e386a8254e0eb885d1125_ribinom_8rZn
#> #define rinorm _rx0e10af87d090fb5e386a8254e0eb885d1126_rinorm_1Fkp
#> #define rxweibull _rx0e10af87d090fb5e386a8254e0eb885d1127_rxweibull_WdUa
#> #define rxunif _rx0e10af87d090fb5e386a8254e0eb885d1128_rxunif_6qOl
#> #define rxt_ _rx0e10af87d090fb5e386a8254e0eb885d1129_rxt__wGpe
#> #define rxpois _rx0e10af87d090fb5e386a8254e0eb885d1130_rxpois_nBkj
#> #define rxbeta _rx0e10af87d090fb5e386a8254e0eb885d1131_rxbeta_kK3f
#> #define rxgamma _rx0e10af87d090fb5e386a8254e0eb885d1132_rxgamma_2dIO
#> #define rxgeom _rx0e10af87d090fb5e386a8254e0eb885d1133_rxgeom_vgii
#> #define rxf _rx0e10af87d090fb5e386a8254e0eb885d1134_rxf_EPZs
#> #define rxexp _rx0e10af87d090fb5e386a8254e0eb885d1135_rxexp_auLF
#> #define rxchisq _rx0e10af87d090fb5e386a8254e0eb885d1136_rxchisq_iDRh
#> #define rxcauchy _rx0e10af87d090fb5e386a8254e0eb885d1137_rxcauchy_fBjK
#> #define rxnbinomMu _rx0e10af87d090fb5e386a8254e0eb885d1138_rxnbinomMu_LmMH
#> #define rxnbinom _rx0e10af87d090fb5e386a8254e0eb885d1139_rxnbinom_G1ek
#> #define rxbinom _rx0e10af87d090fb5e386a8254e0eb885d1140_rxbinom_d3W5
#> #define rxnorm _rx0e10af87d090fb5e386a8254e0eb885d1141_rxnorm_G0hn
#> #define linCmtB _rx0e10af87d090fb5e386a8254e0eb885d1142_linCmtB_8bMg
#> #define linCmtA _rx0e10af87d090fb5e386a8254e0eb885d1143_linCmtA_MnzT
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){}
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){}
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  rx_1366a8fd7c7f2bc5046d3557874abd74__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *rx_1366a8fd7c7f2bc5046d3557874abd74__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP rx_1366a8fd7c7f2bc5046d3557874abd74__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void rx_1366a8fd7c7f2bc5046d3557874abd74__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double r1 = NA_REAL;
#>   double r2 = NA_REAL;
#>   double r3 = NA_REAL;
#>   double r4 = NA_REAL;
#> 
#>   (void)t;
#>   (void)r1;
#>   (void)r2;
#>   (void)r3;
#>   (void)r4;
#> 
#>   r1 = _PL[0];
#>   r2 = _PL[1];
#>   r3 = _PL[2];
#>   r4 = _PL[3];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   r1 =dGELU(t);
#>   r2 =d2GELU(t);
#>   r3 =d3GELU(t);
#>   r4 =d4GELU(t);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void rx_1366a8fd7c7f2bc5046d3557874abd74__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void rx_1366a8fd7c7f2bc5046d3557874abd74__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void rx_1366a8fd7c7f2bc5046d3557874abd74__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double r1 = NA_REAL;
#>   double r2 = NA_REAL;
#>   double r3 = NA_REAL;
#>   double r4 = NA_REAL;
#> 
#>   (void)t;
#>   (void)r1;
#>   (void)r2;
#>   (void)r3;
#>   (void)r4;
#> 
#>   r1 = _PL[0];
#>   r2 = _PL[1];
#>   r3 = _PL[2];
#>   r4 = _PL[3];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   r1 =dGELU(t);
#>   r2 =d2GELU(t);
#>   r3 =d3GELU(t);
#>   r4 =d4GELU(t);
#> 
#>   _lhs[_LHS_0_]=r1;
#>   _lhs[_LHS_1_]=r2;
#>   _lhs[_LHS_2_]=r3;
#>   _lhs[_LHS_3_]=r4;
#> }
#> // Functional based bioavailability
#> double rx_1366a8fd7c7f2bc5046d3557874abd74__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double rx_1366a8fd7c7f2bc5046d3557874abd74__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double rx_1366a8fd7c7f2bc5046d3557874abd74__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double rx_1366a8fd7c7f2bc5046d3557874abd74__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void rx_1366a8fd7c7f2bc5046d3557874abd74__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void rx_1366a8fd7c7f2bc5046d3557874abd74__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void rx_1366a8fd7c7f2bc5046d3557874abd74__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP rx_1366a8fd7c7f2bc5046d3557874abd74__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_1366a8fd7c7f2bc5046d3557874abd74__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAAv(1AAAv7#aT)soSaNAS>8ah\"T*6t1)Fhoe1w@hD_FJe]+I7ZA!y:cYvwN*rX|M5uk|dxRra\"cx$}#%%wmHc[O6v~~f1%%em80KNz*Ix(RA(AW)$OIdIp]BwP\"yXP22Gc`fnd7a34q_\?0qyv!im=nO<[s3Z#,Pq\?,Q$H_~1qG:X3<98/rI\?NYOs!YZ:]LdY|B;U_uz%%.^6W5^$[*5UOp$y.(ykZUJO.;SLT3PljSx5,My^@T:T:dT>@!85bWW4D9c&{dj5CIqd`{o51cM)a=4QsRC[J\"g9U=g>1&rH+toA#:lSvp)pQDi4gk&QDFql/_t@X\"oHd8)<GlL+e75You{qr`E2uVoK58\?9/vcb7:il@g87j_|6l<I%%k&x&8^+;#C<LA}Lu;nnT\?<85@QznRL5{uhGjFQup6],IjzA\";bb=1*ZA)&lnZ(`G7I+5_!uG3%%\?j*/SzjxBc\?vWVBBtb7{jeOY&:RD7s7W9Gf5qx)T]HWriV>iqj3iE;8p\?L+K8D!fV,c}yO+w/_N#0|B|+L6Z^z_O}vVDl3cOx\"8>u:2TYI];7.8I]vPv8si\"w4Vm69IX,pwB)zaS;ALm0TH1w;LCYGTaGbYNM:kb:b),}Ps<u(3l%%\"}`hAH6MN`fIB\?,#8spMp&1ec,Rux||zifi:)!xq354GsVkX0a{(<JSgd7_^R&=%%]f`aw\"^)Gc\"cA$5^IE!9*+9|vj!|1O\?@ta\"tsFO,s~yG!pw[;MwSL+J6T8o7p<~n%%,wfum#,~,Y\?Bh5g@,gyjQ<Y8Y~ve/)Mvz[5WJF\"bLq63jKX,h]+`cHHoLN\?/t@WuZEMK=6C@G@~JzUTk_s]T4kx[s!c<CM<R=O5IW8cwBtu\"E4+419\?9JcW(giXt:4hH1H\?pmal*OZ!+NqfJ*6X>v*UeKlVdd:ik@XeFw%%n7ant;C*(NMmYiU6S:5qh2t.;LEPbuK{XEwj@_e/=T!p\?I.2bkBmgOC^SaKg;*itT1mbfrKqCIDo3yOyQgw,cI*rtd\"\?aH3,0G%%;iVJ3gn7H;h[!z4ezHv)fvbHE}Rc\"oyxpL\"&\"[u\?`FG=pEEN|dH7[+%%L");
#>     char buf[1102];
#> #define __doBufN__ 1102
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void rx_1366a8fd7c7f2bc5046d3557874abd74__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   rx_1366a8fd7c7f2bc5046d3557874abd74__dydt(neq, *t, A, DADT);
#> }
#> extern int rx_1366a8fd7c7f2bc5046d3557874abd74__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   rx_1366a8fd7c7f2bc5046d3557874abd74__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void rx_1366a8fd7c7f2bc5046d3557874abd74__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   rx_1366a8fd7c7f2bc5046d3557874abd74__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void rx_1366a8fd7c7f2bc5046d3557874abd74__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_rx_1366a8fd7c7f2bc5046d3557874abd74_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("rx_1366a8fd7c7f2bc5046d3557874abd74_","rx_1366a8fd7c7f2bc5046d3557874abd74__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("rx_1366a8fd7c7f2bc5046d3557874abd74_","rx_1366a8fd7c7f2bc5046d3557874abd74__assignFuns", (DL_FUNC) rx_1366a8fd7c7f2bc5046d3557874abd74__assignFuns);
#>   R_RegisterCCallable("rx_1366a8fd7c7f2bc5046d3557874abd74_","rx_1366a8fd7c7f2bc5046d3557874abd74__inis",(DL_FUNC) rx_1366a8fd7c7f2bc5046d3557874abd74__inis);
#>   R_RegisterCCallable("rx_1366a8fd7c7f2bc5046d3557874abd74_","rx_1366a8fd7c7f2bc5046d3557874abd74__dydt",(DL_FUNC) rx_1366a8fd7c7f2bc5046d3557874abd74__dydt);
#>   R_RegisterCCallable("rx_1366a8fd7c7f2bc5046d3557874abd74_","rx_1366a8fd7c7f2bc5046d3557874abd74__calc_lhs",(DL_FUNC) rx_1366a8fd7c7f2bc5046d3557874abd74__calc_lhs);
#>   R_RegisterCCallable("rx_1366a8fd7c7f2bc5046d3557874abd74_","rx_1366a8fd7c7f2bc5046d3557874abd74__calc_jac",(DL_FUNC) rx_1366a8fd7c7f2bc5046d3557874abd74__calc_jac);
#>   R_RegisterCCallable("rx_1366a8fd7c7f2bc5046d3557874abd74_","rx_1366a8fd7c7f2bc5046d3557874abd74__dydt_lsoda", (DL_FUNC) rx_1366a8fd7c7f2bc5046d3557874abd74__dydt_lsoda);
#>   R_RegisterCCallable("rx_1366a8fd7c7f2bc5046d3557874abd74_","rx_1366a8fd7c7f2bc5046d3557874abd74__calc_jac_lsoda", (DL_FUNC) rx_1366a8fd7c7f2bc5046d3557874abd74__calc_jac_lsoda);
#>   R_RegisterCCallable("rx_1366a8fd7c7f2bc5046d3557874abd74_","rx_1366a8fd7c7f2bc5046d3557874abd74__ode_solver_solvedata", (DL_FUNC) rx_1366a8fd7c7f2bc5046d3557874abd74__ode_solver_solvedata);
#>   R_RegisterCCallable("rx_1366a8fd7c7f2bc5046d3557874abd74_","rx_1366a8fd7c7f2bc5046d3557874abd74__ode_solver_get_solvedata", (DL_FUNC) rx_1366a8fd7c7f2bc5046d3557874abd74__ode_solver_get_solvedata);
#>   R_RegisterCCallable("rx_1366a8fd7c7f2bc5046d3557874abd74_","rx_1366a8fd7c7f2bc5046d3557874abd74__F", (DL_FUNC) rx_1366a8fd7c7f2bc5046d3557874abd74__F);
#>   R_RegisterCCallable("rx_1366a8fd7c7f2bc5046d3557874abd74_","rx_1366a8fd7c7f2bc5046d3557874abd74__Lag", (DL_FUNC) rx_1366a8fd7c7f2bc5046d3557874abd74__Lag);
#>   R_RegisterCCallable("rx_1366a8fd7c7f2bc5046d3557874abd74_","rx_1366a8fd7c7f2bc5046d3557874abd74__Rate", (DL_FUNC) rx_1366a8fd7c7f2bc5046d3557874abd74__Rate);
#>   R_RegisterCCallable("rx_1366a8fd7c7f2bc5046d3557874abd74_","rx_1366a8fd7c7f2bc5046d3557874abd74__Dur", (DL_FUNC) rx_1366a8fd7c7f2bc5046d3557874abd74__Dur);
#>   R_RegisterCCallable("rx_1366a8fd7c7f2bc5046d3557874abd74_","rx_1366a8fd7c7f2bc5046d3557874abd74__mtime", (DL_FUNC) rx_1366a8fd7c7f2bc5046d3557874abd74__mtime);
#>   R_RegisterCCallable("rx_1366a8fd7c7f2bc5046d3557874abd74_","rx_1366a8fd7c7f2bc5046d3557874abd74__ME", (DL_FUNC) rx_1366a8fd7c7f2bc5046d3557874abd74__ME);
#>   R_RegisterCCallable("rx_1366a8fd7c7f2bc5046d3557874abd74_","rx_1366a8fd7c7f2bc5046d3557874abd74__IndF", (DL_FUNC) rx_1366a8fd7c7f2bc5046d3557874abd74__IndF);
#>   R_RegisterCCallable("rx_1366a8fd7c7f2bc5046d3557874abd74_","rx_1366a8fd7c7f2bc5046d3557874abd74__dydt_liblsoda", (DL_FUNC) rx_1366a8fd7c7f2bc5046d3557874abd74__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_rx_1366a8fd7c7f2bc5046d3557874abd74_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_rx_1366a8fd7c7f2bc5046d3557874abd74_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"rx_1366a8fd7c7f2bc5046d3557874abd74__model_vars", (DL_FUNC) &rx_1366a8fd7c7f2bc5046d3557874abd74__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_rx_1366a8fd7c7f2bc5046d3557874abd74_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_1366a8fd7c7f2bc5046d3557874abd74__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("rx_1366a8fd7c7f2bc5046d3557874abd74__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’
et <- et(c(-2, -1, 0, 1, 2))
rxSolve(r, et)
#> ── Solved rxode2 object ──
#> ── Parameters (value$params): ──
#> # A tibble: 1 × 0
#> ── Initial Conditions (value$inits): ──
#> named numeric(0)
#> ── First part of data (object): ──
#> # A tibble: 5 × 5
#>    time      r1     r2     r3     r4
#>   <dbl>   <dbl>  <dbl>  <dbl>  <dbl>
#> 1    -2 -0.0852 -0.108  0      0.432
#> 2    -1 -0.0833  0.242  0.726  0.484
#> 3     0  0.5     0.798  0     -1.60 
#> 4     1  1.08    0.242 -0.726  0.484
#> 5     2  1.09   -0.108  0      0.432