Derivatives Parametric ReLU Activation Function

Derivatives Parametric ReLU Activation Function

Usage

dPReLU(x, alpha = 1)

dPReLUa(x, alpha = 1)

dPReLUa1(x, alpha = 1)

Arguments

x

A numeric vector. All elements must be finite and non-missing.

alpha

A numeric scalar. All elements must be finite and non-missing.

Value

A numeric vector where the derivative(s) of the ELU function has been applied to each element of x.

See also

Author

Matthew L. Fidler

Examples

dPReLU(c(-1, 0, 1, 2), 2)
#> [1] 2 2 1 1
dPReLUa(c(-1, 0, 1, 2), 2)
#> [1] -1  0  0  0
dPReLUa1(c(-1, 0, 1, 2), 2)
#> [1] 1 1 0 0

# Can also be used in rxode2:
r <- rxode2({
  r1=dPReLU(time, 2)
  r2a=dPReLUa(time, 2)
  ra=dPReLUa1(time, 2)
})
#>  
#>  
#> C model:
#> // Define 3 LHS values
#> #define _LHS_0_ 0 // r1
#> #define _LHS_1_ 1 // r2a
#> #define _LHS_2_ 2 // ra
#> #define _getRxSolve_ _rx4c23a771624566473b9134f1fe585f171144__getRxSolve__TLfd
#> #define _evalUdf _rx4c23a771624566473b9134f1fe585f171145__evalUdf_rwZX
#> #define _solveData _rx4c23a771624566473b9134f1fe585f171146__solveData_wtP3
#> #define _assign_ptr _rx4c23a771624566473b9134f1fe585f171147__assign_ptr_mvNF
#> #define _rxRmModelLib _rx4c23a771624566473b9134f1fe585f171148__rxRmModelLib_5OB8
#> #define _rxGetModelLib _rx4c23a771624566473b9134f1fe585f171149__rxGetModelLib_FY9x
#> #define _old_c _rx4c23a771624566473b9134f1fe585f171150__old_c_wzh8
#> #define _ptrid _rx4c23a771624566473b9134f1fe585f171151__ptrid_0ggu
#> #define _rxIsCurrentC _rx4c23a771624566473b9134f1fe585f171152__rxIsCurrentC_PQXP
#> #define _sumPS _rx4c23a771624566473b9134f1fe585f171153__sumPS_6NCi
#> #define _prodPS _rx4c23a771624566473b9134f1fe585f171154__prodPS_0utk
#> #define _prodType _rx4c23a771624566473b9134f1fe585f171155__prodType_xceN
#> #define _sumType _rx4c23a771624566473b9134f1fe585f171156__sumType_4g7I
#> #define _update_par_ptr _rx4c23a771624566473b9134f1fe585f171157__update_par_ptr_dO4r
#> #define _getParCov _rx4c23a771624566473b9134f1fe585f171158__getParCov_Nf1a
#> #define _rxode2_rxAssignPtr _rx4c23a771624566473b9134f1fe585f171159__rxode2_rxAssignPtr_HO9p
#> #define _rxQr _rx4c23a771624566473b9134f1fe585f171160__rxQr_8utB
#> #define _compareFactorVal _rx4c23a771624566473b9134f1fe585f171161__compareFactorVal_XW9W
#> #define _sum _rx4c23a771624566473b9134f1fe585f171162__sum_DHV9
#> #define _udf _rx4c23a771624566473b9134f1fe585f171163__udf_vUck
#> #define _sign _rx4c23a771624566473b9134f1fe585f171164__sign_cv41
#> #define _prod _rx4c23a771624566473b9134f1fe585f171165__prod_XEdN
#> #define _max _rx4c23a771624566473b9134f1fe585f171166__max_22xE
#> #define _min _rx4c23a771624566473b9134f1fe585f171167__min_0D3E
#> #define _transit4P _rx4c23a771624566473b9134f1fe585f171168__transit4P_O0Ma
#> #define _transit3P _rx4c23a771624566473b9134f1fe585f171169__transit3P_bGfw
#> #define _assignFuns0 _rx4c23a771624566473b9134f1fe585f171170__assignFuns0_0mo4
#> #define _assignFuns _rx4c23a771624566473b9134f1fe585f171171__assignFuns_ZX8H
#> #define _rxord _rx4c23a771624566473b9134f1fe585f171172__rxord_LoAW
#> #define __assignFuns2 _rx4c23a771624566473b9134f1fe585f171173___assignFuns2_lHRf
#> #define _llikCauchyDscale _rx4c23a771624566473b9134f1fe585f171174__llikCauchyDscale_2YvQ
#> #define _llikCauchyDlocation _rx4c23a771624566473b9134f1fe585f171175__llikCauchyDlocation_wyfI
#> #define _llikCauchy _rx4c23a771624566473b9134f1fe585f171176__llikCauchy_CJq3
#> #define _llikGammaDrate _rx4c23a771624566473b9134f1fe585f171177__llikGammaDrate_BTKo
#> #define _llikGammaDshape _rx4c23a771624566473b9134f1fe585f171178__llikGammaDshape_r1pO
#> #define _llikGamma _rx4c23a771624566473b9134f1fe585f171179__llikGamma_f7Tv
#> #define _llikWeibullDscale _rx4c23a771624566473b9134f1fe585f171180__llikWeibullDscale_Zukd
#> #define _llikWeibullDshape _rx4c23a771624566473b9134f1fe585f171181__llikWeibullDshape_nvAB
#> #define _llikWeibull _rx4c23a771624566473b9134f1fe585f171182__llikWeibull_6pnH
#> #define _llikUnifDbeta _rx4c23a771624566473b9134f1fe585f171183__llikUnifDbeta_RcMN
#> #define _llikUnifDalpha _rx4c23a771624566473b9134f1fe585f171184__llikUnifDalpha_uO4h
#> #define _llikUnif _rx4c23a771624566473b9134f1fe585f171185__llikUnif_g52O
#> #define _llikGeomDp _rx4c23a771624566473b9134f1fe585f171186__llikGeomDp_0ZwL
#> #define _llikGeom _rx4c23a771624566473b9134f1fe585f171187__llikGeom_FX7N
#> #define _llikFDdf2 _rx4c23a771624566473b9134f1fe585f171188__llikFDdf2_Kwul
#> #define _llikFDdf1 _rx4c23a771624566473b9134f1fe585f171189__llikFDdf1_lzZD
#> #define _llikF _rx4c23a771624566473b9134f1fe585f171190__llikF_8nyy
#> #define _llikExpDrate _rx4c23a771624566473b9134f1fe585f171191__llikExpDrate_7cr8
#> #define _llikExp _rx4c23a771624566473b9134f1fe585f171192__llikExp_M9RU
#> #define _llikChisqDdf _rx4c23a771624566473b9134f1fe585f171193__llikChisqDdf_wuwl
#> #define _llikChisq _rx4c23a771624566473b9134f1fe585f171194__llikChisq_51KE
#> #define _llikTDsd _rx4c23a771624566473b9134f1fe585f171195__llikTDsd_nUV4
#> #define _llikTDmean _rx4c23a771624566473b9134f1fe585f171196__llikTDmean_LkN0
#> #define _llikTDdf _rx4c23a771624566473b9134f1fe585f171197__llikTDdf_iY4L
#> #define _llikT _rx4c23a771624566473b9134f1fe585f171198__llikT_PJjO
#> #define _llikBetaDshape2 _rx4c23a771624566473b9134f1fe585f171199__llikBetaDshape2_ccne
#> #define _llikBetaDshape1 _rx4c23a771624566473b9134f1fe585f171200__llikBetaDshape1_gPf1
#> #define _llikBeta _rx4c23a771624566473b9134f1fe585f171201__llikBeta_o4t4
#> #define _llikNbinomMuDmu _rx4c23a771624566473b9134f1fe585f171202__llikNbinomMuDmu_BmCP
#> #define _llikNbinomMu _rx4c23a771624566473b9134f1fe585f171203__llikNbinomMu_RbZ3
#> #define _llikNbinomDprob _rx4c23a771624566473b9134f1fe585f171204__llikNbinomDprob_BINi
#> #define _llikNbinom _rx4c23a771624566473b9134f1fe585f171205__llikNbinom_7Q1S
#> #define _llikBinomDprob _rx4c23a771624566473b9134f1fe585f171206__llikBinomDprob_nTEs
#> #define _llikBinom _rx4c23a771624566473b9134f1fe585f171207__llikBinom_7Gsk
#> #define _llikPoisDlambda _rx4c23a771624566473b9134f1fe585f171208__llikPoisDlambda_6or7
#> #define _llikPois _rx4c23a771624566473b9134f1fe585f171209__llikPois_8X3U
#> #define _llikNormDsd _rx4c23a771624566473b9134f1fe585f171210__llikNormDsd_0q3o
#> #define _llikNormDmean _rx4c23a771624566473b9134f1fe585f171211__llikNormDmean_XU1Y
#> #define _llikNorm _rx4c23a771624566473b9134f1fe585f171212__llikNorm_iEoQ
#> #define simeps _rx4c23a771624566473b9134f1fe585f171213_simeps_Z9CA
#> #define simeta _rx4c23a771624566473b9134f1fe585f171214_simeta_p6Gn
#> #define expit _rx4c23a771624566473b9134f1fe585f171215_expit_Ir2L
#> #define logit _rx4c23a771624566473b9134f1fe585f171216_logit_J8NS
#> #define gammapDer _rx4c23a771624566473b9134f1fe585f171217_gammapDer_2zuy
#> #define lowergamma _rx4c23a771624566473b9134f1fe585f171218_lowergamma_A28M
#> #define uppergamma _rx4c23a771624566473b9134f1fe585f171219_uppergamma_XwdU
#> #define gammaqInva _rx4c23a771624566473b9134f1fe585f171220_gammaqInva_5wLM
#> #define gammaqInv _rx4c23a771624566473b9134f1fe585f171221_gammaqInv_cjCa
#> #define gammapInva _rx4c23a771624566473b9134f1fe585f171222_gammapInva_qz0o
#> #define gammapInv _rx4c23a771624566473b9134f1fe585f171223_gammapInv_C7gG
#> #define gammaq _rx4c23a771624566473b9134f1fe585f171224_gammaq_pdW1
#> #define gammap _rx4c23a771624566473b9134f1fe585f171225_gammap_H2t8
#> #define phi _rx4c23a771624566473b9134f1fe585f171226_phi_3Mht
#> #define d2ELUa _rx4c23a771624566473b9134f1fe585f171227_d2ELUa_FOcY
#> #define dELUa _rx4c23a771624566473b9134f1fe585f171228_dELUa_2TwE
#> #define d2aELU _rx4c23a771624566473b9134f1fe585f171229_d2aELU_gu28
#> #define d2ELU _rx4c23a771624566473b9134f1fe585f171230_d2ELU_6PyM
#> #define dELU _rx4c23a771624566473b9134f1fe585f171231_dELU_ih1w
#> #define ELU _rx4c23a771624566473b9134f1fe585f171232_ELU_Xh2u
#> #define dPReLUa1 _rx4c23a771624566473b9134f1fe585f171233_dPReLUa1_2mWG
#> #define dPReLUa _rx4c23a771624566473b9134f1fe585f171234_dPReLUa_VX5u
#> #define dPReLU _rx4c23a771624566473b9134f1fe585f171235_dPReLU_Y3fc
#> #define PReLU _rx4c23a771624566473b9134f1fe585f171236_PReLU_4Rdo
#> #define dSwish _rx4c23a771624566473b9134f1fe585f171237_dSwish_Olsc
#> #define Swish _rx4c23a771624566473b9134f1fe585f171238_Swish_Mvk3
#> #define dlReLU _rx4c23a771624566473b9134f1fe585f171239_dlReLU_6on3
#> #define lReLU _rx4c23a771624566473b9134f1fe585f171240_lReLU_uitX
#> #define dSELU _rx4c23a771624566473b9134f1fe585f171241_dSELU_LpnW
#> #define SELU _rx4c23a771624566473b9134f1fe585f171242_SELU_m3S1
#> #define d4softplus _rx4c23a771624566473b9134f1fe585f171243_d4softplus_jh0T
#> #define d3softplus _rx4c23a771624566473b9134f1fe585f171244_d3softplus_aH3N
#> #define d2softplus _rx4c23a771624566473b9134f1fe585f171245_d2softplus_WFwR
#> #define dsoftplus _rx4c23a771624566473b9134f1fe585f171246_dsoftplus_hAF1
#> #define softplus _rx4c23a771624566473b9134f1fe585f171247_softplus_KP6q
#> #define d4GELU _rx4c23a771624566473b9134f1fe585f171248_d4GELU_0ynV
#> #define d3GELU _rx4c23a771624566473b9134f1fe585f171249_d3GELU_vTVp
#> #define d2GELU _rx4c23a771624566473b9134f1fe585f171250_d2GELU_9WQK
#> #define dGELU _rx4c23a771624566473b9134f1fe585f171251_dGELU_R2F6
#> #define GELU _rx4c23a771624566473b9134f1fe585f171252_GELU_MeQ4
#> #define dReLU _rx4c23a771624566473b9134f1fe585f171253_dReLU_gqr5
#> #define ReLU _rx4c23a771624566473b9134f1fe585f171254_ReLU_0QyU
#> #define riweibull _rx4c23a771624566473b9134f1fe585f171255_riweibull_zU7Y
#> #define riunif _rx4c23a771624566473b9134f1fe585f171256_riunif_EimR
#> #define rit_ _rx4c23a771624566473b9134f1fe585f171257_rit__TOtY
#> #define ripois _rx4c23a771624566473b9134f1fe585f171258_ripois_gK5w
#> #define ribeta _rx4c23a771624566473b9134f1fe585f171259_ribeta_f07q
#> #define rigamma _rx4c23a771624566473b9134f1fe585f171260_rigamma_ROQf
#> #define rigeom _rx4c23a771624566473b9134f1fe585f171261_rigeom_Bpun
#> #define rif _rx4c23a771624566473b9134f1fe585f171262_rif_VFfw
#> #define riexp _rx4c23a771624566473b9134f1fe585f171263_riexp_y3Nn
#> #define richisq _rx4c23a771624566473b9134f1fe585f171264_richisq_XIO1
#> #define ricauchy _rx4c23a771624566473b9134f1fe585f171265_ricauchy_q1eb
#> #define rinbinomMu _rx4c23a771624566473b9134f1fe585f171266_rinbinomMu_Jzpb
#> #define rinbinom _rx4c23a771624566473b9134f1fe585f171267_rinbinom_8rZn
#> #define ribinom _rx4c23a771624566473b9134f1fe585f171268_ribinom_1Fkp
#> #define rinorm _rx4c23a771624566473b9134f1fe585f171269_rinorm_WdUa
#> #define rxweibull _rx4c23a771624566473b9134f1fe585f171270_rxweibull_6qOl
#> #define rxunif _rx4c23a771624566473b9134f1fe585f171271_rxunif_wGpe
#> #define rxt_ _rx4c23a771624566473b9134f1fe585f171272_rxt__nBkj
#> #define rxpois _rx4c23a771624566473b9134f1fe585f171273_rxpois_kK3f
#> #define rxbeta _rx4c23a771624566473b9134f1fe585f171274_rxbeta_2dIO
#> #define rxgamma _rx4c23a771624566473b9134f1fe585f171275_rxgamma_vgii
#> #define rxgeom _rx4c23a771624566473b9134f1fe585f171276_rxgeom_EPZs
#> #define rxf _rx4c23a771624566473b9134f1fe585f171277_rxf_auLF
#> #define rxexp _rx4c23a771624566473b9134f1fe585f171278_rxexp_iDRh
#> #define rxchisq _rx4c23a771624566473b9134f1fe585f171279_rxchisq_fBjK
#> #define rxcauchy _rx4c23a771624566473b9134f1fe585f171280_rxcauchy_LmMH
#> #define rxnbinomMu _rx4c23a771624566473b9134f1fe585f171281_rxnbinomMu_G1ek
#> #define rxnbinom _rx4c23a771624566473b9134f1fe585f171282_rxnbinom_d3W5
#> #define rxbinom _rx4c23a771624566473b9134f1fe585f171283_rxbinom_G0hn
#> #define rxnorm _rx4c23a771624566473b9134f1fe585f171284_rxnorm_8bMg
#> #define linCmtB _rx4c23a771624566473b9134f1fe585f171285_linCmtB_MnzT
#> #define linCmtA _rx4c23a771624566473b9134f1fe585f171286_linCmtA_48ti
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){}
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){}
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  rx_a5bc1722466815bb29f2d63ce8fd3c15__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *rx_a5bc1722466815bb29f2d63ce8fd3c15__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP rx_a5bc1722466815bb29f2d63ce8fd3c15__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void rx_a5bc1722466815bb29f2d63ce8fd3c15__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double r1 = NA_REAL;
#>   double r2a = NA_REAL;
#>   double ra = NA_REAL;
#> 
#>   (void)t;
#>   (void)r1;
#>   (void)r2a;
#>   (void)ra;
#> 
#>   r1 = _PL[0];
#>   r2a = _PL[1];
#>   ra = _PL[2];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   r1=dPReLU(t,2);
#>   r2a=dPReLUa(t,2);
#>   ra=dPReLUa1(t,2);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void rx_a5bc1722466815bb29f2d63ce8fd3c15__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void rx_a5bc1722466815bb29f2d63ce8fd3c15__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void rx_a5bc1722466815bb29f2d63ce8fd3c15__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double r1 = NA_REAL;
#>   double r2a = NA_REAL;
#>   double ra = NA_REAL;
#> 
#>   (void)t;
#>   (void)r1;
#>   (void)r2a;
#>   (void)ra;
#> 
#>   r1 = _PL[0];
#>   r2a = _PL[1];
#>   ra = _PL[2];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   r1=dPReLU(t,2);
#>   r2a=dPReLUa(t,2);
#>   ra=dPReLUa1(t,2);
#> 
#>   _lhs[_LHS_0_]=r1;
#>   _lhs[_LHS_1_]=r2a;
#>   _lhs[_LHS_2_]=ra;
#> }
#> // Functional based bioavailability
#> double rx_a5bc1722466815bb29f2d63ce8fd3c15__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double rx_a5bc1722466815bb29f2d63ce8fd3c15__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double rx_a5bc1722466815bb29f2d63ce8fd3c15__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double rx_a5bc1722466815bb29f2d63ce8fd3c15__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void rx_a5bc1722466815bb29f2d63ce8fd3c15__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void rx_a5bc1722466815bb29f2d63ce8fd3c15__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void rx_a5bc1722466815bb29f2d63ce8fd3c15__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP rx_a5bc1722466815bb29f2d63ce8fd3c15__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_a5bc1722466815bb29f2d63ce8fd3c15__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAAMc1AAAv7#aT)KogQNAj;ca=WdwYB~1u%%C8AeN^_i|CBpu68l<__ZVsoOQvPn`M$2qf>CAyeKeOx1z7={J&x8luKqJ&GvF2$XV@lLJt*A]CJF{mk1]Y0ek)cy!r9<1l8I}=Gk*CytmT4kHy1$t0l)(;_>9o`&4dk\?ugk:d`5zU9q/,<I(CZ&_:X>dn@=Nhq$xsDW*Uyo_%%jiZNh5:`&\?#cS_[gobeYk4KQ%%c)9E@0hotoDCpu4g\"1+=+[x&2I\?]HY=&6;DMVK5ryyG7FyR(@`$YG6CwKhmcj1\"azw\?1~/M\?m9Z5`)[t+Jc/JUtsV|ucSrd!Z5<IIg}:\?1UAD0yTTquG+Mf7hxW]np~Qc66++ft)j%%`CTJ8%%PWUE&\"w+SMA@0_J`R91{lRSY[wAYxVcE*1>SByyB!Wv1Dz6dWq+#Ww:6tW#i@tizn#:w#ecM\?*8*22fSrvHttO{oe5r5!aVsr750Nop:c@<6mrM7,g{4%%$=X~fi%%Y<Je;ht9V0eLF`&2)dPd&WPUb0k}y,w#Tl<nna+LaPFL0]$y@UzaX8@Ui=GP=<0]UIzm\"By#nl`vI$DqkC$L\"MnGjZR^8EzUozr72f@8jgfn&F,Cee]&~e&n/C,rIC0exOskq+L_i1k2pLi#t^bqk*J!n9X04EwbEwF=]`|wF1(m=}sQ*=\"Pilf)$==YEz:&XV{tg0.[Spk)U~d0_po6w$0C1^;s8q}og_LMzZVyJ:V88v^S;+(\?:;aeb6feLC;@[DYqEv\?>3:y:CKeMSGh4qO.zQCPD%%FIG|m1B1Zk:HDAM|XXzYY|L\"CEzym4$xpmo^7PtB6vTWZ+^+TpT=V7)&t9w5#Nx=uCCOQ!hCzQ}eZ@X2T$pVSz3|D#j@1O4:4]>E)QONS2UlDxP<3qn\?hOa24)w99ZH`aYLh)\?:$cwmGOaZ\?R5UWm;Q>2wPgE!bSD`MW(\?/;)pFw`x[SW:iviVhdSgjWg`%%&1%%yKKColxT,fejqV4KK|6vHyC8_Ga4)h7vUQSM1~vXtB)(7G6^V~1\?:=J!*wYAXB7R{vqaNaXQkdlwO5YfeKRZD*qA");
#>     char buf[1094];
#> #define __doBufN__ 1094
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void rx_a5bc1722466815bb29f2d63ce8fd3c15__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   rx_a5bc1722466815bb29f2d63ce8fd3c15__dydt(neq, *t, A, DADT);
#> }
#> extern int rx_a5bc1722466815bb29f2d63ce8fd3c15__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   rx_a5bc1722466815bb29f2d63ce8fd3c15__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void rx_a5bc1722466815bb29f2d63ce8fd3c15__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   rx_a5bc1722466815bb29f2d63ce8fd3c15__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void rx_a5bc1722466815bb29f2d63ce8fd3c15__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_rx_a5bc1722466815bb29f2d63ce8fd3c15_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("rx_a5bc1722466815bb29f2d63ce8fd3c15_","rx_a5bc1722466815bb29f2d63ce8fd3c15__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("rx_a5bc1722466815bb29f2d63ce8fd3c15_","rx_a5bc1722466815bb29f2d63ce8fd3c15__assignFuns", (DL_FUNC) rx_a5bc1722466815bb29f2d63ce8fd3c15__assignFuns);
#>   R_RegisterCCallable("rx_a5bc1722466815bb29f2d63ce8fd3c15_","rx_a5bc1722466815bb29f2d63ce8fd3c15__inis",(DL_FUNC) rx_a5bc1722466815bb29f2d63ce8fd3c15__inis);
#>   R_RegisterCCallable("rx_a5bc1722466815bb29f2d63ce8fd3c15_","rx_a5bc1722466815bb29f2d63ce8fd3c15__dydt",(DL_FUNC) rx_a5bc1722466815bb29f2d63ce8fd3c15__dydt);
#>   R_RegisterCCallable("rx_a5bc1722466815bb29f2d63ce8fd3c15_","rx_a5bc1722466815bb29f2d63ce8fd3c15__calc_lhs",(DL_FUNC) rx_a5bc1722466815bb29f2d63ce8fd3c15__calc_lhs);
#>   R_RegisterCCallable("rx_a5bc1722466815bb29f2d63ce8fd3c15_","rx_a5bc1722466815bb29f2d63ce8fd3c15__calc_jac",(DL_FUNC) rx_a5bc1722466815bb29f2d63ce8fd3c15__calc_jac);
#>   R_RegisterCCallable("rx_a5bc1722466815bb29f2d63ce8fd3c15_","rx_a5bc1722466815bb29f2d63ce8fd3c15__dydt_lsoda", (DL_FUNC) rx_a5bc1722466815bb29f2d63ce8fd3c15__dydt_lsoda);
#>   R_RegisterCCallable("rx_a5bc1722466815bb29f2d63ce8fd3c15_","rx_a5bc1722466815bb29f2d63ce8fd3c15__calc_jac_lsoda", (DL_FUNC) rx_a5bc1722466815bb29f2d63ce8fd3c15__calc_jac_lsoda);
#>   R_RegisterCCallable("rx_a5bc1722466815bb29f2d63ce8fd3c15_","rx_a5bc1722466815bb29f2d63ce8fd3c15__ode_solver_solvedata", (DL_FUNC) rx_a5bc1722466815bb29f2d63ce8fd3c15__ode_solver_solvedata);
#>   R_RegisterCCallable("rx_a5bc1722466815bb29f2d63ce8fd3c15_","rx_a5bc1722466815bb29f2d63ce8fd3c15__ode_solver_get_solvedata", (DL_FUNC) rx_a5bc1722466815bb29f2d63ce8fd3c15__ode_solver_get_solvedata);
#>   R_RegisterCCallable("rx_a5bc1722466815bb29f2d63ce8fd3c15_","rx_a5bc1722466815bb29f2d63ce8fd3c15__F", (DL_FUNC) rx_a5bc1722466815bb29f2d63ce8fd3c15__F);
#>   R_RegisterCCallable("rx_a5bc1722466815bb29f2d63ce8fd3c15_","rx_a5bc1722466815bb29f2d63ce8fd3c15__Lag", (DL_FUNC) rx_a5bc1722466815bb29f2d63ce8fd3c15__Lag);
#>   R_RegisterCCallable("rx_a5bc1722466815bb29f2d63ce8fd3c15_","rx_a5bc1722466815bb29f2d63ce8fd3c15__Rate", (DL_FUNC) rx_a5bc1722466815bb29f2d63ce8fd3c15__Rate);
#>   R_RegisterCCallable("rx_a5bc1722466815bb29f2d63ce8fd3c15_","rx_a5bc1722466815bb29f2d63ce8fd3c15__Dur", (DL_FUNC) rx_a5bc1722466815bb29f2d63ce8fd3c15__Dur);
#>   R_RegisterCCallable("rx_a5bc1722466815bb29f2d63ce8fd3c15_","rx_a5bc1722466815bb29f2d63ce8fd3c15__mtime", (DL_FUNC) rx_a5bc1722466815bb29f2d63ce8fd3c15__mtime);
#>   R_RegisterCCallable("rx_a5bc1722466815bb29f2d63ce8fd3c15_","rx_a5bc1722466815bb29f2d63ce8fd3c15__ME", (DL_FUNC) rx_a5bc1722466815bb29f2d63ce8fd3c15__ME);
#>   R_RegisterCCallable("rx_a5bc1722466815bb29f2d63ce8fd3c15_","rx_a5bc1722466815bb29f2d63ce8fd3c15__IndF", (DL_FUNC) rx_a5bc1722466815bb29f2d63ce8fd3c15__IndF);
#>   R_RegisterCCallable("rx_a5bc1722466815bb29f2d63ce8fd3c15_","rx_a5bc1722466815bb29f2d63ce8fd3c15__dydt_liblsoda", (DL_FUNC) rx_a5bc1722466815bb29f2d63ce8fd3c15__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_rx_a5bc1722466815bb29f2d63ce8fd3c15_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_rx_a5bc1722466815bb29f2d63ce8fd3c15_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"rx_a5bc1722466815bb29f2d63ce8fd3c15__model_vars", (DL_FUNC) &rx_a5bc1722466815bb29f2d63ce8fd3c15__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_rx_a5bc1722466815bb29f2d63ce8fd3c15_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_a5bc1722466815bb29f2d63ce8fd3c15__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("rx_a5bc1722466815bb29f2d63ce8fd3c15__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’

e <- et(c(-1, 0, 1, 2))
rxSolve(r, e)
#> ── Solved rxode2 object ──
#> ── Parameters (value$params): ──
#> # A tibble: 1 × 0
#> ── Initial Conditions (value$inits): ──
#> named numeric(0)
#> ── First part of data (object): ──
#> # A tibble: 4 × 4
#>    time    r1   r2a    ra
#>   <dbl> <dbl> <dbl> <dbl>
#> 1    -1     2    -1     1
#> 2     0     2     0     1
#> 3     1     1     0     0
#> 4     2     1     0     0