Derivative of the Swish Activation Function

Derivative of the Swish Activation Function

Usage

dSwish(x)

Arguments

x

A numeric vector. All elements must be finite and non-missing.

Value

A numeric vector where the derivative of the SELU function has been applied to each element of x.

See also

Author

Matthew Fidler

Examples

dSwish(c(-1, 0, 1, 2))
#> [1] 0.07232949 0.50000000 0.92767051 1.09078425

# Can also be used in rxode2:
x <- rxode2({
  r <- dSwish(time)
})
#>  
#>  
#> C model:
#> // Define 1 LHS values
#> #define _LHS_0_ 0 // r
#> #define _getRxSolve_ _rx0bfdc3cfdb9574599d6c88a668a0a2061573__getRxSolve__mvNF
#> #define _evalUdf _rx0bfdc3cfdb9574599d6c88a668a0a2061574__evalUdf_5OB8
#> #define _solveData _rx0bfdc3cfdb9574599d6c88a668a0a2061575__solveData_FY9x
#> #define _assign_ptr _rx0bfdc3cfdb9574599d6c88a668a0a2061576__assign_ptr_wzh8
#> #define _rxRmModelLib _rx0bfdc3cfdb9574599d6c88a668a0a2061577__rxRmModelLib_0ggu
#> #define _rxGetModelLib _rx0bfdc3cfdb9574599d6c88a668a0a2061578__rxGetModelLib_PQXP
#> #define _old_c _rx0bfdc3cfdb9574599d6c88a668a0a2061579__old_c_6NCi
#> #define _ptrid _rx0bfdc3cfdb9574599d6c88a668a0a2061580__ptrid_0utk
#> #define _rxIsCurrentC _rx0bfdc3cfdb9574599d6c88a668a0a2061581__rxIsCurrentC_xceN
#> #define _sumPS _rx0bfdc3cfdb9574599d6c88a668a0a2061582__sumPS_4g7I
#> #define _prodPS _rx0bfdc3cfdb9574599d6c88a668a0a2061583__prodPS_dO4r
#> #define _prodType _rx0bfdc3cfdb9574599d6c88a668a0a2061584__prodType_Nf1a
#> #define _sumType _rx0bfdc3cfdb9574599d6c88a668a0a2061585__sumType_HO9p
#> #define _update_par_ptr _rx0bfdc3cfdb9574599d6c88a668a0a2061586__update_par_ptr_8utB
#> #define _getParCov _rx0bfdc3cfdb9574599d6c88a668a0a2061587__getParCov_XW9W
#> #define _rxode2_rxAssignPtr _rx0bfdc3cfdb9574599d6c88a668a0a2061588__rxode2_rxAssignPtr_DHV9
#> #define _rxQr _rx0bfdc3cfdb9574599d6c88a668a0a2061589__rxQr_vUck
#> #define _compareFactorVal _rx0bfdc3cfdb9574599d6c88a668a0a2061590__compareFactorVal_cv41
#> #define _sum _rx0bfdc3cfdb9574599d6c88a668a0a2061591__sum_XEdN
#> #define _udf _rx0bfdc3cfdb9574599d6c88a668a0a2061592__udf_22xE
#> #define _sign _rx0bfdc3cfdb9574599d6c88a668a0a2061593__sign_0D3E
#> #define _prod _rx0bfdc3cfdb9574599d6c88a668a0a2061594__prod_O0Ma
#> #define _max _rx0bfdc3cfdb9574599d6c88a668a0a2061595__max_bGfw
#> #define _min _rx0bfdc3cfdb9574599d6c88a668a0a2061596__min_0mo4
#> #define _transit4P _rx0bfdc3cfdb9574599d6c88a668a0a2061597__transit4P_ZX8H
#> #define _transit3P _rx0bfdc3cfdb9574599d6c88a668a0a2061598__transit3P_LoAW
#> #define _assignFuns0 _rx0bfdc3cfdb9574599d6c88a668a0a2061599__assignFuns0_lHRf
#> #define _assignFuns _rx0bfdc3cfdb9574599d6c88a668a0a2061600__assignFuns_2YvQ
#> #define _rxord _rx0bfdc3cfdb9574599d6c88a668a0a2061601__rxord_wyfI
#> #define __assignFuns2 _rx0bfdc3cfdb9574599d6c88a668a0a2061602___assignFuns2_CJq3
#> #define _llikCauchyDscale _rx0bfdc3cfdb9574599d6c88a668a0a2061603__llikCauchyDscale_BTKo
#> #define _llikCauchyDlocation _rx0bfdc3cfdb9574599d6c88a668a0a2061604__llikCauchyDlocation_r1pO
#> #define _llikCauchy _rx0bfdc3cfdb9574599d6c88a668a0a2061605__llikCauchy_f7Tv
#> #define _llikGammaDrate _rx0bfdc3cfdb9574599d6c88a668a0a2061606__llikGammaDrate_Zukd
#> #define _llikGammaDshape _rx0bfdc3cfdb9574599d6c88a668a0a2061607__llikGammaDshape_nvAB
#> #define _llikGamma _rx0bfdc3cfdb9574599d6c88a668a0a2061608__llikGamma_6pnH
#> #define _llikWeibullDscale _rx0bfdc3cfdb9574599d6c88a668a0a2061609__llikWeibullDscale_RcMN
#> #define _llikWeibullDshape _rx0bfdc3cfdb9574599d6c88a668a0a2061610__llikWeibullDshape_uO4h
#> #define _llikWeibull _rx0bfdc3cfdb9574599d6c88a668a0a2061611__llikWeibull_g52O
#> #define _llikUnifDbeta _rx0bfdc3cfdb9574599d6c88a668a0a2061612__llikUnifDbeta_0ZwL
#> #define _llikUnifDalpha _rx0bfdc3cfdb9574599d6c88a668a0a2061613__llikUnifDalpha_FX7N
#> #define _llikUnif _rx0bfdc3cfdb9574599d6c88a668a0a2061614__llikUnif_Kwul
#> #define _llikGeomDp _rx0bfdc3cfdb9574599d6c88a668a0a2061615__llikGeomDp_lzZD
#> #define _llikGeom _rx0bfdc3cfdb9574599d6c88a668a0a2061616__llikGeom_8nyy
#> #define _llikFDdf2 _rx0bfdc3cfdb9574599d6c88a668a0a2061617__llikFDdf2_7cr8
#> #define _llikFDdf1 _rx0bfdc3cfdb9574599d6c88a668a0a2061618__llikFDdf1_M9RU
#> #define _llikF _rx0bfdc3cfdb9574599d6c88a668a0a2061619__llikF_wuwl
#> #define _llikExpDrate _rx0bfdc3cfdb9574599d6c88a668a0a2061620__llikExpDrate_51KE
#> #define _llikExp _rx0bfdc3cfdb9574599d6c88a668a0a2061621__llikExp_nUV4
#> #define _llikChisqDdf _rx0bfdc3cfdb9574599d6c88a668a0a2061622__llikChisqDdf_LkN0
#> #define _llikChisq _rx0bfdc3cfdb9574599d6c88a668a0a2061623__llikChisq_iY4L
#> #define _llikTDsd _rx0bfdc3cfdb9574599d6c88a668a0a2061624__llikTDsd_PJjO
#> #define _llikTDmean _rx0bfdc3cfdb9574599d6c88a668a0a2061625__llikTDmean_ccne
#> #define _llikTDdf _rx0bfdc3cfdb9574599d6c88a668a0a2061626__llikTDdf_gPf1
#> #define _llikT _rx0bfdc3cfdb9574599d6c88a668a0a2061627__llikT_o4t4
#> #define _llikBetaDshape2 _rx0bfdc3cfdb9574599d6c88a668a0a2061628__llikBetaDshape2_BmCP
#> #define _llikBetaDshape1 _rx0bfdc3cfdb9574599d6c88a668a0a2061629__llikBetaDshape1_RbZ3
#> #define _llikBeta _rx0bfdc3cfdb9574599d6c88a668a0a2061630__llikBeta_BINi
#> #define _llikNbinomMuDmu _rx0bfdc3cfdb9574599d6c88a668a0a2061631__llikNbinomMuDmu_7Q1S
#> #define _llikNbinomMu _rx0bfdc3cfdb9574599d6c88a668a0a2061632__llikNbinomMu_nTEs
#> #define _llikNbinomDprob _rx0bfdc3cfdb9574599d6c88a668a0a2061633__llikNbinomDprob_7Gsk
#> #define _llikNbinom _rx0bfdc3cfdb9574599d6c88a668a0a2061634__llikNbinom_6or7
#> #define _llikBinomDprob _rx0bfdc3cfdb9574599d6c88a668a0a2061635__llikBinomDprob_8X3U
#> #define _llikBinom _rx0bfdc3cfdb9574599d6c88a668a0a2061636__llikBinom_0q3o
#> #define _llikPoisDlambda _rx0bfdc3cfdb9574599d6c88a668a0a2061637__llikPoisDlambda_XU1Y
#> #define _llikPois _rx0bfdc3cfdb9574599d6c88a668a0a2061638__llikPois_iEoQ
#> #define _llikNormDsd _rx0bfdc3cfdb9574599d6c88a668a0a2061639__llikNormDsd_Z9CA
#> #define _llikNormDmean _rx0bfdc3cfdb9574599d6c88a668a0a2061640__llikNormDmean_p6Gn
#> #define _llikNorm _rx0bfdc3cfdb9574599d6c88a668a0a2061641__llikNorm_Ir2L
#> #define simeps _rx0bfdc3cfdb9574599d6c88a668a0a2061642_simeps_J8NS
#> #define simeta _rx0bfdc3cfdb9574599d6c88a668a0a2061643_simeta_2zuy
#> #define expit _rx0bfdc3cfdb9574599d6c88a668a0a2061644_expit_A28M
#> #define logit _rx0bfdc3cfdb9574599d6c88a668a0a2061645_logit_XwdU
#> #define gammapDer _rx0bfdc3cfdb9574599d6c88a668a0a2061646_gammapDer_5wLM
#> #define lowergamma _rx0bfdc3cfdb9574599d6c88a668a0a2061647_lowergamma_cjCa
#> #define uppergamma _rx0bfdc3cfdb9574599d6c88a668a0a2061648_uppergamma_qz0o
#> #define gammaqInva _rx0bfdc3cfdb9574599d6c88a668a0a2061649_gammaqInva_C7gG
#> #define gammaqInv _rx0bfdc3cfdb9574599d6c88a668a0a2061650_gammaqInv_pdW1
#> #define gammapInva _rx0bfdc3cfdb9574599d6c88a668a0a2061651_gammapInva_H2t8
#> #define gammapInv _rx0bfdc3cfdb9574599d6c88a668a0a2061652_gammapInv_3Mht
#> #define gammaq _rx0bfdc3cfdb9574599d6c88a668a0a2061653_gammaq_FOcY
#> #define gammap _rx0bfdc3cfdb9574599d6c88a668a0a2061654_gammap_2TwE
#> #define phi _rx0bfdc3cfdb9574599d6c88a668a0a2061655_phi_gu28
#> #define d2ELUa _rx0bfdc3cfdb9574599d6c88a668a0a2061656_d2ELUa_6PyM
#> #define dELUa _rx0bfdc3cfdb9574599d6c88a668a0a2061657_dELUa_ih1w
#> #define d2aELU _rx0bfdc3cfdb9574599d6c88a668a0a2061658_d2aELU_Xh2u
#> #define d2ELU _rx0bfdc3cfdb9574599d6c88a668a0a2061659_d2ELU_2mWG
#> #define dELU _rx0bfdc3cfdb9574599d6c88a668a0a2061660_dELU_VX5u
#> #define ELU _rx0bfdc3cfdb9574599d6c88a668a0a2061661_ELU_Y3fc
#> #define dPReLUa1 _rx0bfdc3cfdb9574599d6c88a668a0a2061662_dPReLUa1_4Rdo
#> #define dPReLUa _rx0bfdc3cfdb9574599d6c88a668a0a2061663_dPReLUa_Olsc
#> #define dPReLU _rx0bfdc3cfdb9574599d6c88a668a0a2061664_dPReLU_Mvk3
#> #define PReLU _rx0bfdc3cfdb9574599d6c88a668a0a2061665_PReLU_6on3
#> #define dSwish _rx0bfdc3cfdb9574599d6c88a668a0a2061666_dSwish_uitX
#> #define Swish _rx0bfdc3cfdb9574599d6c88a668a0a2061667_Swish_LpnW
#> #define dlReLU _rx0bfdc3cfdb9574599d6c88a668a0a2061668_dlReLU_m3S1
#> #define lReLU _rx0bfdc3cfdb9574599d6c88a668a0a2061669_lReLU_jh0T
#> #define dSELU _rx0bfdc3cfdb9574599d6c88a668a0a2061670_dSELU_aH3N
#> #define SELU _rx0bfdc3cfdb9574599d6c88a668a0a2061671_SELU_WFwR
#> #define d4softplus _rx0bfdc3cfdb9574599d6c88a668a0a2061672_d4softplus_hAF1
#> #define d3softplus _rx0bfdc3cfdb9574599d6c88a668a0a2061673_d3softplus_KP6q
#> #define d2softplus _rx0bfdc3cfdb9574599d6c88a668a0a2061674_d2softplus_0ynV
#> #define dsoftplus _rx0bfdc3cfdb9574599d6c88a668a0a2061675_dsoftplus_vTVp
#> #define softplus _rx0bfdc3cfdb9574599d6c88a668a0a2061676_softplus_9WQK
#> #define d4GELU _rx0bfdc3cfdb9574599d6c88a668a0a2061677_d4GELU_R2F6
#> #define d3GELU _rx0bfdc3cfdb9574599d6c88a668a0a2061678_d3GELU_MeQ4
#> #define d2GELU _rx0bfdc3cfdb9574599d6c88a668a0a2061679_d2GELU_gqr5
#> #define dGELU _rx0bfdc3cfdb9574599d6c88a668a0a2061680_dGELU_0QyU
#> #define GELU _rx0bfdc3cfdb9574599d6c88a668a0a2061681_GELU_zU7Y
#> #define dReLU _rx0bfdc3cfdb9574599d6c88a668a0a2061682_dReLU_EimR
#> #define ReLU _rx0bfdc3cfdb9574599d6c88a668a0a2061683_ReLU_TOtY
#> #define riweibull _rx0bfdc3cfdb9574599d6c88a668a0a2061684_riweibull_gK5w
#> #define riunif _rx0bfdc3cfdb9574599d6c88a668a0a2061685_riunif_f07q
#> #define rit_ _rx0bfdc3cfdb9574599d6c88a668a0a2061686_rit__ROQf
#> #define ripois _rx0bfdc3cfdb9574599d6c88a668a0a2061687_ripois_Bpun
#> #define ribeta _rx0bfdc3cfdb9574599d6c88a668a0a2061688_ribeta_VFfw
#> #define rigamma _rx0bfdc3cfdb9574599d6c88a668a0a2061689_rigamma_y3Nn
#> #define rigeom _rx0bfdc3cfdb9574599d6c88a668a0a2061690_rigeom_XIO1
#> #define rif _rx0bfdc3cfdb9574599d6c88a668a0a2061691_rif_q1eb
#> #define riexp _rx0bfdc3cfdb9574599d6c88a668a0a2061692_riexp_Jzpb
#> #define richisq _rx0bfdc3cfdb9574599d6c88a668a0a2061693_richisq_8rZn
#> #define ricauchy _rx0bfdc3cfdb9574599d6c88a668a0a2061694_ricauchy_1Fkp
#> #define rinbinomMu _rx0bfdc3cfdb9574599d6c88a668a0a2061695_rinbinomMu_WdUa
#> #define rinbinom _rx0bfdc3cfdb9574599d6c88a668a0a2061696_rinbinom_6qOl
#> #define ribinom _rx0bfdc3cfdb9574599d6c88a668a0a2061697_ribinom_wGpe
#> #define rinorm _rx0bfdc3cfdb9574599d6c88a668a0a2061698_rinorm_nBkj
#> #define rxweibull _rx0bfdc3cfdb9574599d6c88a668a0a2061699_rxweibull_kK3f
#> #define rxunif _rx0bfdc3cfdb9574599d6c88a668a0a2061700_rxunif_2dIO
#> #define rxt_ _rx0bfdc3cfdb9574599d6c88a668a0a2061701_rxt__vgii
#> #define rxpois _rx0bfdc3cfdb9574599d6c88a668a0a2061702_rxpois_EPZs
#> #define rxbeta _rx0bfdc3cfdb9574599d6c88a668a0a2061703_rxbeta_auLF
#> #define rxgamma _rx0bfdc3cfdb9574599d6c88a668a0a2061704_rxgamma_iDRh
#> #define rxgeom _rx0bfdc3cfdb9574599d6c88a668a0a2061705_rxgeom_fBjK
#> #define rxf _rx0bfdc3cfdb9574599d6c88a668a0a2061706_rxf_LmMH
#> #define rxexp _rx0bfdc3cfdb9574599d6c88a668a0a2061707_rxexp_G1ek
#> #define rxchisq _rx0bfdc3cfdb9574599d6c88a668a0a2061708_rxchisq_d3W5
#> #define rxcauchy _rx0bfdc3cfdb9574599d6c88a668a0a2061709_rxcauchy_G0hn
#> #define rxnbinomMu _rx0bfdc3cfdb9574599d6c88a668a0a2061710_rxnbinomMu_8bMg
#> #define rxnbinom _rx0bfdc3cfdb9574599d6c88a668a0a2061711_rxnbinom_MnzT
#> #define rxbinom _rx0bfdc3cfdb9574599d6c88a668a0a2061712_rxbinom_48ti
#> #define rxnorm _rx0bfdc3cfdb9574599d6c88a668a0a2061713_rxnorm_suT4
#> #define linCmtB _rx0bfdc3cfdb9574599d6c88a668a0a2061714_linCmtB_iGKN
#> #define linCmtA _rx0bfdc3cfdb9574599d6c88a668a0a2061715_linCmtA_K6ak
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){}
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){}
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  rx_65b62eead1e4df27bcbcd47f565c3c36__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *rx_65b62eead1e4df27bcbcd47f565c3c36__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP rx_65b62eead1e4df27bcbcd47f565c3c36__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void rx_65b62eead1e4df27bcbcd47f565c3c36__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double r = NA_REAL;
#> 
#>   (void)t;
#>   (void)r;
#> 
#>   r = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   r =dSwish(t);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void rx_65b62eead1e4df27bcbcd47f565c3c36__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void rx_65b62eead1e4df27bcbcd47f565c3c36__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void rx_65b62eead1e4df27bcbcd47f565c3c36__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double r = NA_REAL;
#> 
#>   (void)t;
#>   (void)r;
#> 
#>   r = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   r =dSwish(t);
#> 
#>   _lhs[_LHS_0_]=r;
#> }
#> // Functional based bioavailability
#> double rx_65b62eead1e4df27bcbcd47f565c3c36__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double rx_65b62eead1e4df27bcbcd47f565c3c36__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double rx_65b62eead1e4df27bcbcd47f565c3c36__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double rx_65b62eead1e4df27bcbcd47f565c3c36__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void rx_65b62eead1e4df27bcbcd47f565c3c36__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void rx_65b62eead1e4df27bcbcd47f565c3c36__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void rx_65b62eead1e4df27bcbcd47f565c3c36__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP rx_65b62eead1e4df27bcbcd47f565c3c36__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_65b62eead1e4df27bcbcd47f565c3c36__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAA6yzAAAv7#aT)gm\"&MAF*4+ftGlYBaieEtYStOXk4)sKAZ*M#6lVw;GvOVjy,Y0e&k))c~]29DUo0`$pMLg~znyZT#/RT!EyLR_jB|LC\"p1+:<>tm.Zj$91`B9F+@]iebt=(U=]Zk:VOL+]WcqSu#UP!TEtf%%Q|^4$nJ|JfRcEVa2/<%%.fQ=\"Tn{zA(j8&WKKM^m+{kaXD,9DND{aG0>N1@bX]gO*1\?8;jG^wgleJ:0UgcG7~kp>%%f,E>V/sux9rIZF,XF\?Kh_[Z~cjib0};W,LC8#feMGeoP!)Ui=k!$kYfPMR01>0UBg{C_d69%%Tk$#+:[n3xTQ=mgcK\"=@0.f+|5T*RDX)7WnLouR#$m.N9:(.lzbCDH)M!\?l[kg<EvXgMmS[w6yt%%,P0Mx;nq^\?^X:0*M=M~QJa}|2@Wr9ib)%%GY\?o!,[=ffCpxwg`Qjvgq<({u`hWm;U}VxOj!.x>iUJ;{boj`|y7(6>r:p$FyRUGBiTR0\"_Y.e4H3ZPd)@800Z~3g5#U,qw~:NEZ6{up5SYe8D^X3e8nn%%2rp|:{~\?]/B`zfl^Cwghm4E}V]qnV7_(7uJct&3>bk_$XKq0(U\?;E6qu9{X3``QrU_(+*jqP4i[~[lV@SKZKK\"xw$q0/rG@yN&+zG;2.V_D%%O<p7_fg.G<Kf(c^aUZASsMwig35=U~yumTkB4IJf)[[($Zn/2M)y|#VZQs2X][5btEe7%%ek>gEb:_BjoQAiC)403:{!CK|RTz3WFksSxoPDM01cSF1B&6~z8I)MlR|x(T\?J{xbvmvoS9*B/:44AfZI>(~W9}l*zz|$EIUVFAd_aQwJbZet|q/xh]36CJ>,/P(AmjCP=Tp^r{P]&njY\?!CIWt`9^3k7,X)0ZrJIB<nNVsQy@xF)j8<xZ%%!_iN29c,sA{;w[98F>$0`aR}FoHXugC4MZ@)W#6,JKarR[[5ux2lKk>5PR{@d&=yP*4`V:k)5,%%E`r_.u]65$t\?Wca;Qc{3y+Ix6gR;DNK)ka{Mb^F,!y*19c_9/B1A");
#>     char buf[1060];
#> #define __doBufN__ 1060
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void rx_65b62eead1e4df27bcbcd47f565c3c36__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   rx_65b62eead1e4df27bcbcd47f565c3c36__dydt(neq, *t, A, DADT);
#> }
#> extern int rx_65b62eead1e4df27bcbcd47f565c3c36__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   rx_65b62eead1e4df27bcbcd47f565c3c36__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void rx_65b62eead1e4df27bcbcd47f565c3c36__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   rx_65b62eead1e4df27bcbcd47f565c3c36__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void rx_65b62eead1e4df27bcbcd47f565c3c36__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_rx_65b62eead1e4df27bcbcd47f565c3c36_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("rx_65b62eead1e4df27bcbcd47f565c3c36_","rx_65b62eead1e4df27bcbcd47f565c3c36__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("rx_65b62eead1e4df27bcbcd47f565c3c36_","rx_65b62eead1e4df27bcbcd47f565c3c36__assignFuns", (DL_FUNC) rx_65b62eead1e4df27bcbcd47f565c3c36__assignFuns);
#>   R_RegisterCCallable("rx_65b62eead1e4df27bcbcd47f565c3c36_","rx_65b62eead1e4df27bcbcd47f565c3c36__inis",(DL_FUNC) rx_65b62eead1e4df27bcbcd47f565c3c36__inis);
#>   R_RegisterCCallable("rx_65b62eead1e4df27bcbcd47f565c3c36_","rx_65b62eead1e4df27bcbcd47f565c3c36__dydt",(DL_FUNC) rx_65b62eead1e4df27bcbcd47f565c3c36__dydt);
#>   R_RegisterCCallable("rx_65b62eead1e4df27bcbcd47f565c3c36_","rx_65b62eead1e4df27bcbcd47f565c3c36__calc_lhs",(DL_FUNC) rx_65b62eead1e4df27bcbcd47f565c3c36__calc_lhs);
#>   R_RegisterCCallable("rx_65b62eead1e4df27bcbcd47f565c3c36_","rx_65b62eead1e4df27bcbcd47f565c3c36__calc_jac",(DL_FUNC) rx_65b62eead1e4df27bcbcd47f565c3c36__calc_jac);
#>   R_RegisterCCallable("rx_65b62eead1e4df27bcbcd47f565c3c36_","rx_65b62eead1e4df27bcbcd47f565c3c36__dydt_lsoda", (DL_FUNC) rx_65b62eead1e4df27bcbcd47f565c3c36__dydt_lsoda);
#>   R_RegisterCCallable("rx_65b62eead1e4df27bcbcd47f565c3c36_","rx_65b62eead1e4df27bcbcd47f565c3c36__calc_jac_lsoda", (DL_FUNC) rx_65b62eead1e4df27bcbcd47f565c3c36__calc_jac_lsoda);
#>   R_RegisterCCallable("rx_65b62eead1e4df27bcbcd47f565c3c36_","rx_65b62eead1e4df27bcbcd47f565c3c36__ode_solver_solvedata", (DL_FUNC) rx_65b62eead1e4df27bcbcd47f565c3c36__ode_solver_solvedata);
#>   R_RegisterCCallable("rx_65b62eead1e4df27bcbcd47f565c3c36_","rx_65b62eead1e4df27bcbcd47f565c3c36__ode_solver_get_solvedata", (DL_FUNC) rx_65b62eead1e4df27bcbcd47f565c3c36__ode_solver_get_solvedata);
#>   R_RegisterCCallable("rx_65b62eead1e4df27bcbcd47f565c3c36_","rx_65b62eead1e4df27bcbcd47f565c3c36__F", (DL_FUNC) rx_65b62eead1e4df27bcbcd47f565c3c36__F);
#>   R_RegisterCCallable("rx_65b62eead1e4df27bcbcd47f565c3c36_","rx_65b62eead1e4df27bcbcd47f565c3c36__Lag", (DL_FUNC) rx_65b62eead1e4df27bcbcd47f565c3c36__Lag);
#>   R_RegisterCCallable("rx_65b62eead1e4df27bcbcd47f565c3c36_","rx_65b62eead1e4df27bcbcd47f565c3c36__Rate", (DL_FUNC) rx_65b62eead1e4df27bcbcd47f565c3c36__Rate);
#>   R_RegisterCCallable("rx_65b62eead1e4df27bcbcd47f565c3c36_","rx_65b62eead1e4df27bcbcd47f565c3c36__Dur", (DL_FUNC) rx_65b62eead1e4df27bcbcd47f565c3c36__Dur);
#>   R_RegisterCCallable("rx_65b62eead1e4df27bcbcd47f565c3c36_","rx_65b62eead1e4df27bcbcd47f565c3c36__mtime", (DL_FUNC) rx_65b62eead1e4df27bcbcd47f565c3c36__mtime);
#>   R_RegisterCCallable("rx_65b62eead1e4df27bcbcd47f565c3c36_","rx_65b62eead1e4df27bcbcd47f565c3c36__ME", (DL_FUNC) rx_65b62eead1e4df27bcbcd47f565c3c36__ME);
#>   R_RegisterCCallable("rx_65b62eead1e4df27bcbcd47f565c3c36_","rx_65b62eead1e4df27bcbcd47f565c3c36__IndF", (DL_FUNC) rx_65b62eead1e4df27bcbcd47f565c3c36__IndF);
#>   R_RegisterCCallable("rx_65b62eead1e4df27bcbcd47f565c3c36_","rx_65b62eead1e4df27bcbcd47f565c3c36__dydt_liblsoda", (DL_FUNC) rx_65b62eead1e4df27bcbcd47f565c3c36__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_rx_65b62eead1e4df27bcbcd47f565c3c36_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_rx_65b62eead1e4df27bcbcd47f565c3c36_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"rx_65b62eead1e4df27bcbcd47f565c3c36__model_vars", (DL_FUNC) &rx_65b62eead1e4df27bcbcd47f565c3c36__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_rx_65b62eead1e4df27bcbcd47f565c3c36_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_65b62eead1e4df27bcbcd47f565c3c36__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("rx_65b62eead1e4df27bcbcd47f565c3c36__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’
e <- et(c(-1, 0, 1, 2))
rxSolve(x, e)
#> ── Solved rxode2 object ──
#> ── Parameters (value$params): ──
#> # A tibble: 1 × 0
#> ── Initial Conditions (value$inits): ──
#> named numeric(0)
#> ── First part of data (object): ──
#> # A tibble: 4 × 2
#>    time      r
#>   <dbl>  <dbl>
#> 1    -1 0.0723
#> 2     0 0.5   
#> 3     1 0.928 
#> 4     2 1.09