Default Softplus Activation Function

Default Softplus Activation Function

Usage

dsoftplus(x)

d2softplus(x)

d3softplus(x)

d4softplus(x)

Arguments

x

numeric vector

Value

numeric vector

See also

Author

Matthew L. Fidler

Examples

dsoftplus(c(-1, 0, 1, 2))
#> [1] 0.2689414 0.5000000 0.7310586 0.8807971

# You can use rxode2 too:

r <- rxode2({
 s1 <- dsoftplus(time)
})
#>  
#>  
#> C model:
#> // Define 1 LHS values
#> #define _LHS_0_ 0 // s1
#> #define _getRxSolve_ _rx4dc9229114df2c2febf5990dfa38e4f61859__getRxSolve__leW9
#> #define _evalUdf _rx4dc9229114df2c2febf5990dfa38e4f61860__evalUdf_goD3
#> #define _solveData _rx4dc9229114df2c2febf5990dfa38e4f61861__solveData_R39J
#> #define _assign_ptr _rx4dc9229114df2c2febf5990dfa38e4f61862__assign_ptr_vcr1
#> #define _rxRmModelLib _rx4dc9229114df2c2febf5990dfa38e4f61863__rxRmModelLib_UmLF
#> #define _rxGetModelLib _rx4dc9229114df2c2febf5990dfa38e4f61864__rxGetModelLib_GL1z
#> #define _old_c _rx4dc9229114df2c2febf5990dfa38e4f61865__old_c_NFD8
#> #define _ptrid _rx4dc9229114df2c2febf5990dfa38e4f61866__ptrid_00Mr
#> #define _rxIsCurrentC _rx4dc9229114df2c2febf5990dfa38e4f61867__rxIsCurrentC_OclB
#> #define _sumPS _rx4dc9229114df2c2febf5990dfa38e4f61868__sumPS_dvcH
#> #define _prodPS _rx4dc9229114df2c2febf5990dfa38e4f61869__prodPS_YHlO
#> #define _prodType _rx4dc9229114df2c2febf5990dfa38e4f61870__prodType_Rzkg
#> #define _sumType _rx4dc9229114df2c2febf5990dfa38e4f61871__sumType_9phZ
#> #define _update_par_ptr _rx4dc9229114df2c2febf5990dfa38e4f61872__update_par_ptr_jbtv
#> #define _getParCov _rx4dc9229114df2c2febf5990dfa38e4f61873__getParCov_sFQk
#> #define _rxode2_rxAssignPtr _rx4dc9229114df2c2febf5990dfa38e4f61874__rxode2_rxAssignPtr_pVHF
#> #define _rxQr _rx4dc9229114df2c2febf5990dfa38e4f61875__rxQr_4tE5
#> #define _compareFactorVal _rx4dc9229114df2c2febf5990dfa38e4f61876__compareFactorVal_bqrW
#> #define _sum _rx4dc9229114df2c2febf5990dfa38e4f61877__sum_j2jI
#> #define _udf _rx4dc9229114df2c2febf5990dfa38e4f61878__udf_RW0c
#> #define _sign _rx4dc9229114df2c2febf5990dfa38e4f61879__sign_rrgc
#> #define _prod _rx4dc9229114df2c2febf5990dfa38e4f61880__prod_GmEs
#> #define _max _rx4dc9229114df2c2febf5990dfa38e4f61881__max_Ju9x
#> #define _min _rx4dc9229114df2c2febf5990dfa38e4f61882__min_ahgB
#> #define _transit4P _rx4dc9229114df2c2febf5990dfa38e4f61883__transit4P_OZl3
#> #define _transit3P _rx4dc9229114df2c2febf5990dfa38e4f61884__transit3P_zYvT
#> #define _assignFuns0 _rx4dc9229114df2c2febf5990dfa38e4f61885__assignFuns0_a97E
#> #define _assignFuns _rx4dc9229114df2c2febf5990dfa38e4f61886__assignFuns_hW3M
#> #define _rxord _rx4dc9229114df2c2febf5990dfa38e4f61887__rxord_G9V3
#> #define __assignFuns2 _rx4dc9229114df2c2febf5990dfa38e4f61888___assignFuns2_pR1a
#> #define _llikCauchyDscale _rx4dc9229114df2c2febf5990dfa38e4f61889__llikCauchyDscale_NkG5
#> #define _llikCauchyDlocation _rx4dc9229114df2c2febf5990dfa38e4f61890__llikCauchyDlocation_JwD9
#> #define _llikCauchy _rx4dc9229114df2c2febf5990dfa38e4f61891__llikCauchy_TK40
#> #define _llikGammaDrate _rx4dc9229114df2c2febf5990dfa38e4f61892__llikGammaDrate_NPJi
#> #define _llikGammaDshape _rx4dc9229114df2c2febf5990dfa38e4f61893__llikGammaDshape_ALZB
#> #define _llikGamma _rx4dc9229114df2c2febf5990dfa38e4f61894__llikGamma_rhef
#> #define _llikWeibullDscale _rx4dc9229114df2c2febf5990dfa38e4f61895__llikWeibullDscale_ghDr
#> #define _llikWeibullDshape _rx4dc9229114df2c2febf5990dfa38e4f61896__llikWeibullDshape_m85i
#> #define _llikWeibull _rx4dc9229114df2c2febf5990dfa38e4f61897__llikWeibull_gvlJ
#> #define _llikUnifDbeta _rx4dc9229114df2c2febf5990dfa38e4f61898__llikUnifDbeta_YgGq
#> #define _llikUnifDalpha _rx4dc9229114df2c2febf5990dfa38e4f61899__llikUnifDalpha_QxLo
#> #define _llikUnif _rx4dc9229114df2c2febf5990dfa38e4f61900__llikUnif_YlFg
#> #define _llikGeomDp _rx4dc9229114df2c2febf5990dfa38e4f61901__llikGeomDp_JdKa
#> #define _llikGeom _rx4dc9229114df2c2febf5990dfa38e4f61902__llikGeom_FgmD
#> #define _llikFDdf2 _rx4dc9229114df2c2febf5990dfa38e4f61903__llikFDdf2_R4jx
#> #define _llikFDdf1 _rx4dc9229114df2c2febf5990dfa38e4f61904__llikFDdf1_uoP8
#> #define _llikF _rx4dc9229114df2c2febf5990dfa38e4f61905__llikF_gmoU
#> #define _llikExpDrate _rx4dc9229114df2c2febf5990dfa38e4f61906__llikExpDrate_omwL
#> #define _llikExp _rx4dc9229114df2c2febf5990dfa38e4f61907__llikExp_tuDg
#> #define _llikChisqDdf _rx4dc9229114df2c2febf5990dfa38e4f61908__llikChisqDdf_OztP
#> #define _llikChisq _rx4dc9229114df2c2febf5990dfa38e4f61909__llikChisq_ZwDq
#> #define _llikTDsd _rx4dc9229114df2c2febf5990dfa38e4f61910__llikTDsd_x2t0
#> #define _llikTDmean _rx4dc9229114df2c2febf5990dfa38e4f61911__llikTDmean_FfCR
#> #define _llikTDdf _rx4dc9229114df2c2febf5990dfa38e4f61912__llikTDdf_upBf
#> #define _llikT _rx4dc9229114df2c2febf5990dfa38e4f61913__llikT_cLLl
#> #define _llikBetaDshape2 _rx4dc9229114df2c2febf5990dfa38e4f61914__llikBetaDshape2_Jk0H
#> #define _llikBetaDshape1 _rx4dc9229114df2c2febf5990dfa38e4f61915__llikBetaDshape1_Ouqu
#> #define _llikBeta _rx4dc9229114df2c2febf5990dfa38e4f61916__llikBeta_MQIP
#> #define _llikNbinomMuDmu _rx4dc9229114df2c2febf5990dfa38e4f61917__llikNbinomMuDmu_4sBK
#> #define _llikNbinomMu _rx4dc9229114df2c2febf5990dfa38e4f61918__llikNbinomMu_wlIO
#> #define _llikNbinomDprob _rx4dc9229114df2c2febf5990dfa38e4f61919__llikNbinomDprob_N1tT
#> #define _llikNbinom _rx4dc9229114df2c2febf5990dfa38e4f61920__llikNbinom_X5Du
#> #define _llikBinomDprob _rx4dc9229114df2c2febf5990dfa38e4f61921__llikBinomDprob_lHLx
#> #define _llikBinom _rx4dc9229114df2c2febf5990dfa38e4f61922__llikBinom_eAUt
#> #define _llikPoisDlambda _rx4dc9229114df2c2febf5990dfa38e4f61923__llikPoisDlambda_0po6
#> #define _llikPois _rx4dc9229114df2c2febf5990dfa38e4f61924__llikPois_AqSQ
#> #define _llikNormDsd _rx4dc9229114df2c2febf5990dfa38e4f61925__llikNormDsd_knH8
#> #define _llikNormDmean _rx4dc9229114df2c2febf5990dfa38e4f61926__llikNormDmean_fPCA
#> #define _llikNorm _rx4dc9229114df2c2febf5990dfa38e4f61927__llikNorm_nNkU
#> #define simeps _rx4dc9229114df2c2febf5990dfa38e4f61928_simeps_mswi
#> #define simeta _rx4dc9229114df2c2febf5990dfa38e4f61929_simeta_OWG1
#> #define expit _rx4dc9229114df2c2febf5990dfa38e4f61930_expit_6HPU
#> #define logit _rx4dc9229114df2c2febf5990dfa38e4f61931_logit_gnCj
#> #define gammapDer _rx4dc9229114df2c2febf5990dfa38e4f61932_gammapDer_uwA8
#> #define lowergamma _rx4dc9229114df2c2febf5990dfa38e4f61933_lowergamma_iwbn
#> #define uppergamma _rx4dc9229114df2c2febf5990dfa38e4f61934_uppergamma_2GTQ
#> #define gammaqInva _rx4dc9229114df2c2febf5990dfa38e4f61935_gammaqInva_jSre
#> #define gammaqInv _rx4dc9229114df2c2febf5990dfa38e4f61936_gammaqInv_r5ze
#> #define gammapInva _rx4dc9229114df2c2febf5990dfa38e4f61937_gammapInva_xLgA
#> #define gammapInv _rx4dc9229114df2c2febf5990dfa38e4f61938_gammapInv_wME6
#> #define gammaq _rx4dc9229114df2c2febf5990dfa38e4f61939_gammaq_nqR3
#> #define gammap _rx4dc9229114df2c2febf5990dfa38e4f61940_gammap_DD9j
#> #define phi _rx4dc9229114df2c2febf5990dfa38e4f61941_phi_N5O4
#> #define d2ELUa _rx4dc9229114df2c2febf5990dfa38e4f61942_d2ELUa_r9D9
#> #define dELUa _rx4dc9229114df2c2febf5990dfa38e4f61943_dELUa_7me5
#> #define d2aELU _rx4dc9229114df2c2febf5990dfa38e4f61944_d2aELU_ybV8
#> #define d2ELU _rx4dc9229114df2c2febf5990dfa38e4f61945_d2ELU_EuZ1
#> #define dELU _rx4dc9229114df2c2febf5990dfa38e4f61946_dELU_xTQY
#> #define ELU _rx4dc9229114df2c2febf5990dfa38e4f61947_ELU_OQe6
#> #define dPReLUa1 _rx4dc9229114df2c2febf5990dfa38e4f61948_dPReLUa1_aqt3
#> #define dPReLUa _rx4dc9229114df2c2febf5990dfa38e4f61949_dPReLUa_Lbsb
#> #define dPReLU _rx4dc9229114df2c2febf5990dfa38e4f61950_dPReLU_eaw5
#> #define PReLU _rx4dc9229114df2c2febf5990dfa38e4f61951_PReLU_m4br
#> #define dSwish _rx4dc9229114df2c2febf5990dfa38e4f61952_dSwish_D9EZ
#> #define Swish _rx4dc9229114df2c2febf5990dfa38e4f61953_Swish_AJNe
#> #define dlReLU _rx4dc9229114df2c2febf5990dfa38e4f61954_dlReLU_7IxU
#> #define lReLU _rx4dc9229114df2c2febf5990dfa38e4f61955_lReLU_FA5R
#> #define dSELU _rx4dc9229114df2c2febf5990dfa38e4f61956_dSELU_2wZw
#> #define SELU _rx4dc9229114df2c2febf5990dfa38e4f61957_SELU_wBgR
#> #define d4softplus _rx4dc9229114df2c2febf5990dfa38e4f61958_d4softplus_vbln
#> #define d3softplus _rx4dc9229114df2c2febf5990dfa38e4f61959_d3softplus_k357
#> #define d2softplus _rx4dc9229114df2c2febf5990dfa38e4f61960_d2softplus_cfUw
#> #define dsoftplus _rx4dc9229114df2c2febf5990dfa38e4f61961_dsoftplus_MarE
#> #define softplus _rx4dc9229114df2c2febf5990dfa38e4f61962_softplus_A9z1
#> #define d4GELU _rx4dc9229114df2c2febf5990dfa38e4f61963_d4GELU_SE1K
#> #define d3GELU _rx4dc9229114df2c2febf5990dfa38e4f61964_d3GELU_zAf2
#> #define d2GELU _rx4dc9229114df2c2febf5990dfa38e4f61965_d2GELU_H4Q2
#> #define dGELU _rx4dc9229114df2c2febf5990dfa38e4f61966_dGELU_oF0j
#> #define GELU _rx4dc9229114df2c2febf5990dfa38e4f61967_GELU_oCGj
#> #define dReLU _rx4dc9229114df2c2febf5990dfa38e4f61968_dReLU_0a8w
#> #define ReLU _rx4dc9229114df2c2febf5990dfa38e4f61969_ReLU_AbQE
#> #define riweibull _rx4dc9229114df2c2febf5990dfa38e4f61970_riweibull_LOlO
#> #define riunif _rx4dc9229114df2c2febf5990dfa38e4f61971_riunif_bqNa
#> #define rit_ _rx4dc9229114df2c2febf5990dfa38e4f61972_rit__thlt
#> #define ripois _rx4dc9229114df2c2febf5990dfa38e4f61973_ripois_iWR2
#> #define ribeta _rx4dc9229114df2c2febf5990dfa38e4f61974_ribeta_IGVL
#> #define rigamma _rx4dc9229114df2c2febf5990dfa38e4f61975_rigamma_QYM8
#> #define rigeom _rx4dc9229114df2c2febf5990dfa38e4f61976_rigeom_4JCn
#> #define rif _rx4dc9229114df2c2febf5990dfa38e4f61977_rif_uyCX
#> #define riexp _rx4dc9229114df2c2febf5990dfa38e4f61978_riexp_PJUj
#> #define richisq _rx4dc9229114df2c2febf5990dfa38e4f61979_richisq_vFk7
#> #define ricauchy _rx4dc9229114df2c2febf5990dfa38e4f61980_ricauchy_KJvD
#> #define rinbinomMu _rx4dc9229114df2c2febf5990dfa38e4f61981_rinbinomMu_xqbV
#> #define rinbinom _rx4dc9229114df2c2febf5990dfa38e4f61982_rinbinom_SKVf
#> #define ribinom _rx4dc9229114df2c2febf5990dfa38e4f61983_ribinom_r7Tc
#> #define rinorm _rx4dc9229114df2c2febf5990dfa38e4f61984_rinorm_aUEg
#> #define rxweibull _rx4dc9229114df2c2febf5990dfa38e4f61985_rxweibull_pYkN
#> #define rxunif _rx4dc9229114df2c2febf5990dfa38e4f61986_rxunif_IqnE
#> #define rxt_ _rx4dc9229114df2c2febf5990dfa38e4f61987_rxt__nmhS
#> #define rxpois _rx4dc9229114df2c2febf5990dfa38e4f61988_rxpois_boMZ
#> #define rxbeta _rx4dc9229114df2c2febf5990dfa38e4f61989_rxbeta_YBjk
#> #define rxgamma _rx4dc9229114df2c2febf5990dfa38e4f61990_rxgamma_iCoU
#> #define rxgeom _rx4dc9229114df2c2febf5990dfa38e4f61991_rxgeom_qvni
#> #define rxf _rx4dc9229114df2c2febf5990dfa38e4f61992_rxf_I75U
#> #define rxexp _rx4dc9229114df2c2febf5990dfa38e4f61993_rxexp_i0GV
#> #define rxchisq _rx4dc9229114df2c2febf5990dfa38e4f61994_rxchisq_Ts34
#> #define rxcauchy _rx4dc9229114df2c2febf5990dfa38e4f61995_rxcauchy_TNHe
#> #define rxnbinomMu _rx4dc9229114df2c2febf5990dfa38e4f61996_rxnbinomMu_Fpue
#> #define rxnbinom _rx4dc9229114df2c2febf5990dfa38e4f61997_rxnbinom_1ZaE
#> #define rxbinom _rx4dc9229114df2c2febf5990dfa38e4f61998_rxbinom_nJR6
#> #define rxnorm _rx4dc9229114df2c2febf5990dfa38e4f61999_rxnorm_UHZK
#> #define linCmtB _rx4dc9229114df2c2febf5990dfa38e4f62000_linCmtB_bSsm
#> #define linCmtA _rx4dc9229114df2c2febf5990dfa38e4f62001_linCmtA_NYlS
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){}
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){}
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  rx_ac4fc6d7cd652c852899096d9f7b7748__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *rx_ac4fc6d7cd652c852899096d9f7b7748__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP rx_ac4fc6d7cd652c852899096d9f7b7748__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void rx_ac4fc6d7cd652c852899096d9f7b7748__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double s1 = NA_REAL;
#> 
#>   (void)t;
#>   (void)s1;
#> 
#>   s1 = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   s1 =dsoftplus(t);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void rx_ac4fc6d7cd652c852899096d9f7b7748__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void rx_ac4fc6d7cd652c852899096d9f7b7748__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void rx_ac4fc6d7cd652c852899096d9f7b7748__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double s1 = NA_REAL;
#> 
#>   (void)t;
#>   (void)s1;
#> 
#>   s1 = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   s1 =dsoftplus(t);
#> 
#>   _lhs[_LHS_0_]=s1;
#> }
#> // Functional based bioavailability
#> double rx_ac4fc6d7cd652c852899096d9f7b7748__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double rx_ac4fc6d7cd652c852899096d9f7b7748__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double rx_ac4fc6d7cd652c852899096d9f7b7748__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double rx_ac4fc6d7cd652c852899096d9f7b7748__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void rx_ac4fc6d7cd652c852899096d9f7b7748__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void rx_ac4fc6d7cd652c852899096d9f7b7748__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void rx_ac4fc6d7cd652c852899096d9f7b7748__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP rx_ac4fc6d7cd652c852899096d9f7b7748__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_ac4fc6d7cd652c852899096d9f7b7748__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAAv(zAAAv7#aT)umT,MA**4+<WdwYB~1tU8RZeQ,`5CL_CB88Pn|qj};DdeEf>,R>1{A5)|LYBBMEMIF){6~T~|:I]racqv<$SgGlB|LC\"p13H!rqm}Zj$91>uw/.pq!Q[IV!.<wli51_Fk{|d4Ejmd,S=_X!=AK`4P4}|be7:2}w<n3PK.jF5*m)R\?[Ne,[*cB1kERTltk<}(PwZb,!\"5mKGZ~,ZETJkG4hXbaiUflj;PbQ~v2hw@B<HWJd}EoV^VqI^L(vFhA%%ZMljk[AZBKb\")dwQ3D^ZEBfi}h=k]kW5fPBiLRzgmxee6>F!,HykHd.ZIv*k$n%%oLY_\"Y1Y%%;vkYWwq2d\?B+o[_(;U\"n}ZS5nhZ+tI*hay2]^4^!Ys#A<by*AllHStg)SR7Xx;1SjD&\"d9fLzN73\?8As1j(^|(8uCM*\?.u*,}u_|9YJxvgfS7vyk_3,Q(.(.$l5E~fnGPd/$|k]P)=1EcJtwXHm.AQTi_sn){J{>A@M,p$0N~/M^I_;,Wdi7f3{7/fIQx8wJK[xgPitO1g/_HyB7X8IyMp$#tpUasDv+i/+<GOAM%%FIEo40y{x[CnB!V_VLZ&CMtn)jKu]]_MVL#43F6W\"@~KF{j}ZAr8xUbN:w5J][SI3u:M1o\"6ejSr~aeL!/R~#O+IgYXgB+_{=iMn_\"_xcyaxShkV:aT%%3qo)~p[1JA=~=r]ayw=0Loq@H;6l*@ob&|Rw#AcaizZ5cAu^5i#dJ_HGJZXCzE0/B.zZBm\"[1m0eBKt~7)ZgL2B|XXzYYMMEwim1K;XaeJCs#/`{=\"|js4s\?n6y5;C:^q*zT{t.i!0mE3+7btC`&|)sTVGpd1C`nuaRjPIV|yYT#8E!Qj<4k|T>PJ5d8qT|k^y+sz9p%%tzW)wxchAX~(f|dy(0bjh]3;LyZjK3r6\")9%%{2RUM6[IXwZV&C~,l~EWK(A#}:Fojl)_VgPaKL8vDUEk4+aVTyy10B%%bBACI}=.,/mCK!7eA3HzPID2&2JOxA~0\?#Z[d0IBF@,Ye9DpMA");
#>     char buf[1059];
#> #define __doBufN__ 1059
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void rx_ac4fc6d7cd652c852899096d9f7b7748__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   rx_ac4fc6d7cd652c852899096d9f7b7748__dydt(neq, *t, A, DADT);
#> }
#> extern int rx_ac4fc6d7cd652c852899096d9f7b7748__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   rx_ac4fc6d7cd652c852899096d9f7b7748__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void rx_ac4fc6d7cd652c852899096d9f7b7748__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   rx_ac4fc6d7cd652c852899096d9f7b7748__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void rx_ac4fc6d7cd652c852899096d9f7b7748__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_rx_ac4fc6d7cd652c852899096d9f7b7748_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("rx_ac4fc6d7cd652c852899096d9f7b7748_","rx_ac4fc6d7cd652c852899096d9f7b7748__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("rx_ac4fc6d7cd652c852899096d9f7b7748_","rx_ac4fc6d7cd652c852899096d9f7b7748__assignFuns", (DL_FUNC) rx_ac4fc6d7cd652c852899096d9f7b7748__assignFuns);
#>   R_RegisterCCallable("rx_ac4fc6d7cd652c852899096d9f7b7748_","rx_ac4fc6d7cd652c852899096d9f7b7748__inis",(DL_FUNC) rx_ac4fc6d7cd652c852899096d9f7b7748__inis);
#>   R_RegisterCCallable("rx_ac4fc6d7cd652c852899096d9f7b7748_","rx_ac4fc6d7cd652c852899096d9f7b7748__dydt",(DL_FUNC) rx_ac4fc6d7cd652c852899096d9f7b7748__dydt);
#>   R_RegisterCCallable("rx_ac4fc6d7cd652c852899096d9f7b7748_","rx_ac4fc6d7cd652c852899096d9f7b7748__calc_lhs",(DL_FUNC) rx_ac4fc6d7cd652c852899096d9f7b7748__calc_lhs);
#>   R_RegisterCCallable("rx_ac4fc6d7cd652c852899096d9f7b7748_","rx_ac4fc6d7cd652c852899096d9f7b7748__calc_jac",(DL_FUNC) rx_ac4fc6d7cd652c852899096d9f7b7748__calc_jac);
#>   R_RegisterCCallable("rx_ac4fc6d7cd652c852899096d9f7b7748_","rx_ac4fc6d7cd652c852899096d9f7b7748__dydt_lsoda", (DL_FUNC) rx_ac4fc6d7cd652c852899096d9f7b7748__dydt_lsoda);
#>   R_RegisterCCallable("rx_ac4fc6d7cd652c852899096d9f7b7748_","rx_ac4fc6d7cd652c852899096d9f7b7748__calc_jac_lsoda", (DL_FUNC) rx_ac4fc6d7cd652c852899096d9f7b7748__calc_jac_lsoda);
#>   R_RegisterCCallable("rx_ac4fc6d7cd652c852899096d9f7b7748_","rx_ac4fc6d7cd652c852899096d9f7b7748__ode_solver_solvedata", (DL_FUNC) rx_ac4fc6d7cd652c852899096d9f7b7748__ode_solver_solvedata);
#>   R_RegisterCCallable("rx_ac4fc6d7cd652c852899096d9f7b7748_","rx_ac4fc6d7cd652c852899096d9f7b7748__ode_solver_get_solvedata", (DL_FUNC) rx_ac4fc6d7cd652c852899096d9f7b7748__ode_solver_get_solvedata);
#>   R_RegisterCCallable("rx_ac4fc6d7cd652c852899096d9f7b7748_","rx_ac4fc6d7cd652c852899096d9f7b7748__F", (DL_FUNC) rx_ac4fc6d7cd652c852899096d9f7b7748__F);
#>   R_RegisterCCallable("rx_ac4fc6d7cd652c852899096d9f7b7748_","rx_ac4fc6d7cd652c852899096d9f7b7748__Lag", (DL_FUNC) rx_ac4fc6d7cd652c852899096d9f7b7748__Lag);
#>   R_RegisterCCallable("rx_ac4fc6d7cd652c852899096d9f7b7748_","rx_ac4fc6d7cd652c852899096d9f7b7748__Rate", (DL_FUNC) rx_ac4fc6d7cd652c852899096d9f7b7748__Rate);
#>   R_RegisterCCallable("rx_ac4fc6d7cd652c852899096d9f7b7748_","rx_ac4fc6d7cd652c852899096d9f7b7748__Dur", (DL_FUNC) rx_ac4fc6d7cd652c852899096d9f7b7748__Dur);
#>   R_RegisterCCallable("rx_ac4fc6d7cd652c852899096d9f7b7748_","rx_ac4fc6d7cd652c852899096d9f7b7748__mtime", (DL_FUNC) rx_ac4fc6d7cd652c852899096d9f7b7748__mtime);
#>   R_RegisterCCallable("rx_ac4fc6d7cd652c852899096d9f7b7748_","rx_ac4fc6d7cd652c852899096d9f7b7748__ME", (DL_FUNC) rx_ac4fc6d7cd652c852899096d9f7b7748__ME);
#>   R_RegisterCCallable("rx_ac4fc6d7cd652c852899096d9f7b7748_","rx_ac4fc6d7cd652c852899096d9f7b7748__IndF", (DL_FUNC) rx_ac4fc6d7cd652c852899096d9f7b7748__IndF);
#>   R_RegisterCCallable("rx_ac4fc6d7cd652c852899096d9f7b7748_","rx_ac4fc6d7cd652c852899096d9f7b7748__dydt_liblsoda", (DL_FUNC) rx_ac4fc6d7cd652c852899096d9f7b7748__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_rx_ac4fc6d7cd652c852899096d9f7b7748_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_rx_ac4fc6d7cd652c852899096d9f7b7748_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"rx_ac4fc6d7cd652c852899096d9f7b7748__model_vars", (DL_FUNC) &rx_ac4fc6d7cd652c852899096d9f7b7748__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_rx_ac4fc6d7cd652c852899096d9f7b7748_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_ac4fc6d7cd652c852899096d9f7b7748__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("rx_ac4fc6d7cd652c852899096d9f7b7748__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’

e <- et(c(-1, 0, 1, 2))

rxSolve(r, e)
#> ── Solved rxode2 object ──
#> ── Parameters (value$params): ──
#> # A tibble: 1 × 0
#> ── Initial Conditions (value$inits): ──
#> named numeric(0)
#> ── First part of data (object): ──
#> # A tibble: 4 × 2
#>    time    s1
#>   <dbl> <dbl>
#> 1    -1 0.269
#> 2     0 0.5  
#> 3     1 0.731
#> 4     2 0.881