Calculate the log likelihood of the binomial function (and its derivatives)

Calculate the log likelihood of the binomial function (and its derivatives)

Usage

llikBinom(x, size, prob, full = FALSE)

Arguments

x

Number of successes

size

Size of trial

prob

probability of success

full

Add the data frame showing x, mean, sd as well as the fx and derivatives

Value

data frame with fx for the pdf value of with dProb that has the derivatives with respect to the parameters at the observation time-point

Details

In an rxode2() model, you can use llikBinom() but you have to use all arguments. You can also get the derivative of prob with llikBinomDprob()

Author

Matthew L. Fidler

Examples

# \donttest{
llikBinom(46:54, 100, 0.5)
#>          fx dProb
#> 1 -2.848030   -16
#> 2 -2.709193   -12
#> 3 -2.610102    -8
#> 4 -2.550679    -4
#> 5 -2.530876     0
#> 6 -2.550679     4
#> 7 -2.610102     8
#> 8 -2.709193    12
#> 9 -2.848030    16

llikBinom(46:54, 100, 0.5, TRUE)
#>    x size prob        fx dProb
#> 1 46  100  0.5 -2.848030   -16
#> 2 47  100  0.5 -2.709193   -12
#> 3 48  100  0.5 -2.610102    -8
#> 4 49  100  0.5 -2.550679    -4
#> 5 50  100  0.5 -2.530876     0
#> 6 51  100  0.5 -2.550679     4
#> 7 52  100  0.5 -2.610102     8
#> 8 53  100  0.5 -2.709193    12
#> 9 54  100  0.5 -2.848030    16

# In rxode2 you can use:

et <- et(46:54)
et$size <- 100
et$prob <-0.5

model <- function() {
  model({
    fx <- llikBinom(time, size, prob)
    dProb <- llikBinomDprob(time, size, prob)
  })
}

rxSolve(model, et)
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#>  
#>  
#> C model:
#> // Define 2 LHS values
#> #define _LHS_0_ 0 // fx
#> #define _LHS_1_ 1 // dProb
#> #define _getRxSolve_ _rxb6bc541f52db4a756ac2c9813808a98b2431__getRxSolve__kg9p
#> #define _evalUdf _rxb6bc541f52db4a756ac2c9813808a98b2432__evalUdf_hZjb
#> #define _solveData _rxb6bc541f52db4a756ac2c9813808a98b2433__solveData_tvsF
#> #define _assign_ptr _rxb6bc541f52db4a756ac2c9813808a98b2434__assign_ptr_QkpV
#> #define _rxRmModelLib _rxb6bc541f52db4a756ac2c9813808a98b2435__rxRmModelLib_HF4t
#> #define _rxGetModelLib _rxb6bc541f52db4a756ac2c9813808a98b2436__rxGetModelLib_E5bq
#> #define _old_c _rxb6bc541f52db4a756ac2c9813808a98b2437__old_c_rWj2
#> #define _ptrid _rxb6bc541f52db4a756ac2c9813808a98b2438__ptrid_jIRW
#> #define _rxIsCurrentC _rxb6bc541f52db4a756ac2c9813808a98b2439__rxIsCurrentC_0crr
#> #define _sumPS _rxb6bc541f52db4a756ac2c9813808a98b2440__sumPS_gcGm
#> #define _prodPS _rxb6bc541f52db4a756ac2c9813808a98b2441__prodPS_EsJu
#> #define _prodType _rxb6bc541f52db4a756ac2c9813808a98b2442__prodType_9xah
#> #define _sumType _rxb6bc541f52db4a756ac2c9813808a98b2443__sumType_gBOZ
#> #define _update_par_ptr _rxb6bc541f52db4a756ac2c9813808a98b2444__update_par_ptr_l3zY
#> #define _getParCov _rxb6bc541f52db4a756ac2c9813808a98b2445__getParCov_vTa9
#> #define _rxode2_rxAssignPtr _rxb6bc541f52db4a756ac2c9813808a98b2446__rxode2_rxAssignPtr_7EhW
#> #define _rxQr _rxb6bc541f52db4a756ac2c9813808a98b2447__rxQr_3MG9
#> #define _compareFactorVal _rxb6bc541f52db4a756ac2c9813808a98b2448__compareFactorVal_V3pR
#> #define _sum _rxb6bc541f52db4a756ac2c9813808a98b2449__sum_1aNk
#> #define _udf _rxb6bc541f52db4a756ac2c9813808a98b2450__udf_G5Jw
#> #define _sign _rxb6bc541f52db4a756ac2c9813808a98b2451__sign_D9TK
#> #define _prod _rxb6bc541f52db4a756ac2c9813808a98b2452__prod_40NP
#> #define _max _rxb6bc541f52db4a756ac2c9813808a98b2453__max_JiAL
#> #define _min _rxb6bc541f52db4a756ac2c9813808a98b2454__min_ZBrh
#> #define _transit4P _rxb6bc541f52db4a756ac2c9813808a98b2455__transit4P_efgh
#> #define _transit3P _rxb6bc541f52db4a756ac2c9813808a98b2456__transit3P_Drm8
#> #define _assignFuns0 _rxb6bc541f52db4a756ac2c9813808a98b2457__assignFuns0_5igv
#> #define _assignFuns _rxb6bc541f52db4a756ac2c9813808a98b2458__assignFuns_lJYg
#> #define _rxord _rxb6bc541f52db4a756ac2c9813808a98b2459__rxord_GqQx
#> #define __assignFuns2 _rxb6bc541f52db4a756ac2c9813808a98b2460___assignFuns2_LoYl
#> #define _llikCauchyDscale _rxb6bc541f52db4a756ac2c9813808a98b2461__llikCauchyDscale_FgJd
#> #define _llikCauchyDlocation _rxb6bc541f52db4a756ac2c9813808a98b2462__llikCauchyDlocation_KaFg
#> #define _llikCauchy _rxb6bc541f52db4a756ac2c9813808a98b2463__llikCauchy_mDR4
#> #define _llikGammaDrate _rxb6bc541f52db4a756ac2c9813808a98b2464__llikGammaDrate_jxuo
#> #define _llikGammaDshape _rxb6bc541f52db4a756ac2c9813808a98b2465__llikGammaDshape_P8gm
#> #define _llikGamma _rxb6bc541f52db4a756ac2c9813808a98b2466__llikGamma_oUom
#> #define _llikWeibullDscale _rxb6bc541f52db4a756ac2c9813808a98b2467__llikWeibullDscale_wLtu
#> #define _llikWeibullDshape _rxb6bc541f52db4a756ac2c9813808a98b2468__llikWeibullDshape_DgOz
#> #define _llikWeibull _rxb6bc541f52db4a756ac2c9813808a98b2469__llikWeibull_tPZw
#> #define _llikUnifDbeta _rxb6bc541f52db4a756ac2c9813808a98b2470__llikUnifDbeta_Dqx2
#> #define _llikUnifDalpha _rxb6bc541f52db4a756ac2c9813808a98b2471__llikUnifDalpha_t0Ff
#> #define _llikUnif _rxb6bc541f52db4a756ac2c9813808a98b2472__llikUnif_CRup
#> #define _llikGeomDp _rxb6bc541f52db4a756ac2c9813808a98b2473__llikGeomDp_BfcL
#> #define _llikGeom _rxb6bc541f52db4a756ac2c9813808a98b2474__llikGeom_LlJk
#> #define _llikFDdf2 _rxb6bc541f52db4a756ac2c9813808a98b2475__llikFDdf2_0HOu
#> #define _llikFDdf1 _rxb6bc541f52db4a756ac2c9813808a98b2476__llikFDdf1_quMQ
#> #define _llikF _rxb6bc541f52db4a756ac2c9813808a98b2477__llikF_IP4s
#> #define _llikExpDrate _rxb6bc541f52db4a756ac2c9813808a98b2478__llikExpDrate_BKwl
#> #define _llikExp _rxb6bc541f52db4a756ac2c9813808a98b2479__llikExp_ION1
#> #define _llikChisqDdf _rxb6bc541f52db4a756ac2c9813808a98b2480__llikChisqDdf_tTX5
#> #define _llikChisq _rxb6bc541f52db4a756ac2c9813808a98b2481__llikChisq_DulH
#> #define _llikTDsd _rxb6bc541f52db4a756ac2c9813808a98b2482__llikTDsd_LxeA
#> #define _llikTDmean _rxb6bc541f52db4a756ac2c9813808a98b2483__llikTDmean_Ut0p
#> #define _llikTDdf _rxb6bc541f52db4a756ac2c9813808a98b2484__llikTDdf_o6Aq
#> #define _llikT _rxb6bc541f52db4a756ac2c9813808a98b2485__llikT_SQkn
#> #define _llikBetaDshape2 _rxb6bc541f52db4a756ac2c9813808a98b2486__llikBetaDshape2_H8fP
#> #define _llikBetaDshape1 _rxb6bc541f52db4a756ac2c9813808a98b2487__llikBetaDshape1_CAnN
#> #define _llikBeta _rxb6bc541f52db4a756ac2c9813808a98b2488__llikBeta_kUms
#> #define _llikNbinomMuDmu _rxb6bc541f52db4a756ac2c9813808a98b2489__llikNbinomMuDmu_wiOW
#> #define _llikNbinomMu _rxb6bc541f52db4a756ac2c9813808a98b2490__llikNbinomMu_G16H
#> #define _llikNbinomDprob _rxb6bc541f52db4a756ac2c9813808a98b2491__llikNbinomDprob_PUgn
#> #define _llikNbinom _rxb6bc541f52db4a756ac2c9813808a98b2492__llikNbinom_Cjuw
#> #define _llikBinomDprob _rxb6bc541f52db4a756ac2c9813808a98b2493__llikBinomDprob_A8iw
#> #define _llikBinom _rxb6bc541f52db4a756ac2c9813808a98b2494__llikBinom_bn2G
#> #define _llikPoisDlambda _rxb6bc541f52db4a756ac2c9813808a98b2495__llikPoisDlambda_TQjS
#> #define _llikPois _rxb6bc541f52db4a756ac2c9813808a98b2496__llikPois_rer5
#> #define _llikNormDsd _rxb6bc541f52db4a756ac2c9813808a98b2497__llikNormDsd_zexL
#> #define _llikNormDmean _rxb6bc541f52db4a756ac2c9813808a98b2498__llikNormDmean_gAwM
#> #define _llikNorm _rxb6bc541f52db4a756ac2c9813808a98b2499__llikNorm_E6nq
#> #define simeps _rxb6bc541f52db4a756ac2c9813808a98b2500_simeps_R3DD
#> #define simeta _rxb6bc541f52db4a756ac2c9813808a98b2501_simeta_9jN5
#> #define expit _rxb6bc541f52db4a756ac2c9813808a98b2502_expit_O4r9
#> #define logit _rxb6bc541f52db4a756ac2c9813808a98b2503_logit_D97m
#> #define gammapDer _rxb6bc541f52db4a756ac2c9813808a98b2504_gammapDer_e5yb
#> #define lowergamma _rxb6bc541f52db4a756ac2c9813808a98b2505_lowergamma_V8Eu
#> #define uppergamma _rxb6bc541f52db4a756ac2c9813808a98b2506_uppergamma_Z1xT
#> #define gammaqInva _rxb6bc541f52db4a756ac2c9813808a98b2507_gammaqInva_QYOQ
#> #define gammaqInv _rxb6bc541f52db4a756ac2c9813808a98b2508_gammaqInv_e6aq
#> #define gammapInva _rxb6bc541f52db4a756ac2c9813808a98b2509_gammapInva_t3Lb
#> #define gammapInv _rxb6bc541f52db4a756ac2c9813808a98b2510_gammapInv_sbea
#> #define gammaq _rxb6bc541f52db4a756ac2c9813808a98b2511_gammaq_w5m4
#> #define gammap _rxb6bc541f52db4a756ac2c9813808a98b2512_gammap_brD9
#> #define phi _rxb6bc541f52db4a756ac2c9813808a98b2513_phi_EZAJ
#> #define d2ELUa _rxb6bc541f52db4a756ac2c9813808a98b2514_d2ELUa_Ne7I
#> #define dELUa _rxb6bc541f52db4a756ac2c9813808a98b2515_dELUa_xUFA
#> #define d2aELU _rxb6bc541f52db4a756ac2c9813808a98b2516_d2aELU_5R2w
#> #define d2ELU _rxb6bc541f52db4a756ac2c9813808a98b2517_d2ELU_ZwwB
#> #define dELU _rxb6bc541f52db4a756ac2c9813808a98b2518_dELU_gRvb
#> #define ELU _rxb6bc541f52db4a756ac2c9813808a98b2519_ELU_lnk3
#> #define dPReLUa1 _rxb6bc541f52db4a756ac2c9813808a98b2520_dPReLUa1_57cf
#> #define dPReLUa _rxb6bc541f52db4a756ac2c9813808a98b2521_dPReLUa_UwMa
#> #define dPReLU _rxb6bc541f52db4a756ac2c9813808a98b2522_dPReLU_rEA9
#> #define PReLU _rxb6bc541f52db4a756ac2c9813808a98b2523_PReLU_z1SE
#> #define dSwish _rxb6bc541f52db4a756ac2c9813808a98b2524_dSwish_1KzA
#> #define Swish _rxb6bc541f52db4a756ac2c9813808a98b2525_Swish_f2H4
#> #define dlReLU _rxb6bc541f52db4a756ac2c9813808a98b2526_dlReLU_Q2oF
#> #define lReLU _rxb6bc541f52db4a756ac2c9813808a98b2527_lReLU_0joC
#> #define dSELU _rxb6bc541f52db4a756ac2c9813808a98b2528_dSELU_Gj0a
#> #define SELU _rxb6bc541f52db4a756ac2c9813808a98b2529_SELU_8wAb
#> #define d4softplus _rxb6bc541f52db4a756ac2c9813808a98b2530_d4softplus_QELO
#> #define d3softplus _rxb6bc541f52db4a756ac2c9813808a98b2531_d3softplus_lObq
#> #define d2softplus _rxb6bc541f52db4a756ac2c9813808a98b2532_d2softplus_Nath
#> #define dsoftplus _rxb6bc541f52db4a756ac2c9813808a98b2533_dsoftplus_ltiW
#> #define softplus _rxb6bc541f52db4a756ac2c9813808a98b2534_softplus_R2IG
#> #define d4GELU _rxb6bc541f52db4a756ac2c9813808a98b2535_d4GELU_VLQY
#> #define d3GELU _rxb6bc541f52db4a756ac2c9813808a98b2536_d3GELU_M84J
#> #define d2GELU _rxb6bc541f52db4a756ac2c9813808a98b2537_d2GELU_Cnuy
#> #define dGELU _rxb6bc541f52db4a756ac2c9813808a98b2538_dGELU_CXPJ
#> #define GELU _rxb6bc541f52db4a756ac2c9813808a98b2539_GELU_UjvF
#> #define dReLU _rxb6bc541f52db4a756ac2c9813808a98b2540_dReLU_k7KJ
#> #define ReLU _rxb6bc541f52db4a756ac2c9813808a98b2541_ReLU_vDxq
#> #define riweibull _rxb6bc541f52db4a756ac2c9813808a98b2542_riweibull_bVSK
#> #define riunif _rxb6bc541f52db4a756ac2c9813808a98b2543_riunif_Vfr7
#> #define rit_ _rxb6bc541f52db4a756ac2c9813808a98b2544_rit__TcaU
#> #define ripois _rxb6bc541f52db4a756ac2c9813808a98b2545_ripois_EgpY
#> #define ribeta _rxb6bc541f52db4a756ac2c9813808a98b2546_ribeta_kNIq
#> #define rigamma _rxb6bc541f52db4a756ac2c9813808a98b2547_rigamma_nEnm
#> #define rigeom _rxb6bc541f52db4a756ac2c9813808a98b2548_rigeom_hSbo
#> #define rif _rxb6bc541f52db4a756ac2c9813808a98b2549_rif_MZYB
#> #define riexp _rxb6bc541f52db4a756ac2c9813808a98b2550_riexp_jkiC
#> #define richisq _rxb6bc541f52db4a756ac2c9813808a98b2551_richisq_oUqv
#> #define ricauchy _rxb6bc541f52db4a756ac2c9813808a98b2552_ricauchy_niI7
#> #define rinbinomMu _rxb6bc541f52db4a756ac2c9813808a98b2553_rinbinomMu_5Ui0
#> #define rinbinom _rxb6bc541f52db4a756ac2c9813808a98b2554_rinbinom_GVTs
#> #define ribinom _rxb6bc541f52db4a756ac2c9813808a98b2555_ribinom_34TN
#> #define rinorm _rxb6bc541f52db4a756ac2c9813808a98b2556_rinorm_HeFp
#> #define rxweibull _rxb6bc541f52db4a756ac2c9813808a98b2557_rxweibull_ue1Z
#> #define rxunif _rxb6bc541f52db4a756ac2c9813808a98b2558_rxunif_aEnJ
#> #define rxt_ _rxb6bc541f52db4a756ac2c9813808a98b2559_rxt__R6UH
#> #define rxpois _rxb6bc541f52db4a756ac2c9813808a98b2560_rxpois_ZKbS
#> #define rxbeta _rxb6bc541f52db4a756ac2c9813808a98b2561_rxbeta_smNY
#> #define rxgamma _rxb6bc541f52db4a756ac2c9813808a98b2562_rxgamma_lS2w
#> #define rxgeom _rxb6bc541f52db4a756ac2c9813808a98b2563_rxgeom_zYqw
#> #define rxf _rxb6bc541f52db4a756ac2c9813808a98b2564_rxf_2z8b
#> #define rxexp _rxb6bc541f52db4a756ac2c9813808a98b2565_rxexp_GMdj
#> #define rxchisq _rxb6bc541f52db4a756ac2c9813808a98b2566_rxchisq_LB3L
#> #define rxcauchy _rxb6bc541f52db4a756ac2c9813808a98b2567_rxcauchy_sOUR
#> #define rxnbinomMu _rxb6bc541f52db4a756ac2c9813808a98b2568_rxnbinomMu_bPYU
#> #define rxnbinom _rxb6bc541f52db4a756ac2c9813808a98b2569_rxnbinom_PmW8
#> #define rxbinom _rxb6bc541f52db4a756ac2c9813808a98b2570_rxbinom_sOmG
#> #define rxnorm _rxb6bc541f52db4a756ac2c9813808a98b2571_rxnorm_ypO5
#> #define linCmtB _rxb6bc541f52db4a756ac2c9813808a98b2572_linCmtB_5OnO
#> #define linCmtA _rxb6bc541f52db4a756ac2c9813808a98b2573_linCmtA_E9TF
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){  if (_solveData->svar[_svari] == 0) {size = _PP[0];};   if (_solveData->svar[_svari] == 1) {prob = _PP[1];}; }
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){  if (_solveData->ovar[_ovari] == 0) {size = _PP[0];};   if (_solveData->ovar[_ovari] == 1) {prob = _PP[1];}; }
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  rx_ab17931ce35cd48dce5b22def5a99928__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *rx_ab17931ce35cd48dce5b22def5a99928__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP rx_ab17931ce35cd48dce5b22def5a99928__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void rx_ab17931ce35cd48dce5b22def5a99928__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double size = NA_REAL;
#>   double prob = NA_REAL;
#>   double fx = NA_REAL;
#>   double dProb = NA_REAL;
#> 
#>   (void)t;
#>   (void)size;
#>   (void)prob;
#>   (void)fx;
#>   (void)dProb;
#> 
#>   fx = _PL[0];
#>   dProb = _PL[1];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#>   size = _PP[0];
#>   prob = _PP[1];
#> 
#> 
#>   fx =llikBinom(t,size,prob);
#>   dProb =llikBinomDprob(t,size,prob);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void rx_ab17931ce35cd48dce5b22def5a99928__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void rx_ab17931ce35cd48dce5b22def5a99928__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void rx_ab17931ce35cd48dce5b22def5a99928__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double size = NA_REAL;
#>   double prob = NA_REAL;
#>   double fx = NA_REAL;
#>   double dProb = NA_REAL;
#> 
#>   (void)t;
#>   (void)size;
#>   (void)prob;
#>   (void)fx;
#>   (void)dProb;
#> 
#>   fx = _PL[0];
#>   dProb = _PL[1];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#>   size = _PP[0];
#>   prob = _PP[1];
#> 
#> 
#>   fx =llikBinom(t,size,prob);
#>   dProb =llikBinomDprob(t,size,prob);
#> 
#>   _lhs[_LHS_0_]=fx;
#>   _lhs[_LHS_1_]=dProb;
#> }
#> // Functional based bioavailability
#> double rx_ab17931ce35cd48dce5b22def5a99928__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double rx_ab17931ce35cd48dce5b22def5a99928__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double rx_ab17931ce35cd48dce5b22def5a99928__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double rx_ab17931ce35cd48dce5b22def5a99928__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void rx_ab17931ce35cd48dce5b22def5a99928__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void rx_ab17931ce35cd48dce5b22def5a99928__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void rx_ab17931ce35cd48dce5b22def5a99928__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP rx_ab17931ce35cd48dce5b22def5a99928__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_ab17931ce35cd48dce5b22def5a99928__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAA6y2AAAv7#aT)JpssNAy>0ah\"W$Tu{RP{@;O5Vxz4SVg6YSoMZ8|19Q:e97^%%CdjTEN4dHFNR+I|,mquu9Ra(>9t~P4*\?&&|JsjFKFtjApB|jjhUM9yDs|+Qk\?f%%>@mD.i*mrSjjx5QpODdT4.(WnH.+bc;]}>&$quwh8\?`K}|l6za3vKn]5iO,q;M6$t8!:@K&KJ{R_6]`+Q~8Gi)k4co9QhZ}5w28mou)P,T1y#koh),[9n9_8X>bs.(1@T*FXYm&;&Qcg|5A)5h+j+aQ6Is,*weXLh1g%%\?td#PTxUQ.G\"hYyh]cR~al0*cSrf!gow+R$DSQOc42^}PyK]Z(GJ5$[\"_I@&tX1<iPd[IDJ|YGoQMZrr*.dYLn;PBRRGOtZlN^.lbH_MFWM8FKYM3TV`R$*<1oI\?Djw_QKYX8qL3w+`(ZRLxR+5_z7[l{hOu:32`BJ\?7\?YK=oLpk1Fn*:!wl|si]$[j`eAKHc}Q#;k[N6W&XKij8fn8PHli.xvIqUrMg49I)/@sW$0j!1|:9P{*+&DvQ8RYGQVQx0GQg#:zd%%t@I/K9/By#mUyWMgNEe\?b4%%C&tqA3!PKtq>WnL~61#`>F4xQ}:QLT3WF,jq1$1tEkmzfCz+&=8Xa(ykpEgi_[Ov2lV)q.X~89XF.h!,zQS<Pg<ZHo&W~pxEl]\?@,*>.zkysk:24b640h7YR*9Lf6HY[;Fq6>P9|kb{<P%%){]y>r7]8rhf!Dy;klouHTI38qz\"o6d2n3\?`d|!Rd%%YBYMh*RfvKe+YlEqNvz/B{/a\"1dgY6%%/AB5/PsB^C}L6YAYvZeM3u%%)9SB4tDJ:Nc`\?m7x:;~4zsL~#3e9fjW7c~x[n{x\?_otP\?GCa@P^J8{5uc4FDPn`MTpmPuMk*m{ck6bw<*4*55[Pe7Fc!I:+G\"UE%%FC8H|SJXjlHU@*aOhc7ODW6s#yen&;sjyruYbURTb.GKD*ru27G%%zr7qIduTFf{WZbs0sLz3hFqp7ZgD_3N_>}Zy\?Ig&A]HzqI(.1}GrHXjy\"G0p.oO/*H(\?V@{)L@<5$)\?diY}S7jKPCx*!N:J9Saq|z|0c$!SIJhMa8VvX)=c");
#>     char buf[1118];
#> #define __doBufN__ 1118
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void rx_ab17931ce35cd48dce5b22def5a99928__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   rx_ab17931ce35cd48dce5b22def5a99928__dydt(neq, *t, A, DADT);
#> }
#> extern int rx_ab17931ce35cd48dce5b22def5a99928__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   rx_ab17931ce35cd48dce5b22def5a99928__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void rx_ab17931ce35cd48dce5b22def5a99928__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   rx_ab17931ce35cd48dce5b22def5a99928__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void rx_ab17931ce35cd48dce5b22def5a99928__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_rx_ab17931ce35cd48dce5b22def5a99928_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("rx_ab17931ce35cd48dce5b22def5a99928_","rx_ab17931ce35cd48dce5b22def5a99928__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("rx_ab17931ce35cd48dce5b22def5a99928_","rx_ab17931ce35cd48dce5b22def5a99928__assignFuns", (DL_FUNC) rx_ab17931ce35cd48dce5b22def5a99928__assignFuns);
#>   R_RegisterCCallable("rx_ab17931ce35cd48dce5b22def5a99928_","rx_ab17931ce35cd48dce5b22def5a99928__inis",(DL_FUNC) rx_ab17931ce35cd48dce5b22def5a99928__inis);
#>   R_RegisterCCallable("rx_ab17931ce35cd48dce5b22def5a99928_","rx_ab17931ce35cd48dce5b22def5a99928__dydt",(DL_FUNC) rx_ab17931ce35cd48dce5b22def5a99928__dydt);
#>   R_RegisterCCallable("rx_ab17931ce35cd48dce5b22def5a99928_","rx_ab17931ce35cd48dce5b22def5a99928__calc_lhs",(DL_FUNC) rx_ab17931ce35cd48dce5b22def5a99928__calc_lhs);
#>   R_RegisterCCallable("rx_ab17931ce35cd48dce5b22def5a99928_","rx_ab17931ce35cd48dce5b22def5a99928__calc_jac",(DL_FUNC) rx_ab17931ce35cd48dce5b22def5a99928__calc_jac);
#>   R_RegisterCCallable("rx_ab17931ce35cd48dce5b22def5a99928_","rx_ab17931ce35cd48dce5b22def5a99928__dydt_lsoda", (DL_FUNC) rx_ab17931ce35cd48dce5b22def5a99928__dydt_lsoda);
#>   R_RegisterCCallable("rx_ab17931ce35cd48dce5b22def5a99928_","rx_ab17931ce35cd48dce5b22def5a99928__calc_jac_lsoda", (DL_FUNC) rx_ab17931ce35cd48dce5b22def5a99928__calc_jac_lsoda);
#>   R_RegisterCCallable("rx_ab17931ce35cd48dce5b22def5a99928_","rx_ab17931ce35cd48dce5b22def5a99928__ode_solver_solvedata", (DL_FUNC) rx_ab17931ce35cd48dce5b22def5a99928__ode_solver_solvedata);
#>   R_RegisterCCallable("rx_ab17931ce35cd48dce5b22def5a99928_","rx_ab17931ce35cd48dce5b22def5a99928__ode_solver_get_solvedata", (DL_FUNC) rx_ab17931ce35cd48dce5b22def5a99928__ode_solver_get_solvedata);
#>   R_RegisterCCallable("rx_ab17931ce35cd48dce5b22def5a99928_","rx_ab17931ce35cd48dce5b22def5a99928__F", (DL_FUNC) rx_ab17931ce35cd48dce5b22def5a99928__F);
#>   R_RegisterCCallable("rx_ab17931ce35cd48dce5b22def5a99928_","rx_ab17931ce35cd48dce5b22def5a99928__Lag", (DL_FUNC) rx_ab17931ce35cd48dce5b22def5a99928__Lag);
#>   R_RegisterCCallable("rx_ab17931ce35cd48dce5b22def5a99928_","rx_ab17931ce35cd48dce5b22def5a99928__Rate", (DL_FUNC) rx_ab17931ce35cd48dce5b22def5a99928__Rate);
#>   R_RegisterCCallable("rx_ab17931ce35cd48dce5b22def5a99928_","rx_ab17931ce35cd48dce5b22def5a99928__Dur", (DL_FUNC) rx_ab17931ce35cd48dce5b22def5a99928__Dur);
#>   R_RegisterCCallable("rx_ab17931ce35cd48dce5b22def5a99928_","rx_ab17931ce35cd48dce5b22def5a99928__mtime", (DL_FUNC) rx_ab17931ce35cd48dce5b22def5a99928__mtime);
#>   R_RegisterCCallable("rx_ab17931ce35cd48dce5b22def5a99928_","rx_ab17931ce35cd48dce5b22def5a99928__ME", (DL_FUNC) rx_ab17931ce35cd48dce5b22def5a99928__ME);
#>   R_RegisterCCallable("rx_ab17931ce35cd48dce5b22def5a99928_","rx_ab17931ce35cd48dce5b22def5a99928__IndF", (DL_FUNC) rx_ab17931ce35cd48dce5b22def5a99928__IndF);
#>   R_RegisterCCallable("rx_ab17931ce35cd48dce5b22def5a99928_","rx_ab17931ce35cd48dce5b22def5a99928__dydt_liblsoda", (DL_FUNC) rx_ab17931ce35cd48dce5b22def5a99928__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_rx_ab17931ce35cd48dce5b22def5a99928_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_rx_ab17931ce35cd48dce5b22def5a99928_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"rx_ab17931ce35cd48dce5b22def5a99928__model_vars", (DL_FUNC) &rx_ab17931ce35cd48dce5b22def5a99928__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_rx_ab17931ce35cd48dce5b22def5a99928_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_ab17931ce35cd48dce5b22def5a99928__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("rx_ab17931ce35cd48dce5b22def5a99928__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’
#> ── Solved rxode2 object ──
#> ── Parameters (value$params): ──
#> # A tibble: 1 × 0
#> ── Initial Conditions (value$inits): ──
#> named numeric(0)
#> ── First part of data (object): ──
#> # A tibble: 9 × 5
#>    time    fx dProb  size  prob
#>   <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1    46 -2.85   -16   100   0.5
#> 2    47 -2.71   -12   100   0.5
#> 3    48 -2.61    -8   100   0.5
#> 4    49 -2.55    -4   100   0.5
#> 5    50 -2.53     0   100   0.5
#> 6    51 -2.55     4   100   0.5
#> # ℹ 3 more rows
# }