log likelihood of Cauchy distribution and it's derivatives (from stan)

log likelihood of Cauchy distribution and it's derivatives (from stan)

Usage

llikCauchy(x, location = 0, scale = 1, full = FALSE)

Arguments

x

Observation

location, scale

location and scale parameters.

full

Add the data frame showing x, mean, sd as well as the fx and derivatives

Value

data frame with fx for the log pdf value of with dLocation and dScale that has the derivatives with respect to the parameters at the observation time-point

Details

In an rxode2() model, you can use llikCauchy() but you have to use all arguments. You can also get the derivative of location and scale with llikCauchyDlocation() and llikCauchyDscale().

Author

Matthew L. Fidler

Examples

# \donttest{
x <- seq(-3, 3, length.out = 21)

llikCauchy(x, 0, 1)
#>           fx  dLocation     dScale
#> 1  -3.447315 -0.6000000  0.8000000
#> 2  -3.259780 -0.6513872  0.7587455
#> 3  -3.055753 -0.7100592  0.7041420
#> 4  -2.832979 -0.7763401  0.6303142
#> 5  -2.589293 -0.8490566  0.5283019
#> 6  -2.323385 -0.9230769  0.3846154
#> 7  -2.036728 -0.9836066  0.1803279
#> 8  -1.738057 -0.9944751 -0.1049724
#> 9  -1.452215 -0.8823529 -0.4705882
#> 10 -1.230908 -0.5504587 -0.8348624
#> 11 -1.144730  0.0000000 -1.0000000
#> 12 -1.230908  0.5504587 -0.8348624
#> 13 -1.452215  0.8823529 -0.4705882
#> 14 -1.738057  0.9944751 -0.1049724
#> 15 -2.036728  0.9836066  0.1803279
#> 16 -2.323385  0.9230769  0.3846154
#> 17 -2.589293  0.8490566  0.5283019
#> 18 -2.832979  0.7763401  0.6303142
#> 19 -3.055753  0.7100592  0.7041420
#> 20 -3.259780  0.6513872  0.7587455
#> 21 -3.447315  0.6000000  0.8000000

llikCauchy(x, 3, 1, full=TRUE)
#>       x location scale        fx  dLocation     dScale
#> 1  -3.0        3     1 -4.755648 -0.3243243  0.9459459
#> 2  -2.7        3     1 -4.655977 -0.3404001  0.9402807
#> 3  -2.4        3     1 -4.551246 -0.3580902  0.9336870
#> 4  -2.1        3     1 -4.440937 -0.3776379  0.9259534
#> 5  -1.8        3     1 -4.324449 -0.3993344  0.9168053
#> 6  -1.5        3     1 -4.201087 -0.4235294  0.9058824
#> 7  -1.2        3     1 -4.070040 -0.4506438  0.8927039
#> 8  -0.9        3     1 -3.930358 -0.4811845  0.8766194
#> 9  -0.6        3     1 -3.780926 -0.5157593  0.8567335
#> 10 -0.3        3     1 -3.620428 -0.5550883  0.8317914
#> 11  0.0        3     1 -3.447315 -0.6000000  0.8000000
#> 12  0.3        3     1 -3.259780 -0.6513872  0.7587455
#> 13  0.6        3     1 -3.055753 -0.7100592  0.7041420
#> 14  0.9        3     1 -2.832979 -0.7763401  0.6303142
#> 15  1.2        3     1 -2.589293 -0.8490566  0.5283019
#> 16  1.5        3     1 -2.323385 -0.9230769  0.3846154
#> 17  1.8        3     1 -2.036728 -0.9836066  0.1803279
#> 18  2.1        3     1 -1.738057 -0.9944751 -0.1049724
#> 19  2.4        3     1 -1.452215 -0.8823529 -0.4705882
#> 20  2.7        3     1 -1.230908 -0.5504587 -0.8348624
#> 21  3.0        3     1 -1.144730  0.0000000 -1.0000000

et <- et(-3, 3, length.out=10)
et$location <- 0
et$scale <- 1

model <- function() {
  model({
    fx <- llikCauchy(time, location, scale)
    dLocation <- llikCauchyDlocation(time, location, scale)
    dScale <- llikCauchyDscale(time, location, scale)
  })
}

rxSolve(model, et)
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#>  
#>  
#> C model:
#> // Define 3 LHS values
#> #define _LHS_0_ 0 // fx
#> #define _LHS_1_ 1 // dLocation
#> #define _LHS_2_ 2 // dScale
#> #define _getRxSolve_ _rx1e8db89411dbf3747ea7dcb64db521e22574__getRxSolve__hZjb
#> #define _evalUdf _rx1e8db89411dbf3747ea7dcb64db521e22575__evalUdf_tvsF
#> #define _solveData _rx1e8db89411dbf3747ea7dcb64db521e22576__solveData_QkpV
#> #define _assign_ptr _rx1e8db89411dbf3747ea7dcb64db521e22577__assign_ptr_HF4t
#> #define _rxRmModelLib _rx1e8db89411dbf3747ea7dcb64db521e22578__rxRmModelLib_E5bq
#> #define _rxGetModelLib _rx1e8db89411dbf3747ea7dcb64db521e22579__rxGetModelLib_rWj2
#> #define _old_c _rx1e8db89411dbf3747ea7dcb64db521e22580__old_c_jIRW
#> #define _ptrid _rx1e8db89411dbf3747ea7dcb64db521e22581__ptrid_0crr
#> #define _rxIsCurrentC _rx1e8db89411dbf3747ea7dcb64db521e22582__rxIsCurrentC_gcGm
#> #define _sumPS _rx1e8db89411dbf3747ea7dcb64db521e22583__sumPS_EsJu
#> #define _prodPS _rx1e8db89411dbf3747ea7dcb64db521e22584__prodPS_9xah
#> #define _prodType _rx1e8db89411dbf3747ea7dcb64db521e22585__prodType_gBOZ
#> #define _sumType _rx1e8db89411dbf3747ea7dcb64db521e22586__sumType_l3zY
#> #define _update_par_ptr _rx1e8db89411dbf3747ea7dcb64db521e22587__update_par_ptr_vTa9
#> #define _getParCov _rx1e8db89411dbf3747ea7dcb64db521e22588__getParCov_7EhW
#> #define _rxode2_rxAssignPtr _rx1e8db89411dbf3747ea7dcb64db521e22589__rxode2_rxAssignPtr_3MG9
#> #define _rxQr _rx1e8db89411dbf3747ea7dcb64db521e22590__rxQr_V3pR
#> #define _compareFactorVal _rx1e8db89411dbf3747ea7dcb64db521e22591__compareFactorVal_1aNk
#> #define _sum _rx1e8db89411dbf3747ea7dcb64db521e22592__sum_G5Jw
#> #define _udf _rx1e8db89411dbf3747ea7dcb64db521e22593__udf_D9TK
#> #define _sign _rx1e8db89411dbf3747ea7dcb64db521e22594__sign_40NP
#> #define _prod _rx1e8db89411dbf3747ea7dcb64db521e22595__prod_JiAL
#> #define _max _rx1e8db89411dbf3747ea7dcb64db521e22596__max_ZBrh
#> #define _min _rx1e8db89411dbf3747ea7dcb64db521e22597__min_efgh
#> #define _transit4P _rx1e8db89411dbf3747ea7dcb64db521e22598__transit4P_Drm8
#> #define _transit3P _rx1e8db89411dbf3747ea7dcb64db521e22599__transit3P_5igv
#> #define _assignFuns0 _rx1e8db89411dbf3747ea7dcb64db521e22600__assignFuns0_lJYg
#> #define _assignFuns _rx1e8db89411dbf3747ea7dcb64db521e22601__assignFuns_GqQx
#> #define _rxord _rx1e8db89411dbf3747ea7dcb64db521e22602__rxord_LoYl
#> #define __assignFuns2 _rx1e8db89411dbf3747ea7dcb64db521e22603___assignFuns2_FgJd
#> #define _llikCauchyDscale _rx1e8db89411dbf3747ea7dcb64db521e22604__llikCauchyDscale_KaFg
#> #define _llikCauchyDlocation _rx1e8db89411dbf3747ea7dcb64db521e22605__llikCauchyDlocation_mDR4
#> #define _llikCauchy _rx1e8db89411dbf3747ea7dcb64db521e22606__llikCauchy_jxuo
#> #define _llikGammaDrate _rx1e8db89411dbf3747ea7dcb64db521e22607__llikGammaDrate_P8gm
#> #define _llikGammaDshape _rx1e8db89411dbf3747ea7dcb64db521e22608__llikGammaDshape_oUom
#> #define _llikGamma _rx1e8db89411dbf3747ea7dcb64db521e22609__llikGamma_wLtu
#> #define _llikWeibullDscale _rx1e8db89411dbf3747ea7dcb64db521e22610__llikWeibullDscale_DgOz
#> #define _llikWeibullDshape _rx1e8db89411dbf3747ea7dcb64db521e22611__llikWeibullDshape_tPZw
#> #define _llikWeibull _rx1e8db89411dbf3747ea7dcb64db521e22612__llikWeibull_Dqx2
#> #define _llikUnifDbeta _rx1e8db89411dbf3747ea7dcb64db521e22613__llikUnifDbeta_t0Ff
#> #define _llikUnifDalpha _rx1e8db89411dbf3747ea7dcb64db521e22614__llikUnifDalpha_CRup
#> #define _llikUnif _rx1e8db89411dbf3747ea7dcb64db521e22615__llikUnif_BfcL
#> #define _llikGeomDp _rx1e8db89411dbf3747ea7dcb64db521e22616__llikGeomDp_LlJk
#> #define _llikGeom _rx1e8db89411dbf3747ea7dcb64db521e22617__llikGeom_0HOu
#> #define _llikFDdf2 _rx1e8db89411dbf3747ea7dcb64db521e22618__llikFDdf2_quMQ
#> #define _llikFDdf1 _rx1e8db89411dbf3747ea7dcb64db521e22619__llikFDdf1_IP4s
#> #define _llikF _rx1e8db89411dbf3747ea7dcb64db521e22620__llikF_BKwl
#> #define _llikExpDrate _rx1e8db89411dbf3747ea7dcb64db521e22621__llikExpDrate_ION1
#> #define _llikExp _rx1e8db89411dbf3747ea7dcb64db521e22622__llikExp_tTX5
#> #define _llikChisqDdf _rx1e8db89411dbf3747ea7dcb64db521e22623__llikChisqDdf_DulH
#> #define _llikChisq _rx1e8db89411dbf3747ea7dcb64db521e22624__llikChisq_LxeA
#> #define _llikTDsd _rx1e8db89411dbf3747ea7dcb64db521e22625__llikTDsd_Ut0p
#> #define _llikTDmean _rx1e8db89411dbf3747ea7dcb64db521e22626__llikTDmean_o6Aq
#> #define _llikTDdf _rx1e8db89411dbf3747ea7dcb64db521e22627__llikTDdf_SQkn
#> #define _llikT _rx1e8db89411dbf3747ea7dcb64db521e22628__llikT_H8fP
#> #define _llikBetaDshape2 _rx1e8db89411dbf3747ea7dcb64db521e22629__llikBetaDshape2_CAnN
#> #define _llikBetaDshape1 _rx1e8db89411dbf3747ea7dcb64db521e22630__llikBetaDshape1_kUms
#> #define _llikBeta _rx1e8db89411dbf3747ea7dcb64db521e22631__llikBeta_wiOW
#> #define _llikNbinomMuDmu _rx1e8db89411dbf3747ea7dcb64db521e22632__llikNbinomMuDmu_G16H
#> #define _llikNbinomMu _rx1e8db89411dbf3747ea7dcb64db521e22633__llikNbinomMu_PUgn
#> #define _llikNbinomDprob _rx1e8db89411dbf3747ea7dcb64db521e22634__llikNbinomDprob_Cjuw
#> #define _llikNbinom _rx1e8db89411dbf3747ea7dcb64db521e22635__llikNbinom_A8iw
#> #define _llikBinomDprob _rx1e8db89411dbf3747ea7dcb64db521e22636__llikBinomDprob_bn2G
#> #define _llikBinom _rx1e8db89411dbf3747ea7dcb64db521e22637__llikBinom_TQjS
#> #define _llikPoisDlambda _rx1e8db89411dbf3747ea7dcb64db521e22638__llikPoisDlambda_rer5
#> #define _llikPois _rx1e8db89411dbf3747ea7dcb64db521e22639__llikPois_zexL
#> #define _llikNormDsd _rx1e8db89411dbf3747ea7dcb64db521e22640__llikNormDsd_gAwM
#> #define _llikNormDmean _rx1e8db89411dbf3747ea7dcb64db521e22641__llikNormDmean_E6nq
#> #define _llikNorm _rx1e8db89411dbf3747ea7dcb64db521e22642__llikNorm_R3DD
#> #define simeps _rx1e8db89411dbf3747ea7dcb64db521e22643_simeps_9jN5
#> #define simeta _rx1e8db89411dbf3747ea7dcb64db521e22644_simeta_O4r9
#> #define expit _rx1e8db89411dbf3747ea7dcb64db521e22645_expit_D97m
#> #define logit _rx1e8db89411dbf3747ea7dcb64db521e22646_logit_e5yb
#> #define gammapDer _rx1e8db89411dbf3747ea7dcb64db521e22647_gammapDer_V8Eu
#> #define lowergamma _rx1e8db89411dbf3747ea7dcb64db521e22648_lowergamma_Z1xT
#> #define uppergamma _rx1e8db89411dbf3747ea7dcb64db521e22649_uppergamma_QYOQ
#> #define gammaqInva _rx1e8db89411dbf3747ea7dcb64db521e22650_gammaqInva_e6aq
#> #define gammaqInv _rx1e8db89411dbf3747ea7dcb64db521e22651_gammaqInv_t3Lb
#> #define gammapInva _rx1e8db89411dbf3747ea7dcb64db521e22652_gammapInva_sbea
#> #define gammapInv _rx1e8db89411dbf3747ea7dcb64db521e22653_gammapInv_w5m4
#> #define gammaq _rx1e8db89411dbf3747ea7dcb64db521e22654_gammaq_brD9
#> #define gammap _rx1e8db89411dbf3747ea7dcb64db521e22655_gammap_EZAJ
#> #define phi _rx1e8db89411dbf3747ea7dcb64db521e22656_phi_Ne7I
#> #define d2ELUa _rx1e8db89411dbf3747ea7dcb64db521e22657_d2ELUa_xUFA
#> #define dELUa _rx1e8db89411dbf3747ea7dcb64db521e22658_dELUa_5R2w
#> #define d2aELU _rx1e8db89411dbf3747ea7dcb64db521e22659_d2aELU_ZwwB
#> #define d2ELU _rx1e8db89411dbf3747ea7dcb64db521e22660_d2ELU_gRvb
#> #define dELU _rx1e8db89411dbf3747ea7dcb64db521e22661_dELU_lnk3
#> #define ELU _rx1e8db89411dbf3747ea7dcb64db521e22662_ELU_57cf
#> #define dPReLUa1 _rx1e8db89411dbf3747ea7dcb64db521e22663_dPReLUa1_UwMa
#> #define dPReLUa _rx1e8db89411dbf3747ea7dcb64db521e22664_dPReLUa_rEA9
#> #define dPReLU _rx1e8db89411dbf3747ea7dcb64db521e22665_dPReLU_z1SE
#> #define PReLU _rx1e8db89411dbf3747ea7dcb64db521e22666_PReLU_1KzA
#> #define dSwish _rx1e8db89411dbf3747ea7dcb64db521e22667_dSwish_f2H4
#> #define Swish _rx1e8db89411dbf3747ea7dcb64db521e22668_Swish_Q2oF
#> #define dlReLU _rx1e8db89411dbf3747ea7dcb64db521e22669_dlReLU_0joC
#> #define lReLU _rx1e8db89411dbf3747ea7dcb64db521e22670_lReLU_Gj0a
#> #define dSELU _rx1e8db89411dbf3747ea7dcb64db521e22671_dSELU_8wAb
#> #define SELU _rx1e8db89411dbf3747ea7dcb64db521e22672_SELU_QELO
#> #define d4softplus _rx1e8db89411dbf3747ea7dcb64db521e22673_d4softplus_lObq
#> #define d3softplus _rx1e8db89411dbf3747ea7dcb64db521e22674_d3softplus_Nath
#> #define d2softplus _rx1e8db89411dbf3747ea7dcb64db521e22675_d2softplus_ltiW
#> #define dsoftplus _rx1e8db89411dbf3747ea7dcb64db521e22676_dsoftplus_R2IG
#> #define softplus _rx1e8db89411dbf3747ea7dcb64db521e22677_softplus_VLQY
#> #define d4GELU _rx1e8db89411dbf3747ea7dcb64db521e22678_d4GELU_M84J
#> #define d3GELU _rx1e8db89411dbf3747ea7dcb64db521e22679_d3GELU_Cnuy
#> #define d2GELU _rx1e8db89411dbf3747ea7dcb64db521e22680_d2GELU_CXPJ
#> #define dGELU _rx1e8db89411dbf3747ea7dcb64db521e22681_dGELU_UjvF
#> #define GELU _rx1e8db89411dbf3747ea7dcb64db521e22682_GELU_k7KJ
#> #define dReLU _rx1e8db89411dbf3747ea7dcb64db521e22683_dReLU_vDxq
#> #define ReLU _rx1e8db89411dbf3747ea7dcb64db521e22684_ReLU_bVSK
#> #define riweibull _rx1e8db89411dbf3747ea7dcb64db521e22685_riweibull_Vfr7
#> #define riunif _rx1e8db89411dbf3747ea7dcb64db521e22686_riunif_TcaU
#> #define rit_ _rx1e8db89411dbf3747ea7dcb64db521e22687_rit__EgpY
#> #define ripois _rx1e8db89411dbf3747ea7dcb64db521e22688_ripois_kNIq
#> #define ribeta _rx1e8db89411dbf3747ea7dcb64db521e22689_ribeta_nEnm
#> #define rigamma _rx1e8db89411dbf3747ea7dcb64db521e22690_rigamma_hSbo
#> #define rigeom _rx1e8db89411dbf3747ea7dcb64db521e22691_rigeom_MZYB
#> #define rif _rx1e8db89411dbf3747ea7dcb64db521e22692_rif_jkiC
#> #define riexp _rx1e8db89411dbf3747ea7dcb64db521e22693_riexp_oUqv
#> #define richisq _rx1e8db89411dbf3747ea7dcb64db521e22694_richisq_niI7
#> #define ricauchy _rx1e8db89411dbf3747ea7dcb64db521e22695_ricauchy_5Ui0
#> #define rinbinomMu _rx1e8db89411dbf3747ea7dcb64db521e22696_rinbinomMu_GVTs
#> #define rinbinom _rx1e8db89411dbf3747ea7dcb64db521e22697_rinbinom_34TN
#> #define ribinom _rx1e8db89411dbf3747ea7dcb64db521e22698_ribinom_HeFp
#> #define rinorm _rx1e8db89411dbf3747ea7dcb64db521e22699_rinorm_ue1Z
#> #define rxweibull _rx1e8db89411dbf3747ea7dcb64db521e22700_rxweibull_aEnJ
#> #define rxunif _rx1e8db89411dbf3747ea7dcb64db521e22701_rxunif_R6UH
#> #define rxt_ _rx1e8db89411dbf3747ea7dcb64db521e22702_rxt__ZKbS
#> #define rxpois _rx1e8db89411dbf3747ea7dcb64db521e22703_rxpois_smNY
#> #define rxbeta _rx1e8db89411dbf3747ea7dcb64db521e22704_rxbeta_lS2w
#> #define rxgamma _rx1e8db89411dbf3747ea7dcb64db521e22705_rxgamma_zYqw
#> #define rxgeom _rx1e8db89411dbf3747ea7dcb64db521e22706_rxgeom_2z8b
#> #define rxf _rx1e8db89411dbf3747ea7dcb64db521e22707_rxf_GMdj
#> #define rxexp _rx1e8db89411dbf3747ea7dcb64db521e22708_rxexp_LB3L
#> #define rxchisq _rx1e8db89411dbf3747ea7dcb64db521e22709_rxchisq_sOUR
#> #define rxcauchy _rx1e8db89411dbf3747ea7dcb64db521e22710_rxcauchy_bPYU
#> #define rxnbinomMu _rx1e8db89411dbf3747ea7dcb64db521e22711_rxnbinomMu_PmW8
#> #define rxnbinom _rx1e8db89411dbf3747ea7dcb64db521e22712_rxnbinom_sOmG
#> #define rxbinom _rx1e8db89411dbf3747ea7dcb64db521e22713_rxbinom_ypO5
#> #define rxnorm _rx1e8db89411dbf3747ea7dcb64db521e22714_rxnorm_5OnO
#> #define linCmtB _rx1e8db89411dbf3747ea7dcb64db521e22715_linCmtB_E9TF
#> #define linCmtA _rx1e8db89411dbf3747ea7dcb64db521e22716_linCmtA_yOK7
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){  if (_solveData->svar[_svari] == 0) {location = _PP[0];};   if (_solveData->svar[_svari] == 1) {scale = _PP[1];}; }
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){  if (_solveData->ovar[_ovari] == 0) {location = _PP[0];};   if (_solveData->ovar[_ovari] == 1) {scale = _PP[1];}; }
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  rx_8711d080fd11f3466e2b0dcee8d5ab4b__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *rx_8711d080fd11f3466e2b0dcee8d5ab4b__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP rx_8711d080fd11f3466e2b0dcee8d5ab4b__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void rx_8711d080fd11f3466e2b0dcee8d5ab4b__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double location = NA_REAL;
#>   double scale = NA_REAL;
#>   double fx = NA_REAL;
#>   double dLocation = NA_REAL;
#>   double dScale = NA_REAL;
#> 
#>   (void)t;
#>   (void)location;
#>   (void)scale;
#>   (void)fx;
#>   (void)dLocation;
#>   (void)dScale;
#> 
#>   fx = _PL[0];
#>   dLocation = _PL[1];
#>   dScale = _PL[2];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#>   location = _PP[0];
#>   scale = _PP[1];
#> 
#> 
#>   fx =llikCauchy(t,location,scale);
#>   dLocation =llikCauchyDlocation(t,location,scale);
#>   dScale =llikCauchyDscale(t,location,scale);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void rx_8711d080fd11f3466e2b0dcee8d5ab4b__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void rx_8711d080fd11f3466e2b0dcee8d5ab4b__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void rx_8711d080fd11f3466e2b0dcee8d5ab4b__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double location = NA_REAL;
#>   double scale = NA_REAL;
#>   double fx = NA_REAL;
#>   double dLocation = NA_REAL;
#>   double dScale = NA_REAL;
#> 
#>   (void)t;
#>   (void)location;
#>   (void)scale;
#>   (void)fx;
#>   (void)dLocation;
#>   (void)dScale;
#> 
#>   fx = _PL[0];
#>   dLocation = _PL[1];
#>   dScale = _PL[2];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#>   location = _PP[0];
#>   scale = _PP[1];
#> 
#> 
#>   fx =llikCauchy(t,location,scale);
#>   dLocation =llikCauchyDlocation(t,location,scale);
#>   dScale =llikCauchyDscale(t,location,scale);
#> 
#>   _lhs[_LHS_0_]=fx;
#>   _lhs[_LHS_1_]=dLocation;
#>   _lhs[_LHS_2_]=dScale;
#> }
#> // Functional based bioavailability
#> double rx_8711d080fd11f3466e2b0dcee8d5ab4b__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double rx_8711d080fd11f3466e2b0dcee8d5ab4b__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double rx_8711d080fd11f3466e2b0dcee8d5ab4b__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double rx_8711d080fd11f3466e2b0dcee8d5ab4b__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void rx_8711d080fd11f3466e2b0dcee8d5ab4b__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void rx_8711d080fd11f3466e2b0dcee8d5ab4b__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void rx_8711d080fd11f3466e2b0dcee8d5ab4b__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP rx_8711d080fd11f3466e2b0dcee8d5ab4b__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_8711d080fd11f3466e2b0dcee8d5ab4b__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAA[93AAAv7#aT)%%r/*NAy>5,h\"\?*7\"Aj:&v51/%%U,E_l~&VJxN4!E#KY]Z1c:+H1vOJwtW>aWxOvE,8wbgKXg9cb=~ghgKOj}4pf3B%%FayXLEAdyng</]m6IP4~jy}i*`{Bm]x|6L6zio(<(g[8>7k)iU>HSf&Wg\"+98l3#ge@n>[v+|wa5q5=D^>>o.&MZ1:Js+uv9,doqs_K@k0#)us`)[.ke>kaYapp1MtI&}npMv]^FrPRQOIh}\?}oPG7_,E47VP$J%%bHM:>o4r1GhlcCI^PFauIq~O!;gSNC^Ea%%v<)BnB>\"9&b6.pDAv+k&nhoqj]A^bD[:>$zMM%%Yx)b3jXCNf7.mSeq)q&5HAG:ODzrmWu5SfSZ[9tjRUE|L9<2^,u0/,z`QAGOp+7\"ix;\"TSG$A%%RFMa7}QQln]\?\?Wr~(zB(Mq@W6U|>%%E:]]wPCA|z)UKbeO$cy0.K_5wSIgfY<U.=L/zUb~VE%%`_|z]OkSy$P;\?)uv#([]GfxXhClw.^OJ_\?@e8HSUa<w~j]]f]E=eC_a1cZk|aqt7gm9`8r~Vb{Ps0kdnM;1Z;[IRXxr>MsH%%ZwM@ohz%%Z2UjDUpPrQq*#l|RXvg95G\"8Jn%%7AKveT$@7&Q\"R[[hJ#^twx*_4sy0S5P@v;8VpolqI>USao(y*!)MEdXnB|LRW8`le\?@9]MDQg(jpX0:\"CL;<NZANN6GHt^Cb*o[iz1Ip2eG\?+r>3pNg0L4#/,[BTJ.ufHXh6o1V1H5V(9j@}@v0/{2_Q7vK<dJoeGJ6RCzE0qO,zZB844!:1lG44kyMS**/Fk\"P7EYvZeMFkAjZe1F_Yw<l_I~M~Ho)~Qo\"W,lT/BD@K:J|_{D\?KYsV;n2:]__q+:reT$<;.[%%/YQn4}6GW1T+G!wgPeCzoCGkRhPkmZ%%/UuJsDlU)+Lp=:\?{7@w~R>y#&[P&%%s0Z4MLOBx]j=$rEX>6FL7cYMoGx~lq>jMd\?IfZr;4psNjN~ctI).nK@&sF}{eZ0+*q)rvZn&q%%JWsCN3MNp#Q\?>.0Db8z1Ir75PJBivD<\"V6x1%%1_xUxq#XLh[S[NnrD)*#xiOE)m=\"/(Iyh,7qz=+mZAeB.J96:f[41v5O27F");
#>     char buf[1146];
#> #define __doBufN__ 1146
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void rx_8711d080fd11f3466e2b0dcee8d5ab4b__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   rx_8711d080fd11f3466e2b0dcee8d5ab4b__dydt(neq, *t, A, DADT);
#> }
#> extern int rx_8711d080fd11f3466e2b0dcee8d5ab4b__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   rx_8711d080fd11f3466e2b0dcee8d5ab4b__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void rx_8711d080fd11f3466e2b0dcee8d5ab4b__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   rx_8711d080fd11f3466e2b0dcee8d5ab4b__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void rx_8711d080fd11f3466e2b0dcee8d5ab4b__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_rx_8711d080fd11f3466e2b0dcee8d5ab4b_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("rx_8711d080fd11f3466e2b0dcee8d5ab4b_","rx_8711d080fd11f3466e2b0dcee8d5ab4b__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("rx_8711d080fd11f3466e2b0dcee8d5ab4b_","rx_8711d080fd11f3466e2b0dcee8d5ab4b__assignFuns", (DL_FUNC) rx_8711d080fd11f3466e2b0dcee8d5ab4b__assignFuns);
#>   R_RegisterCCallable("rx_8711d080fd11f3466e2b0dcee8d5ab4b_","rx_8711d080fd11f3466e2b0dcee8d5ab4b__inis",(DL_FUNC) rx_8711d080fd11f3466e2b0dcee8d5ab4b__inis);
#>   R_RegisterCCallable("rx_8711d080fd11f3466e2b0dcee8d5ab4b_","rx_8711d080fd11f3466e2b0dcee8d5ab4b__dydt",(DL_FUNC) rx_8711d080fd11f3466e2b0dcee8d5ab4b__dydt);
#>   R_RegisterCCallable("rx_8711d080fd11f3466e2b0dcee8d5ab4b_","rx_8711d080fd11f3466e2b0dcee8d5ab4b__calc_lhs",(DL_FUNC) rx_8711d080fd11f3466e2b0dcee8d5ab4b__calc_lhs);
#>   R_RegisterCCallable("rx_8711d080fd11f3466e2b0dcee8d5ab4b_","rx_8711d080fd11f3466e2b0dcee8d5ab4b__calc_jac",(DL_FUNC) rx_8711d080fd11f3466e2b0dcee8d5ab4b__calc_jac);
#>   R_RegisterCCallable("rx_8711d080fd11f3466e2b0dcee8d5ab4b_","rx_8711d080fd11f3466e2b0dcee8d5ab4b__dydt_lsoda", (DL_FUNC) rx_8711d080fd11f3466e2b0dcee8d5ab4b__dydt_lsoda);
#>   R_RegisterCCallable("rx_8711d080fd11f3466e2b0dcee8d5ab4b_","rx_8711d080fd11f3466e2b0dcee8d5ab4b__calc_jac_lsoda", (DL_FUNC) rx_8711d080fd11f3466e2b0dcee8d5ab4b__calc_jac_lsoda);
#>   R_RegisterCCallable("rx_8711d080fd11f3466e2b0dcee8d5ab4b_","rx_8711d080fd11f3466e2b0dcee8d5ab4b__ode_solver_solvedata", (DL_FUNC) rx_8711d080fd11f3466e2b0dcee8d5ab4b__ode_solver_solvedata);
#>   R_RegisterCCallable("rx_8711d080fd11f3466e2b0dcee8d5ab4b_","rx_8711d080fd11f3466e2b0dcee8d5ab4b__ode_solver_get_solvedata", (DL_FUNC) rx_8711d080fd11f3466e2b0dcee8d5ab4b__ode_solver_get_solvedata);
#>   R_RegisterCCallable("rx_8711d080fd11f3466e2b0dcee8d5ab4b_","rx_8711d080fd11f3466e2b0dcee8d5ab4b__F", (DL_FUNC) rx_8711d080fd11f3466e2b0dcee8d5ab4b__F);
#>   R_RegisterCCallable("rx_8711d080fd11f3466e2b0dcee8d5ab4b_","rx_8711d080fd11f3466e2b0dcee8d5ab4b__Lag", (DL_FUNC) rx_8711d080fd11f3466e2b0dcee8d5ab4b__Lag);
#>   R_RegisterCCallable("rx_8711d080fd11f3466e2b0dcee8d5ab4b_","rx_8711d080fd11f3466e2b0dcee8d5ab4b__Rate", (DL_FUNC) rx_8711d080fd11f3466e2b0dcee8d5ab4b__Rate);
#>   R_RegisterCCallable("rx_8711d080fd11f3466e2b0dcee8d5ab4b_","rx_8711d080fd11f3466e2b0dcee8d5ab4b__Dur", (DL_FUNC) rx_8711d080fd11f3466e2b0dcee8d5ab4b__Dur);
#>   R_RegisterCCallable("rx_8711d080fd11f3466e2b0dcee8d5ab4b_","rx_8711d080fd11f3466e2b0dcee8d5ab4b__mtime", (DL_FUNC) rx_8711d080fd11f3466e2b0dcee8d5ab4b__mtime);
#>   R_RegisterCCallable("rx_8711d080fd11f3466e2b0dcee8d5ab4b_","rx_8711d080fd11f3466e2b0dcee8d5ab4b__ME", (DL_FUNC) rx_8711d080fd11f3466e2b0dcee8d5ab4b__ME);
#>   R_RegisterCCallable("rx_8711d080fd11f3466e2b0dcee8d5ab4b_","rx_8711d080fd11f3466e2b0dcee8d5ab4b__IndF", (DL_FUNC) rx_8711d080fd11f3466e2b0dcee8d5ab4b__IndF);
#>   R_RegisterCCallable("rx_8711d080fd11f3466e2b0dcee8d5ab4b_","rx_8711d080fd11f3466e2b0dcee8d5ab4b__dydt_liblsoda", (DL_FUNC) rx_8711d080fd11f3466e2b0dcee8d5ab4b__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_rx_8711d080fd11f3466e2b0dcee8d5ab4b_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_rx_8711d080fd11f3466e2b0dcee8d5ab4b_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"rx_8711d080fd11f3466e2b0dcee8d5ab4b__model_vars", (DL_FUNC) &rx_8711d080fd11f3466e2b0dcee8d5ab4b__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_rx_8711d080fd11f3466e2b0dcee8d5ab4b_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_8711d080fd11f3466e2b0dcee8d5ab4b__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("rx_8711d080fd11f3466e2b0dcee8d5ab4b__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’
#> ── Solved rxode2 object ──
#> ── Parameters (value$params): ──
#> # A tibble: 1 × 0
#> ── Initial Conditions (value$inits): ──
#> named numeric(0)
#> ── First part of data (object): ──
#> # A tibble: 10 × 6
#>     time    fx dLocation dScale location scale
#>    <dbl> <dbl>     <dbl>  <dbl>    <dbl> <dbl>
#> 1 -3     -3.45    -0.6    0.8          0     1
#> 2 -2.33  -3.01    -0.724  0.690        0     1
#> 3 -1.67  -2.47    -0.882  0.471        0     1
#> 4 -1     -1.84    -1      0            0     1
#> 5 -0.333 -1.25    -0.6   -0.8          0     1
#> 6  0.333 -1.25     0.600 -0.8          0     1
#> # ℹ 4 more rows
# }