Calculate the log likelihood of the negative binomial function (and its derivatives)

Calculate the log likelihood of the negative binomial function (and its derivatives)

Usage

llikNbinom(x, size, prob, full = FALSE)

Arguments

x

Number of successes

size

Size of trial

prob

probability of success

full

Add the data frame showing x, mean, sd as well as the fx and derivatives

Value

data frame with fx for the pdf value of with dProb that has the derivatives with respect to the parameters at the observation time-point

Details

In an rxode2() model, you can use llikNbinom() but you have to use all arguments. You can also get the derivative of prob with llikNbinomDprob()

Author

Matthew L. Fidler

Examples

# \donttest{
llikNbinom(46:54, 100, 0.5)
#>          fx dProb
#> 1 -13.25200   108
#> 2 -12.81168   106
#> 3 -12.38560   104
#> 4 -11.97335   102
#> 5 -11.57458   100
#> 6 -11.18892    98
#> 7 -10.81603    96
#> 8 -10.45559    94
#> 9 -10.10728    92

llikNbinom(46:54, 100, 0.5, TRUE)
#>    x size prob        fx dProb
#> 1 46  100  0.5 -13.25200   108
#> 2 47  100  0.5 -12.81168   106
#> 3 48  100  0.5 -12.38560   104
#> 4 49  100  0.5 -11.97335   102
#> 5 50  100  0.5 -11.57458   100
#> 6 51  100  0.5 -11.18892    98
#> 7 52  100  0.5 -10.81603    96
#> 8 53  100  0.5 -10.45559    94
#> 9 54  100  0.5 -10.10728    92

# In rxode2 you can use:

et <- et(46:54)
et$size <- 100
et$prob <-0.5

model <- function() {
  model({
    fx <- llikNbinom(time, size, prob)
    dProb <- llikNbinomDprob(time, size, prob)
  })
}

rxSolve(model, et)
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#>  
#>  
#> C model:
#> // Define 2 LHS values
#> #define _LHS_0_ 0 // fx
#> #define _LHS_1_ 1 // dProb
#> #define _getRxSolve_ _rx2dd6512f8f40bf129950e57dd80c7fc33432__getRxSolve__jIRW
#> #define _evalUdf _rx2dd6512f8f40bf129950e57dd80c7fc33433__evalUdf_0crr
#> #define _solveData _rx2dd6512f8f40bf129950e57dd80c7fc33434__solveData_gcGm
#> #define _assign_ptr _rx2dd6512f8f40bf129950e57dd80c7fc33435__assign_ptr_EsJu
#> #define _rxRmModelLib _rx2dd6512f8f40bf129950e57dd80c7fc33436__rxRmModelLib_9xah
#> #define _rxGetModelLib _rx2dd6512f8f40bf129950e57dd80c7fc33437__rxGetModelLib_gBOZ
#> #define _old_c _rx2dd6512f8f40bf129950e57dd80c7fc33438__old_c_l3zY
#> #define _ptrid _rx2dd6512f8f40bf129950e57dd80c7fc33439__ptrid_vTa9
#> #define _rxIsCurrentC _rx2dd6512f8f40bf129950e57dd80c7fc33440__rxIsCurrentC_7EhW
#> #define _sumPS _rx2dd6512f8f40bf129950e57dd80c7fc33441__sumPS_3MG9
#> #define _prodPS _rx2dd6512f8f40bf129950e57dd80c7fc33442__prodPS_V3pR
#> #define _prodType _rx2dd6512f8f40bf129950e57dd80c7fc33443__prodType_1aNk
#> #define _sumType _rx2dd6512f8f40bf129950e57dd80c7fc33444__sumType_G5Jw
#> #define _update_par_ptr _rx2dd6512f8f40bf129950e57dd80c7fc33445__update_par_ptr_D9TK
#> #define _getParCov _rx2dd6512f8f40bf129950e57dd80c7fc33446__getParCov_40NP
#> #define _rxode2_rxAssignPtr _rx2dd6512f8f40bf129950e57dd80c7fc33447__rxode2_rxAssignPtr_JiAL
#> #define _rxQr _rx2dd6512f8f40bf129950e57dd80c7fc33448__rxQr_ZBrh
#> #define _compareFactorVal _rx2dd6512f8f40bf129950e57dd80c7fc33449__compareFactorVal_efgh
#> #define _sum _rx2dd6512f8f40bf129950e57dd80c7fc33450__sum_Drm8
#> #define _udf _rx2dd6512f8f40bf129950e57dd80c7fc33451__udf_5igv
#> #define _sign _rx2dd6512f8f40bf129950e57dd80c7fc33452__sign_lJYg
#> #define _prod _rx2dd6512f8f40bf129950e57dd80c7fc33453__prod_GqQx
#> #define _max _rx2dd6512f8f40bf129950e57dd80c7fc33454__max_LoYl
#> #define _min _rx2dd6512f8f40bf129950e57dd80c7fc33455__min_FgJd
#> #define _transit4P _rx2dd6512f8f40bf129950e57dd80c7fc33456__transit4P_KaFg
#> #define _transit3P _rx2dd6512f8f40bf129950e57dd80c7fc33457__transit3P_mDR4
#> #define _assignFuns0 _rx2dd6512f8f40bf129950e57dd80c7fc33458__assignFuns0_jxuo
#> #define _assignFuns _rx2dd6512f8f40bf129950e57dd80c7fc33459__assignFuns_P8gm
#> #define _rxord _rx2dd6512f8f40bf129950e57dd80c7fc33460__rxord_oUom
#> #define __assignFuns2 _rx2dd6512f8f40bf129950e57dd80c7fc33461___assignFuns2_wLtu
#> #define _llikCauchyDscale _rx2dd6512f8f40bf129950e57dd80c7fc33462__llikCauchyDscale_DgOz
#> #define _llikCauchyDlocation _rx2dd6512f8f40bf129950e57dd80c7fc33463__llikCauchyDlocation_tPZw
#> #define _llikCauchy _rx2dd6512f8f40bf129950e57dd80c7fc33464__llikCauchy_Dqx2
#> #define _llikGammaDrate _rx2dd6512f8f40bf129950e57dd80c7fc33465__llikGammaDrate_t0Ff
#> #define _llikGammaDshape _rx2dd6512f8f40bf129950e57dd80c7fc33466__llikGammaDshape_CRup
#> #define _llikGamma _rx2dd6512f8f40bf129950e57dd80c7fc33467__llikGamma_BfcL
#> #define _llikWeibullDscale _rx2dd6512f8f40bf129950e57dd80c7fc33468__llikWeibullDscale_LlJk
#> #define _llikWeibullDshape _rx2dd6512f8f40bf129950e57dd80c7fc33469__llikWeibullDshape_0HOu
#> #define _llikWeibull _rx2dd6512f8f40bf129950e57dd80c7fc33470__llikWeibull_quMQ
#> #define _llikUnifDbeta _rx2dd6512f8f40bf129950e57dd80c7fc33471__llikUnifDbeta_IP4s
#> #define _llikUnifDalpha _rx2dd6512f8f40bf129950e57dd80c7fc33472__llikUnifDalpha_BKwl
#> #define _llikUnif _rx2dd6512f8f40bf129950e57dd80c7fc33473__llikUnif_ION1
#> #define _llikGeomDp _rx2dd6512f8f40bf129950e57dd80c7fc33474__llikGeomDp_tTX5
#> #define _llikGeom _rx2dd6512f8f40bf129950e57dd80c7fc33475__llikGeom_DulH
#> #define _llikFDdf2 _rx2dd6512f8f40bf129950e57dd80c7fc33476__llikFDdf2_LxeA
#> #define _llikFDdf1 _rx2dd6512f8f40bf129950e57dd80c7fc33477__llikFDdf1_Ut0p
#> #define _llikF _rx2dd6512f8f40bf129950e57dd80c7fc33478__llikF_o6Aq
#> #define _llikExpDrate _rx2dd6512f8f40bf129950e57dd80c7fc33479__llikExpDrate_SQkn
#> #define _llikExp _rx2dd6512f8f40bf129950e57dd80c7fc33480__llikExp_H8fP
#> #define _llikChisqDdf _rx2dd6512f8f40bf129950e57dd80c7fc33481__llikChisqDdf_CAnN
#> #define _llikChisq _rx2dd6512f8f40bf129950e57dd80c7fc33482__llikChisq_kUms
#> #define _llikTDsd _rx2dd6512f8f40bf129950e57dd80c7fc33483__llikTDsd_wiOW
#> #define _llikTDmean _rx2dd6512f8f40bf129950e57dd80c7fc33484__llikTDmean_G16H
#> #define _llikTDdf _rx2dd6512f8f40bf129950e57dd80c7fc33485__llikTDdf_PUgn
#> #define _llikT _rx2dd6512f8f40bf129950e57dd80c7fc33486__llikT_Cjuw
#> #define _llikBetaDshape2 _rx2dd6512f8f40bf129950e57dd80c7fc33487__llikBetaDshape2_A8iw
#> #define _llikBetaDshape1 _rx2dd6512f8f40bf129950e57dd80c7fc33488__llikBetaDshape1_bn2G
#> #define _llikBeta _rx2dd6512f8f40bf129950e57dd80c7fc33489__llikBeta_TQjS
#> #define _llikNbinomMuDmu _rx2dd6512f8f40bf129950e57dd80c7fc33490__llikNbinomMuDmu_rer5
#> #define _llikNbinomMu _rx2dd6512f8f40bf129950e57dd80c7fc33491__llikNbinomMu_zexL
#> #define _llikNbinomDprob _rx2dd6512f8f40bf129950e57dd80c7fc33492__llikNbinomDprob_gAwM
#> #define _llikNbinom _rx2dd6512f8f40bf129950e57dd80c7fc33493__llikNbinom_E6nq
#> #define _llikBinomDprob _rx2dd6512f8f40bf129950e57dd80c7fc33494__llikBinomDprob_R3DD
#> #define _llikBinom _rx2dd6512f8f40bf129950e57dd80c7fc33495__llikBinom_9jN5
#> #define _llikPoisDlambda _rx2dd6512f8f40bf129950e57dd80c7fc33496__llikPoisDlambda_O4r9
#> #define _llikPois _rx2dd6512f8f40bf129950e57dd80c7fc33497__llikPois_D97m
#> #define _llikNormDsd _rx2dd6512f8f40bf129950e57dd80c7fc33498__llikNormDsd_e5yb
#> #define _llikNormDmean _rx2dd6512f8f40bf129950e57dd80c7fc33499__llikNormDmean_V8Eu
#> #define _llikNorm _rx2dd6512f8f40bf129950e57dd80c7fc33500__llikNorm_Z1xT
#> #define simeps _rx2dd6512f8f40bf129950e57dd80c7fc33501_simeps_QYOQ
#> #define simeta _rx2dd6512f8f40bf129950e57dd80c7fc33502_simeta_e6aq
#> #define expit _rx2dd6512f8f40bf129950e57dd80c7fc33503_expit_t3Lb
#> #define logit _rx2dd6512f8f40bf129950e57dd80c7fc33504_logit_sbea
#> #define gammapDer _rx2dd6512f8f40bf129950e57dd80c7fc33505_gammapDer_w5m4
#> #define lowergamma _rx2dd6512f8f40bf129950e57dd80c7fc33506_lowergamma_brD9
#> #define uppergamma _rx2dd6512f8f40bf129950e57dd80c7fc33507_uppergamma_EZAJ
#> #define gammaqInva _rx2dd6512f8f40bf129950e57dd80c7fc33508_gammaqInva_Ne7I
#> #define gammaqInv _rx2dd6512f8f40bf129950e57dd80c7fc33509_gammaqInv_xUFA
#> #define gammapInva _rx2dd6512f8f40bf129950e57dd80c7fc33510_gammapInva_5R2w
#> #define gammapInv _rx2dd6512f8f40bf129950e57dd80c7fc33511_gammapInv_ZwwB
#> #define gammaq _rx2dd6512f8f40bf129950e57dd80c7fc33512_gammaq_gRvb
#> #define gammap _rx2dd6512f8f40bf129950e57dd80c7fc33513_gammap_lnk3
#> #define phi _rx2dd6512f8f40bf129950e57dd80c7fc33514_phi_57cf
#> #define d2ELUa _rx2dd6512f8f40bf129950e57dd80c7fc33515_d2ELUa_UwMa
#> #define dELUa _rx2dd6512f8f40bf129950e57dd80c7fc33516_dELUa_rEA9
#> #define d2aELU _rx2dd6512f8f40bf129950e57dd80c7fc33517_d2aELU_z1SE
#> #define d2ELU _rx2dd6512f8f40bf129950e57dd80c7fc33518_d2ELU_1KzA
#> #define dELU _rx2dd6512f8f40bf129950e57dd80c7fc33519_dELU_f2H4
#> #define ELU _rx2dd6512f8f40bf129950e57dd80c7fc33520_ELU_Q2oF
#> #define dPReLUa1 _rx2dd6512f8f40bf129950e57dd80c7fc33521_dPReLUa1_0joC
#> #define dPReLUa _rx2dd6512f8f40bf129950e57dd80c7fc33522_dPReLUa_Gj0a
#> #define dPReLU _rx2dd6512f8f40bf129950e57dd80c7fc33523_dPReLU_8wAb
#> #define PReLU _rx2dd6512f8f40bf129950e57dd80c7fc33524_PReLU_QELO
#> #define dSwish _rx2dd6512f8f40bf129950e57dd80c7fc33525_dSwish_lObq
#> #define Swish _rx2dd6512f8f40bf129950e57dd80c7fc33526_Swish_Nath
#> #define dlReLU _rx2dd6512f8f40bf129950e57dd80c7fc33527_dlReLU_ltiW
#> #define lReLU _rx2dd6512f8f40bf129950e57dd80c7fc33528_lReLU_R2IG
#> #define dSELU _rx2dd6512f8f40bf129950e57dd80c7fc33529_dSELU_VLQY
#> #define SELU _rx2dd6512f8f40bf129950e57dd80c7fc33530_SELU_M84J
#> #define d4softplus _rx2dd6512f8f40bf129950e57dd80c7fc33531_d4softplus_Cnuy
#> #define d3softplus _rx2dd6512f8f40bf129950e57dd80c7fc33532_d3softplus_CXPJ
#> #define d2softplus _rx2dd6512f8f40bf129950e57dd80c7fc33533_d2softplus_UjvF
#> #define dsoftplus _rx2dd6512f8f40bf129950e57dd80c7fc33534_dsoftplus_k7KJ
#> #define softplus _rx2dd6512f8f40bf129950e57dd80c7fc33535_softplus_vDxq
#> #define d4GELU _rx2dd6512f8f40bf129950e57dd80c7fc33536_d4GELU_bVSK
#> #define d3GELU _rx2dd6512f8f40bf129950e57dd80c7fc33537_d3GELU_Vfr7
#> #define d2GELU _rx2dd6512f8f40bf129950e57dd80c7fc33538_d2GELU_TcaU
#> #define dGELU _rx2dd6512f8f40bf129950e57dd80c7fc33539_dGELU_EgpY
#> #define GELU _rx2dd6512f8f40bf129950e57dd80c7fc33540_GELU_kNIq
#> #define dReLU _rx2dd6512f8f40bf129950e57dd80c7fc33541_dReLU_nEnm
#> #define ReLU _rx2dd6512f8f40bf129950e57dd80c7fc33542_ReLU_hSbo
#> #define riweibull _rx2dd6512f8f40bf129950e57dd80c7fc33543_riweibull_MZYB
#> #define riunif _rx2dd6512f8f40bf129950e57dd80c7fc33544_riunif_jkiC
#> #define rit_ _rx2dd6512f8f40bf129950e57dd80c7fc33545_rit__oUqv
#> #define ripois _rx2dd6512f8f40bf129950e57dd80c7fc33546_ripois_niI7
#> #define ribeta _rx2dd6512f8f40bf129950e57dd80c7fc33547_ribeta_5Ui0
#> #define rigamma _rx2dd6512f8f40bf129950e57dd80c7fc33548_rigamma_GVTs
#> #define rigeom _rx2dd6512f8f40bf129950e57dd80c7fc33549_rigeom_34TN
#> #define rif _rx2dd6512f8f40bf129950e57dd80c7fc33550_rif_HeFp
#> #define riexp _rx2dd6512f8f40bf129950e57dd80c7fc33551_riexp_ue1Z
#> #define richisq _rx2dd6512f8f40bf129950e57dd80c7fc33552_richisq_aEnJ
#> #define ricauchy _rx2dd6512f8f40bf129950e57dd80c7fc33553_ricauchy_R6UH
#> #define rinbinomMu _rx2dd6512f8f40bf129950e57dd80c7fc33554_rinbinomMu_ZKbS
#> #define rinbinom _rx2dd6512f8f40bf129950e57dd80c7fc33555_rinbinom_smNY
#> #define ribinom _rx2dd6512f8f40bf129950e57dd80c7fc33556_ribinom_lS2w
#> #define rinorm _rx2dd6512f8f40bf129950e57dd80c7fc33557_rinorm_zYqw
#> #define rxweibull _rx2dd6512f8f40bf129950e57dd80c7fc33558_rxweibull_2z8b
#> #define rxunif _rx2dd6512f8f40bf129950e57dd80c7fc33559_rxunif_GMdj
#> #define rxt_ _rx2dd6512f8f40bf129950e57dd80c7fc33560_rxt__LB3L
#> #define rxpois _rx2dd6512f8f40bf129950e57dd80c7fc33561_rxpois_sOUR
#> #define rxbeta _rx2dd6512f8f40bf129950e57dd80c7fc33562_rxbeta_bPYU
#> #define rxgamma _rx2dd6512f8f40bf129950e57dd80c7fc33563_rxgamma_PmW8
#> #define rxgeom _rx2dd6512f8f40bf129950e57dd80c7fc33564_rxgeom_sOmG
#> #define rxf _rx2dd6512f8f40bf129950e57dd80c7fc33565_rxf_ypO5
#> #define rxexp _rx2dd6512f8f40bf129950e57dd80c7fc33566_rxexp_5OnO
#> #define rxchisq _rx2dd6512f8f40bf129950e57dd80c7fc33567_rxchisq_E9TF
#> #define rxcauchy _rx2dd6512f8f40bf129950e57dd80c7fc33568_rxcauchy_yOK7
#> #define rxnbinomMu _rx2dd6512f8f40bf129950e57dd80c7fc33569_rxnbinomMu_OqEK
#> #define rxnbinom _rx2dd6512f8f40bf129950e57dd80c7fc33570_rxnbinom_wH4T
#> #define rxbinom _rx2dd6512f8f40bf129950e57dd80c7fc33571_rxbinom_tSJj
#> #define rxnorm _rx2dd6512f8f40bf129950e57dd80c7fc33572_rxnorm_0v8j
#> #define linCmtB _rx2dd6512f8f40bf129950e57dd80c7fc33573_linCmtB_xwnn
#> #define linCmtA _rx2dd6512f8f40bf129950e57dd80c7fc33574_linCmtA_6QBD
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){  if (_solveData->svar[_svari] == 0) {size = _PP[0];};   if (_solveData->svar[_svari] == 1) {prob = _PP[1];}; }
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){  if (_solveData->ovar[_ovari] == 0) {size = _PP[0];};   if (_solveData->ovar[_ovari] == 1) {prob = _PP[1];}; }
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  rx_7be4440b671f3678cb2bbb37611f8bc5__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *rx_7be4440b671f3678cb2bbb37611f8bc5__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP rx_7be4440b671f3678cb2bbb37611f8bc5__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void rx_7be4440b671f3678cb2bbb37611f8bc5__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double size = NA_REAL;
#>   double prob = NA_REAL;
#>   double fx = NA_REAL;
#>   double dProb = NA_REAL;
#> 
#>   (void)t;
#>   (void)size;
#>   (void)prob;
#>   (void)fx;
#>   (void)dProb;
#> 
#>   fx = _PL[0];
#>   dProb = _PL[1];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#>   size = _PP[0];
#>   prob = _PP[1];
#> 
#> 
#>   fx =llikNbinom(t,size,prob);
#>   dProb =llikNbinomDprob(t,size,prob);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void rx_7be4440b671f3678cb2bbb37611f8bc5__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void rx_7be4440b671f3678cb2bbb37611f8bc5__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void rx_7be4440b671f3678cb2bbb37611f8bc5__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double size = NA_REAL;
#>   double prob = NA_REAL;
#>   double fx = NA_REAL;
#>   double dProb = NA_REAL;
#> 
#>   (void)t;
#>   (void)size;
#>   (void)prob;
#>   (void)fx;
#>   (void)dProb;
#> 
#>   fx = _PL[0];
#>   dProb = _PL[1];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#>   size = _PP[0];
#>   prob = _PP[1];
#> 
#> 
#>   fx =llikNbinom(t,size,prob);
#>   dProb =llikNbinomDprob(t,size,prob);
#> 
#>   _lhs[_LHS_0_]=fx;
#>   _lhs[_LHS_1_]=dProb;
#> }
#> // Functional based bioavailability
#> double rx_7be4440b671f3678cb2bbb37611f8bc5__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double rx_7be4440b671f3678cb2bbb37611f8bc5__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double rx_7be4440b671f3678cb2bbb37611f8bc5__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double rx_7be4440b671f3678cb2bbb37611f8bc5__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void rx_7be4440b671f3678cb2bbb37611f8bc5__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void rx_7be4440b671f3678cb2bbb37611f8bc5__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void rx_7be4440b671f3678cb2bbb37611f8bc5__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP rx_7be4440b671f3678cb2bbb37611f8bc5__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_7be4440b671f3678cb2bbb37611f8bc5__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAA[92AAAv7#aT)NpbvNA]>8a=WdwkC~1u%%C8Ae68jRE)Pk5X[f,k}#;M7V}0AM90^}iKP!xA0(iAw+|_:~da:8_(FOFY_\?B.3t6NEHx(T\"%%AXz6HK!<)uh\?T#Is;!$V=Mu:_\"^p1odBo{W^l0Rmn9^*|:7_<c1+2}21i3]r&D(7<H*dkt+L}O<5\?P#\?[|>18CU!mNa2u:qL%%#aWPx`/.DGES%%4#z>BW)AR/zj6\?ZTC3k(i%%@{RcmG/%%)EM{2/W:<9L0\"zfz}h4:QtCl\?rI>8H5Ez,xD+Y)58nSZ`\?gmqu1e93P`&>peNH1M%%xTKg@N`T<\"vjeHqj&Armjk,>;Z/Z`1aiuyEM:F_o=Ub1.!Y.gInxnNY!mf!aJIGuse\";mAKABY#]q96t)*iy*HAwP,HbB`&qAF>A0W^wDz9OV3re8J`}%%x^`rn:\"\?#kC9$HV:hpxmU{>}>k5yIjSJ(Pf@55HF*anx\?j1&hK`i1ejh9s/Idrq\"|Rp8f_l%%Bo\?tywoV,qDe#!o)Z0[62Ye]P36G]V@m*iB7@2!ONHl$#ON9V!#3mr>n.zO*$i{uRW!}\?v:pFF3\?M#!9ZkK~_uZ>MRohQp;t8+`0*4U]XlUSLv,ZL&LAJG:4oxWF|=oYa^W1;)ko(1O7OJop`+s1fxs(1#.oz&MlOe<0Go&Ts`wkl]\?p+*>pzkyskRxTa640hBaD3syPdoN|]wrdn2;Apw]y6LXrp\?lxg@8shdkgvMNCk=P:5B[y$q<0o+R<Pt6v`y@Xxe&9nID4XqZ7@PH^y>\">MiZ;*dZ>u7d(2yW6:(@CK5LAYHDuZ#Z.MI!fq1Nv:S\?#R]q:~,_m;sW@5@veTGac8@.}J.;mNEog~8y.UwtJU!&a{S8i]p\">*hekvFQ+Fa^fbLaG{6qt|$1%%:)@8SC,Iu\?1aLhBouAPZs<*<1=Bd<$zb11wyB\?/(f|dG^TLRjkoN=jc7jkDR*@VI.<+>ZEe6FduTFu/WZbs0sLz3hz$@cw$S_,d;(D0s\?jfVX/PK`\?~9jjF_d@wMvq4wU^oVE,Ut|uqrX(#Wns*7)u{jc,UImFxib5Tfl0_[N]P3milQS&Yye^+ktgJ");
#>     char buf[1119];
#> #define __doBufN__ 1119
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void rx_7be4440b671f3678cb2bbb37611f8bc5__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   rx_7be4440b671f3678cb2bbb37611f8bc5__dydt(neq, *t, A, DADT);
#> }
#> extern int rx_7be4440b671f3678cb2bbb37611f8bc5__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   rx_7be4440b671f3678cb2bbb37611f8bc5__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void rx_7be4440b671f3678cb2bbb37611f8bc5__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   rx_7be4440b671f3678cb2bbb37611f8bc5__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void rx_7be4440b671f3678cb2bbb37611f8bc5__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_rx_7be4440b671f3678cb2bbb37611f8bc5_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("rx_7be4440b671f3678cb2bbb37611f8bc5_","rx_7be4440b671f3678cb2bbb37611f8bc5__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("rx_7be4440b671f3678cb2bbb37611f8bc5_","rx_7be4440b671f3678cb2bbb37611f8bc5__assignFuns", (DL_FUNC) rx_7be4440b671f3678cb2bbb37611f8bc5__assignFuns);
#>   R_RegisterCCallable("rx_7be4440b671f3678cb2bbb37611f8bc5_","rx_7be4440b671f3678cb2bbb37611f8bc5__inis",(DL_FUNC) rx_7be4440b671f3678cb2bbb37611f8bc5__inis);
#>   R_RegisterCCallable("rx_7be4440b671f3678cb2bbb37611f8bc5_","rx_7be4440b671f3678cb2bbb37611f8bc5__dydt",(DL_FUNC) rx_7be4440b671f3678cb2bbb37611f8bc5__dydt);
#>   R_RegisterCCallable("rx_7be4440b671f3678cb2bbb37611f8bc5_","rx_7be4440b671f3678cb2bbb37611f8bc5__calc_lhs",(DL_FUNC) rx_7be4440b671f3678cb2bbb37611f8bc5__calc_lhs);
#>   R_RegisterCCallable("rx_7be4440b671f3678cb2bbb37611f8bc5_","rx_7be4440b671f3678cb2bbb37611f8bc5__calc_jac",(DL_FUNC) rx_7be4440b671f3678cb2bbb37611f8bc5__calc_jac);
#>   R_RegisterCCallable("rx_7be4440b671f3678cb2bbb37611f8bc5_","rx_7be4440b671f3678cb2bbb37611f8bc5__dydt_lsoda", (DL_FUNC) rx_7be4440b671f3678cb2bbb37611f8bc5__dydt_lsoda);
#>   R_RegisterCCallable("rx_7be4440b671f3678cb2bbb37611f8bc5_","rx_7be4440b671f3678cb2bbb37611f8bc5__calc_jac_lsoda", (DL_FUNC) rx_7be4440b671f3678cb2bbb37611f8bc5__calc_jac_lsoda);
#>   R_RegisterCCallable("rx_7be4440b671f3678cb2bbb37611f8bc5_","rx_7be4440b671f3678cb2bbb37611f8bc5__ode_solver_solvedata", (DL_FUNC) rx_7be4440b671f3678cb2bbb37611f8bc5__ode_solver_solvedata);
#>   R_RegisterCCallable("rx_7be4440b671f3678cb2bbb37611f8bc5_","rx_7be4440b671f3678cb2bbb37611f8bc5__ode_solver_get_solvedata", (DL_FUNC) rx_7be4440b671f3678cb2bbb37611f8bc5__ode_solver_get_solvedata);
#>   R_RegisterCCallable("rx_7be4440b671f3678cb2bbb37611f8bc5_","rx_7be4440b671f3678cb2bbb37611f8bc5__F", (DL_FUNC) rx_7be4440b671f3678cb2bbb37611f8bc5__F);
#>   R_RegisterCCallable("rx_7be4440b671f3678cb2bbb37611f8bc5_","rx_7be4440b671f3678cb2bbb37611f8bc5__Lag", (DL_FUNC) rx_7be4440b671f3678cb2bbb37611f8bc5__Lag);
#>   R_RegisterCCallable("rx_7be4440b671f3678cb2bbb37611f8bc5_","rx_7be4440b671f3678cb2bbb37611f8bc5__Rate", (DL_FUNC) rx_7be4440b671f3678cb2bbb37611f8bc5__Rate);
#>   R_RegisterCCallable("rx_7be4440b671f3678cb2bbb37611f8bc5_","rx_7be4440b671f3678cb2bbb37611f8bc5__Dur", (DL_FUNC) rx_7be4440b671f3678cb2bbb37611f8bc5__Dur);
#>   R_RegisterCCallable("rx_7be4440b671f3678cb2bbb37611f8bc5_","rx_7be4440b671f3678cb2bbb37611f8bc5__mtime", (DL_FUNC) rx_7be4440b671f3678cb2bbb37611f8bc5__mtime);
#>   R_RegisterCCallable("rx_7be4440b671f3678cb2bbb37611f8bc5_","rx_7be4440b671f3678cb2bbb37611f8bc5__ME", (DL_FUNC) rx_7be4440b671f3678cb2bbb37611f8bc5__ME);
#>   R_RegisterCCallable("rx_7be4440b671f3678cb2bbb37611f8bc5_","rx_7be4440b671f3678cb2bbb37611f8bc5__IndF", (DL_FUNC) rx_7be4440b671f3678cb2bbb37611f8bc5__IndF);
#>   R_RegisterCCallable("rx_7be4440b671f3678cb2bbb37611f8bc5_","rx_7be4440b671f3678cb2bbb37611f8bc5__dydt_liblsoda", (DL_FUNC) rx_7be4440b671f3678cb2bbb37611f8bc5__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_rx_7be4440b671f3678cb2bbb37611f8bc5_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_rx_7be4440b671f3678cb2bbb37611f8bc5_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"rx_7be4440b671f3678cb2bbb37611f8bc5__model_vars", (DL_FUNC) &rx_7be4440b671f3678cb2bbb37611f8bc5__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_rx_7be4440b671f3678cb2bbb37611f8bc5_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_7be4440b671f3678cb2bbb37611f8bc5__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("rx_7be4440b671f3678cb2bbb37611f8bc5__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’
#> ── Solved rxode2 object ──
#> ── Parameters (value$params): ──
#> # A tibble: 1 × 0
#> ── Initial Conditions (value$inits): ──
#> named numeric(0)
#> ── First part of data (object): ──
#> # A tibble: 9 × 5
#>    time    fx dProb  size  prob
#>   <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1    46 -13.3   108   100   0.5
#> 2    47 -12.8   106   100   0.5
#> 3    48 -12.4   104   100   0.5
#> 4    49 -12.0   102   100   0.5
#> 5    50 -11.6   100   100   0.5
#> 6    51 -11.2    98   100   0.5
#> # ℹ 3 more rows
# }