Calculate the log likelihood of the negative binomial function (and its derivatives)

Calculate the log likelihood of the negative binomial function (and its derivatives)

Usage

llikNbinomMu(x, size, mu, full = FALSE)

Arguments

x

Number of successes

size

Size of trial

mu

mu parameter for negative binomial

full

Add the data frame showing x, mean, sd as well as the fx and derivatives

Value

data frame with fx for the pdf value of with dProb that has the derivatives with respect to the parameters at the observation time-point

Details

In an rxode2() model, you can use llikNbinomMu() but you have to use all arguments. You can also get the derivative of mu with llikNbinomMuDmu()

Author

Matthew L. Fidler

Examples

# \donttest{
llikNbinomMu(46:54, 100, 40)
#>          fx       dMu
#> 1 -3.326828 0.1071429
#> 2 -3.446132 0.1250000
#> 3 -3.579663 0.1428571
#> 4 -3.727034 0.1607143
#> 5 -3.887874 0.1785714
#> 6 -4.061827 0.1964286
#> 7 -4.248554 0.2142857
#> 8 -4.447728 0.2321429
#> 9 -4.659037 0.2500000

llikNbinomMu(46:54, 100, 40, TRUE)
#>    x size mu        fx       dMu
#> 1 46  100 40 -3.326828 0.1071429
#> 2 47  100 40 -3.446132 0.1250000
#> 3 48  100 40 -3.579663 0.1428571
#> 4 49  100 40 -3.727034 0.1607143
#> 5 50  100 40 -3.887874 0.1785714
#> 6 51  100 40 -4.061827 0.1964286
#> 7 52  100 40 -4.248554 0.2142857
#> 8 53  100 40 -4.447728 0.2321429
#> 9 54  100 40 -4.659037 0.2500000

et <- et(46:54)
et$size <- 100
et$mu <- 40

model <- function() {
  model({
    fx <- llikNbinomMu(time, size, mu)
    dProb <- llikNbinomMuDmu(time, size, mu)
  })
}

rxSolve(model, et)
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#>  
#>  
#> C model:
#> // Define 2 LHS values
#> #define _LHS_0_ 0 // fx
#> #define _LHS_1_ 1 // dProb
#> #define _getRxSolve_ _rxd9eff4deffcdd856f90a213d60c09f9c3575__getRxSolve__0crr
#> #define _evalUdf _rxd9eff4deffcdd856f90a213d60c09f9c3576__evalUdf_gcGm
#> #define _solveData _rxd9eff4deffcdd856f90a213d60c09f9c3577__solveData_EsJu
#> #define _assign_ptr _rxd9eff4deffcdd856f90a213d60c09f9c3578__assign_ptr_9xah
#> #define _rxRmModelLib _rxd9eff4deffcdd856f90a213d60c09f9c3579__rxRmModelLib_gBOZ
#> #define _rxGetModelLib _rxd9eff4deffcdd856f90a213d60c09f9c3580__rxGetModelLib_l3zY
#> #define _old_c _rxd9eff4deffcdd856f90a213d60c09f9c3581__old_c_vTa9
#> #define _ptrid _rxd9eff4deffcdd856f90a213d60c09f9c3582__ptrid_7EhW
#> #define _rxIsCurrentC _rxd9eff4deffcdd856f90a213d60c09f9c3583__rxIsCurrentC_3MG9
#> #define _sumPS _rxd9eff4deffcdd856f90a213d60c09f9c3584__sumPS_V3pR
#> #define _prodPS _rxd9eff4deffcdd856f90a213d60c09f9c3585__prodPS_1aNk
#> #define _prodType _rxd9eff4deffcdd856f90a213d60c09f9c3586__prodType_G5Jw
#> #define _sumType _rxd9eff4deffcdd856f90a213d60c09f9c3587__sumType_D9TK
#> #define _update_par_ptr _rxd9eff4deffcdd856f90a213d60c09f9c3588__update_par_ptr_40NP
#> #define _getParCov _rxd9eff4deffcdd856f90a213d60c09f9c3589__getParCov_JiAL
#> #define _rxode2_rxAssignPtr _rxd9eff4deffcdd856f90a213d60c09f9c3590__rxode2_rxAssignPtr_ZBrh
#> #define _rxQr _rxd9eff4deffcdd856f90a213d60c09f9c3591__rxQr_efgh
#> #define _compareFactorVal _rxd9eff4deffcdd856f90a213d60c09f9c3592__compareFactorVal_Drm8
#> #define _sum _rxd9eff4deffcdd856f90a213d60c09f9c3593__sum_5igv
#> #define _udf _rxd9eff4deffcdd856f90a213d60c09f9c3594__udf_lJYg
#> #define _sign _rxd9eff4deffcdd856f90a213d60c09f9c3595__sign_GqQx
#> #define _prod _rxd9eff4deffcdd856f90a213d60c09f9c3596__prod_LoYl
#> #define _max _rxd9eff4deffcdd856f90a213d60c09f9c3597__max_FgJd
#> #define _min _rxd9eff4deffcdd856f90a213d60c09f9c3598__min_KaFg
#> #define _transit4P _rxd9eff4deffcdd856f90a213d60c09f9c3599__transit4P_mDR4
#> #define _transit3P _rxd9eff4deffcdd856f90a213d60c09f9c3600__transit3P_jxuo
#> #define _assignFuns0 _rxd9eff4deffcdd856f90a213d60c09f9c3601__assignFuns0_P8gm
#> #define _assignFuns _rxd9eff4deffcdd856f90a213d60c09f9c3602__assignFuns_oUom
#> #define _rxord _rxd9eff4deffcdd856f90a213d60c09f9c3603__rxord_wLtu
#> #define __assignFuns2 _rxd9eff4deffcdd856f90a213d60c09f9c3604___assignFuns2_DgOz
#> #define _llikCauchyDscale _rxd9eff4deffcdd856f90a213d60c09f9c3605__llikCauchyDscale_tPZw
#> #define _llikCauchyDlocation _rxd9eff4deffcdd856f90a213d60c09f9c3606__llikCauchyDlocation_Dqx2
#> #define _llikCauchy _rxd9eff4deffcdd856f90a213d60c09f9c3607__llikCauchy_t0Ff
#> #define _llikGammaDrate _rxd9eff4deffcdd856f90a213d60c09f9c3608__llikGammaDrate_CRup
#> #define _llikGammaDshape _rxd9eff4deffcdd856f90a213d60c09f9c3609__llikGammaDshape_BfcL
#> #define _llikGamma _rxd9eff4deffcdd856f90a213d60c09f9c3610__llikGamma_LlJk
#> #define _llikWeibullDscale _rxd9eff4deffcdd856f90a213d60c09f9c3611__llikWeibullDscale_0HOu
#> #define _llikWeibullDshape _rxd9eff4deffcdd856f90a213d60c09f9c3612__llikWeibullDshape_quMQ
#> #define _llikWeibull _rxd9eff4deffcdd856f90a213d60c09f9c3613__llikWeibull_IP4s
#> #define _llikUnifDbeta _rxd9eff4deffcdd856f90a213d60c09f9c3614__llikUnifDbeta_BKwl
#> #define _llikUnifDalpha _rxd9eff4deffcdd856f90a213d60c09f9c3615__llikUnifDalpha_ION1
#> #define _llikUnif _rxd9eff4deffcdd856f90a213d60c09f9c3616__llikUnif_tTX5
#> #define _llikGeomDp _rxd9eff4deffcdd856f90a213d60c09f9c3617__llikGeomDp_DulH
#> #define _llikGeom _rxd9eff4deffcdd856f90a213d60c09f9c3618__llikGeom_LxeA
#> #define _llikFDdf2 _rxd9eff4deffcdd856f90a213d60c09f9c3619__llikFDdf2_Ut0p
#> #define _llikFDdf1 _rxd9eff4deffcdd856f90a213d60c09f9c3620__llikFDdf1_o6Aq
#> #define _llikF _rxd9eff4deffcdd856f90a213d60c09f9c3621__llikF_SQkn
#> #define _llikExpDrate _rxd9eff4deffcdd856f90a213d60c09f9c3622__llikExpDrate_H8fP
#> #define _llikExp _rxd9eff4deffcdd856f90a213d60c09f9c3623__llikExp_CAnN
#> #define _llikChisqDdf _rxd9eff4deffcdd856f90a213d60c09f9c3624__llikChisqDdf_kUms
#> #define _llikChisq _rxd9eff4deffcdd856f90a213d60c09f9c3625__llikChisq_wiOW
#> #define _llikTDsd _rxd9eff4deffcdd856f90a213d60c09f9c3626__llikTDsd_G16H
#> #define _llikTDmean _rxd9eff4deffcdd856f90a213d60c09f9c3627__llikTDmean_PUgn
#> #define _llikTDdf _rxd9eff4deffcdd856f90a213d60c09f9c3628__llikTDdf_Cjuw
#> #define _llikT _rxd9eff4deffcdd856f90a213d60c09f9c3629__llikT_A8iw
#> #define _llikBetaDshape2 _rxd9eff4deffcdd856f90a213d60c09f9c3630__llikBetaDshape2_bn2G
#> #define _llikBetaDshape1 _rxd9eff4deffcdd856f90a213d60c09f9c3631__llikBetaDshape1_TQjS
#> #define _llikBeta _rxd9eff4deffcdd856f90a213d60c09f9c3632__llikBeta_rer5
#> #define _llikNbinomMuDmu _rxd9eff4deffcdd856f90a213d60c09f9c3633__llikNbinomMuDmu_zexL
#> #define _llikNbinomMu _rxd9eff4deffcdd856f90a213d60c09f9c3634__llikNbinomMu_gAwM
#> #define _llikNbinomDprob _rxd9eff4deffcdd856f90a213d60c09f9c3635__llikNbinomDprob_E6nq
#> #define _llikNbinom _rxd9eff4deffcdd856f90a213d60c09f9c3636__llikNbinom_R3DD
#> #define _llikBinomDprob _rxd9eff4deffcdd856f90a213d60c09f9c3637__llikBinomDprob_9jN5
#> #define _llikBinom _rxd9eff4deffcdd856f90a213d60c09f9c3638__llikBinom_O4r9
#> #define _llikPoisDlambda _rxd9eff4deffcdd856f90a213d60c09f9c3639__llikPoisDlambda_D97m
#> #define _llikPois _rxd9eff4deffcdd856f90a213d60c09f9c3640__llikPois_e5yb
#> #define _llikNormDsd _rxd9eff4deffcdd856f90a213d60c09f9c3641__llikNormDsd_V8Eu
#> #define _llikNormDmean _rxd9eff4deffcdd856f90a213d60c09f9c3642__llikNormDmean_Z1xT
#> #define _llikNorm _rxd9eff4deffcdd856f90a213d60c09f9c3643__llikNorm_QYOQ
#> #define simeps _rxd9eff4deffcdd856f90a213d60c09f9c3644_simeps_e6aq
#> #define simeta _rxd9eff4deffcdd856f90a213d60c09f9c3645_simeta_t3Lb
#> #define expit _rxd9eff4deffcdd856f90a213d60c09f9c3646_expit_sbea
#> #define logit _rxd9eff4deffcdd856f90a213d60c09f9c3647_logit_w5m4
#> #define gammapDer _rxd9eff4deffcdd856f90a213d60c09f9c3648_gammapDer_brD9
#> #define lowergamma _rxd9eff4deffcdd856f90a213d60c09f9c3649_lowergamma_EZAJ
#> #define uppergamma _rxd9eff4deffcdd856f90a213d60c09f9c3650_uppergamma_Ne7I
#> #define gammaqInva _rxd9eff4deffcdd856f90a213d60c09f9c3651_gammaqInva_xUFA
#> #define gammaqInv _rxd9eff4deffcdd856f90a213d60c09f9c3652_gammaqInv_5R2w
#> #define gammapInva _rxd9eff4deffcdd856f90a213d60c09f9c3653_gammapInva_ZwwB
#> #define gammapInv _rxd9eff4deffcdd856f90a213d60c09f9c3654_gammapInv_gRvb
#> #define gammaq _rxd9eff4deffcdd856f90a213d60c09f9c3655_gammaq_lnk3
#> #define gammap _rxd9eff4deffcdd856f90a213d60c09f9c3656_gammap_57cf
#> #define phi _rxd9eff4deffcdd856f90a213d60c09f9c3657_phi_UwMa
#> #define d2ELUa _rxd9eff4deffcdd856f90a213d60c09f9c3658_d2ELUa_rEA9
#> #define dELUa _rxd9eff4deffcdd856f90a213d60c09f9c3659_dELUa_z1SE
#> #define d2aELU _rxd9eff4deffcdd856f90a213d60c09f9c3660_d2aELU_1KzA
#> #define d2ELU _rxd9eff4deffcdd856f90a213d60c09f9c3661_d2ELU_f2H4
#> #define dELU _rxd9eff4deffcdd856f90a213d60c09f9c3662_dELU_Q2oF
#> #define ELU _rxd9eff4deffcdd856f90a213d60c09f9c3663_ELU_0joC
#> #define dPReLUa1 _rxd9eff4deffcdd856f90a213d60c09f9c3664_dPReLUa1_Gj0a
#> #define dPReLUa _rxd9eff4deffcdd856f90a213d60c09f9c3665_dPReLUa_8wAb
#> #define dPReLU _rxd9eff4deffcdd856f90a213d60c09f9c3666_dPReLU_QELO
#> #define PReLU _rxd9eff4deffcdd856f90a213d60c09f9c3667_PReLU_lObq
#> #define dSwish _rxd9eff4deffcdd856f90a213d60c09f9c3668_dSwish_Nath
#> #define Swish _rxd9eff4deffcdd856f90a213d60c09f9c3669_Swish_ltiW
#> #define dlReLU _rxd9eff4deffcdd856f90a213d60c09f9c3670_dlReLU_R2IG
#> #define lReLU _rxd9eff4deffcdd856f90a213d60c09f9c3671_lReLU_VLQY
#> #define dSELU _rxd9eff4deffcdd856f90a213d60c09f9c3672_dSELU_M84J
#> #define SELU _rxd9eff4deffcdd856f90a213d60c09f9c3673_SELU_Cnuy
#> #define d4softplus _rxd9eff4deffcdd856f90a213d60c09f9c3674_d4softplus_CXPJ
#> #define d3softplus _rxd9eff4deffcdd856f90a213d60c09f9c3675_d3softplus_UjvF
#> #define d2softplus _rxd9eff4deffcdd856f90a213d60c09f9c3676_d2softplus_k7KJ
#> #define dsoftplus _rxd9eff4deffcdd856f90a213d60c09f9c3677_dsoftplus_vDxq
#> #define softplus _rxd9eff4deffcdd856f90a213d60c09f9c3678_softplus_bVSK
#> #define d4GELU _rxd9eff4deffcdd856f90a213d60c09f9c3679_d4GELU_Vfr7
#> #define d3GELU _rxd9eff4deffcdd856f90a213d60c09f9c3680_d3GELU_TcaU
#> #define d2GELU _rxd9eff4deffcdd856f90a213d60c09f9c3681_d2GELU_EgpY
#> #define dGELU _rxd9eff4deffcdd856f90a213d60c09f9c3682_dGELU_kNIq
#> #define GELU _rxd9eff4deffcdd856f90a213d60c09f9c3683_GELU_nEnm
#> #define dReLU _rxd9eff4deffcdd856f90a213d60c09f9c3684_dReLU_hSbo
#> #define ReLU _rxd9eff4deffcdd856f90a213d60c09f9c3685_ReLU_MZYB
#> #define riweibull _rxd9eff4deffcdd856f90a213d60c09f9c3686_riweibull_jkiC
#> #define riunif _rxd9eff4deffcdd856f90a213d60c09f9c3687_riunif_oUqv
#> #define rit_ _rxd9eff4deffcdd856f90a213d60c09f9c3688_rit__niI7
#> #define ripois _rxd9eff4deffcdd856f90a213d60c09f9c3689_ripois_5Ui0
#> #define ribeta _rxd9eff4deffcdd856f90a213d60c09f9c3690_ribeta_GVTs
#> #define rigamma _rxd9eff4deffcdd856f90a213d60c09f9c3691_rigamma_34TN
#> #define rigeom _rxd9eff4deffcdd856f90a213d60c09f9c3692_rigeom_HeFp
#> #define rif _rxd9eff4deffcdd856f90a213d60c09f9c3693_rif_ue1Z
#> #define riexp _rxd9eff4deffcdd856f90a213d60c09f9c3694_riexp_aEnJ
#> #define richisq _rxd9eff4deffcdd856f90a213d60c09f9c3695_richisq_R6UH
#> #define ricauchy _rxd9eff4deffcdd856f90a213d60c09f9c3696_ricauchy_ZKbS
#> #define rinbinomMu _rxd9eff4deffcdd856f90a213d60c09f9c3697_rinbinomMu_smNY
#> #define rinbinom _rxd9eff4deffcdd856f90a213d60c09f9c3698_rinbinom_lS2w
#> #define ribinom _rxd9eff4deffcdd856f90a213d60c09f9c3699_ribinom_zYqw
#> #define rinorm _rxd9eff4deffcdd856f90a213d60c09f9c3700_rinorm_2z8b
#> #define rxweibull _rxd9eff4deffcdd856f90a213d60c09f9c3701_rxweibull_GMdj
#> #define rxunif _rxd9eff4deffcdd856f90a213d60c09f9c3702_rxunif_LB3L
#> #define rxt_ _rxd9eff4deffcdd856f90a213d60c09f9c3703_rxt__sOUR
#> #define rxpois _rxd9eff4deffcdd856f90a213d60c09f9c3704_rxpois_bPYU
#> #define rxbeta _rxd9eff4deffcdd856f90a213d60c09f9c3705_rxbeta_PmW8
#> #define rxgamma _rxd9eff4deffcdd856f90a213d60c09f9c3706_rxgamma_sOmG
#> #define rxgeom _rxd9eff4deffcdd856f90a213d60c09f9c3707_rxgeom_ypO5
#> #define rxf _rxd9eff4deffcdd856f90a213d60c09f9c3708_rxf_5OnO
#> #define rxexp _rxd9eff4deffcdd856f90a213d60c09f9c3709_rxexp_E9TF
#> #define rxchisq _rxd9eff4deffcdd856f90a213d60c09f9c3710_rxchisq_yOK7
#> #define rxcauchy _rxd9eff4deffcdd856f90a213d60c09f9c3711_rxcauchy_OqEK
#> #define rxnbinomMu _rxd9eff4deffcdd856f90a213d60c09f9c3712_rxnbinomMu_wH4T
#> #define rxnbinom _rxd9eff4deffcdd856f90a213d60c09f9c3713_rxnbinom_tSJj
#> #define rxbinom _rxd9eff4deffcdd856f90a213d60c09f9c3714_rxbinom_0v8j
#> #define rxnorm _rxd9eff4deffcdd856f90a213d60c09f9c3715_rxnorm_xwnn
#> #define linCmtB _rxd9eff4deffcdd856f90a213d60c09f9c3716_linCmtB_6QBD
#> #define linCmtA _rxd9eff4deffcdd856f90a213d60c09f9c3717_linCmtA_Uzzs
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){  if (_solveData->svar[_svari] == 0) {size = _PP[0];};   if (_solveData->svar[_svari] == 1) {mu = _PP[1];}; }
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){  if (_solveData->ovar[_ovari] == 0) {size = _PP[0];};   if (_solveData->ovar[_ovari] == 1) {mu = _PP[1];}; }
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  rx_5ee0cd88ae5461dd482ad497ead9b818__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *rx_5ee0cd88ae5461dd482ad497ead9b818__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP rx_5ee0cd88ae5461dd482ad497ead9b818__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void rx_5ee0cd88ae5461dd482ad497ead9b818__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double size = NA_REAL;
#>   double mu = NA_REAL;
#>   double fx = NA_REAL;
#>   double dProb = NA_REAL;
#> 
#>   (void)t;
#>   (void)size;
#>   (void)mu;
#>   (void)fx;
#>   (void)dProb;
#> 
#>   fx = _PL[0];
#>   dProb = _PL[1];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#>   size = _PP[0];
#>   mu = _PP[1];
#> 
#> 
#>   fx =llikNbinomMu(t,size,mu);
#>   dProb =llikNbinomMuDmu(t,size,mu);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void rx_5ee0cd88ae5461dd482ad497ead9b818__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void rx_5ee0cd88ae5461dd482ad497ead9b818__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void rx_5ee0cd88ae5461dd482ad497ead9b818__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double size = NA_REAL;
#>   double mu = NA_REAL;
#>   double fx = NA_REAL;
#>   double dProb = NA_REAL;
#> 
#>   (void)t;
#>   (void)size;
#>   (void)mu;
#>   (void)fx;
#>   (void)dProb;
#> 
#>   fx = _PL[0];
#>   dProb = _PL[1];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#>   size = _PP[0];
#>   mu = _PP[1];
#> 
#> 
#>   fx =llikNbinomMu(t,size,mu);
#>   dProb =llikNbinomMuDmu(t,size,mu);
#> 
#>   _lhs[_LHS_0_]=fx;
#>   _lhs[_LHS_1_]=dProb;
#> }
#> // Functional based bioavailability
#> double rx_5ee0cd88ae5461dd482ad497ead9b818__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double rx_5ee0cd88ae5461dd482ad497ead9b818__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double rx_5ee0cd88ae5461dd482ad497ead9b818__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double rx_5ee0cd88ae5461dd482ad497ead9b818__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void rx_5ee0cd88ae5461dd482ad497ead9b818__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void rx_5ee0cd88ae5461dd482ad497ead9b818__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void rx_5ee0cd88ae5461dd482ad497ead9b818__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP rx_5ee0cd88ae5461dd482ad497ead9b818__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_5ee0cd88ae5461dd482ad497ead9b818__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAAjn2AAAv7#aT)}o9pNA|\?0a=WdwYB~1u%%C8Ae68jRE)Pk5X[fT<XNSlDdQObrt6jj(7;I;x`AFDQwgU|^6JokcD\"s!,}\?B.JX\?MYLz(jApB#;[I:#h*gyng.j)SexHc\"R=@<_W~7fcd<j3j.e|lr_f%%Q|N2#1Fq=az9FIXoQ8h[I11q4$q\?{&N:7KdT*U%%TL#rC;j8BX@LS=@OUM%%&CvI\"QORdW+hnqz)#\?R,oH`&F)@4[s:Tu90$4L*R`]/VKxu+}Y*_qLT6jLzv@Q7fTt(W\"QHs),rfNR[}\?gv#UQoQ^99|M>Hq\"^Wo:p\"o9b^P@Mj$na<z4DXr7`Y\"5akaV,YEOp&\"Cacc,:k8##r&5HnxMOBMdP7JKg\"R1(D39CyY@Q:CUJ%%wZn\?)wQ6LsH/W5Rlo3n`/nJ8R[Qp1+MuZPMZ=NN_z=s`4)U\?JhjN#kCD7e/1DEix5f3{OpBQOZaly}/]IF9OMn3@mFORy;sJB{}rs3%%DJyMJE}$S\?|,nUetf5;;xIjd3Ux}v$OawxyGQ5aI3lJ}PT0tW:6d{ZX:\?>{Y>lG8i(cd~D6@+:Ztq6Gedv\"vjhaO}uGGU7KF0caOQu\?n1vq[R6VPV_.`>>mv\?^Ltv)8Jd`dhItR[[><z^twx*_4sy0R57o&]C1w$0P6lKJM%%a+{djz:N5Tv}42@OA=e|}l(+JZ=9E;FN$(vy&ZNZ^MN6%%Gt^kV*o]\?bqIpRdH\?+r>3zPR,FcUf~_W!;,d#[nVc/#wP.i6*SR@&1G`KyT^RnReCzjZvlEqNBN4D|5F)P7gEU^7B(vV#7LDM<cpbAYBz8YRDmv:,DT,iD;J\"n^1r.v7[HIiBR@Q!_hv]TI1YGIKCarLq6ZWhnu[[=1^*K}g9|9Qr~I`5I:.I[xhyJ/{OV:*pf{BW(G;ap.\"x5p<8dD.e6%%_.CYWc}4`[d.\"HJGT38Z~m8be{Ro8kH<)_s:so]KL,PiXw`O#^O$SR/,5:[)H(9Ro2OTK}#U]mR}:&}ot5nGGYyeG|{EWdol{\"|t_bdV)HWoq!)K%%iNO6[l3h]=V6TsWhJ7clt9:J];J|]VhUwq\?iS^G|Lm,|d%%t=#oI");
#>     char buf[1116];
#> #define __doBufN__ 1116
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void rx_5ee0cd88ae5461dd482ad497ead9b818__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   rx_5ee0cd88ae5461dd482ad497ead9b818__dydt(neq, *t, A, DADT);
#> }
#> extern int rx_5ee0cd88ae5461dd482ad497ead9b818__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   rx_5ee0cd88ae5461dd482ad497ead9b818__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void rx_5ee0cd88ae5461dd482ad497ead9b818__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   rx_5ee0cd88ae5461dd482ad497ead9b818__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void rx_5ee0cd88ae5461dd482ad497ead9b818__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_rx_5ee0cd88ae5461dd482ad497ead9b818_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("rx_5ee0cd88ae5461dd482ad497ead9b818_","rx_5ee0cd88ae5461dd482ad497ead9b818__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("rx_5ee0cd88ae5461dd482ad497ead9b818_","rx_5ee0cd88ae5461dd482ad497ead9b818__assignFuns", (DL_FUNC) rx_5ee0cd88ae5461dd482ad497ead9b818__assignFuns);
#>   R_RegisterCCallable("rx_5ee0cd88ae5461dd482ad497ead9b818_","rx_5ee0cd88ae5461dd482ad497ead9b818__inis",(DL_FUNC) rx_5ee0cd88ae5461dd482ad497ead9b818__inis);
#>   R_RegisterCCallable("rx_5ee0cd88ae5461dd482ad497ead9b818_","rx_5ee0cd88ae5461dd482ad497ead9b818__dydt",(DL_FUNC) rx_5ee0cd88ae5461dd482ad497ead9b818__dydt);
#>   R_RegisterCCallable("rx_5ee0cd88ae5461dd482ad497ead9b818_","rx_5ee0cd88ae5461dd482ad497ead9b818__calc_lhs",(DL_FUNC) rx_5ee0cd88ae5461dd482ad497ead9b818__calc_lhs);
#>   R_RegisterCCallable("rx_5ee0cd88ae5461dd482ad497ead9b818_","rx_5ee0cd88ae5461dd482ad497ead9b818__calc_jac",(DL_FUNC) rx_5ee0cd88ae5461dd482ad497ead9b818__calc_jac);
#>   R_RegisterCCallable("rx_5ee0cd88ae5461dd482ad497ead9b818_","rx_5ee0cd88ae5461dd482ad497ead9b818__dydt_lsoda", (DL_FUNC) rx_5ee0cd88ae5461dd482ad497ead9b818__dydt_lsoda);
#>   R_RegisterCCallable("rx_5ee0cd88ae5461dd482ad497ead9b818_","rx_5ee0cd88ae5461dd482ad497ead9b818__calc_jac_lsoda", (DL_FUNC) rx_5ee0cd88ae5461dd482ad497ead9b818__calc_jac_lsoda);
#>   R_RegisterCCallable("rx_5ee0cd88ae5461dd482ad497ead9b818_","rx_5ee0cd88ae5461dd482ad497ead9b818__ode_solver_solvedata", (DL_FUNC) rx_5ee0cd88ae5461dd482ad497ead9b818__ode_solver_solvedata);
#>   R_RegisterCCallable("rx_5ee0cd88ae5461dd482ad497ead9b818_","rx_5ee0cd88ae5461dd482ad497ead9b818__ode_solver_get_solvedata", (DL_FUNC) rx_5ee0cd88ae5461dd482ad497ead9b818__ode_solver_get_solvedata);
#>   R_RegisterCCallable("rx_5ee0cd88ae5461dd482ad497ead9b818_","rx_5ee0cd88ae5461dd482ad497ead9b818__F", (DL_FUNC) rx_5ee0cd88ae5461dd482ad497ead9b818__F);
#>   R_RegisterCCallable("rx_5ee0cd88ae5461dd482ad497ead9b818_","rx_5ee0cd88ae5461dd482ad497ead9b818__Lag", (DL_FUNC) rx_5ee0cd88ae5461dd482ad497ead9b818__Lag);
#>   R_RegisterCCallable("rx_5ee0cd88ae5461dd482ad497ead9b818_","rx_5ee0cd88ae5461dd482ad497ead9b818__Rate", (DL_FUNC) rx_5ee0cd88ae5461dd482ad497ead9b818__Rate);
#>   R_RegisterCCallable("rx_5ee0cd88ae5461dd482ad497ead9b818_","rx_5ee0cd88ae5461dd482ad497ead9b818__Dur", (DL_FUNC) rx_5ee0cd88ae5461dd482ad497ead9b818__Dur);
#>   R_RegisterCCallable("rx_5ee0cd88ae5461dd482ad497ead9b818_","rx_5ee0cd88ae5461dd482ad497ead9b818__mtime", (DL_FUNC) rx_5ee0cd88ae5461dd482ad497ead9b818__mtime);
#>   R_RegisterCCallable("rx_5ee0cd88ae5461dd482ad497ead9b818_","rx_5ee0cd88ae5461dd482ad497ead9b818__ME", (DL_FUNC) rx_5ee0cd88ae5461dd482ad497ead9b818__ME);
#>   R_RegisterCCallable("rx_5ee0cd88ae5461dd482ad497ead9b818_","rx_5ee0cd88ae5461dd482ad497ead9b818__IndF", (DL_FUNC) rx_5ee0cd88ae5461dd482ad497ead9b818__IndF);
#>   R_RegisterCCallable("rx_5ee0cd88ae5461dd482ad497ead9b818_","rx_5ee0cd88ae5461dd482ad497ead9b818__dydt_liblsoda", (DL_FUNC) rx_5ee0cd88ae5461dd482ad497ead9b818__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_rx_5ee0cd88ae5461dd482ad497ead9b818_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_rx_5ee0cd88ae5461dd482ad497ead9b818_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"rx_5ee0cd88ae5461dd482ad497ead9b818__model_vars", (DL_FUNC) &rx_5ee0cd88ae5461dd482ad497ead9b818__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_rx_5ee0cd88ae5461dd482ad497ead9b818_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_5ee0cd88ae5461dd482ad497ead9b818__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("rx_5ee0cd88ae5461dd482ad497ead9b818__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’
#> ── Solved rxode2 object ──
#> ── Parameters (value$params): ──
#> # A tibble: 1 × 0
#> ── Initial Conditions (value$inits): ──
#> named numeric(0)
#> ── First part of data (object): ──
#> # A tibble: 9 × 5
#>    time    fx dProb  size    mu
#>   <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1    46 -3.33 0.107   100    40
#> 2    47 -3.45 0.125   100    40
#> 3    48 -3.58 0.143   100    40
#> 4    49 -3.73 0.161   100    40
#> 5    50 -3.89 0.179   100    40
#> 6    51 -4.06 0.196   100    40
#> # ℹ 3 more rows
# }