Log likelihood for normal distribution

Log likelihood for normal distribution

Usage

llikNorm(x, mean = 0, sd = 1, full = FALSE)

Arguments

x

Observation

mean

Mean for the likelihood

sd

Standard deviation for the likelihood

full

Add the data frame showing x, mean, sd as well as the fx and derivatives

Value

data frame with fx for the pdf value of with dMean and dSd that has the derivatives with respect to the parameters at the observation time-point

Details

In an rxode2() model, you can use llikNorm() but you have to use all arguments. You can also get the derivatives with llikNormDmean() and llikNormDsd()

Author

Matthew L. Fidler

Examples

# \donttest{

llikNorm(0)
#>           fx dMean dSd
#> 1 -0.9189385     0  -1

llikNorm(seq(-2,2,length.out=10), full=TRUE)
#>             x mean sd         fx      dMean        dSd
#> 1  -2.0000000    0  1 -2.9189385 -2.0000000  3.0000000
#> 2  -1.5555556    0  1 -2.1288151 -1.5555556  1.4197531
#> 3  -1.1111111    0  1 -1.5362225 -1.1111111  0.2345679
#> 4  -0.6666667    0  1 -1.1411608 -0.6666667 -0.5555556
#> 5  -0.2222222    0  1 -0.9436299 -0.2222222 -0.9506173
#> 6   0.2222222    0  1 -0.9436299  0.2222222 -0.9506173
#> 7   0.6666667    0  1 -1.1411608  0.6666667 -0.5555556
#> 8   1.1111111    0  1 -1.5362225  1.1111111  0.2345679
#> 9   1.5555556    0  1 -2.1288151  1.5555556  1.4197531
#> 10  2.0000000    0  1 -2.9189385  2.0000000  3.0000000

# With rxode2 you can use:
 
et <- et(-3, 3, length.out=10)
et$mu <- 0
et$sigma <- 1

model <- function(){
  model({
    fx <- llikNorm(time, mu, sigma)
    dMean <- llikNormDmean(time, mu, sigma)
    dSd <- llikNormDsd(time, mu, sigma)
  })
 }

ret <- rxSolve(model, et)
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#>  
#>  
#> C model:
#> // Define 3 LHS values
#> #define _LHS_0_ 0 // fx
#> #define _LHS_1_ 1 // dMean
#> #define _LHS_2_ 2 // dSd
#> #define _getRxSolve_ _rxf53dabf3a77f3951d3aaee55fb36977c3718__getRxSolve__gcGm
#> #define _evalUdf _rxf53dabf3a77f3951d3aaee55fb36977c3719__evalUdf_EsJu
#> #define _solveData _rxf53dabf3a77f3951d3aaee55fb36977c3720__solveData_9xah
#> #define _assign_ptr _rxf53dabf3a77f3951d3aaee55fb36977c3721__assign_ptr_gBOZ
#> #define _rxRmModelLib _rxf53dabf3a77f3951d3aaee55fb36977c3722__rxRmModelLib_l3zY
#> #define _rxGetModelLib _rxf53dabf3a77f3951d3aaee55fb36977c3723__rxGetModelLib_vTa9
#> #define _old_c _rxf53dabf3a77f3951d3aaee55fb36977c3724__old_c_7EhW
#> #define _ptrid _rxf53dabf3a77f3951d3aaee55fb36977c3725__ptrid_3MG9
#> #define _rxIsCurrentC _rxf53dabf3a77f3951d3aaee55fb36977c3726__rxIsCurrentC_V3pR
#> #define _sumPS _rxf53dabf3a77f3951d3aaee55fb36977c3727__sumPS_1aNk
#> #define _prodPS _rxf53dabf3a77f3951d3aaee55fb36977c3728__prodPS_G5Jw
#> #define _prodType _rxf53dabf3a77f3951d3aaee55fb36977c3729__prodType_D9TK
#> #define _sumType _rxf53dabf3a77f3951d3aaee55fb36977c3730__sumType_40NP
#> #define _update_par_ptr _rxf53dabf3a77f3951d3aaee55fb36977c3731__update_par_ptr_JiAL
#> #define _getParCov _rxf53dabf3a77f3951d3aaee55fb36977c3732__getParCov_ZBrh
#> #define _rxode2_rxAssignPtr _rxf53dabf3a77f3951d3aaee55fb36977c3733__rxode2_rxAssignPtr_efgh
#> #define _rxQr _rxf53dabf3a77f3951d3aaee55fb36977c3734__rxQr_Drm8
#> #define _compareFactorVal _rxf53dabf3a77f3951d3aaee55fb36977c3735__compareFactorVal_5igv
#> #define _sum _rxf53dabf3a77f3951d3aaee55fb36977c3736__sum_lJYg
#> #define _udf _rxf53dabf3a77f3951d3aaee55fb36977c3737__udf_GqQx
#> #define _sign _rxf53dabf3a77f3951d3aaee55fb36977c3738__sign_LoYl
#> #define _prod _rxf53dabf3a77f3951d3aaee55fb36977c3739__prod_FgJd
#> #define _max _rxf53dabf3a77f3951d3aaee55fb36977c3740__max_KaFg
#> #define _min _rxf53dabf3a77f3951d3aaee55fb36977c3741__min_mDR4
#> #define _transit4P _rxf53dabf3a77f3951d3aaee55fb36977c3742__transit4P_jxuo
#> #define _transit3P _rxf53dabf3a77f3951d3aaee55fb36977c3743__transit3P_P8gm
#> #define _assignFuns0 _rxf53dabf3a77f3951d3aaee55fb36977c3744__assignFuns0_oUom
#> #define _assignFuns _rxf53dabf3a77f3951d3aaee55fb36977c3745__assignFuns_wLtu
#> #define _rxord _rxf53dabf3a77f3951d3aaee55fb36977c3746__rxord_DgOz
#> #define __assignFuns2 _rxf53dabf3a77f3951d3aaee55fb36977c3747___assignFuns2_tPZw
#> #define _llikCauchyDscale _rxf53dabf3a77f3951d3aaee55fb36977c3748__llikCauchyDscale_Dqx2
#> #define _llikCauchyDlocation _rxf53dabf3a77f3951d3aaee55fb36977c3749__llikCauchyDlocation_t0Ff
#> #define _llikCauchy _rxf53dabf3a77f3951d3aaee55fb36977c3750__llikCauchy_CRup
#> #define _llikGammaDrate _rxf53dabf3a77f3951d3aaee55fb36977c3751__llikGammaDrate_BfcL
#> #define _llikGammaDshape _rxf53dabf3a77f3951d3aaee55fb36977c3752__llikGammaDshape_LlJk
#> #define _llikGamma _rxf53dabf3a77f3951d3aaee55fb36977c3753__llikGamma_0HOu
#> #define _llikWeibullDscale _rxf53dabf3a77f3951d3aaee55fb36977c3754__llikWeibullDscale_quMQ
#> #define _llikWeibullDshape _rxf53dabf3a77f3951d3aaee55fb36977c3755__llikWeibullDshape_IP4s
#> #define _llikWeibull _rxf53dabf3a77f3951d3aaee55fb36977c3756__llikWeibull_BKwl
#> #define _llikUnifDbeta _rxf53dabf3a77f3951d3aaee55fb36977c3757__llikUnifDbeta_ION1
#> #define _llikUnifDalpha _rxf53dabf3a77f3951d3aaee55fb36977c3758__llikUnifDalpha_tTX5
#> #define _llikUnif _rxf53dabf3a77f3951d3aaee55fb36977c3759__llikUnif_DulH
#> #define _llikGeomDp _rxf53dabf3a77f3951d3aaee55fb36977c3760__llikGeomDp_LxeA
#> #define _llikGeom _rxf53dabf3a77f3951d3aaee55fb36977c3761__llikGeom_Ut0p
#> #define _llikFDdf2 _rxf53dabf3a77f3951d3aaee55fb36977c3762__llikFDdf2_o6Aq
#> #define _llikFDdf1 _rxf53dabf3a77f3951d3aaee55fb36977c3763__llikFDdf1_SQkn
#> #define _llikF _rxf53dabf3a77f3951d3aaee55fb36977c3764__llikF_H8fP
#> #define _llikExpDrate _rxf53dabf3a77f3951d3aaee55fb36977c3765__llikExpDrate_CAnN
#> #define _llikExp _rxf53dabf3a77f3951d3aaee55fb36977c3766__llikExp_kUms
#> #define _llikChisqDdf _rxf53dabf3a77f3951d3aaee55fb36977c3767__llikChisqDdf_wiOW
#> #define _llikChisq _rxf53dabf3a77f3951d3aaee55fb36977c3768__llikChisq_G16H
#> #define _llikTDsd _rxf53dabf3a77f3951d3aaee55fb36977c3769__llikTDsd_PUgn
#> #define _llikTDmean _rxf53dabf3a77f3951d3aaee55fb36977c3770__llikTDmean_Cjuw
#> #define _llikTDdf _rxf53dabf3a77f3951d3aaee55fb36977c3771__llikTDdf_A8iw
#> #define _llikT _rxf53dabf3a77f3951d3aaee55fb36977c3772__llikT_bn2G
#> #define _llikBetaDshape2 _rxf53dabf3a77f3951d3aaee55fb36977c3773__llikBetaDshape2_TQjS
#> #define _llikBetaDshape1 _rxf53dabf3a77f3951d3aaee55fb36977c3774__llikBetaDshape1_rer5
#> #define _llikBeta _rxf53dabf3a77f3951d3aaee55fb36977c3775__llikBeta_zexL
#> #define _llikNbinomMuDmu _rxf53dabf3a77f3951d3aaee55fb36977c3776__llikNbinomMuDmu_gAwM
#> #define _llikNbinomMu _rxf53dabf3a77f3951d3aaee55fb36977c3777__llikNbinomMu_E6nq
#> #define _llikNbinomDprob _rxf53dabf3a77f3951d3aaee55fb36977c3778__llikNbinomDprob_R3DD
#> #define _llikNbinom _rxf53dabf3a77f3951d3aaee55fb36977c3779__llikNbinom_9jN5
#> #define _llikBinomDprob _rxf53dabf3a77f3951d3aaee55fb36977c3780__llikBinomDprob_O4r9
#> #define _llikBinom _rxf53dabf3a77f3951d3aaee55fb36977c3781__llikBinom_D97m
#> #define _llikPoisDlambda _rxf53dabf3a77f3951d3aaee55fb36977c3782__llikPoisDlambda_e5yb
#> #define _llikPois _rxf53dabf3a77f3951d3aaee55fb36977c3783__llikPois_V8Eu
#> #define _llikNormDsd _rxf53dabf3a77f3951d3aaee55fb36977c3784__llikNormDsd_Z1xT
#> #define _llikNormDmean _rxf53dabf3a77f3951d3aaee55fb36977c3785__llikNormDmean_QYOQ
#> #define _llikNorm _rxf53dabf3a77f3951d3aaee55fb36977c3786__llikNorm_e6aq
#> #define simeps _rxf53dabf3a77f3951d3aaee55fb36977c3787_simeps_t3Lb
#> #define simeta _rxf53dabf3a77f3951d3aaee55fb36977c3788_simeta_sbea
#> #define expit _rxf53dabf3a77f3951d3aaee55fb36977c3789_expit_w5m4
#> #define logit _rxf53dabf3a77f3951d3aaee55fb36977c3790_logit_brD9
#> #define gammapDer _rxf53dabf3a77f3951d3aaee55fb36977c3791_gammapDer_EZAJ
#> #define lowergamma _rxf53dabf3a77f3951d3aaee55fb36977c3792_lowergamma_Ne7I
#> #define uppergamma _rxf53dabf3a77f3951d3aaee55fb36977c3793_uppergamma_xUFA
#> #define gammaqInva _rxf53dabf3a77f3951d3aaee55fb36977c3794_gammaqInva_5R2w
#> #define gammaqInv _rxf53dabf3a77f3951d3aaee55fb36977c3795_gammaqInv_ZwwB
#> #define gammapInva _rxf53dabf3a77f3951d3aaee55fb36977c3796_gammapInva_gRvb
#> #define gammapInv _rxf53dabf3a77f3951d3aaee55fb36977c3797_gammapInv_lnk3
#> #define gammaq _rxf53dabf3a77f3951d3aaee55fb36977c3798_gammaq_57cf
#> #define gammap _rxf53dabf3a77f3951d3aaee55fb36977c3799_gammap_UwMa
#> #define phi _rxf53dabf3a77f3951d3aaee55fb36977c3800_phi_rEA9
#> #define d2ELUa _rxf53dabf3a77f3951d3aaee55fb36977c3801_d2ELUa_z1SE
#> #define dELUa _rxf53dabf3a77f3951d3aaee55fb36977c3802_dELUa_1KzA
#> #define d2aELU _rxf53dabf3a77f3951d3aaee55fb36977c3803_d2aELU_f2H4
#> #define d2ELU _rxf53dabf3a77f3951d3aaee55fb36977c3804_d2ELU_Q2oF
#> #define dELU _rxf53dabf3a77f3951d3aaee55fb36977c3805_dELU_0joC
#> #define ELU _rxf53dabf3a77f3951d3aaee55fb36977c3806_ELU_Gj0a
#> #define dPReLUa1 _rxf53dabf3a77f3951d3aaee55fb36977c3807_dPReLUa1_8wAb
#> #define dPReLUa _rxf53dabf3a77f3951d3aaee55fb36977c3808_dPReLUa_QELO
#> #define dPReLU _rxf53dabf3a77f3951d3aaee55fb36977c3809_dPReLU_lObq
#> #define PReLU _rxf53dabf3a77f3951d3aaee55fb36977c3810_PReLU_Nath
#> #define dSwish _rxf53dabf3a77f3951d3aaee55fb36977c3811_dSwish_ltiW
#> #define Swish _rxf53dabf3a77f3951d3aaee55fb36977c3812_Swish_R2IG
#> #define dlReLU _rxf53dabf3a77f3951d3aaee55fb36977c3813_dlReLU_VLQY
#> #define lReLU _rxf53dabf3a77f3951d3aaee55fb36977c3814_lReLU_M84J
#> #define dSELU _rxf53dabf3a77f3951d3aaee55fb36977c3815_dSELU_Cnuy
#> #define SELU _rxf53dabf3a77f3951d3aaee55fb36977c3816_SELU_CXPJ
#> #define d4softplus _rxf53dabf3a77f3951d3aaee55fb36977c3817_d4softplus_UjvF
#> #define d3softplus _rxf53dabf3a77f3951d3aaee55fb36977c3818_d3softplus_k7KJ
#> #define d2softplus _rxf53dabf3a77f3951d3aaee55fb36977c3819_d2softplus_vDxq
#> #define dsoftplus _rxf53dabf3a77f3951d3aaee55fb36977c3820_dsoftplus_bVSK
#> #define softplus _rxf53dabf3a77f3951d3aaee55fb36977c3821_softplus_Vfr7
#> #define d4GELU _rxf53dabf3a77f3951d3aaee55fb36977c3822_d4GELU_TcaU
#> #define d3GELU _rxf53dabf3a77f3951d3aaee55fb36977c3823_d3GELU_EgpY
#> #define d2GELU _rxf53dabf3a77f3951d3aaee55fb36977c3824_d2GELU_kNIq
#> #define dGELU _rxf53dabf3a77f3951d3aaee55fb36977c3825_dGELU_nEnm
#> #define GELU _rxf53dabf3a77f3951d3aaee55fb36977c3826_GELU_hSbo
#> #define dReLU _rxf53dabf3a77f3951d3aaee55fb36977c3827_dReLU_MZYB
#> #define ReLU _rxf53dabf3a77f3951d3aaee55fb36977c3828_ReLU_jkiC
#> #define riweibull _rxf53dabf3a77f3951d3aaee55fb36977c3829_riweibull_oUqv
#> #define riunif _rxf53dabf3a77f3951d3aaee55fb36977c3830_riunif_niI7
#> #define rit_ _rxf53dabf3a77f3951d3aaee55fb36977c3831_rit__5Ui0
#> #define ripois _rxf53dabf3a77f3951d3aaee55fb36977c3832_ripois_GVTs
#> #define ribeta _rxf53dabf3a77f3951d3aaee55fb36977c3833_ribeta_34TN
#> #define rigamma _rxf53dabf3a77f3951d3aaee55fb36977c3834_rigamma_HeFp
#> #define rigeom _rxf53dabf3a77f3951d3aaee55fb36977c3835_rigeom_ue1Z
#> #define rif _rxf53dabf3a77f3951d3aaee55fb36977c3836_rif_aEnJ
#> #define riexp _rxf53dabf3a77f3951d3aaee55fb36977c3837_riexp_R6UH
#> #define richisq _rxf53dabf3a77f3951d3aaee55fb36977c3838_richisq_ZKbS
#> #define ricauchy _rxf53dabf3a77f3951d3aaee55fb36977c3839_ricauchy_smNY
#> #define rinbinomMu _rxf53dabf3a77f3951d3aaee55fb36977c3840_rinbinomMu_lS2w
#> #define rinbinom _rxf53dabf3a77f3951d3aaee55fb36977c3841_rinbinom_zYqw
#> #define ribinom _rxf53dabf3a77f3951d3aaee55fb36977c3842_ribinom_2z8b
#> #define rinorm _rxf53dabf3a77f3951d3aaee55fb36977c3843_rinorm_GMdj
#> #define rxweibull _rxf53dabf3a77f3951d3aaee55fb36977c3844_rxweibull_LB3L
#> #define rxunif _rxf53dabf3a77f3951d3aaee55fb36977c3845_rxunif_sOUR
#> #define rxt_ _rxf53dabf3a77f3951d3aaee55fb36977c3846_rxt__bPYU
#> #define rxpois _rxf53dabf3a77f3951d3aaee55fb36977c3847_rxpois_PmW8
#> #define rxbeta _rxf53dabf3a77f3951d3aaee55fb36977c3848_rxbeta_sOmG
#> #define rxgamma _rxf53dabf3a77f3951d3aaee55fb36977c3849_rxgamma_ypO5
#> #define rxgeom _rxf53dabf3a77f3951d3aaee55fb36977c3850_rxgeom_5OnO
#> #define rxf _rxf53dabf3a77f3951d3aaee55fb36977c3851_rxf_E9TF
#> #define rxexp _rxf53dabf3a77f3951d3aaee55fb36977c3852_rxexp_yOK7
#> #define rxchisq _rxf53dabf3a77f3951d3aaee55fb36977c3853_rxchisq_OqEK
#> #define rxcauchy _rxf53dabf3a77f3951d3aaee55fb36977c3854_rxcauchy_wH4T
#> #define rxnbinomMu _rxf53dabf3a77f3951d3aaee55fb36977c3855_rxnbinomMu_tSJj
#> #define rxnbinom _rxf53dabf3a77f3951d3aaee55fb36977c3856_rxnbinom_0v8j
#> #define rxbinom _rxf53dabf3a77f3951d3aaee55fb36977c3857_rxbinom_xwnn
#> #define rxnorm _rxf53dabf3a77f3951d3aaee55fb36977c3858_rxnorm_6QBD
#> #define linCmtB _rxf53dabf3a77f3951d3aaee55fb36977c3859_linCmtB_Uzzs
#> #define linCmtA _rxf53dabf3a77f3951d3aaee55fb36977c3860_linCmtA_oz5n
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){  if (_solveData->svar[_svari] == 0) {mu = _PP[0];};   if (_solveData->svar[_svari] == 1) {sigma = _PP[1];}; }
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){  if (_solveData->ovar[_ovari] == 0) {mu = _PP[0];};   if (_solveData->ovar[_ovari] == 1) {sigma = _PP[1];}; }
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  rx_2caa923c2e1313664984f5d665cc9cff__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *rx_2caa923c2e1313664984f5d665cc9cff__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP rx_2caa923c2e1313664984f5d665cc9cff__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void rx_2caa923c2e1313664984f5d665cc9cff__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double mu = NA_REAL;
#>   double sigma = NA_REAL;
#>   double fx = NA_REAL;
#>   double dMean = NA_REAL;
#>   double dSd = NA_REAL;
#> 
#>   (void)t;
#>   (void)mu;
#>   (void)sigma;
#>   (void)fx;
#>   (void)dMean;
#>   (void)dSd;
#> 
#>   fx = _PL[0];
#>   dMean = _PL[1];
#>   dSd = _PL[2];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#>   mu = _PP[0];
#>   sigma = _PP[1];
#> 
#> 
#>   fx =llikNorm(t,mu,sigma);
#>   dMean =llikNormDmean(t,mu,sigma);
#>   dSd =llikNormDsd(t,mu,sigma);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void rx_2caa923c2e1313664984f5d665cc9cff__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void rx_2caa923c2e1313664984f5d665cc9cff__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void rx_2caa923c2e1313664984f5d665cc9cff__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double mu = NA_REAL;
#>   double sigma = NA_REAL;
#>   double fx = NA_REAL;
#>   double dMean = NA_REAL;
#>   double dSd = NA_REAL;
#> 
#>   (void)t;
#>   (void)mu;
#>   (void)sigma;
#>   (void)fx;
#>   (void)dMean;
#>   (void)dSd;
#> 
#>   fx = _PL[0];
#>   dMean = _PL[1];
#>   dSd = _PL[2];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#>   mu = _PP[0];
#>   sigma = _PP[1];
#> 
#> 
#>   fx =llikNorm(t,mu,sigma);
#>   dMean =llikNormDmean(t,mu,sigma);
#>   dSd =llikNormDsd(t,mu,sigma);
#> 
#>   _lhs[_LHS_0_]=fx;
#>   _lhs[_LHS_1_]=dMean;
#>   _lhs[_LHS_2_]=dSd;
#> }
#> // Functional based bioavailability
#> double rx_2caa923c2e1313664984f5d665cc9cff__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double rx_2caa923c2e1313664984f5d665cc9cff__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double rx_2caa923c2e1313664984f5d665cc9cff__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double rx_2caa923c2e1313664984f5d665cc9cff__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void rx_2caa923c2e1313664984f5d665cc9cff__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void rx_2caa923c2e1313664984f5d665cc9cff__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void rx_2caa923c2e1313664984f5d665cc9cff__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP rx_2caa923c2e1313664984f5d665cc9cff__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_2caa923c2e1313664984f5d665cc9cff__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAAMc3AAAv7#aT)Cqh9NAr^@a=WdwYB~1u%%C8Ae68jR&AE/iengB+vZ;l@cQObrt6jj(7L$:lY\"*ADzEZw{!Nk;,t*>mov)A.>4KobEl_Jt(ABlz;lHlVW!|6]=bR:sF.<;c*n$Bm]xM7L6Sho(ZA!of_\"XliV>:jCiEVm]0#m3#gzBo>s+{|v9@\"KKLms_[/AKp17G`FhFu1>NT{{K7Z&1!*ARr|]/Sw\?HBRDPDPqP,%%fpMh`r`GXvJYYyliI/idAN5RnpU*e&_C~JjvDH5/Nf%%Y^i9#Q0qG]}\?gvS.ahPzwr>K&k>H&zg9>O6`zqQH5J@n[_+pBDXH2gBdNZY[Dqx\"VSwXXPz!=3^l*z&.PsI<TIqk}xOBzE;~mRQsO2t)RvDeGy9G8e=(n4IKbA\?|Lc]35\"ix;\"T+qa\"vPKYyc{h$:I{Q,JC*w|CrZbbVPn_X$;mBTe/C\"9\?jMyIjSK$lK[V{Pj!m3>irU;{Kr$|c+:hzFWlW=D+&8u\"#2]c~T*`uZxbZ6jRoTa0_~1,UTlf:n5UDo\"DZ+=79M!b#bFP!bfTx]%%ee3)hte\?hB]D(^X(x_G>o1BCN`ciZs]9xHY>M`c^o,erUt~)!!Nb~&KJw;&,Cw9$$9}ztRcY=6y8=[@U8)$%%IP/xk2fu%%F+46sDw5{G1;zz\".,riYJ%%L54`3$u!e`);>~{ff/H*<pX=@M~;H`TDI]16Kpk)Ss67>^>apoNFea7}UnGHy*0VzU[8;htNnE*ad&l`CnRhaNkN[+pR`8Z<7!<v9c.I`\?+RA*TviGTH6OnzQCcMnXEYWga4t](iCK@WAYHD;y)MfZUwCm1K}4N[B!l2p1>,CnlO0~z+BX\?%%F^oBz~(Y+LjD\?KbdS)tGgTUXS0l|%%ER~by[J5t,pG(NY(+/yB[aM\?lig<>n^q89[{TB+u60k{1_ktl*dO)pNqXLXVe@;8]*tTl4E|7#_4LkCxDSu\"KLoHL@d<F|IZ>8B)Z~e3!>|\?)FNrcqFWv]/^;D7>ON9J&rOr_Wj^n8IK1h$l&~k.1=|$D<d5=D,+!PFyS;R{3~ek\"D1]eQ`QRp`K7>JF$SxTNULhmP&@Ng1UY~O+hp!]ezW@)\",B");
#>     char buf[1131];
#> #define __doBufN__ 1131
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void rx_2caa923c2e1313664984f5d665cc9cff__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   rx_2caa923c2e1313664984f5d665cc9cff__dydt(neq, *t, A, DADT);
#> }
#> extern int rx_2caa923c2e1313664984f5d665cc9cff__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   rx_2caa923c2e1313664984f5d665cc9cff__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void rx_2caa923c2e1313664984f5d665cc9cff__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   rx_2caa923c2e1313664984f5d665cc9cff__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void rx_2caa923c2e1313664984f5d665cc9cff__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_rx_2caa923c2e1313664984f5d665cc9cff_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("rx_2caa923c2e1313664984f5d665cc9cff_","rx_2caa923c2e1313664984f5d665cc9cff__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("rx_2caa923c2e1313664984f5d665cc9cff_","rx_2caa923c2e1313664984f5d665cc9cff__assignFuns", (DL_FUNC) rx_2caa923c2e1313664984f5d665cc9cff__assignFuns);
#>   R_RegisterCCallable("rx_2caa923c2e1313664984f5d665cc9cff_","rx_2caa923c2e1313664984f5d665cc9cff__inis",(DL_FUNC) rx_2caa923c2e1313664984f5d665cc9cff__inis);
#>   R_RegisterCCallable("rx_2caa923c2e1313664984f5d665cc9cff_","rx_2caa923c2e1313664984f5d665cc9cff__dydt",(DL_FUNC) rx_2caa923c2e1313664984f5d665cc9cff__dydt);
#>   R_RegisterCCallable("rx_2caa923c2e1313664984f5d665cc9cff_","rx_2caa923c2e1313664984f5d665cc9cff__calc_lhs",(DL_FUNC) rx_2caa923c2e1313664984f5d665cc9cff__calc_lhs);
#>   R_RegisterCCallable("rx_2caa923c2e1313664984f5d665cc9cff_","rx_2caa923c2e1313664984f5d665cc9cff__calc_jac",(DL_FUNC) rx_2caa923c2e1313664984f5d665cc9cff__calc_jac);
#>   R_RegisterCCallable("rx_2caa923c2e1313664984f5d665cc9cff_","rx_2caa923c2e1313664984f5d665cc9cff__dydt_lsoda", (DL_FUNC) rx_2caa923c2e1313664984f5d665cc9cff__dydt_lsoda);
#>   R_RegisterCCallable("rx_2caa923c2e1313664984f5d665cc9cff_","rx_2caa923c2e1313664984f5d665cc9cff__calc_jac_lsoda", (DL_FUNC) rx_2caa923c2e1313664984f5d665cc9cff__calc_jac_lsoda);
#>   R_RegisterCCallable("rx_2caa923c2e1313664984f5d665cc9cff_","rx_2caa923c2e1313664984f5d665cc9cff__ode_solver_solvedata", (DL_FUNC) rx_2caa923c2e1313664984f5d665cc9cff__ode_solver_solvedata);
#>   R_RegisterCCallable("rx_2caa923c2e1313664984f5d665cc9cff_","rx_2caa923c2e1313664984f5d665cc9cff__ode_solver_get_solvedata", (DL_FUNC) rx_2caa923c2e1313664984f5d665cc9cff__ode_solver_get_solvedata);
#>   R_RegisterCCallable("rx_2caa923c2e1313664984f5d665cc9cff_","rx_2caa923c2e1313664984f5d665cc9cff__F", (DL_FUNC) rx_2caa923c2e1313664984f5d665cc9cff__F);
#>   R_RegisterCCallable("rx_2caa923c2e1313664984f5d665cc9cff_","rx_2caa923c2e1313664984f5d665cc9cff__Lag", (DL_FUNC) rx_2caa923c2e1313664984f5d665cc9cff__Lag);
#>   R_RegisterCCallable("rx_2caa923c2e1313664984f5d665cc9cff_","rx_2caa923c2e1313664984f5d665cc9cff__Rate", (DL_FUNC) rx_2caa923c2e1313664984f5d665cc9cff__Rate);
#>   R_RegisterCCallable("rx_2caa923c2e1313664984f5d665cc9cff_","rx_2caa923c2e1313664984f5d665cc9cff__Dur", (DL_FUNC) rx_2caa923c2e1313664984f5d665cc9cff__Dur);
#>   R_RegisterCCallable("rx_2caa923c2e1313664984f5d665cc9cff_","rx_2caa923c2e1313664984f5d665cc9cff__mtime", (DL_FUNC) rx_2caa923c2e1313664984f5d665cc9cff__mtime);
#>   R_RegisterCCallable("rx_2caa923c2e1313664984f5d665cc9cff_","rx_2caa923c2e1313664984f5d665cc9cff__ME", (DL_FUNC) rx_2caa923c2e1313664984f5d665cc9cff__ME);
#>   R_RegisterCCallable("rx_2caa923c2e1313664984f5d665cc9cff_","rx_2caa923c2e1313664984f5d665cc9cff__IndF", (DL_FUNC) rx_2caa923c2e1313664984f5d665cc9cff__IndF);
#>   R_RegisterCCallable("rx_2caa923c2e1313664984f5d665cc9cff_","rx_2caa923c2e1313664984f5d665cc9cff__dydt_liblsoda", (DL_FUNC) rx_2caa923c2e1313664984f5d665cc9cff__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_rx_2caa923c2e1313664984f5d665cc9cff_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_rx_2caa923c2e1313664984f5d665cc9cff_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"rx_2caa923c2e1313664984f5d665cc9cff__model_vars", (DL_FUNC) &rx_2caa923c2e1313664984f5d665cc9cff__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_rx_2caa923c2e1313664984f5d665cc9cff_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_2caa923c2e1313664984f5d665cc9cff__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("rx_2caa923c2e1313664984f5d665cc9cff__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’
ret
#> ── Solved rxode2 object ──
#> ── Parameters (value$params): ──
#> # A tibble: 1 × 0
#> ── Initial Conditions (value$inits): ──
#> named numeric(0)
#> ── First part of data (object): ──
#> # A tibble: 10 × 6
#>     time     fx  dMean    dSd    mu sigma
#>    <dbl>  <dbl>  <dbl>  <dbl> <dbl> <dbl>
#> 1 -3     -5.42  -3      8         0     1
#> 2 -2.33  -3.64  -2.33   4.44      0     1
#> 3 -1.67  -2.31  -1.67   1.78      0     1
#> 4 -1     -1.42  -1      0         0     1
#> 5 -0.333 -0.974 -0.333 -0.889     0     1
#> 6  0.333 -0.974  0.333 -0.889     0     1
#> # ℹ 4 more rows
# }