log-likelihood for the Poisson distribution

log-likelihood for the Poisson distribution

Usage

llikPois(x, lambda, full = FALSE)

Arguments

x

non negative integers

lambda

non-negative means

full

Add the data frame showing x, mean, sd as well as the fx and derivatives

Value

data frame with fx for the pdf value of with dLambda that has the derivatives with respect to the parameters at the observation time-point

Details

In an rxode2() model, you can use llikPois() but you have to use all arguments. You can also get the derivatives with llikPoisDlambda()

Author

Matthew L. Fidler

Examples

# \donttest{
llikPois(0:7, lambda = 1)
#>          fx dLambda
#> 1 -1.000000      -1
#> 2 -1.000000       0
#> 3 -1.693147       1
#> 4 -2.791759       2
#> 5 -4.178054       3
#> 6 -5.787492       4
#> 7 -7.579251       5
#> 8 -9.525161       6

llikPois(0:7, lambda = 4, full=TRUE)
#>   x lambda        fx dLambda
#> 1 0      4 -4.000000   -1.00
#> 2 1      4 -2.613706   -0.75
#> 3 2      4 -1.920558   -0.50
#> 4 3      4 -1.632876   -0.25
#> 5 4      4 -1.632876    0.00
#> 6 5      4 -1.856020    0.25
#> 7 6      4 -2.261485    0.50
#> 8 7      4 -2.821101    0.75

# In rxode2 you can use:

et <- et(0:10)
et$lambda <- 0.5

model <- function() {
  model({
    fx <- llikPois(time, lambda)
    dLambda <- llikPoisDlambda(time, lambda)
  })
}

rxSolve(model, et)
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#>  
#>  
#> C model:
#> // Define 2 LHS values
#> #define _LHS_0_ 0 // fx
#> #define _LHS_1_ 1 // dLambda
#> #define _getRxSolve_ _rx8e3c873363a323b877c88406bf62d6263861__getRxSolve__EsJu
#> #define _evalUdf _rx8e3c873363a323b877c88406bf62d6263862__evalUdf_9xah
#> #define _solveData _rx8e3c873363a323b877c88406bf62d6263863__solveData_gBOZ
#> #define _assign_ptr _rx8e3c873363a323b877c88406bf62d6263864__assign_ptr_l3zY
#> #define _rxRmModelLib _rx8e3c873363a323b877c88406bf62d6263865__rxRmModelLib_vTa9
#> #define _rxGetModelLib _rx8e3c873363a323b877c88406bf62d6263866__rxGetModelLib_7EhW
#> #define _old_c _rx8e3c873363a323b877c88406bf62d6263867__old_c_3MG9
#> #define _ptrid _rx8e3c873363a323b877c88406bf62d6263868__ptrid_V3pR
#> #define _rxIsCurrentC _rx8e3c873363a323b877c88406bf62d6263869__rxIsCurrentC_1aNk
#> #define _sumPS _rx8e3c873363a323b877c88406bf62d6263870__sumPS_G5Jw
#> #define _prodPS _rx8e3c873363a323b877c88406bf62d6263871__prodPS_D9TK
#> #define _prodType _rx8e3c873363a323b877c88406bf62d6263872__prodType_40NP
#> #define _sumType _rx8e3c873363a323b877c88406bf62d6263873__sumType_JiAL
#> #define _update_par_ptr _rx8e3c873363a323b877c88406bf62d6263874__update_par_ptr_ZBrh
#> #define _getParCov _rx8e3c873363a323b877c88406bf62d6263875__getParCov_efgh
#> #define _rxode2_rxAssignPtr _rx8e3c873363a323b877c88406bf62d6263876__rxode2_rxAssignPtr_Drm8
#> #define _rxQr _rx8e3c873363a323b877c88406bf62d6263877__rxQr_5igv
#> #define _compareFactorVal _rx8e3c873363a323b877c88406bf62d6263878__compareFactorVal_lJYg
#> #define _sum _rx8e3c873363a323b877c88406bf62d6263879__sum_GqQx
#> #define _udf _rx8e3c873363a323b877c88406bf62d6263880__udf_LoYl
#> #define _sign _rx8e3c873363a323b877c88406bf62d6263881__sign_FgJd
#> #define _prod _rx8e3c873363a323b877c88406bf62d6263882__prod_KaFg
#> #define _max _rx8e3c873363a323b877c88406bf62d6263883__max_mDR4
#> #define _min _rx8e3c873363a323b877c88406bf62d6263884__min_jxuo
#> #define _transit4P _rx8e3c873363a323b877c88406bf62d6263885__transit4P_P8gm
#> #define _transit3P _rx8e3c873363a323b877c88406bf62d6263886__transit3P_oUom
#> #define _assignFuns0 _rx8e3c873363a323b877c88406bf62d6263887__assignFuns0_wLtu
#> #define _assignFuns _rx8e3c873363a323b877c88406bf62d6263888__assignFuns_DgOz
#> #define _rxord _rx8e3c873363a323b877c88406bf62d6263889__rxord_tPZw
#> #define __assignFuns2 _rx8e3c873363a323b877c88406bf62d6263890___assignFuns2_Dqx2
#> #define _llikCauchyDscale _rx8e3c873363a323b877c88406bf62d6263891__llikCauchyDscale_t0Ff
#> #define _llikCauchyDlocation _rx8e3c873363a323b877c88406bf62d6263892__llikCauchyDlocation_CRup
#> #define _llikCauchy _rx8e3c873363a323b877c88406bf62d6263893__llikCauchy_BfcL
#> #define _llikGammaDrate _rx8e3c873363a323b877c88406bf62d6263894__llikGammaDrate_LlJk
#> #define _llikGammaDshape _rx8e3c873363a323b877c88406bf62d6263895__llikGammaDshape_0HOu
#> #define _llikGamma _rx8e3c873363a323b877c88406bf62d6263896__llikGamma_quMQ
#> #define _llikWeibullDscale _rx8e3c873363a323b877c88406bf62d6263897__llikWeibullDscale_IP4s
#> #define _llikWeibullDshape _rx8e3c873363a323b877c88406bf62d6263898__llikWeibullDshape_BKwl
#> #define _llikWeibull _rx8e3c873363a323b877c88406bf62d6263899__llikWeibull_ION1
#> #define _llikUnifDbeta _rx8e3c873363a323b877c88406bf62d6263900__llikUnifDbeta_tTX5
#> #define _llikUnifDalpha _rx8e3c873363a323b877c88406bf62d6263901__llikUnifDalpha_DulH
#> #define _llikUnif _rx8e3c873363a323b877c88406bf62d6263902__llikUnif_LxeA
#> #define _llikGeomDp _rx8e3c873363a323b877c88406bf62d6263903__llikGeomDp_Ut0p
#> #define _llikGeom _rx8e3c873363a323b877c88406bf62d6263904__llikGeom_o6Aq
#> #define _llikFDdf2 _rx8e3c873363a323b877c88406bf62d6263905__llikFDdf2_SQkn
#> #define _llikFDdf1 _rx8e3c873363a323b877c88406bf62d6263906__llikFDdf1_H8fP
#> #define _llikF _rx8e3c873363a323b877c88406bf62d6263907__llikF_CAnN
#> #define _llikExpDrate _rx8e3c873363a323b877c88406bf62d6263908__llikExpDrate_kUms
#> #define _llikExp _rx8e3c873363a323b877c88406bf62d6263909__llikExp_wiOW
#> #define _llikChisqDdf _rx8e3c873363a323b877c88406bf62d6263910__llikChisqDdf_G16H
#> #define _llikChisq _rx8e3c873363a323b877c88406bf62d6263911__llikChisq_PUgn
#> #define _llikTDsd _rx8e3c873363a323b877c88406bf62d6263912__llikTDsd_Cjuw
#> #define _llikTDmean _rx8e3c873363a323b877c88406bf62d6263913__llikTDmean_A8iw
#> #define _llikTDdf _rx8e3c873363a323b877c88406bf62d6263914__llikTDdf_bn2G
#> #define _llikT _rx8e3c873363a323b877c88406bf62d6263915__llikT_TQjS
#> #define _llikBetaDshape2 _rx8e3c873363a323b877c88406bf62d6263916__llikBetaDshape2_rer5
#> #define _llikBetaDshape1 _rx8e3c873363a323b877c88406bf62d6263917__llikBetaDshape1_zexL
#> #define _llikBeta _rx8e3c873363a323b877c88406bf62d6263918__llikBeta_gAwM
#> #define _llikNbinomMuDmu _rx8e3c873363a323b877c88406bf62d6263919__llikNbinomMuDmu_E6nq
#> #define _llikNbinomMu _rx8e3c873363a323b877c88406bf62d6263920__llikNbinomMu_R3DD
#> #define _llikNbinomDprob _rx8e3c873363a323b877c88406bf62d6263921__llikNbinomDprob_9jN5
#> #define _llikNbinom _rx8e3c873363a323b877c88406bf62d6263922__llikNbinom_O4r9
#> #define _llikBinomDprob _rx8e3c873363a323b877c88406bf62d6263923__llikBinomDprob_D97m
#> #define _llikBinom _rx8e3c873363a323b877c88406bf62d6263924__llikBinom_e5yb
#> #define _llikPoisDlambda _rx8e3c873363a323b877c88406bf62d6263925__llikPoisDlambda_V8Eu
#> #define _llikPois _rx8e3c873363a323b877c88406bf62d6263926__llikPois_Z1xT
#> #define _llikNormDsd _rx8e3c873363a323b877c88406bf62d6263927__llikNormDsd_QYOQ
#> #define _llikNormDmean _rx8e3c873363a323b877c88406bf62d6263928__llikNormDmean_e6aq
#> #define _llikNorm _rx8e3c873363a323b877c88406bf62d6263929__llikNorm_t3Lb
#> #define simeps _rx8e3c873363a323b877c88406bf62d6263930_simeps_sbea
#> #define simeta _rx8e3c873363a323b877c88406bf62d6263931_simeta_w5m4
#> #define expit _rx8e3c873363a323b877c88406bf62d6263932_expit_brD9
#> #define logit _rx8e3c873363a323b877c88406bf62d6263933_logit_EZAJ
#> #define gammapDer _rx8e3c873363a323b877c88406bf62d6263934_gammapDer_Ne7I
#> #define lowergamma _rx8e3c873363a323b877c88406bf62d6263935_lowergamma_xUFA
#> #define uppergamma _rx8e3c873363a323b877c88406bf62d6263936_uppergamma_5R2w
#> #define gammaqInva _rx8e3c873363a323b877c88406bf62d6263937_gammaqInva_ZwwB
#> #define gammaqInv _rx8e3c873363a323b877c88406bf62d6263938_gammaqInv_gRvb
#> #define gammapInva _rx8e3c873363a323b877c88406bf62d6263939_gammapInva_lnk3
#> #define gammapInv _rx8e3c873363a323b877c88406bf62d6263940_gammapInv_57cf
#> #define gammaq _rx8e3c873363a323b877c88406bf62d6263941_gammaq_UwMa
#> #define gammap _rx8e3c873363a323b877c88406bf62d6263942_gammap_rEA9
#> #define phi _rx8e3c873363a323b877c88406bf62d6263943_phi_z1SE
#> #define d2ELUa _rx8e3c873363a323b877c88406bf62d6263944_d2ELUa_1KzA
#> #define dELUa _rx8e3c873363a323b877c88406bf62d6263945_dELUa_f2H4
#> #define d2aELU _rx8e3c873363a323b877c88406bf62d6263946_d2aELU_Q2oF
#> #define d2ELU _rx8e3c873363a323b877c88406bf62d6263947_d2ELU_0joC
#> #define dELU _rx8e3c873363a323b877c88406bf62d6263948_dELU_Gj0a
#> #define ELU _rx8e3c873363a323b877c88406bf62d6263949_ELU_8wAb
#> #define dPReLUa1 _rx8e3c873363a323b877c88406bf62d6263950_dPReLUa1_QELO
#> #define dPReLUa _rx8e3c873363a323b877c88406bf62d6263951_dPReLUa_lObq
#> #define dPReLU _rx8e3c873363a323b877c88406bf62d6263952_dPReLU_Nath
#> #define PReLU _rx8e3c873363a323b877c88406bf62d6263953_PReLU_ltiW
#> #define dSwish _rx8e3c873363a323b877c88406bf62d6263954_dSwish_R2IG
#> #define Swish _rx8e3c873363a323b877c88406bf62d6263955_Swish_VLQY
#> #define dlReLU _rx8e3c873363a323b877c88406bf62d6263956_dlReLU_M84J
#> #define lReLU _rx8e3c873363a323b877c88406bf62d6263957_lReLU_Cnuy
#> #define dSELU _rx8e3c873363a323b877c88406bf62d6263958_dSELU_CXPJ
#> #define SELU _rx8e3c873363a323b877c88406bf62d6263959_SELU_UjvF
#> #define d4softplus _rx8e3c873363a323b877c88406bf62d6263960_d4softplus_k7KJ
#> #define d3softplus _rx8e3c873363a323b877c88406bf62d6263961_d3softplus_vDxq
#> #define d2softplus _rx8e3c873363a323b877c88406bf62d6263962_d2softplus_bVSK
#> #define dsoftplus _rx8e3c873363a323b877c88406bf62d6263963_dsoftplus_Vfr7
#> #define softplus _rx8e3c873363a323b877c88406bf62d6263964_softplus_TcaU
#> #define d4GELU _rx8e3c873363a323b877c88406bf62d6263965_d4GELU_EgpY
#> #define d3GELU _rx8e3c873363a323b877c88406bf62d6263966_d3GELU_kNIq
#> #define d2GELU _rx8e3c873363a323b877c88406bf62d6263967_d2GELU_nEnm
#> #define dGELU _rx8e3c873363a323b877c88406bf62d6263968_dGELU_hSbo
#> #define GELU _rx8e3c873363a323b877c88406bf62d6263969_GELU_MZYB
#> #define dReLU _rx8e3c873363a323b877c88406bf62d6263970_dReLU_jkiC
#> #define ReLU _rx8e3c873363a323b877c88406bf62d6263971_ReLU_oUqv
#> #define riweibull _rx8e3c873363a323b877c88406bf62d6263972_riweibull_niI7
#> #define riunif _rx8e3c873363a323b877c88406bf62d6263973_riunif_5Ui0
#> #define rit_ _rx8e3c873363a323b877c88406bf62d6263974_rit__GVTs
#> #define ripois _rx8e3c873363a323b877c88406bf62d6263975_ripois_34TN
#> #define ribeta _rx8e3c873363a323b877c88406bf62d6263976_ribeta_HeFp
#> #define rigamma _rx8e3c873363a323b877c88406bf62d6263977_rigamma_ue1Z
#> #define rigeom _rx8e3c873363a323b877c88406bf62d6263978_rigeom_aEnJ
#> #define rif _rx8e3c873363a323b877c88406bf62d6263979_rif_R6UH
#> #define riexp _rx8e3c873363a323b877c88406bf62d6263980_riexp_ZKbS
#> #define richisq _rx8e3c873363a323b877c88406bf62d6263981_richisq_smNY
#> #define ricauchy _rx8e3c873363a323b877c88406bf62d6263982_ricauchy_lS2w
#> #define rinbinomMu _rx8e3c873363a323b877c88406bf62d6263983_rinbinomMu_zYqw
#> #define rinbinom _rx8e3c873363a323b877c88406bf62d6263984_rinbinom_2z8b
#> #define ribinom _rx8e3c873363a323b877c88406bf62d6263985_ribinom_GMdj
#> #define rinorm _rx8e3c873363a323b877c88406bf62d6263986_rinorm_LB3L
#> #define rxweibull _rx8e3c873363a323b877c88406bf62d6263987_rxweibull_sOUR
#> #define rxunif _rx8e3c873363a323b877c88406bf62d6263988_rxunif_bPYU
#> #define rxt_ _rx8e3c873363a323b877c88406bf62d6263989_rxt__PmW8
#> #define rxpois _rx8e3c873363a323b877c88406bf62d6263990_rxpois_sOmG
#> #define rxbeta _rx8e3c873363a323b877c88406bf62d6263991_rxbeta_ypO5
#> #define rxgamma _rx8e3c873363a323b877c88406bf62d6263992_rxgamma_5OnO
#> #define rxgeom _rx8e3c873363a323b877c88406bf62d6263993_rxgeom_E9TF
#> #define rxf _rx8e3c873363a323b877c88406bf62d6263994_rxf_yOK7
#> #define rxexp _rx8e3c873363a323b877c88406bf62d6263995_rxexp_OqEK
#> #define rxchisq _rx8e3c873363a323b877c88406bf62d6263996_rxchisq_wH4T
#> #define rxcauchy _rx8e3c873363a323b877c88406bf62d6263997_rxcauchy_tSJj
#> #define rxnbinomMu _rx8e3c873363a323b877c88406bf62d6263998_rxnbinomMu_0v8j
#> #define rxnbinom _rx8e3c873363a323b877c88406bf62d6263999_rxnbinom_xwnn
#> #define rxbinom _rx8e3c873363a323b877c88406bf62d6264000_rxbinom_6QBD
#> #define rxnorm _rx8e3c873363a323b877c88406bf62d6264001_rxnorm_Uzzs
#> #define linCmtB _rx8e3c873363a323b877c88406bf62d6264002_linCmtB_oz5n
#> #define linCmtA _rx8e3c873363a323b877c88406bf62d6264003_linCmtA_36W2
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){  if (_solveData->svar[_svari] == 0) {lambda = _PP[0];}; }
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){  if (_solveData->ovar[_ovari] == 0) {lambda = _PP[0];}; }
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  rx_6cb39a2d79ab9aa20622e79fc5d5998a__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *rx_6cb39a2d79ab9aa20622e79fc5d5998a__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP rx_6cb39a2d79ab9aa20622e79fc5d5998a__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void rx_6cb39a2d79ab9aa20622e79fc5d5998a__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double lambda = NA_REAL;
#>   double fx = NA_REAL;
#>   double dLambda = NA_REAL;
#> 
#>   (void)t;
#>   (void)lambda;
#>   (void)fx;
#>   (void)dLambda;
#> 
#>   fx = _PL[0];
#>   dLambda = _PL[1];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#>   lambda = _PP[0];
#> 
#> 
#>   fx =llikPois(t,lambda);
#>   dLambda =llikPoisDlambda(t,lambda);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void rx_6cb39a2d79ab9aa20622e79fc5d5998a__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void rx_6cb39a2d79ab9aa20622e79fc5d5998a__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void rx_6cb39a2d79ab9aa20622e79fc5d5998a__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double lambda = NA_REAL;
#>   double fx = NA_REAL;
#>   double dLambda = NA_REAL;
#> 
#>   (void)t;
#>   (void)lambda;
#>   (void)fx;
#>   (void)dLambda;
#> 
#>   fx = _PL[0];
#>   dLambda = _PL[1];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#>   lambda = _PP[0];
#> 
#> 
#>   fx =llikPois(t,lambda);
#>   dLambda =llikPoisDlambda(t,lambda);
#> 
#>   _lhs[_LHS_0_]=fx;
#>   _lhs[_LHS_1_]=dLambda;
#> }
#> // Functional based bioavailability
#> double rx_6cb39a2d79ab9aa20622e79fc5d5998a__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double rx_6cb39a2d79ab9aa20622e79fc5d5998a__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double rx_6cb39a2d79ab9aa20622e79fc5d5998a__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double rx_6cb39a2d79ab9aa20622e79fc5d5998a__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void rx_6cb39a2d79ab9aa20622e79fc5d5998a__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void rx_6cb39a2d79ab9aa20622e79fc5d5998a__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void rx_6cb39a2d79ab9aa20622e79fc5d5998a__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP rx_6cb39a2d79ab9aa20622e79fc5d5998a__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_6cb39a2d79ab9aa20622e79fc5d5998a__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAA@Q2AAAv7#aT)>oEkNA(;Ua=WjmlC~15w`Qv<9*u/kY6u~dT*Jnv<6sajVG~mQ^@6Wt7UX<7l,(eCa3>IO9\?kwyi&N4pjueY,[95GJt*A>C4}KdASFS`EDzer@+,P(sImE`TZd#CHo}U7A.c.w]GKv\?_~!6#ge94bF.&}R./SZmy3@]KYb~^kEp_iGMBqr>Q2Ys0vGWmRTw|5;3R_|6\"4z!o<u`dfdiU8$e|:a#F4{RHs#ld&Qs,cyu{>)G(9FXRWi\?y3FX!G,h4T,C+E/w~k.ree!hSgYaEm}5B8R&gP[ui*c]08:6!$w7kW&r5O^f)!r:VgRHvBnr,&uX+S4j]w/xiKnL\"\"s<ISy^}E1;XG<Tx/3bRg9wH<PE&9x(^XT\?7sd29x9/(T+y)<7ZaGp}#K)P6\">R=M>MF_17A4P6h>3|oq&!ehIGRlMwJDLQ66zKx;n+3LO!HgQn)i428.}7D(#{Yh~&5b6K1hzFw3jp\?}$O[2@9C{h/zC}8QSw%%6&qjy`{pFW{%%R<~bY$|R:zM\?IS~oMwjrUpLDi_Cs[aGVQ|q:Qj.J=Ql/$jzBkHZ*]CS@5F[J,/;C&tpV7,gsfr}Uh;M|\"(pKRdY%%St_69Oe<$HC>sj*%%QZ6m=3\?6U]fu%%>I*AhOwr}GP/<DkjDtUH*C|T0(Np]$%%%%zGk/zen%%=I2m31tYdND~!+Yy<`Lm[D_+kRqewB&z_nGsO7suzsN>Ytm1@S&:gK;6%%J8~t$zCiumBPOn^!Wi=ZOHO,dJleGJ6R=1E0%%59ZLF74>c:!5GXLoG+X|v/Fxwmu)M=MkGF<PV/1Y$4/)SAA!&(<b<1~96@XvDd]Z4saxY_C=_,MB,D<(feH8UG,cdi|+5T%%*pj6/l.<gW,H8LQ/VX!DqSr0u3F|eXE7+wcWSc)r$v&h3||,pj%%H9Fgf8rX!MMx`|kP84:R=FC\"=+&#W(_GbH%%%%r/gMi71{\";pd^s<\?r=/G^6H[IM[Dw1=K~Wg+TI4tUt5lD2T,Q+&/ng]_vhL~(aEGK>0Kt.i|I,Ia\"K]n@/iqd*xHK<L+BFsZ4v%%$Y!R#Qi@&nIMVQ~,bX+VEd\"JXLp");
#>     char buf[1115];
#> #define __doBufN__ 1115
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void rx_6cb39a2d79ab9aa20622e79fc5d5998a__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   rx_6cb39a2d79ab9aa20622e79fc5d5998a__dydt(neq, *t, A, DADT);
#> }
#> extern int rx_6cb39a2d79ab9aa20622e79fc5d5998a__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   rx_6cb39a2d79ab9aa20622e79fc5d5998a__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void rx_6cb39a2d79ab9aa20622e79fc5d5998a__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   rx_6cb39a2d79ab9aa20622e79fc5d5998a__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void rx_6cb39a2d79ab9aa20622e79fc5d5998a__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_rx_6cb39a2d79ab9aa20622e79fc5d5998a_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("rx_6cb39a2d79ab9aa20622e79fc5d5998a_","rx_6cb39a2d79ab9aa20622e79fc5d5998a__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("rx_6cb39a2d79ab9aa20622e79fc5d5998a_","rx_6cb39a2d79ab9aa20622e79fc5d5998a__assignFuns", (DL_FUNC) rx_6cb39a2d79ab9aa20622e79fc5d5998a__assignFuns);
#>   R_RegisterCCallable("rx_6cb39a2d79ab9aa20622e79fc5d5998a_","rx_6cb39a2d79ab9aa20622e79fc5d5998a__inis",(DL_FUNC) rx_6cb39a2d79ab9aa20622e79fc5d5998a__inis);
#>   R_RegisterCCallable("rx_6cb39a2d79ab9aa20622e79fc5d5998a_","rx_6cb39a2d79ab9aa20622e79fc5d5998a__dydt",(DL_FUNC) rx_6cb39a2d79ab9aa20622e79fc5d5998a__dydt);
#>   R_RegisterCCallable("rx_6cb39a2d79ab9aa20622e79fc5d5998a_","rx_6cb39a2d79ab9aa20622e79fc5d5998a__calc_lhs",(DL_FUNC) rx_6cb39a2d79ab9aa20622e79fc5d5998a__calc_lhs);
#>   R_RegisterCCallable("rx_6cb39a2d79ab9aa20622e79fc5d5998a_","rx_6cb39a2d79ab9aa20622e79fc5d5998a__calc_jac",(DL_FUNC) rx_6cb39a2d79ab9aa20622e79fc5d5998a__calc_jac);
#>   R_RegisterCCallable("rx_6cb39a2d79ab9aa20622e79fc5d5998a_","rx_6cb39a2d79ab9aa20622e79fc5d5998a__dydt_lsoda", (DL_FUNC) rx_6cb39a2d79ab9aa20622e79fc5d5998a__dydt_lsoda);
#>   R_RegisterCCallable("rx_6cb39a2d79ab9aa20622e79fc5d5998a_","rx_6cb39a2d79ab9aa20622e79fc5d5998a__calc_jac_lsoda", (DL_FUNC) rx_6cb39a2d79ab9aa20622e79fc5d5998a__calc_jac_lsoda);
#>   R_RegisterCCallable("rx_6cb39a2d79ab9aa20622e79fc5d5998a_","rx_6cb39a2d79ab9aa20622e79fc5d5998a__ode_solver_solvedata", (DL_FUNC) rx_6cb39a2d79ab9aa20622e79fc5d5998a__ode_solver_solvedata);
#>   R_RegisterCCallable("rx_6cb39a2d79ab9aa20622e79fc5d5998a_","rx_6cb39a2d79ab9aa20622e79fc5d5998a__ode_solver_get_solvedata", (DL_FUNC) rx_6cb39a2d79ab9aa20622e79fc5d5998a__ode_solver_get_solvedata);
#>   R_RegisterCCallable("rx_6cb39a2d79ab9aa20622e79fc5d5998a_","rx_6cb39a2d79ab9aa20622e79fc5d5998a__F", (DL_FUNC) rx_6cb39a2d79ab9aa20622e79fc5d5998a__F);
#>   R_RegisterCCallable("rx_6cb39a2d79ab9aa20622e79fc5d5998a_","rx_6cb39a2d79ab9aa20622e79fc5d5998a__Lag", (DL_FUNC) rx_6cb39a2d79ab9aa20622e79fc5d5998a__Lag);
#>   R_RegisterCCallable("rx_6cb39a2d79ab9aa20622e79fc5d5998a_","rx_6cb39a2d79ab9aa20622e79fc5d5998a__Rate", (DL_FUNC) rx_6cb39a2d79ab9aa20622e79fc5d5998a__Rate);
#>   R_RegisterCCallable("rx_6cb39a2d79ab9aa20622e79fc5d5998a_","rx_6cb39a2d79ab9aa20622e79fc5d5998a__Dur", (DL_FUNC) rx_6cb39a2d79ab9aa20622e79fc5d5998a__Dur);
#>   R_RegisterCCallable("rx_6cb39a2d79ab9aa20622e79fc5d5998a_","rx_6cb39a2d79ab9aa20622e79fc5d5998a__mtime", (DL_FUNC) rx_6cb39a2d79ab9aa20622e79fc5d5998a__mtime);
#>   R_RegisterCCallable("rx_6cb39a2d79ab9aa20622e79fc5d5998a_","rx_6cb39a2d79ab9aa20622e79fc5d5998a__ME", (DL_FUNC) rx_6cb39a2d79ab9aa20622e79fc5d5998a__ME);
#>   R_RegisterCCallable("rx_6cb39a2d79ab9aa20622e79fc5d5998a_","rx_6cb39a2d79ab9aa20622e79fc5d5998a__IndF", (DL_FUNC) rx_6cb39a2d79ab9aa20622e79fc5d5998a__IndF);
#>   R_RegisterCCallable("rx_6cb39a2d79ab9aa20622e79fc5d5998a_","rx_6cb39a2d79ab9aa20622e79fc5d5998a__dydt_liblsoda", (DL_FUNC) rx_6cb39a2d79ab9aa20622e79fc5d5998a__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_rx_6cb39a2d79ab9aa20622e79fc5d5998a_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_rx_6cb39a2d79ab9aa20622e79fc5d5998a_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"rx_6cb39a2d79ab9aa20622e79fc5d5998a__model_vars", (DL_FUNC) &rx_6cb39a2d79ab9aa20622e79fc5d5998a__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_rx_6cb39a2d79ab9aa20622e79fc5d5998a_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_6cb39a2d79ab9aa20622e79fc5d5998a__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("rx_6cb39a2d79ab9aa20622e79fc5d5998a__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’
#> ── Solved rxode2 object ──
#> ── Parameters (value$params): ──
#> # A tibble: 1 × 0
#> ── Initial Conditions (value$inits): ──
#> named numeric(0)
#> ── First part of data (object): ──
#> # A tibble: 11 × 4
#>    time    fx dLambda lambda
#>   <dbl> <dbl>   <dbl>  <dbl>
#> 1     0 -0.5       -1    0.5
#> 2     1 -1.19       1    0.5
#> 3     2 -2.58       3    0.5
#> 4     3 -4.37       5    0.5
#> 5     4 -6.45       7    0.5
#> 6     5 -8.75       9    0.5
#> # ℹ 5 more rows
# }