Return the C file associated with the rxode2 object

This will return C code for generating the rxode2 DLL.

Usage

rxC(obj)

Arguments

obj

A rxode2 family of objects or a character string of the model specification or location of a file with a model specification.

Value

a path of the library

Author

Matthew L.Fidler