Simulate beta variable from threefry generator

Care should be taken with this method not to encounter the birthday problem, described https://www.johndcook.com/blog/2016/01/29/random-number-generator-seed-mistakes/. Since the sitmo threefry, this currently generates one random deviate from the uniform distribution to seed the engine threefry and then run the code.

Usage

rxbeta(shape1, shape2, n = 1L, ncores = 1L)

Arguments

shape1, shape2

non-negative parameters of the Beta distribution.

n

number of observations. If length(n) > 1, the length is taken to be the number required.

ncores

Number of cores for the simulation

rxnorm simulates using the threefry sitmo generator.

rxnormV used to simulate with the vandercorput simulator, but since it didn't satisfy the normal properties it was changed to simple be an alias of rxnorm. It is no longer supported in rxode2({}) blocks

Value

beta random deviates

Details

Therefore, a simple call to the random number generated followed by a second call to random number generated may have identical seeds. As the number of random number generator calls are increased the probability that the birthday problem will increase.

The key to avoid this problem is to either run all simulations in the rxode2 environment once (therefore one seed or series of seeds for the whole simulation), pre-generate all random variables used for the simulation, or seed the rxode2 engine with rxSetSeed()

Internally each ID is seeded with a unique number so that the results do not depend on the number of cores used.

Examples

# \donttest{

## Use threefry engine

rxbeta(0.5, 0.5, n = 10) # with rxbeta you have to explicitly state n
#>  [1] 0.20010253 0.57312552 0.37768796 0.86754361 0.85391446 0.99893102
#>  [7] 0.99309318 0.02719286 0.00314745 0.42356861
rxbeta(5, 1, n = 10, ncores = 2) # You can parallelize the simulation using openMP
#>  [1] 0.9998949 0.9416763 0.8655410 0.8965935 0.9685887 0.9084803 0.8582096
#>  [8] 0.7147322 0.9299513 0.9665686

rxbeta(1, 3)
#> [1] 0.1473567

## This example uses `rxbeta` directly in the model

rx <- function() {
  model({
    a <- rxbeta(2, 2)
  })
}

et <- et(1, id = 1:2)

s <- rxSolve(rx, et)
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#>  
#>  
#> C model:
#> // Define 1 LHS values
#> #define _LHS_0_ 0 // a
#> #define _getRxSolve_ _rxf0af58525eeda19624ba3ff7f41489125434__getRxSolve__aMe0
#> #define _evalUdf _rxf0af58525eeda19624ba3ff7f41489125435__evalUdf_0W5E
#> #define _solveData _rxf0af58525eeda19624ba3ff7f41489125436__solveData_vb1N
#> #define _assign_ptr _rxf0af58525eeda19624ba3ff7f41489125437__assign_ptr_sbx6
#> #define _rxRmModelLib _rxf0af58525eeda19624ba3ff7f41489125438__rxRmModelLib_oc8H
#> #define _rxGetModelLib _rxf0af58525eeda19624ba3ff7f41489125439__rxGetModelLib_DHkE
#> #define _old_c _rxf0af58525eeda19624ba3ff7f41489125440__old_c_xm1J
#> #define _ptrid _rxf0af58525eeda19624ba3ff7f41489125441__ptrid_d4ay
#> #define _rxIsCurrentC _rxf0af58525eeda19624ba3ff7f41489125442__rxIsCurrentC_leW9
#> #define _sumPS _rxf0af58525eeda19624ba3ff7f41489125443__sumPS_goD3
#> #define _prodPS _rxf0af58525eeda19624ba3ff7f41489125444__prodPS_R39J
#> #define _prodType _rxf0af58525eeda19624ba3ff7f41489125445__prodType_vcr1
#> #define _sumType _rxf0af58525eeda19624ba3ff7f41489125446__sumType_UmLF
#> #define _update_par_ptr _rxf0af58525eeda19624ba3ff7f41489125447__update_par_ptr_GL1z
#> #define _getParCov _rxf0af58525eeda19624ba3ff7f41489125448__getParCov_NFD8
#> #define _rxode2_rxAssignPtr _rxf0af58525eeda19624ba3ff7f41489125449__rxode2_rxAssignPtr_00Mr
#> #define _rxQr _rxf0af58525eeda19624ba3ff7f41489125450__rxQr_OclB
#> #define _compareFactorVal _rxf0af58525eeda19624ba3ff7f41489125451__compareFactorVal_dvcH
#> #define _sum _rxf0af58525eeda19624ba3ff7f41489125452__sum_YHlO
#> #define _udf _rxf0af58525eeda19624ba3ff7f41489125453__udf_Rzkg
#> #define _sign _rxf0af58525eeda19624ba3ff7f41489125454__sign_9phZ
#> #define _prod _rxf0af58525eeda19624ba3ff7f41489125455__prod_jbtv
#> #define _max _rxf0af58525eeda19624ba3ff7f41489125456__max_sFQk
#> #define _min _rxf0af58525eeda19624ba3ff7f41489125457__min_pVHF
#> #define _transit4P _rxf0af58525eeda19624ba3ff7f41489125458__transit4P_4tE5
#> #define _transit3P _rxf0af58525eeda19624ba3ff7f41489125459__transit3P_bqrW
#> #define _assignFuns0 _rxf0af58525eeda19624ba3ff7f41489125460__assignFuns0_j2jI
#> #define _assignFuns _rxf0af58525eeda19624ba3ff7f41489125461__assignFuns_RW0c
#> #define _rxord _rxf0af58525eeda19624ba3ff7f41489125462__rxord_rrgc
#> #define __assignFuns2 _rxf0af58525eeda19624ba3ff7f41489125463___assignFuns2_GmEs
#> #define _llikCauchyDscale _rxf0af58525eeda19624ba3ff7f41489125464__llikCauchyDscale_Ju9x
#> #define _llikCauchyDlocation _rxf0af58525eeda19624ba3ff7f41489125465__llikCauchyDlocation_ahgB
#> #define _llikCauchy _rxf0af58525eeda19624ba3ff7f41489125466__llikCauchy_OZl3
#> #define _llikGammaDrate _rxf0af58525eeda19624ba3ff7f41489125467__llikGammaDrate_zYvT
#> #define _llikGammaDshape _rxf0af58525eeda19624ba3ff7f41489125468__llikGammaDshape_a97E
#> #define _llikGamma _rxf0af58525eeda19624ba3ff7f41489125469__llikGamma_hW3M
#> #define _llikWeibullDscale _rxf0af58525eeda19624ba3ff7f41489125470__llikWeibullDscale_G9V3
#> #define _llikWeibullDshape _rxf0af58525eeda19624ba3ff7f41489125471__llikWeibullDshape_pR1a
#> #define _llikWeibull _rxf0af58525eeda19624ba3ff7f41489125472__llikWeibull_NkG5
#> #define _llikUnifDbeta _rxf0af58525eeda19624ba3ff7f41489125473__llikUnifDbeta_JwD9
#> #define _llikUnifDalpha _rxf0af58525eeda19624ba3ff7f41489125474__llikUnifDalpha_TK40
#> #define _llikUnif _rxf0af58525eeda19624ba3ff7f41489125475__llikUnif_NPJi
#> #define _llikGeomDp _rxf0af58525eeda19624ba3ff7f41489125476__llikGeomDp_ALZB
#> #define _llikGeom _rxf0af58525eeda19624ba3ff7f41489125477__llikGeom_rhef
#> #define _llikFDdf2 _rxf0af58525eeda19624ba3ff7f41489125478__llikFDdf2_ghDr
#> #define _llikFDdf1 _rxf0af58525eeda19624ba3ff7f41489125479__llikFDdf1_m85i
#> #define _llikF _rxf0af58525eeda19624ba3ff7f41489125480__llikF_gvlJ
#> #define _llikExpDrate _rxf0af58525eeda19624ba3ff7f41489125481__llikExpDrate_YgGq
#> #define _llikExp _rxf0af58525eeda19624ba3ff7f41489125482__llikExp_QxLo
#> #define _llikChisqDdf _rxf0af58525eeda19624ba3ff7f41489125483__llikChisqDdf_YlFg
#> #define _llikChisq _rxf0af58525eeda19624ba3ff7f41489125484__llikChisq_JdKa
#> #define _llikTDsd _rxf0af58525eeda19624ba3ff7f41489125485__llikTDsd_FgmD
#> #define _llikTDmean _rxf0af58525eeda19624ba3ff7f41489125486__llikTDmean_R4jx
#> #define _llikTDdf _rxf0af58525eeda19624ba3ff7f41489125487__llikTDdf_uoP8
#> #define _llikT _rxf0af58525eeda19624ba3ff7f41489125488__llikT_gmoU
#> #define _llikBetaDshape2 _rxf0af58525eeda19624ba3ff7f41489125489__llikBetaDshape2_omwL
#> #define _llikBetaDshape1 _rxf0af58525eeda19624ba3ff7f41489125490__llikBetaDshape1_tuDg
#> #define _llikBeta _rxf0af58525eeda19624ba3ff7f41489125491__llikBeta_OztP
#> #define _llikNbinomMuDmu _rxf0af58525eeda19624ba3ff7f41489125492__llikNbinomMuDmu_ZwDq
#> #define _llikNbinomMu _rxf0af58525eeda19624ba3ff7f41489125493__llikNbinomMu_x2t0
#> #define _llikNbinomDprob _rxf0af58525eeda19624ba3ff7f41489125494__llikNbinomDprob_FfCR
#> #define _llikNbinom _rxf0af58525eeda19624ba3ff7f41489125495__llikNbinom_upBf
#> #define _llikBinomDprob _rxf0af58525eeda19624ba3ff7f41489125496__llikBinomDprob_cLLl
#> #define _llikBinom _rxf0af58525eeda19624ba3ff7f41489125497__llikBinom_Jk0H
#> #define _llikPoisDlambda _rxf0af58525eeda19624ba3ff7f41489125498__llikPoisDlambda_Ouqu
#> #define _llikPois _rxf0af58525eeda19624ba3ff7f41489125499__llikPois_MQIP
#> #define _llikNormDsd _rxf0af58525eeda19624ba3ff7f41489125500__llikNormDsd_4sBK
#> #define _llikNormDmean _rxf0af58525eeda19624ba3ff7f41489125501__llikNormDmean_wlIO
#> #define _llikNorm _rxf0af58525eeda19624ba3ff7f41489125502__llikNorm_N1tT
#> #define simeps _rxf0af58525eeda19624ba3ff7f41489125503_simeps_X5Du
#> #define simeta _rxf0af58525eeda19624ba3ff7f41489125504_simeta_lHLx
#> #define expit _rxf0af58525eeda19624ba3ff7f41489125505_expit_eAUt
#> #define logit _rxf0af58525eeda19624ba3ff7f41489125506_logit_0po6
#> #define gammapDer _rxf0af58525eeda19624ba3ff7f41489125507_gammapDer_AqSQ
#> #define lowergamma _rxf0af58525eeda19624ba3ff7f41489125508_lowergamma_knH8
#> #define uppergamma _rxf0af58525eeda19624ba3ff7f41489125509_uppergamma_fPCA
#> #define gammaqInva _rxf0af58525eeda19624ba3ff7f41489125510_gammaqInva_nNkU
#> #define gammaqInv _rxf0af58525eeda19624ba3ff7f41489125511_gammaqInv_mswi
#> #define gammapInva _rxf0af58525eeda19624ba3ff7f41489125512_gammapInva_OWG1
#> #define gammapInv _rxf0af58525eeda19624ba3ff7f41489125513_gammapInv_6HPU
#> #define gammaq _rxf0af58525eeda19624ba3ff7f41489125514_gammaq_gnCj
#> #define gammap _rxf0af58525eeda19624ba3ff7f41489125515_gammap_uwA8
#> #define phi _rxf0af58525eeda19624ba3ff7f41489125516_phi_iwbn
#> #define d2ELUa _rxf0af58525eeda19624ba3ff7f41489125517_d2ELUa_2GTQ
#> #define dELUa _rxf0af58525eeda19624ba3ff7f41489125518_dELUa_jSre
#> #define d2aELU _rxf0af58525eeda19624ba3ff7f41489125519_d2aELU_r5ze
#> #define d2ELU _rxf0af58525eeda19624ba3ff7f41489125520_d2ELU_xLgA
#> #define dELU _rxf0af58525eeda19624ba3ff7f41489125521_dELU_wME6
#> #define ELU _rxf0af58525eeda19624ba3ff7f41489125522_ELU_nqR3
#> #define dPReLUa1 _rxf0af58525eeda19624ba3ff7f41489125523_dPReLUa1_DD9j
#> #define dPReLUa _rxf0af58525eeda19624ba3ff7f41489125524_dPReLUa_N5O4
#> #define dPReLU _rxf0af58525eeda19624ba3ff7f41489125525_dPReLU_r9D9
#> #define PReLU _rxf0af58525eeda19624ba3ff7f41489125526_PReLU_7me5
#> #define dSwish _rxf0af58525eeda19624ba3ff7f41489125527_dSwish_ybV8
#> #define Swish _rxf0af58525eeda19624ba3ff7f41489125528_Swish_EuZ1
#> #define dlReLU _rxf0af58525eeda19624ba3ff7f41489125529_dlReLU_xTQY
#> #define lReLU _rxf0af58525eeda19624ba3ff7f41489125530_lReLU_OQe6
#> #define dSELU _rxf0af58525eeda19624ba3ff7f41489125531_dSELU_aqt3
#> #define SELU _rxf0af58525eeda19624ba3ff7f41489125532_SELU_Lbsb
#> #define d4softplus _rxf0af58525eeda19624ba3ff7f41489125533_d4softplus_eaw5
#> #define d3softplus _rxf0af58525eeda19624ba3ff7f41489125534_d3softplus_m4br
#> #define d2softplus _rxf0af58525eeda19624ba3ff7f41489125535_d2softplus_D9EZ
#> #define dsoftplus _rxf0af58525eeda19624ba3ff7f41489125536_dsoftplus_AJNe
#> #define softplus _rxf0af58525eeda19624ba3ff7f41489125537_softplus_7IxU
#> #define d4GELU _rxf0af58525eeda19624ba3ff7f41489125538_d4GELU_FA5R
#> #define d3GELU _rxf0af58525eeda19624ba3ff7f41489125539_d3GELU_2wZw
#> #define d2GELU _rxf0af58525eeda19624ba3ff7f41489125540_d2GELU_wBgR
#> #define dGELU _rxf0af58525eeda19624ba3ff7f41489125541_dGELU_vbln
#> #define GELU _rxf0af58525eeda19624ba3ff7f41489125542_GELU_k357
#> #define dReLU _rxf0af58525eeda19624ba3ff7f41489125543_dReLU_cfUw
#> #define ReLU _rxf0af58525eeda19624ba3ff7f41489125544_ReLU_MarE
#> #define riweibull _rxf0af58525eeda19624ba3ff7f41489125545_riweibull_A9z1
#> #define riunif _rxf0af58525eeda19624ba3ff7f41489125546_riunif_SE1K
#> #define rit_ _rxf0af58525eeda19624ba3ff7f41489125547_rit__zAf2
#> #define ripois _rxf0af58525eeda19624ba3ff7f41489125548_ripois_H4Q2
#> #define ribeta _rxf0af58525eeda19624ba3ff7f41489125549_ribeta_oF0j
#> #define rigamma _rxf0af58525eeda19624ba3ff7f41489125550_rigamma_oCGj
#> #define rigeom _rxf0af58525eeda19624ba3ff7f41489125551_rigeom_0a8w
#> #define rif _rxf0af58525eeda19624ba3ff7f41489125552_rif_AbQE
#> #define riexp _rxf0af58525eeda19624ba3ff7f41489125553_riexp_LOlO
#> #define richisq _rxf0af58525eeda19624ba3ff7f41489125554_richisq_bqNa
#> #define ricauchy _rxf0af58525eeda19624ba3ff7f41489125555_ricauchy_thlt
#> #define rinbinomMu _rxf0af58525eeda19624ba3ff7f41489125556_rinbinomMu_iWR2
#> #define rinbinom _rxf0af58525eeda19624ba3ff7f41489125557_rinbinom_IGVL
#> #define ribinom _rxf0af58525eeda19624ba3ff7f41489125558_ribinom_QYM8
#> #define rinorm _rxf0af58525eeda19624ba3ff7f41489125559_rinorm_4JCn
#> #define rxweibull _rxf0af58525eeda19624ba3ff7f41489125560_rxweibull_uyCX
#> #define rxunif _rxf0af58525eeda19624ba3ff7f41489125561_rxunif_PJUj
#> #define rxt_ _rxf0af58525eeda19624ba3ff7f41489125562_rxt__vFk7
#> #define rxpois _rxf0af58525eeda19624ba3ff7f41489125563_rxpois_KJvD
#> #define rxbeta _rxf0af58525eeda19624ba3ff7f41489125564_rxbeta_xqbV
#> #define rxgamma _rxf0af58525eeda19624ba3ff7f41489125565_rxgamma_SKVf
#> #define rxgeom _rxf0af58525eeda19624ba3ff7f41489125566_rxgeom_r7Tc
#> #define rxf _rxf0af58525eeda19624ba3ff7f41489125567_rxf_aUEg
#> #define rxexp _rxf0af58525eeda19624ba3ff7f41489125568_rxexp_pYkN
#> #define rxchisq _rxf0af58525eeda19624ba3ff7f41489125569_rxchisq_IqnE
#> #define rxcauchy _rxf0af58525eeda19624ba3ff7f41489125570_rxcauchy_nmhS
#> #define rxnbinomMu _rxf0af58525eeda19624ba3ff7f41489125571_rxnbinomMu_boMZ
#> #define rxnbinom _rxf0af58525eeda19624ba3ff7f41489125572_rxnbinom_YBjk
#> #define rxbinom _rxf0af58525eeda19624ba3ff7f41489125573_rxbinom_iCoU
#> #define rxnorm _rxf0af58525eeda19624ba3ff7f41489125574_rxnorm_qvni
#> #define linCmtB _rxf0af58525eeda19624ba3ff7f41489125575_linCmtB_I75U
#> #define linCmtA _rxf0af58525eeda19624ba3ff7f41489125576_linCmtA_i0GV
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){}
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){}
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  rx_75e02901f2375e3ae5a3571ce0b4d400__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *rx_75e02901f2375e3ae5a3571ce0b4d400__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP rx_75e02901f2375e3ae5a3571ce0b4d400__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void rx_75e02901f2375e3ae5a3571ce0b4d400__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double a = NA_REAL;
#> 
#>   double __udf[2];
#>   a = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   a =rxbeta(&_solveData->subjects[_cSub], 2,(double)2);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void rx_75e02901f2375e3ae5a3571ce0b4d400__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void rx_75e02901f2375e3ae5a3571ce0b4d400__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void rx_75e02901f2375e3ae5a3571ce0b4d400__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double a = NA_REAL;
#> 
#>   double __udf[2];
#>   a = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   a =rxbeta(&_solveData->subjects[_cSub], 2,(double)2);
#> 
#>   _lhs[_LHS_0_]=a;
#> }
#> // Functional based bioavailability
#> double rx_75e02901f2375e3ae5a3571ce0b4d400__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double rx_75e02901f2375e3ae5a3571ce0b4d400__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double rx_75e02901f2375e3ae5a3571ce0b4d400__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double rx_75e02901f2375e3ae5a3571ce0b4d400__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void rx_75e02901f2375e3ae5a3571ce0b4d400__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void rx_75e02901f2375e3ae5a3571ce0b4d400__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void rx_75e02901f2375e3ae5a3571ce0b4d400__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP rx_75e02901f2375e3ae5a3571ce0b4d400__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_75e02901f2375e3ae5a3571ce0b4d400__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAA[9zAAAv7#aT)km/*MAg)4+=WjmaBg@|dt6B;;#mY2Hby\"1!Fu$_fRo*5xZ{CAzPmTV*UtI:p0)i6*T|y\"t&]<~YR\"\?y#oquoVpR^OW=JCAIA~8]WQ9+b4+7wsnSw*9V2ly(lfg4m8Vl|yh..!7{k;n;;1rAwd=u`wdf{f1_<Ez1i&%%Whw>intGdkzkHsHmbwj(&.j!h8+.u/P*U0$<eYk1seDky2_DM*zf,.)aLxer\?@diT%%[Px#i;7)AEV6mG2wbZ\")a4~1\"uGu:P,id3X4HT1fEp8UlSo~KyHks,\?z~]r*%%q9p;iYrE57Q*)f,HK><agq*jOlVeT;e4jaS&&.V&gPf~hSTzp0w2EW&%%J7a<nGS&08)RA8y1lOeM|HE:L#HixzNiinj4{.Ah@4T)k%%nSCyj{kSYr8]5}oYS\?7TgTXJI=~&e}aQ]46Cz#*(Es!]K3y;TeOv=aHH0U$5;mhE#U9.!r8@V_R<Co{]|@OH_|V4j%%lu^Tp&4NycU5@Kp8fc%%@5.`ebl8|0bdx#L`9\?x\?ulg<UGy#><UG3~$DcTUg_CHll@BiY$u9DDgQAAnB9ARjrUYG\?WnL#Ic^*F`D=5^,Vc`).D&+!H#hSpyqta@@O(+:C{r\?l8}/L8^5\?qQ3~[^cT*F=a)Fq/9d%%%%pX9.}Q9v7pv.*/KH6_|~#YE#Oy/6D{54V{b!^Srv~%%Lx;lK!SK`Q#[2(zX=J:c,+[qYVC}BU*0.~2wpQXs*X:bl7f\?vHDFYOc\?59ZLF1BH,b72C.A[7\"M,\".A\"F:MGGDDBMK=TnB*+~kcWAB:Z={~D_\"HQJC@6WS\"w1*ZuAqYSzCpjBBJ.&GzN&@<FkkB#b:(\"\"jKI(`AJ@E3[2]4J8L_7OjX8q$}YXkxL\?<t5B^3\"^@IA5HFF[_/Sg~_W.|zAxt#%%wo<)H_J&`_z`LOPIupJSG#!uX)2rnmOlRt5fGKM`7G|ojl7V!R7miaF\?iqYm\?S<DyKvBQ2:4|b%%sk1xLG`\?wwL(8E*=x52KN%%@8\?T<ZuRJUyk1N1wCVB");
#>     char buf[1063];
#> #define __doBufN__ 1063
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void rx_75e02901f2375e3ae5a3571ce0b4d400__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   rx_75e02901f2375e3ae5a3571ce0b4d400__dydt(neq, *t, A, DADT);
#> }
#> extern int rx_75e02901f2375e3ae5a3571ce0b4d400__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   rx_75e02901f2375e3ae5a3571ce0b4d400__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void rx_75e02901f2375e3ae5a3571ce0b4d400__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   rx_75e02901f2375e3ae5a3571ce0b4d400__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void rx_75e02901f2375e3ae5a3571ce0b4d400__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_rx_75e02901f2375e3ae5a3571ce0b4d400_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("rx_75e02901f2375e3ae5a3571ce0b4d400_","rx_75e02901f2375e3ae5a3571ce0b4d400__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("rx_75e02901f2375e3ae5a3571ce0b4d400_","rx_75e02901f2375e3ae5a3571ce0b4d400__assignFuns", (DL_FUNC) rx_75e02901f2375e3ae5a3571ce0b4d400__assignFuns);
#>   R_RegisterCCallable("rx_75e02901f2375e3ae5a3571ce0b4d400_","rx_75e02901f2375e3ae5a3571ce0b4d400__inis",(DL_FUNC) rx_75e02901f2375e3ae5a3571ce0b4d400__inis);
#>   R_RegisterCCallable("rx_75e02901f2375e3ae5a3571ce0b4d400_","rx_75e02901f2375e3ae5a3571ce0b4d400__dydt",(DL_FUNC) rx_75e02901f2375e3ae5a3571ce0b4d400__dydt);
#>   R_RegisterCCallable("rx_75e02901f2375e3ae5a3571ce0b4d400_","rx_75e02901f2375e3ae5a3571ce0b4d400__calc_lhs",(DL_FUNC) rx_75e02901f2375e3ae5a3571ce0b4d400__calc_lhs);
#>   R_RegisterCCallable("rx_75e02901f2375e3ae5a3571ce0b4d400_","rx_75e02901f2375e3ae5a3571ce0b4d400__calc_jac",(DL_FUNC) rx_75e02901f2375e3ae5a3571ce0b4d400__calc_jac);
#>   R_RegisterCCallable("rx_75e02901f2375e3ae5a3571ce0b4d400_","rx_75e02901f2375e3ae5a3571ce0b4d400__dydt_lsoda", (DL_FUNC) rx_75e02901f2375e3ae5a3571ce0b4d400__dydt_lsoda);
#>   R_RegisterCCallable("rx_75e02901f2375e3ae5a3571ce0b4d400_","rx_75e02901f2375e3ae5a3571ce0b4d400__calc_jac_lsoda", (DL_FUNC) rx_75e02901f2375e3ae5a3571ce0b4d400__calc_jac_lsoda);
#>   R_RegisterCCallable("rx_75e02901f2375e3ae5a3571ce0b4d400_","rx_75e02901f2375e3ae5a3571ce0b4d400__ode_solver_solvedata", (DL_FUNC) rx_75e02901f2375e3ae5a3571ce0b4d400__ode_solver_solvedata);
#>   R_RegisterCCallable("rx_75e02901f2375e3ae5a3571ce0b4d400_","rx_75e02901f2375e3ae5a3571ce0b4d400__ode_solver_get_solvedata", (DL_FUNC) rx_75e02901f2375e3ae5a3571ce0b4d400__ode_solver_get_solvedata);
#>   R_RegisterCCallable("rx_75e02901f2375e3ae5a3571ce0b4d400_","rx_75e02901f2375e3ae5a3571ce0b4d400__F", (DL_FUNC) rx_75e02901f2375e3ae5a3571ce0b4d400__F);
#>   R_RegisterCCallable("rx_75e02901f2375e3ae5a3571ce0b4d400_","rx_75e02901f2375e3ae5a3571ce0b4d400__Lag", (DL_FUNC) rx_75e02901f2375e3ae5a3571ce0b4d400__Lag);
#>   R_RegisterCCallable("rx_75e02901f2375e3ae5a3571ce0b4d400_","rx_75e02901f2375e3ae5a3571ce0b4d400__Rate", (DL_FUNC) rx_75e02901f2375e3ae5a3571ce0b4d400__Rate);
#>   R_RegisterCCallable("rx_75e02901f2375e3ae5a3571ce0b4d400_","rx_75e02901f2375e3ae5a3571ce0b4d400__Dur", (DL_FUNC) rx_75e02901f2375e3ae5a3571ce0b4d400__Dur);
#>   R_RegisterCCallable("rx_75e02901f2375e3ae5a3571ce0b4d400_","rx_75e02901f2375e3ae5a3571ce0b4d400__mtime", (DL_FUNC) rx_75e02901f2375e3ae5a3571ce0b4d400__mtime);
#>   R_RegisterCCallable("rx_75e02901f2375e3ae5a3571ce0b4d400_","rx_75e02901f2375e3ae5a3571ce0b4d400__ME", (DL_FUNC) rx_75e02901f2375e3ae5a3571ce0b4d400__ME);
#>   R_RegisterCCallable("rx_75e02901f2375e3ae5a3571ce0b4d400_","rx_75e02901f2375e3ae5a3571ce0b4d400__IndF", (DL_FUNC) rx_75e02901f2375e3ae5a3571ce0b4d400__IndF);
#>   R_RegisterCCallable("rx_75e02901f2375e3ae5a3571ce0b4d400_","rx_75e02901f2375e3ae5a3571ce0b4d400__dydt_liblsoda", (DL_FUNC) rx_75e02901f2375e3ae5a3571ce0b4d400__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_rx_75e02901f2375e3ae5a3571ce0b4d400_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_rx_75e02901f2375e3ae5a3571ce0b4d400_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"rx_75e02901f2375e3ae5a3571ce0b4d400__model_vars", (DL_FUNC) &rx_75e02901f2375e3ae5a3571ce0b4d400__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_rx_75e02901f2375e3ae5a3571ce0b4d400_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_75e02901f2375e3ae5a3571ce0b4d400__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("rx_75e02901f2375e3ae5a3571ce0b4d400__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’
# }