Simulate Binomial variable from threefry generator

Care should be taken with this method not to encounter the birthday problem, described https://www.johndcook.com/blog/2016/01/29/random-number-generator-seed-mistakes/. Since the sitmo threefry, this currently generates one random deviate from the uniform distribution to seed the engine threefry and then run the code.

Usage

rxbinom(size, prob, n = 1L, ncores = 1L)

Arguments

size

number of trials (zero or more).

prob

probability of success on each trial.

n

number of observations. If length(n) > 1, the length is taken to be the number required.

ncores

Number of cores for the simulation

rxnorm simulates using the threefry sitmo generator.

rxnormV used to simulate with the vandercorput simulator, but since it didn't satisfy the normal properties it was changed to simple be an alias of rxnorm. It is no longer supported in rxode2({}) blocks

Value

binomial random deviates

Details

Therefore, a simple call to the random number generated followed by a second call to random number generated may have identical seeds. As the number of random number generator calls are increased the probability that the birthday problem will increase.

The key to avoid this problem is to either run all simulations in the rxode2 environment once (therefore one seed or series of seeds for the whole simulation), pre-generate all random variables used for the simulation, or seed the rxode2 engine with rxSetSeed()

Internally each ID is seeded with a unique number so that the results do not depend on the number of cores used.

Examples

# \donttest{
## Use threefry engine

rxbinom(10, 0.9, n = 10) # with rxbinom you have to explicitly state n
#>  [1] 10  8  9  7 10  9  9  8 10  8
rxbinom(3, 0.5, n = 10, ncores = 2) # You can parallelize the simulation using openMP
#>  [1] 3 3 2 3 2 3 2 1 3 1

rxbinom(4, 0.7)
#> [1] 3

## This example uses `rxbinom` directly in the model

rx <- function() {
  model({
    a <- rxbinom(1, 0.5)
  })
}

et <- et(1, id = 1:2)

s <- rxSolve(rx, et)
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#>  
#>  
#> C model:
#> // Define 1 LHS values
#> #define _LHS_0_ 0 // a
#> #define _getRxSolve_ _rxee3bca887c02b45602f539b3985d18b65577__getRxSolve__1Nsb
#> #define _evalUdf _rxee3bca887c02b45602f539b3985d18b65578__evalUdf_x6oc
#> #define _solveData _rxee3bca887c02b45602f539b3985d18b65579__solveData_8HDH
#> #define _assign_ptr _rxee3bca887c02b45602f539b3985d18b65580__assign_ptr_kExm
#> #define _rxRmModelLib _rxee3bca887c02b45602f539b3985d18b65581__rxRmModelLib_1Jd4
#> #define _rxGetModelLib _rxee3bca887c02b45602f539b3985d18b65582__rxGetModelLib_ayle
#> #define _old_c _rxee3bca887c02b45602f539b3985d18b65583__old_c_W9go
#> #define _ptrid _rxee3bca887c02b45602f539b3985d18b65584__ptrid_D3R3
#> #define _rxIsCurrentC _rxee3bca887c02b45602f539b3985d18b65585__rxIsCurrentC_9Jvc
#> #define _sumPS _rxee3bca887c02b45602f539b3985d18b65586__sumPS_r1Um
#> #define _prodPS _rxee3bca887c02b45602f539b3985d18b65587__prodPS_LFGL
#> #define _prodType _rxee3bca887c02b45602f539b3985d18b65588__prodType_1zNF
#> #define _sumType _rxee3bca887c02b45602f539b3985d18b65589__sumType_D800
#> #define _update_par_ptr _rxee3bca887c02b45602f539b3985d18b65590__update_par_ptr_MrOc
#> #define _getParCov _rxee3bca887c02b45602f539b3985d18b65591__getParCov_lBdv
#> #define _rxode2_rxAssignPtr _rxee3bca887c02b45602f539b3985d18b65592__rxode2_rxAssignPtr_cHYH
#> #define _rxQr _rxee3bca887c02b45602f539b3985d18b65593__rxQr_lORz
#> #define _compareFactorVal _rxee3bca887c02b45602f539b3985d18b65594__compareFactorVal_kg9p
#> #define _sum _rxee3bca887c02b45602f539b3985d18b65595__sum_hZjb
#> #define _udf _rxee3bca887c02b45602f539b3985d18b65596__udf_tvsF
#> #define _sign _rxee3bca887c02b45602f539b3985d18b65597__sign_QkpV
#> #define _prod _rxee3bca887c02b45602f539b3985d18b65598__prod_HF4t
#> #define _max _rxee3bca887c02b45602f539b3985d18b65599__max_E5bq
#> #define _min _rxee3bca887c02b45602f539b3985d18b65600__min_rWj2
#> #define _transit4P _rxee3bca887c02b45602f539b3985d18b65601__transit4P_jIRW
#> #define _transit3P _rxee3bca887c02b45602f539b3985d18b65602__transit3P_0crr
#> #define _assignFuns0 _rxee3bca887c02b45602f539b3985d18b65603__assignFuns0_gcGm
#> #define _assignFuns _rxee3bca887c02b45602f539b3985d18b65604__assignFuns_EsJu
#> #define _rxord _rxee3bca887c02b45602f539b3985d18b65605__rxord_9xah
#> #define __assignFuns2 _rxee3bca887c02b45602f539b3985d18b65606___assignFuns2_gBOZ
#> #define _llikCauchyDscale _rxee3bca887c02b45602f539b3985d18b65607__llikCauchyDscale_l3zY
#> #define _llikCauchyDlocation _rxee3bca887c02b45602f539b3985d18b65608__llikCauchyDlocation_vTa9
#> #define _llikCauchy _rxee3bca887c02b45602f539b3985d18b65609__llikCauchy_7EhW
#> #define _llikGammaDrate _rxee3bca887c02b45602f539b3985d18b65610__llikGammaDrate_3MG9
#> #define _llikGammaDshape _rxee3bca887c02b45602f539b3985d18b65611__llikGammaDshape_V3pR
#> #define _llikGamma _rxee3bca887c02b45602f539b3985d18b65612__llikGamma_1aNk
#> #define _llikWeibullDscale _rxee3bca887c02b45602f539b3985d18b65613__llikWeibullDscale_G5Jw
#> #define _llikWeibullDshape _rxee3bca887c02b45602f539b3985d18b65614__llikWeibullDshape_D9TK
#> #define _llikWeibull _rxee3bca887c02b45602f539b3985d18b65615__llikWeibull_40NP
#> #define _llikUnifDbeta _rxee3bca887c02b45602f539b3985d18b65616__llikUnifDbeta_JiAL
#> #define _llikUnifDalpha _rxee3bca887c02b45602f539b3985d18b65617__llikUnifDalpha_ZBrh
#> #define _llikUnif _rxee3bca887c02b45602f539b3985d18b65618__llikUnif_efgh
#> #define _llikGeomDp _rxee3bca887c02b45602f539b3985d18b65619__llikGeomDp_Drm8
#> #define _llikGeom _rxee3bca887c02b45602f539b3985d18b65620__llikGeom_5igv
#> #define _llikFDdf2 _rxee3bca887c02b45602f539b3985d18b65621__llikFDdf2_lJYg
#> #define _llikFDdf1 _rxee3bca887c02b45602f539b3985d18b65622__llikFDdf1_GqQx
#> #define _llikF _rxee3bca887c02b45602f539b3985d18b65623__llikF_LoYl
#> #define _llikExpDrate _rxee3bca887c02b45602f539b3985d18b65624__llikExpDrate_FgJd
#> #define _llikExp _rxee3bca887c02b45602f539b3985d18b65625__llikExp_KaFg
#> #define _llikChisqDdf _rxee3bca887c02b45602f539b3985d18b65626__llikChisqDdf_mDR4
#> #define _llikChisq _rxee3bca887c02b45602f539b3985d18b65627__llikChisq_jxuo
#> #define _llikTDsd _rxee3bca887c02b45602f539b3985d18b65628__llikTDsd_P8gm
#> #define _llikTDmean _rxee3bca887c02b45602f539b3985d18b65629__llikTDmean_oUom
#> #define _llikTDdf _rxee3bca887c02b45602f539b3985d18b65630__llikTDdf_wLtu
#> #define _llikT _rxee3bca887c02b45602f539b3985d18b65631__llikT_DgOz
#> #define _llikBetaDshape2 _rxee3bca887c02b45602f539b3985d18b65632__llikBetaDshape2_tPZw
#> #define _llikBetaDshape1 _rxee3bca887c02b45602f539b3985d18b65633__llikBetaDshape1_Dqx2
#> #define _llikBeta _rxee3bca887c02b45602f539b3985d18b65634__llikBeta_t0Ff
#> #define _llikNbinomMuDmu _rxee3bca887c02b45602f539b3985d18b65635__llikNbinomMuDmu_CRup
#> #define _llikNbinomMu _rxee3bca887c02b45602f539b3985d18b65636__llikNbinomMu_BfcL
#> #define _llikNbinomDprob _rxee3bca887c02b45602f539b3985d18b65637__llikNbinomDprob_LlJk
#> #define _llikNbinom _rxee3bca887c02b45602f539b3985d18b65638__llikNbinom_0HOu
#> #define _llikBinomDprob _rxee3bca887c02b45602f539b3985d18b65639__llikBinomDprob_quMQ
#> #define _llikBinom _rxee3bca887c02b45602f539b3985d18b65640__llikBinom_IP4s
#> #define _llikPoisDlambda _rxee3bca887c02b45602f539b3985d18b65641__llikPoisDlambda_BKwl
#> #define _llikPois _rxee3bca887c02b45602f539b3985d18b65642__llikPois_ION1
#> #define _llikNormDsd _rxee3bca887c02b45602f539b3985d18b65643__llikNormDsd_tTX5
#> #define _llikNormDmean _rxee3bca887c02b45602f539b3985d18b65644__llikNormDmean_DulH
#> #define _llikNorm _rxee3bca887c02b45602f539b3985d18b65645__llikNorm_LxeA
#> #define simeps _rxee3bca887c02b45602f539b3985d18b65646_simeps_Ut0p
#> #define simeta _rxee3bca887c02b45602f539b3985d18b65647_simeta_o6Aq
#> #define expit _rxee3bca887c02b45602f539b3985d18b65648_expit_SQkn
#> #define logit _rxee3bca887c02b45602f539b3985d18b65649_logit_H8fP
#> #define gammapDer _rxee3bca887c02b45602f539b3985d18b65650_gammapDer_CAnN
#> #define lowergamma _rxee3bca887c02b45602f539b3985d18b65651_lowergamma_kUms
#> #define uppergamma _rxee3bca887c02b45602f539b3985d18b65652_uppergamma_wiOW
#> #define gammaqInva _rxee3bca887c02b45602f539b3985d18b65653_gammaqInva_G16H
#> #define gammaqInv _rxee3bca887c02b45602f539b3985d18b65654_gammaqInv_PUgn
#> #define gammapInva _rxee3bca887c02b45602f539b3985d18b65655_gammapInva_Cjuw
#> #define gammapInv _rxee3bca887c02b45602f539b3985d18b65656_gammapInv_A8iw
#> #define gammaq _rxee3bca887c02b45602f539b3985d18b65657_gammaq_bn2G
#> #define gammap _rxee3bca887c02b45602f539b3985d18b65658_gammap_TQjS
#> #define phi _rxee3bca887c02b45602f539b3985d18b65659_phi_rer5
#> #define d2ELUa _rxee3bca887c02b45602f539b3985d18b65660_d2ELUa_zexL
#> #define dELUa _rxee3bca887c02b45602f539b3985d18b65661_dELUa_gAwM
#> #define d2aELU _rxee3bca887c02b45602f539b3985d18b65662_d2aELU_E6nq
#> #define d2ELU _rxee3bca887c02b45602f539b3985d18b65663_d2ELU_R3DD
#> #define dELU _rxee3bca887c02b45602f539b3985d18b65664_dELU_9jN5
#> #define ELU _rxee3bca887c02b45602f539b3985d18b65665_ELU_O4r9
#> #define dPReLUa1 _rxee3bca887c02b45602f539b3985d18b65666_dPReLUa1_D97m
#> #define dPReLUa _rxee3bca887c02b45602f539b3985d18b65667_dPReLUa_e5yb
#> #define dPReLU _rxee3bca887c02b45602f539b3985d18b65668_dPReLU_V8Eu
#> #define PReLU _rxee3bca887c02b45602f539b3985d18b65669_PReLU_Z1xT
#> #define dSwish _rxee3bca887c02b45602f539b3985d18b65670_dSwish_QYOQ
#> #define Swish _rxee3bca887c02b45602f539b3985d18b65671_Swish_e6aq
#> #define dlReLU _rxee3bca887c02b45602f539b3985d18b65672_dlReLU_t3Lb
#> #define lReLU _rxee3bca887c02b45602f539b3985d18b65673_lReLU_sbea
#> #define dSELU _rxee3bca887c02b45602f539b3985d18b65674_dSELU_w5m4
#> #define SELU _rxee3bca887c02b45602f539b3985d18b65675_SELU_brD9
#> #define d4softplus _rxee3bca887c02b45602f539b3985d18b65676_d4softplus_EZAJ
#> #define d3softplus _rxee3bca887c02b45602f539b3985d18b65677_d3softplus_Ne7I
#> #define d2softplus _rxee3bca887c02b45602f539b3985d18b65678_d2softplus_xUFA
#> #define dsoftplus _rxee3bca887c02b45602f539b3985d18b65679_dsoftplus_5R2w
#> #define softplus _rxee3bca887c02b45602f539b3985d18b65680_softplus_ZwwB
#> #define d4GELU _rxee3bca887c02b45602f539b3985d18b65681_d4GELU_gRvb
#> #define d3GELU _rxee3bca887c02b45602f539b3985d18b65682_d3GELU_lnk3
#> #define d2GELU _rxee3bca887c02b45602f539b3985d18b65683_d2GELU_57cf
#> #define dGELU _rxee3bca887c02b45602f539b3985d18b65684_dGELU_UwMa
#> #define GELU _rxee3bca887c02b45602f539b3985d18b65685_GELU_rEA9
#> #define dReLU _rxee3bca887c02b45602f539b3985d18b65686_dReLU_z1SE
#> #define ReLU _rxee3bca887c02b45602f539b3985d18b65687_ReLU_1KzA
#> #define riweibull _rxee3bca887c02b45602f539b3985d18b65688_riweibull_f2H4
#> #define riunif _rxee3bca887c02b45602f539b3985d18b65689_riunif_Q2oF
#> #define rit_ _rxee3bca887c02b45602f539b3985d18b65690_rit__0joC
#> #define ripois _rxee3bca887c02b45602f539b3985d18b65691_ripois_Gj0a
#> #define ribeta _rxee3bca887c02b45602f539b3985d18b65692_ribeta_8wAb
#> #define rigamma _rxee3bca887c02b45602f539b3985d18b65693_rigamma_QELO
#> #define rigeom _rxee3bca887c02b45602f539b3985d18b65694_rigeom_lObq
#> #define rif _rxee3bca887c02b45602f539b3985d18b65695_rif_Nath
#> #define riexp _rxee3bca887c02b45602f539b3985d18b65696_riexp_ltiW
#> #define richisq _rxee3bca887c02b45602f539b3985d18b65697_richisq_R2IG
#> #define ricauchy _rxee3bca887c02b45602f539b3985d18b65698_ricauchy_VLQY
#> #define rinbinomMu _rxee3bca887c02b45602f539b3985d18b65699_rinbinomMu_M84J
#> #define rinbinom _rxee3bca887c02b45602f539b3985d18b65700_rinbinom_Cnuy
#> #define ribinom _rxee3bca887c02b45602f539b3985d18b65701_ribinom_CXPJ
#> #define rinorm _rxee3bca887c02b45602f539b3985d18b65702_rinorm_UjvF
#> #define rxweibull _rxee3bca887c02b45602f539b3985d18b65703_rxweibull_k7KJ
#> #define rxunif _rxee3bca887c02b45602f539b3985d18b65704_rxunif_vDxq
#> #define rxt_ _rxee3bca887c02b45602f539b3985d18b65705_rxt__bVSK
#> #define rxpois _rxee3bca887c02b45602f539b3985d18b65706_rxpois_Vfr7
#> #define rxbeta _rxee3bca887c02b45602f539b3985d18b65707_rxbeta_TcaU
#> #define rxgamma _rxee3bca887c02b45602f539b3985d18b65708_rxgamma_EgpY
#> #define rxgeom _rxee3bca887c02b45602f539b3985d18b65709_rxgeom_kNIq
#> #define rxf _rxee3bca887c02b45602f539b3985d18b65710_rxf_nEnm
#> #define rxexp _rxee3bca887c02b45602f539b3985d18b65711_rxexp_hSbo
#> #define rxchisq _rxee3bca887c02b45602f539b3985d18b65712_rxchisq_MZYB
#> #define rxcauchy _rxee3bca887c02b45602f539b3985d18b65713_rxcauchy_jkiC
#> #define rxnbinomMu _rxee3bca887c02b45602f539b3985d18b65714_rxnbinomMu_oUqv
#> #define rxnbinom _rxee3bca887c02b45602f539b3985d18b65715_rxnbinom_niI7
#> #define rxbinom _rxee3bca887c02b45602f539b3985d18b65716_rxbinom_5Ui0
#> #define rxnorm _rxee3bca887c02b45602f539b3985d18b65717_rxnorm_GVTs
#> #define linCmtB _rxee3bca887c02b45602f539b3985d18b65718_linCmtB_34TN
#> #define linCmtA _rxee3bca887c02b45602f539b3985d18b65719_linCmtA_HeFp
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){}
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){}
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  rx_44005ac63828fefe2da429f50e0ab26f__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *rx_44005ac63828fefe2da429f50e0ab26f__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP rx_44005ac63828fefe2da429f50e0ab26f__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void rx_44005ac63828fefe2da429f50e0ab26f__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double a = NA_REAL;
#> 
#>   double __udf[2];
#>   a = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   a =(double)rxbinom(&_solveData->subjects[_cSub], (int)1,(double)0.5);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void rx_44005ac63828fefe2da429f50e0ab26f__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void rx_44005ac63828fefe2da429f50e0ab26f__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void rx_44005ac63828fefe2da429f50e0ab26f__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double a = NA_REAL;
#> 
#>   double __udf[2];
#>   a = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   a =(double)rxbinom(&_solveData->subjects[_cSub], (int)1,(double)0.5);
#> 
#>   _lhs[_LHS_0_]=a;
#> }
#> // Functional based bioavailability
#> double rx_44005ac63828fefe2da429f50e0ab26f__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double rx_44005ac63828fefe2da429f50e0ab26f__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double rx_44005ac63828fefe2da429f50e0ab26f__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double rx_44005ac63828fefe2da429f50e0ab26f__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void rx_44005ac63828fefe2da429f50e0ab26f__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void rx_44005ac63828fefe2da429f50e0ab26f__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void rx_44005ac63828fefe2da429f50e0ab26f__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP rx_44005ac63828fefe2da429f50e0ab26f__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_44005ac63828fefe2da429f50e0ab26f__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAAAAaAAA+:g6JiVTmmGAlj_5!(S$5A11<i}WDX7tN[5}EAhVz$}8ZRHdZjyuv,Dz<de=\?b}Dc5U:b4]0.&gM;IFohbJM`\?q\?{!WXE\"uWyWgA_TcaDueg(0Nz6HK!lBB(8g]U{8gnp$C[^(^U;oiE6[iwt&l(4Bq>|1y328*qT&|\"6G&;%%[T{Q|XN}oIq57fw>Nr#kJWFiw,Kj\"XNi+g4.Mg}p;y.$MMly*K970;mMRh&,`&#+f/0@.nfO!{&LDFvo48t]91RG\"{5iB4`+/n6=|8tjXKeC*oO+PYNT=qm~To(6c#HL$JGM!UBo{l|!3f,m73&OfXfB@Z=`=(zO/a^`.2[5jp)wcP7;CcS0_8i0Qsjv2Un<IyBht!!h;JO#Pg6L(bD(0GR.u]JDzWUt_6:uRvgdtbv0zjzx/JP:[W3p&cHWFNDdxy61Ma,<J*R_qStjC)[^hbX}QL(mq$ZzOMG|YRG\?Y2~HK{rmZ6[e_{]px052bl/fhENnSl5$Tu7$0ft$iPN+~Pnp]<6|:!PIvN9Dvg8PY1^UQg31^f#/2j%%1@}~4^pBy#P`qnY$wT}u0J+>&`g)R+EmNB@qnVr/f>lhY\"]uieZl[2uLc&,4U;zCy$dj;}k1c2F3rUd{)<C;~o49yzj^$Y>yN=bq7~i0)^n3%%Q%%~27B;(_s|z5,KY_IZnL2j{qALA{q[H]Tstafh@XSrZnQi,rc^4Z\?ZD.J:@7f+pbbYY:Me2mkjc1nabx:j7!dJ>H*XWCcjOcC2bbMKXL<v4Om<7y+>LrVH>v_Qa)2B.t](e/R=Ky\")G{{R^TPjt#I9r8e|u*qucg#Fw1QV+Znt0w+gElOqD#Lsj)Yxt_OuYFk{pnh!H@,f7WIC!]&6Yl_[Ys*C!_9v$QzN&*<cwE!(DYy`hlCtgS{0D;HD+Y)9^Mgn,\"4ClFB&!aLLQ{*n88|eUwHIt%%\"dgZbm[JvKrCYkR&pKjYrWoXQF+f[16PuEi*2oIc<@W)z@:DcP8EFS9x1OF|el}o`w6I5l~kCoKE(GQo**8Eq6yTUA");
#>     char buf[1062];
#> #define __doBufN__ 1062
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void rx_44005ac63828fefe2da429f50e0ab26f__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   rx_44005ac63828fefe2da429f50e0ab26f__dydt(neq, *t, A, DADT);
#> }
#> extern int rx_44005ac63828fefe2da429f50e0ab26f__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   rx_44005ac63828fefe2da429f50e0ab26f__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void rx_44005ac63828fefe2da429f50e0ab26f__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   rx_44005ac63828fefe2da429f50e0ab26f__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void rx_44005ac63828fefe2da429f50e0ab26f__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_rx_44005ac63828fefe2da429f50e0ab26f_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("rx_44005ac63828fefe2da429f50e0ab26f_","rx_44005ac63828fefe2da429f50e0ab26f__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("rx_44005ac63828fefe2da429f50e0ab26f_","rx_44005ac63828fefe2da429f50e0ab26f__assignFuns", (DL_FUNC) rx_44005ac63828fefe2da429f50e0ab26f__assignFuns);
#>   R_RegisterCCallable("rx_44005ac63828fefe2da429f50e0ab26f_","rx_44005ac63828fefe2da429f50e0ab26f__inis",(DL_FUNC) rx_44005ac63828fefe2da429f50e0ab26f__inis);
#>   R_RegisterCCallable("rx_44005ac63828fefe2da429f50e0ab26f_","rx_44005ac63828fefe2da429f50e0ab26f__dydt",(DL_FUNC) rx_44005ac63828fefe2da429f50e0ab26f__dydt);
#>   R_RegisterCCallable("rx_44005ac63828fefe2da429f50e0ab26f_","rx_44005ac63828fefe2da429f50e0ab26f__calc_lhs",(DL_FUNC) rx_44005ac63828fefe2da429f50e0ab26f__calc_lhs);
#>   R_RegisterCCallable("rx_44005ac63828fefe2da429f50e0ab26f_","rx_44005ac63828fefe2da429f50e0ab26f__calc_jac",(DL_FUNC) rx_44005ac63828fefe2da429f50e0ab26f__calc_jac);
#>   R_RegisterCCallable("rx_44005ac63828fefe2da429f50e0ab26f_","rx_44005ac63828fefe2da429f50e0ab26f__dydt_lsoda", (DL_FUNC) rx_44005ac63828fefe2da429f50e0ab26f__dydt_lsoda);
#>   R_RegisterCCallable("rx_44005ac63828fefe2da429f50e0ab26f_","rx_44005ac63828fefe2da429f50e0ab26f__calc_jac_lsoda", (DL_FUNC) rx_44005ac63828fefe2da429f50e0ab26f__calc_jac_lsoda);
#>   R_RegisterCCallable("rx_44005ac63828fefe2da429f50e0ab26f_","rx_44005ac63828fefe2da429f50e0ab26f__ode_solver_solvedata", (DL_FUNC) rx_44005ac63828fefe2da429f50e0ab26f__ode_solver_solvedata);
#>   R_RegisterCCallable("rx_44005ac63828fefe2da429f50e0ab26f_","rx_44005ac63828fefe2da429f50e0ab26f__ode_solver_get_solvedata", (DL_FUNC) rx_44005ac63828fefe2da429f50e0ab26f__ode_solver_get_solvedata);
#>   R_RegisterCCallable("rx_44005ac63828fefe2da429f50e0ab26f_","rx_44005ac63828fefe2da429f50e0ab26f__F", (DL_FUNC) rx_44005ac63828fefe2da429f50e0ab26f__F);
#>   R_RegisterCCallable("rx_44005ac63828fefe2da429f50e0ab26f_","rx_44005ac63828fefe2da429f50e0ab26f__Lag", (DL_FUNC) rx_44005ac63828fefe2da429f50e0ab26f__Lag);
#>   R_RegisterCCallable("rx_44005ac63828fefe2da429f50e0ab26f_","rx_44005ac63828fefe2da429f50e0ab26f__Rate", (DL_FUNC) rx_44005ac63828fefe2da429f50e0ab26f__Rate);
#>   R_RegisterCCallable("rx_44005ac63828fefe2da429f50e0ab26f_","rx_44005ac63828fefe2da429f50e0ab26f__Dur", (DL_FUNC) rx_44005ac63828fefe2da429f50e0ab26f__Dur);
#>   R_RegisterCCallable("rx_44005ac63828fefe2da429f50e0ab26f_","rx_44005ac63828fefe2da429f50e0ab26f__mtime", (DL_FUNC) rx_44005ac63828fefe2da429f50e0ab26f__mtime);
#>   R_RegisterCCallable("rx_44005ac63828fefe2da429f50e0ab26f_","rx_44005ac63828fefe2da429f50e0ab26f__ME", (DL_FUNC) rx_44005ac63828fefe2da429f50e0ab26f__ME);
#>   R_RegisterCCallable("rx_44005ac63828fefe2da429f50e0ab26f_","rx_44005ac63828fefe2da429f50e0ab26f__IndF", (DL_FUNC) rx_44005ac63828fefe2da429f50e0ab26f__IndF);
#>   R_RegisterCCallable("rx_44005ac63828fefe2da429f50e0ab26f_","rx_44005ac63828fefe2da429f50e0ab26f__dydt_liblsoda", (DL_FUNC) rx_44005ac63828fefe2da429f50e0ab26f__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_rx_44005ac63828fefe2da429f50e0ab26f_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_rx_44005ac63828fefe2da429f50e0ab26f_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"rx_44005ac63828fefe2da429f50e0ab26f__model_vars", (DL_FUNC) &rx_44005ac63828fefe2da429f50e0ab26f__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_rx_44005ac63828fefe2da429f50e0ab26f_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_44005ac63828fefe2da429f50e0ab26f__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("rx_44005ac63828fefe2da429f50e0ab26f__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’
# }