Simulate Cauchy variable from threefry generator

Care should be taken with this method not to encounter the birthday problem, described https://www.johndcook.com/blog/2016/01/29/random-number-generator-seed-mistakes/. Since the sitmo threefry, this currently generates one random deviate from the uniform distribution to seed the engine threefry and then run the code.

Usage

rxcauchy(location = 0, scale = 1, n = 1L, ncores = 1L)

Arguments

location, scale

location and scale parameters.

n

number of observations. If length(n) > 1, the length is taken to be the number required.

ncores

Number of cores for the simulation

rxnorm simulates using the threefry sitmo generator.

rxnormV used to simulate with the vandercorput simulator, but since it didn't satisfy the normal properties it was changed to simple be an alias of rxnorm. It is no longer supported in rxode2({}) blocks

Value

Cauchy random deviates

Details

Therefore, a simple call to the random number generated followed by a second call to random number generated may have identical seeds. As the number of random number generator calls are increased the probability that the birthday problem will increase.

The key to avoid this problem is to either run all simulations in the rxode2 environment once (therefore one seed or series of seeds for the whole simulation), pre-generate all random variables used for the simulation, or seed the rxode2 engine with rxSetSeed()

Internally each ID is seeded with a unique number so that the results do not depend on the number of cores used.

Examples

# \donttest{

## Use threefry engine

rxcauchy(0, 1, n = 10) # with rxcauchy you have to explicitly state n
#>  [1] -0.3161956 -9.8293808  0.2817785  0.5647767 -0.8987297 -0.2183899
#>  [7] 18.6560233  2.9999284 -0.5308393 -1.9472665
rxcauchy(0.5, n = 10, ncores = 2) # You can parallelize the simulation using openMP
#>  [1]   1.55072313   1.22780297 -49.40914679  32.47667344   2.33699668
#>  [6]   4.99563885   2.63606252  -1.93851598  -0.05421445 -11.45911180

rxcauchy(3)
#> [1] 2.112529

## This example uses `rxcauchy` directly in the model

rx <- function() {
  model({
    a <- rxcauchy(2)
  })
}

et <- et(1, id = 1:2)

s <- rxSolve(rx, et)
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#>  
#>  
#> C model:
#> // Define 1 LHS values
#> #define _LHS_0_ 0 // a
#> #define _getRxSolve_ _rxd73c089c366b97a8e54a2804498d7a325720__getRxSolve__DHkE
#> #define _evalUdf _rxd73c089c366b97a8e54a2804498d7a325721__evalUdf_xm1J
#> #define _solveData _rxd73c089c366b97a8e54a2804498d7a325722__solveData_d4ay
#> #define _assign_ptr _rxd73c089c366b97a8e54a2804498d7a325723__assign_ptr_leW9
#> #define _rxRmModelLib _rxd73c089c366b97a8e54a2804498d7a325724__rxRmModelLib_goD3
#> #define _rxGetModelLib _rxd73c089c366b97a8e54a2804498d7a325725__rxGetModelLib_R39J
#> #define _old_c _rxd73c089c366b97a8e54a2804498d7a325726__old_c_vcr1
#> #define _ptrid _rxd73c089c366b97a8e54a2804498d7a325727__ptrid_UmLF
#> #define _rxIsCurrentC _rxd73c089c366b97a8e54a2804498d7a325728__rxIsCurrentC_GL1z
#> #define _sumPS _rxd73c089c366b97a8e54a2804498d7a325729__sumPS_NFD8
#> #define _prodPS _rxd73c089c366b97a8e54a2804498d7a325730__prodPS_00Mr
#> #define _prodType _rxd73c089c366b97a8e54a2804498d7a325731__prodType_OclB
#> #define _sumType _rxd73c089c366b97a8e54a2804498d7a325732__sumType_dvcH
#> #define _update_par_ptr _rxd73c089c366b97a8e54a2804498d7a325733__update_par_ptr_YHlO
#> #define _getParCov _rxd73c089c366b97a8e54a2804498d7a325734__getParCov_Rzkg
#> #define _rxode2_rxAssignPtr _rxd73c089c366b97a8e54a2804498d7a325735__rxode2_rxAssignPtr_9phZ
#> #define _rxQr _rxd73c089c366b97a8e54a2804498d7a325736__rxQr_jbtv
#> #define _compareFactorVal _rxd73c089c366b97a8e54a2804498d7a325737__compareFactorVal_sFQk
#> #define _sum _rxd73c089c366b97a8e54a2804498d7a325738__sum_pVHF
#> #define _udf _rxd73c089c366b97a8e54a2804498d7a325739__udf_4tE5
#> #define _sign _rxd73c089c366b97a8e54a2804498d7a325740__sign_bqrW
#> #define _prod _rxd73c089c366b97a8e54a2804498d7a325741__prod_j2jI
#> #define _max _rxd73c089c366b97a8e54a2804498d7a325742__max_RW0c
#> #define _min _rxd73c089c366b97a8e54a2804498d7a325743__min_rrgc
#> #define _transit4P _rxd73c089c366b97a8e54a2804498d7a325744__transit4P_GmEs
#> #define _transit3P _rxd73c089c366b97a8e54a2804498d7a325745__transit3P_Ju9x
#> #define _assignFuns0 _rxd73c089c366b97a8e54a2804498d7a325746__assignFuns0_ahgB
#> #define _assignFuns _rxd73c089c366b97a8e54a2804498d7a325747__assignFuns_OZl3
#> #define _rxord _rxd73c089c366b97a8e54a2804498d7a325748__rxord_zYvT
#> #define __assignFuns2 _rxd73c089c366b97a8e54a2804498d7a325749___assignFuns2_a97E
#> #define _llikCauchyDscale _rxd73c089c366b97a8e54a2804498d7a325750__llikCauchyDscale_hW3M
#> #define _llikCauchyDlocation _rxd73c089c366b97a8e54a2804498d7a325751__llikCauchyDlocation_G9V3
#> #define _llikCauchy _rxd73c089c366b97a8e54a2804498d7a325752__llikCauchy_pR1a
#> #define _llikGammaDrate _rxd73c089c366b97a8e54a2804498d7a325753__llikGammaDrate_NkG5
#> #define _llikGammaDshape _rxd73c089c366b97a8e54a2804498d7a325754__llikGammaDshape_JwD9
#> #define _llikGamma _rxd73c089c366b97a8e54a2804498d7a325755__llikGamma_TK40
#> #define _llikWeibullDscale _rxd73c089c366b97a8e54a2804498d7a325756__llikWeibullDscale_NPJi
#> #define _llikWeibullDshape _rxd73c089c366b97a8e54a2804498d7a325757__llikWeibullDshape_ALZB
#> #define _llikWeibull _rxd73c089c366b97a8e54a2804498d7a325758__llikWeibull_rhef
#> #define _llikUnifDbeta _rxd73c089c366b97a8e54a2804498d7a325759__llikUnifDbeta_ghDr
#> #define _llikUnifDalpha _rxd73c089c366b97a8e54a2804498d7a325760__llikUnifDalpha_m85i
#> #define _llikUnif _rxd73c089c366b97a8e54a2804498d7a325761__llikUnif_gvlJ
#> #define _llikGeomDp _rxd73c089c366b97a8e54a2804498d7a325762__llikGeomDp_YgGq
#> #define _llikGeom _rxd73c089c366b97a8e54a2804498d7a325763__llikGeom_QxLo
#> #define _llikFDdf2 _rxd73c089c366b97a8e54a2804498d7a325764__llikFDdf2_YlFg
#> #define _llikFDdf1 _rxd73c089c366b97a8e54a2804498d7a325765__llikFDdf1_JdKa
#> #define _llikF _rxd73c089c366b97a8e54a2804498d7a325766__llikF_FgmD
#> #define _llikExpDrate _rxd73c089c366b97a8e54a2804498d7a325767__llikExpDrate_R4jx
#> #define _llikExp _rxd73c089c366b97a8e54a2804498d7a325768__llikExp_uoP8
#> #define _llikChisqDdf _rxd73c089c366b97a8e54a2804498d7a325769__llikChisqDdf_gmoU
#> #define _llikChisq _rxd73c089c366b97a8e54a2804498d7a325770__llikChisq_omwL
#> #define _llikTDsd _rxd73c089c366b97a8e54a2804498d7a325771__llikTDsd_tuDg
#> #define _llikTDmean _rxd73c089c366b97a8e54a2804498d7a325772__llikTDmean_OztP
#> #define _llikTDdf _rxd73c089c366b97a8e54a2804498d7a325773__llikTDdf_ZwDq
#> #define _llikT _rxd73c089c366b97a8e54a2804498d7a325774__llikT_x2t0
#> #define _llikBetaDshape2 _rxd73c089c366b97a8e54a2804498d7a325775__llikBetaDshape2_FfCR
#> #define _llikBetaDshape1 _rxd73c089c366b97a8e54a2804498d7a325776__llikBetaDshape1_upBf
#> #define _llikBeta _rxd73c089c366b97a8e54a2804498d7a325777__llikBeta_cLLl
#> #define _llikNbinomMuDmu _rxd73c089c366b97a8e54a2804498d7a325778__llikNbinomMuDmu_Jk0H
#> #define _llikNbinomMu _rxd73c089c366b97a8e54a2804498d7a325779__llikNbinomMu_Ouqu
#> #define _llikNbinomDprob _rxd73c089c366b97a8e54a2804498d7a325780__llikNbinomDprob_MQIP
#> #define _llikNbinom _rxd73c089c366b97a8e54a2804498d7a325781__llikNbinom_4sBK
#> #define _llikBinomDprob _rxd73c089c366b97a8e54a2804498d7a325782__llikBinomDprob_wlIO
#> #define _llikBinom _rxd73c089c366b97a8e54a2804498d7a325783__llikBinom_N1tT
#> #define _llikPoisDlambda _rxd73c089c366b97a8e54a2804498d7a325784__llikPoisDlambda_X5Du
#> #define _llikPois _rxd73c089c366b97a8e54a2804498d7a325785__llikPois_lHLx
#> #define _llikNormDsd _rxd73c089c366b97a8e54a2804498d7a325786__llikNormDsd_eAUt
#> #define _llikNormDmean _rxd73c089c366b97a8e54a2804498d7a325787__llikNormDmean_0po6
#> #define _llikNorm _rxd73c089c366b97a8e54a2804498d7a325788__llikNorm_AqSQ
#> #define simeps _rxd73c089c366b97a8e54a2804498d7a325789_simeps_knH8
#> #define simeta _rxd73c089c366b97a8e54a2804498d7a325790_simeta_fPCA
#> #define expit _rxd73c089c366b97a8e54a2804498d7a325791_expit_nNkU
#> #define logit _rxd73c089c366b97a8e54a2804498d7a325792_logit_mswi
#> #define gammapDer _rxd73c089c366b97a8e54a2804498d7a325793_gammapDer_OWG1
#> #define lowergamma _rxd73c089c366b97a8e54a2804498d7a325794_lowergamma_6HPU
#> #define uppergamma _rxd73c089c366b97a8e54a2804498d7a325795_uppergamma_gnCj
#> #define gammaqInva _rxd73c089c366b97a8e54a2804498d7a325796_gammaqInva_uwA8
#> #define gammaqInv _rxd73c089c366b97a8e54a2804498d7a325797_gammaqInv_iwbn
#> #define gammapInva _rxd73c089c366b97a8e54a2804498d7a325798_gammapInva_2GTQ
#> #define gammapInv _rxd73c089c366b97a8e54a2804498d7a325799_gammapInv_jSre
#> #define gammaq _rxd73c089c366b97a8e54a2804498d7a325800_gammaq_r5ze
#> #define gammap _rxd73c089c366b97a8e54a2804498d7a325801_gammap_xLgA
#> #define phi _rxd73c089c366b97a8e54a2804498d7a325802_phi_wME6
#> #define d2ELUa _rxd73c089c366b97a8e54a2804498d7a325803_d2ELUa_nqR3
#> #define dELUa _rxd73c089c366b97a8e54a2804498d7a325804_dELUa_DD9j
#> #define d2aELU _rxd73c089c366b97a8e54a2804498d7a325805_d2aELU_N5O4
#> #define d2ELU _rxd73c089c366b97a8e54a2804498d7a325806_d2ELU_r9D9
#> #define dELU _rxd73c089c366b97a8e54a2804498d7a325807_dELU_7me5
#> #define ELU _rxd73c089c366b97a8e54a2804498d7a325808_ELU_ybV8
#> #define dPReLUa1 _rxd73c089c366b97a8e54a2804498d7a325809_dPReLUa1_EuZ1
#> #define dPReLUa _rxd73c089c366b97a8e54a2804498d7a325810_dPReLUa_xTQY
#> #define dPReLU _rxd73c089c366b97a8e54a2804498d7a325811_dPReLU_OQe6
#> #define PReLU _rxd73c089c366b97a8e54a2804498d7a325812_PReLU_aqt3
#> #define dSwish _rxd73c089c366b97a8e54a2804498d7a325813_dSwish_Lbsb
#> #define Swish _rxd73c089c366b97a8e54a2804498d7a325814_Swish_eaw5
#> #define dlReLU _rxd73c089c366b97a8e54a2804498d7a325815_dlReLU_m4br
#> #define lReLU _rxd73c089c366b97a8e54a2804498d7a325816_lReLU_D9EZ
#> #define dSELU _rxd73c089c366b97a8e54a2804498d7a325817_dSELU_AJNe
#> #define SELU _rxd73c089c366b97a8e54a2804498d7a325818_SELU_7IxU
#> #define d4softplus _rxd73c089c366b97a8e54a2804498d7a325819_d4softplus_FA5R
#> #define d3softplus _rxd73c089c366b97a8e54a2804498d7a325820_d3softplus_2wZw
#> #define d2softplus _rxd73c089c366b97a8e54a2804498d7a325821_d2softplus_wBgR
#> #define dsoftplus _rxd73c089c366b97a8e54a2804498d7a325822_dsoftplus_vbln
#> #define softplus _rxd73c089c366b97a8e54a2804498d7a325823_softplus_k357
#> #define d4GELU _rxd73c089c366b97a8e54a2804498d7a325824_d4GELU_cfUw
#> #define d3GELU _rxd73c089c366b97a8e54a2804498d7a325825_d3GELU_MarE
#> #define d2GELU _rxd73c089c366b97a8e54a2804498d7a325826_d2GELU_A9z1
#> #define dGELU _rxd73c089c366b97a8e54a2804498d7a325827_dGELU_SE1K
#> #define GELU _rxd73c089c366b97a8e54a2804498d7a325828_GELU_zAf2
#> #define dReLU _rxd73c089c366b97a8e54a2804498d7a325829_dReLU_H4Q2
#> #define ReLU _rxd73c089c366b97a8e54a2804498d7a325830_ReLU_oF0j
#> #define riweibull _rxd73c089c366b97a8e54a2804498d7a325831_riweibull_oCGj
#> #define riunif _rxd73c089c366b97a8e54a2804498d7a325832_riunif_0a8w
#> #define rit_ _rxd73c089c366b97a8e54a2804498d7a325833_rit__AbQE
#> #define ripois _rxd73c089c366b97a8e54a2804498d7a325834_ripois_LOlO
#> #define ribeta _rxd73c089c366b97a8e54a2804498d7a325835_ribeta_bqNa
#> #define rigamma _rxd73c089c366b97a8e54a2804498d7a325836_rigamma_thlt
#> #define rigeom _rxd73c089c366b97a8e54a2804498d7a325837_rigeom_iWR2
#> #define rif _rxd73c089c366b97a8e54a2804498d7a325838_rif_IGVL
#> #define riexp _rxd73c089c366b97a8e54a2804498d7a325839_riexp_QYM8
#> #define richisq _rxd73c089c366b97a8e54a2804498d7a325840_richisq_4JCn
#> #define ricauchy _rxd73c089c366b97a8e54a2804498d7a325841_ricauchy_uyCX
#> #define rinbinomMu _rxd73c089c366b97a8e54a2804498d7a325842_rinbinomMu_PJUj
#> #define rinbinom _rxd73c089c366b97a8e54a2804498d7a325843_rinbinom_vFk7
#> #define ribinom _rxd73c089c366b97a8e54a2804498d7a325844_ribinom_KJvD
#> #define rinorm _rxd73c089c366b97a8e54a2804498d7a325845_rinorm_xqbV
#> #define rxweibull _rxd73c089c366b97a8e54a2804498d7a325846_rxweibull_SKVf
#> #define rxunif _rxd73c089c366b97a8e54a2804498d7a325847_rxunif_r7Tc
#> #define rxt_ _rxd73c089c366b97a8e54a2804498d7a325848_rxt__aUEg
#> #define rxpois _rxd73c089c366b97a8e54a2804498d7a325849_rxpois_pYkN
#> #define rxbeta _rxd73c089c366b97a8e54a2804498d7a325850_rxbeta_IqnE
#> #define rxgamma _rxd73c089c366b97a8e54a2804498d7a325851_rxgamma_nmhS
#> #define rxgeom _rxd73c089c366b97a8e54a2804498d7a325852_rxgeom_boMZ
#> #define rxf _rxd73c089c366b97a8e54a2804498d7a325853_rxf_YBjk
#> #define rxexp _rxd73c089c366b97a8e54a2804498d7a325854_rxexp_iCoU
#> #define rxchisq _rxd73c089c366b97a8e54a2804498d7a325855_rxchisq_qvni
#> #define rxcauchy _rxd73c089c366b97a8e54a2804498d7a325856_rxcauchy_I75U
#> #define rxnbinomMu _rxd73c089c366b97a8e54a2804498d7a325857_rxnbinomMu_i0GV
#> #define rxnbinom _rxd73c089c366b97a8e54a2804498d7a325858_rxnbinom_Ts34
#> #define rxbinom _rxd73c089c366b97a8e54a2804498d7a325859_rxbinom_TNHe
#> #define rxnorm _rxd73c089c366b97a8e54a2804498d7a325860_rxnorm_Fpue
#> #define linCmtB _rxd73c089c366b97a8e54a2804498d7a325861_linCmtB_1ZaE
#> #define linCmtA _rxd73c089c366b97a8e54a2804498d7a325862_linCmtA_nJR6
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){}
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){}
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  rx_dbb1a66380c6c71e0e945a65a37ff414__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *rx_dbb1a66380c6c71e0e945a65a37ff414__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP rx_dbb1a66380c6c71e0e945a65a37ff414__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void rx_dbb1a66380c6c71e0e945a65a37ff414__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double a = NA_REAL;
#> 
#>   double __udf[2];
#>   a = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   a =rxcauchy(&_solveData->subjects[_cSub], 2,(double)1);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void rx_dbb1a66380c6c71e0e945a65a37ff414__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void rx_dbb1a66380c6c71e0e945a65a37ff414__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void rx_dbb1a66380c6c71e0e945a65a37ff414__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double a = NA_REAL;
#> 
#>   double __udf[2];
#>   a = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   a =rxcauchy(&_solveData->subjects[_cSub], 2,(double)1);
#> 
#>   _lhs[_LHS_0_]=a;
#> }
#> // Functional based bioavailability
#> double rx_dbb1a66380c6c71e0e945a65a37ff414__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double rx_dbb1a66380c6c71e0e945a65a37ff414__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double rx_dbb1a66380c6c71e0e945a65a37ff414__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double rx_dbb1a66380c6c71e0e945a65a37ff414__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void rx_dbb1a66380c6c71e0e945a65a37ff414__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void rx_dbb1a66380c6c71e0e945a65a37ff414__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void rx_dbb1a66380c6c71e0e945a65a37ff414__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP rx_dbb1a66380c6c71e0e945a65a37ff414__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_dbb1a66380c6c71e0e945a65a37ff414__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAA5F0AAAv7#aT)omW>MAq+@+=WjmaB~15wu1tE9*u/kYHjAknu5m8Nwb|5yilZ5z_crD%%O=[!)IBgJJp|jl1>YExi&2h`fK2tE{akKeW=JwWQAF;g2}}Ry~k<_$Mmw=/V5C<C;3~kpTkPG!^]>#O{WBbI8S~@Kz|o:g_~tb=_tqQvbalM94e8~mrfZer12udw$fZa#[mvmS}ZBbbEXmt^ZPT/#$i1+/,u82v/@mjSEl^1&u1XmZ,f,X%%]Js%%@,p4POB};I4EnX+4@Fb\?EZCi9.+bxJK;lw.fKecmk:Dl8nXn\?!/9a2,aE^et1V)T8e8@f{*\?aQ`Z&Nf=ahoHP/\?{R|PjCd]t*<++!ZbVuGId7\?Tl%%2rijSLD@A0PD!Si:2~r[,ynlS&YO<aioOgx8Mn[Vtx`(XuI)z`/||V63x{=yv16ABEa%%#^7NEd]95a`ZuyB6a7McFpzkT^B@/8a#2(n]md#,!Q9M_4SBD.{=q%%Ok=Z>Q,n;]U.ntt|Eq9N6K:<7:BOYV4!^83\"pZ{Ug~W]NsM:8lzREhg\?6REg^%%6ts8|4yHVlA^<ksMvGVtMbquA..9C13n$auM`&`Eo0\"8{uI9C3bEkd4(tNE&+!HS4(uyqta]@}*md=Ot\?K9_/;8;5i%%>No;&kUw3{ruL_}bTum^C!;`.1b*,JLJy{wOa//9Ni9p$cW9tZW<z$6>P9\"[]H~r5<qb&W(U\?)ZQGqc=UyXBfpuRBJ1IBPm,[gLJ#YM1:%%df9x_D@LMu7RHOBbqSxo6y^XcHU^dZjnzV5w_XqIO\?/t](|>8k2mFZtiE}}I|J1R*<y;W;9})iVXOQ>vh+23@M%%WaY16vSHVw<G(S\?6*GqI)5CSr($!1Do]8dLEBfpi:6#}`Ys*C!_9v$Q06%%*<cwE!(DYy`hlCtgS{0D;2+*YYh<Z$>vBsK5VA_&Qts6:E*3!;h=Meg`[|dkLQ#0nGq[JD3%%_;pvH]~(D%%UFlwj0!CSnHj.fWa1`CT0aPV96SU.6Sg6O>e7>/$P0is8]5H+pQ{a!+~I9Sv(\?LuC");
#>     char buf[1072];
#> #define __doBufN__ 1072
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void rx_dbb1a66380c6c71e0e945a65a37ff414__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   rx_dbb1a66380c6c71e0e945a65a37ff414__dydt(neq, *t, A, DADT);
#> }
#> extern int rx_dbb1a66380c6c71e0e945a65a37ff414__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   rx_dbb1a66380c6c71e0e945a65a37ff414__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void rx_dbb1a66380c6c71e0e945a65a37ff414__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   rx_dbb1a66380c6c71e0e945a65a37ff414__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void rx_dbb1a66380c6c71e0e945a65a37ff414__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_rx_dbb1a66380c6c71e0e945a65a37ff414_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("rx_dbb1a66380c6c71e0e945a65a37ff414_","rx_dbb1a66380c6c71e0e945a65a37ff414__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("rx_dbb1a66380c6c71e0e945a65a37ff414_","rx_dbb1a66380c6c71e0e945a65a37ff414__assignFuns", (DL_FUNC) rx_dbb1a66380c6c71e0e945a65a37ff414__assignFuns);
#>   R_RegisterCCallable("rx_dbb1a66380c6c71e0e945a65a37ff414_","rx_dbb1a66380c6c71e0e945a65a37ff414__inis",(DL_FUNC) rx_dbb1a66380c6c71e0e945a65a37ff414__inis);
#>   R_RegisterCCallable("rx_dbb1a66380c6c71e0e945a65a37ff414_","rx_dbb1a66380c6c71e0e945a65a37ff414__dydt",(DL_FUNC) rx_dbb1a66380c6c71e0e945a65a37ff414__dydt);
#>   R_RegisterCCallable("rx_dbb1a66380c6c71e0e945a65a37ff414_","rx_dbb1a66380c6c71e0e945a65a37ff414__calc_lhs",(DL_FUNC) rx_dbb1a66380c6c71e0e945a65a37ff414__calc_lhs);
#>   R_RegisterCCallable("rx_dbb1a66380c6c71e0e945a65a37ff414_","rx_dbb1a66380c6c71e0e945a65a37ff414__calc_jac",(DL_FUNC) rx_dbb1a66380c6c71e0e945a65a37ff414__calc_jac);
#>   R_RegisterCCallable("rx_dbb1a66380c6c71e0e945a65a37ff414_","rx_dbb1a66380c6c71e0e945a65a37ff414__dydt_lsoda", (DL_FUNC) rx_dbb1a66380c6c71e0e945a65a37ff414__dydt_lsoda);
#>   R_RegisterCCallable("rx_dbb1a66380c6c71e0e945a65a37ff414_","rx_dbb1a66380c6c71e0e945a65a37ff414__calc_jac_lsoda", (DL_FUNC) rx_dbb1a66380c6c71e0e945a65a37ff414__calc_jac_lsoda);
#>   R_RegisterCCallable("rx_dbb1a66380c6c71e0e945a65a37ff414_","rx_dbb1a66380c6c71e0e945a65a37ff414__ode_solver_solvedata", (DL_FUNC) rx_dbb1a66380c6c71e0e945a65a37ff414__ode_solver_solvedata);
#>   R_RegisterCCallable("rx_dbb1a66380c6c71e0e945a65a37ff414_","rx_dbb1a66380c6c71e0e945a65a37ff414__ode_solver_get_solvedata", (DL_FUNC) rx_dbb1a66380c6c71e0e945a65a37ff414__ode_solver_get_solvedata);
#>   R_RegisterCCallable("rx_dbb1a66380c6c71e0e945a65a37ff414_","rx_dbb1a66380c6c71e0e945a65a37ff414__F", (DL_FUNC) rx_dbb1a66380c6c71e0e945a65a37ff414__F);
#>   R_RegisterCCallable("rx_dbb1a66380c6c71e0e945a65a37ff414_","rx_dbb1a66380c6c71e0e945a65a37ff414__Lag", (DL_FUNC) rx_dbb1a66380c6c71e0e945a65a37ff414__Lag);
#>   R_RegisterCCallable("rx_dbb1a66380c6c71e0e945a65a37ff414_","rx_dbb1a66380c6c71e0e945a65a37ff414__Rate", (DL_FUNC) rx_dbb1a66380c6c71e0e945a65a37ff414__Rate);
#>   R_RegisterCCallable("rx_dbb1a66380c6c71e0e945a65a37ff414_","rx_dbb1a66380c6c71e0e945a65a37ff414__Dur", (DL_FUNC) rx_dbb1a66380c6c71e0e945a65a37ff414__Dur);
#>   R_RegisterCCallable("rx_dbb1a66380c6c71e0e945a65a37ff414_","rx_dbb1a66380c6c71e0e945a65a37ff414__mtime", (DL_FUNC) rx_dbb1a66380c6c71e0e945a65a37ff414__mtime);
#>   R_RegisterCCallable("rx_dbb1a66380c6c71e0e945a65a37ff414_","rx_dbb1a66380c6c71e0e945a65a37ff414__ME", (DL_FUNC) rx_dbb1a66380c6c71e0e945a65a37ff414__ME);
#>   R_RegisterCCallable("rx_dbb1a66380c6c71e0e945a65a37ff414_","rx_dbb1a66380c6c71e0e945a65a37ff414__IndF", (DL_FUNC) rx_dbb1a66380c6c71e0e945a65a37ff414__IndF);
#>   R_RegisterCCallable("rx_dbb1a66380c6c71e0e945a65a37ff414_","rx_dbb1a66380c6c71e0e945a65a37ff414__dydt_liblsoda", (DL_FUNC) rx_dbb1a66380c6c71e0e945a65a37ff414__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_rx_dbb1a66380c6c71e0e945a65a37ff414_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_rx_dbb1a66380c6c71e0e945a65a37ff414_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"rx_dbb1a66380c6c71e0e945a65a37ff414__model_vars", (DL_FUNC) &rx_dbb1a66380c6c71e0e945a65a37ff414__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_rx_dbb1a66380c6c71e0e945a65a37ff414_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_dbb1a66380c6c71e0e945a65a37ff414__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("rx_dbb1a66380c6c71e0e945a65a37ff414__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’
# }