Simulate exponential variable from threefry generator

Care should be taken with this method not to encounter the birthday problem, described https://www.johndcook.com/blog/2016/01/29/random-number-generator-seed-mistakes/. Since the sitmo threefry, this currently generates one random deviate from the uniform distribution to seed the engine threefry and then run the code.

Usage

rxexp(rate, n = 1L, ncores = 1L)

Arguments

rate

vector of rates.

n

number of observations. If length(n) > 1, the length is taken to be the number required.

ncores

Number of cores for the simulation

rxnorm simulates using the threefry sitmo generator.

rxnormV used to simulate with the vandercorput simulator, but since it didn't satisfy the normal properties it was changed to simple be an alias of rxnorm. It is no longer supported in rxode2({}) blocks

Value

exponential random deviates

Details

Therefore, a simple call to the random number generated followed by a second call to random number generated may have identical seeds. As the number of random number generator calls are increased the probability that the birthday problem will increase.

The key to avoid this problem is to either run all simulations in the rxode2 environment once (therefore one seed or series of seeds for the whole simulation), pre-generate all random variables used for the simulation, or seed the rxode2 engine with rxSetSeed()

Internally each ID is seeded with a unique number so that the results do not depend on the number of cores used.

Examples

# \donttest{

## Use threefry engine

rxexp(0.5, n = 10) # with rxexp you have to explicitly state n
#>  [1] 0.05410404 2.25964412 0.40951127 4.24314765 0.58305288 0.56626127
#>  [7] 0.39817970 5.21608196 2.82629689 0.52366837
rxexp(5, n = 10, ncores = 2) # You can parallelize the simulation using openMP
#>  [1] 0.208739640 0.145190818 0.136720742 0.007704255 0.027296188 0.394830887
#>  [7] 0.491186947 0.184064949 0.265036398 0.312613464

rxexp(1)
#> [1] 2.190768

## This example uses `rxexp` directly in the model

rx <- function() {
  model({
    a <- rxexp(2)
  })
}

et <- et(1, id = 1:2)

s <- rxSolve(rx, et)
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#>  
#>  
#> C model:
#> // Define 1 LHS values
#> #define _LHS_0_ 0 // a
#> #define _getRxSolve_ _rx609cb368c568b97e818f77d02631ea876006__getRxSolve__R39J
#> #define _evalUdf _rx609cb368c568b97e818f77d02631ea876007__evalUdf_vcr1
#> #define _solveData _rx609cb368c568b97e818f77d02631ea876008__solveData_UmLF
#> #define _assign_ptr _rx609cb368c568b97e818f77d02631ea876009__assign_ptr_GL1z
#> #define _rxRmModelLib _rx609cb368c568b97e818f77d02631ea876010__rxRmModelLib_NFD8
#> #define _rxGetModelLib _rx609cb368c568b97e818f77d02631ea876011__rxGetModelLib_00Mr
#> #define _old_c _rx609cb368c568b97e818f77d02631ea876012__old_c_OclB
#> #define _ptrid _rx609cb368c568b97e818f77d02631ea876013__ptrid_dvcH
#> #define _rxIsCurrentC _rx609cb368c568b97e818f77d02631ea876014__rxIsCurrentC_YHlO
#> #define _sumPS _rx609cb368c568b97e818f77d02631ea876015__sumPS_Rzkg
#> #define _prodPS _rx609cb368c568b97e818f77d02631ea876016__prodPS_9phZ
#> #define _prodType _rx609cb368c568b97e818f77d02631ea876017__prodType_jbtv
#> #define _sumType _rx609cb368c568b97e818f77d02631ea876018__sumType_sFQk
#> #define _update_par_ptr _rx609cb368c568b97e818f77d02631ea876019__update_par_ptr_pVHF
#> #define _getParCov _rx609cb368c568b97e818f77d02631ea876020__getParCov_4tE5
#> #define _rxode2_rxAssignPtr _rx609cb368c568b97e818f77d02631ea876021__rxode2_rxAssignPtr_bqrW
#> #define _rxQr _rx609cb368c568b97e818f77d02631ea876022__rxQr_j2jI
#> #define _compareFactorVal _rx609cb368c568b97e818f77d02631ea876023__compareFactorVal_RW0c
#> #define _sum _rx609cb368c568b97e818f77d02631ea876024__sum_rrgc
#> #define _udf _rx609cb368c568b97e818f77d02631ea876025__udf_GmEs
#> #define _sign _rx609cb368c568b97e818f77d02631ea876026__sign_Ju9x
#> #define _prod _rx609cb368c568b97e818f77d02631ea876027__prod_ahgB
#> #define _max _rx609cb368c568b97e818f77d02631ea876028__max_OZl3
#> #define _min _rx609cb368c568b97e818f77d02631ea876029__min_zYvT
#> #define _transit4P _rx609cb368c568b97e818f77d02631ea876030__transit4P_a97E
#> #define _transit3P _rx609cb368c568b97e818f77d02631ea876031__transit3P_hW3M
#> #define _assignFuns0 _rx609cb368c568b97e818f77d02631ea876032__assignFuns0_G9V3
#> #define _assignFuns _rx609cb368c568b97e818f77d02631ea876033__assignFuns_pR1a
#> #define _rxord _rx609cb368c568b97e818f77d02631ea876034__rxord_NkG5
#> #define __assignFuns2 _rx609cb368c568b97e818f77d02631ea876035___assignFuns2_JwD9
#> #define _llikCauchyDscale _rx609cb368c568b97e818f77d02631ea876036__llikCauchyDscale_TK40
#> #define _llikCauchyDlocation _rx609cb368c568b97e818f77d02631ea876037__llikCauchyDlocation_NPJi
#> #define _llikCauchy _rx609cb368c568b97e818f77d02631ea876038__llikCauchy_ALZB
#> #define _llikGammaDrate _rx609cb368c568b97e818f77d02631ea876039__llikGammaDrate_rhef
#> #define _llikGammaDshape _rx609cb368c568b97e818f77d02631ea876040__llikGammaDshape_ghDr
#> #define _llikGamma _rx609cb368c568b97e818f77d02631ea876041__llikGamma_m85i
#> #define _llikWeibullDscale _rx609cb368c568b97e818f77d02631ea876042__llikWeibullDscale_gvlJ
#> #define _llikWeibullDshape _rx609cb368c568b97e818f77d02631ea876043__llikWeibullDshape_YgGq
#> #define _llikWeibull _rx609cb368c568b97e818f77d02631ea876044__llikWeibull_QxLo
#> #define _llikUnifDbeta _rx609cb368c568b97e818f77d02631ea876045__llikUnifDbeta_YlFg
#> #define _llikUnifDalpha _rx609cb368c568b97e818f77d02631ea876046__llikUnifDalpha_JdKa
#> #define _llikUnif _rx609cb368c568b97e818f77d02631ea876047__llikUnif_FgmD
#> #define _llikGeomDp _rx609cb368c568b97e818f77d02631ea876048__llikGeomDp_R4jx
#> #define _llikGeom _rx609cb368c568b97e818f77d02631ea876049__llikGeom_uoP8
#> #define _llikFDdf2 _rx609cb368c568b97e818f77d02631ea876050__llikFDdf2_gmoU
#> #define _llikFDdf1 _rx609cb368c568b97e818f77d02631ea876051__llikFDdf1_omwL
#> #define _llikF _rx609cb368c568b97e818f77d02631ea876052__llikF_tuDg
#> #define _llikExpDrate _rx609cb368c568b97e818f77d02631ea876053__llikExpDrate_OztP
#> #define _llikExp _rx609cb368c568b97e818f77d02631ea876054__llikExp_ZwDq
#> #define _llikChisqDdf _rx609cb368c568b97e818f77d02631ea876055__llikChisqDdf_x2t0
#> #define _llikChisq _rx609cb368c568b97e818f77d02631ea876056__llikChisq_FfCR
#> #define _llikTDsd _rx609cb368c568b97e818f77d02631ea876057__llikTDsd_upBf
#> #define _llikTDmean _rx609cb368c568b97e818f77d02631ea876058__llikTDmean_cLLl
#> #define _llikTDdf _rx609cb368c568b97e818f77d02631ea876059__llikTDdf_Jk0H
#> #define _llikT _rx609cb368c568b97e818f77d02631ea876060__llikT_Ouqu
#> #define _llikBetaDshape2 _rx609cb368c568b97e818f77d02631ea876061__llikBetaDshape2_MQIP
#> #define _llikBetaDshape1 _rx609cb368c568b97e818f77d02631ea876062__llikBetaDshape1_4sBK
#> #define _llikBeta _rx609cb368c568b97e818f77d02631ea876063__llikBeta_wlIO
#> #define _llikNbinomMuDmu _rx609cb368c568b97e818f77d02631ea876064__llikNbinomMuDmu_N1tT
#> #define _llikNbinomMu _rx609cb368c568b97e818f77d02631ea876065__llikNbinomMu_X5Du
#> #define _llikNbinomDprob _rx609cb368c568b97e818f77d02631ea876066__llikNbinomDprob_lHLx
#> #define _llikNbinom _rx609cb368c568b97e818f77d02631ea876067__llikNbinom_eAUt
#> #define _llikBinomDprob _rx609cb368c568b97e818f77d02631ea876068__llikBinomDprob_0po6
#> #define _llikBinom _rx609cb368c568b97e818f77d02631ea876069__llikBinom_AqSQ
#> #define _llikPoisDlambda _rx609cb368c568b97e818f77d02631ea876070__llikPoisDlambda_knH8
#> #define _llikPois _rx609cb368c568b97e818f77d02631ea876071__llikPois_fPCA
#> #define _llikNormDsd _rx609cb368c568b97e818f77d02631ea876072__llikNormDsd_nNkU
#> #define _llikNormDmean _rx609cb368c568b97e818f77d02631ea876073__llikNormDmean_mswi
#> #define _llikNorm _rx609cb368c568b97e818f77d02631ea876074__llikNorm_OWG1
#> #define simeps _rx609cb368c568b97e818f77d02631ea876075_simeps_6HPU
#> #define simeta _rx609cb368c568b97e818f77d02631ea876076_simeta_gnCj
#> #define expit _rx609cb368c568b97e818f77d02631ea876077_expit_uwA8
#> #define logit _rx609cb368c568b97e818f77d02631ea876078_logit_iwbn
#> #define gammapDer _rx609cb368c568b97e818f77d02631ea876079_gammapDer_2GTQ
#> #define lowergamma _rx609cb368c568b97e818f77d02631ea876080_lowergamma_jSre
#> #define uppergamma _rx609cb368c568b97e818f77d02631ea876081_uppergamma_r5ze
#> #define gammaqInva _rx609cb368c568b97e818f77d02631ea876082_gammaqInva_xLgA
#> #define gammaqInv _rx609cb368c568b97e818f77d02631ea876083_gammaqInv_wME6
#> #define gammapInva _rx609cb368c568b97e818f77d02631ea876084_gammapInva_nqR3
#> #define gammapInv _rx609cb368c568b97e818f77d02631ea876085_gammapInv_DD9j
#> #define gammaq _rx609cb368c568b97e818f77d02631ea876086_gammaq_N5O4
#> #define gammap _rx609cb368c568b97e818f77d02631ea876087_gammap_r9D9
#> #define phi _rx609cb368c568b97e818f77d02631ea876088_phi_7me5
#> #define d2ELUa _rx609cb368c568b97e818f77d02631ea876089_d2ELUa_ybV8
#> #define dELUa _rx609cb368c568b97e818f77d02631ea876090_dELUa_EuZ1
#> #define d2aELU _rx609cb368c568b97e818f77d02631ea876091_d2aELU_xTQY
#> #define d2ELU _rx609cb368c568b97e818f77d02631ea876092_d2ELU_OQe6
#> #define dELU _rx609cb368c568b97e818f77d02631ea876093_dELU_aqt3
#> #define ELU _rx609cb368c568b97e818f77d02631ea876094_ELU_Lbsb
#> #define dPReLUa1 _rx609cb368c568b97e818f77d02631ea876095_dPReLUa1_eaw5
#> #define dPReLUa _rx609cb368c568b97e818f77d02631ea876096_dPReLUa_m4br
#> #define dPReLU _rx609cb368c568b97e818f77d02631ea876097_dPReLU_D9EZ
#> #define PReLU _rx609cb368c568b97e818f77d02631ea876098_PReLU_AJNe
#> #define dSwish _rx609cb368c568b97e818f77d02631ea876099_dSwish_7IxU
#> #define Swish _rx609cb368c568b97e818f77d02631ea876100_Swish_FA5R
#> #define dlReLU _rx609cb368c568b97e818f77d02631ea876101_dlReLU_2wZw
#> #define lReLU _rx609cb368c568b97e818f77d02631ea876102_lReLU_wBgR
#> #define dSELU _rx609cb368c568b97e818f77d02631ea876103_dSELU_vbln
#> #define SELU _rx609cb368c568b97e818f77d02631ea876104_SELU_k357
#> #define d4softplus _rx609cb368c568b97e818f77d02631ea876105_d4softplus_cfUw
#> #define d3softplus _rx609cb368c568b97e818f77d02631ea876106_d3softplus_MarE
#> #define d2softplus _rx609cb368c568b97e818f77d02631ea876107_d2softplus_A9z1
#> #define dsoftplus _rx609cb368c568b97e818f77d02631ea876108_dsoftplus_SE1K
#> #define softplus _rx609cb368c568b97e818f77d02631ea876109_softplus_zAf2
#> #define d4GELU _rx609cb368c568b97e818f77d02631ea876110_d4GELU_H4Q2
#> #define d3GELU _rx609cb368c568b97e818f77d02631ea876111_d3GELU_oF0j
#> #define d2GELU _rx609cb368c568b97e818f77d02631ea876112_d2GELU_oCGj
#> #define dGELU _rx609cb368c568b97e818f77d02631ea876113_dGELU_0a8w
#> #define GELU _rx609cb368c568b97e818f77d02631ea876114_GELU_AbQE
#> #define dReLU _rx609cb368c568b97e818f77d02631ea876115_dReLU_LOlO
#> #define ReLU _rx609cb368c568b97e818f77d02631ea876116_ReLU_bqNa
#> #define riweibull _rx609cb368c568b97e818f77d02631ea876117_riweibull_thlt
#> #define riunif _rx609cb368c568b97e818f77d02631ea876118_riunif_iWR2
#> #define rit_ _rx609cb368c568b97e818f77d02631ea876119_rit__IGVL
#> #define ripois _rx609cb368c568b97e818f77d02631ea876120_ripois_QYM8
#> #define ribeta _rx609cb368c568b97e818f77d02631ea876121_ribeta_4JCn
#> #define rigamma _rx609cb368c568b97e818f77d02631ea876122_rigamma_uyCX
#> #define rigeom _rx609cb368c568b97e818f77d02631ea876123_rigeom_PJUj
#> #define rif _rx609cb368c568b97e818f77d02631ea876124_rif_vFk7
#> #define riexp _rx609cb368c568b97e818f77d02631ea876125_riexp_KJvD
#> #define richisq _rx609cb368c568b97e818f77d02631ea876126_richisq_xqbV
#> #define ricauchy _rx609cb368c568b97e818f77d02631ea876127_ricauchy_SKVf
#> #define rinbinomMu _rx609cb368c568b97e818f77d02631ea876128_rinbinomMu_r7Tc
#> #define rinbinom _rx609cb368c568b97e818f77d02631ea876129_rinbinom_aUEg
#> #define ribinom _rx609cb368c568b97e818f77d02631ea876130_ribinom_pYkN
#> #define rinorm _rx609cb368c568b97e818f77d02631ea876131_rinorm_IqnE
#> #define rxweibull _rx609cb368c568b97e818f77d02631ea876132_rxweibull_nmhS
#> #define rxunif _rx609cb368c568b97e818f77d02631ea876133_rxunif_boMZ
#> #define rxt_ _rx609cb368c568b97e818f77d02631ea876134_rxt__YBjk
#> #define rxpois _rx609cb368c568b97e818f77d02631ea876135_rxpois_iCoU
#> #define rxbeta _rx609cb368c568b97e818f77d02631ea876136_rxbeta_qvni
#> #define rxgamma _rx609cb368c568b97e818f77d02631ea876137_rxgamma_I75U
#> #define rxgeom _rx609cb368c568b97e818f77d02631ea876138_rxgeom_i0GV
#> #define rxf _rx609cb368c568b97e818f77d02631ea876139_rxf_Ts34
#> #define rxexp _rx609cb368c568b97e818f77d02631ea876140_rxexp_TNHe
#> #define rxchisq _rx609cb368c568b97e818f77d02631ea876141_rxchisq_Fpue
#> #define rxcauchy _rx609cb368c568b97e818f77d02631ea876142_rxcauchy_1ZaE
#> #define rxnbinomMu _rx609cb368c568b97e818f77d02631ea876143_rxnbinomMu_nJR6
#> #define rxnbinom _rx609cb368c568b97e818f77d02631ea876144_rxnbinom_UHZK
#> #define rxbinom _rx609cb368c568b97e818f77d02631ea876145_rxbinom_bSsm
#> #define rxnorm _rx609cb368c568b97e818f77d02631ea876146_rxnorm_NYlS
#> #define linCmtB _rx609cb368c568b97e818f77d02631ea876147_linCmtB_2wzY
#> #define linCmtA _rx609cb368c568b97e818f77d02631ea876148_linCmtA_qw2z
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){}
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){}
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  rx_0ebbc306248d0ec9a7605d5868898467__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *rx_0ebbc306248d0ec9a7605d5868898467__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP rx_0ebbc306248d0ec9a7605d5868898467__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void rx_0ebbc306248d0ec9a7605d5868898467__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double a = NA_REAL;
#> 
#>   double __udf[2];
#>   a = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   a =rxexp(&_solveData->subjects[_cSub], 2);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void rx_0ebbc306248d0ec9a7605d5868898467__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void rx_0ebbc306248d0ec9a7605d5868898467__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void rx_0ebbc306248d0ec9a7605d5868898467__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double a = NA_REAL;
#> 
#>   double __udf[2];
#>   a = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   a =rxexp(&_solveData->subjects[_cSub], 2);
#> 
#>   _lhs[_LHS_0_]=a;
#> }
#> // Functional based bioavailability
#> double rx_0ebbc306248d0ec9a7605d5868898467__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double rx_0ebbc306248d0ec9a7605d5868898467__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double rx_0ebbc306248d0ec9a7605d5868898467__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double rx_0ebbc306248d0ec9a7605d5868898467__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void rx_0ebbc306248d0ec9a7605d5868898467__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void rx_0ebbc306248d0ec9a7605d5868898467__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void rx_0ebbc306248d0ec9a7605d5868898467__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP rx_0ebbc306248d0ec9a7605d5868898467__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_0ebbc306248d0ec9a7605d5868898467__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAA[9zAAAv7#aT)em/*MA$)$+=WjmaB~15wu1tEOC~4/wO+A.<CY=}8IU@c+Msu\?P*M43j3Xx.SxR37ob>7:~|}eg_@*}p;:.FF8bkKOWoIZLQA`f$XQ9+bD.bwrnSw*912ly(lfg4m8Vj|yh..zc%%<7\?59Q|`F5$E[G78]bW<#HMY*npt&D(*>\"M6.L/O}P<0G*sH:h!feVmj#3QHNR$2/g6M;jQZbXE`a\?0k,A2UfySVbP9tEHEgph@&&FQ@0gmPH2F|v`Q4L$10#`hI4*g^7@~Dw(5Kx8YX5horzHks,g.sp)\?BiqQ4PS>+[GcRRfowvU$0ei!:uo[96JPpEaCK|UZJx]w*,5xjKO5NBzQW*e6^TDJhad\?IAeZEe0)41I\?q+eO+t~ur6lFbLbV)By2UJci[1W2z/4lrOj^LJFN$HccwY7Rwp&NwfKK+Ris8(5@A~VEvI\"xk1Iu*:%%\?._,:0\?wOq(K@5s.\"$gX4R6{_E{RHh#<e$7yDKpxp][v3\"W<N@6+~fnc`>U]7kf6wkP!J,eWzN=ngcaO=7kcZ=oe,lWUg_CHll@RfY$.evBrHAA&teA3!VKNwnLoAc^*F`Dc0A.Vc`).D&+eH#hSpyqta@@O(*dR]`)/e_5Ve:Z)`}m|5GbHGRm*CV>Q(X[><&1r`#$E6HJLJWrXH7[];iQ;%%ZwPdpN|]wrdn2;B{+DR$d^mOy1i3I\?=1>%%oP(eXG.z#>GE5,<:a:A#brnS{I)Wic+,Ko;,.Yo\"K.}kH[eGWOQ5kKPDnb{NeRtZ:i[Jsn(T\?J1E;E[kdxB44XWBfLbq^}|$\?9AbS@=Sg@<U~Z!R9MkS^e_\?:XPs0r0KeqHZ7}A!yJ2mN:6`.#z|%%zAv*Kg}^,7$n4xQ~CbMIjgEd*1_`%%M0RW}01kPM;+#x<jRvJ~{/aPx(8Wj\?W.Q*\?X*d}EnJqyDzAB(2NnmO]*t5tGMm7pvrx!!:wd=oV!\"u*|V5|yJ^M+>(OH13s>`v~WIf%%W)rOMfQyo#eP0iC2P.zL#RO%%WBj!YqJ+KkG");
#>     char buf[1059];
#> #define __doBufN__ 1059
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void rx_0ebbc306248d0ec9a7605d5868898467__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   rx_0ebbc306248d0ec9a7605d5868898467__dydt(neq, *t, A, DADT);
#> }
#> extern int rx_0ebbc306248d0ec9a7605d5868898467__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   rx_0ebbc306248d0ec9a7605d5868898467__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void rx_0ebbc306248d0ec9a7605d5868898467__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   rx_0ebbc306248d0ec9a7605d5868898467__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void rx_0ebbc306248d0ec9a7605d5868898467__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_rx_0ebbc306248d0ec9a7605d5868898467_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("rx_0ebbc306248d0ec9a7605d5868898467_","rx_0ebbc306248d0ec9a7605d5868898467__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("rx_0ebbc306248d0ec9a7605d5868898467_","rx_0ebbc306248d0ec9a7605d5868898467__assignFuns", (DL_FUNC) rx_0ebbc306248d0ec9a7605d5868898467__assignFuns);
#>   R_RegisterCCallable("rx_0ebbc306248d0ec9a7605d5868898467_","rx_0ebbc306248d0ec9a7605d5868898467__inis",(DL_FUNC) rx_0ebbc306248d0ec9a7605d5868898467__inis);
#>   R_RegisterCCallable("rx_0ebbc306248d0ec9a7605d5868898467_","rx_0ebbc306248d0ec9a7605d5868898467__dydt",(DL_FUNC) rx_0ebbc306248d0ec9a7605d5868898467__dydt);
#>   R_RegisterCCallable("rx_0ebbc306248d0ec9a7605d5868898467_","rx_0ebbc306248d0ec9a7605d5868898467__calc_lhs",(DL_FUNC) rx_0ebbc306248d0ec9a7605d5868898467__calc_lhs);
#>   R_RegisterCCallable("rx_0ebbc306248d0ec9a7605d5868898467_","rx_0ebbc306248d0ec9a7605d5868898467__calc_jac",(DL_FUNC) rx_0ebbc306248d0ec9a7605d5868898467__calc_jac);
#>   R_RegisterCCallable("rx_0ebbc306248d0ec9a7605d5868898467_","rx_0ebbc306248d0ec9a7605d5868898467__dydt_lsoda", (DL_FUNC) rx_0ebbc306248d0ec9a7605d5868898467__dydt_lsoda);
#>   R_RegisterCCallable("rx_0ebbc306248d0ec9a7605d5868898467_","rx_0ebbc306248d0ec9a7605d5868898467__calc_jac_lsoda", (DL_FUNC) rx_0ebbc306248d0ec9a7605d5868898467__calc_jac_lsoda);
#>   R_RegisterCCallable("rx_0ebbc306248d0ec9a7605d5868898467_","rx_0ebbc306248d0ec9a7605d5868898467__ode_solver_solvedata", (DL_FUNC) rx_0ebbc306248d0ec9a7605d5868898467__ode_solver_solvedata);
#>   R_RegisterCCallable("rx_0ebbc306248d0ec9a7605d5868898467_","rx_0ebbc306248d0ec9a7605d5868898467__ode_solver_get_solvedata", (DL_FUNC) rx_0ebbc306248d0ec9a7605d5868898467__ode_solver_get_solvedata);
#>   R_RegisterCCallable("rx_0ebbc306248d0ec9a7605d5868898467_","rx_0ebbc306248d0ec9a7605d5868898467__F", (DL_FUNC) rx_0ebbc306248d0ec9a7605d5868898467__F);
#>   R_RegisterCCallable("rx_0ebbc306248d0ec9a7605d5868898467_","rx_0ebbc306248d0ec9a7605d5868898467__Lag", (DL_FUNC) rx_0ebbc306248d0ec9a7605d5868898467__Lag);
#>   R_RegisterCCallable("rx_0ebbc306248d0ec9a7605d5868898467_","rx_0ebbc306248d0ec9a7605d5868898467__Rate", (DL_FUNC) rx_0ebbc306248d0ec9a7605d5868898467__Rate);
#>   R_RegisterCCallable("rx_0ebbc306248d0ec9a7605d5868898467_","rx_0ebbc306248d0ec9a7605d5868898467__Dur", (DL_FUNC) rx_0ebbc306248d0ec9a7605d5868898467__Dur);
#>   R_RegisterCCallable("rx_0ebbc306248d0ec9a7605d5868898467_","rx_0ebbc306248d0ec9a7605d5868898467__mtime", (DL_FUNC) rx_0ebbc306248d0ec9a7605d5868898467__mtime);
#>   R_RegisterCCallable("rx_0ebbc306248d0ec9a7605d5868898467_","rx_0ebbc306248d0ec9a7605d5868898467__ME", (DL_FUNC) rx_0ebbc306248d0ec9a7605d5868898467__ME);
#>   R_RegisterCCallable("rx_0ebbc306248d0ec9a7605d5868898467_","rx_0ebbc306248d0ec9a7605d5868898467__IndF", (DL_FUNC) rx_0ebbc306248d0ec9a7605d5868898467__IndF);
#>   R_RegisterCCallable("rx_0ebbc306248d0ec9a7605d5868898467_","rx_0ebbc306248d0ec9a7605d5868898467__dydt_liblsoda", (DL_FUNC) rx_0ebbc306248d0ec9a7605d5868898467__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_rx_0ebbc306248d0ec9a7605d5868898467_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_rx_0ebbc306248d0ec9a7605d5868898467_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"rx_0ebbc306248d0ec9a7605d5868898467__model_vars", (DL_FUNC) &rx_0ebbc306248d0ec9a7605d5868898467__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_rx_0ebbc306248d0ec9a7605d5868898467_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_0ebbc306248d0ec9a7605d5868898467__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("rx_0ebbc306248d0ec9a7605d5868898467__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’
# }