Simulate F variable from threefry generator

Care should be taken with this method not to encounter the birthday problem, described https://www.johndcook.com/blog/2016/01/29/random-number-generator-seed-mistakes/. Since the sitmo threefry, this currently generates one random deviate from the uniform distribution to seed the engine threefry and then run the code.

Usage

rxf(df1, df2, n = 1L, ncores = 1L)

Arguments

df1, df2

degrees of freedom. Inf is allowed.

n

number of observations. If length(n) > 1, the length is taken to be the number required.

ncores

Number of cores for the simulation

rxnorm simulates using the threefry sitmo generator.

rxnormV used to simulate with the vandercorput simulator, but since it didn't satisfy the normal properties it was changed to simple be an alias of rxnorm. It is no longer supported in rxode2({}) blocks

Value

f random deviates

Details

Therefore, a simple call to the random number generated followed by a second call to random number generated may have identical seeds. As the number of random number generator calls are increased the probability that the birthday problem will increase.

The key to avoid this problem is to either run all simulations in the rxode2 environment once (therefore one seed or series of seeds for the whole simulation), pre-generate all random variables used for the simulation, or seed the rxode2 engine with rxSetSeed()

Internally each ID is seeded with a unique number so that the results do not depend on the number of cores used.

Examples

# \donttest{

## Use threefry engine

rxf(0.5, 0.5, n = 10) # with rxf you have to explicitly state n
#>  [1] 2.268848e-06 1.315195e+00 1.918315e+02 5.195747e-05 1.602946e+00
#>  [6] 2.437402e+00 5.475269e+02 2.882414e+04 2.329001e-07 5.059613e+00
rxf(5, 1, n = 10, ncores = 2) # You can parallelize the simulation using openMP
#>  [1] 23.7578920 28.9085893  0.5882514  0.3202233 20.7032261  0.2102623
#>  [7]  6.0936919  0.7095009  0.6808133  1.2529923

rxf(1, 3)
#> [1] 0.8087568

## This example uses `rxf` directly in the model

rx <- function() {
  model({
    a <- rxf(2, 2)
  })
}

et <- et(1, id = 1:2)

s <- rxSolve(rx, et)
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#>  
#>  
#> C model:
#> // Define 1 LHS values
#> #define _LHS_0_ 0 // a
#> #define _getRxSolve_ _rx59d6a5739290b219e55861cb60f680c36149__getRxSolve__LFGL
#> #define _evalUdf _rx59d6a5739290b219e55861cb60f680c36150__evalUdf_1zNF
#> #define _solveData _rx59d6a5739290b219e55861cb60f680c36151__solveData_D800
#> #define _assign_ptr _rx59d6a5739290b219e55861cb60f680c36152__assign_ptr_MrOc
#> #define _rxRmModelLib _rx59d6a5739290b219e55861cb60f680c36153__rxRmModelLib_lBdv
#> #define _rxGetModelLib _rx59d6a5739290b219e55861cb60f680c36154__rxGetModelLib_cHYH
#> #define _old_c _rx59d6a5739290b219e55861cb60f680c36155__old_c_lORz
#> #define _ptrid _rx59d6a5739290b219e55861cb60f680c36156__ptrid_kg9p
#> #define _rxIsCurrentC _rx59d6a5739290b219e55861cb60f680c36157__rxIsCurrentC_hZjb
#> #define _sumPS _rx59d6a5739290b219e55861cb60f680c36158__sumPS_tvsF
#> #define _prodPS _rx59d6a5739290b219e55861cb60f680c36159__prodPS_QkpV
#> #define _prodType _rx59d6a5739290b219e55861cb60f680c36160__prodType_HF4t
#> #define _sumType _rx59d6a5739290b219e55861cb60f680c36161__sumType_E5bq
#> #define _update_par_ptr _rx59d6a5739290b219e55861cb60f680c36162__update_par_ptr_rWj2
#> #define _getParCov _rx59d6a5739290b219e55861cb60f680c36163__getParCov_jIRW
#> #define _rxode2_rxAssignPtr _rx59d6a5739290b219e55861cb60f680c36164__rxode2_rxAssignPtr_0crr
#> #define _rxQr _rx59d6a5739290b219e55861cb60f680c36165__rxQr_gcGm
#> #define _compareFactorVal _rx59d6a5739290b219e55861cb60f680c36166__compareFactorVal_EsJu
#> #define _sum _rx59d6a5739290b219e55861cb60f680c36167__sum_9xah
#> #define _udf _rx59d6a5739290b219e55861cb60f680c36168__udf_gBOZ
#> #define _sign _rx59d6a5739290b219e55861cb60f680c36169__sign_l3zY
#> #define _prod _rx59d6a5739290b219e55861cb60f680c36170__prod_vTa9
#> #define _max _rx59d6a5739290b219e55861cb60f680c36171__max_7EhW
#> #define _min _rx59d6a5739290b219e55861cb60f680c36172__min_3MG9
#> #define _transit4P _rx59d6a5739290b219e55861cb60f680c36173__transit4P_V3pR
#> #define _transit3P _rx59d6a5739290b219e55861cb60f680c36174__transit3P_1aNk
#> #define _assignFuns0 _rx59d6a5739290b219e55861cb60f680c36175__assignFuns0_G5Jw
#> #define _assignFuns _rx59d6a5739290b219e55861cb60f680c36176__assignFuns_D9TK
#> #define _rxord _rx59d6a5739290b219e55861cb60f680c36177__rxord_40NP
#> #define __assignFuns2 _rx59d6a5739290b219e55861cb60f680c36178___assignFuns2_JiAL
#> #define _llikCauchyDscale _rx59d6a5739290b219e55861cb60f680c36179__llikCauchyDscale_ZBrh
#> #define _llikCauchyDlocation _rx59d6a5739290b219e55861cb60f680c36180__llikCauchyDlocation_efgh
#> #define _llikCauchy _rx59d6a5739290b219e55861cb60f680c36181__llikCauchy_Drm8
#> #define _llikGammaDrate _rx59d6a5739290b219e55861cb60f680c36182__llikGammaDrate_5igv
#> #define _llikGammaDshape _rx59d6a5739290b219e55861cb60f680c36183__llikGammaDshape_lJYg
#> #define _llikGamma _rx59d6a5739290b219e55861cb60f680c36184__llikGamma_GqQx
#> #define _llikWeibullDscale _rx59d6a5739290b219e55861cb60f680c36185__llikWeibullDscale_LoYl
#> #define _llikWeibullDshape _rx59d6a5739290b219e55861cb60f680c36186__llikWeibullDshape_FgJd
#> #define _llikWeibull _rx59d6a5739290b219e55861cb60f680c36187__llikWeibull_KaFg
#> #define _llikUnifDbeta _rx59d6a5739290b219e55861cb60f680c36188__llikUnifDbeta_mDR4
#> #define _llikUnifDalpha _rx59d6a5739290b219e55861cb60f680c36189__llikUnifDalpha_jxuo
#> #define _llikUnif _rx59d6a5739290b219e55861cb60f680c36190__llikUnif_P8gm
#> #define _llikGeomDp _rx59d6a5739290b219e55861cb60f680c36191__llikGeomDp_oUom
#> #define _llikGeom _rx59d6a5739290b219e55861cb60f680c36192__llikGeom_wLtu
#> #define _llikFDdf2 _rx59d6a5739290b219e55861cb60f680c36193__llikFDdf2_DgOz
#> #define _llikFDdf1 _rx59d6a5739290b219e55861cb60f680c36194__llikFDdf1_tPZw
#> #define _llikF _rx59d6a5739290b219e55861cb60f680c36195__llikF_Dqx2
#> #define _llikExpDrate _rx59d6a5739290b219e55861cb60f680c36196__llikExpDrate_t0Ff
#> #define _llikExp _rx59d6a5739290b219e55861cb60f680c36197__llikExp_CRup
#> #define _llikChisqDdf _rx59d6a5739290b219e55861cb60f680c36198__llikChisqDdf_BfcL
#> #define _llikChisq _rx59d6a5739290b219e55861cb60f680c36199__llikChisq_LlJk
#> #define _llikTDsd _rx59d6a5739290b219e55861cb60f680c36200__llikTDsd_0HOu
#> #define _llikTDmean _rx59d6a5739290b219e55861cb60f680c36201__llikTDmean_quMQ
#> #define _llikTDdf _rx59d6a5739290b219e55861cb60f680c36202__llikTDdf_IP4s
#> #define _llikT _rx59d6a5739290b219e55861cb60f680c36203__llikT_BKwl
#> #define _llikBetaDshape2 _rx59d6a5739290b219e55861cb60f680c36204__llikBetaDshape2_ION1
#> #define _llikBetaDshape1 _rx59d6a5739290b219e55861cb60f680c36205__llikBetaDshape1_tTX5
#> #define _llikBeta _rx59d6a5739290b219e55861cb60f680c36206__llikBeta_DulH
#> #define _llikNbinomMuDmu _rx59d6a5739290b219e55861cb60f680c36207__llikNbinomMuDmu_LxeA
#> #define _llikNbinomMu _rx59d6a5739290b219e55861cb60f680c36208__llikNbinomMu_Ut0p
#> #define _llikNbinomDprob _rx59d6a5739290b219e55861cb60f680c36209__llikNbinomDprob_o6Aq
#> #define _llikNbinom _rx59d6a5739290b219e55861cb60f680c36210__llikNbinom_SQkn
#> #define _llikBinomDprob _rx59d6a5739290b219e55861cb60f680c36211__llikBinomDprob_H8fP
#> #define _llikBinom _rx59d6a5739290b219e55861cb60f680c36212__llikBinom_CAnN
#> #define _llikPoisDlambda _rx59d6a5739290b219e55861cb60f680c36213__llikPoisDlambda_kUms
#> #define _llikPois _rx59d6a5739290b219e55861cb60f680c36214__llikPois_wiOW
#> #define _llikNormDsd _rx59d6a5739290b219e55861cb60f680c36215__llikNormDsd_G16H
#> #define _llikNormDmean _rx59d6a5739290b219e55861cb60f680c36216__llikNormDmean_PUgn
#> #define _llikNorm _rx59d6a5739290b219e55861cb60f680c36217__llikNorm_Cjuw
#> #define simeps _rx59d6a5739290b219e55861cb60f680c36218_simeps_A8iw
#> #define simeta _rx59d6a5739290b219e55861cb60f680c36219_simeta_bn2G
#> #define expit _rx59d6a5739290b219e55861cb60f680c36220_expit_TQjS
#> #define logit _rx59d6a5739290b219e55861cb60f680c36221_logit_rer5
#> #define gammapDer _rx59d6a5739290b219e55861cb60f680c36222_gammapDer_zexL
#> #define lowergamma _rx59d6a5739290b219e55861cb60f680c36223_lowergamma_gAwM
#> #define uppergamma _rx59d6a5739290b219e55861cb60f680c36224_uppergamma_E6nq
#> #define gammaqInva _rx59d6a5739290b219e55861cb60f680c36225_gammaqInva_R3DD
#> #define gammaqInv _rx59d6a5739290b219e55861cb60f680c36226_gammaqInv_9jN5
#> #define gammapInva _rx59d6a5739290b219e55861cb60f680c36227_gammapInva_O4r9
#> #define gammapInv _rx59d6a5739290b219e55861cb60f680c36228_gammapInv_D97m
#> #define gammaq _rx59d6a5739290b219e55861cb60f680c36229_gammaq_e5yb
#> #define gammap _rx59d6a5739290b219e55861cb60f680c36230_gammap_V8Eu
#> #define phi _rx59d6a5739290b219e55861cb60f680c36231_phi_Z1xT
#> #define d2ELUa _rx59d6a5739290b219e55861cb60f680c36232_d2ELUa_QYOQ
#> #define dELUa _rx59d6a5739290b219e55861cb60f680c36233_dELUa_e6aq
#> #define d2aELU _rx59d6a5739290b219e55861cb60f680c36234_d2aELU_t3Lb
#> #define d2ELU _rx59d6a5739290b219e55861cb60f680c36235_d2ELU_sbea
#> #define dELU _rx59d6a5739290b219e55861cb60f680c36236_dELU_w5m4
#> #define ELU _rx59d6a5739290b219e55861cb60f680c36237_ELU_brD9
#> #define dPReLUa1 _rx59d6a5739290b219e55861cb60f680c36238_dPReLUa1_EZAJ
#> #define dPReLUa _rx59d6a5739290b219e55861cb60f680c36239_dPReLUa_Ne7I
#> #define dPReLU _rx59d6a5739290b219e55861cb60f680c36240_dPReLU_xUFA
#> #define PReLU _rx59d6a5739290b219e55861cb60f680c36241_PReLU_5R2w
#> #define dSwish _rx59d6a5739290b219e55861cb60f680c36242_dSwish_ZwwB
#> #define Swish _rx59d6a5739290b219e55861cb60f680c36243_Swish_gRvb
#> #define dlReLU _rx59d6a5739290b219e55861cb60f680c36244_dlReLU_lnk3
#> #define lReLU _rx59d6a5739290b219e55861cb60f680c36245_lReLU_57cf
#> #define dSELU _rx59d6a5739290b219e55861cb60f680c36246_dSELU_UwMa
#> #define SELU _rx59d6a5739290b219e55861cb60f680c36247_SELU_rEA9
#> #define d4softplus _rx59d6a5739290b219e55861cb60f680c36248_d4softplus_z1SE
#> #define d3softplus _rx59d6a5739290b219e55861cb60f680c36249_d3softplus_1KzA
#> #define d2softplus _rx59d6a5739290b219e55861cb60f680c36250_d2softplus_f2H4
#> #define dsoftplus _rx59d6a5739290b219e55861cb60f680c36251_dsoftplus_Q2oF
#> #define softplus _rx59d6a5739290b219e55861cb60f680c36252_softplus_0joC
#> #define d4GELU _rx59d6a5739290b219e55861cb60f680c36253_d4GELU_Gj0a
#> #define d3GELU _rx59d6a5739290b219e55861cb60f680c36254_d3GELU_8wAb
#> #define d2GELU _rx59d6a5739290b219e55861cb60f680c36255_d2GELU_QELO
#> #define dGELU _rx59d6a5739290b219e55861cb60f680c36256_dGELU_lObq
#> #define GELU _rx59d6a5739290b219e55861cb60f680c36257_GELU_Nath
#> #define dReLU _rx59d6a5739290b219e55861cb60f680c36258_dReLU_ltiW
#> #define ReLU _rx59d6a5739290b219e55861cb60f680c36259_ReLU_R2IG
#> #define riweibull _rx59d6a5739290b219e55861cb60f680c36260_riweibull_VLQY
#> #define riunif _rx59d6a5739290b219e55861cb60f680c36261_riunif_M84J
#> #define rit_ _rx59d6a5739290b219e55861cb60f680c36262_rit__Cnuy
#> #define ripois _rx59d6a5739290b219e55861cb60f680c36263_ripois_CXPJ
#> #define ribeta _rx59d6a5739290b219e55861cb60f680c36264_ribeta_UjvF
#> #define rigamma _rx59d6a5739290b219e55861cb60f680c36265_rigamma_k7KJ
#> #define rigeom _rx59d6a5739290b219e55861cb60f680c36266_rigeom_vDxq
#> #define rif _rx59d6a5739290b219e55861cb60f680c36267_rif_bVSK
#> #define riexp _rx59d6a5739290b219e55861cb60f680c36268_riexp_Vfr7
#> #define richisq _rx59d6a5739290b219e55861cb60f680c36269_richisq_TcaU
#> #define ricauchy _rx59d6a5739290b219e55861cb60f680c36270_ricauchy_EgpY
#> #define rinbinomMu _rx59d6a5739290b219e55861cb60f680c36271_rinbinomMu_kNIq
#> #define rinbinom _rx59d6a5739290b219e55861cb60f680c36272_rinbinom_nEnm
#> #define ribinom _rx59d6a5739290b219e55861cb60f680c36273_ribinom_hSbo
#> #define rinorm _rx59d6a5739290b219e55861cb60f680c36274_rinorm_MZYB
#> #define rxweibull _rx59d6a5739290b219e55861cb60f680c36275_rxweibull_jkiC
#> #define rxunif _rx59d6a5739290b219e55861cb60f680c36276_rxunif_oUqv
#> #define rxt_ _rx59d6a5739290b219e55861cb60f680c36277_rxt__niI7
#> #define rxpois _rx59d6a5739290b219e55861cb60f680c36278_rxpois_5Ui0
#> #define rxbeta _rx59d6a5739290b219e55861cb60f680c36279_rxbeta_GVTs
#> #define rxgamma _rx59d6a5739290b219e55861cb60f680c36280_rxgamma_34TN
#> #define rxgeom _rx59d6a5739290b219e55861cb60f680c36281_rxgeom_HeFp
#> #define rxf _rx59d6a5739290b219e55861cb60f680c36282_rxf_ue1Z
#> #define rxexp _rx59d6a5739290b219e55861cb60f680c36283_rxexp_aEnJ
#> #define rxchisq _rx59d6a5739290b219e55861cb60f680c36284_rxchisq_R6UH
#> #define rxcauchy _rx59d6a5739290b219e55861cb60f680c36285_rxcauchy_ZKbS
#> #define rxnbinomMu _rx59d6a5739290b219e55861cb60f680c36286_rxnbinomMu_smNY
#> #define rxnbinom _rx59d6a5739290b219e55861cb60f680c36287_rxnbinom_lS2w
#> #define rxbinom _rx59d6a5739290b219e55861cb60f680c36288_rxbinom_zYqw
#> #define rxnorm _rx59d6a5739290b219e55861cb60f680c36289_rxnorm_2z8b
#> #define linCmtB _rx59d6a5739290b219e55861cb60f680c36290_linCmtB_GMdj
#> #define linCmtA _rx59d6a5739290b219e55861cb60f680c36291_linCmtA_LB3L
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){}
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){}
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  rx_67a40bcab4e1e4a21ceb6d41f380b984__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *rx_67a40bcab4e1e4a21ceb6d41f380b984__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP rx_67a40bcab4e1e4a21ceb6d41f380b984__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void rx_67a40bcab4e1e4a21ceb6d41f380b984__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double a = NA_REAL;
#> 
#>   double __udf[2];
#>   a = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   a =rxf(&_solveData->subjects[_cSub], 2,(double)2);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void rx_67a40bcab4e1e4a21ceb6d41f380b984__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void rx_67a40bcab4e1e4a21ceb6d41f380b984__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void rx_67a40bcab4e1e4a21ceb6d41f380b984__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double a = NA_REAL;
#> 
#>   double __udf[2];
#>   a = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   a =rxf(&_solveData->subjects[_cSub], 2,(double)2);
#> 
#>   _lhs[_LHS_0_]=a;
#> }
#> // Functional based bioavailability
#> double rx_67a40bcab4e1e4a21ceb6d41f380b984__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double rx_67a40bcab4e1e4a21ceb6d41f380b984__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double rx_67a40bcab4e1e4a21ceb6d41f380b984__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double rx_67a40bcab4e1e4a21ceb6d41f380b984__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void rx_67a40bcab4e1e4a21ceb6d41f380b984__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void rx_67a40bcab4e1e4a21ceb6d41f380b984__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void rx_67a40bcab4e1e4a21ceb6d41f380b984__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP rx_67a40bcab4e1e4a21ceb6d41f380b984__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_67a40bcab4e1e4a21ceb6d41f380b984__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAA[9zAAAv7#aT)em/*MA$)$+=WjmaB~15wu1tEOC~4/wO+A.<CY=}8IU@c+Msu\?P*M43j3Xx.SxR37obWj(f3}0$F/u|[843LKd3.UOWoIZLQA`f$XQ9+bD.bwrnSw*912ly(lfg4m8Vj|yh..zc%%<7\?59Q|`F5$E[G78]bW;SIMY*npt&D(*>\"M6.L/O}P<4G*sH:h!femjj#3QHNR$2/g6M;jQ+E;YXkPu{pAT{n%%q/@1~Q+d%%Ykq`Vjv)yO^*1/>(Lf!GZL$10#`hI4*g^7@~Dw(5axlN8/!q)+E#Nlg.sp.\?njr9Bfe>+[GcRRjowvU$~fi!]uo[96/PqEaCK|UZg8O8Am;*Sy0c1ta14iPd|JwxQN8nG\"8MdOD8Yicuyz9cvB{D`a;K2W2qtDHTEROu:RrSK5W;{c)<XSKalPn7z,5EevVV=:`nH,r|3\"+,\"|rIDR|J,QPCu7m*0\?y7M*Gd]sT+Y~xB:o~s&0(*N:,cl/fh,6UOm:N<y5Fq|Bg3r3D~i(Q,q^z71jvO39&mJh`Ywm&nUPzm34ULi/YP8~>m$La60z*iC=PhHG[dAA>CfB&y][FZpBwDP#GX(PRd|R`W!FSP<Mgd)s,:R@}Pq13}Dd#6$FUh8z_eWN`.@%%.zYSdY&!EL>q#}}2kl[i=,*>u0<6WSs{TP/[];iQ;%%ZwPdpN|]wrdn2;B{+DR$d^mOy1i3I\?=1>%%oP(eXGdMGlfp;RX`_a`a39loM#~BsXCRp+WRQJ!B.k(ZtS%%y^XcHU^dZjnzV5w_XqIO\?/t](|>8k2mFZtiE}_I|J&\"P\?2`:\?p(AS(ym;fzi^[N(+!y1/5h6xYF7{Z_bqb\?bKx_\"(Umg$vav.O;S@=a~HFrf^4R1>#tQ&^LRxeymRTHb)&^uc+};fe49wVN*Sf1%%Ig{)ZQ9K\"et))r2fw)vv7`J>S]KFMBCpTb><c,x~YANY<b^D|Fjg7G7#[oi}Du`pY]LR<Xy$tdONV|&>F}tQ!ntt|cY!gJ[j9cO*ERfzNVkjcntC+hxw!\?iwP");
#>     char buf[1059];
#> #define __doBufN__ 1059
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void rx_67a40bcab4e1e4a21ceb6d41f380b984__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   rx_67a40bcab4e1e4a21ceb6d41f380b984__dydt(neq, *t, A, DADT);
#> }
#> extern int rx_67a40bcab4e1e4a21ceb6d41f380b984__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   rx_67a40bcab4e1e4a21ceb6d41f380b984__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void rx_67a40bcab4e1e4a21ceb6d41f380b984__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   rx_67a40bcab4e1e4a21ceb6d41f380b984__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void rx_67a40bcab4e1e4a21ceb6d41f380b984__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_rx_67a40bcab4e1e4a21ceb6d41f380b984_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("rx_67a40bcab4e1e4a21ceb6d41f380b984_","rx_67a40bcab4e1e4a21ceb6d41f380b984__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("rx_67a40bcab4e1e4a21ceb6d41f380b984_","rx_67a40bcab4e1e4a21ceb6d41f380b984__assignFuns", (DL_FUNC) rx_67a40bcab4e1e4a21ceb6d41f380b984__assignFuns);
#>   R_RegisterCCallable("rx_67a40bcab4e1e4a21ceb6d41f380b984_","rx_67a40bcab4e1e4a21ceb6d41f380b984__inis",(DL_FUNC) rx_67a40bcab4e1e4a21ceb6d41f380b984__inis);
#>   R_RegisterCCallable("rx_67a40bcab4e1e4a21ceb6d41f380b984_","rx_67a40bcab4e1e4a21ceb6d41f380b984__dydt",(DL_FUNC) rx_67a40bcab4e1e4a21ceb6d41f380b984__dydt);
#>   R_RegisterCCallable("rx_67a40bcab4e1e4a21ceb6d41f380b984_","rx_67a40bcab4e1e4a21ceb6d41f380b984__calc_lhs",(DL_FUNC) rx_67a40bcab4e1e4a21ceb6d41f380b984__calc_lhs);
#>   R_RegisterCCallable("rx_67a40bcab4e1e4a21ceb6d41f380b984_","rx_67a40bcab4e1e4a21ceb6d41f380b984__calc_jac",(DL_FUNC) rx_67a40bcab4e1e4a21ceb6d41f380b984__calc_jac);
#>   R_RegisterCCallable("rx_67a40bcab4e1e4a21ceb6d41f380b984_","rx_67a40bcab4e1e4a21ceb6d41f380b984__dydt_lsoda", (DL_FUNC) rx_67a40bcab4e1e4a21ceb6d41f380b984__dydt_lsoda);
#>   R_RegisterCCallable("rx_67a40bcab4e1e4a21ceb6d41f380b984_","rx_67a40bcab4e1e4a21ceb6d41f380b984__calc_jac_lsoda", (DL_FUNC) rx_67a40bcab4e1e4a21ceb6d41f380b984__calc_jac_lsoda);
#>   R_RegisterCCallable("rx_67a40bcab4e1e4a21ceb6d41f380b984_","rx_67a40bcab4e1e4a21ceb6d41f380b984__ode_solver_solvedata", (DL_FUNC) rx_67a40bcab4e1e4a21ceb6d41f380b984__ode_solver_solvedata);
#>   R_RegisterCCallable("rx_67a40bcab4e1e4a21ceb6d41f380b984_","rx_67a40bcab4e1e4a21ceb6d41f380b984__ode_solver_get_solvedata", (DL_FUNC) rx_67a40bcab4e1e4a21ceb6d41f380b984__ode_solver_get_solvedata);
#>   R_RegisterCCallable("rx_67a40bcab4e1e4a21ceb6d41f380b984_","rx_67a40bcab4e1e4a21ceb6d41f380b984__F", (DL_FUNC) rx_67a40bcab4e1e4a21ceb6d41f380b984__F);
#>   R_RegisterCCallable("rx_67a40bcab4e1e4a21ceb6d41f380b984_","rx_67a40bcab4e1e4a21ceb6d41f380b984__Lag", (DL_FUNC) rx_67a40bcab4e1e4a21ceb6d41f380b984__Lag);
#>   R_RegisterCCallable("rx_67a40bcab4e1e4a21ceb6d41f380b984_","rx_67a40bcab4e1e4a21ceb6d41f380b984__Rate", (DL_FUNC) rx_67a40bcab4e1e4a21ceb6d41f380b984__Rate);
#>   R_RegisterCCallable("rx_67a40bcab4e1e4a21ceb6d41f380b984_","rx_67a40bcab4e1e4a21ceb6d41f380b984__Dur", (DL_FUNC) rx_67a40bcab4e1e4a21ceb6d41f380b984__Dur);
#>   R_RegisterCCallable("rx_67a40bcab4e1e4a21ceb6d41f380b984_","rx_67a40bcab4e1e4a21ceb6d41f380b984__mtime", (DL_FUNC) rx_67a40bcab4e1e4a21ceb6d41f380b984__mtime);
#>   R_RegisterCCallable("rx_67a40bcab4e1e4a21ceb6d41f380b984_","rx_67a40bcab4e1e4a21ceb6d41f380b984__ME", (DL_FUNC) rx_67a40bcab4e1e4a21ceb6d41f380b984__ME);
#>   R_RegisterCCallable("rx_67a40bcab4e1e4a21ceb6d41f380b984_","rx_67a40bcab4e1e4a21ceb6d41f380b984__IndF", (DL_FUNC) rx_67a40bcab4e1e4a21ceb6d41f380b984__IndF);
#>   R_RegisterCCallable("rx_67a40bcab4e1e4a21ceb6d41f380b984_","rx_67a40bcab4e1e4a21ceb6d41f380b984__dydt_liblsoda", (DL_FUNC) rx_67a40bcab4e1e4a21ceb6d41f380b984__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_rx_67a40bcab4e1e4a21ceb6d41f380b984_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_rx_67a40bcab4e1e4a21ceb6d41f380b984_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"rx_67a40bcab4e1e4a21ceb6d41f380b984__model_vars", (DL_FUNC) &rx_67a40bcab4e1e4a21ceb6d41f380b984__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_rx_67a40bcab4e1e4a21ceb6d41f380b984_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_67a40bcab4e1e4a21ceb6d41f380b984__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("rx_67a40bcab4e1e4a21ceb6d41f380b984__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’
# }