Simulate geometric variable from threefry generator

Care should be taken with this method not to encounter the birthday problem, described https://www.johndcook.com/blog/2016/01/29/random-number-generator-seed-mistakes/. Since the sitmo threefry, this currently generates one random deviate from the uniform distribution to seed the engine threefry and then run the code.

Usage

rxgeom(prob, n = 1L, ncores = 1L)

Arguments

prob

probability of success in each trial. 0 < prob <= 1.

n

number of observations. If length(n) > 1, the length is taken to be the number required.

ncores

Number of cores for the simulation

rxnorm simulates using the threefry sitmo generator.

rxnormV used to simulate with the vandercorput simulator, but since it didn't satisfy the normal properties it was changed to simple be an alias of rxnorm. It is no longer supported in rxode2({}) blocks

Value

geometric random deviates

Details

Therefore, a simple call to the random number generated followed by a second call to random number generated may have identical seeds. As the number of random number generator calls are increased the probability that the birthday problem will increase.

The key to avoid this problem is to either run all simulations in the rxode2 environment once (therefore one seed or series of seeds for the whole simulation), pre-generate all random variables used for the simulation, or seed the rxode2 engine with rxSetSeed()

Internally each ID is seeded with a unique number so that the results do not depend on the number of cores used.

Examples

# \donttest{

## Use threefry engine

rxgeom(0.5, n = 10) # with rxgeom you have to explicitly state n
#>  [1] 0 0 0 0 3 0 3 0 1 0
rxgeom(0.25, n = 10, ncores = 2) # You can parallelize the simulation using openMP
#>  [1] 2 0 4 0 1 1 0 2 0 2

rxgeom(0.75)
#> [1] 0

## This example uses `rxgeom` directly in the model

rx <- function() {
  model({
    a <- rxgeom(0.24)
  })
}

et <- et(1, id = 1:2)

s <- rxSolve(rx, et)
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#>  
#>  
#> C model:
#> // Define 1 LHS values
#> #define _LHS_0_ 0 // a
#> #define _getRxSolve_ _rxf4eb35859ab32ea3bb429f73470be70e6435__getRxSolve__cHYH
#> #define _evalUdf _rxf4eb35859ab32ea3bb429f73470be70e6436__evalUdf_lORz
#> #define _solveData _rxf4eb35859ab32ea3bb429f73470be70e6437__solveData_kg9p
#> #define _assign_ptr _rxf4eb35859ab32ea3bb429f73470be70e6438__assign_ptr_hZjb
#> #define _rxRmModelLib _rxf4eb35859ab32ea3bb429f73470be70e6439__rxRmModelLib_tvsF
#> #define _rxGetModelLib _rxf4eb35859ab32ea3bb429f73470be70e6440__rxGetModelLib_QkpV
#> #define _old_c _rxf4eb35859ab32ea3bb429f73470be70e6441__old_c_HF4t
#> #define _ptrid _rxf4eb35859ab32ea3bb429f73470be70e6442__ptrid_E5bq
#> #define _rxIsCurrentC _rxf4eb35859ab32ea3bb429f73470be70e6443__rxIsCurrentC_rWj2
#> #define _sumPS _rxf4eb35859ab32ea3bb429f73470be70e6444__sumPS_jIRW
#> #define _prodPS _rxf4eb35859ab32ea3bb429f73470be70e6445__prodPS_0crr
#> #define _prodType _rxf4eb35859ab32ea3bb429f73470be70e6446__prodType_gcGm
#> #define _sumType _rxf4eb35859ab32ea3bb429f73470be70e6447__sumType_EsJu
#> #define _update_par_ptr _rxf4eb35859ab32ea3bb429f73470be70e6448__update_par_ptr_9xah
#> #define _getParCov _rxf4eb35859ab32ea3bb429f73470be70e6449__getParCov_gBOZ
#> #define _rxode2_rxAssignPtr _rxf4eb35859ab32ea3bb429f73470be70e6450__rxode2_rxAssignPtr_l3zY
#> #define _rxQr _rxf4eb35859ab32ea3bb429f73470be70e6451__rxQr_vTa9
#> #define _compareFactorVal _rxf4eb35859ab32ea3bb429f73470be70e6452__compareFactorVal_7EhW
#> #define _sum _rxf4eb35859ab32ea3bb429f73470be70e6453__sum_3MG9
#> #define _udf _rxf4eb35859ab32ea3bb429f73470be70e6454__udf_V3pR
#> #define _sign _rxf4eb35859ab32ea3bb429f73470be70e6455__sign_1aNk
#> #define _prod _rxf4eb35859ab32ea3bb429f73470be70e6456__prod_G5Jw
#> #define _max _rxf4eb35859ab32ea3bb429f73470be70e6457__max_D9TK
#> #define _min _rxf4eb35859ab32ea3bb429f73470be70e6458__min_40NP
#> #define _transit4P _rxf4eb35859ab32ea3bb429f73470be70e6459__transit4P_JiAL
#> #define _transit3P _rxf4eb35859ab32ea3bb429f73470be70e6460__transit3P_ZBrh
#> #define _assignFuns0 _rxf4eb35859ab32ea3bb429f73470be70e6461__assignFuns0_efgh
#> #define _assignFuns _rxf4eb35859ab32ea3bb429f73470be70e6462__assignFuns_Drm8
#> #define _rxord _rxf4eb35859ab32ea3bb429f73470be70e6463__rxord_5igv
#> #define __assignFuns2 _rxf4eb35859ab32ea3bb429f73470be70e6464___assignFuns2_lJYg
#> #define _llikCauchyDscale _rxf4eb35859ab32ea3bb429f73470be70e6465__llikCauchyDscale_GqQx
#> #define _llikCauchyDlocation _rxf4eb35859ab32ea3bb429f73470be70e6466__llikCauchyDlocation_LoYl
#> #define _llikCauchy _rxf4eb35859ab32ea3bb429f73470be70e6467__llikCauchy_FgJd
#> #define _llikGammaDrate _rxf4eb35859ab32ea3bb429f73470be70e6468__llikGammaDrate_KaFg
#> #define _llikGammaDshape _rxf4eb35859ab32ea3bb429f73470be70e6469__llikGammaDshape_mDR4
#> #define _llikGamma _rxf4eb35859ab32ea3bb429f73470be70e6470__llikGamma_jxuo
#> #define _llikWeibullDscale _rxf4eb35859ab32ea3bb429f73470be70e6471__llikWeibullDscale_P8gm
#> #define _llikWeibullDshape _rxf4eb35859ab32ea3bb429f73470be70e6472__llikWeibullDshape_oUom
#> #define _llikWeibull _rxf4eb35859ab32ea3bb429f73470be70e6473__llikWeibull_wLtu
#> #define _llikUnifDbeta _rxf4eb35859ab32ea3bb429f73470be70e6474__llikUnifDbeta_DgOz
#> #define _llikUnifDalpha _rxf4eb35859ab32ea3bb429f73470be70e6475__llikUnifDalpha_tPZw
#> #define _llikUnif _rxf4eb35859ab32ea3bb429f73470be70e6476__llikUnif_Dqx2
#> #define _llikGeomDp _rxf4eb35859ab32ea3bb429f73470be70e6477__llikGeomDp_t0Ff
#> #define _llikGeom _rxf4eb35859ab32ea3bb429f73470be70e6478__llikGeom_CRup
#> #define _llikFDdf2 _rxf4eb35859ab32ea3bb429f73470be70e6479__llikFDdf2_BfcL
#> #define _llikFDdf1 _rxf4eb35859ab32ea3bb429f73470be70e6480__llikFDdf1_LlJk
#> #define _llikF _rxf4eb35859ab32ea3bb429f73470be70e6481__llikF_0HOu
#> #define _llikExpDrate _rxf4eb35859ab32ea3bb429f73470be70e6482__llikExpDrate_quMQ
#> #define _llikExp _rxf4eb35859ab32ea3bb429f73470be70e6483__llikExp_IP4s
#> #define _llikChisqDdf _rxf4eb35859ab32ea3bb429f73470be70e6484__llikChisqDdf_BKwl
#> #define _llikChisq _rxf4eb35859ab32ea3bb429f73470be70e6485__llikChisq_ION1
#> #define _llikTDsd _rxf4eb35859ab32ea3bb429f73470be70e6486__llikTDsd_tTX5
#> #define _llikTDmean _rxf4eb35859ab32ea3bb429f73470be70e6487__llikTDmean_DulH
#> #define _llikTDdf _rxf4eb35859ab32ea3bb429f73470be70e6488__llikTDdf_LxeA
#> #define _llikT _rxf4eb35859ab32ea3bb429f73470be70e6489__llikT_Ut0p
#> #define _llikBetaDshape2 _rxf4eb35859ab32ea3bb429f73470be70e6490__llikBetaDshape2_o6Aq
#> #define _llikBetaDshape1 _rxf4eb35859ab32ea3bb429f73470be70e6491__llikBetaDshape1_SQkn
#> #define _llikBeta _rxf4eb35859ab32ea3bb429f73470be70e6492__llikBeta_H8fP
#> #define _llikNbinomMuDmu _rxf4eb35859ab32ea3bb429f73470be70e6493__llikNbinomMuDmu_CAnN
#> #define _llikNbinomMu _rxf4eb35859ab32ea3bb429f73470be70e6494__llikNbinomMu_kUms
#> #define _llikNbinomDprob _rxf4eb35859ab32ea3bb429f73470be70e6495__llikNbinomDprob_wiOW
#> #define _llikNbinom _rxf4eb35859ab32ea3bb429f73470be70e6496__llikNbinom_G16H
#> #define _llikBinomDprob _rxf4eb35859ab32ea3bb429f73470be70e6497__llikBinomDprob_PUgn
#> #define _llikBinom _rxf4eb35859ab32ea3bb429f73470be70e6498__llikBinom_Cjuw
#> #define _llikPoisDlambda _rxf4eb35859ab32ea3bb429f73470be70e6499__llikPoisDlambda_A8iw
#> #define _llikPois _rxf4eb35859ab32ea3bb429f73470be70e6500__llikPois_bn2G
#> #define _llikNormDsd _rxf4eb35859ab32ea3bb429f73470be70e6501__llikNormDsd_TQjS
#> #define _llikNormDmean _rxf4eb35859ab32ea3bb429f73470be70e6502__llikNormDmean_rer5
#> #define _llikNorm _rxf4eb35859ab32ea3bb429f73470be70e6503__llikNorm_zexL
#> #define simeps _rxf4eb35859ab32ea3bb429f73470be70e6504_simeps_gAwM
#> #define simeta _rxf4eb35859ab32ea3bb429f73470be70e6505_simeta_E6nq
#> #define expit _rxf4eb35859ab32ea3bb429f73470be70e6506_expit_R3DD
#> #define logit _rxf4eb35859ab32ea3bb429f73470be70e6507_logit_9jN5
#> #define gammapDer _rxf4eb35859ab32ea3bb429f73470be70e6508_gammapDer_O4r9
#> #define lowergamma _rxf4eb35859ab32ea3bb429f73470be70e6509_lowergamma_D97m
#> #define uppergamma _rxf4eb35859ab32ea3bb429f73470be70e6510_uppergamma_e5yb
#> #define gammaqInva _rxf4eb35859ab32ea3bb429f73470be70e6511_gammaqInva_V8Eu
#> #define gammaqInv _rxf4eb35859ab32ea3bb429f73470be70e6512_gammaqInv_Z1xT
#> #define gammapInva _rxf4eb35859ab32ea3bb429f73470be70e6513_gammapInva_QYOQ
#> #define gammapInv _rxf4eb35859ab32ea3bb429f73470be70e6514_gammapInv_e6aq
#> #define gammaq _rxf4eb35859ab32ea3bb429f73470be70e6515_gammaq_t3Lb
#> #define gammap _rxf4eb35859ab32ea3bb429f73470be70e6516_gammap_sbea
#> #define phi _rxf4eb35859ab32ea3bb429f73470be70e6517_phi_w5m4
#> #define d2ELUa _rxf4eb35859ab32ea3bb429f73470be70e6518_d2ELUa_brD9
#> #define dELUa _rxf4eb35859ab32ea3bb429f73470be70e6519_dELUa_EZAJ
#> #define d2aELU _rxf4eb35859ab32ea3bb429f73470be70e6520_d2aELU_Ne7I
#> #define d2ELU _rxf4eb35859ab32ea3bb429f73470be70e6521_d2ELU_xUFA
#> #define dELU _rxf4eb35859ab32ea3bb429f73470be70e6522_dELU_5R2w
#> #define ELU _rxf4eb35859ab32ea3bb429f73470be70e6523_ELU_ZwwB
#> #define dPReLUa1 _rxf4eb35859ab32ea3bb429f73470be70e6524_dPReLUa1_gRvb
#> #define dPReLUa _rxf4eb35859ab32ea3bb429f73470be70e6525_dPReLUa_lnk3
#> #define dPReLU _rxf4eb35859ab32ea3bb429f73470be70e6526_dPReLU_57cf
#> #define PReLU _rxf4eb35859ab32ea3bb429f73470be70e6527_PReLU_UwMa
#> #define dSwish _rxf4eb35859ab32ea3bb429f73470be70e6528_dSwish_rEA9
#> #define Swish _rxf4eb35859ab32ea3bb429f73470be70e6529_Swish_z1SE
#> #define dlReLU _rxf4eb35859ab32ea3bb429f73470be70e6530_dlReLU_1KzA
#> #define lReLU _rxf4eb35859ab32ea3bb429f73470be70e6531_lReLU_f2H4
#> #define dSELU _rxf4eb35859ab32ea3bb429f73470be70e6532_dSELU_Q2oF
#> #define SELU _rxf4eb35859ab32ea3bb429f73470be70e6533_SELU_0joC
#> #define d4softplus _rxf4eb35859ab32ea3bb429f73470be70e6534_d4softplus_Gj0a
#> #define d3softplus _rxf4eb35859ab32ea3bb429f73470be70e6535_d3softplus_8wAb
#> #define d2softplus _rxf4eb35859ab32ea3bb429f73470be70e6536_d2softplus_QELO
#> #define dsoftplus _rxf4eb35859ab32ea3bb429f73470be70e6537_dsoftplus_lObq
#> #define softplus _rxf4eb35859ab32ea3bb429f73470be70e6538_softplus_Nath
#> #define d4GELU _rxf4eb35859ab32ea3bb429f73470be70e6539_d4GELU_ltiW
#> #define d3GELU _rxf4eb35859ab32ea3bb429f73470be70e6540_d3GELU_R2IG
#> #define d2GELU _rxf4eb35859ab32ea3bb429f73470be70e6541_d2GELU_VLQY
#> #define dGELU _rxf4eb35859ab32ea3bb429f73470be70e6542_dGELU_M84J
#> #define GELU _rxf4eb35859ab32ea3bb429f73470be70e6543_GELU_Cnuy
#> #define dReLU _rxf4eb35859ab32ea3bb429f73470be70e6544_dReLU_CXPJ
#> #define ReLU _rxf4eb35859ab32ea3bb429f73470be70e6545_ReLU_UjvF
#> #define riweibull _rxf4eb35859ab32ea3bb429f73470be70e6546_riweibull_k7KJ
#> #define riunif _rxf4eb35859ab32ea3bb429f73470be70e6547_riunif_vDxq
#> #define rit_ _rxf4eb35859ab32ea3bb429f73470be70e6548_rit__bVSK
#> #define ripois _rxf4eb35859ab32ea3bb429f73470be70e6549_ripois_Vfr7
#> #define ribeta _rxf4eb35859ab32ea3bb429f73470be70e6550_ribeta_TcaU
#> #define rigamma _rxf4eb35859ab32ea3bb429f73470be70e6551_rigamma_EgpY
#> #define rigeom _rxf4eb35859ab32ea3bb429f73470be70e6552_rigeom_kNIq
#> #define rif _rxf4eb35859ab32ea3bb429f73470be70e6553_rif_nEnm
#> #define riexp _rxf4eb35859ab32ea3bb429f73470be70e6554_riexp_hSbo
#> #define richisq _rxf4eb35859ab32ea3bb429f73470be70e6555_richisq_MZYB
#> #define ricauchy _rxf4eb35859ab32ea3bb429f73470be70e6556_ricauchy_jkiC
#> #define rinbinomMu _rxf4eb35859ab32ea3bb429f73470be70e6557_rinbinomMu_oUqv
#> #define rinbinom _rxf4eb35859ab32ea3bb429f73470be70e6558_rinbinom_niI7
#> #define ribinom _rxf4eb35859ab32ea3bb429f73470be70e6559_ribinom_5Ui0
#> #define rinorm _rxf4eb35859ab32ea3bb429f73470be70e6560_rinorm_GVTs
#> #define rxweibull _rxf4eb35859ab32ea3bb429f73470be70e6561_rxweibull_34TN
#> #define rxunif _rxf4eb35859ab32ea3bb429f73470be70e6562_rxunif_HeFp
#> #define rxt_ _rxf4eb35859ab32ea3bb429f73470be70e6563_rxt__ue1Z
#> #define rxpois _rxf4eb35859ab32ea3bb429f73470be70e6564_rxpois_aEnJ
#> #define rxbeta _rxf4eb35859ab32ea3bb429f73470be70e6565_rxbeta_R6UH
#> #define rxgamma _rxf4eb35859ab32ea3bb429f73470be70e6566_rxgamma_ZKbS
#> #define rxgeom _rxf4eb35859ab32ea3bb429f73470be70e6567_rxgeom_smNY
#> #define rxf _rxf4eb35859ab32ea3bb429f73470be70e6568_rxf_lS2w
#> #define rxexp _rxf4eb35859ab32ea3bb429f73470be70e6569_rxexp_zYqw
#> #define rxchisq _rxf4eb35859ab32ea3bb429f73470be70e6570_rxchisq_2z8b
#> #define rxcauchy _rxf4eb35859ab32ea3bb429f73470be70e6571_rxcauchy_GMdj
#> #define rxnbinomMu _rxf4eb35859ab32ea3bb429f73470be70e6572_rxnbinomMu_LB3L
#> #define rxnbinom _rxf4eb35859ab32ea3bb429f73470be70e6573_rxnbinom_sOUR
#> #define rxbinom _rxf4eb35859ab32ea3bb429f73470be70e6574_rxbinom_bPYU
#> #define rxnorm _rxf4eb35859ab32ea3bb429f73470be70e6575_rxnorm_PmW8
#> #define linCmtB _rxf4eb35859ab32ea3bb429f73470be70e6576_linCmtB_sOmG
#> #define linCmtA _rxf4eb35859ab32ea3bb429f73470be70e6577_linCmtA_ypO5
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){}
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){}
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  rx_d5f4a6a811173ed3f50ca51782347c59__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *rx_d5f4a6a811173ed3f50ca51782347c59__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP rx_d5f4a6a811173ed3f50ca51782347c59__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void rx_d5f4a6a811173ed3f50ca51782347c59__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double a = NA_REAL;
#> 
#>   double __udf[2];
#>   a = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   a =(double)rxgeom(&_solveData->subjects[_cSub], 0.24);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void rx_d5f4a6a811173ed3f50ca51782347c59__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void rx_d5f4a6a811173ed3f50ca51782347c59__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void rx_d5f4a6a811173ed3f50ca51782347c59__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double a = NA_REAL;
#> 
#>   double __udf[2];
#>   a = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   a =(double)rxgeom(&_solveData->subjects[_cSub], 0.24);
#> 
#>   _lhs[_LHS_0_]=a;
#> }
#> // Functional based bioavailability
#> double rx_d5f4a6a811173ed3f50ca51782347c59__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double rx_d5f4a6a811173ed3f50ca51782347c59__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double rx_d5f4a6a811173ed3f50ca51782347c59__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double rx_d5f4a6a811173ed3f50ca51782347c59__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void rx_d5f4a6a811173ed3f50ca51782347c59__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void rx_d5f4a6a811173ed3f50ca51782347c59__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void rx_d5f4a6a811173ed3f50ca51782347c59__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP rx_d5f4a6a811173ed3f50ca51782347c59__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_d5f4a6a811173ed3f50ca51782347c59__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAA+>zAAAv7#aT)mmC:MAK+(Z=WjmaB~15wu1tE9*u/\"OYSASY4}HWPCFV0:IX0=N>5{Gnd9/gvAFi2s0}j3!M=<B[,.Q}PG.oc~M7AFthAnB}Y=6Y//])/u`|tqI6Kp8/@35{<I1P=Tb6}QSl[#1!g,nr;1rkqd=w`gdf{O4_<T+d!\?=4QZ_#$3qP#K#y~E$cV#hCpD<Q8eX4<<FiMP$@.g6LeMFZbsN\"(pf,d++]!h;r2r8xDIHUV|x4,m$}I&oe5`pzXtt{C}!h}KCMJz*4U!Tcc88;Zrz>8PZQoFP@H3`ofn|EDjeoHZny!BH:,Of@LG9`G2P<n[r=:qjyAwn;sXVd7NHGw7*EYO5Wx/:r`UaU<(r%%+kM:C*tW5Rbp|Iv3GJPfMzD~JUw<St_Q@JQD%%6\"`BENEHG5W;Vc)<XS*ZsKPX@PUN0*z!(x\?IS(3(pt|o|QOi^TzheE\"crwM*jcXw(4/~#xv}mZ6[$>o8pxZ6{z.4#&&Y=kpNZ!IZKn!CL;hTI{s~XN:#V65zv&i&`$iqe1I$W*[8S$(u]slXq30}o%%7uJ1pX.CK<f`4!UkDAAUvNAoK+7RUKBycF4.oC|!(,=,[B%%Wm&\"yYs~}Xa%%ykp/hi_[O(2jV>q.X#hgv9Vu^Jbi.|X#+p3l[%%_RiQ6yxkRAtzcy~}.\?;F6240h)ts<gXd1cGuhV>ts7$;5W;i#pkNO1){]2C(+l^,9`==*)98qEeB<XaP.7*{%%fn.\?;j7S%%PvvsE4+c4CS>/M.vW7FWd:!aLw(X{qO#F>hyuSDzB.t74h$TK}vL]]j;ne+}kPg`e7_Ax^C5$jL,Fhqxz>\"0wvVElOqD#Lsj)Yx.^OuYFk{pnh!G@,f7WIC!]&6Yl_[x}vFf\?eFmhLblx!5FJftHwH[(:E\"gS{0D;Iw*Y(g<Z$>vBsK5VA_&Qts6:E*3!;h=Meg`[|dkLQ#0nGq[JD3%%_;pvH]~(D%%UFlwj0!CSnHj.fWa1`CT0aPV96SU.6Sg6O>e7>/$P0is8]5H+pQ{a!+~I9S,PVBEA");
#>     char buf[1065];
#> #define __doBufN__ 1065
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void rx_d5f4a6a811173ed3f50ca51782347c59__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   rx_d5f4a6a811173ed3f50ca51782347c59__dydt(neq, *t, A, DADT);
#> }
#> extern int rx_d5f4a6a811173ed3f50ca51782347c59__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   rx_d5f4a6a811173ed3f50ca51782347c59__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void rx_d5f4a6a811173ed3f50ca51782347c59__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   rx_d5f4a6a811173ed3f50ca51782347c59__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void rx_d5f4a6a811173ed3f50ca51782347c59__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_rx_d5f4a6a811173ed3f50ca51782347c59_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("rx_d5f4a6a811173ed3f50ca51782347c59_","rx_d5f4a6a811173ed3f50ca51782347c59__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("rx_d5f4a6a811173ed3f50ca51782347c59_","rx_d5f4a6a811173ed3f50ca51782347c59__assignFuns", (DL_FUNC) rx_d5f4a6a811173ed3f50ca51782347c59__assignFuns);
#>   R_RegisterCCallable("rx_d5f4a6a811173ed3f50ca51782347c59_","rx_d5f4a6a811173ed3f50ca51782347c59__inis",(DL_FUNC) rx_d5f4a6a811173ed3f50ca51782347c59__inis);
#>   R_RegisterCCallable("rx_d5f4a6a811173ed3f50ca51782347c59_","rx_d5f4a6a811173ed3f50ca51782347c59__dydt",(DL_FUNC) rx_d5f4a6a811173ed3f50ca51782347c59__dydt);
#>   R_RegisterCCallable("rx_d5f4a6a811173ed3f50ca51782347c59_","rx_d5f4a6a811173ed3f50ca51782347c59__calc_lhs",(DL_FUNC) rx_d5f4a6a811173ed3f50ca51782347c59__calc_lhs);
#>   R_RegisterCCallable("rx_d5f4a6a811173ed3f50ca51782347c59_","rx_d5f4a6a811173ed3f50ca51782347c59__calc_jac",(DL_FUNC) rx_d5f4a6a811173ed3f50ca51782347c59__calc_jac);
#>   R_RegisterCCallable("rx_d5f4a6a811173ed3f50ca51782347c59_","rx_d5f4a6a811173ed3f50ca51782347c59__dydt_lsoda", (DL_FUNC) rx_d5f4a6a811173ed3f50ca51782347c59__dydt_lsoda);
#>   R_RegisterCCallable("rx_d5f4a6a811173ed3f50ca51782347c59_","rx_d5f4a6a811173ed3f50ca51782347c59__calc_jac_lsoda", (DL_FUNC) rx_d5f4a6a811173ed3f50ca51782347c59__calc_jac_lsoda);
#>   R_RegisterCCallable("rx_d5f4a6a811173ed3f50ca51782347c59_","rx_d5f4a6a811173ed3f50ca51782347c59__ode_solver_solvedata", (DL_FUNC) rx_d5f4a6a811173ed3f50ca51782347c59__ode_solver_solvedata);
#>   R_RegisterCCallable("rx_d5f4a6a811173ed3f50ca51782347c59_","rx_d5f4a6a811173ed3f50ca51782347c59__ode_solver_get_solvedata", (DL_FUNC) rx_d5f4a6a811173ed3f50ca51782347c59__ode_solver_get_solvedata);
#>   R_RegisterCCallable("rx_d5f4a6a811173ed3f50ca51782347c59_","rx_d5f4a6a811173ed3f50ca51782347c59__F", (DL_FUNC) rx_d5f4a6a811173ed3f50ca51782347c59__F);
#>   R_RegisterCCallable("rx_d5f4a6a811173ed3f50ca51782347c59_","rx_d5f4a6a811173ed3f50ca51782347c59__Lag", (DL_FUNC) rx_d5f4a6a811173ed3f50ca51782347c59__Lag);
#>   R_RegisterCCallable("rx_d5f4a6a811173ed3f50ca51782347c59_","rx_d5f4a6a811173ed3f50ca51782347c59__Rate", (DL_FUNC) rx_d5f4a6a811173ed3f50ca51782347c59__Rate);
#>   R_RegisterCCallable("rx_d5f4a6a811173ed3f50ca51782347c59_","rx_d5f4a6a811173ed3f50ca51782347c59__Dur", (DL_FUNC) rx_d5f4a6a811173ed3f50ca51782347c59__Dur);
#>   R_RegisterCCallable("rx_d5f4a6a811173ed3f50ca51782347c59_","rx_d5f4a6a811173ed3f50ca51782347c59__mtime", (DL_FUNC) rx_d5f4a6a811173ed3f50ca51782347c59__mtime);
#>   R_RegisterCCallable("rx_d5f4a6a811173ed3f50ca51782347c59_","rx_d5f4a6a811173ed3f50ca51782347c59__ME", (DL_FUNC) rx_d5f4a6a811173ed3f50ca51782347c59__ME);
#>   R_RegisterCCallable("rx_d5f4a6a811173ed3f50ca51782347c59_","rx_d5f4a6a811173ed3f50ca51782347c59__IndF", (DL_FUNC) rx_d5f4a6a811173ed3f50ca51782347c59__IndF);
#>   R_RegisterCCallable("rx_d5f4a6a811173ed3f50ca51782347c59_","rx_d5f4a6a811173ed3f50ca51782347c59__dydt_liblsoda", (DL_FUNC) rx_d5f4a6a811173ed3f50ca51782347c59__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_rx_d5f4a6a811173ed3f50ca51782347c59_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_rx_d5f4a6a811173ed3f50ca51782347c59_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"rx_d5f4a6a811173ed3f50ca51782347c59__model_vars", (DL_FUNC) &rx_d5f4a6a811173ed3f50ca51782347c59__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_rx_d5f4a6a811173ed3f50ca51782347c59_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_d5f4a6a811173ed3f50ca51782347c59__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("rx_d5f4a6a811173ed3f50ca51782347c59__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’
# }