Simulate random normal variable from threefry generator

Simulate random normal variable from threefry generator

Usage

rxnormV(mean = 0, sd = 1, n = 1L, ncores = 1L)

rxnorm(mean = 0, sd = 1, n = 1L, ncores = 1L)

Arguments

mean

vector of means.

sd

vector of standard deviations.

n

number of observations

ncores

Number of cores for the simulation

rxnorm simulates using the threefry sitmo generator.

rxnormV used to simulate with the vandercorput simulator, but since it didn't satisfy the normal properties it was changed to simple be an alias of rxnorm. It is no longer supported in rxode2({}) blocks

Value

normal random number deviates

Examples

# \donttest{
## Use threefry engine

rxnorm(n = 10) # with rxnorm you have to explicitly state n
#>  [1]  0.172888709  0.284784040 -0.005692539 -0.421214620  0.452990248
#>  [6]  0.012650216 -0.934263347 -1.208389993  0.298874284 -0.397359774
rxnorm(n = 10, ncores = 2) # You can parallelize the simulation using openMP
#>  [1]  0.06786949 -0.28052929 -0.66690901  0.33965344 -1.16260363  0.25713640
#>  [7] -1.12190228  0.79853569 -0.53656934  0.07836036

rxnorm(2, 3) ## The first 2 arguments are the mean and standard deviation
#> [1] -0.7994892

## This example uses `rxnorm` directly in the model

rx <- function() {
  model({
    a <- rxnorm()
  })
}

et <- et(1, id = 1:2)

s <- rxSolve(rx, et)
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#>  
#>  
#> C model:
#> // Define 1 LHS values
#> #define _LHS_0_ 0 // a
#> #define _getRxSolve_ _rxdac23bc34facc9a82bfdee69bf82d91e6864__getRxSolve__HF4t
#> #define _evalUdf _rxdac23bc34facc9a82bfdee69bf82d91e6865__evalUdf_E5bq
#> #define _solveData _rxdac23bc34facc9a82bfdee69bf82d91e6866__solveData_rWj2
#> #define _assign_ptr _rxdac23bc34facc9a82bfdee69bf82d91e6867__assign_ptr_jIRW
#> #define _rxRmModelLib _rxdac23bc34facc9a82bfdee69bf82d91e6868__rxRmModelLib_0crr
#> #define _rxGetModelLib _rxdac23bc34facc9a82bfdee69bf82d91e6869__rxGetModelLib_gcGm
#> #define _old_c _rxdac23bc34facc9a82bfdee69bf82d91e6870__old_c_EsJu
#> #define _ptrid _rxdac23bc34facc9a82bfdee69bf82d91e6871__ptrid_9xah
#> #define _rxIsCurrentC _rxdac23bc34facc9a82bfdee69bf82d91e6872__rxIsCurrentC_gBOZ
#> #define _sumPS _rxdac23bc34facc9a82bfdee69bf82d91e6873__sumPS_l3zY
#> #define _prodPS _rxdac23bc34facc9a82bfdee69bf82d91e6874__prodPS_vTa9
#> #define _prodType _rxdac23bc34facc9a82bfdee69bf82d91e6875__prodType_7EhW
#> #define _sumType _rxdac23bc34facc9a82bfdee69bf82d91e6876__sumType_3MG9
#> #define _update_par_ptr _rxdac23bc34facc9a82bfdee69bf82d91e6877__update_par_ptr_V3pR
#> #define _getParCov _rxdac23bc34facc9a82bfdee69bf82d91e6878__getParCov_1aNk
#> #define _rxode2_rxAssignPtr _rxdac23bc34facc9a82bfdee69bf82d91e6879__rxode2_rxAssignPtr_G5Jw
#> #define _rxQr _rxdac23bc34facc9a82bfdee69bf82d91e6880__rxQr_D9TK
#> #define _compareFactorVal _rxdac23bc34facc9a82bfdee69bf82d91e6881__compareFactorVal_40NP
#> #define _sum _rxdac23bc34facc9a82bfdee69bf82d91e6882__sum_JiAL
#> #define _udf _rxdac23bc34facc9a82bfdee69bf82d91e6883__udf_ZBrh
#> #define _sign _rxdac23bc34facc9a82bfdee69bf82d91e6884__sign_efgh
#> #define _prod _rxdac23bc34facc9a82bfdee69bf82d91e6885__prod_Drm8
#> #define _max _rxdac23bc34facc9a82bfdee69bf82d91e6886__max_5igv
#> #define _min _rxdac23bc34facc9a82bfdee69bf82d91e6887__min_lJYg
#> #define _transit4P _rxdac23bc34facc9a82bfdee69bf82d91e6888__transit4P_GqQx
#> #define _transit3P _rxdac23bc34facc9a82bfdee69bf82d91e6889__transit3P_LoYl
#> #define _assignFuns0 _rxdac23bc34facc9a82bfdee69bf82d91e6890__assignFuns0_FgJd
#> #define _assignFuns _rxdac23bc34facc9a82bfdee69bf82d91e6891__assignFuns_KaFg
#> #define _rxord _rxdac23bc34facc9a82bfdee69bf82d91e6892__rxord_mDR4
#> #define __assignFuns2 _rxdac23bc34facc9a82bfdee69bf82d91e6893___assignFuns2_jxuo
#> #define _llikCauchyDscale _rxdac23bc34facc9a82bfdee69bf82d91e6894__llikCauchyDscale_P8gm
#> #define _llikCauchyDlocation _rxdac23bc34facc9a82bfdee69bf82d91e6895__llikCauchyDlocation_oUom
#> #define _llikCauchy _rxdac23bc34facc9a82bfdee69bf82d91e6896__llikCauchy_wLtu
#> #define _llikGammaDrate _rxdac23bc34facc9a82bfdee69bf82d91e6897__llikGammaDrate_DgOz
#> #define _llikGammaDshape _rxdac23bc34facc9a82bfdee69bf82d91e6898__llikGammaDshape_tPZw
#> #define _llikGamma _rxdac23bc34facc9a82bfdee69bf82d91e6899__llikGamma_Dqx2
#> #define _llikWeibullDscale _rxdac23bc34facc9a82bfdee69bf82d91e6900__llikWeibullDscale_t0Ff
#> #define _llikWeibullDshape _rxdac23bc34facc9a82bfdee69bf82d91e6901__llikWeibullDshape_CRup
#> #define _llikWeibull _rxdac23bc34facc9a82bfdee69bf82d91e6902__llikWeibull_BfcL
#> #define _llikUnifDbeta _rxdac23bc34facc9a82bfdee69bf82d91e6903__llikUnifDbeta_LlJk
#> #define _llikUnifDalpha _rxdac23bc34facc9a82bfdee69bf82d91e6904__llikUnifDalpha_0HOu
#> #define _llikUnif _rxdac23bc34facc9a82bfdee69bf82d91e6905__llikUnif_quMQ
#> #define _llikGeomDp _rxdac23bc34facc9a82bfdee69bf82d91e6906__llikGeomDp_IP4s
#> #define _llikGeom _rxdac23bc34facc9a82bfdee69bf82d91e6907__llikGeom_BKwl
#> #define _llikFDdf2 _rxdac23bc34facc9a82bfdee69bf82d91e6908__llikFDdf2_ION1
#> #define _llikFDdf1 _rxdac23bc34facc9a82bfdee69bf82d91e6909__llikFDdf1_tTX5
#> #define _llikF _rxdac23bc34facc9a82bfdee69bf82d91e6910__llikF_DulH
#> #define _llikExpDrate _rxdac23bc34facc9a82bfdee69bf82d91e6911__llikExpDrate_LxeA
#> #define _llikExp _rxdac23bc34facc9a82bfdee69bf82d91e6912__llikExp_Ut0p
#> #define _llikChisqDdf _rxdac23bc34facc9a82bfdee69bf82d91e6913__llikChisqDdf_o6Aq
#> #define _llikChisq _rxdac23bc34facc9a82bfdee69bf82d91e6914__llikChisq_SQkn
#> #define _llikTDsd _rxdac23bc34facc9a82bfdee69bf82d91e6915__llikTDsd_H8fP
#> #define _llikTDmean _rxdac23bc34facc9a82bfdee69bf82d91e6916__llikTDmean_CAnN
#> #define _llikTDdf _rxdac23bc34facc9a82bfdee69bf82d91e6917__llikTDdf_kUms
#> #define _llikT _rxdac23bc34facc9a82bfdee69bf82d91e6918__llikT_wiOW
#> #define _llikBetaDshape2 _rxdac23bc34facc9a82bfdee69bf82d91e6919__llikBetaDshape2_G16H
#> #define _llikBetaDshape1 _rxdac23bc34facc9a82bfdee69bf82d91e6920__llikBetaDshape1_PUgn
#> #define _llikBeta _rxdac23bc34facc9a82bfdee69bf82d91e6921__llikBeta_Cjuw
#> #define _llikNbinomMuDmu _rxdac23bc34facc9a82bfdee69bf82d91e6922__llikNbinomMuDmu_A8iw
#> #define _llikNbinomMu _rxdac23bc34facc9a82bfdee69bf82d91e6923__llikNbinomMu_bn2G
#> #define _llikNbinomDprob _rxdac23bc34facc9a82bfdee69bf82d91e6924__llikNbinomDprob_TQjS
#> #define _llikNbinom _rxdac23bc34facc9a82bfdee69bf82d91e6925__llikNbinom_rer5
#> #define _llikBinomDprob _rxdac23bc34facc9a82bfdee69bf82d91e6926__llikBinomDprob_zexL
#> #define _llikBinom _rxdac23bc34facc9a82bfdee69bf82d91e6927__llikBinom_gAwM
#> #define _llikPoisDlambda _rxdac23bc34facc9a82bfdee69bf82d91e6928__llikPoisDlambda_E6nq
#> #define _llikPois _rxdac23bc34facc9a82bfdee69bf82d91e6929__llikPois_R3DD
#> #define _llikNormDsd _rxdac23bc34facc9a82bfdee69bf82d91e6930__llikNormDsd_9jN5
#> #define _llikNormDmean _rxdac23bc34facc9a82bfdee69bf82d91e6931__llikNormDmean_O4r9
#> #define _llikNorm _rxdac23bc34facc9a82bfdee69bf82d91e6932__llikNorm_D97m
#> #define simeps _rxdac23bc34facc9a82bfdee69bf82d91e6933_simeps_e5yb
#> #define simeta _rxdac23bc34facc9a82bfdee69bf82d91e6934_simeta_V8Eu
#> #define expit _rxdac23bc34facc9a82bfdee69bf82d91e6935_expit_Z1xT
#> #define logit _rxdac23bc34facc9a82bfdee69bf82d91e6936_logit_QYOQ
#> #define gammapDer _rxdac23bc34facc9a82bfdee69bf82d91e6937_gammapDer_e6aq
#> #define lowergamma _rxdac23bc34facc9a82bfdee69bf82d91e6938_lowergamma_t3Lb
#> #define uppergamma _rxdac23bc34facc9a82bfdee69bf82d91e6939_uppergamma_sbea
#> #define gammaqInva _rxdac23bc34facc9a82bfdee69bf82d91e6940_gammaqInva_w5m4
#> #define gammaqInv _rxdac23bc34facc9a82bfdee69bf82d91e6941_gammaqInv_brD9
#> #define gammapInva _rxdac23bc34facc9a82bfdee69bf82d91e6942_gammapInva_EZAJ
#> #define gammapInv _rxdac23bc34facc9a82bfdee69bf82d91e6943_gammapInv_Ne7I
#> #define gammaq _rxdac23bc34facc9a82bfdee69bf82d91e6944_gammaq_xUFA
#> #define gammap _rxdac23bc34facc9a82bfdee69bf82d91e6945_gammap_5R2w
#> #define phi _rxdac23bc34facc9a82bfdee69bf82d91e6946_phi_ZwwB
#> #define d2ELUa _rxdac23bc34facc9a82bfdee69bf82d91e6947_d2ELUa_gRvb
#> #define dELUa _rxdac23bc34facc9a82bfdee69bf82d91e6948_dELUa_lnk3
#> #define d2aELU _rxdac23bc34facc9a82bfdee69bf82d91e6949_d2aELU_57cf
#> #define d2ELU _rxdac23bc34facc9a82bfdee69bf82d91e6950_d2ELU_UwMa
#> #define dELU _rxdac23bc34facc9a82bfdee69bf82d91e6951_dELU_rEA9
#> #define ELU _rxdac23bc34facc9a82bfdee69bf82d91e6952_ELU_z1SE
#> #define dPReLUa1 _rxdac23bc34facc9a82bfdee69bf82d91e6953_dPReLUa1_1KzA
#> #define dPReLUa _rxdac23bc34facc9a82bfdee69bf82d91e6954_dPReLUa_f2H4
#> #define dPReLU _rxdac23bc34facc9a82bfdee69bf82d91e6955_dPReLU_Q2oF
#> #define PReLU _rxdac23bc34facc9a82bfdee69bf82d91e6956_PReLU_0joC
#> #define dSwish _rxdac23bc34facc9a82bfdee69bf82d91e6957_dSwish_Gj0a
#> #define Swish _rxdac23bc34facc9a82bfdee69bf82d91e6958_Swish_8wAb
#> #define dlReLU _rxdac23bc34facc9a82bfdee69bf82d91e6959_dlReLU_QELO
#> #define lReLU _rxdac23bc34facc9a82bfdee69bf82d91e6960_lReLU_lObq
#> #define dSELU _rxdac23bc34facc9a82bfdee69bf82d91e6961_dSELU_Nath
#> #define SELU _rxdac23bc34facc9a82bfdee69bf82d91e6962_SELU_ltiW
#> #define d4softplus _rxdac23bc34facc9a82bfdee69bf82d91e6963_d4softplus_R2IG
#> #define d3softplus _rxdac23bc34facc9a82bfdee69bf82d91e6964_d3softplus_VLQY
#> #define d2softplus _rxdac23bc34facc9a82bfdee69bf82d91e6965_d2softplus_M84J
#> #define dsoftplus _rxdac23bc34facc9a82bfdee69bf82d91e6966_dsoftplus_Cnuy
#> #define softplus _rxdac23bc34facc9a82bfdee69bf82d91e6967_softplus_CXPJ
#> #define d4GELU _rxdac23bc34facc9a82bfdee69bf82d91e6968_d4GELU_UjvF
#> #define d3GELU _rxdac23bc34facc9a82bfdee69bf82d91e6969_d3GELU_k7KJ
#> #define d2GELU _rxdac23bc34facc9a82bfdee69bf82d91e6970_d2GELU_vDxq
#> #define dGELU _rxdac23bc34facc9a82bfdee69bf82d91e6971_dGELU_bVSK
#> #define GELU _rxdac23bc34facc9a82bfdee69bf82d91e6972_GELU_Vfr7
#> #define dReLU _rxdac23bc34facc9a82bfdee69bf82d91e6973_dReLU_TcaU
#> #define ReLU _rxdac23bc34facc9a82bfdee69bf82d91e6974_ReLU_EgpY
#> #define riweibull _rxdac23bc34facc9a82bfdee69bf82d91e6975_riweibull_kNIq
#> #define riunif _rxdac23bc34facc9a82bfdee69bf82d91e6976_riunif_nEnm
#> #define rit_ _rxdac23bc34facc9a82bfdee69bf82d91e6977_rit__hSbo
#> #define ripois _rxdac23bc34facc9a82bfdee69bf82d91e6978_ripois_MZYB
#> #define ribeta _rxdac23bc34facc9a82bfdee69bf82d91e6979_ribeta_jkiC
#> #define rigamma _rxdac23bc34facc9a82bfdee69bf82d91e6980_rigamma_oUqv
#> #define rigeom _rxdac23bc34facc9a82bfdee69bf82d91e6981_rigeom_niI7
#> #define rif _rxdac23bc34facc9a82bfdee69bf82d91e6982_rif_5Ui0
#> #define riexp _rxdac23bc34facc9a82bfdee69bf82d91e6983_riexp_GVTs
#> #define richisq _rxdac23bc34facc9a82bfdee69bf82d91e6984_richisq_34TN
#> #define ricauchy _rxdac23bc34facc9a82bfdee69bf82d91e6985_ricauchy_HeFp
#> #define rinbinomMu _rxdac23bc34facc9a82bfdee69bf82d91e6986_rinbinomMu_ue1Z
#> #define rinbinom _rxdac23bc34facc9a82bfdee69bf82d91e6987_rinbinom_aEnJ
#> #define ribinom _rxdac23bc34facc9a82bfdee69bf82d91e6988_ribinom_R6UH
#> #define rinorm _rxdac23bc34facc9a82bfdee69bf82d91e6989_rinorm_ZKbS
#> #define rxweibull _rxdac23bc34facc9a82bfdee69bf82d91e6990_rxweibull_smNY
#> #define rxunif _rxdac23bc34facc9a82bfdee69bf82d91e6991_rxunif_lS2w
#> #define rxt_ _rxdac23bc34facc9a82bfdee69bf82d91e6992_rxt__zYqw
#> #define rxpois _rxdac23bc34facc9a82bfdee69bf82d91e6993_rxpois_2z8b
#> #define rxbeta _rxdac23bc34facc9a82bfdee69bf82d91e6994_rxbeta_GMdj
#> #define rxgamma _rxdac23bc34facc9a82bfdee69bf82d91e6995_rxgamma_LB3L
#> #define rxgeom _rxdac23bc34facc9a82bfdee69bf82d91e6996_rxgeom_sOUR
#> #define rxf _rxdac23bc34facc9a82bfdee69bf82d91e6997_rxf_bPYU
#> #define rxexp _rxdac23bc34facc9a82bfdee69bf82d91e6998_rxexp_PmW8
#> #define rxchisq _rxdac23bc34facc9a82bfdee69bf82d91e6999_rxchisq_sOmG
#> #define rxcauchy _rxdac23bc34facc9a82bfdee69bf82d91e7000_rxcauchy_ypO5
#> #define rxnbinomMu _rxdac23bc34facc9a82bfdee69bf82d91e7001_rxnbinomMu_5OnO
#> #define rxnbinom _rxdac23bc34facc9a82bfdee69bf82d91e7002_rxnbinom_E9TF
#> #define rxbinom _rxdac23bc34facc9a82bfdee69bf82d91e7003_rxbinom_yOK7
#> #define rxnorm _rxdac23bc34facc9a82bfdee69bf82d91e7004_rxnorm_OqEK
#> #define linCmtB _rxdac23bc34facc9a82bfdee69bf82d91e7005_linCmtB_wH4T
#> #define linCmtA _rxdac23bc34facc9a82bfdee69bf82d91e7006_linCmtA_tSJj
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){}
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){}
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  rx_c28902dc9c62ce7aa5f89f58135dca7b__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *rx_c28902dc9c62ce7aa5f89f58135dca7b__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP rx_c28902dc9c62ce7aa5f89f58135dca7b__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void rx_c28902dc9c62ce7aa5f89f58135dca7b__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double a = NA_REAL;
#> 
#>   double __udf[2];
#>   a = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   a =rxnorm(&_solveData->subjects[_cSub], 0.0, 1.0);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void rx_c28902dc9c62ce7aa5f89f58135dca7b__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void rx_c28902dc9c62ce7aa5f89f58135dca7b__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void rx_c28902dc9c62ce7aa5f89f58135dca7b__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double a = NA_REAL;
#> 
#>   double __udf[2];
#>   a = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   a =rxnorm(&_solveData->subjects[_cSub], 0.0, 1.0);
#> 
#>   _lhs[_LHS_0_]=a;
#> }
#> // Functional based bioavailability
#> double rx_c28902dc9c62ce7aa5f89f58135dca7b__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double rx_c28902dc9c62ce7aa5f89f58135dca7b__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double rx_c28902dc9c62ce7aa5f89f58135dca7b__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double rx_c28902dc9c62ce7aa5f89f58135dca7b__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void rx_c28902dc9c62ce7aa5f89f58135dca7b__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void rx_c28902dc9c62ce7aa5f89f58135dca7b__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void rx_c28902dc9c62ce7aa5f89f58135dca7b__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP rx_c28902dc9c62ce7aa5f89f58135dca7b__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_c28902dc9c62ce7aa5f89f58135dca7b__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAAMczAAAv7#aT)emh9MAb(jZ=W6xYB~15w`Qv<9*u/\"O+YnXEIMG72J%%JMjG5!+jMa]8K&T*I=JDVpWecT9^$lYFPngh~PL/2qLN./It(A.CZl`A}lMdC%%]BOVuZAIOycT_n~]d9v4$r!Qv;M!RpDl/ehn;yf%%Q|A5$nJ|2I]cEVa2/<%%.fQLMn>UO\"r*eKzUUW<2+{k@VL,oCewpxW0>N:B%%YLl*[#;oo~YTwtMP]\")MhTY5/qf^nn*y>~MoX~9s/{eII2w8S9xqX<qIV+pm|[XuZj#!GjwBiE{yc]j6*G[~]c6t<*%%Ki~@|y4&v}ZP{^yfnm$#.5NB0,Tgp0w(mo_{}o7iCNFaW)9Wo4nuo:$mSP;M!nG\"8Mw]hkF>j#CTpb4_eic]cSf+F9v\"9`dB\?IrNa7~QglF}2@Wr#iIBywDJ\?R*afv8H\"^r#;yYBmL.)[XcF(t|!c^\?|,#t*7pnx][V=!Dt,UZo\"XW{6^mJK[!cC)m3Zju_39WNUG1(*##S6Z~3g/#U,qwC;NEI9*xp56Ze8D^@8e8nnf8rpM;{~>lqB`zo]mS7;<UkXoE5F1w5L7dnU$\"[A4$E\?!Hx92>`9pp+DJ%%hIKKapfTlaC>P$30auc<1#RiRHj)p2{cLJ~b916e]`q~H1H$_&tQ~~~Hg#%%_q3cT<nA$YV2tD}Nl!LK!z>m]|UUdKKqB*{kr$PPLf$nwLh8Vh0IKv~*,]n!S62J&%%pj:&in#dJme@L7RCzE0VNszOATt$4m0f\"yWA;@5fLTt\"ClGDDwuBzym3Tv\?@sT7J\"J]7^y|(|F)nu.kcek#)1v7XAW5c(d3NOdHe(sVb}j]0<h\"r!w(`f\?xe=4Wuqs@>{b/0O`6PJ\?LeVr~*nf56NzE)oWycio]xtUW{0H8_LeHh0ma>M4h]99\"/XmqGoAW:i6RwgsliGWgS^0G:W\?h]>[tZxNhbqF]/19HlH2cs)CaqjbC.Q&cGn{nru>UeTq|dLAQylAY9s7c@5U*YwNc:Zra/fDdua+}Q7*}j5eC");
#>     char buf[1052];
#> #define __doBufN__ 1052
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void rx_c28902dc9c62ce7aa5f89f58135dca7b__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   rx_c28902dc9c62ce7aa5f89f58135dca7b__dydt(neq, *t, A, DADT);
#> }
#> extern int rx_c28902dc9c62ce7aa5f89f58135dca7b__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   rx_c28902dc9c62ce7aa5f89f58135dca7b__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void rx_c28902dc9c62ce7aa5f89f58135dca7b__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   rx_c28902dc9c62ce7aa5f89f58135dca7b__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void rx_c28902dc9c62ce7aa5f89f58135dca7b__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_rx_c28902dc9c62ce7aa5f89f58135dca7b_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("rx_c28902dc9c62ce7aa5f89f58135dca7b_","rx_c28902dc9c62ce7aa5f89f58135dca7b__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("rx_c28902dc9c62ce7aa5f89f58135dca7b_","rx_c28902dc9c62ce7aa5f89f58135dca7b__assignFuns", (DL_FUNC) rx_c28902dc9c62ce7aa5f89f58135dca7b__assignFuns);
#>   R_RegisterCCallable("rx_c28902dc9c62ce7aa5f89f58135dca7b_","rx_c28902dc9c62ce7aa5f89f58135dca7b__inis",(DL_FUNC) rx_c28902dc9c62ce7aa5f89f58135dca7b__inis);
#>   R_RegisterCCallable("rx_c28902dc9c62ce7aa5f89f58135dca7b_","rx_c28902dc9c62ce7aa5f89f58135dca7b__dydt",(DL_FUNC) rx_c28902dc9c62ce7aa5f89f58135dca7b__dydt);
#>   R_RegisterCCallable("rx_c28902dc9c62ce7aa5f89f58135dca7b_","rx_c28902dc9c62ce7aa5f89f58135dca7b__calc_lhs",(DL_FUNC) rx_c28902dc9c62ce7aa5f89f58135dca7b__calc_lhs);
#>   R_RegisterCCallable("rx_c28902dc9c62ce7aa5f89f58135dca7b_","rx_c28902dc9c62ce7aa5f89f58135dca7b__calc_jac",(DL_FUNC) rx_c28902dc9c62ce7aa5f89f58135dca7b__calc_jac);
#>   R_RegisterCCallable("rx_c28902dc9c62ce7aa5f89f58135dca7b_","rx_c28902dc9c62ce7aa5f89f58135dca7b__dydt_lsoda", (DL_FUNC) rx_c28902dc9c62ce7aa5f89f58135dca7b__dydt_lsoda);
#>   R_RegisterCCallable("rx_c28902dc9c62ce7aa5f89f58135dca7b_","rx_c28902dc9c62ce7aa5f89f58135dca7b__calc_jac_lsoda", (DL_FUNC) rx_c28902dc9c62ce7aa5f89f58135dca7b__calc_jac_lsoda);
#>   R_RegisterCCallable("rx_c28902dc9c62ce7aa5f89f58135dca7b_","rx_c28902dc9c62ce7aa5f89f58135dca7b__ode_solver_solvedata", (DL_FUNC) rx_c28902dc9c62ce7aa5f89f58135dca7b__ode_solver_solvedata);
#>   R_RegisterCCallable("rx_c28902dc9c62ce7aa5f89f58135dca7b_","rx_c28902dc9c62ce7aa5f89f58135dca7b__ode_solver_get_solvedata", (DL_FUNC) rx_c28902dc9c62ce7aa5f89f58135dca7b__ode_solver_get_solvedata);
#>   R_RegisterCCallable("rx_c28902dc9c62ce7aa5f89f58135dca7b_","rx_c28902dc9c62ce7aa5f89f58135dca7b__F", (DL_FUNC) rx_c28902dc9c62ce7aa5f89f58135dca7b__F);
#>   R_RegisterCCallable("rx_c28902dc9c62ce7aa5f89f58135dca7b_","rx_c28902dc9c62ce7aa5f89f58135dca7b__Lag", (DL_FUNC) rx_c28902dc9c62ce7aa5f89f58135dca7b__Lag);
#>   R_RegisterCCallable("rx_c28902dc9c62ce7aa5f89f58135dca7b_","rx_c28902dc9c62ce7aa5f89f58135dca7b__Rate", (DL_FUNC) rx_c28902dc9c62ce7aa5f89f58135dca7b__Rate);
#>   R_RegisterCCallable("rx_c28902dc9c62ce7aa5f89f58135dca7b_","rx_c28902dc9c62ce7aa5f89f58135dca7b__Dur", (DL_FUNC) rx_c28902dc9c62ce7aa5f89f58135dca7b__Dur);
#>   R_RegisterCCallable("rx_c28902dc9c62ce7aa5f89f58135dca7b_","rx_c28902dc9c62ce7aa5f89f58135dca7b__mtime", (DL_FUNC) rx_c28902dc9c62ce7aa5f89f58135dca7b__mtime);
#>   R_RegisterCCallable("rx_c28902dc9c62ce7aa5f89f58135dca7b_","rx_c28902dc9c62ce7aa5f89f58135dca7b__ME", (DL_FUNC) rx_c28902dc9c62ce7aa5f89f58135dca7b__ME);
#>   R_RegisterCCallable("rx_c28902dc9c62ce7aa5f89f58135dca7b_","rx_c28902dc9c62ce7aa5f89f58135dca7b__IndF", (DL_FUNC) rx_c28902dc9c62ce7aa5f89f58135dca7b__IndF);
#>   R_RegisterCCallable("rx_c28902dc9c62ce7aa5f89f58135dca7b_","rx_c28902dc9c62ce7aa5f89f58135dca7b__dydt_liblsoda", (DL_FUNC) rx_c28902dc9c62ce7aa5f89f58135dca7b__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_rx_c28902dc9c62ce7aa5f89f58135dca7b_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_rx_c28902dc9c62ce7aa5f89f58135dca7b_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"rx_c28902dc9c62ce7aa5f89f58135dca7b__model_vars", (DL_FUNC) &rx_c28902dc9c62ce7aa5f89f58135dca7b__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_rx_c28902dc9c62ce7aa5f89f58135dca7b_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_c28902dc9c62ce7aa5f89f58135dca7b__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("rx_c28902dc9c62ce7aa5f89f58135dca7b__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’

# }