Simulate random Poisson variable from threefry generator

Care should be taken with this method not to encounter the birthday problem, described https://www.johndcook.com/blog/2016/01/29/random-number-generator-seed-mistakes/. Since the sitmo threefry, this currently generates one random deviate from the uniform distribution to seed the engine threefry and then run the code.

Usage

rxpois(lambda, n = 1L, ncores = 1L)

Arguments

lambda

vector of (non-negative) means.

n

number of random values to return.

ncores

Number of cores for the simulation

rxnorm simulates using the threefry sitmo generator.

rxnormV used to simulate with the vandercorput simulator, but since it didn't satisfy the normal properties it was changed to simple be an alias of rxnorm. It is no longer supported in rxode2({}) blocks

Value

poission random number deviates

Details

Therefore, a simple call to the random number generated followed by a second call to random number generated may have identical seeds. As the number of random number generator calls are increased the probability that the birthday problem will increase.

The key to avoid this problem is to either run all simulations in the rxode2 environment once (therefore one seed or series of seeds for the whole simulation), pre-generate all random variables used for the simulation, or seed the rxode2 engine with rxSetSeed()

Internally each ID is seeded with a unique number so that the results do not depend on the number of cores used.

Examples

# \donttest{
## Use threefry engine

rxpois(lambda = 3, n = 10) # with rxpois you have to explicitly state n
#>  [1] 3 4 1 2 3 2 1 1 7 1
rxpois(lambda = 3, n = 10, ncores = 2) # You can parallelize the simulation using openMP
#>  [1] 2 1 3 3 5 2 6 6 8 2

rxpois(4) ## The first arguments are the lambda parameter
#> [1] 6

## This example uses `rxpois` directly in the model

rx <- function() {
  model({
    a <- rxpois(3)
  })
}

et <- et(1, id = 1:2)

s <- rxSolve(rx, et)
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#>  
#>  
#> C model:
#> // Define 1 LHS values
#> #define _LHS_0_ 0 // a
#> #define _getRxSolve_ _rx949477af372afb091f0698028887286a7150__getRxSolve__EhW3
#> #define _evalUdf _rx949477af372afb091f0698028887286a7151__evalUdf_MG9V
#> #define _solveData _rx949477af372afb091f0698028887286a7152__solveData_3pR1
#> #define _assign_ptr _rx949477af372afb091f0698028887286a7153__assign_ptr_aNkG
#> #define _rxRmModelLib _rx949477af372afb091f0698028887286a7154__rxRmModelLib_5JwD
#> #define _rxGetModelLib _rx949477af372afb091f0698028887286a7155__rxGetModelLib_9TK4
#> #define _old_c _rx949477af372afb091f0698028887286a7156__old_c_0NPJ
#> #define _ptrid _rx949477af372afb091f0698028887286a7157__ptrid_iALZ
#> #define _rxIsCurrentC _rx949477af372afb091f0698028887286a7158__rxIsCurrentC_Brhe
#> #define _sumPS _rx949477af372afb091f0698028887286a7159__sumPS_fghD
#> #define _prodPS _rx949477af372afb091f0698028887286a7160__prodPS_rm85
#> #define _prodType _rx949477af372afb091f0698028887286a7161__prodType_igvl
#> #define _sumType _rx949477af372afb091f0698028887286a7162__sumType_JYgG
#> #define _update_par_ptr _rx949477af372afb091f0698028887286a7163__update_par_ptr_qQxL
#> #define _getParCov _rx949477af372afb091f0698028887286a7164__getParCov_oYlF
#> #define _rxode2_rxAssignPtr _rx949477af372afb091f0698028887286a7165__rxode2_rxAssignPtr_gJdK
#> #define _rxQr _rx949477af372afb091f0698028887286a7166__rxQr_aFgm
#> #define _compareFactorVal _rx949477af372afb091f0698028887286a7167__compareFactorVal_DR4j
#> #define _sum _rx949477af372afb091f0698028887286a7168__sum_xuoP
#> #define _udf _rx949477af372afb091f0698028887286a7169__udf_8gmo
#> #define _sign _rx949477af372afb091f0698028887286a7170__sign_Uomw
#> #define _prod _rx949477af372afb091f0698028887286a7171__prod_LtuD
#> #define _max _rx949477af372afb091f0698028887286a7172__max_gOzt
#> #define _min _rx949477af372afb091f0698028887286a7173__min_PZwD
#> #define _transit4P _rx949477af372afb091f0698028887286a7174__transit4P_qx2t
#> #define _transit3P _rx949477af372afb091f0698028887286a7175__transit3P_0FfC
#> #define _assignFuns0 _rx949477af372afb091f0698028887286a7176__assignFuns0_RupB
#> #define _assignFuns _rx949477af372afb091f0698028887286a7177__assignFuns_fcLL
#> #define _rxord _rx949477af372afb091f0698028887286a7178__rxord_lJk0
#> #define __assignFuns2 _rx949477af372afb091f0698028887286a7179___assignFuns2_HOuq
#> #define _llikCauchyDscale _rx949477af372afb091f0698028887286a7180__llikCauchyDscale_uMQI
#> #define _llikCauchyDlocation _rx949477af372afb091f0698028887286a7181__llikCauchyDlocation_P4sB
#> #define _llikCauchy _rx949477af372afb091f0698028887286a7182__llikCauchy_KwlI
#> #define _llikGammaDrate _rx949477af372afb091f0698028887286a7183__llikGammaDrate_ON1t
#> #define _llikGammaDshape _rx949477af372afb091f0698028887286a7184__llikGammaDshape_TX5D
#> #define _llikGamma _rx949477af372afb091f0698028887286a7185__llikGamma_ulHL
#> #define _llikWeibullDscale _rx949477af372afb091f0698028887286a7186__llikWeibullDscale_xeAU
#> #define _llikWeibullDshape _rx949477af372afb091f0698028887286a7187__llikWeibullDshape_t0po
#> #define _llikWeibull _rx949477af372afb091f0698028887286a7188__llikWeibull_6AqS
#> #define _llikUnifDbeta _rx949477af372afb091f0698028887286a7189__llikUnifDbeta_QknH
#> #define _llikUnifDalpha _rx949477af372afb091f0698028887286a7190__llikUnifDalpha_8fPC
#> #define _llikUnif _rx949477af372afb091f0698028887286a7191__llikUnif_AnNk
#> #define _llikGeomDp _rx949477af372afb091f0698028887286a7192__llikGeomDp_Umsw
#> #define _llikGeom _rx949477af372afb091f0698028887286a7193__llikGeom_iOWG
#> #define _llikFDdf2 _rx949477af372afb091f0698028887286a7194__llikFDdf2_16HP
#> #define _llikFDdf1 _rx949477af372afb091f0698028887286a7195__llikFDdf1_UgnC
#> #define _llikF _rx949477af372afb091f0698028887286a7196__llikF_juwA
#> #define _llikExpDrate _rx949477af372afb091f0698028887286a7197__llikExpDrate_8iwb
#> #define _llikExp _rx949477af372afb091f0698028887286a7198__llikExp_n2GT
#> #define _llikChisqDdf _rx949477af372afb091f0698028887286a7199__llikChisqDdf_QjSr
#> #define _llikChisq _rx949477af372afb091f0698028887286a7200__llikChisq_er5z
#> #define _llikTDsd _rx949477af372afb091f0698028887286a7201__llikTDsd_exLg
#> #define _llikTDmean _rx949477af372afb091f0698028887286a7202__llikTDmean_AwME
#> #define _llikTDdf _rx949477af372afb091f0698028887286a7203__llikTDdf_6nqR
#> #define _llikT _rx949477af372afb091f0698028887286a7204__llikT_3DD9
#> #define _llikBetaDshape2 _rx949477af372afb091f0698028887286a7205__llikBetaDshape2_jN5O
#> #define _llikBetaDshape1 _rx949477af372afb091f0698028887286a7206__llikBetaDshape1_4r9D
#> #define _llikBeta _rx949477af372afb091f0698028887286a7207__llikBeta_97me
#> #define _llikNbinomMuDmu _rx949477af372afb091f0698028887286a7208__llikNbinomMuDmu_5ybV
#> #define _llikNbinomMu _rx949477af372afb091f0698028887286a7209__llikNbinomMu_8EuZ
#> #define _llikNbinomDprob _rx949477af372afb091f0698028887286a7210__llikNbinomDprob_1xTQ
#> #define _llikNbinom _rx949477af372afb091f0698028887286a7211__llikNbinom_YOQe
#> #define _llikBinomDprob _rx949477af372afb091f0698028887286a7212__llikBinomDprob_6aqt
#> #define _llikBinom _rx949477af372afb091f0698028887286a7213__llikBinom_3Lbs
#> #define _llikPoisDlambda _rx949477af372afb091f0698028887286a7214__llikPoisDlambda_beaw
#> #define _llikPois _rx949477af372afb091f0698028887286a7215__llikPois_5m4b
#> #define _llikNormDsd _rx949477af372afb091f0698028887286a7216__llikNormDsd_rD9E
#> #define _llikNormDmean _rx949477af372afb091f0698028887286a7217__llikNormDmean_ZAJN
#> #define _llikNorm _rx949477af372afb091f0698028887286a7218__llikNorm_e7Ix
#> #define simeps _rx949477af372afb091f0698028887286a7219_simeps_UFA5
#> #define simeta _rx949477af372afb091f0698028887286a7220_simeta_R2wZ
#> #define expit _rx949477af372afb091f0698028887286a7221_expit_wwBg
#> #define logit _rx949477af372afb091f0698028887286a7222_logit_Rvbl
#> #define gammapDer _rx949477af372afb091f0698028887286a7223_gammapDer_nk35
#> #define lowergamma _rx949477af372afb091f0698028887286a7224_lowergamma_7cfU
#> #define uppergamma _rx949477af372afb091f0698028887286a7225_uppergamma_wMar
#> #define gammaqInva _rx949477af372afb091f0698028887286a7226_gammaqInva_EA9z
#> #define gammaqInv _rx949477af372afb091f0698028887286a7227_gammaqInv_1SE1
#> #define gammapInva _rx949477af372afb091f0698028887286a7228_gammapInva_KzAf
#> #define gammapInv _rx949477af372afb091f0698028887286a7229_gammapInv_2H4Q
#> #define gammaq _rx949477af372afb091f0698028887286a7230_gammaq_2oF0
#> #define gammap _rx949477af372afb091f0698028887286a7231_gammap_joCG
#> #define phi _rx949477af372afb091f0698028887286a7232_phi_j0a8
#> #define d2ELUa _rx949477af372afb091f0698028887286a7233_d2ELUa_wAbQ
#> #define dELUa _rx949477af372afb091f0698028887286a7234_dELUa_ELOl
#> #define d2aELU _rx949477af372afb091f0698028887286a7235_d2aELU_ObqN
#> #define d2ELU _rx949477af372afb091f0698028887286a7236_d2ELU_athl
#> #define dELU _rx949477af372afb091f0698028887286a7237_dELU_tiWR
#> #define ELU _rx949477af372afb091f0698028887286a7238_ELU_2IGV
#> #define dPReLUa1 _rx949477af372afb091f0698028887286a7239_dPReLUa1_LQYM
#> #define dPReLUa _rx949477af372afb091f0698028887286a7240_dPReLUa_84JC
#> #define dPReLU _rx949477af372afb091f0698028887286a7241_dPReLU_nuyC
#> #define PReLU _rx949477af372afb091f0698028887286a7242_PReLU_XPJU
#> #define dSwish _rx949477af372afb091f0698028887286a7243_dSwish_jvFk
#> #define Swish _rx949477af372afb091f0698028887286a7244_Swish_7KJv
#> #define dlReLU _rx949477af372afb091f0698028887286a7245_dlReLU_Dxqb
#> #define lReLU _rx949477af372afb091f0698028887286a7246_lReLU_VSKV
#> #define dSELU _rx949477af372afb091f0698028887286a7247_dSELU_fr7T
#> #define SELU _rx949477af372afb091f0698028887286a7248_SELU_caUE
#> #define d4softplus _rx949477af372afb091f0698028887286a7249_d4softplus_gpYk
#> #define d3softplus _rx949477af372afb091f0698028887286a7250_d3softplus_NIqn
#> #define d2softplus _rx949477af372afb091f0698028887286a7251_d2softplus_Enmh
#> #define dsoftplus _rx949477af372afb091f0698028887286a7252_dsoftplus_SboM
#> #define softplus _rx949477af372afb091f0698028887286a7253_softplus_ZYBj
#> #define d4GELU _rx949477af372afb091f0698028887286a7254_d4GELU_kiCo
#> #define d3GELU _rx949477af372afb091f0698028887286a7255_d3GELU_Uqvn
#> #define d2GELU _rx949477af372afb091f0698028887286a7256_d2GELU_iI75
#> #define dGELU _rx949477af372afb091f0698028887286a7257_dGELU_Ui0G
#> #define GELU _rx949477af372afb091f0698028887286a7258_GELU_VTs3
#> #define dReLU _rx949477af372afb091f0698028887286a7259_dReLU_4TNH
#> #define ReLU _rx949477af372afb091f0698028887286a7260_ReLU_eFpu
#> #define riweibull _rx949477af372afb091f0698028887286a7261_riweibull_e1Za
#> #define riunif _rx949477af372afb091f0698028887286a7262_riunif_EnJR
#> #define rit_ _rx949477af372afb091f0698028887286a7263_rit__6UHZ
#> #define ripois _rx949477af372afb091f0698028887286a7264_ripois_KbSs
#> #define ribeta _rx949477af372afb091f0698028887286a7265_ribeta_mNYl
#> #define rigamma _rx949477af372afb091f0698028887286a7266_rigamma_S2wz
#> #define rigeom _rx949477af372afb091f0698028887286a7267_rigeom_Yqw2
#> #define rif _rx949477af372afb091f0698028887286a7268_rif_z8bG
#> #define riexp _rx949477af372afb091f0698028887286a7269_riexp_MdjL
#> #define richisq _rx949477af372afb091f0698028887286a7270_richisq_B3Ls
#> #define ricauchy _rx949477af372afb091f0698028887286a7271_ricauchy_OURb
#> #define rinbinomMu _rx949477af372afb091f0698028887286a7272_rinbinomMu_PYUP
#> #define rinbinom _rx949477af372afb091f0698028887286a7273_rinbinom_mW8s
#> #define ribinom _rx949477af372afb091f0698028887286a7274_ribinom_OmGy
#> #define rinorm _rx949477af372afb091f0698028887286a7275_rinorm_pO55
#> #define rxweibull _rx949477af372afb091f0698028887286a7276_rxweibull_OnOE
#> #define rxunif _rx949477af372afb091f0698028887286a7277_rxunif_9TFy
#> #define rxt_ _rx949477af372afb091f0698028887286a7278_rxt__OK7O
#> #define rxpois _rx949477af372afb091f0698028887286a7279_rxpois_qEKw
#> #define rxbeta _rx949477af372afb091f0698028887286a7280_rxbeta_H4Tt
#> #define rxgamma _rx949477af372afb091f0698028887286a7281_rxgamma_SJj0
#> #define rxgeom _rx949477af372afb091f0698028887286a7282_rxgeom_v8jx
#> #define rxf _rx949477af372afb091f0698028887286a7283_rxf_wnn6
#> #define rxexp _rx949477af372afb091f0698028887286a7284_rxexp_QBDU
#> #define rxchisq _rx949477af372afb091f0698028887286a7285_rxchisq_zzso
#> #define rxcauchy _rx949477af372afb091f0698028887286a7286_rxcauchy_z5n3
#> #define rxnbinomMu _rx949477af372afb091f0698028887286a7287_rxnbinomMu_6W22
#> #define rxnbinom _rx949477af372afb091f0698028887286a7288_rxnbinom_yJ1K
#> #define rxbinom _rx949477af372afb091f0698028887286a7289_rxbinom_elqu
#> #define rxnorm _rx949477af372afb091f0698028887286a7290_rxnorm_4n1M
#> #define linCmtB _rx949477af372afb091f0698028887286a7291_linCmtB_VL5w
#> #define linCmtA _rx949477af372afb091f0698028887286a7292_linCmtA_svsA
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){}
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){}
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  rx_5b297d5b4f7309198e8635fe4c84a0f1__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *rx_5b297d5b4f7309198e8635fe4c84a0f1__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP rx_5b297d5b4f7309198e8635fe4c84a0f1__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void rx_5b297d5b4f7309198e8635fe4c84a0f1__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double a = NA_REAL;
#> 
#>   double __udf[2];
#>   a = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   a =(double)rxpois(&_solveData->subjects[_cSub], 3);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void rx_5b297d5b4f7309198e8635fe4c84a0f1__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void rx_5b297d5b4f7309198e8635fe4c84a0f1__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void rx_5b297d5b4f7309198e8635fe4c84a0f1__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double a = NA_REAL;
#> 
#>   double __udf[2];
#>   a = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   a =(double)rxpois(&_solveData->subjects[_cSub], 3);
#> 
#>   _lhs[_LHS_0_]=a;
#> }
#> // Functional based bioavailability
#> double rx_5b297d5b4f7309198e8635fe4c84a0f1__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double rx_5b297d5b4f7309198e8635fe4c84a0f1__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double rx_5b297d5b4f7309198e8635fe4c84a0f1__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double rx_5b297d5b4f7309198e8635fe4c84a0f1__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void rx_5b297d5b4f7309198e8635fe4c84a0f1__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void rx_5b297d5b4f7309198e8635fe4c84a0f1__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void rx_5b297d5b4f7309198e8635fe4c84a0f1__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP rx_5b297d5b4f7309198e8635fe4c84a0f1__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_5b297d5b4f7309198e8635fe4c84a0f1__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAA[9zAAAv7#aT)gm/*MAF*#Z=WjmaB~15wu1tE9*u/kY(vAkz1@l1eEKoOjGWXXjy2/b#b5YkJ+I::&6!#<38~sWr`is&llkxv;N2~J\"hAmB<8RMX(7>2JkPADKM\"`pk=]ql/sw%%I#$F!^[>\"NiYnc[[S~QK1|5,~~!i8^b!4aN19%%r5/:_~>3[/2~ve9Gxe|[:#/e+]0&!XEQ+LVwdb#{[8.33+p,)@>!K<5i<+c}a2T2Kx)Guv(:=%%iiBvKdDDbY7Mti(aYQeM)LD\?pa$1f[FO,xl;XK!.K.mkCo*P`I6|nQ(}nx4a]DEV&gfw^@b9{yx;;w)^T\?\?|97ZaKBI)a~bGa5V5u2rRyMNH]he6=yb0A+[\?bw^JvWfz[O9wMvSFoCixOMzfnj4{$Ah@\"Nc<//kEyj{kSYr8]5}oYS\?7/eTX\?{M]>kJlySDMixsJWu^K3y;TeOv=aHH0U$5;mhE#U9.!r8@V_R8Co{]|@OH_|V4j%%lu^Tp$4NycU5@sj8f:n)Z0[92.e]P36yOV@O*G*<=!!ELHl$#ELR~mH4mr>s/zO)$KeSgWi.p1GyWVA:Cou;CkK][uu>W@A2<wL\?HmO:1r4=v2Hnz9Oj(k]J`\"0(,bsw7D]`)/e_5Ve:Z)`eU{/<5EwJ$juL_1h5oo^>a0|fgCdxE0x4V5D<`{lRIjdH5aH\?ZWr[3f=l{B~oXF9;UKaS@fk+T5Y}mhaSPDMwNd6=0:%%7j\?R46T2(m:Q>t)4v1V@512x<AT2{/M.9MsceY/UeG>2[b8i\"y5*,T}>rn(TPJ7J,/4IBWWvsC!W2_/|]n&hB2i+$l!+>p|zhjfZ.kQ!;*6ve}N|NV8_QyZ|\"hKTR=97Y[0lK`mN\"DwV!{:1an=WIh|F2Y/;$I4wO\?>pXOjs^PQ/eY5ziJI/hxS}@51eFtet))r2fw)vv7`J>S]KFMBCpTb><c,x~YANY<b^D|Fjg7G7#[oi}Du`pY]LR<Xy$tdONV|&>F}tQ!ntt|cY!gJ[j9cO*ERfzNVkjcntC+hxN3T0/U");
#>     char buf[1056];
#> #define __doBufN__ 1056
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void rx_5b297d5b4f7309198e8635fe4c84a0f1__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   rx_5b297d5b4f7309198e8635fe4c84a0f1__dydt(neq, *t, A, DADT);
#> }
#> extern int rx_5b297d5b4f7309198e8635fe4c84a0f1__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   rx_5b297d5b4f7309198e8635fe4c84a0f1__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void rx_5b297d5b4f7309198e8635fe4c84a0f1__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   rx_5b297d5b4f7309198e8635fe4c84a0f1__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void rx_5b297d5b4f7309198e8635fe4c84a0f1__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_rx_5b297d5b4f7309198e8635fe4c84a0f1_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("rx_5b297d5b4f7309198e8635fe4c84a0f1_","rx_5b297d5b4f7309198e8635fe4c84a0f1__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("rx_5b297d5b4f7309198e8635fe4c84a0f1_","rx_5b297d5b4f7309198e8635fe4c84a0f1__assignFuns", (DL_FUNC) rx_5b297d5b4f7309198e8635fe4c84a0f1__assignFuns);
#>   R_RegisterCCallable("rx_5b297d5b4f7309198e8635fe4c84a0f1_","rx_5b297d5b4f7309198e8635fe4c84a0f1__inis",(DL_FUNC) rx_5b297d5b4f7309198e8635fe4c84a0f1__inis);
#>   R_RegisterCCallable("rx_5b297d5b4f7309198e8635fe4c84a0f1_","rx_5b297d5b4f7309198e8635fe4c84a0f1__dydt",(DL_FUNC) rx_5b297d5b4f7309198e8635fe4c84a0f1__dydt);
#>   R_RegisterCCallable("rx_5b297d5b4f7309198e8635fe4c84a0f1_","rx_5b297d5b4f7309198e8635fe4c84a0f1__calc_lhs",(DL_FUNC) rx_5b297d5b4f7309198e8635fe4c84a0f1__calc_lhs);
#>   R_RegisterCCallable("rx_5b297d5b4f7309198e8635fe4c84a0f1_","rx_5b297d5b4f7309198e8635fe4c84a0f1__calc_jac",(DL_FUNC) rx_5b297d5b4f7309198e8635fe4c84a0f1__calc_jac);
#>   R_RegisterCCallable("rx_5b297d5b4f7309198e8635fe4c84a0f1_","rx_5b297d5b4f7309198e8635fe4c84a0f1__dydt_lsoda", (DL_FUNC) rx_5b297d5b4f7309198e8635fe4c84a0f1__dydt_lsoda);
#>   R_RegisterCCallable("rx_5b297d5b4f7309198e8635fe4c84a0f1_","rx_5b297d5b4f7309198e8635fe4c84a0f1__calc_jac_lsoda", (DL_FUNC) rx_5b297d5b4f7309198e8635fe4c84a0f1__calc_jac_lsoda);
#>   R_RegisterCCallable("rx_5b297d5b4f7309198e8635fe4c84a0f1_","rx_5b297d5b4f7309198e8635fe4c84a0f1__ode_solver_solvedata", (DL_FUNC) rx_5b297d5b4f7309198e8635fe4c84a0f1__ode_solver_solvedata);
#>   R_RegisterCCallable("rx_5b297d5b4f7309198e8635fe4c84a0f1_","rx_5b297d5b4f7309198e8635fe4c84a0f1__ode_solver_get_solvedata", (DL_FUNC) rx_5b297d5b4f7309198e8635fe4c84a0f1__ode_solver_get_solvedata);
#>   R_RegisterCCallable("rx_5b297d5b4f7309198e8635fe4c84a0f1_","rx_5b297d5b4f7309198e8635fe4c84a0f1__F", (DL_FUNC) rx_5b297d5b4f7309198e8635fe4c84a0f1__F);
#>   R_RegisterCCallable("rx_5b297d5b4f7309198e8635fe4c84a0f1_","rx_5b297d5b4f7309198e8635fe4c84a0f1__Lag", (DL_FUNC) rx_5b297d5b4f7309198e8635fe4c84a0f1__Lag);
#>   R_RegisterCCallable("rx_5b297d5b4f7309198e8635fe4c84a0f1_","rx_5b297d5b4f7309198e8635fe4c84a0f1__Rate", (DL_FUNC) rx_5b297d5b4f7309198e8635fe4c84a0f1__Rate);
#>   R_RegisterCCallable("rx_5b297d5b4f7309198e8635fe4c84a0f1_","rx_5b297d5b4f7309198e8635fe4c84a0f1__Dur", (DL_FUNC) rx_5b297d5b4f7309198e8635fe4c84a0f1__Dur);
#>   R_RegisterCCallable("rx_5b297d5b4f7309198e8635fe4c84a0f1_","rx_5b297d5b4f7309198e8635fe4c84a0f1__mtime", (DL_FUNC) rx_5b297d5b4f7309198e8635fe4c84a0f1__mtime);
#>   R_RegisterCCallable("rx_5b297d5b4f7309198e8635fe4c84a0f1_","rx_5b297d5b4f7309198e8635fe4c84a0f1__ME", (DL_FUNC) rx_5b297d5b4f7309198e8635fe4c84a0f1__ME);
#>   R_RegisterCCallable("rx_5b297d5b4f7309198e8635fe4c84a0f1_","rx_5b297d5b4f7309198e8635fe4c84a0f1__IndF", (DL_FUNC) rx_5b297d5b4f7309198e8635fe4c84a0f1__IndF);
#>   R_RegisterCCallable("rx_5b297d5b4f7309198e8635fe4c84a0f1_","rx_5b297d5b4f7309198e8635fe4c84a0f1__dydt_liblsoda", (DL_FUNC) rx_5b297d5b4f7309198e8635fe4c84a0f1__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_rx_5b297d5b4f7309198e8635fe4c84a0f1_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_rx_5b297d5b4f7309198e8635fe4c84a0f1_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"rx_5b297d5b4f7309198e8635fe4c84a0f1__model_vars", (DL_FUNC) &rx_5b297d5b4f7309198e8635fe4c84a0f1__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_rx_5b297d5b4f7309198e8635fe4c84a0f1_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_5b297d5b4f7309198e8635fe4c84a0f1__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("rx_5b297d5b4f7309198e8635fe4c84a0f1__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’
# }