Simulate uniform variable from threefry generator

Care should be taken with this method not to encounter the birthday problem, described https://www.johndcook.com/blog/2016/01/29/random-number-generator-seed-mistakes/. Since the sitmo threefry, this currently generates one random deviate from the uniform distribution to seed the engine threefry and then run the code.

Usage

rxunif(min = 0, max = 1, n = 1L, ncores = 1L)

Arguments

min, max

lower and upper limits of the distribution. Must be finite.

n

number of observations. If length(n) > 1, the length is taken to be the number required.

ncores

Number of cores for the simulation

rxnorm simulates using the threefry sitmo generator.

rxnormV used to simulate with the vandercorput simulator, but since it didn't satisfy the normal properties it was changed to simple be an alias of rxnorm. It is no longer supported in rxode2({}) blocks

Value

uniform random numbers

Details

Therefore, a simple call to the random number generated followed by a second call to random number generated may have identical seeds. As the number of random number generator calls are increased the probability that the birthday problem will increase.

The key to avoid this problem is to either run all simulations in the rxode2 environment once (therefore one seed or series of seeds for the whole simulation), pre-generate all random variables used for the simulation, or seed the rxode2 engine with rxSetSeed()

Internally each ID is seeded with a unique number so that the results do not depend on the number of cores used.

Examples

# \donttest{

## Use threefry engine

rxunif(min = 0, max = 4, n = 10) # with rxunif you have to explicitly state n
#>  [1] 2.4079607 1.7783947 1.2768194 0.4902107 1.5013125 3.5942682 1.4827298
#>  [8] 1.9568479 2.4242906 3.6372585
rxunif(min = 0, max = 4, n = 10, ncores = 2) # You can parallelize the simulation using openMP
#>  [1] 0.3825280 0.8600095 1.7186870 2.4484074 2.2692561 0.4697962 3.7543812
#>  [8] 0.4969432 2.6547598 2.1530229

rxunif()
#> [1] 0.5617085

## This example uses `rxunif` directly in the model

rx <- function() {
  model({
    a <- rxunif(0, 3)
  })
}

et <- et(1, id = 1:2)

s <- rxSolve(rx, et)
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#>  
#>  
#> C model:
#> // Define 1 LHS values
#> #define _LHS_0_ 0 // a
#> #define _getRxSolve_ _rx17272c8803d4472c38849f1599fff7e47436__getRxSolve__9TK4
#> #define _evalUdf _rx17272c8803d4472c38849f1599fff7e47437__evalUdf_0NPJ
#> #define _solveData _rx17272c8803d4472c38849f1599fff7e47438__solveData_iALZ
#> #define _assign_ptr _rx17272c8803d4472c38849f1599fff7e47439__assign_ptr_Brhe
#> #define _rxRmModelLib _rx17272c8803d4472c38849f1599fff7e47440__rxRmModelLib_fghD
#> #define _rxGetModelLib _rx17272c8803d4472c38849f1599fff7e47441__rxGetModelLib_rm85
#> #define _old_c _rx17272c8803d4472c38849f1599fff7e47442__old_c_igvl
#> #define _ptrid _rx17272c8803d4472c38849f1599fff7e47443__ptrid_JYgG
#> #define _rxIsCurrentC _rx17272c8803d4472c38849f1599fff7e47444__rxIsCurrentC_qQxL
#> #define _sumPS _rx17272c8803d4472c38849f1599fff7e47445__sumPS_oYlF
#> #define _prodPS _rx17272c8803d4472c38849f1599fff7e47446__prodPS_gJdK
#> #define _prodType _rx17272c8803d4472c38849f1599fff7e47447__prodType_aFgm
#> #define _sumType _rx17272c8803d4472c38849f1599fff7e47448__sumType_DR4j
#> #define _update_par_ptr _rx17272c8803d4472c38849f1599fff7e47449__update_par_ptr_xuoP
#> #define _getParCov _rx17272c8803d4472c38849f1599fff7e47450__getParCov_8gmo
#> #define _rxode2_rxAssignPtr _rx17272c8803d4472c38849f1599fff7e47451__rxode2_rxAssignPtr_Uomw
#> #define _rxQr _rx17272c8803d4472c38849f1599fff7e47452__rxQr_LtuD
#> #define _compareFactorVal _rx17272c8803d4472c38849f1599fff7e47453__compareFactorVal_gOzt
#> #define _sum _rx17272c8803d4472c38849f1599fff7e47454__sum_PZwD
#> #define _udf _rx17272c8803d4472c38849f1599fff7e47455__udf_qx2t
#> #define _sign _rx17272c8803d4472c38849f1599fff7e47456__sign_0FfC
#> #define _prod _rx17272c8803d4472c38849f1599fff7e47457__prod_RupB
#> #define _max _rx17272c8803d4472c38849f1599fff7e47458__max_fcLL
#> #define _min _rx17272c8803d4472c38849f1599fff7e47459__min_lJk0
#> #define _transit4P _rx17272c8803d4472c38849f1599fff7e47460__transit4P_HOuq
#> #define _transit3P _rx17272c8803d4472c38849f1599fff7e47461__transit3P_uMQI
#> #define _assignFuns0 _rx17272c8803d4472c38849f1599fff7e47462__assignFuns0_P4sB
#> #define _assignFuns _rx17272c8803d4472c38849f1599fff7e47463__assignFuns_KwlI
#> #define _rxord _rx17272c8803d4472c38849f1599fff7e47464__rxord_ON1t
#> #define __assignFuns2 _rx17272c8803d4472c38849f1599fff7e47465___assignFuns2_TX5D
#> #define _llikCauchyDscale _rx17272c8803d4472c38849f1599fff7e47466__llikCauchyDscale_ulHL
#> #define _llikCauchyDlocation _rx17272c8803d4472c38849f1599fff7e47467__llikCauchyDlocation_xeAU
#> #define _llikCauchy _rx17272c8803d4472c38849f1599fff7e47468__llikCauchy_t0po
#> #define _llikGammaDrate _rx17272c8803d4472c38849f1599fff7e47469__llikGammaDrate_6AqS
#> #define _llikGammaDshape _rx17272c8803d4472c38849f1599fff7e47470__llikGammaDshape_QknH
#> #define _llikGamma _rx17272c8803d4472c38849f1599fff7e47471__llikGamma_8fPC
#> #define _llikWeibullDscale _rx17272c8803d4472c38849f1599fff7e47472__llikWeibullDscale_AnNk
#> #define _llikWeibullDshape _rx17272c8803d4472c38849f1599fff7e47473__llikWeibullDshape_Umsw
#> #define _llikWeibull _rx17272c8803d4472c38849f1599fff7e47474__llikWeibull_iOWG
#> #define _llikUnifDbeta _rx17272c8803d4472c38849f1599fff7e47475__llikUnifDbeta_16HP
#> #define _llikUnifDalpha _rx17272c8803d4472c38849f1599fff7e47476__llikUnifDalpha_UgnC
#> #define _llikUnif _rx17272c8803d4472c38849f1599fff7e47477__llikUnif_juwA
#> #define _llikGeomDp _rx17272c8803d4472c38849f1599fff7e47478__llikGeomDp_8iwb
#> #define _llikGeom _rx17272c8803d4472c38849f1599fff7e47479__llikGeom_n2GT
#> #define _llikFDdf2 _rx17272c8803d4472c38849f1599fff7e47480__llikFDdf2_QjSr
#> #define _llikFDdf1 _rx17272c8803d4472c38849f1599fff7e47481__llikFDdf1_er5z
#> #define _llikF _rx17272c8803d4472c38849f1599fff7e47482__llikF_exLg
#> #define _llikExpDrate _rx17272c8803d4472c38849f1599fff7e47483__llikExpDrate_AwME
#> #define _llikExp _rx17272c8803d4472c38849f1599fff7e47484__llikExp_6nqR
#> #define _llikChisqDdf _rx17272c8803d4472c38849f1599fff7e47485__llikChisqDdf_3DD9
#> #define _llikChisq _rx17272c8803d4472c38849f1599fff7e47486__llikChisq_jN5O
#> #define _llikTDsd _rx17272c8803d4472c38849f1599fff7e47487__llikTDsd_4r9D
#> #define _llikTDmean _rx17272c8803d4472c38849f1599fff7e47488__llikTDmean_97me
#> #define _llikTDdf _rx17272c8803d4472c38849f1599fff7e47489__llikTDdf_5ybV
#> #define _llikT _rx17272c8803d4472c38849f1599fff7e47490__llikT_8EuZ
#> #define _llikBetaDshape2 _rx17272c8803d4472c38849f1599fff7e47491__llikBetaDshape2_1xTQ
#> #define _llikBetaDshape1 _rx17272c8803d4472c38849f1599fff7e47492__llikBetaDshape1_YOQe
#> #define _llikBeta _rx17272c8803d4472c38849f1599fff7e47493__llikBeta_6aqt
#> #define _llikNbinomMuDmu _rx17272c8803d4472c38849f1599fff7e47494__llikNbinomMuDmu_3Lbs
#> #define _llikNbinomMu _rx17272c8803d4472c38849f1599fff7e47495__llikNbinomMu_beaw
#> #define _llikNbinomDprob _rx17272c8803d4472c38849f1599fff7e47496__llikNbinomDprob_5m4b
#> #define _llikNbinom _rx17272c8803d4472c38849f1599fff7e47497__llikNbinom_rD9E
#> #define _llikBinomDprob _rx17272c8803d4472c38849f1599fff7e47498__llikBinomDprob_ZAJN
#> #define _llikBinom _rx17272c8803d4472c38849f1599fff7e47499__llikBinom_e7Ix
#> #define _llikPoisDlambda _rx17272c8803d4472c38849f1599fff7e47500__llikPoisDlambda_UFA5
#> #define _llikPois _rx17272c8803d4472c38849f1599fff7e47501__llikPois_R2wZ
#> #define _llikNormDsd _rx17272c8803d4472c38849f1599fff7e47502__llikNormDsd_wwBg
#> #define _llikNormDmean _rx17272c8803d4472c38849f1599fff7e47503__llikNormDmean_Rvbl
#> #define _llikNorm _rx17272c8803d4472c38849f1599fff7e47504__llikNorm_nk35
#> #define simeps _rx17272c8803d4472c38849f1599fff7e47505_simeps_7cfU
#> #define simeta _rx17272c8803d4472c38849f1599fff7e47506_simeta_wMar
#> #define expit _rx17272c8803d4472c38849f1599fff7e47507_expit_EA9z
#> #define logit _rx17272c8803d4472c38849f1599fff7e47508_logit_1SE1
#> #define gammapDer _rx17272c8803d4472c38849f1599fff7e47509_gammapDer_KzAf
#> #define lowergamma _rx17272c8803d4472c38849f1599fff7e47510_lowergamma_2H4Q
#> #define uppergamma _rx17272c8803d4472c38849f1599fff7e47511_uppergamma_2oF0
#> #define gammaqInva _rx17272c8803d4472c38849f1599fff7e47512_gammaqInva_joCG
#> #define gammaqInv _rx17272c8803d4472c38849f1599fff7e47513_gammaqInv_j0a8
#> #define gammapInva _rx17272c8803d4472c38849f1599fff7e47514_gammapInva_wAbQ
#> #define gammapInv _rx17272c8803d4472c38849f1599fff7e47515_gammapInv_ELOl
#> #define gammaq _rx17272c8803d4472c38849f1599fff7e47516_gammaq_ObqN
#> #define gammap _rx17272c8803d4472c38849f1599fff7e47517_gammap_athl
#> #define phi _rx17272c8803d4472c38849f1599fff7e47518_phi_tiWR
#> #define d2ELUa _rx17272c8803d4472c38849f1599fff7e47519_d2ELUa_2IGV
#> #define dELUa _rx17272c8803d4472c38849f1599fff7e47520_dELUa_LQYM
#> #define d2aELU _rx17272c8803d4472c38849f1599fff7e47521_d2aELU_84JC
#> #define d2ELU _rx17272c8803d4472c38849f1599fff7e47522_d2ELU_nuyC
#> #define dELU _rx17272c8803d4472c38849f1599fff7e47523_dELU_XPJU
#> #define ELU _rx17272c8803d4472c38849f1599fff7e47524_ELU_jvFk
#> #define dPReLUa1 _rx17272c8803d4472c38849f1599fff7e47525_dPReLUa1_7KJv
#> #define dPReLUa _rx17272c8803d4472c38849f1599fff7e47526_dPReLUa_Dxqb
#> #define dPReLU _rx17272c8803d4472c38849f1599fff7e47527_dPReLU_VSKV
#> #define PReLU _rx17272c8803d4472c38849f1599fff7e47528_PReLU_fr7T
#> #define dSwish _rx17272c8803d4472c38849f1599fff7e47529_dSwish_caUE
#> #define Swish _rx17272c8803d4472c38849f1599fff7e47530_Swish_gpYk
#> #define dlReLU _rx17272c8803d4472c38849f1599fff7e47531_dlReLU_NIqn
#> #define lReLU _rx17272c8803d4472c38849f1599fff7e47532_lReLU_Enmh
#> #define dSELU _rx17272c8803d4472c38849f1599fff7e47533_dSELU_SboM
#> #define SELU _rx17272c8803d4472c38849f1599fff7e47534_SELU_ZYBj
#> #define d4softplus _rx17272c8803d4472c38849f1599fff7e47535_d4softplus_kiCo
#> #define d3softplus _rx17272c8803d4472c38849f1599fff7e47536_d3softplus_Uqvn
#> #define d2softplus _rx17272c8803d4472c38849f1599fff7e47537_d2softplus_iI75
#> #define dsoftplus _rx17272c8803d4472c38849f1599fff7e47538_dsoftplus_Ui0G
#> #define softplus _rx17272c8803d4472c38849f1599fff7e47539_softplus_VTs3
#> #define d4GELU _rx17272c8803d4472c38849f1599fff7e47540_d4GELU_4TNH
#> #define d3GELU _rx17272c8803d4472c38849f1599fff7e47541_d3GELU_eFpu
#> #define d2GELU _rx17272c8803d4472c38849f1599fff7e47542_d2GELU_e1Za
#> #define dGELU _rx17272c8803d4472c38849f1599fff7e47543_dGELU_EnJR
#> #define GELU _rx17272c8803d4472c38849f1599fff7e47544_GELU_6UHZ
#> #define dReLU _rx17272c8803d4472c38849f1599fff7e47545_dReLU_KbSs
#> #define ReLU _rx17272c8803d4472c38849f1599fff7e47546_ReLU_mNYl
#> #define riweibull _rx17272c8803d4472c38849f1599fff7e47547_riweibull_S2wz
#> #define riunif _rx17272c8803d4472c38849f1599fff7e47548_riunif_Yqw2
#> #define rit_ _rx17272c8803d4472c38849f1599fff7e47549_rit__z8bG
#> #define ripois _rx17272c8803d4472c38849f1599fff7e47550_ripois_MdjL
#> #define ribeta _rx17272c8803d4472c38849f1599fff7e47551_ribeta_B3Ls
#> #define rigamma _rx17272c8803d4472c38849f1599fff7e47552_rigamma_OURb
#> #define rigeom _rx17272c8803d4472c38849f1599fff7e47553_rigeom_PYUP
#> #define rif _rx17272c8803d4472c38849f1599fff7e47554_rif_mW8s
#> #define riexp _rx17272c8803d4472c38849f1599fff7e47555_riexp_OmGy
#> #define richisq _rx17272c8803d4472c38849f1599fff7e47556_richisq_pO55
#> #define ricauchy _rx17272c8803d4472c38849f1599fff7e47557_ricauchy_OnOE
#> #define rinbinomMu _rx17272c8803d4472c38849f1599fff7e47558_rinbinomMu_9TFy
#> #define rinbinom _rx17272c8803d4472c38849f1599fff7e47559_rinbinom_OK7O
#> #define ribinom _rx17272c8803d4472c38849f1599fff7e47560_ribinom_qEKw
#> #define rinorm _rx17272c8803d4472c38849f1599fff7e47561_rinorm_H4Tt
#> #define rxweibull _rx17272c8803d4472c38849f1599fff7e47562_rxweibull_SJj0
#> #define rxunif _rx17272c8803d4472c38849f1599fff7e47563_rxunif_v8jx
#> #define rxt_ _rx17272c8803d4472c38849f1599fff7e47564_rxt__wnn6
#> #define rxpois _rx17272c8803d4472c38849f1599fff7e47565_rxpois_QBDU
#> #define rxbeta _rx17272c8803d4472c38849f1599fff7e47566_rxbeta_zzso
#> #define rxgamma _rx17272c8803d4472c38849f1599fff7e47567_rxgamma_z5n3
#> #define rxgeom _rx17272c8803d4472c38849f1599fff7e47568_rxgeom_6W22
#> #define rxf _rx17272c8803d4472c38849f1599fff7e47569_rxf_yJ1K
#> #define rxexp _rx17272c8803d4472c38849f1599fff7e47570_rxexp_elqu
#> #define rxchisq _rx17272c8803d4472c38849f1599fff7e47571_rxchisq_4n1M
#> #define rxcauchy _rx17272c8803d4472c38849f1599fff7e47572_rxcauchy_VL5w
#> #define rxnbinomMu _rx17272c8803d4472c38849f1599fff7e47573_rxnbinomMu_svsA
#> #define rxnbinom _rx17272c8803d4472c38849f1599fff7e47574_rxnbinom_VKVQ
#> #define rxbinom _rx17272c8803d4472c38849f1599fff7e47575_rxbinom_gWt0
#> #define rxnorm _rx17272c8803d4472c38849f1599fff7e47576_rxnorm_j8P1
#> #define linCmtB _rx17272c8803d4472c38849f1599fff7e47577_linCmtB_Sql1
#> #define linCmtA _rx17272c8803d4472c38849f1599fff7e47578_linCmtA_4DAb
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){}
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){}
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  rx_d4ffee5fd240abf037ffcbdd4c87af0f__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *rx_d4ffee5fd240abf037ffcbdd4c87af0f__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP rx_d4ffee5fd240abf037ffcbdd4c87af0f__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void rx_d4ffee5fd240abf037ffcbdd4c87af0f__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double a = NA_REAL;
#> 
#>   double __udf[2];
#>   a = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   a =rxunif(&_solveData->subjects[_cSub], 0,(double)3);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void rx_d4ffee5fd240abf037ffcbdd4c87af0f__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void rx_d4ffee5fd240abf037ffcbdd4c87af0f__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void rx_d4ffee5fd240abf037ffcbdd4c87af0f__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double a = NA_REAL;
#> 
#>   double __udf[2];
#>   a = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   a =rxunif(&_solveData->subjects[_cSub], 0,(double)3);
#> 
#>   _lhs[_LHS_0_]=a;
#> }
#> // Functional based bioavailability
#> double rx_d4ffee5fd240abf037ffcbdd4c87af0f__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double rx_d4ffee5fd240abf037ffcbdd4c87af0f__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double rx_d4ffee5fd240abf037ffcbdd4c87af0f__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double rx_d4ffee5fd240abf037ffcbdd4c87af0f__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void rx_d4ffee5fd240abf037ffcbdd4c87af0f__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void rx_d4ffee5fd240abf037ffcbdd4c87af0f__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void rx_d4ffee5fd240abf037ffcbdd4c87af0f__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP rx_d4ffee5fd240abf037ffcbdd4c87af0f__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_d4ffee5fd240abf037ffcbdd4c87af0f__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAA+>zAAAv7#aT)kmC:MA**)+=WjmaB~15wu1tE|n)])SI4\"1D\")S62IU.zSw4L5R*M43j3Xx]UaiRd[2s+@_P|_~61A`+fUVXU4U2pHL_EYLJtE;1z}}RylmbRmNqw[/D\?v1C;3~kpTkuG!^6(dUS5U7b{2s,F.rvk\"s1Vs^x&V1[VWl|84e8~lrvZf}V2Jcv$vZS#@mIb#%%,X9VmRVwbbQT/#tLFY!#xV4xDTmTLaY%%V58=vp~cp~vO97.,OMxI}|;I[DnXgg^jxXYs:PYr5UtWPm(f0Z~/`,gW+m*8a22`\?gp|eFq93PQ_@oE7MRfowv<;4e3VM,|9s9HPpEt~IqUZJx]wAm^*iy0c1t[$MScP9.7{RN9nG\"8MxSE8TPuI5XjP&IMb.)>+oE:AK*dUOm%%naCyj{kSYS#]500,`JaKQ44MIYJd@},qx\"xVJ.%%7[=t}`8RGi^T~\?O_,8O;A]jaG:30P~AXE~ID4k}~MQU^vaJ6:4{=xbZfO+Czp;SVa5io]6+~\"m!\?G]E;]mm*m6`xL#4ZZnnnX6Znf#05`|9@SL:QqBP;T`ARxh0P*AECBt>WR\"%%=KFHY%%FKO4.YC|!&a=,%%BDWqf9MD&n%%V6lL.l^&&>80V33L9=bN*.saN3N:E#YSMb&!rM>q#}cKller!1d+bOc>}.TRZaPuiscD&zQa`$3R^iV>vs7$A62f[`pkcZl){]/~`(RhQnElCZD^;4uSWv|u6)%%/YM!jai&j7S%%PvvsE4+c4CSX/M.vW7FWd:!aLw(X{qO#F>hyuSDzB.t74h$TK}vL]]j;ne+}kPg`e7_Ax^C5$jLQ\"QV+Znt0w[PElOqD#Lsj)Yxo^OuYFk{pngRG@,f7WIC!]&6Yl_[x}vFf\?eFmhLbkx!5FJftHwH[(:E\"gS{0D;1Z*Yzg<Z$>vBsK5VA_&Qts6:E*3!;h=Meg`[|dkLQ#0nGq[JD3%%_;pvH]~(D%%UFlwj0!CSnHj.fWa1`CT0aPV96SU.6Sg6O>e7>/$P0is8]5H+pQ{a!+~I9S!&QS&C");
#>     char buf[1067];
#> #define __doBufN__ 1067
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void rx_d4ffee5fd240abf037ffcbdd4c87af0f__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   rx_d4ffee5fd240abf037ffcbdd4c87af0f__dydt(neq, *t, A, DADT);
#> }
#> extern int rx_d4ffee5fd240abf037ffcbdd4c87af0f__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   rx_d4ffee5fd240abf037ffcbdd4c87af0f__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void rx_d4ffee5fd240abf037ffcbdd4c87af0f__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   rx_d4ffee5fd240abf037ffcbdd4c87af0f__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void rx_d4ffee5fd240abf037ffcbdd4c87af0f__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_rx_d4ffee5fd240abf037ffcbdd4c87af0f_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("rx_d4ffee5fd240abf037ffcbdd4c87af0f_","rx_d4ffee5fd240abf037ffcbdd4c87af0f__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("rx_d4ffee5fd240abf037ffcbdd4c87af0f_","rx_d4ffee5fd240abf037ffcbdd4c87af0f__assignFuns", (DL_FUNC) rx_d4ffee5fd240abf037ffcbdd4c87af0f__assignFuns);
#>   R_RegisterCCallable("rx_d4ffee5fd240abf037ffcbdd4c87af0f_","rx_d4ffee5fd240abf037ffcbdd4c87af0f__inis",(DL_FUNC) rx_d4ffee5fd240abf037ffcbdd4c87af0f__inis);
#>   R_RegisterCCallable("rx_d4ffee5fd240abf037ffcbdd4c87af0f_","rx_d4ffee5fd240abf037ffcbdd4c87af0f__dydt",(DL_FUNC) rx_d4ffee5fd240abf037ffcbdd4c87af0f__dydt);
#>   R_RegisterCCallable("rx_d4ffee5fd240abf037ffcbdd4c87af0f_","rx_d4ffee5fd240abf037ffcbdd4c87af0f__calc_lhs",(DL_FUNC) rx_d4ffee5fd240abf037ffcbdd4c87af0f__calc_lhs);
#>   R_RegisterCCallable("rx_d4ffee5fd240abf037ffcbdd4c87af0f_","rx_d4ffee5fd240abf037ffcbdd4c87af0f__calc_jac",(DL_FUNC) rx_d4ffee5fd240abf037ffcbdd4c87af0f__calc_jac);
#>   R_RegisterCCallable("rx_d4ffee5fd240abf037ffcbdd4c87af0f_","rx_d4ffee5fd240abf037ffcbdd4c87af0f__dydt_lsoda", (DL_FUNC) rx_d4ffee5fd240abf037ffcbdd4c87af0f__dydt_lsoda);
#>   R_RegisterCCallable("rx_d4ffee5fd240abf037ffcbdd4c87af0f_","rx_d4ffee5fd240abf037ffcbdd4c87af0f__calc_jac_lsoda", (DL_FUNC) rx_d4ffee5fd240abf037ffcbdd4c87af0f__calc_jac_lsoda);
#>   R_RegisterCCallable("rx_d4ffee5fd240abf037ffcbdd4c87af0f_","rx_d4ffee5fd240abf037ffcbdd4c87af0f__ode_solver_solvedata", (DL_FUNC) rx_d4ffee5fd240abf037ffcbdd4c87af0f__ode_solver_solvedata);
#>   R_RegisterCCallable("rx_d4ffee5fd240abf037ffcbdd4c87af0f_","rx_d4ffee5fd240abf037ffcbdd4c87af0f__ode_solver_get_solvedata", (DL_FUNC) rx_d4ffee5fd240abf037ffcbdd4c87af0f__ode_solver_get_solvedata);
#>   R_RegisterCCallable("rx_d4ffee5fd240abf037ffcbdd4c87af0f_","rx_d4ffee5fd240abf037ffcbdd4c87af0f__F", (DL_FUNC) rx_d4ffee5fd240abf037ffcbdd4c87af0f__F);
#>   R_RegisterCCallable("rx_d4ffee5fd240abf037ffcbdd4c87af0f_","rx_d4ffee5fd240abf037ffcbdd4c87af0f__Lag", (DL_FUNC) rx_d4ffee5fd240abf037ffcbdd4c87af0f__Lag);
#>   R_RegisterCCallable("rx_d4ffee5fd240abf037ffcbdd4c87af0f_","rx_d4ffee5fd240abf037ffcbdd4c87af0f__Rate", (DL_FUNC) rx_d4ffee5fd240abf037ffcbdd4c87af0f__Rate);
#>   R_RegisterCCallable("rx_d4ffee5fd240abf037ffcbdd4c87af0f_","rx_d4ffee5fd240abf037ffcbdd4c87af0f__Dur", (DL_FUNC) rx_d4ffee5fd240abf037ffcbdd4c87af0f__Dur);
#>   R_RegisterCCallable("rx_d4ffee5fd240abf037ffcbdd4c87af0f_","rx_d4ffee5fd240abf037ffcbdd4c87af0f__mtime", (DL_FUNC) rx_d4ffee5fd240abf037ffcbdd4c87af0f__mtime);
#>   R_RegisterCCallable("rx_d4ffee5fd240abf037ffcbdd4c87af0f_","rx_d4ffee5fd240abf037ffcbdd4c87af0f__ME", (DL_FUNC) rx_d4ffee5fd240abf037ffcbdd4c87af0f__ME);
#>   R_RegisterCCallable("rx_d4ffee5fd240abf037ffcbdd4c87af0f_","rx_d4ffee5fd240abf037ffcbdd4c87af0f__IndF", (DL_FUNC) rx_d4ffee5fd240abf037ffcbdd4c87af0f__IndF);
#>   R_RegisterCCallable("rx_d4ffee5fd240abf037ffcbdd4c87af0f_","rx_d4ffee5fd240abf037ffcbdd4c87af0f__dydt_liblsoda", (DL_FUNC) rx_d4ffee5fd240abf037ffcbdd4c87af0f__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_rx_d4ffee5fd240abf037ffcbdd4c87af0f_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_rx_d4ffee5fd240abf037ffcbdd4c87af0f_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"rx_d4ffee5fd240abf037ffcbdd4c87af0f__model_vars", (DL_FUNC) &rx_d4ffee5fd240abf037ffcbdd4c87af0f__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_rx_d4ffee5fd240abf037ffcbdd4c87af0f_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_d4ffee5fd240abf037ffcbdd4c87af0f__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("rx_d4ffee5fd240abf037ffcbdd4c87af0f__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’
# }