Simulate Weibull variable from threefry generator

Care should be taken with this method not to encounter the birthday problem, described https://www.johndcook.com/blog/2016/01/29/random-number-generator-seed-mistakes/. Since the sitmo threefry, this currently generates one random deviate from the uniform distribution to seed the engine threefry and then run the code.

Usage

rxweibull(shape, scale = 1, n = 1L, ncores = 1L)

Arguments

shape, scale

shape and scale parameters, the latter defaulting to 1.

n

number of observations. If length(n) > 1, the length is taken to be the number required.

ncores

Number of cores for the simulation

rxnorm simulates using the threefry sitmo generator.

rxnormV used to simulate with the vandercorput simulator, but since it didn't satisfy the normal properties it was changed to simple be an alias of rxnorm. It is no longer supported in rxode2({}) blocks

Value

Weibull random deviates

Details

Therefore, a simple call to the random number generated followed by a second call to random number generated may have identical seeds. As the number of random number generator calls are increased the probability that the birthday problem will increase.

The key to avoid this problem is to either run all simulations in the rxode2 environment once (therefore one seed or series of seeds for the whole simulation), pre-generate all random variables used for the simulation, or seed the rxode2 engine with rxSetSeed()

Internally each ID is seeded with a unique number so that the results do not depend on the number of cores used.

Examples

# \donttest{

## Use threefry engine

# with rxweibull you have to explicitly state n
rxweibull(shape = 1, scale = 4, n = 10)
#>  [1]  1.953503  0.124592 12.502286  2.284222  4.342211  2.337220  3.743978
#>  [8]  1.087069  1.155109  7.962468

# You can parallelize the simulation using openMP
rxweibull(shape = 1, scale = 4, n = 10, ncores = 2)
#>  [1]  8.673411 11.228521  3.515282  8.043888  2.588675  9.741515  2.163835
#>  [8]  1.233657  6.839772  3.783971

rxweibull(3)
#> [1] 0.5160208

## This example uses `rxweibull` directly in the model

rx <- function() {
  model({
    a <- rxweibull(1, 3)
  })
}

et <- et(1, id = 1:2)

s <- rxSolve(rx, et)
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#>  
#>  
#> C model:
#> // Define 1 LHS values
#> #define _LHS_0_ 0 // a
#> #define _getRxSolve_ _rxd0205c04379a33168bab6df21f717fe67579__getRxSolve__LZBr
#> #define _evalUdf _rxd0205c04379a33168bab6df21f717fe67580__evalUdf_hefg
#> #define _solveData _rxd0205c04379a33168bab6df21f717fe67581__solveData_hDrm
#> #define _assign_ptr _rxd0205c04379a33168bab6df21f717fe67582__assign_ptr_85ig
#> #define _rxRmModelLib _rxd0205c04379a33168bab6df21f717fe67583__rxRmModelLib_vlJY
#> #define _rxGetModelLib _rxd0205c04379a33168bab6df21f717fe67584__rxGetModelLib_gGqQ
#> #define _old_c _rxd0205c04379a33168bab6df21f717fe67585__old_c_xLoY
#> #define _ptrid _rxd0205c04379a33168bab6df21f717fe67586__ptrid_lFgJ
#> #define _rxIsCurrentC _rxd0205c04379a33168bab6df21f717fe67587__rxIsCurrentC_dKaF
#> #define _sumPS _rxd0205c04379a33168bab6df21f717fe67588__sumPS_gmDR
#> #define _prodPS _rxd0205c04379a33168bab6df21f717fe67589__prodPS_4jxu
#> #define _prodType _rxd0205c04379a33168bab6df21f717fe67590__prodType_oP8g
#> #define _sumType _rxd0205c04379a33168bab6df21f717fe67591__sumType_moUo
#> #define _update_par_ptr _rxd0205c04379a33168bab6df21f717fe67592__update_par_ptr_mwLt
#> #define _getParCov _rxd0205c04379a33168bab6df21f717fe67593__getParCov_uDgO
#> #define _rxode2_rxAssignPtr _rxd0205c04379a33168bab6df21f717fe67594__rxode2_rxAssignPtr_ztPZ
#> #define _rxQr _rxd0205c04379a33168bab6df21f717fe67595__rxQr_wDqx
#> #define _compareFactorVal _rxd0205c04379a33168bab6df21f717fe67596__compareFactorVal_2t0F
#> #define _sum _rxd0205c04379a33168bab6df21f717fe67597__sum_fCRu
#> #define _udf _rxd0205c04379a33168bab6df21f717fe67598__udf_pBfc
#> #define _sign _rxd0205c04379a33168bab6df21f717fe67599__sign_LLlJ
#> #define _prod _rxd0205c04379a33168bab6df21f717fe67600__prod_k0HO
#> #define _max _rxd0205c04379a33168bab6df21f717fe67601__max_uquM
#> #define _min _rxd0205c04379a33168bab6df21f717fe67602__min_QIP4
#> #define _transit4P _rxd0205c04379a33168bab6df21f717fe67603__transit4P_sBKw
#> #define _transit3P _rxd0205c04379a33168bab6df21f717fe67604__transit3P_lION
#> #define _assignFuns0 _rxd0205c04379a33168bab6df21f717fe67605__assignFuns0_1tTX
#> #define _assignFuns _rxd0205c04379a33168bab6df21f717fe67606__assignFuns_5Dul
#> #define _rxord _rxd0205c04379a33168bab6df21f717fe67607__rxord_HLxe
#> #define __assignFuns2 _rxd0205c04379a33168bab6df21f717fe67608___assignFuns2_AUt0
#> #define _llikCauchyDscale _rxd0205c04379a33168bab6df21f717fe67609__llikCauchyDscale_po6A
#> #define _llikCauchyDlocation _rxd0205c04379a33168bab6df21f717fe67610__llikCauchyDlocation_qSQk
#> #define _llikCauchy _rxd0205c04379a33168bab6df21f717fe67611__llikCauchy_nH8f
#> #define _llikGammaDrate _rxd0205c04379a33168bab6df21f717fe67612__llikGammaDrate_PCAn
#> #define _llikGammaDshape _rxd0205c04379a33168bab6df21f717fe67613__llikGammaDshape_NkUm
#> #define _llikGamma _rxd0205c04379a33168bab6df21f717fe67614__llikGamma_swiO
#> #define _llikWeibullDscale _rxd0205c04379a33168bab6df21f717fe67615__llikWeibullDscale_WG16
#> #define _llikWeibullDshape _rxd0205c04379a33168bab6df21f717fe67616__llikWeibullDshape_HPUg
#> #define _llikWeibull _rxd0205c04379a33168bab6df21f717fe67617__llikWeibull_nCju
#> #define _llikUnifDbeta _rxd0205c04379a33168bab6df21f717fe67618__llikUnifDbeta_wA8i
#> #define _llikUnifDalpha _rxd0205c04379a33168bab6df21f717fe67619__llikUnifDalpha_wbn2
#> #define _llikUnif _rxd0205c04379a33168bab6df21f717fe67620__llikUnif_GTQj
#> #define _llikGeomDp _rxd0205c04379a33168bab6df21f717fe67621__llikGeomDp_Srer
#> #define _llikGeom _rxd0205c04379a33168bab6df21f717fe67622__llikGeom_5zex
#> #define _llikFDdf2 _rxd0205c04379a33168bab6df21f717fe67623__llikFDdf2_LgAw
#> #define _llikFDdf1 _rxd0205c04379a33168bab6df21f717fe67624__llikFDdf1_ME6n
#> #define _llikF _rxd0205c04379a33168bab6df21f717fe67625__llikF_qR3D
#> #define _llikExpDrate _rxd0205c04379a33168bab6df21f717fe67626__llikExpDrate_D9jN
#> #define _llikExp _rxd0205c04379a33168bab6df21f717fe67627__llikExp_5O4r
#> #define _llikChisqDdf _rxd0205c04379a33168bab6df21f717fe67628__llikChisqDdf_9D97
#> #define _llikChisq _rxd0205c04379a33168bab6df21f717fe67629__llikChisq_me5y
#> #define _llikTDsd _rxd0205c04379a33168bab6df21f717fe67630__llikTDsd_bV8E
#> #define _llikTDmean _rxd0205c04379a33168bab6df21f717fe67631__llikTDmean_uZ1x
#> #define _llikTDdf _rxd0205c04379a33168bab6df21f717fe67632__llikTDdf_TQYO
#> #define _llikT _rxd0205c04379a33168bab6df21f717fe67633__llikT_Qe6a
#> #define _llikBetaDshape2 _rxd0205c04379a33168bab6df21f717fe67634__llikBetaDshape2_qt3L
#> #define _llikBetaDshape1 _rxd0205c04379a33168bab6df21f717fe67635__llikBetaDshape1_bsbe
#> #define _llikBeta _rxd0205c04379a33168bab6df21f717fe67636__llikBeta_aw5m
#> #define _llikNbinomMuDmu _rxd0205c04379a33168bab6df21f717fe67637__llikNbinomMuDmu_4brD
#> #define _llikNbinomMu _rxd0205c04379a33168bab6df21f717fe67638__llikNbinomMu_9EZA
#> #define _llikNbinomDprob _rxd0205c04379a33168bab6df21f717fe67639__llikNbinomDprob_JNe7
#> #define _llikNbinom _rxd0205c04379a33168bab6df21f717fe67640__llikNbinom_IxUF
#> #define _llikBinomDprob _rxd0205c04379a33168bab6df21f717fe67641__llikBinomDprob_A5R2
#> #define _llikBinom _rxd0205c04379a33168bab6df21f717fe67642__llikBinom_wZww
#> #define _llikPoisDlambda _rxd0205c04379a33168bab6df21f717fe67643__llikPoisDlambda_BgRv
#> #define _llikPois _rxd0205c04379a33168bab6df21f717fe67644__llikPois_blnk
#> #define _llikNormDsd _rxd0205c04379a33168bab6df21f717fe67645__llikNormDsd_357c
#> #define _llikNormDmean _rxd0205c04379a33168bab6df21f717fe67646__llikNormDmean_fUwM
#> #define _llikNorm _rxd0205c04379a33168bab6df21f717fe67647__llikNorm_arEA
#> #define simeps _rxd0205c04379a33168bab6df21f717fe67648_simeps_9z1S
#> #define simeta _rxd0205c04379a33168bab6df21f717fe67649_simeta_E1Kz
#> #define expit _rxd0205c04379a33168bab6df21f717fe67650_expit_Af2H
#> #define logit _rxd0205c04379a33168bab6df21f717fe67651_logit_4Q2o
#> #define gammapDer _rxd0205c04379a33168bab6df21f717fe67652_gammapDer_F0jo
#> #define lowergamma _rxd0205c04379a33168bab6df21f717fe67653_lowergamma_CGj0
#> #define uppergamma _rxd0205c04379a33168bab6df21f717fe67654_uppergamma_a8wA
#> #define gammaqInva _rxd0205c04379a33168bab6df21f717fe67655_gammaqInva_bQEL
#> #define gammaqInv _rxd0205c04379a33168bab6df21f717fe67656_gammaqInv_OlOb
#> #define gammapInva _rxd0205c04379a33168bab6df21f717fe67657_gammapInva_qNat
#> #define gammapInv _rxd0205c04379a33168bab6df21f717fe67658_gammapInv_hlti
#> #define gammaq _rxd0205c04379a33168bab6df21f717fe67659_gammaq_WR2I
#> #define gammap _rxd0205c04379a33168bab6df21f717fe67660_gammap_GVLQ
#> #define phi _rxd0205c04379a33168bab6df21f717fe67661_phi_YM84
#> #define d2ELUa _rxd0205c04379a33168bab6df21f717fe67662_d2ELUa_JCnu
#> #define dELUa _rxd0205c04379a33168bab6df21f717fe67663_dELUa_yCXP
#> #define d2aELU _rxd0205c04379a33168bab6df21f717fe67664_d2aELU_JUjv
#> #define d2ELU _rxd0205c04379a33168bab6df21f717fe67665_d2ELU_Fk7K
#> #define dELU _rxd0205c04379a33168bab6df21f717fe67666_dELU_JvDx
#> #define ELU _rxd0205c04379a33168bab6df21f717fe67667_ELU_qbVS
#> #define dPReLUa1 _rxd0205c04379a33168bab6df21f717fe67668_dPReLUa1_KVfr
#> #define dPReLUa _rxd0205c04379a33168bab6df21f717fe67669_dPReLUa_7Tca
#> #define dPReLU _rxd0205c04379a33168bab6df21f717fe67670_dPReLU_UEgp
#> #define PReLU _rxd0205c04379a33168bab6df21f717fe67671_PReLU_YkNI
#> #define dSwish _rxd0205c04379a33168bab6df21f717fe67672_dSwish_qnEn
#> #define Swish _rxd0205c04379a33168bab6df21f717fe67673_Swish_mhSb
#> #define dlReLU _rxd0205c04379a33168bab6df21f717fe67674_dlReLU_oMZY
#> #define lReLU _rxd0205c04379a33168bab6df21f717fe67675_lReLU_Bjki
#> #define dSELU _rxd0205c04379a33168bab6df21f717fe67676_dSELU_CoUq
#> #define SELU _rxd0205c04379a33168bab6df21f717fe67677_SELU_vniI
#> #define d4softplus _rxd0205c04379a33168bab6df21f717fe67678_d4softplus_75Ui
#> #define d3softplus _rxd0205c04379a33168bab6df21f717fe67679_d3softplus_0GVT
#> #define d2softplus _rxd0205c04379a33168bab6df21f717fe67680_d2softplus_s34T
#> #define dsoftplus _rxd0205c04379a33168bab6df21f717fe67681_dsoftplus_NHeF
#> #define softplus _rxd0205c04379a33168bab6df21f717fe67682_softplus_pue1
#> #define d4GELU _rxd0205c04379a33168bab6df21f717fe67683_d4GELU_ZaEn
#> #define d3GELU _rxd0205c04379a33168bab6df21f717fe67684_d3GELU_JR6U
#> #define d2GELU _rxd0205c04379a33168bab6df21f717fe67685_d2GELU_HZKb
#> #define dGELU _rxd0205c04379a33168bab6df21f717fe67686_dGELU_SsmN
#> #define GELU _rxd0205c04379a33168bab6df21f717fe67687_GELU_YlS2
#> #define dReLU _rxd0205c04379a33168bab6df21f717fe67688_dReLU_wzYq
#> #define ReLU _rxd0205c04379a33168bab6df21f717fe67689_ReLU_w2z8
#> #define riweibull _rxd0205c04379a33168bab6df21f717fe67690_riweibull_bGMd
#> #define riunif _rxd0205c04379a33168bab6df21f717fe67691_riunif_jLB3
#> #define rit_ _rxd0205c04379a33168bab6df21f717fe67692_rit__LsOU
#> #define ripois _rxd0205c04379a33168bab6df21f717fe67693_ripois_RbPY
#> #define ribeta _rxd0205c04379a33168bab6df21f717fe67694_ribeta_UPmW
#> #define rigamma _rxd0205c04379a33168bab6df21f717fe67695_rigamma_8sOm
#> #define rigeom _rxd0205c04379a33168bab6df21f717fe67696_rigeom_GypO
#> #define rif _rxd0205c04379a33168bab6df21f717fe67697_rif_55On
#> #define riexp _rxd0205c04379a33168bab6df21f717fe67698_riexp_OE9T
#> #define richisq _rxd0205c04379a33168bab6df21f717fe67699_richisq_FyOK
#> #define ricauchy _rxd0205c04379a33168bab6df21f717fe67700_ricauchy_7OqE
#> #define rinbinomMu _rxd0205c04379a33168bab6df21f717fe67701_rinbinomMu_KwH4
#> #define rinbinom _rxd0205c04379a33168bab6df21f717fe67702_rinbinom_TtSJ
#> #define ribinom _rxd0205c04379a33168bab6df21f717fe67703_ribinom_j0v8
#> #define rinorm _rxd0205c04379a33168bab6df21f717fe67704_rinorm_jxwn
#> #define rxweibull _rxd0205c04379a33168bab6df21f717fe67705_rxweibull_n6QB
#> #define rxunif _rxd0205c04379a33168bab6df21f717fe67706_rxunif_DUzz
#> #define rxt_ _rxd0205c04379a33168bab6df21f717fe67707_rxt__soz5
#> #define rxpois _rxd0205c04379a33168bab6df21f717fe67708_rxpois_n36W
#> #define rxbeta _rxd0205c04379a33168bab6df21f717fe67709_rxbeta_22yJ
#> #define rxgamma _rxd0205c04379a33168bab6df21f717fe67710_rxgamma_1Kel
#> #define rxgeom _rxd0205c04379a33168bab6df21f717fe67711_rxgeom_qu4n
#> #define rxf _rxd0205c04379a33168bab6df21f717fe67712_rxf_1MVL
#> #define rxexp _rxd0205c04379a33168bab6df21f717fe67713_rxexp_5wsv
#> #define rxchisq _rxd0205c04379a33168bab6df21f717fe67714_rxchisq_sAVK
#> #define rxcauchy _rxd0205c04379a33168bab6df21f717fe67715_rxcauchy_VQgW
#> #define rxnbinomMu _rxd0205c04379a33168bab6df21f717fe67716_rxnbinomMu_t0j8
#> #define rxnbinom _rxd0205c04379a33168bab6df21f717fe67717_rxnbinom_P1Sq
#> #define rxbinom _rxd0205c04379a33168bab6df21f717fe67718_rxbinom_l14D
#> #define rxnorm _rxd0205c04379a33168bab6df21f717fe67719_rxnorm_Ab0z
#> #define linCmtB _rxd0205c04379a33168bab6df21f717fe67720_linCmtB_qPsd
#> #define linCmtA _rxd0205c04379a33168bab6df21f717fe67721_linCmtA_CCuo
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){}
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){}
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  rx_4239c6a8c2257921babd7950b768e607__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *rx_4239c6a8c2257921babd7950b768e607__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP rx_4239c6a8c2257921babd7950b768e607__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void rx_4239c6a8c2257921babd7950b768e607__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double a = NA_REAL;
#> 
#>   double __udf[2];
#>   a = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   a =rxweibull(&_solveData->subjects[_cSub], 1,(double)3);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void rx_4239c6a8c2257921babd7950b768e607__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void rx_4239c6a8c2257921babd7950b768e607__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void rx_4239c6a8c2257921babd7950b768e607__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double a = NA_REAL;
#> 
#>   double __udf[2];
#>   a = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   a =rxweibull(&_solveData->subjects[_cSub], 1,(double)3);
#> 
#>   _lhs[_LHS_0_]=a;
#> }
#> // Functional based bioavailability
#> double rx_4239c6a8c2257921babd7950b768e607__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double rx_4239c6a8c2257921babd7950b768e607__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double rx_4239c6a8c2257921babd7950b768e607__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double rx_4239c6a8c2257921babd7950b768e607__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void rx_4239c6a8c2257921babd7950b768e607__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void rx_4239c6a8c2257921babd7950b768e607__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void rx_4239c6a8c2257921babd7950b768e607__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP rx_4239c6a8c2257921babd7950b768e607__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_4239c6a8c2257921babd7950b768e607__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAA@Q0AAAv7#aT)qmF[MA:+}+=WjmaB~15wu1}LVR%%FjeNV`j)W=xZf.ZA:cY1X`/KY9\?F<}(o$rI{QJ)0v3NRGK~Gw+Jh6u*B2:7ij1WXL,hQAv%%Kg$X@9:bmcrDsnSwq!UJ=9(l*>Ws8Vk|yh}.!7{kVp<;\"s\"vd=y`Ojcl`b%%mUM7EFQ,wHsT)\"ZX3UL8{9)da4E%%b_BQlv:u/w&k0v\?eYA%%`fZ&,TjeD/W:1fWm>kL_R%%Pl4gLk53;mp9;$sfl#wziWDX)W}X1X+La*^0))d3X4iS[P1`LlNkU#|m>#X:u~4,A+wmxSU1zMo{f\?4g90e;AxQ%%&NY3=>L~x7Nt{$L)E5OHP0sNR0|y%%ZJO;]JHId9=`Fqz~Y5F1BRup0j{g8pCs`7}&Y+Z\?J.a1jLM\"y[|#F#Ks+vXRY0SbCYP^iI|VMP]WtRJLxlwepW5GYfOlUrvM&iW__BNho6;MM3&6%%]FU`5XxO%%>K@[3kG9<sWxawvt:/cl/}m/65Pm:{[l8mqqY{=,bx~E\?U{QrNe4%%YHmq%%T\".mMI(#(~0J(d/.y\?[70<~X(=F2e/+B)Msd!TDgQAAnB%%A:ovU[L\?WnLQt7puI}B3bF[L7Pv.D&+eHy4Spyqta_@}*Z0h]R@Tg_58f:Z,oOn>8rcnGS^`CV>S(sC\?<@L6[Ho1w0T)o$}Gdy42hw%%O{K6qg:1q!L_Y(dgB:;P||V#.5ntrpLx}h4Q3$ahCZT^Wl:k/PgV+8FqojAi\?riLS7$jFU^j.Yo\"K.imH[C\"bLs63j1WDtt]`cjL%%(HDlGLDzBaX%%mnU^FV::,7\?6y{/e$\?9Z\?\"H<FVn,WGi&_FN!BYOHpP63>J.s{sDER`(3EHVv;:(m*YyMl3\"QEf;obx:;/x}vFf\?eFmhLbnx!5FJftHwH[(:E\"gS{0D;HD+Y)9^Mgn,\"4ClFB&!aLLQ{*n88|eUwHIt%%\"dgZbm[JvKrCYkR&pKjYrWoXQF+f[16PuEi*2oIc<@W)z@:DcP8EFS9x1OF|el}o`w6I5l~kCoKE(GQo**8E~1B+KB");
#>     char buf[1078];
#> #define __doBufN__ 1078
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void rx_4239c6a8c2257921babd7950b768e607__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   rx_4239c6a8c2257921babd7950b768e607__dydt(neq, *t, A, DADT);
#> }
#> extern int rx_4239c6a8c2257921babd7950b768e607__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   rx_4239c6a8c2257921babd7950b768e607__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void rx_4239c6a8c2257921babd7950b768e607__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   rx_4239c6a8c2257921babd7950b768e607__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void rx_4239c6a8c2257921babd7950b768e607__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_rx_4239c6a8c2257921babd7950b768e607_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("rx_4239c6a8c2257921babd7950b768e607_","rx_4239c6a8c2257921babd7950b768e607__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("rx_4239c6a8c2257921babd7950b768e607_","rx_4239c6a8c2257921babd7950b768e607__assignFuns", (DL_FUNC) rx_4239c6a8c2257921babd7950b768e607__assignFuns);
#>   R_RegisterCCallable("rx_4239c6a8c2257921babd7950b768e607_","rx_4239c6a8c2257921babd7950b768e607__inis",(DL_FUNC) rx_4239c6a8c2257921babd7950b768e607__inis);
#>   R_RegisterCCallable("rx_4239c6a8c2257921babd7950b768e607_","rx_4239c6a8c2257921babd7950b768e607__dydt",(DL_FUNC) rx_4239c6a8c2257921babd7950b768e607__dydt);
#>   R_RegisterCCallable("rx_4239c6a8c2257921babd7950b768e607_","rx_4239c6a8c2257921babd7950b768e607__calc_lhs",(DL_FUNC) rx_4239c6a8c2257921babd7950b768e607__calc_lhs);
#>   R_RegisterCCallable("rx_4239c6a8c2257921babd7950b768e607_","rx_4239c6a8c2257921babd7950b768e607__calc_jac",(DL_FUNC) rx_4239c6a8c2257921babd7950b768e607__calc_jac);
#>   R_RegisterCCallable("rx_4239c6a8c2257921babd7950b768e607_","rx_4239c6a8c2257921babd7950b768e607__dydt_lsoda", (DL_FUNC) rx_4239c6a8c2257921babd7950b768e607__dydt_lsoda);
#>   R_RegisterCCallable("rx_4239c6a8c2257921babd7950b768e607_","rx_4239c6a8c2257921babd7950b768e607__calc_jac_lsoda", (DL_FUNC) rx_4239c6a8c2257921babd7950b768e607__calc_jac_lsoda);
#>   R_RegisterCCallable("rx_4239c6a8c2257921babd7950b768e607_","rx_4239c6a8c2257921babd7950b768e607__ode_solver_solvedata", (DL_FUNC) rx_4239c6a8c2257921babd7950b768e607__ode_solver_solvedata);
#>   R_RegisterCCallable("rx_4239c6a8c2257921babd7950b768e607_","rx_4239c6a8c2257921babd7950b768e607__ode_solver_get_solvedata", (DL_FUNC) rx_4239c6a8c2257921babd7950b768e607__ode_solver_get_solvedata);
#>   R_RegisterCCallable("rx_4239c6a8c2257921babd7950b768e607_","rx_4239c6a8c2257921babd7950b768e607__F", (DL_FUNC) rx_4239c6a8c2257921babd7950b768e607__F);
#>   R_RegisterCCallable("rx_4239c6a8c2257921babd7950b768e607_","rx_4239c6a8c2257921babd7950b768e607__Lag", (DL_FUNC) rx_4239c6a8c2257921babd7950b768e607__Lag);
#>   R_RegisterCCallable("rx_4239c6a8c2257921babd7950b768e607_","rx_4239c6a8c2257921babd7950b768e607__Rate", (DL_FUNC) rx_4239c6a8c2257921babd7950b768e607__Rate);
#>   R_RegisterCCallable("rx_4239c6a8c2257921babd7950b768e607_","rx_4239c6a8c2257921babd7950b768e607__Dur", (DL_FUNC) rx_4239c6a8c2257921babd7950b768e607__Dur);
#>   R_RegisterCCallable("rx_4239c6a8c2257921babd7950b768e607_","rx_4239c6a8c2257921babd7950b768e607__mtime", (DL_FUNC) rx_4239c6a8c2257921babd7950b768e607__mtime);
#>   R_RegisterCCallable("rx_4239c6a8c2257921babd7950b768e607_","rx_4239c6a8c2257921babd7950b768e607__ME", (DL_FUNC) rx_4239c6a8c2257921babd7950b768e607__ME);
#>   R_RegisterCCallable("rx_4239c6a8c2257921babd7950b768e607_","rx_4239c6a8c2257921babd7950b768e607__IndF", (DL_FUNC) rx_4239c6a8c2257921babd7950b768e607__IndF);
#>   R_RegisterCCallable("rx_4239c6a8c2257921babd7950b768e607_","rx_4239c6a8c2257921babd7950b768e607__dydt_liblsoda", (DL_FUNC) rx_4239c6a8c2257921babd7950b768e607__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_rx_4239c6a8c2257921babd7950b768e607_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_rx_4239c6a8c2257921babd7950b768e607_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"rx_4239c6a8c2257921babd7950b768e607__model_vars", (DL_FUNC) &rx_4239c6a8c2257921babd7950b768e607__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_rx_4239c6a8c2257921babd7950b768e607_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_4239c6a8c2257921babd7950b768e607__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("rx_4239c6a8c2257921babd7950b768e607__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’
# }