Softplus Activation Function

Softplus Activation Function

Usage

softplus(x)

Arguments

x

numeric vector

Value

numeric vector

See also

Author

Matthew L. Fidler

Examples

softplus(c(-1, 0, 1, 2))
#> [1] 0.3132617 0.6931472 1.3132617 2.1269280

# You can use rxode2 too:

r <- rxode2({
 s <- softplus(time)
})
#>  
#>  
#> C model:
#> // Define 1 LHS values
#> #define _LHS_0_ 0 // s
#> #define _getRxSolve_ _rx961b336d3c362fbedecba47449bdc3687722__getRxSolve__hefg
#> #define _evalUdf _rx961b336d3c362fbedecba47449bdc3687723__evalUdf_hDrm
#> #define _solveData _rx961b336d3c362fbedecba47449bdc3687724__solveData_85ig
#> #define _assign_ptr _rx961b336d3c362fbedecba47449bdc3687725__assign_ptr_vlJY
#> #define _rxRmModelLib _rx961b336d3c362fbedecba47449bdc3687726__rxRmModelLib_gGqQ
#> #define _rxGetModelLib _rx961b336d3c362fbedecba47449bdc3687727__rxGetModelLib_xLoY
#> #define _old_c _rx961b336d3c362fbedecba47449bdc3687728__old_c_lFgJ
#> #define _ptrid _rx961b336d3c362fbedecba47449bdc3687729__ptrid_dKaF
#> #define _rxIsCurrentC _rx961b336d3c362fbedecba47449bdc3687730__rxIsCurrentC_gmDR
#> #define _sumPS _rx961b336d3c362fbedecba47449bdc3687731__sumPS_4jxu
#> #define _prodPS _rx961b336d3c362fbedecba47449bdc3687732__prodPS_oP8g
#> #define _prodType _rx961b336d3c362fbedecba47449bdc3687733__prodType_moUo
#> #define _sumType _rx961b336d3c362fbedecba47449bdc3687734__sumType_mwLt
#> #define _update_par_ptr _rx961b336d3c362fbedecba47449bdc3687735__update_par_ptr_uDgO
#> #define _getParCov _rx961b336d3c362fbedecba47449bdc3687736__getParCov_ztPZ
#> #define _rxode2_rxAssignPtr _rx961b336d3c362fbedecba47449bdc3687737__rxode2_rxAssignPtr_wDqx
#> #define _rxQr _rx961b336d3c362fbedecba47449bdc3687738__rxQr_2t0F
#> #define _compareFactorVal _rx961b336d3c362fbedecba47449bdc3687739__compareFactorVal_fCRu
#> #define _sum _rx961b336d3c362fbedecba47449bdc3687740__sum_pBfc
#> #define _udf _rx961b336d3c362fbedecba47449bdc3687741__udf_LLlJ
#> #define _sign _rx961b336d3c362fbedecba47449bdc3687742__sign_k0HO
#> #define _prod _rx961b336d3c362fbedecba47449bdc3687743__prod_uquM
#> #define _max _rx961b336d3c362fbedecba47449bdc3687744__max_QIP4
#> #define _min _rx961b336d3c362fbedecba47449bdc3687745__min_sBKw
#> #define _transit4P _rx961b336d3c362fbedecba47449bdc3687746__transit4P_lION
#> #define _transit3P _rx961b336d3c362fbedecba47449bdc3687747__transit3P_1tTX
#> #define _assignFuns0 _rx961b336d3c362fbedecba47449bdc3687748__assignFuns0_5Dul
#> #define _assignFuns _rx961b336d3c362fbedecba47449bdc3687749__assignFuns_HLxe
#> #define _rxord _rx961b336d3c362fbedecba47449bdc3687750__rxord_AUt0
#> #define __assignFuns2 _rx961b336d3c362fbedecba47449bdc3687751___assignFuns2_po6A
#> #define _llikCauchyDscale _rx961b336d3c362fbedecba47449bdc3687752__llikCauchyDscale_qSQk
#> #define _llikCauchyDlocation _rx961b336d3c362fbedecba47449bdc3687753__llikCauchyDlocation_nH8f
#> #define _llikCauchy _rx961b336d3c362fbedecba47449bdc3687754__llikCauchy_PCAn
#> #define _llikGammaDrate _rx961b336d3c362fbedecba47449bdc3687755__llikGammaDrate_NkUm
#> #define _llikGammaDshape _rx961b336d3c362fbedecba47449bdc3687756__llikGammaDshape_swiO
#> #define _llikGamma _rx961b336d3c362fbedecba47449bdc3687757__llikGamma_WG16
#> #define _llikWeibullDscale _rx961b336d3c362fbedecba47449bdc3687758__llikWeibullDscale_HPUg
#> #define _llikWeibullDshape _rx961b336d3c362fbedecba47449bdc3687759__llikWeibullDshape_nCju
#> #define _llikWeibull _rx961b336d3c362fbedecba47449bdc3687760__llikWeibull_wA8i
#> #define _llikUnifDbeta _rx961b336d3c362fbedecba47449bdc3687761__llikUnifDbeta_wbn2
#> #define _llikUnifDalpha _rx961b336d3c362fbedecba47449bdc3687762__llikUnifDalpha_GTQj
#> #define _llikUnif _rx961b336d3c362fbedecba47449bdc3687763__llikUnif_Srer
#> #define _llikGeomDp _rx961b336d3c362fbedecba47449bdc3687764__llikGeomDp_5zex
#> #define _llikGeom _rx961b336d3c362fbedecba47449bdc3687765__llikGeom_LgAw
#> #define _llikFDdf2 _rx961b336d3c362fbedecba47449bdc3687766__llikFDdf2_ME6n
#> #define _llikFDdf1 _rx961b336d3c362fbedecba47449bdc3687767__llikFDdf1_qR3D
#> #define _llikF _rx961b336d3c362fbedecba47449bdc3687768__llikF_D9jN
#> #define _llikExpDrate _rx961b336d3c362fbedecba47449bdc3687769__llikExpDrate_5O4r
#> #define _llikExp _rx961b336d3c362fbedecba47449bdc3687770__llikExp_9D97
#> #define _llikChisqDdf _rx961b336d3c362fbedecba47449bdc3687771__llikChisqDdf_me5y
#> #define _llikChisq _rx961b336d3c362fbedecba47449bdc3687772__llikChisq_bV8E
#> #define _llikTDsd _rx961b336d3c362fbedecba47449bdc3687773__llikTDsd_uZ1x
#> #define _llikTDmean _rx961b336d3c362fbedecba47449bdc3687774__llikTDmean_TQYO
#> #define _llikTDdf _rx961b336d3c362fbedecba47449bdc3687775__llikTDdf_Qe6a
#> #define _llikT _rx961b336d3c362fbedecba47449bdc3687776__llikT_qt3L
#> #define _llikBetaDshape2 _rx961b336d3c362fbedecba47449bdc3687777__llikBetaDshape2_bsbe
#> #define _llikBetaDshape1 _rx961b336d3c362fbedecba47449bdc3687778__llikBetaDshape1_aw5m
#> #define _llikBeta _rx961b336d3c362fbedecba47449bdc3687779__llikBeta_4brD
#> #define _llikNbinomMuDmu _rx961b336d3c362fbedecba47449bdc3687780__llikNbinomMuDmu_9EZA
#> #define _llikNbinomMu _rx961b336d3c362fbedecba47449bdc3687781__llikNbinomMu_JNe7
#> #define _llikNbinomDprob _rx961b336d3c362fbedecba47449bdc3687782__llikNbinomDprob_IxUF
#> #define _llikNbinom _rx961b336d3c362fbedecba47449bdc3687783__llikNbinom_A5R2
#> #define _llikBinomDprob _rx961b336d3c362fbedecba47449bdc3687784__llikBinomDprob_wZww
#> #define _llikBinom _rx961b336d3c362fbedecba47449bdc3687785__llikBinom_BgRv
#> #define _llikPoisDlambda _rx961b336d3c362fbedecba47449bdc3687786__llikPoisDlambda_blnk
#> #define _llikPois _rx961b336d3c362fbedecba47449bdc3687787__llikPois_357c
#> #define _llikNormDsd _rx961b336d3c362fbedecba47449bdc3687788__llikNormDsd_fUwM
#> #define _llikNormDmean _rx961b336d3c362fbedecba47449bdc3687789__llikNormDmean_arEA
#> #define _llikNorm _rx961b336d3c362fbedecba47449bdc3687790__llikNorm_9z1S
#> #define simeps _rx961b336d3c362fbedecba47449bdc3687791_simeps_E1Kz
#> #define simeta _rx961b336d3c362fbedecba47449bdc3687792_simeta_Af2H
#> #define expit _rx961b336d3c362fbedecba47449bdc3687793_expit_4Q2o
#> #define logit _rx961b336d3c362fbedecba47449bdc3687794_logit_F0jo
#> #define gammapDer _rx961b336d3c362fbedecba47449bdc3687795_gammapDer_CGj0
#> #define lowergamma _rx961b336d3c362fbedecba47449bdc3687796_lowergamma_a8wA
#> #define uppergamma _rx961b336d3c362fbedecba47449bdc3687797_uppergamma_bQEL
#> #define gammaqInva _rx961b336d3c362fbedecba47449bdc3687798_gammaqInva_OlOb
#> #define gammaqInv _rx961b336d3c362fbedecba47449bdc3687799_gammaqInv_qNat
#> #define gammapInva _rx961b336d3c362fbedecba47449bdc3687800_gammapInva_hlti
#> #define gammapInv _rx961b336d3c362fbedecba47449bdc3687801_gammapInv_WR2I
#> #define gammaq _rx961b336d3c362fbedecba47449bdc3687802_gammaq_GVLQ
#> #define gammap _rx961b336d3c362fbedecba47449bdc3687803_gammap_YM84
#> #define phi _rx961b336d3c362fbedecba47449bdc3687804_phi_JCnu
#> #define d2ELUa _rx961b336d3c362fbedecba47449bdc3687805_d2ELUa_yCXP
#> #define dELUa _rx961b336d3c362fbedecba47449bdc3687806_dELUa_JUjv
#> #define d2aELU _rx961b336d3c362fbedecba47449bdc3687807_d2aELU_Fk7K
#> #define d2ELU _rx961b336d3c362fbedecba47449bdc3687808_d2ELU_JvDx
#> #define dELU _rx961b336d3c362fbedecba47449bdc3687809_dELU_qbVS
#> #define ELU _rx961b336d3c362fbedecba47449bdc3687810_ELU_KVfr
#> #define dPReLUa1 _rx961b336d3c362fbedecba47449bdc3687811_dPReLUa1_7Tca
#> #define dPReLUa _rx961b336d3c362fbedecba47449bdc3687812_dPReLUa_UEgp
#> #define dPReLU _rx961b336d3c362fbedecba47449bdc3687813_dPReLU_YkNI
#> #define PReLU _rx961b336d3c362fbedecba47449bdc3687814_PReLU_qnEn
#> #define dSwish _rx961b336d3c362fbedecba47449bdc3687815_dSwish_mhSb
#> #define Swish _rx961b336d3c362fbedecba47449bdc3687816_Swish_oMZY
#> #define dlReLU _rx961b336d3c362fbedecba47449bdc3687817_dlReLU_Bjki
#> #define lReLU _rx961b336d3c362fbedecba47449bdc3687818_lReLU_CoUq
#> #define dSELU _rx961b336d3c362fbedecba47449bdc3687819_dSELU_vniI
#> #define SELU _rx961b336d3c362fbedecba47449bdc3687820_SELU_75Ui
#> #define d4softplus _rx961b336d3c362fbedecba47449bdc3687821_d4softplus_0GVT
#> #define d3softplus _rx961b336d3c362fbedecba47449bdc3687822_d3softplus_s34T
#> #define d2softplus _rx961b336d3c362fbedecba47449bdc3687823_d2softplus_NHeF
#> #define dsoftplus _rx961b336d3c362fbedecba47449bdc3687824_dsoftplus_pue1
#> #define softplus _rx961b336d3c362fbedecba47449bdc3687825_softplus_ZaEn
#> #define d4GELU _rx961b336d3c362fbedecba47449bdc3687826_d4GELU_JR6U
#> #define d3GELU _rx961b336d3c362fbedecba47449bdc3687827_d3GELU_HZKb
#> #define d2GELU _rx961b336d3c362fbedecba47449bdc3687828_d2GELU_SsmN
#> #define dGELU _rx961b336d3c362fbedecba47449bdc3687829_dGELU_YlS2
#> #define GELU _rx961b336d3c362fbedecba47449bdc3687830_GELU_wzYq
#> #define dReLU _rx961b336d3c362fbedecba47449bdc3687831_dReLU_w2z8
#> #define ReLU _rx961b336d3c362fbedecba47449bdc3687832_ReLU_bGMd
#> #define riweibull _rx961b336d3c362fbedecba47449bdc3687833_riweibull_jLB3
#> #define riunif _rx961b336d3c362fbedecba47449bdc3687834_riunif_LsOU
#> #define rit_ _rx961b336d3c362fbedecba47449bdc3687835_rit__RbPY
#> #define ripois _rx961b336d3c362fbedecba47449bdc3687836_ripois_UPmW
#> #define ribeta _rx961b336d3c362fbedecba47449bdc3687837_ribeta_8sOm
#> #define rigamma _rx961b336d3c362fbedecba47449bdc3687838_rigamma_GypO
#> #define rigeom _rx961b336d3c362fbedecba47449bdc3687839_rigeom_55On
#> #define rif _rx961b336d3c362fbedecba47449bdc3687840_rif_OE9T
#> #define riexp _rx961b336d3c362fbedecba47449bdc3687841_riexp_FyOK
#> #define richisq _rx961b336d3c362fbedecba47449bdc3687842_richisq_7OqE
#> #define ricauchy _rx961b336d3c362fbedecba47449bdc3687843_ricauchy_KwH4
#> #define rinbinomMu _rx961b336d3c362fbedecba47449bdc3687844_rinbinomMu_TtSJ
#> #define rinbinom _rx961b336d3c362fbedecba47449bdc3687845_rinbinom_j0v8
#> #define ribinom _rx961b336d3c362fbedecba47449bdc3687846_ribinom_jxwn
#> #define rinorm _rx961b336d3c362fbedecba47449bdc3687847_rinorm_n6QB
#> #define rxweibull _rx961b336d3c362fbedecba47449bdc3687848_rxweibull_DUzz
#> #define rxunif _rx961b336d3c362fbedecba47449bdc3687849_rxunif_soz5
#> #define rxt_ _rx961b336d3c362fbedecba47449bdc3687850_rxt__n36W
#> #define rxpois _rx961b336d3c362fbedecba47449bdc3687851_rxpois_22yJ
#> #define rxbeta _rx961b336d3c362fbedecba47449bdc3687852_rxbeta_1Kel
#> #define rxgamma _rx961b336d3c362fbedecba47449bdc3687853_rxgamma_qu4n
#> #define rxgeom _rx961b336d3c362fbedecba47449bdc3687854_rxgeom_1MVL
#> #define rxf _rx961b336d3c362fbedecba47449bdc3687855_rxf_5wsv
#> #define rxexp _rx961b336d3c362fbedecba47449bdc3687856_rxexp_sAVK
#> #define rxchisq _rx961b336d3c362fbedecba47449bdc3687857_rxchisq_VQgW
#> #define rxcauchy _rx961b336d3c362fbedecba47449bdc3687858_rxcauchy_t0j8
#> #define rxnbinomMu _rx961b336d3c362fbedecba47449bdc3687859_rxnbinomMu_P1Sq
#> #define rxnbinom _rx961b336d3c362fbedecba47449bdc3687860_rxnbinom_l14D
#> #define rxbinom _rx961b336d3c362fbedecba47449bdc3687861_rxbinom_Ab0z
#> #define rxnorm _rx961b336d3c362fbedecba47449bdc3687862_rxnorm_qPsd
#> #define linCmtB _rx961b336d3c362fbedecba47449bdc3687863_linCmtB_CCuo
#> #define linCmtA _rx961b336d3c362fbedecba47449bdc3687864_linCmtA_zWfS
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){}
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){}
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  rx_fcc96f22ad1b918fc501e9380f394892__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *rx_fcc96f22ad1b918fc501e9380f394892__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP rx_fcc96f22ad1b918fc501e9380f394892__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void rx_fcc96f22ad1b918fc501e9380f394892__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double s = NA_REAL;
#> 
#>   double __udf[2];
#>   s = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   s =softplus(t);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void rx_fcc96f22ad1b918fc501e9380f394892__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void rx_fcc96f22ad1b918fc501e9380f394892__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void rx_fcc96f22ad1b918fc501e9380f394892__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double s = NA_REAL;
#> 
#>   double __udf[2];
#>   s = _PL[0];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#> 
#> 
#>   s =softplus(t);
#> 
#>   _lhs[_LHS_0_]=s;
#> }
#> // Functional based bioavailability
#> double rx_fcc96f22ad1b918fc501e9380f394892__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double rx_fcc96f22ad1b918fc501e9380f394892__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double rx_fcc96f22ad1b918fc501e9380f394892__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double rx_fcc96f22ad1b918fc501e9380f394892__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void rx_fcc96f22ad1b918fc501e9380f394892__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void rx_fcc96f22ad1b918fc501e9380f394892__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void rx_fcc96f22ad1b918fc501e9380f394892__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP rx_fcc96f22ad1b918fc501e9380f394892__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_fcc96f22ad1b918fc501e9380f394892__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAAv(zAAAv7#aT)kmT,MA**8+ftGlYBaieE6LIA!xz4x`[W}YS^(l9\?Z6XH5!+jw+szE]y!S!q{uv:!5;X~VN@%%5Y5+qai7Emr#cBkB|L&\"N/sO~E$9YP#GzOFzeBb=CRX=}s`]d9+2q_nvB9~WV)~/S8y>QA9oxiyv6sgy$1M7H_Hk$pR#2,#Ww=rj+K/]R/w5AQ/Iim.AMUlXpBrNE,`z@m{KEPV101&Zy*]@hck\"b1\"Qx0p9,ZAROOgpsM`J@q}*P&X|r#Hwy]3^AU8&tyQ38R<cA)B!XxS$ZleOoQX5D8Dl>HCeNcSUPcfypK~Zm=*8%%pJrCNZTl`bTS*:&z.pJ{OPOyg,Ru:%%`&OskSPh^g0i*\"HWEMAcA_a8T_gg&0BuD9ryNhDAwO&OBE5=grC6E&$7ziG5wgU@OekuUtTB$(TTD<8}wnak*Xx;C<VxH{_t~^xGft[$,#`Lx30cCtmTDIll=z%%ra+r8)3UubS\?vEjoz*=2Eabt.Y:by.TPh)@800D=W%%Vln0!`~:~wqaZ}@Yyt6eF<tU7e$r%%2U][79&k97I#bh7!LIog#QTrD2[&`w)`6Nj^\?\?qnVzKzMaucLOM+U]:.\?ItUkUVrUtAVWWKqkfU&*\"Jp2z5EZ</.lwOsb4ikwQoNYY[B3(*R,\?;v/G14u$CB`{=HNn_z~XOH,<k%%b)51m,P%%oW(p[aKA=~=n]*[w=ZMBwD26b5w2[]wtIfio9&$nJRZ4Nmf[B}uVf@yti6k&GN5fO,zwCPDM01cSF1B&6~z8I)MlR|x(T\?J{xbvmvoS9*tF;44A\?Zg>z~lQ^;$GL``I`JyKA6=1eGT2&8}_^4I(,VaFo6z5dsD<,JJ9U<LS,Sz|8ybuw~Qsp[e=U/Z1tuOzvVQC9\?eqFhGaQZ~DaSTzlj>*hlU2Tr{/1eR/eto\?X1fw[XZh}E<S@w._ABoTuln7.xt5zMdtb^D|Fjc7G7#[ki}Du`pY:LR<Hyvw3gaq;r&L^BfiXC~_dYN(8Ej94cHB}8Lbv.+4>\"C+hxy4nuXWD");
#>     char buf[1060];
#> #define __doBufN__ 1060
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void rx_fcc96f22ad1b918fc501e9380f394892__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   rx_fcc96f22ad1b918fc501e9380f394892__dydt(neq, *t, A, DADT);
#> }
#> extern int rx_fcc96f22ad1b918fc501e9380f394892__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   rx_fcc96f22ad1b918fc501e9380f394892__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void rx_fcc96f22ad1b918fc501e9380f394892__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   rx_fcc96f22ad1b918fc501e9380f394892__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void rx_fcc96f22ad1b918fc501e9380f394892__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_rx_fcc96f22ad1b918fc501e9380f394892_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("rx_fcc96f22ad1b918fc501e9380f394892_","rx_fcc96f22ad1b918fc501e9380f394892__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("rx_fcc96f22ad1b918fc501e9380f394892_","rx_fcc96f22ad1b918fc501e9380f394892__assignFuns", (DL_FUNC) rx_fcc96f22ad1b918fc501e9380f394892__assignFuns);
#>   R_RegisterCCallable("rx_fcc96f22ad1b918fc501e9380f394892_","rx_fcc96f22ad1b918fc501e9380f394892__inis",(DL_FUNC) rx_fcc96f22ad1b918fc501e9380f394892__inis);
#>   R_RegisterCCallable("rx_fcc96f22ad1b918fc501e9380f394892_","rx_fcc96f22ad1b918fc501e9380f394892__dydt",(DL_FUNC) rx_fcc96f22ad1b918fc501e9380f394892__dydt);
#>   R_RegisterCCallable("rx_fcc96f22ad1b918fc501e9380f394892_","rx_fcc96f22ad1b918fc501e9380f394892__calc_lhs",(DL_FUNC) rx_fcc96f22ad1b918fc501e9380f394892__calc_lhs);
#>   R_RegisterCCallable("rx_fcc96f22ad1b918fc501e9380f394892_","rx_fcc96f22ad1b918fc501e9380f394892__calc_jac",(DL_FUNC) rx_fcc96f22ad1b918fc501e9380f394892__calc_jac);
#>   R_RegisterCCallable("rx_fcc96f22ad1b918fc501e9380f394892_","rx_fcc96f22ad1b918fc501e9380f394892__dydt_lsoda", (DL_FUNC) rx_fcc96f22ad1b918fc501e9380f394892__dydt_lsoda);
#>   R_RegisterCCallable("rx_fcc96f22ad1b918fc501e9380f394892_","rx_fcc96f22ad1b918fc501e9380f394892__calc_jac_lsoda", (DL_FUNC) rx_fcc96f22ad1b918fc501e9380f394892__calc_jac_lsoda);
#>   R_RegisterCCallable("rx_fcc96f22ad1b918fc501e9380f394892_","rx_fcc96f22ad1b918fc501e9380f394892__ode_solver_solvedata", (DL_FUNC) rx_fcc96f22ad1b918fc501e9380f394892__ode_solver_solvedata);
#>   R_RegisterCCallable("rx_fcc96f22ad1b918fc501e9380f394892_","rx_fcc96f22ad1b918fc501e9380f394892__ode_solver_get_solvedata", (DL_FUNC) rx_fcc96f22ad1b918fc501e9380f394892__ode_solver_get_solvedata);
#>   R_RegisterCCallable("rx_fcc96f22ad1b918fc501e9380f394892_","rx_fcc96f22ad1b918fc501e9380f394892__F", (DL_FUNC) rx_fcc96f22ad1b918fc501e9380f394892__F);
#>   R_RegisterCCallable("rx_fcc96f22ad1b918fc501e9380f394892_","rx_fcc96f22ad1b918fc501e9380f394892__Lag", (DL_FUNC) rx_fcc96f22ad1b918fc501e9380f394892__Lag);
#>   R_RegisterCCallable("rx_fcc96f22ad1b918fc501e9380f394892_","rx_fcc96f22ad1b918fc501e9380f394892__Rate", (DL_FUNC) rx_fcc96f22ad1b918fc501e9380f394892__Rate);
#>   R_RegisterCCallable("rx_fcc96f22ad1b918fc501e9380f394892_","rx_fcc96f22ad1b918fc501e9380f394892__Dur", (DL_FUNC) rx_fcc96f22ad1b918fc501e9380f394892__Dur);
#>   R_RegisterCCallable("rx_fcc96f22ad1b918fc501e9380f394892_","rx_fcc96f22ad1b918fc501e9380f394892__mtime", (DL_FUNC) rx_fcc96f22ad1b918fc501e9380f394892__mtime);
#>   R_RegisterCCallable("rx_fcc96f22ad1b918fc501e9380f394892_","rx_fcc96f22ad1b918fc501e9380f394892__ME", (DL_FUNC) rx_fcc96f22ad1b918fc501e9380f394892__ME);
#>   R_RegisterCCallable("rx_fcc96f22ad1b918fc501e9380f394892_","rx_fcc96f22ad1b918fc501e9380f394892__IndF", (DL_FUNC) rx_fcc96f22ad1b918fc501e9380f394892__IndF);
#>   R_RegisterCCallable("rx_fcc96f22ad1b918fc501e9380f394892_","rx_fcc96f22ad1b918fc501e9380f394892__dydt_liblsoda", (DL_FUNC) rx_fcc96f22ad1b918fc501e9380f394892__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_rx_fcc96f22ad1b918fc501e9380f394892_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_rx_fcc96f22ad1b918fc501e9380f394892_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"rx_fcc96f22ad1b918fc501e9380f394892__model_vars", (DL_FUNC) &rx_fcc96f22ad1b918fc501e9380f394892__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_rx_fcc96f22ad1b918fc501e9380f394892_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_fcc96f22ad1b918fc501e9380f394892__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("rx_fcc96f22ad1b918fc501e9380f394892__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’

e <- et(c(-1, 0, 1, 2))

rxSolve(r, e)
#> ── Solved rxode2 object ──
#> ── Parameters (value$params): ──
#> # A tibble: 1 × 0
#> ── Initial Conditions (value$inits): ──
#> named numeric(0)
#> ── First part of data (object): ──
#> # A tibble: 4 × 2
#>    time     s
#>   <dbl> <dbl>
#> 1    -1 0.313
#> 2     0 0.693
#> 3     1 1.31 
#> 4     2 2.13