Generate an example (template) of a dosing regimen shiny app

Create a complete shiny application for exploring dosing regimens given a (hardcoded) PK/PD model.

Usage

genShinyApp.template(
  appDir = "shinyExample",
  verbose = TRUE,
  ODE.config = list(ode = "model", params = c(KA = 0.294), inits = c(eff = 1), method =
    "lsoda", atol = 1e-08, rtol = 1e-06)
)

write.template.server(appDir)

write.template.ui(appDir, statevars)

Arguments

appDir

a string with a directory where to store the shiny app, by default is "shinyExample". The directory appDir will be created if it does not exist.

verbose

logical specifying whether to write messages as the shiny app is generated. Defaults to TRUE.

ODE.config

model name compiled and list of parameters sent to rxSolve().

statevars

List of statevars passed to to the write.template.ui() function. This usually isn't called directly.

A PK/PD model is defined using rxode2(), and a set of parameters and initial values are defined. Then the appropriate R scripts for the shiny's user interface ui.R and the server logic server.R are created in the directory appDir.

The function evaluates the following PK/PD model by default:

    C2 = centr/V2;
    C3 = peri/V3;
    d/dt(depot) =-KA*depot;
    d/dt(centr) = KA*depot - CL*C2 - Q*C2 + Q*C3;
    d/dt(peri)  =                    Q*C2 - Q*C3;
    d/dt(eff)  = Kin - Kout*(1-C2/(EC50+C2))*eff;

This can be changed by the ODE.config parameter.

To launch the shiny app, simply issue the runApp(appDir) R command.

Value

None, these functions are used for their side effects.

Note

These functions create a simple, but working example of a dosing regimen simulation web application. Users may want to modify the code to experiment creating shiny applications for their specific rxode2 models.

See also

Examples

# \donttest{
# remove myapp when the example is complete
on.exit(unlink("myapp", recursive = TRUE, force = TRUE))
# create the shiny app example (template)
genShinyApp.template(appDir = "myapp")
#> 
#> Generating an example (template) for a dosing regimen shiny app
#> 
#> Using the following PK/PD model:
#>    C2 = centr/V2;
#>    C3 = peri/V3;
#>    d/dt(depot) =-KA*depot;
#>    d/dt(centr) = KA*depot - CL*C2 - Q*C2 + Q*C3;
#>    d/dt(peri)  =                    Q*C2 - Q*C3;
#>    d/dt(eff)  = Kin - Kout*(1-C2/(EC50+C2))*eff;
#>     
#> Translating the PK/PD ODE model into C, compiling, etc.
#> 
#>  
#>  
#> C model:
#> // Define translation state order for 4 states
#> #define __DDT0__ 0 // depot
#> #define _DEPOT_ 0 // depot
#> #define __DDT1__ 1 // centr
#> #define __DDT2__ 2 // peri
#> #define __DDT3__ 3 // eff
#> // Define 2 LHS values
#> #define _LHS_0_ 0 // C2
#> #define _LHS_1_ 1 // C3
#> #define _getRxSolve_ _rxmod1_b8dc402a58f4ef28428c503ca0c870192002__getRxSolve__MrOc
#> #define _evalUdf _rxmod1_b8dc402a58f4ef28428c503ca0c870192003__evalUdf_lBdv
#> #define _solveData _rxmod1_b8dc402a58f4ef28428c503ca0c870192004__solveData_cHYH
#> #define _assign_ptr _rxmod1_b8dc402a58f4ef28428c503ca0c870192005__assign_ptr_lORz
#> #define _rxRmModelLib _rxmod1_b8dc402a58f4ef28428c503ca0c870192006__rxRmModelLib_kg9p
#> #define _rxGetModelLib _rxmod1_b8dc402a58f4ef28428c503ca0c870192007__rxGetModelLib_hZjb
#> #define _old_c _rxmod1_b8dc402a58f4ef28428c503ca0c870192008__old_c_tvsF
#> #define _ptrid _rxmod1_b8dc402a58f4ef28428c503ca0c870192009__ptrid_QkpV
#> #define _rxIsCurrentC _rxmod1_b8dc402a58f4ef28428c503ca0c870192010__rxIsCurrentC_HF4t
#> #define _sumPS _rxmod1_b8dc402a58f4ef28428c503ca0c870192011__sumPS_E5bq
#> #define _prodPS _rxmod1_b8dc402a58f4ef28428c503ca0c870192012__prodPS_rWj2
#> #define _prodType _rxmod1_b8dc402a58f4ef28428c503ca0c870192013__prodType_jIRW
#> #define _sumType _rxmod1_b8dc402a58f4ef28428c503ca0c870192014__sumType_0crr
#> #define _update_par_ptr _rxmod1_b8dc402a58f4ef28428c503ca0c870192015__update_par_ptr_gcGm
#> #define _getParCov _rxmod1_b8dc402a58f4ef28428c503ca0c870192016__getParCov_EsJu
#> #define _rxode2_rxAssignPtr _rxmod1_b8dc402a58f4ef28428c503ca0c870192017__rxode2_rxAssignPtr_9xah
#> #define _rxQr _rxmod1_b8dc402a58f4ef28428c503ca0c870192018__rxQr_gBOZ
#> #define _compareFactorVal _rxmod1_b8dc402a58f4ef28428c503ca0c870192019__compareFactorVal_l3zY
#> #define _sum _rxmod1_b8dc402a58f4ef28428c503ca0c870192020__sum_vTa9
#> #define _udf _rxmod1_b8dc402a58f4ef28428c503ca0c870192021__udf_7EhW
#> #define _sign _rxmod1_b8dc402a58f4ef28428c503ca0c870192022__sign_3MG9
#> #define _prod _rxmod1_b8dc402a58f4ef28428c503ca0c870192023__prod_V3pR
#> #define _max _rxmod1_b8dc402a58f4ef28428c503ca0c870192024__max_1aNk
#> #define _min _rxmod1_b8dc402a58f4ef28428c503ca0c870192025__min_G5Jw
#> #define _transit4P _rxmod1_b8dc402a58f4ef28428c503ca0c870192026__transit4P_D9TK
#> #define _transit3P _rxmod1_b8dc402a58f4ef28428c503ca0c870192027__transit3P_40NP
#> #define _assignFuns0 _rxmod1_b8dc402a58f4ef28428c503ca0c870192028__assignFuns0_JiAL
#> #define _assignFuns _rxmod1_b8dc402a58f4ef28428c503ca0c870192029__assignFuns_ZBrh
#> #define _rxord _rxmod1_b8dc402a58f4ef28428c503ca0c870192030__rxord_efgh
#> #define __assignFuns2 _rxmod1_b8dc402a58f4ef28428c503ca0c870192031___assignFuns2_Drm8
#> #define _llikCauchyDscale _rxmod1_b8dc402a58f4ef28428c503ca0c870192032__llikCauchyDscale_5igv
#> #define _llikCauchyDlocation _rxmod1_b8dc402a58f4ef28428c503ca0c870192033__llikCauchyDlocation_lJYg
#> #define _llikCauchy _rxmod1_b8dc402a58f4ef28428c503ca0c870192034__llikCauchy_GqQx
#> #define _llikGammaDrate _rxmod1_b8dc402a58f4ef28428c503ca0c870192035__llikGammaDrate_LoYl
#> #define _llikGammaDshape _rxmod1_b8dc402a58f4ef28428c503ca0c870192036__llikGammaDshape_FgJd
#> #define _llikGamma _rxmod1_b8dc402a58f4ef28428c503ca0c870192037__llikGamma_KaFg
#> #define _llikWeibullDscale _rxmod1_b8dc402a58f4ef28428c503ca0c870192038__llikWeibullDscale_mDR4
#> #define _llikWeibullDshape _rxmod1_b8dc402a58f4ef28428c503ca0c870192039__llikWeibullDshape_jxuo
#> #define _llikWeibull _rxmod1_b8dc402a58f4ef28428c503ca0c870192040__llikWeibull_P8gm
#> #define _llikUnifDbeta _rxmod1_b8dc402a58f4ef28428c503ca0c870192041__llikUnifDbeta_oUom
#> #define _llikUnifDalpha _rxmod1_b8dc402a58f4ef28428c503ca0c870192042__llikUnifDalpha_wLtu
#> #define _llikUnif _rxmod1_b8dc402a58f4ef28428c503ca0c870192043__llikUnif_DgOz
#> #define _llikGeomDp _rxmod1_b8dc402a58f4ef28428c503ca0c870192044__llikGeomDp_tPZw
#> #define _llikGeom _rxmod1_b8dc402a58f4ef28428c503ca0c870192045__llikGeom_Dqx2
#> #define _llikFDdf2 _rxmod1_b8dc402a58f4ef28428c503ca0c870192046__llikFDdf2_t0Ff
#> #define _llikFDdf1 _rxmod1_b8dc402a58f4ef28428c503ca0c870192047__llikFDdf1_CRup
#> #define _llikF _rxmod1_b8dc402a58f4ef28428c503ca0c870192048__llikF_BfcL
#> #define _llikExpDrate _rxmod1_b8dc402a58f4ef28428c503ca0c870192049__llikExpDrate_LlJk
#> #define _llikExp _rxmod1_b8dc402a58f4ef28428c503ca0c870192050__llikExp_0HOu
#> #define _llikChisqDdf _rxmod1_b8dc402a58f4ef28428c503ca0c870192051__llikChisqDdf_quMQ
#> #define _llikChisq _rxmod1_b8dc402a58f4ef28428c503ca0c870192052__llikChisq_IP4s
#> #define _llikTDsd _rxmod1_b8dc402a58f4ef28428c503ca0c870192053__llikTDsd_BKwl
#> #define _llikTDmean _rxmod1_b8dc402a58f4ef28428c503ca0c870192054__llikTDmean_ION1
#> #define _llikTDdf _rxmod1_b8dc402a58f4ef28428c503ca0c870192055__llikTDdf_tTX5
#> #define _llikT _rxmod1_b8dc402a58f4ef28428c503ca0c870192056__llikT_DulH
#> #define _llikBetaDshape2 _rxmod1_b8dc402a58f4ef28428c503ca0c870192057__llikBetaDshape2_LxeA
#> #define _llikBetaDshape1 _rxmod1_b8dc402a58f4ef28428c503ca0c870192058__llikBetaDshape1_Ut0p
#> #define _llikBeta _rxmod1_b8dc402a58f4ef28428c503ca0c870192059__llikBeta_o6Aq
#> #define _llikNbinomMuDmu _rxmod1_b8dc402a58f4ef28428c503ca0c870192060__llikNbinomMuDmu_SQkn
#> #define _llikNbinomMu _rxmod1_b8dc402a58f4ef28428c503ca0c870192061__llikNbinomMu_H8fP
#> #define _llikNbinomDprob _rxmod1_b8dc402a58f4ef28428c503ca0c870192062__llikNbinomDprob_CAnN
#> #define _llikNbinom _rxmod1_b8dc402a58f4ef28428c503ca0c870192063__llikNbinom_kUms
#> #define _llikBinomDprob _rxmod1_b8dc402a58f4ef28428c503ca0c870192064__llikBinomDprob_wiOW
#> #define _llikBinom _rxmod1_b8dc402a58f4ef28428c503ca0c870192065__llikBinom_G16H
#> #define _llikPoisDlambda _rxmod1_b8dc402a58f4ef28428c503ca0c870192066__llikPoisDlambda_PUgn
#> #define _llikPois _rxmod1_b8dc402a58f4ef28428c503ca0c870192067__llikPois_Cjuw
#> #define _llikNormDsd _rxmod1_b8dc402a58f4ef28428c503ca0c870192068__llikNormDsd_A8iw
#> #define _llikNormDmean _rxmod1_b8dc402a58f4ef28428c503ca0c870192069__llikNormDmean_bn2G
#> #define _llikNorm _rxmod1_b8dc402a58f4ef28428c503ca0c870192070__llikNorm_TQjS
#> #define simeps _rxmod1_b8dc402a58f4ef28428c503ca0c870192071_simeps_rer5
#> #define simeta _rxmod1_b8dc402a58f4ef28428c503ca0c870192072_simeta_zexL
#> #define expit _rxmod1_b8dc402a58f4ef28428c503ca0c870192073_expit_gAwM
#> #define logit _rxmod1_b8dc402a58f4ef28428c503ca0c870192074_logit_E6nq
#> #define gammapDer _rxmod1_b8dc402a58f4ef28428c503ca0c870192075_gammapDer_R3DD
#> #define lowergamma _rxmod1_b8dc402a58f4ef28428c503ca0c870192076_lowergamma_9jN5
#> #define uppergamma _rxmod1_b8dc402a58f4ef28428c503ca0c870192077_uppergamma_O4r9
#> #define gammaqInva _rxmod1_b8dc402a58f4ef28428c503ca0c870192078_gammaqInva_D97m
#> #define gammaqInv _rxmod1_b8dc402a58f4ef28428c503ca0c870192079_gammaqInv_e5yb
#> #define gammapInva _rxmod1_b8dc402a58f4ef28428c503ca0c870192080_gammapInva_V8Eu
#> #define gammapInv _rxmod1_b8dc402a58f4ef28428c503ca0c870192081_gammapInv_Z1xT
#> #define gammaq _rxmod1_b8dc402a58f4ef28428c503ca0c870192082_gammaq_QYOQ
#> #define gammap _rxmod1_b8dc402a58f4ef28428c503ca0c870192083_gammap_e6aq
#> #define phi _rxmod1_b8dc402a58f4ef28428c503ca0c870192084_phi_t3Lb
#> #define d2ELUa _rxmod1_b8dc402a58f4ef28428c503ca0c870192085_d2ELUa_sbea
#> #define dELUa _rxmod1_b8dc402a58f4ef28428c503ca0c870192086_dELUa_w5m4
#> #define d2aELU _rxmod1_b8dc402a58f4ef28428c503ca0c870192087_d2aELU_brD9
#> #define d2ELU _rxmod1_b8dc402a58f4ef28428c503ca0c870192088_d2ELU_EZAJ
#> #define dELU _rxmod1_b8dc402a58f4ef28428c503ca0c870192089_dELU_Ne7I
#> #define ELU _rxmod1_b8dc402a58f4ef28428c503ca0c870192090_ELU_xUFA
#> #define dPReLUa1 _rxmod1_b8dc402a58f4ef28428c503ca0c870192091_dPReLUa1_5R2w
#> #define dPReLUa _rxmod1_b8dc402a58f4ef28428c503ca0c870192092_dPReLUa_ZwwB
#> #define dPReLU _rxmod1_b8dc402a58f4ef28428c503ca0c870192093_dPReLU_gRvb
#> #define PReLU _rxmod1_b8dc402a58f4ef28428c503ca0c870192094_PReLU_lnk3
#> #define dSwish _rxmod1_b8dc402a58f4ef28428c503ca0c870192095_dSwish_57cf
#> #define Swish _rxmod1_b8dc402a58f4ef28428c503ca0c870192096_Swish_UwMa
#> #define dlReLU _rxmod1_b8dc402a58f4ef28428c503ca0c870192097_dlReLU_rEA9
#> #define lReLU _rxmod1_b8dc402a58f4ef28428c503ca0c870192098_lReLU_z1SE
#> #define dSELU _rxmod1_b8dc402a58f4ef28428c503ca0c870192099_dSELU_1KzA
#> #define SELU _rxmod1_b8dc402a58f4ef28428c503ca0c870192100_SELU_f2H4
#> #define d4softplus _rxmod1_b8dc402a58f4ef28428c503ca0c870192101_d4softplus_Q2oF
#> #define d3softplus _rxmod1_b8dc402a58f4ef28428c503ca0c870192102_d3softplus_0joC
#> #define d2softplus _rxmod1_b8dc402a58f4ef28428c503ca0c870192103_d2softplus_Gj0a
#> #define dsoftplus _rxmod1_b8dc402a58f4ef28428c503ca0c870192104_dsoftplus_8wAb
#> #define softplus _rxmod1_b8dc402a58f4ef28428c503ca0c870192105_softplus_QELO
#> #define d4GELU _rxmod1_b8dc402a58f4ef28428c503ca0c870192106_d4GELU_lObq
#> #define d3GELU _rxmod1_b8dc402a58f4ef28428c503ca0c870192107_d3GELU_Nath
#> #define d2GELU _rxmod1_b8dc402a58f4ef28428c503ca0c870192108_d2GELU_ltiW
#> #define dGELU _rxmod1_b8dc402a58f4ef28428c503ca0c870192109_dGELU_R2IG
#> #define GELU _rxmod1_b8dc402a58f4ef28428c503ca0c870192110_GELU_VLQY
#> #define dReLU _rxmod1_b8dc402a58f4ef28428c503ca0c870192111_dReLU_M84J
#> #define ReLU _rxmod1_b8dc402a58f4ef28428c503ca0c870192112_ReLU_Cnuy
#> #define riweibull _rxmod1_b8dc402a58f4ef28428c503ca0c870192113_riweibull_CXPJ
#> #define riunif _rxmod1_b8dc402a58f4ef28428c503ca0c870192114_riunif_UjvF
#> #define rit_ _rxmod1_b8dc402a58f4ef28428c503ca0c870192115_rit__k7KJ
#> #define ripois _rxmod1_b8dc402a58f4ef28428c503ca0c870192116_ripois_vDxq
#> #define ribeta _rxmod1_b8dc402a58f4ef28428c503ca0c870192117_ribeta_bVSK
#> #define rigamma _rxmod1_b8dc402a58f4ef28428c503ca0c870192118_rigamma_Vfr7
#> #define rigeom _rxmod1_b8dc402a58f4ef28428c503ca0c870192119_rigeom_TcaU
#> #define rif _rxmod1_b8dc402a58f4ef28428c503ca0c870192120_rif_EgpY
#> #define riexp _rxmod1_b8dc402a58f4ef28428c503ca0c870192121_riexp_kNIq
#> #define richisq _rxmod1_b8dc402a58f4ef28428c503ca0c870192122_richisq_nEnm
#> #define ricauchy _rxmod1_b8dc402a58f4ef28428c503ca0c870192123_ricauchy_hSbo
#> #define rinbinomMu _rxmod1_b8dc402a58f4ef28428c503ca0c870192124_rinbinomMu_MZYB
#> #define rinbinom _rxmod1_b8dc402a58f4ef28428c503ca0c870192125_rinbinom_jkiC
#> #define ribinom _rxmod1_b8dc402a58f4ef28428c503ca0c870192126_ribinom_oUqv
#> #define rinorm _rxmod1_b8dc402a58f4ef28428c503ca0c870192127_rinorm_niI7
#> #define rxweibull _rxmod1_b8dc402a58f4ef28428c503ca0c870192128_rxweibull_5Ui0
#> #define rxunif _rxmod1_b8dc402a58f4ef28428c503ca0c870192129_rxunif_GVTs
#> #define rxt_ _rxmod1_b8dc402a58f4ef28428c503ca0c870192130_rxt__34TN
#> #define rxpois _rxmod1_b8dc402a58f4ef28428c503ca0c870192131_rxpois_HeFp
#> #define rxbeta _rxmod1_b8dc402a58f4ef28428c503ca0c870192132_rxbeta_ue1Z
#> #define rxgamma _rxmod1_b8dc402a58f4ef28428c503ca0c870192133_rxgamma_aEnJ
#> #define rxgeom _rxmod1_b8dc402a58f4ef28428c503ca0c870192134_rxgeom_R6UH
#> #define rxf _rxmod1_b8dc402a58f4ef28428c503ca0c870192135_rxf_ZKbS
#> #define rxexp _rxmod1_b8dc402a58f4ef28428c503ca0c870192136_rxexp_smNY
#> #define rxchisq _rxmod1_b8dc402a58f4ef28428c503ca0c870192137_rxchisq_lS2w
#> #define rxcauchy _rxmod1_b8dc402a58f4ef28428c503ca0c870192138_rxcauchy_zYqw
#> #define rxnbinomMu _rxmod1_b8dc402a58f4ef28428c503ca0c870192139_rxnbinomMu_2z8b
#> #define rxnbinom _rxmod1_b8dc402a58f4ef28428c503ca0c870192140_rxnbinom_GMdj
#> #define rxbinom _rxmod1_b8dc402a58f4ef28428c503ca0c870192141_rxbinom_LB3L
#> #define rxnorm _rxmod1_b8dc402a58f4ef28428c503ca0c870192142_rxnorm_sOUR
#> #define linCmtB _rxmod1_b8dc402a58f4ef28428c503ca0c870192143_linCmtB_bPYU
#> #define linCmtA _rxmod1_b8dc402a58f4ef28428c503ca0c870192144_linCmtA_PmW8
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){  if (_solveData->svar[_svari] == 0) {V2 = _PP[0];};   if (_solveData->svar[_svari] == 1) {V3 = _PP[1];};   if (_solveData->svar[_svari] == 2) {KA = _PP[2];};   if (_solveData->svar[_svari] == 3) {CL = _PP[3];};   if (_solveData->svar[_svari] == 4) {Q = _PP[4];};   if (_solveData->svar[_svari] == 5) {Kin = _PP[5];};   if (_solveData->svar[_svari] == 6) {Kout = _PP[6];};   if (_solveData->svar[_svari] == 7) {EC50 = _PP[7];}; }
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){  if (_solveData->ovar[_ovari] == 0) {V2 = _PP[0];};   if (_solveData->ovar[_ovari] == 1) {V3 = _PP[1];};   if (_solveData->ovar[_ovari] == 2) {KA = _PP[2];};   if (_solveData->ovar[_ovari] == 3) {CL = _PP[3];};   if (_solveData->ovar[_ovari] == 4) {Q = _PP[4];};   if (_solveData->ovar[_ovari] == 5) {Kin = _PP[5];};   if (_solveData->ovar[_ovari] == 6) {Kout = _PP[6];};   if (_solveData->ovar[_ovari] == 7) {EC50 = _PP[7];}; }
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  mod1__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *mod1__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP mod1__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void mod1__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double C2 = NA_REAL;
#>   double centr = NA_REAL;
#>   double V2 = NA_REAL;
#>   double C3 = NA_REAL;
#>   double peri = NA_REAL;
#>   double V3 = NA_REAL;
#>   double depot = NA_REAL;
#>   double KA = NA_REAL;
#>   double CL = NA_REAL;
#>   double Q = NA_REAL;
#>   double eff = NA_REAL;
#>   double Kin = NA_REAL;
#>   double Kout = NA_REAL;
#>   double EC50 = NA_REAL;
#> 
#>   (void)t;
#>   (void)C2;
#>   (void)centr;
#>   (void)V2;
#>   (void)C3;
#>   (void)peri;
#>   (void)V3;
#>   (void)depot;
#>   (void)KA;
#>   (void)CL;
#>   (void)Q;
#>   (void)eff;
#>   (void)Kin;
#>   (void)Kout;
#>   (void)EC50;
#> 
#>   C2 = _PL[0];
#>   C3 = _PL[1];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#>   V2 = _PP[0];
#>   V3 = _PP[1];
#>   KA = _PP[2];
#>   CL = _PP[3];
#>   Q = _PP[4];
#>   Kin = _PP[5];
#>   Kout = _PP[6];
#>   EC50 = _PP[7];
#> 
#>   depot = __zzStateVar__[__DDT0__]*((double)(_ON[__DDT0__]));
#>   centr = __zzStateVar__[__DDT1__]*((double)(_ON[__DDT1__]));
#>   peri = __zzStateVar__[__DDT2__]*((double)(_ON[__DDT2__]));
#>   eff = __zzStateVar__[__DDT3__]*((double)(_ON[__DDT3__]));
#> 
#>   C2=centr/_div0(V2);
#>   C3=peri/_div0(V3);
#>   __DDtStateVar__[__DDT0__] = ((double)(_ON[__DDT0__]))*(_IR[__DDT0__] -KA*depot);
#>   __DDtStateVar__[__DDT1__] = ((double)(_ON[__DDT1__]))*(_IR[__DDT1__] + KA*depot-CL*C2-Q*C2+Q*C3);
#>   __DDtStateVar__[__DDT2__] = ((double)(_ON[__DDT2__]))*(_IR[__DDT2__] + Q*C2-Q*C3);
#>   __DDtStateVar__[__DDT3__] = ((double)(_ON[__DDT3__]))*(_IR[__DDT3__] + Kin-Kout*(1-C2/_div0((EC50+C2)))*eff);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void mod1__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void mod1__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void mod1__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double  __DDtStateVar_0__;
#>   double  __DDtStateVar_1__;
#>   double  __DDtStateVar_2__;
#>   double  __DDtStateVar_3__;
#>   double C2 = NA_REAL;
#>   double centr = NA_REAL;
#>   double V2 = NA_REAL;
#>   double C3 = NA_REAL;
#>   double peri = NA_REAL;
#>   double V3 = NA_REAL;
#>   double depot = NA_REAL;
#>   double KA = NA_REAL;
#>   double CL = NA_REAL;
#>   double Q = NA_REAL;
#>   double eff = NA_REAL;
#>   double Kin = NA_REAL;
#>   double Kout = NA_REAL;
#>   double EC50 = NA_REAL;
#> 
#>   (void)t;
#>   (void)__DDtStateVar_0__;
#>   (void)__DDtStateVar_1__;
#>   (void)__DDtStateVar_2__;
#>   (void)__DDtStateVar_3__;
#>   (void)C2;
#>   (void)centr;
#>   (void)V2;
#>   (void)C3;
#>   (void)peri;
#>   (void)V3;
#>   (void)depot;
#>   (void)KA;
#>   (void)CL;
#>   (void)Q;
#>   (void)eff;
#>   (void)Kin;
#>   (void)Kout;
#>   (void)EC50;
#> 
#>   C2 = _PL[0];
#>   C3 = _PL[1];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#>   V2 = _PP[0];
#>   V3 = _PP[1];
#>   KA = _PP[2];
#>   CL = _PP[3];
#>   Q = _PP[4];
#>   Kin = _PP[5];
#>   Kout = _PP[6];
#>   EC50 = _PP[7];
#> 
#>   depot = __zzStateVar__[__DDT0__]*((double)(_ON[__DDT0__]));
#>   centr = __zzStateVar__[__DDT1__]*((double)(_ON[__DDT1__]));
#>   peri = __zzStateVar__[__DDT2__]*((double)(_ON[__DDT2__]));
#>   eff = __zzStateVar__[__DDT3__]*((double)(_ON[__DDT3__]));
#> 
#>   C2=centr/_div0(V2);
#>   C3=peri/_div0(V3);
#>   __DDtStateVar_0__ = ((double)(_ON[__DDT0__]))*(_IR[__DDT0__] -KA*depot);
#>   __DDtStateVar_1__ = ((double)(_ON[__DDT1__]))*(_IR[__DDT1__] + KA*depot-CL*C2-Q*C2+Q*C3);
#>   __DDtStateVar_2__ = ((double)(_ON[__DDT2__]))*(_IR[__DDT2__] + Q*C2-Q*C3);
#>   __DDtStateVar_3__ = ((double)(_ON[__DDT3__]))*(_IR[__DDT3__] + Kin-Kout*(1-C2/_div0((EC50+C2)))*eff);
#> 
#>   _lhs[_LHS_0_]=C2;
#>   _lhs[_LHS_1_]=C3;
#> }
#> // Functional based bioavailability
#> double mod1__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double mod1__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double mod1__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double mod1__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void mod1__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void mod1__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void mod1__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP mod1__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("mod1__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAAAAeAAA+:g6JifQmmHA\?1qlo4P*5AAkj9*wuNr2lvyPZ=!c_nX;1ez=Fz8i2{}aH%%fM:J_Qds[Q&Mb|X<tKGS|{_*EM=kNkekW)EG}PJ15F0(lAWaD{+X}/=WUawl]nt~J+eVm|*H}$}_6K99(l*{kpU[UHY}Qn..V)K2g8X^rk:&XSBau4;[{(g+~Nmm<7P;_~>34.tz2{{(Sba#v~|L(]20KJ{Rx+^*s;GYFG]RH>|JL5)QR.~yQ>.Qpmb\?leZtJIZ@e<Q$M#7#/\"rcm{$Z&Mr\?p7~N)aHzkcRJ_>)tbx)Yy)G[bLDo|nQ]UWLh)I.>dl(j}yy\?@t0j1i17~fv,LOF]~|i8|w:B.)a~0\?Mfner6FvFK0N`arnT9Hb1l<\?MuU<dLp4C6zE+Hgb]!Mo0g9j\">GcvIBsuU}$_IDw]!YcH%%jLS0E!vHrcO)t^s,xB`QcfO)FxYYq+5bBZbyA3jfcM!88KqtI0xC0\?2Bz2]Og7C}9&Am|[6Xz<.c(Y<mNXa<S]5uYVBV;CC\"|@=*,G(R4#cP7nc>e3\?v1cW/V,!$HUK/!RJ=B]~8vnvU)j8fM98@I@=C%%kQY(]V944Auc8+~fnk`G]E;_mbu$rKVlkR)+Qng~8$,h}T8T(>r%%a|llA]fDFAzoP47wnqu}xcqAo@t8)q>39/hF/JRE8L4]+j+T3`4b`_@_jEu&LA.{e;p}}_fz+oJ01<J]f8oUwn$Qaa#e0r[x#+8Lb=H,{a[D&j(xts^Z/[@I`*wuFskM<|8Ntis^OjN\"5=nO*}Sr[3]Ts%%eS4NYWr%%v6U{&8{s<.q|qda]n8Sd$(r4x{|U*2`My4xzJHg_Y=U~:J`V[Jb.xJ[6I729N)G[44K#*wI2xbjsSiZnxeZ}4b7L^8M4Y_TqHtxpZjNAXgJoJ%%{40Ua;gaLyNX]e=r|G,\?W9hPcC%%si&ge)xwGVSjqW=/R<kRBjiFx,aWt8ZQFi}H;C;J3cg3i0DI&~JxDvplf:aY|R1bnmhdE.dY98@Dk[g5`eP#[kHFlA$D.cLvf[uuGH%%Rz@9x(`Gd1#U\?w^QCuc#qywb]mVYBHA&$gk4nHqbm)Y*.*x7B[zrL>*H1]3;kJv!s06lhFCnHryYtJ+wxQQaSsB)*EfDK%%}hXM[adp>!5))M}$c;x<W.wMz/i:=AD+:#X/Qkl3qjzKl`tI8dYWsqLUe*:V}V");
#>     char buf[1224];
#> #define __doBufN__ 1224
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void mod1__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   mod1__dydt(neq, *t, A, DADT);
#> }
#> extern int mod1__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   mod1__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void mod1__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   mod1__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void mod1__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_mod1_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("mod1_","mod1__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("mod1_","mod1__assignFuns", (DL_FUNC) mod1__assignFuns);
#>   R_RegisterCCallable("mod1_","mod1__inis",(DL_FUNC) mod1__inis);
#>   R_RegisterCCallable("mod1_","mod1__dydt",(DL_FUNC) mod1__dydt);
#>   R_RegisterCCallable("mod1_","mod1__calc_lhs",(DL_FUNC) mod1__calc_lhs);
#>   R_RegisterCCallable("mod1_","mod1__calc_jac",(DL_FUNC) mod1__calc_jac);
#>   R_RegisterCCallable("mod1_","mod1__dydt_lsoda", (DL_FUNC) mod1__dydt_lsoda);
#>   R_RegisterCCallable("mod1_","mod1__calc_jac_lsoda", (DL_FUNC) mod1__calc_jac_lsoda);
#>   R_RegisterCCallable("mod1_","mod1__ode_solver_solvedata", (DL_FUNC) mod1__ode_solver_solvedata);
#>   R_RegisterCCallable("mod1_","mod1__ode_solver_get_solvedata", (DL_FUNC) mod1__ode_solver_get_solvedata);
#>   R_RegisterCCallable("mod1_","mod1__F", (DL_FUNC) mod1__F);
#>   R_RegisterCCallable("mod1_","mod1__Lag", (DL_FUNC) mod1__Lag);
#>   R_RegisterCCallable("mod1_","mod1__Rate", (DL_FUNC) mod1__Rate);
#>   R_RegisterCCallable("mod1_","mod1__Dur", (DL_FUNC) mod1__Dur);
#>   R_RegisterCCallable("mod1_","mod1__mtime", (DL_FUNC) mod1__mtime);
#>   R_RegisterCCallable("mod1_","mod1__ME", (DL_FUNC) mod1__ME);
#>   R_RegisterCCallable("mod1_","mod1__IndF", (DL_FUNC) mod1__IndF);
#>   R_RegisterCCallable("mod1_","mod1__dydt_liblsoda", (DL_FUNC) mod1__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_mod1_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_mod1_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"mod1__model_vars", (DL_FUNC) &mod1__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_mod1_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("mod1__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("mod1__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’
#> 
#> Parameters and their values:
#>      KA      CL      V2       Q      V3     Kin    Kout    EC50 
#>   0.294  18.600  40.200  10.500 297.000   1.000   1.000 200.000 
#> 
#> Initial values in each compartment:
#> depot centr  pari   eff 
#>     0     0     0     1 
#> 
#> Warning: 2 arguments not used by format '
#>       #
#>       # Dosing regimen template generated by rxode2::genShinyApp.template()
#>       #
#> 
#>       debug = TRUE
#>       #wd = sprintf("%%s/../", getwd())
#>       #setwd(wd)
#> 
#>       # Server inputs: Dose, dosing regimen, dosing frequency,
#>       # dosing cycle definition, number of dosing cycles
#> 
#>       library(shiny)
#>       library(rxode2)
#> 
#>       # read objects from "rx_shiny_data.rda" in the  AppDir folder,
#>       # objects include, mod1, params, inits, method, atol, rtol.]
#> 
#>       load("./rx_shiny_data.rda")
#>       if (!rxDynLoad(mod1)) mod1 <- rxode2(mod1, modName="mod1")
#>       # Define server logic
#>       shinyServer(function(input, output) {
#> 
#>         get.cp <- reactive({
#>           ds <- input$Dose
#>           reg <- switch(input$regimen, "QD"=1, "BID"=2)
#>           cyc <- switch(input$cycle,
#>               "continous"=c(7,0),
#>               "1wkon 1wkoff"=c(7,7),
#>               "2wkon 1wkoff"=c(14,7),
#>               "3wkon 1wkoff"=c(21,7)
#>           )
#>           cyc <- [... truncated]
#> Shiny files (ui.R, server.R) plus R data saved.
#> 
#> To launch the Shiny app, type the following two R commands:
#> 
#> 	library(shiny)
#> 	runApp("myapp")
#> 
# run the shiny app
if (requireNamespace("shiny", quietly=TRUE)) {
  library(shiny)
  # runApp("myapp") # Won't launch in environments without browsers
}
# }