Calculate the log likelihood of the binomial function (and its derivatives)

Calculate the log likelihood of the binomial function (and its derivatives)

Usage

llikBeta(x, shape1, shape2, full = FALSE)

Arguments

x

Observation

shape1, shape2

non-negative parameters of the Beta distribution.

full

Add the data frame showing x, mean, sd as well as the fx and derivatives

Value

data frame with fx for the log pdf value of with dShape1 and dShape2 that has the derivatives with respect to the parameters at the observation time-point

Details

In an rxode2() model, you can use llikBeta() but you have to use all arguments. You can also get the derivative of shape1 and shape2 with llikBetaDshape1() and llikBetaDshape2().

Author

Matthew L. Fidler

Examples

# \donttest{

x <- seq(1e-4, 1 - 1e-4, length.out = 21)

llikBeta(x, 0.5, 0.5)
#>             fx     dShape1     dShape2
#> 1   3.46049030 -7.82404601  1.38619436
#> 2   0.37793108 -1.60763953  1.33490633
#> 3   0.05888752 -0.91549105  1.28084495
#> 4  -0.11510253 -0.51035907  1.22369308
#> 5  -0.22855163 -0.22284360  1.16307581
#> 6  -0.30780832  0.00019998  1.09854562
#> 7  -0.36444410  0.18245488  1.02956227
#> 8  -0.40444714  0.33655795  0.95546529
#> 9  -0.43118004  0.47005363  0.87543540
#> 10 -0.44655956  0.58780889  0.78843918
#> 11 -0.45158271  0.69314718  0.69314718
#> 12 -0.44655956  0.78843918  0.58780889
#> 13 -0.43118004  0.87543540  0.47005363
#> 14 -0.40444714  0.95546529  0.33655795
#> 15 -0.36444410  1.02956227  0.18245488
#> 16 -0.30780832  1.09854562  0.00019998
#> 17 -0.22855163  1.16307581 -0.22284360
#> 18 -0.11510253  1.22369308 -0.51035907
#> 19  0.05888752  1.28084495 -0.91549105
#> 20  0.37793108  1.33490633 -1.60763953
#> 21  3.46049030  1.38619436 -7.82404601

llikBeta(x, 1, 3, TRUE)
#>          x shape1 shape2           fx     dShape1     dShape2
#> 1  0.00010      1      3   1.09841228 -7.37700704  0.33323333
#> 2  0.05009      1      3   0.99583622 -1.16060056  0.28194530
#> 3  0.10008      1      3   0.88771347 -0.46845208  0.22788392
#> 4  0.15007      1      3   0.77340972 -0.06332009  0.17073205
#> 5  0.20006      1      3   0.65217518  0.22419538  0.11011478
#> 6  0.25005      1      3   0.52311481  0.44723895  0.04558459
#> 7  0.30004      1      3   0.38514811  0.62949385 -0.02339876
#> 8  0.35003      1      3   0.23695415  0.78359692 -0.09749574
#> 9  0.40002      1      3   0.07689437  0.91709260 -0.17752562
#> 10 0.45001      1      3  -0.09709808  1.03484786 -0.26452185
#> 11 0.50000      1      3  -0.28768207  1.14018615 -0.35981385
#> 12 0.54999      1      3  -0.49835866  1.23547815 -0.46515214
#> 13 0.59998      1      3  -0.73386918  1.32247438 -0.58290740
#> 14 0.64997      1      3  -1.00086054  1.40250426 -0.71640308
#> 15 0.69996      1      3  -1.30906667  1.47660124 -0.87050615
#> 16 0.74995      1      3  -1.67357647  1.54558459 -1.05276105
#> 17 0.79994      1      3  -2.11966363  1.61011478 -1.27580462
#> 18 0.84993      1      3  -2.69469457  1.67073205 -1.56332009
#> 19 0.89992      1      3  -3.50495854  1.72788392 -1.96845208
#> 20 0.94991      1      3  -4.88925549  1.78194530 -2.66060056
#> 21 0.99990      1      3 -17.32206846  1.83323333 -8.87700704

et <- et(seq(1e-4, 1-1e-4, length.out=21))
et$shape1 <- 0.5
et$shape2 <- 1.5

model <- function() {
  model({
    fx <- llikBeta(time, shape1, shape2)
    dShape1 <- llikBetaDshape1(time, shape1, shape2)
    dShape2 <- llikBetaDshape2(time, shape1, shape2)
  })
}

rxSolve(model, et)
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#>  
#>  
#> C model:
#> // Define 3 LHS values
#> #define _LHS_0_ 0 // fx
#> #define _LHS_1_ 1 // dShape1
#> #define _LHS_2_ 2 // dShape2
#> #define _getRxSolve_ _rx94eed3fa5a22c6e0c4e5731d282af1ca2288__getRxSolve__lORz
#> #define _evalUdf _rx94eed3fa5a22c6e0c4e5731d282af1ca2289__evalUdf_kg9p
#> #define _solveData _rx94eed3fa5a22c6e0c4e5731d282af1ca2290__solveData_hZjb
#> #define _assign_ptr _rx94eed3fa5a22c6e0c4e5731d282af1ca2291__assign_ptr_tvsF
#> #define _rxRmModelLib _rx94eed3fa5a22c6e0c4e5731d282af1ca2292__rxRmModelLib_QkpV
#> #define _rxGetModelLib _rx94eed3fa5a22c6e0c4e5731d282af1ca2293__rxGetModelLib_HF4t
#> #define _old_c _rx94eed3fa5a22c6e0c4e5731d282af1ca2294__old_c_E5bq
#> #define _ptrid _rx94eed3fa5a22c6e0c4e5731d282af1ca2295__ptrid_rWj2
#> #define _rxIsCurrentC _rx94eed3fa5a22c6e0c4e5731d282af1ca2296__rxIsCurrentC_jIRW
#> #define _sumPS _rx94eed3fa5a22c6e0c4e5731d282af1ca2297__sumPS_0crr
#> #define _prodPS _rx94eed3fa5a22c6e0c4e5731d282af1ca2298__prodPS_gcGm
#> #define _prodType _rx94eed3fa5a22c6e0c4e5731d282af1ca2299__prodType_EsJu
#> #define _sumType _rx94eed3fa5a22c6e0c4e5731d282af1ca2300__sumType_9xah
#> #define _update_par_ptr _rx94eed3fa5a22c6e0c4e5731d282af1ca2301__update_par_ptr_gBOZ
#> #define _getParCov _rx94eed3fa5a22c6e0c4e5731d282af1ca2302__getParCov_l3zY
#> #define _rxode2_rxAssignPtr _rx94eed3fa5a22c6e0c4e5731d282af1ca2303__rxode2_rxAssignPtr_vTa9
#> #define _rxQr _rx94eed3fa5a22c6e0c4e5731d282af1ca2304__rxQr_7EhW
#> #define _compareFactorVal _rx94eed3fa5a22c6e0c4e5731d282af1ca2305__compareFactorVal_3MG9
#> #define _sum _rx94eed3fa5a22c6e0c4e5731d282af1ca2306__sum_V3pR
#> #define _udf _rx94eed3fa5a22c6e0c4e5731d282af1ca2307__udf_1aNk
#> #define _sign _rx94eed3fa5a22c6e0c4e5731d282af1ca2308__sign_G5Jw
#> #define _prod _rx94eed3fa5a22c6e0c4e5731d282af1ca2309__prod_D9TK
#> #define _max _rx94eed3fa5a22c6e0c4e5731d282af1ca2310__max_40NP
#> #define _min _rx94eed3fa5a22c6e0c4e5731d282af1ca2311__min_JiAL
#> #define _transit4P _rx94eed3fa5a22c6e0c4e5731d282af1ca2312__transit4P_ZBrh
#> #define _transit3P _rx94eed3fa5a22c6e0c4e5731d282af1ca2313__transit3P_efgh
#> #define _assignFuns0 _rx94eed3fa5a22c6e0c4e5731d282af1ca2314__assignFuns0_Drm8
#> #define _assignFuns _rx94eed3fa5a22c6e0c4e5731d282af1ca2315__assignFuns_5igv
#> #define _rxord _rx94eed3fa5a22c6e0c4e5731d282af1ca2316__rxord_lJYg
#> #define __assignFuns2 _rx94eed3fa5a22c6e0c4e5731d282af1ca2317___assignFuns2_GqQx
#> #define _llikCauchyDscale _rx94eed3fa5a22c6e0c4e5731d282af1ca2318__llikCauchyDscale_LoYl
#> #define _llikCauchyDlocation _rx94eed3fa5a22c6e0c4e5731d282af1ca2319__llikCauchyDlocation_FgJd
#> #define _llikCauchy _rx94eed3fa5a22c6e0c4e5731d282af1ca2320__llikCauchy_KaFg
#> #define _llikGammaDrate _rx94eed3fa5a22c6e0c4e5731d282af1ca2321__llikGammaDrate_mDR4
#> #define _llikGammaDshape _rx94eed3fa5a22c6e0c4e5731d282af1ca2322__llikGammaDshape_jxuo
#> #define _llikGamma _rx94eed3fa5a22c6e0c4e5731d282af1ca2323__llikGamma_P8gm
#> #define _llikWeibullDscale _rx94eed3fa5a22c6e0c4e5731d282af1ca2324__llikWeibullDscale_oUom
#> #define _llikWeibullDshape _rx94eed3fa5a22c6e0c4e5731d282af1ca2325__llikWeibullDshape_wLtu
#> #define _llikWeibull _rx94eed3fa5a22c6e0c4e5731d282af1ca2326__llikWeibull_DgOz
#> #define _llikUnifDbeta _rx94eed3fa5a22c6e0c4e5731d282af1ca2327__llikUnifDbeta_tPZw
#> #define _llikUnifDalpha _rx94eed3fa5a22c6e0c4e5731d282af1ca2328__llikUnifDalpha_Dqx2
#> #define _llikUnif _rx94eed3fa5a22c6e0c4e5731d282af1ca2329__llikUnif_t0Ff
#> #define _llikGeomDp _rx94eed3fa5a22c6e0c4e5731d282af1ca2330__llikGeomDp_CRup
#> #define _llikGeom _rx94eed3fa5a22c6e0c4e5731d282af1ca2331__llikGeom_BfcL
#> #define _llikFDdf2 _rx94eed3fa5a22c6e0c4e5731d282af1ca2332__llikFDdf2_LlJk
#> #define _llikFDdf1 _rx94eed3fa5a22c6e0c4e5731d282af1ca2333__llikFDdf1_0HOu
#> #define _llikF _rx94eed3fa5a22c6e0c4e5731d282af1ca2334__llikF_quMQ
#> #define _llikExpDrate _rx94eed3fa5a22c6e0c4e5731d282af1ca2335__llikExpDrate_IP4s
#> #define _llikExp _rx94eed3fa5a22c6e0c4e5731d282af1ca2336__llikExp_BKwl
#> #define _llikChisqDdf _rx94eed3fa5a22c6e0c4e5731d282af1ca2337__llikChisqDdf_ION1
#> #define _llikChisq _rx94eed3fa5a22c6e0c4e5731d282af1ca2338__llikChisq_tTX5
#> #define _llikTDsd _rx94eed3fa5a22c6e0c4e5731d282af1ca2339__llikTDsd_DulH
#> #define _llikTDmean _rx94eed3fa5a22c6e0c4e5731d282af1ca2340__llikTDmean_LxeA
#> #define _llikTDdf _rx94eed3fa5a22c6e0c4e5731d282af1ca2341__llikTDdf_Ut0p
#> #define _llikT _rx94eed3fa5a22c6e0c4e5731d282af1ca2342__llikT_o6Aq
#> #define _llikBetaDshape2 _rx94eed3fa5a22c6e0c4e5731d282af1ca2343__llikBetaDshape2_SQkn
#> #define _llikBetaDshape1 _rx94eed3fa5a22c6e0c4e5731d282af1ca2344__llikBetaDshape1_H8fP
#> #define _llikBeta _rx94eed3fa5a22c6e0c4e5731d282af1ca2345__llikBeta_CAnN
#> #define _llikNbinomMuDmu _rx94eed3fa5a22c6e0c4e5731d282af1ca2346__llikNbinomMuDmu_kUms
#> #define _llikNbinomMu _rx94eed3fa5a22c6e0c4e5731d282af1ca2347__llikNbinomMu_wiOW
#> #define _llikNbinomDprob _rx94eed3fa5a22c6e0c4e5731d282af1ca2348__llikNbinomDprob_G16H
#> #define _llikNbinom _rx94eed3fa5a22c6e0c4e5731d282af1ca2349__llikNbinom_PUgn
#> #define _llikBinomDprob _rx94eed3fa5a22c6e0c4e5731d282af1ca2350__llikBinomDprob_Cjuw
#> #define _llikBinom _rx94eed3fa5a22c6e0c4e5731d282af1ca2351__llikBinom_A8iw
#> #define _llikPoisDlambda _rx94eed3fa5a22c6e0c4e5731d282af1ca2352__llikPoisDlambda_bn2G
#> #define _llikPois _rx94eed3fa5a22c6e0c4e5731d282af1ca2353__llikPois_TQjS
#> #define _llikNormDsd _rx94eed3fa5a22c6e0c4e5731d282af1ca2354__llikNormDsd_rer5
#> #define _llikNormDmean _rx94eed3fa5a22c6e0c4e5731d282af1ca2355__llikNormDmean_zexL
#> #define _llikNorm _rx94eed3fa5a22c6e0c4e5731d282af1ca2356__llikNorm_gAwM
#> #define simeps _rx94eed3fa5a22c6e0c4e5731d282af1ca2357_simeps_E6nq
#> #define simeta _rx94eed3fa5a22c6e0c4e5731d282af1ca2358_simeta_R3DD
#> #define expit _rx94eed3fa5a22c6e0c4e5731d282af1ca2359_expit_9jN5
#> #define logit _rx94eed3fa5a22c6e0c4e5731d282af1ca2360_logit_O4r9
#> #define gammapDer _rx94eed3fa5a22c6e0c4e5731d282af1ca2361_gammapDer_D97m
#> #define lowergamma _rx94eed3fa5a22c6e0c4e5731d282af1ca2362_lowergamma_e5yb
#> #define uppergamma _rx94eed3fa5a22c6e0c4e5731d282af1ca2363_uppergamma_V8Eu
#> #define gammaqInva _rx94eed3fa5a22c6e0c4e5731d282af1ca2364_gammaqInva_Z1xT
#> #define gammaqInv _rx94eed3fa5a22c6e0c4e5731d282af1ca2365_gammaqInv_QYOQ
#> #define gammapInva _rx94eed3fa5a22c6e0c4e5731d282af1ca2366_gammapInva_e6aq
#> #define gammapInv _rx94eed3fa5a22c6e0c4e5731d282af1ca2367_gammapInv_t3Lb
#> #define gammaq _rx94eed3fa5a22c6e0c4e5731d282af1ca2368_gammaq_sbea
#> #define gammap _rx94eed3fa5a22c6e0c4e5731d282af1ca2369_gammap_w5m4
#> #define phi _rx94eed3fa5a22c6e0c4e5731d282af1ca2370_phi_brD9
#> #define d2ELUa _rx94eed3fa5a22c6e0c4e5731d282af1ca2371_d2ELUa_EZAJ
#> #define dELUa _rx94eed3fa5a22c6e0c4e5731d282af1ca2372_dELUa_Ne7I
#> #define d2aELU _rx94eed3fa5a22c6e0c4e5731d282af1ca2373_d2aELU_xUFA
#> #define d2ELU _rx94eed3fa5a22c6e0c4e5731d282af1ca2374_d2ELU_5R2w
#> #define dELU _rx94eed3fa5a22c6e0c4e5731d282af1ca2375_dELU_ZwwB
#> #define ELU _rx94eed3fa5a22c6e0c4e5731d282af1ca2376_ELU_gRvb
#> #define dPReLUa1 _rx94eed3fa5a22c6e0c4e5731d282af1ca2377_dPReLUa1_lnk3
#> #define dPReLUa _rx94eed3fa5a22c6e0c4e5731d282af1ca2378_dPReLUa_57cf
#> #define dPReLU _rx94eed3fa5a22c6e0c4e5731d282af1ca2379_dPReLU_UwMa
#> #define PReLU _rx94eed3fa5a22c6e0c4e5731d282af1ca2380_PReLU_rEA9
#> #define dSwish _rx94eed3fa5a22c6e0c4e5731d282af1ca2381_dSwish_z1SE
#> #define Swish _rx94eed3fa5a22c6e0c4e5731d282af1ca2382_Swish_1KzA
#> #define dlReLU _rx94eed3fa5a22c6e0c4e5731d282af1ca2383_dlReLU_f2H4
#> #define lReLU _rx94eed3fa5a22c6e0c4e5731d282af1ca2384_lReLU_Q2oF
#> #define dSELU _rx94eed3fa5a22c6e0c4e5731d282af1ca2385_dSELU_0joC
#> #define SELU _rx94eed3fa5a22c6e0c4e5731d282af1ca2386_SELU_Gj0a
#> #define d4softplus _rx94eed3fa5a22c6e0c4e5731d282af1ca2387_d4softplus_8wAb
#> #define d3softplus _rx94eed3fa5a22c6e0c4e5731d282af1ca2388_d3softplus_QELO
#> #define d2softplus _rx94eed3fa5a22c6e0c4e5731d282af1ca2389_d2softplus_lObq
#> #define dsoftplus _rx94eed3fa5a22c6e0c4e5731d282af1ca2390_dsoftplus_Nath
#> #define softplus _rx94eed3fa5a22c6e0c4e5731d282af1ca2391_softplus_ltiW
#> #define d4GELU _rx94eed3fa5a22c6e0c4e5731d282af1ca2392_d4GELU_R2IG
#> #define d3GELU _rx94eed3fa5a22c6e0c4e5731d282af1ca2393_d3GELU_VLQY
#> #define d2GELU _rx94eed3fa5a22c6e0c4e5731d282af1ca2394_d2GELU_M84J
#> #define dGELU _rx94eed3fa5a22c6e0c4e5731d282af1ca2395_dGELU_Cnuy
#> #define GELU _rx94eed3fa5a22c6e0c4e5731d282af1ca2396_GELU_CXPJ
#> #define dReLU _rx94eed3fa5a22c6e0c4e5731d282af1ca2397_dReLU_UjvF
#> #define ReLU _rx94eed3fa5a22c6e0c4e5731d282af1ca2398_ReLU_k7KJ
#> #define riweibull _rx94eed3fa5a22c6e0c4e5731d282af1ca2399_riweibull_vDxq
#> #define riunif _rx94eed3fa5a22c6e0c4e5731d282af1ca2400_riunif_bVSK
#> #define rit_ _rx94eed3fa5a22c6e0c4e5731d282af1ca2401_rit__Vfr7
#> #define ripois _rx94eed3fa5a22c6e0c4e5731d282af1ca2402_ripois_TcaU
#> #define ribeta _rx94eed3fa5a22c6e0c4e5731d282af1ca2403_ribeta_EgpY
#> #define rigamma _rx94eed3fa5a22c6e0c4e5731d282af1ca2404_rigamma_kNIq
#> #define rigeom _rx94eed3fa5a22c6e0c4e5731d282af1ca2405_rigeom_nEnm
#> #define rif _rx94eed3fa5a22c6e0c4e5731d282af1ca2406_rif_hSbo
#> #define riexp _rx94eed3fa5a22c6e0c4e5731d282af1ca2407_riexp_MZYB
#> #define richisq _rx94eed3fa5a22c6e0c4e5731d282af1ca2408_richisq_jkiC
#> #define ricauchy _rx94eed3fa5a22c6e0c4e5731d282af1ca2409_ricauchy_oUqv
#> #define rinbinomMu _rx94eed3fa5a22c6e0c4e5731d282af1ca2410_rinbinomMu_niI7
#> #define rinbinom _rx94eed3fa5a22c6e0c4e5731d282af1ca2411_rinbinom_5Ui0
#> #define ribinom _rx94eed3fa5a22c6e0c4e5731d282af1ca2412_ribinom_GVTs
#> #define rinorm _rx94eed3fa5a22c6e0c4e5731d282af1ca2413_rinorm_34TN
#> #define rxweibull _rx94eed3fa5a22c6e0c4e5731d282af1ca2414_rxweibull_HeFp
#> #define rxunif _rx94eed3fa5a22c6e0c4e5731d282af1ca2415_rxunif_ue1Z
#> #define rxt_ _rx94eed3fa5a22c6e0c4e5731d282af1ca2416_rxt__aEnJ
#> #define rxpois _rx94eed3fa5a22c6e0c4e5731d282af1ca2417_rxpois_R6UH
#> #define rxbeta _rx94eed3fa5a22c6e0c4e5731d282af1ca2418_rxbeta_ZKbS
#> #define rxgamma _rx94eed3fa5a22c6e0c4e5731d282af1ca2419_rxgamma_smNY
#> #define rxgeom _rx94eed3fa5a22c6e0c4e5731d282af1ca2420_rxgeom_lS2w
#> #define rxf _rx94eed3fa5a22c6e0c4e5731d282af1ca2421_rxf_zYqw
#> #define rxexp _rx94eed3fa5a22c6e0c4e5731d282af1ca2422_rxexp_2z8b
#> #define rxchisq _rx94eed3fa5a22c6e0c4e5731d282af1ca2423_rxchisq_GMdj
#> #define rxcauchy _rx94eed3fa5a22c6e0c4e5731d282af1ca2424_rxcauchy_LB3L
#> #define rxnbinomMu _rx94eed3fa5a22c6e0c4e5731d282af1ca2425_rxnbinomMu_sOUR
#> #define rxnbinom _rx94eed3fa5a22c6e0c4e5731d282af1ca2426_rxnbinom_bPYU
#> #define rxbinom _rx94eed3fa5a22c6e0c4e5731d282af1ca2427_rxbinom_PmW8
#> #define rxnorm _rx94eed3fa5a22c6e0c4e5731d282af1ca2428_rxnorm_sOmG
#> #define linCmtB _rx94eed3fa5a22c6e0c4e5731d282af1ca2429_linCmtB_ypO5
#> #define linCmtA _rx94eed3fa5a22c6e0c4e5731d282af1ca2430_linCmtA_5OnO
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){  if (_solveData->svar[_svari] == 0) {shape1 = _PP[0];};   if (_solveData->svar[_svari] == 1) {shape2 = _PP[1];}; }
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){  if (_solveData->ovar[_ovari] == 0) {shape1 = _PP[0];};   if (_solveData->ovar[_ovari] == 1) {shape2 = _PP[1];}; }
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  rx_ed69a861a9ee57524f69f40f16a8beef__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *rx_ed69a861a9ee57524f69f40f16a8beef__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP rx_ed69a861a9ee57524f69f40f16a8beef__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void rx_ed69a861a9ee57524f69f40f16a8beef__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double shape1 = NA_REAL;
#>   double shape2 = NA_REAL;
#>   double fx = NA_REAL;
#>   double dShape1 = NA_REAL;
#>   double dShape2 = NA_REAL;
#> 
#>   (void)t;
#>   (void)shape1;
#>   (void)shape2;
#>   (void)fx;
#>   (void)dShape1;
#>   (void)dShape2;
#> 
#>   fx = _PL[0];
#>   dShape1 = _PL[1];
#>   dShape2 = _PL[2];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#>   shape1 = _PP[0];
#>   shape2 = _PP[1];
#> 
#> 
#>   fx =llikBeta(t,shape1,shape2);
#>   dShape1 =llikBetaDshape1(t,shape1,shape2);
#>   dShape2 =llikBetaDshape2(t,shape1,shape2);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void rx_ed69a861a9ee57524f69f40f16a8beef__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void rx_ed69a861a9ee57524f69f40f16a8beef__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void rx_ed69a861a9ee57524f69f40f16a8beef__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double shape1 = NA_REAL;
#>   double shape2 = NA_REAL;
#>   double fx = NA_REAL;
#>   double dShape1 = NA_REAL;
#>   double dShape2 = NA_REAL;
#> 
#>   (void)t;
#>   (void)shape1;
#>   (void)shape2;
#>   (void)fx;
#>   (void)dShape1;
#>   (void)dShape2;
#> 
#>   fx = _PL[0];
#>   dShape1 = _PL[1];
#>   dShape2 = _PL[2];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#>   shape1 = _PP[0];
#>   shape2 = _PP[1];
#> 
#> 
#>   fx =llikBeta(t,shape1,shape2);
#>   dShape1 =llikBetaDshape1(t,shape1,shape2);
#>   dShape2 =llikBetaDshape2(t,shape1,shape2);
#> 
#>   _lhs[_LHS_0_]=fx;
#>   _lhs[_LHS_1_]=dShape1;
#>   _lhs[_LHS_2_]=dShape2;
#> }
#> // Functional based bioavailability
#> double rx_ed69a861a9ee57524f69f40f16a8beef__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double rx_ed69a861a9ee57524f69f40f16a8beef__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double rx_ed69a861a9ee57524f69f40f16a8beef__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double rx_ed69a861a9ee57524f69f40f16a8beef__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void rx_ed69a861a9ee57524f69f40f16a8beef__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void rx_ed69a861a9ee57524f69f40f16a8beef__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void rx_ed69a861a9ee57524f69f40f16a8beef__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP rx_ed69a861a9ee57524f69f40f16a8beef__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_ed69a861a9ee57524f69f40f16a8beef__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAA5F3AAAv7#aT)fr(3NAo<h,h\"P*Tu1)3rMqbOr1l4zIF5E##HnOc</lcbH%%>m_Z9j^g84h.<=zy*<J8KkmBHCn(ywpVl@bS&}c4CQnLoINcK\":bBbBxtnftt|3Ese@#inM(QL&}kWF[<ddJlfu=u^Y4n}os7P8xDO6,3[G3v*=nZ_f$kt2<V).s|%%%%Wl.Br/hp;UD#^YD<FGmj=P;2s\"K=[a#~J&De*s&.4%%\?jG\?}o`\"Qs|.,Lj_!L7mhe!+|NZAzMfLqlbn)%%trZku8+=5UM;+aQ|J<1a5f,y2,I>nARU5KOF(pID0c1Xweh}#|zybpzY#(e<z4D`pyrTAQ0MM;S@\?_NgcXDp\?i.V9^TF2Ovd55If!!|>QwY@#=>bCwACGm9er+D4IB6tn~Fd0|4(XH[\"TSG&\"3PFM06[`GgS`E]_p[5.tizD;I:}}i\?PB=\?aTVEM)BA$Z}Vd|`|~n8;k;GPj@RN.+ba2~NLE~(C6k*CNQ},vaJ6BX@%%I3)qO+<F[8lqyY{=,bw~a=I6L:H8:=&.E$csD#.Wa_nnX6a_f#s5\?|q[/:.>VP/QW!QNa0^K@WeOZjtcs}g\?PONHO`gQ:;S9vky1C;2[3vwQ!(t*:m|7M1dAP>][[0^tN9{[>s5e]oO@F[mQMn[fp;F1c2t4w7BzGV/.h>&\?RdJ2fkVL2zPr=V,)G8CGt|WNy1ldj*PyWJ)`M{K({Nl.]4nm\?rc^ZM\",K^YPE;.o.DT+3\?l*SwR>\?5xRnR:Qnl[}6H!X7W8@PH^y=t{/rL{/7yT,jv`VKtwN#xGRWX|vrvAwiGyix/\?n#X#iiLv.[B9hf&.q6u.QYepBKi{qc<J*3&e<g4&=6|8vwP[x/|j3(%%IBQ!Sjrh9d:Sch&>!PhfM{J<QkC.pXtLSLAbBIUx$cb#re#(Wr8ByD]u/1>Alt`]>Q3pXu:}W2[dC\?YJ:wZV]D&S.+G+O!@G)BoRa/+RS*\"5$hmxm\?UFgtd_SWr(FD_f>f%%+M9i^r%%[hO7)T.Uzqn,O&X`qC;elHg6iaXOevD2vtRFQC;(ICCBnk3Saiqoc7zK.HPEv7Xn#<)WBzS{gT7Ovuq!D@}I~N.$x4L2AK");
#>     char buf[1126];
#> #define __doBufN__ 1126
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void rx_ed69a861a9ee57524f69f40f16a8beef__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   rx_ed69a861a9ee57524f69f40f16a8beef__dydt(neq, *t, A, DADT);
#> }
#> extern int rx_ed69a861a9ee57524f69f40f16a8beef__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   rx_ed69a861a9ee57524f69f40f16a8beef__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void rx_ed69a861a9ee57524f69f40f16a8beef__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   rx_ed69a861a9ee57524f69f40f16a8beef__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void rx_ed69a861a9ee57524f69f40f16a8beef__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_rx_ed69a861a9ee57524f69f40f16a8beef_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("rx_ed69a861a9ee57524f69f40f16a8beef_","rx_ed69a861a9ee57524f69f40f16a8beef__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("rx_ed69a861a9ee57524f69f40f16a8beef_","rx_ed69a861a9ee57524f69f40f16a8beef__assignFuns", (DL_FUNC) rx_ed69a861a9ee57524f69f40f16a8beef__assignFuns);
#>   R_RegisterCCallable("rx_ed69a861a9ee57524f69f40f16a8beef_","rx_ed69a861a9ee57524f69f40f16a8beef__inis",(DL_FUNC) rx_ed69a861a9ee57524f69f40f16a8beef__inis);
#>   R_RegisterCCallable("rx_ed69a861a9ee57524f69f40f16a8beef_","rx_ed69a861a9ee57524f69f40f16a8beef__dydt",(DL_FUNC) rx_ed69a861a9ee57524f69f40f16a8beef__dydt);
#>   R_RegisterCCallable("rx_ed69a861a9ee57524f69f40f16a8beef_","rx_ed69a861a9ee57524f69f40f16a8beef__calc_lhs",(DL_FUNC) rx_ed69a861a9ee57524f69f40f16a8beef__calc_lhs);
#>   R_RegisterCCallable("rx_ed69a861a9ee57524f69f40f16a8beef_","rx_ed69a861a9ee57524f69f40f16a8beef__calc_jac",(DL_FUNC) rx_ed69a861a9ee57524f69f40f16a8beef__calc_jac);
#>   R_RegisterCCallable("rx_ed69a861a9ee57524f69f40f16a8beef_","rx_ed69a861a9ee57524f69f40f16a8beef__dydt_lsoda", (DL_FUNC) rx_ed69a861a9ee57524f69f40f16a8beef__dydt_lsoda);
#>   R_RegisterCCallable("rx_ed69a861a9ee57524f69f40f16a8beef_","rx_ed69a861a9ee57524f69f40f16a8beef__calc_jac_lsoda", (DL_FUNC) rx_ed69a861a9ee57524f69f40f16a8beef__calc_jac_lsoda);
#>   R_RegisterCCallable("rx_ed69a861a9ee57524f69f40f16a8beef_","rx_ed69a861a9ee57524f69f40f16a8beef__ode_solver_solvedata", (DL_FUNC) rx_ed69a861a9ee57524f69f40f16a8beef__ode_solver_solvedata);
#>   R_RegisterCCallable("rx_ed69a861a9ee57524f69f40f16a8beef_","rx_ed69a861a9ee57524f69f40f16a8beef__ode_solver_get_solvedata", (DL_FUNC) rx_ed69a861a9ee57524f69f40f16a8beef__ode_solver_get_solvedata);
#>   R_RegisterCCallable("rx_ed69a861a9ee57524f69f40f16a8beef_","rx_ed69a861a9ee57524f69f40f16a8beef__F", (DL_FUNC) rx_ed69a861a9ee57524f69f40f16a8beef__F);
#>   R_RegisterCCallable("rx_ed69a861a9ee57524f69f40f16a8beef_","rx_ed69a861a9ee57524f69f40f16a8beef__Lag", (DL_FUNC) rx_ed69a861a9ee57524f69f40f16a8beef__Lag);
#>   R_RegisterCCallable("rx_ed69a861a9ee57524f69f40f16a8beef_","rx_ed69a861a9ee57524f69f40f16a8beef__Rate", (DL_FUNC) rx_ed69a861a9ee57524f69f40f16a8beef__Rate);
#>   R_RegisterCCallable("rx_ed69a861a9ee57524f69f40f16a8beef_","rx_ed69a861a9ee57524f69f40f16a8beef__Dur", (DL_FUNC) rx_ed69a861a9ee57524f69f40f16a8beef__Dur);
#>   R_RegisterCCallable("rx_ed69a861a9ee57524f69f40f16a8beef_","rx_ed69a861a9ee57524f69f40f16a8beef__mtime", (DL_FUNC) rx_ed69a861a9ee57524f69f40f16a8beef__mtime);
#>   R_RegisterCCallable("rx_ed69a861a9ee57524f69f40f16a8beef_","rx_ed69a861a9ee57524f69f40f16a8beef__ME", (DL_FUNC) rx_ed69a861a9ee57524f69f40f16a8beef__ME);
#>   R_RegisterCCallable("rx_ed69a861a9ee57524f69f40f16a8beef_","rx_ed69a861a9ee57524f69f40f16a8beef__IndF", (DL_FUNC) rx_ed69a861a9ee57524f69f40f16a8beef__IndF);
#>   R_RegisterCCallable("rx_ed69a861a9ee57524f69f40f16a8beef_","rx_ed69a861a9ee57524f69f40f16a8beef__dydt_liblsoda", (DL_FUNC) rx_ed69a861a9ee57524f69f40f16a8beef__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_rx_ed69a861a9ee57524f69f40f16a8beef_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_rx_ed69a861a9ee57524f69f40f16a8beef_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"rx_ed69a861a9ee57524f69f40f16a8beef__model_vars", (DL_FUNC) &rx_ed69a861a9ee57524f69f40f16a8beef__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_rx_ed69a861a9ee57524f69f40f16a8beef_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_ed69a861a9ee57524f69f40f16a8beef__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("rx_ed69a861a9ee57524f69f40f16a8beef__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’
#> ── Solved rxode2 object ──
#> ── Parameters (value$params): ──
#> # A tibble: 1 × 0
#> ── Initial Conditions (value$inits): ──
#> named numeric(0)
#> ── First part of data (object): ──
#> # A tibble: 21 × 6
#>     time     fx dShape1 dShape2 shape1 shape2
#>    <dbl>  <dbl>   <dbl>   <dbl>  <dbl>  <dbl>
#> 1 0.0001 4.15   -6.82    0.386     0.5    1.5
#> 2 0.0501 1.02   -0.608   0.335     0.5    1.5
#> 3 0.100  0.647   0.0845  0.281     0.5    1.5
#> 4 0.150  0.415   0.490   0.224     0.5    1.5
#> 5 0.200  0.241   0.777   0.163     0.5    1.5
#> 6 0.250  0.0976  1.00    0.0985    0.5    1.5
#> # ℹ 15 more rows
# }