log likelihood and derivatives for chi-squared distribution

log likelihood and derivatives for chi-squared distribution

Usage

llikChisq(x, df, full = FALSE)

Arguments

x

variable that is distributed by chi-squared distribution

df

degrees of freedom (non-negative, but can be non-integer).

full

Add the data frame showing x, mean, sd as well as the fx and derivatives

Value

data frame with fx for the log pdf value of with dDf that has the derivatives with respect to the df parameter the observation time-point

Details

In an rxode2() model, you can use llikChisq() but you have to use the x and df arguments. You can also get the derivative of df with llikChisqDdf().

Author

Matthew L. Fidler

Examples

# \donttest{
llikChisq(1, df = 1:3, full=TRUE)
#>   x df        fx         dDf
#> 1 1  1 -1.418939  0.63518142
#> 2 1  2 -1.193147 -0.05796576
#> 3 1  3 -1.418939 -0.36481858

llikChisq(1, df = 6:9)
#>          fx        dDf
#> 1 -3.272589 -0.8079658
#> 2 -4.126989 -0.8981519
#> 3 -5.064348 -0.9746324
#> 4 -6.072899 -1.0410091

et <- et(1:3)
et$x <- 1

model <- function() {
  model({
   fx <- llikChisq(x, time)
   dDf <- llikChisqDdf(x, time)
  })
}

rxSolve(model, et)
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#>  
#>  
#> C model:
#> // Define 2 LHS values
#> #define _LHS_0_ 0 // fx
#> #define _LHS_1_ 1 // dDf
#> #define _getRxSolve_ _rxfb73207dcb1fbde27dd3d5d7aba46a692717__getRxSolve__tvsF
#> #define _evalUdf _rxfb73207dcb1fbde27dd3d5d7aba46a692718__evalUdf_QkpV
#> #define _solveData _rxfb73207dcb1fbde27dd3d5d7aba46a692719__solveData_HF4t
#> #define _assign_ptr _rxfb73207dcb1fbde27dd3d5d7aba46a692720__assign_ptr_E5bq
#> #define _rxRmModelLib _rxfb73207dcb1fbde27dd3d5d7aba46a692721__rxRmModelLib_rWj2
#> #define _rxGetModelLib _rxfb73207dcb1fbde27dd3d5d7aba46a692722__rxGetModelLib_jIRW
#> #define _old_c _rxfb73207dcb1fbde27dd3d5d7aba46a692723__old_c_0crr
#> #define _ptrid _rxfb73207dcb1fbde27dd3d5d7aba46a692724__ptrid_gcGm
#> #define _rxIsCurrentC _rxfb73207dcb1fbde27dd3d5d7aba46a692725__rxIsCurrentC_EsJu
#> #define _sumPS _rxfb73207dcb1fbde27dd3d5d7aba46a692726__sumPS_9xah
#> #define _prodPS _rxfb73207dcb1fbde27dd3d5d7aba46a692727__prodPS_gBOZ
#> #define _prodType _rxfb73207dcb1fbde27dd3d5d7aba46a692728__prodType_l3zY
#> #define _sumType _rxfb73207dcb1fbde27dd3d5d7aba46a692729__sumType_vTa9
#> #define _update_par_ptr _rxfb73207dcb1fbde27dd3d5d7aba46a692730__update_par_ptr_7EhW
#> #define _getParCov _rxfb73207dcb1fbde27dd3d5d7aba46a692731__getParCov_3MG9
#> #define _rxode2_rxAssignPtr _rxfb73207dcb1fbde27dd3d5d7aba46a692732__rxode2_rxAssignPtr_V3pR
#> #define _rxQr _rxfb73207dcb1fbde27dd3d5d7aba46a692733__rxQr_1aNk
#> #define _compareFactorVal _rxfb73207dcb1fbde27dd3d5d7aba46a692734__compareFactorVal_G5Jw
#> #define _sum _rxfb73207dcb1fbde27dd3d5d7aba46a692735__sum_D9TK
#> #define _udf _rxfb73207dcb1fbde27dd3d5d7aba46a692736__udf_40NP
#> #define _sign _rxfb73207dcb1fbde27dd3d5d7aba46a692737__sign_JiAL
#> #define _prod _rxfb73207dcb1fbde27dd3d5d7aba46a692738__prod_ZBrh
#> #define _max _rxfb73207dcb1fbde27dd3d5d7aba46a692739__max_efgh
#> #define _min _rxfb73207dcb1fbde27dd3d5d7aba46a692740__min_Drm8
#> #define _transit4P _rxfb73207dcb1fbde27dd3d5d7aba46a692741__transit4P_5igv
#> #define _transit3P _rxfb73207dcb1fbde27dd3d5d7aba46a692742__transit3P_lJYg
#> #define _assignFuns0 _rxfb73207dcb1fbde27dd3d5d7aba46a692743__assignFuns0_GqQx
#> #define _assignFuns _rxfb73207dcb1fbde27dd3d5d7aba46a692744__assignFuns_LoYl
#> #define _rxord _rxfb73207dcb1fbde27dd3d5d7aba46a692745__rxord_FgJd
#> #define __assignFuns2 _rxfb73207dcb1fbde27dd3d5d7aba46a692746___assignFuns2_KaFg
#> #define _llikCauchyDscale _rxfb73207dcb1fbde27dd3d5d7aba46a692747__llikCauchyDscale_mDR4
#> #define _llikCauchyDlocation _rxfb73207dcb1fbde27dd3d5d7aba46a692748__llikCauchyDlocation_jxuo
#> #define _llikCauchy _rxfb73207dcb1fbde27dd3d5d7aba46a692749__llikCauchy_P8gm
#> #define _llikGammaDrate _rxfb73207dcb1fbde27dd3d5d7aba46a692750__llikGammaDrate_oUom
#> #define _llikGammaDshape _rxfb73207dcb1fbde27dd3d5d7aba46a692751__llikGammaDshape_wLtu
#> #define _llikGamma _rxfb73207dcb1fbde27dd3d5d7aba46a692752__llikGamma_DgOz
#> #define _llikWeibullDscale _rxfb73207dcb1fbde27dd3d5d7aba46a692753__llikWeibullDscale_tPZw
#> #define _llikWeibullDshape _rxfb73207dcb1fbde27dd3d5d7aba46a692754__llikWeibullDshape_Dqx2
#> #define _llikWeibull _rxfb73207dcb1fbde27dd3d5d7aba46a692755__llikWeibull_t0Ff
#> #define _llikUnifDbeta _rxfb73207dcb1fbde27dd3d5d7aba46a692756__llikUnifDbeta_CRup
#> #define _llikUnifDalpha _rxfb73207dcb1fbde27dd3d5d7aba46a692757__llikUnifDalpha_BfcL
#> #define _llikUnif _rxfb73207dcb1fbde27dd3d5d7aba46a692758__llikUnif_LlJk
#> #define _llikGeomDp _rxfb73207dcb1fbde27dd3d5d7aba46a692759__llikGeomDp_0HOu
#> #define _llikGeom _rxfb73207dcb1fbde27dd3d5d7aba46a692760__llikGeom_quMQ
#> #define _llikFDdf2 _rxfb73207dcb1fbde27dd3d5d7aba46a692761__llikFDdf2_IP4s
#> #define _llikFDdf1 _rxfb73207dcb1fbde27dd3d5d7aba46a692762__llikFDdf1_BKwl
#> #define _llikF _rxfb73207dcb1fbde27dd3d5d7aba46a692763__llikF_ION1
#> #define _llikExpDrate _rxfb73207dcb1fbde27dd3d5d7aba46a692764__llikExpDrate_tTX5
#> #define _llikExp _rxfb73207dcb1fbde27dd3d5d7aba46a692765__llikExp_DulH
#> #define _llikChisqDdf _rxfb73207dcb1fbde27dd3d5d7aba46a692766__llikChisqDdf_LxeA
#> #define _llikChisq _rxfb73207dcb1fbde27dd3d5d7aba46a692767__llikChisq_Ut0p
#> #define _llikTDsd _rxfb73207dcb1fbde27dd3d5d7aba46a692768__llikTDsd_o6Aq
#> #define _llikTDmean _rxfb73207dcb1fbde27dd3d5d7aba46a692769__llikTDmean_SQkn
#> #define _llikTDdf _rxfb73207dcb1fbde27dd3d5d7aba46a692770__llikTDdf_H8fP
#> #define _llikT _rxfb73207dcb1fbde27dd3d5d7aba46a692771__llikT_CAnN
#> #define _llikBetaDshape2 _rxfb73207dcb1fbde27dd3d5d7aba46a692772__llikBetaDshape2_kUms
#> #define _llikBetaDshape1 _rxfb73207dcb1fbde27dd3d5d7aba46a692773__llikBetaDshape1_wiOW
#> #define _llikBeta _rxfb73207dcb1fbde27dd3d5d7aba46a692774__llikBeta_G16H
#> #define _llikNbinomMuDmu _rxfb73207dcb1fbde27dd3d5d7aba46a692775__llikNbinomMuDmu_PUgn
#> #define _llikNbinomMu _rxfb73207dcb1fbde27dd3d5d7aba46a692776__llikNbinomMu_Cjuw
#> #define _llikNbinomDprob _rxfb73207dcb1fbde27dd3d5d7aba46a692777__llikNbinomDprob_A8iw
#> #define _llikNbinom _rxfb73207dcb1fbde27dd3d5d7aba46a692778__llikNbinom_bn2G
#> #define _llikBinomDprob _rxfb73207dcb1fbde27dd3d5d7aba46a692779__llikBinomDprob_TQjS
#> #define _llikBinom _rxfb73207dcb1fbde27dd3d5d7aba46a692780__llikBinom_rer5
#> #define _llikPoisDlambda _rxfb73207dcb1fbde27dd3d5d7aba46a692781__llikPoisDlambda_zexL
#> #define _llikPois _rxfb73207dcb1fbde27dd3d5d7aba46a692782__llikPois_gAwM
#> #define _llikNormDsd _rxfb73207dcb1fbde27dd3d5d7aba46a692783__llikNormDsd_E6nq
#> #define _llikNormDmean _rxfb73207dcb1fbde27dd3d5d7aba46a692784__llikNormDmean_R3DD
#> #define _llikNorm _rxfb73207dcb1fbde27dd3d5d7aba46a692785__llikNorm_9jN5
#> #define simeps _rxfb73207dcb1fbde27dd3d5d7aba46a692786_simeps_O4r9
#> #define simeta _rxfb73207dcb1fbde27dd3d5d7aba46a692787_simeta_D97m
#> #define expit _rxfb73207dcb1fbde27dd3d5d7aba46a692788_expit_e5yb
#> #define logit _rxfb73207dcb1fbde27dd3d5d7aba46a692789_logit_V8Eu
#> #define gammapDer _rxfb73207dcb1fbde27dd3d5d7aba46a692790_gammapDer_Z1xT
#> #define lowergamma _rxfb73207dcb1fbde27dd3d5d7aba46a692791_lowergamma_QYOQ
#> #define uppergamma _rxfb73207dcb1fbde27dd3d5d7aba46a692792_uppergamma_e6aq
#> #define gammaqInva _rxfb73207dcb1fbde27dd3d5d7aba46a692793_gammaqInva_t3Lb
#> #define gammaqInv _rxfb73207dcb1fbde27dd3d5d7aba46a692794_gammaqInv_sbea
#> #define gammapInva _rxfb73207dcb1fbde27dd3d5d7aba46a692795_gammapInva_w5m4
#> #define gammapInv _rxfb73207dcb1fbde27dd3d5d7aba46a692796_gammapInv_brD9
#> #define gammaq _rxfb73207dcb1fbde27dd3d5d7aba46a692797_gammaq_EZAJ
#> #define gammap _rxfb73207dcb1fbde27dd3d5d7aba46a692798_gammap_Ne7I
#> #define phi _rxfb73207dcb1fbde27dd3d5d7aba46a692799_phi_xUFA
#> #define d2ELUa _rxfb73207dcb1fbde27dd3d5d7aba46a692800_d2ELUa_5R2w
#> #define dELUa _rxfb73207dcb1fbde27dd3d5d7aba46a692801_dELUa_ZwwB
#> #define d2aELU _rxfb73207dcb1fbde27dd3d5d7aba46a692802_d2aELU_gRvb
#> #define d2ELU _rxfb73207dcb1fbde27dd3d5d7aba46a692803_d2ELU_lnk3
#> #define dELU _rxfb73207dcb1fbde27dd3d5d7aba46a692804_dELU_57cf
#> #define ELU _rxfb73207dcb1fbde27dd3d5d7aba46a692805_ELU_UwMa
#> #define dPReLUa1 _rxfb73207dcb1fbde27dd3d5d7aba46a692806_dPReLUa1_rEA9
#> #define dPReLUa _rxfb73207dcb1fbde27dd3d5d7aba46a692807_dPReLUa_z1SE
#> #define dPReLU _rxfb73207dcb1fbde27dd3d5d7aba46a692808_dPReLU_1KzA
#> #define PReLU _rxfb73207dcb1fbde27dd3d5d7aba46a692809_PReLU_f2H4
#> #define dSwish _rxfb73207dcb1fbde27dd3d5d7aba46a692810_dSwish_Q2oF
#> #define Swish _rxfb73207dcb1fbde27dd3d5d7aba46a692811_Swish_0joC
#> #define dlReLU _rxfb73207dcb1fbde27dd3d5d7aba46a692812_dlReLU_Gj0a
#> #define lReLU _rxfb73207dcb1fbde27dd3d5d7aba46a692813_lReLU_8wAb
#> #define dSELU _rxfb73207dcb1fbde27dd3d5d7aba46a692814_dSELU_QELO
#> #define SELU _rxfb73207dcb1fbde27dd3d5d7aba46a692815_SELU_lObq
#> #define d4softplus _rxfb73207dcb1fbde27dd3d5d7aba46a692816_d4softplus_Nath
#> #define d3softplus _rxfb73207dcb1fbde27dd3d5d7aba46a692817_d3softplus_ltiW
#> #define d2softplus _rxfb73207dcb1fbde27dd3d5d7aba46a692818_d2softplus_R2IG
#> #define dsoftplus _rxfb73207dcb1fbde27dd3d5d7aba46a692819_dsoftplus_VLQY
#> #define softplus _rxfb73207dcb1fbde27dd3d5d7aba46a692820_softplus_M84J
#> #define d4GELU _rxfb73207dcb1fbde27dd3d5d7aba46a692821_d4GELU_Cnuy
#> #define d3GELU _rxfb73207dcb1fbde27dd3d5d7aba46a692822_d3GELU_CXPJ
#> #define d2GELU _rxfb73207dcb1fbde27dd3d5d7aba46a692823_d2GELU_UjvF
#> #define dGELU _rxfb73207dcb1fbde27dd3d5d7aba46a692824_dGELU_k7KJ
#> #define GELU _rxfb73207dcb1fbde27dd3d5d7aba46a692825_GELU_vDxq
#> #define dReLU _rxfb73207dcb1fbde27dd3d5d7aba46a692826_dReLU_bVSK
#> #define ReLU _rxfb73207dcb1fbde27dd3d5d7aba46a692827_ReLU_Vfr7
#> #define riweibull _rxfb73207dcb1fbde27dd3d5d7aba46a692828_riweibull_TcaU
#> #define riunif _rxfb73207dcb1fbde27dd3d5d7aba46a692829_riunif_EgpY
#> #define rit_ _rxfb73207dcb1fbde27dd3d5d7aba46a692830_rit__kNIq
#> #define ripois _rxfb73207dcb1fbde27dd3d5d7aba46a692831_ripois_nEnm
#> #define ribeta _rxfb73207dcb1fbde27dd3d5d7aba46a692832_ribeta_hSbo
#> #define rigamma _rxfb73207dcb1fbde27dd3d5d7aba46a692833_rigamma_MZYB
#> #define rigeom _rxfb73207dcb1fbde27dd3d5d7aba46a692834_rigeom_jkiC
#> #define rif _rxfb73207dcb1fbde27dd3d5d7aba46a692835_rif_oUqv
#> #define riexp _rxfb73207dcb1fbde27dd3d5d7aba46a692836_riexp_niI7
#> #define richisq _rxfb73207dcb1fbde27dd3d5d7aba46a692837_richisq_5Ui0
#> #define ricauchy _rxfb73207dcb1fbde27dd3d5d7aba46a692838_ricauchy_GVTs
#> #define rinbinomMu _rxfb73207dcb1fbde27dd3d5d7aba46a692839_rinbinomMu_34TN
#> #define rinbinom _rxfb73207dcb1fbde27dd3d5d7aba46a692840_rinbinom_HeFp
#> #define ribinom _rxfb73207dcb1fbde27dd3d5d7aba46a692841_ribinom_ue1Z
#> #define rinorm _rxfb73207dcb1fbde27dd3d5d7aba46a692842_rinorm_aEnJ
#> #define rxweibull _rxfb73207dcb1fbde27dd3d5d7aba46a692843_rxweibull_R6UH
#> #define rxunif _rxfb73207dcb1fbde27dd3d5d7aba46a692844_rxunif_ZKbS
#> #define rxt_ _rxfb73207dcb1fbde27dd3d5d7aba46a692845_rxt__smNY
#> #define rxpois _rxfb73207dcb1fbde27dd3d5d7aba46a692846_rxpois_lS2w
#> #define rxbeta _rxfb73207dcb1fbde27dd3d5d7aba46a692847_rxbeta_zYqw
#> #define rxgamma _rxfb73207dcb1fbde27dd3d5d7aba46a692848_rxgamma_2z8b
#> #define rxgeom _rxfb73207dcb1fbde27dd3d5d7aba46a692849_rxgeom_GMdj
#> #define rxf _rxfb73207dcb1fbde27dd3d5d7aba46a692850_rxf_LB3L
#> #define rxexp _rxfb73207dcb1fbde27dd3d5d7aba46a692851_rxexp_sOUR
#> #define rxchisq _rxfb73207dcb1fbde27dd3d5d7aba46a692852_rxchisq_bPYU
#> #define rxcauchy _rxfb73207dcb1fbde27dd3d5d7aba46a692853_rxcauchy_PmW8
#> #define rxnbinomMu _rxfb73207dcb1fbde27dd3d5d7aba46a692854_rxnbinomMu_sOmG
#> #define rxnbinom _rxfb73207dcb1fbde27dd3d5d7aba46a692855_rxnbinom_ypO5
#> #define rxbinom _rxfb73207dcb1fbde27dd3d5d7aba46a692856_rxbinom_5OnO
#> #define rxnorm _rxfb73207dcb1fbde27dd3d5d7aba46a692857_rxnorm_E9TF
#> #define linCmtB _rxfb73207dcb1fbde27dd3d5d7aba46a692858_linCmtB_yOK7
#> #define linCmtA _rxfb73207dcb1fbde27dd3d5d7aba46a692859_linCmtA_OqEK
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){  if (_solveData->svar[_svari] == 0) {x = _PP[0];}; }
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){  if (_solveData->ovar[_ovari] == 0) {x = _PP[0];}; }
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  rx_a238fb4244abacc22ca4458c4df93549__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *rx_a238fb4244abacc22ca4458c4df93549__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP rx_a238fb4244abacc22ca4458c4df93549__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void rx_a238fb4244abacc22ca4458c4df93549__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double x = NA_REAL;
#>   double fx = NA_REAL;
#>   double dDf = NA_REAL;
#> 
#>   (void)t;
#>   (void)x;
#>   (void)fx;
#>   (void)dDf;
#> 
#>   fx = _PL[0];
#>   dDf = _PL[1];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#>   x = _PP[0];
#> 
#> 
#>   fx =llikChisq(x,t);
#>   dDf =llikChisqDdf(x,t);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void rx_a238fb4244abacc22ca4458c4df93549__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void rx_a238fb4244abacc22ca4458c4df93549__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void rx_a238fb4244abacc22ca4458c4df93549__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double x = NA_REAL;
#>   double fx = NA_REAL;
#>   double dDf = NA_REAL;
#> 
#>   (void)t;
#>   (void)x;
#>   (void)fx;
#>   (void)dDf;
#> 
#>   fx = _PL[0];
#>   dDf = _PL[1];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#>   x = _PP[0];
#> 
#> 
#>   fx =llikChisq(x,t);
#>   dDf =llikChisqDdf(x,t);
#> 
#>   _lhs[_LHS_0_]=fx;
#>   _lhs[_LHS_1_]=dDf;
#> }
#> // Functional based bioavailability
#> double rx_a238fb4244abacc22ca4458c4df93549__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double rx_a238fb4244abacc22ca4458c4df93549__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double rx_a238fb4244abacc22ca4458c4df93549__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double rx_a238fb4244abacc22ca4458c4df93549__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void rx_a238fb4244abacc22ca4458c4df93549__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void rx_a238fb4244abacc22ca4458c4df93549__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void rx_a238fb4244abacc22ca4458c4df93549__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP rx_a238fb4244abacc22ca4458c4df93549__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_a238fb4244abacc22ca4458c4df93549__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAA[91AAAv7#aT)^n.YNAX\?0agtjmYBaifx>BLCzCw2r_TADv}&G=cY024w2yOm]1W0.dj{o^yTeiOEwE\"%%=:D^5H.=1NAGVLmK!MzMFtjAqBmeZ8Y/:;^6#K9ilHHYM^jCb8{<KnM(;KC[QSN[=d!J*f@#n$=v9^oV1!Kp/|T!H5IV72$\?w>0=:MekFOm3D_sGZUF].y)faH@$}F}yWmr=P;Rr\"KP!\?0\"Qp(~R2g3g@!G60`U13aL.!iU#qfU&/F{:55}`/IRKNRkE/o{}PXgz&wdNkt8\?0Z1IAf&<:fSM$0ARzIb9[Y>SF>gV10W;AxM%%\"5S2o=\?=\?1Nt2$bUE5UDkz/Gruet\?(=CcPhyy7r=VFizbC{l&S5CDt#Xo8<Wfo]$(#AC:c$FQ@Yp5<Ch~ID66d^L4O45*M1oL5\"J8yaNBYdHG[XPuhY3;K\"[,F`c)YYQ%%T+>r#hX^CMvA%%#`s{I:.Ucw3pHdCcly7.85d#`},fP%%E1Qa0X@%%I3iUb\?n/zCG6vy,g\?fy_H|B.ed+PO9<:.#Hj@>GT\?Ec}<\"{PRs94PKC/aP]3|}UaggON4*x{O*9PRB8I\?WxN%%t&)qU3L%%FuC$|i7yTUNt9T/f>lhPtg*SP4@4,N\"Hn#k]0(tN9*_n[X9/[~G1oU1YU{PMnwaDsy[Mb6!QN>q=2AP37!=)\?.NPNep;,VFb8$W^\?_PV)C()wo5WJ)`4STs]bN3:5TiEcc^aM>}MTYPE;B*]_bQM[i&qhnWd;`lSGq@2XHe5Yhm)QeC(jbvlE*Mvz/BeR3)P7THU^7B^va@9BIY%%t]2\"CBYXRvZZT$0E$+i\?dJ\"&pOD])p}#Yeva]Vjhg+sT5`CA*<>b+iHqS7C=[U970$6)[kotJyaN^:!8~_19]<0Vw4zY%%G%%!w,0x:>);r^G&Kxd7{}HXjuEYE!9m(o<YG[Z|k\"7gSoJU\"lHqtq|cV>DD&J>=Ta[.NiB*8RV`S5F2pvYrTK\?J`(@*)QD}beINE{p57[2G<]u&VoE`wh|a!gybwr.`X8z5a@u+O`18EXOZ=0(A]fB^WYs*+n@]yh^2tvGFPYY24P");
#>     char buf[1101];
#> #define __doBufN__ 1101
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void rx_a238fb4244abacc22ca4458c4df93549__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   rx_a238fb4244abacc22ca4458c4df93549__dydt(neq, *t, A, DADT);
#> }
#> extern int rx_a238fb4244abacc22ca4458c4df93549__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   rx_a238fb4244abacc22ca4458c4df93549__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void rx_a238fb4244abacc22ca4458c4df93549__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   rx_a238fb4244abacc22ca4458c4df93549__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void rx_a238fb4244abacc22ca4458c4df93549__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_rx_a238fb4244abacc22ca4458c4df93549_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("rx_a238fb4244abacc22ca4458c4df93549_","rx_a238fb4244abacc22ca4458c4df93549__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("rx_a238fb4244abacc22ca4458c4df93549_","rx_a238fb4244abacc22ca4458c4df93549__assignFuns", (DL_FUNC) rx_a238fb4244abacc22ca4458c4df93549__assignFuns);
#>   R_RegisterCCallable("rx_a238fb4244abacc22ca4458c4df93549_","rx_a238fb4244abacc22ca4458c4df93549__inis",(DL_FUNC) rx_a238fb4244abacc22ca4458c4df93549__inis);
#>   R_RegisterCCallable("rx_a238fb4244abacc22ca4458c4df93549_","rx_a238fb4244abacc22ca4458c4df93549__dydt",(DL_FUNC) rx_a238fb4244abacc22ca4458c4df93549__dydt);
#>   R_RegisterCCallable("rx_a238fb4244abacc22ca4458c4df93549_","rx_a238fb4244abacc22ca4458c4df93549__calc_lhs",(DL_FUNC) rx_a238fb4244abacc22ca4458c4df93549__calc_lhs);
#>   R_RegisterCCallable("rx_a238fb4244abacc22ca4458c4df93549_","rx_a238fb4244abacc22ca4458c4df93549__calc_jac",(DL_FUNC) rx_a238fb4244abacc22ca4458c4df93549__calc_jac);
#>   R_RegisterCCallable("rx_a238fb4244abacc22ca4458c4df93549_","rx_a238fb4244abacc22ca4458c4df93549__dydt_lsoda", (DL_FUNC) rx_a238fb4244abacc22ca4458c4df93549__dydt_lsoda);
#>   R_RegisterCCallable("rx_a238fb4244abacc22ca4458c4df93549_","rx_a238fb4244abacc22ca4458c4df93549__calc_jac_lsoda", (DL_FUNC) rx_a238fb4244abacc22ca4458c4df93549__calc_jac_lsoda);
#>   R_RegisterCCallable("rx_a238fb4244abacc22ca4458c4df93549_","rx_a238fb4244abacc22ca4458c4df93549__ode_solver_solvedata", (DL_FUNC) rx_a238fb4244abacc22ca4458c4df93549__ode_solver_solvedata);
#>   R_RegisterCCallable("rx_a238fb4244abacc22ca4458c4df93549_","rx_a238fb4244abacc22ca4458c4df93549__ode_solver_get_solvedata", (DL_FUNC) rx_a238fb4244abacc22ca4458c4df93549__ode_solver_get_solvedata);
#>   R_RegisterCCallable("rx_a238fb4244abacc22ca4458c4df93549_","rx_a238fb4244abacc22ca4458c4df93549__F", (DL_FUNC) rx_a238fb4244abacc22ca4458c4df93549__F);
#>   R_RegisterCCallable("rx_a238fb4244abacc22ca4458c4df93549_","rx_a238fb4244abacc22ca4458c4df93549__Lag", (DL_FUNC) rx_a238fb4244abacc22ca4458c4df93549__Lag);
#>   R_RegisterCCallable("rx_a238fb4244abacc22ca4458c4df93549_","rx_a238fb4244abacc22ca4458c4df93549__Rate", (DL_FUNC) rx_a238fb4244abacc22ca4458c4df93549__Rate);
#>   R_RegisterCCallable("rx_a238fb4244abacc22ca4458c4df93549_","rx_a238fb4244abacc22ca4458c4df93549__Dur", (DL_FUNC) rx_a238fb4244abacc22ca4458c4df93549__Dur);
#>   R_RegisterCCallable("rx_a238fb4244abacc22ca4458c4df93549_","rx_a238fb4244abacc22ca4458c4df93549__mtime", (DL_FUNC) rx_a238fb4244abacc22ca4458c4df93549__mtime);
#>   R_RegisterCCallable("rx_a238fb4244abacc22ca4458c4df93549_","rx_a238fb4244abacc22ca4458c4df93549__ME", (DL_FUNC) rx_a238fb4244abacc22ca4458c4df93549__ME);
#>   R_RegisterCCallable("rx_a238fb4244abacc22ca4458c4df93549_","rx_a238fb4244abacc22ca4458c4df93549__IndF", (DL_FUNC) rx_a238fb4244abacc22ca4458c4df93549__IndF);
#>   R_RegisterCCallable("rx_a238fb4244abacc22ca4458c4df93549_","rx_a238fb4244abacc22ca4458c4df93549__dydt_liblsoda", (DL_FUNC) rx_a238fb4244abacc22ca4458c4df93549__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_rx_a238fb4244abacc22ca4458c4df93549_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_rx_a238fb4244abacc22ca4458c4df93549_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"rx_a238fb4244abacc22ca4458c4df93549__model_vars", (DL_FUNC) &rx_a238fb4244abacc22ca4458c4df93549__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_rx_a238fb4244abacc22ca4458c4df93549_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_a238fb4244abacc22ca4458c4df93549__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("rx_a238fb4244abacc22ca4458c4df93549__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’
#> ── Solved rxode2 object ──
#> ── Parameters (value$params): ──
#> # A tibble: 1 × 0
#> ── Initial Conditions (value$inits): ──
#> named numeric(0)
#> ── First part of data (object): ──
#> # A tibble: 3 × 4
#>    time    fx     dDf     x
#>   <dbl> <dbl>   <dbl> <dbl>
#> 1     1 -1.42  0.635      1
#> 2     2 -1.19 -0.0580     1
#> 3     3 -1.42 -0.365      1
# }