log likelihood and derivatives for exponential distribution

log likelihood and derivatives for exponential distribution

Usage

llikExp(x, rate, full = FALSE)

Arguments

x

variable that is distributed by exponential distribution

rate

vector of rates.

full

Add the data frame showing x, mean, sd as well as the fx and derivatives

Value

data frame with fx for the log pdf value of with dRate that has the derivatives with respect to the rate parameter the observation time-point

Details

In an rxode2() model, you can use llikExp() but you have to use the x and rate arguments. You can also get the derivative of rate with llikExpDrate().

Author

Matthew L. Fidler

Examples

# \donttest{
llikExp(1, 1:3)
#>          fx      dRate
#> 1 -1.000000  0.0000000
#> 2 -1.306853 -0.5000000
#> 3 -1.901388 -0.6666667

llikExp(1, 1:3, full=TRUE)
#>   x rate        fx      dRate
#> 1 1    1 -1.000000  0.0000000
#> 2 1    2 -1.306853 -0.5000000
#> 3 1    3 -1.901388 -0.6666667

# You can use rxode2 for these too:

et <- et(1:3)
et$x <- 1

model <- function() {
  model({
    fx <- llikExp(x, time)
    dRate <- llikExpDrate(x, time)
  })
}

rxSolve(model, et)
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#>  
#>  
#> C model:
#> // Define 2 LHS values
#> #define _LHS_0_ 0 // fx
#> #define _LHS_1_ 1 // dRate
#> #define _getRxSolve_ _rx4e6ed916a2098ff9787dc3ff4af05bf82860__getRxSolve__QkpV
#> #define _evalUdf _rx4e6ed916a2098ff9787dc3ff4af05bf82861__evalUdf_HF4t
#> #define _solveData _rx4e6ed916a2098ff9787dc3ff4af05bf82862__solveData_E5bq
#> #define _assign_ptr _rx4e6ed916a2098ff9787dc3ff4af05bf82863__assign_ptr_rWj2
#> #define _rxRmModelLib _rx4e6ed916a2098ff9787dc3ff4af05bf82864__rxRmModelLib_jIRW
#> #define _rxGetModelLib _rx4e6ed916a2098ff9787dc3ff4af05bf82865__rxGetModelLib_0crr
#> #define _old_c _rx4e6ed916a2098ff9787dc3ff4af05bf82866__old_c_gcGm
#> #define _ptrid _rx4e6ed916a2098ff9787dc3ff4af05bf82867__ptrid_EsJu
#> #define _rxIsCurrentC _rx4e6ed916a2098ff9787dc3ff4af05bf82868__rxIsCurrentC_9xah
#> #define _sumPS _rx4e6ed916a2098ff9787dc3ff4af05bf82869__sumPS_gBOZ
#> #define _prodPS _rx4e6ed916a2098ff9787dc3ff4af05bf82870__prodPS_l3zY
#> #define _prodType _rx4e6ed916a2098ff9787dc3ff4af05bf82871__prodType_vTa9
#> #define _sumType _rx4e6ed916a2098ff9787dc3ff4af05bf82872__sumType_7EhW
#> #define _update_par_ptr _rx4e6ed916a2098ff9787dc3ff4af05bf82873__update_par_ptr_3MG9
#> #define _getParCov _rx4e6ed916a2098ff9787dc3ff4af05bf82874__getParCov_V3pR
#> #define _rxode2_rxAssignPtr _rx4e6ed916a2098ff9787dc3ff4af05bf82875__rxode2_rxAssignPtr_1aNk
#> #define _rxQr _rx4e6ed916a2098ff9787dc3ff4af05bf82876__rxQr_G5Jw
#> #define _compareFactorVal _rx4e6ed916a2098ff9787dc3ff4af05bf82877__compareFactorVal_D9TK
#> #define _sum _rx4e6ed916a2098ff9787dc3ff4af05bf82878__sum_40NP
#> #define _udf _rx4e6ed916a2098ff9787dc3ff4af05bf82879__udf_JiAL
#> #define _sign _rx4e6ed916a2098ff9787dc3ff4af05bf82880__sign_ZBrh
#> #define _prod _rx4e6ed916a2098ff9787dc3ff4af05bf82881__prod_efgh
#> #define _max _rx4e6ed916a2098ff9787dc3ff4af05bf82882__max_Drm8
#> #define _min _rx4e6ed916a2098ff9787dc3ff4af05bf82883__min_5igv
#> #define _transit4P _rx4e6ed916a2098ff9787dc3ff4af05bf82884__transit4P_lJYg
#> #define _transit3P _rx4e6ed916a2098ff9787dc3ff4af05bf82885__transit3P_GqQx
#> #define _assignFuns0 _rx4e6ed916a2098ff9787dc3ff4af05bf82886__assignFuns0_LoYl
#> #define _assignFuns _rx4e6ed916a2098ff9787dc3ff4af05bf82887__assignFuns_FgJd
#> #define _rxord _rx4e6ed916a2098ff9787dc3ff4af05bf82888__rxord_KaFg
#> #define __assignFuns2 _rx4e6ed916a2098ff9787dc3ff4af05bf82889___assignFuns2_mDR4
#> #define _llikCauchyDscale _rx4e6ed916a2098ff9787dc3ff4af05bf82890__llikCauchyDscale_jxuo
#> #define _llikCauchyDlocation _rx4e6ed916a2098ff9787dc3ff4af05bf82891__llikCauchyDlocation_P8gm
#> #define _llikCauchy _rx4e6ed916a2098ff9787dc3ff4af05bf82892__llikCauchy_oUom
#> #define _llikGammaDrate _rx4e6ed916a2098ff9787dc3ff4af05bf82893__llikGammaDrate_wLtu
#> #define _llikGammaDshape _rx4e6ed916a2098ff9787dc3ff4af05bf82894__llikGammaDshape_DgOz
#> #define _llikGamma _rx4e6ed916a2098ff9787dc3ff4af05bf82895__llikGamma_tPZw
#> #define _llikWeibullDscale _rx4e6ed916a2098ff9787dc3ff4af05bf82896__llikWeibullDscale_Dqx2
#> #define _llikWeibullDshape _rx4e6ed916a2098ff9787dc3ff4af05bf82897__llikWeibullDshape_t0Ff
#> #define _llikWeibull _rx4e6ed916a2098ff9787dc3ff4af05bf82898__llikWeibull_CRup
#> #define _llikUnifDbeta _rx4e6ed916a2098ff9787dc3ff4af05bf82899__llikUnifDbeta_BfcL
#> #define _llikUnifDalpha _rx4e6ed916a2098ff9787dc3ff4af05bf82900__llikUnifDalpha_LlJk
#> #define _llikUnif _rx4e6ed916a2098ff9787dc3ff4af05bf82901__llikUnif_0HOu
#> #define _llikGeomDp _rx4e6ed916a2098ff9787dc3ff4af05bf82902__llikGeomDp_quMQ
#> #define _llikGeom _rx4e6ed916a2098ff9787dc3ff4af05bf82903__llikGeom_IP4s
#> #define _llikFDdf2 _rx4e6ed916a2098ff9787dc3ff4af05bf82904__llikFDdf2_BKwl
#> #define _llikFDdf1 _rx4e6ed916a2098ff9787dc3ff4af05bf82905__llikFDdf1_ION1
#> #define _llikF _rx4e6ed916a2098ff9787dc3ff4af05bf82906__llikF_tTX5
#> #define _llikExpDrate _rx4e6ed916a2098ff9787dc3ff4af05bf82907__llikExpDrate_DulH
#> #define _llikExp _rx4e6ed916a2098ff9787dc3ff4af05bf82908__llikExp_LxeA
#> #define _llikChisqDdf _rx4e6ed916a2098ff9787dc3ff4af05bf82909__llikChisqDdf_Ut0p
#> #define _llikChisq _rx4e6ed916a2098ff9787dc3ff4af05bf82910__llikChisq_o6Aq
#> #define _llikTDsd _rx4e6ed916a2098ff9787dc3ff4af05bf82911__llikTDsd_SQkn
#> #define _llikTDmean _rx4e6ed916a2098ff9787dc3ff4af05bf82912__llikTDmean_H8fP
#> #define _llikTDdf _rx4e6ed916a2098ff9787dc3ff4af05bf82913__llikTDdf_CAnN
#> #define _llikT _rx4e6ed916a2098ff9787dc3ff4af05bf82914__llikT_kUms
#> #define _llikBetaDshape2 _rx4e6ed916a2098ff9787dc3ff4af05bf82915__llikBetaDshape2_wiOW
#> #define _llikBetaDshape1 _rx4e6ed916a2098ff9787dc3ff4af05bf82916__llikBetaDshape1_G16H
#> #define _llikBeta _rx4e6ed916a2098ff9787dc3ff4af05bf82917__llikBeta_PUgn
#> #define _llikNbinomMuDmu _rx4e6ed916a2098ff9787dc3ff4af05bf82918__llikNbinomMuDmu_Cjuw
#> #define _llikNbinomMu _rx4e6ed916a2098ff9787dc3ff4af05bf82919__llikNbinomMu_A8iw
#> #define _llikNbinomDprob _rx4e6ed916a2098ff9787dc3ff4af05bf82920__llikNbinomDprob_bn2G
#> #define _llikNbinom _rx4e6ed916a2098ff9787dc3ff4af05bf82921__llikNbinom_TQjS
#> #define _llikBinomDprob _rx4e6ed916a2098ff9787dc3ff4af05bf82922__llikBinomDprob_rer5
#> #define _llikBinom _rx4e6ed916a2098ff9787dc3ff4af05bf82923__llikBinom_zexL
#> #define _llikPoisDlambda _rx4e6ed916a2098ff9787dc3ff4af05bf82924__llikPoisDlambda_gAwM
#> #define _llikPois _rx4e6ed916a2098ff9787dc3ff4af05bf82925__llikPois_E6nq
#> #define _llikNormDsd _rx4e6ed916a2098ff9787dc3ff4af05bf82926__llikNormDsd_R3DD
#> #define _llikNormDmean _rx4e6ed916a2098ff9787dc3ff4af05bf82927__llikNormDmean_9jN5
#> #define _llikNorm _rx4e6ed916a2098ff9787dc3ff4af05bf82928__llikNorm_O4r9
#> #define simeps _rx4e6ed916a2098ff9787dc3ff4af05bf82929_simeps_D97m
#> #define simeta _rx4e6ed916a2098ff9787dc3ff4af05bf82930_simeta_e5yb
#> #define expit _rx4e6ed916a2098ff9787dc3ff4af05bf82931_expit_V8Eu
#> #define logit _rx4e6ed916a2098ff9787dc3ff4af05bf82932_logit_Z1xT
#> #define gammapDer _rx4e6ed916a2098ff9787dc3ff4af05bf82933_gammapDer_QYOQ
#> #define lowergamma _rx4e6ed916a2098ff9787dc3ff4af05bf82934_lowergamma_e6aq
#> #define uppergamma _rx4e6ed916a2098ff9787dc3ff4af05bf82935_uppergamma_t3Lb
#> #define gammaqInva _rx4e6ed916a2098ff9787dc3ff4af05bf82936_gammaqInva_sbea
#> #define gammaqInv _rx4e6ed916a2098ff9787dc3ff4af05bf82937_gammaqInv_w5m4
#> #define gammapInva _rx4e6ed916a2098ff9787dc3ff4af05bf82938_gammapInva_brD9
#> #define gammapInv _rx4e6ed916a2098ff9787dc3ff4af05bf82939_gammapInv_EZAJ
#> #define gammaq _rx4e6ed916a2098ff9787dc3ff4af05bf82940_gammaq_Ne7I
#> #define gammap _rx4e6ed916a2098ff9787dc3ff4af05bf82941_gammap_xUFA
#> #define phi _rx4e6ed916a2098ff9787dc3ff4af05bf82942_phi_5R2w
#> #define d2ELUa _rx4e6ed916a2098ff9787dc3ff4af05bf82943_d2ELUa_ZwwB
#> #define dELUa _rx4e6ed916a2098ff9787dc3ff4af05bf82944_dELUa_gRvb
#> #define d2aELU _rx4e6ed916a2098ff9787dc3ff4af05bf82945_d2aELU_lnk3
#> #define d2ELU _rx4e6ed916a2098ff9787dc3ff4af05bf82946_d2ELU_57cf
#> #define dELU _rx4e6ed916a2098ff9787dc3ff4af05bf82947_dELU_UwMa
#> #define ELU _rx4e6ed916a2098ff9787dc3ff4af05bf82948_ELU_rEA9
#> #define dPReLUa1 _rx4e6ed916a2098ff9787dc3ff4af05bf82949_dPReLUa1_z1SE
#> #define dPReLUa _rx4e6ed916a2098ff9787dc3ff4af05bf82950_dPReLUa_1KzA
#> #define dPReLU _rx4e6ed916a2098ff9787dc3ff4af05bf82951_dPReLU_f2H4
#> #define PReLU _rx4e6ed916a2098ff9787dc3ff4af05bf82952_PReLU_Q2oF
#> #define dSwish _rx4e6ed916a2098ff9787dc3ff4af05bf82953_dSwish_0joC
#> #define Swish _rx4e6ed916a2098ff9787dc3ff4af05bf82954_Swish_Gj0a
#> #define dlReLU _rx4e6ed916a2098ff9787dc3ff4af05bf82955_dlReLU_8wAb
#> #define lReLU _rx4e6ed916a2098ff9787dc3ff4af05bf82956_lReLU_QELO
#> #define dSELU _rx4e6ed916a2098ff9787dc3ff4af05bf82957_dSELU_lObq
#> #define SELU _rx4e6ed916a2098ff9787dc3ff4af05bf82958_SELU_Nath
#> #define d4softplus _rx4e6ed916a2098ff9787dc3ff4af05bf82959_d4softplus_ltiW
#> #define d3softplus _rx4e6ed916a2098ff9787dc3ff4af05bf82960_d3softplus_R2IG
#> #define d2softplus _rx4e6ed916a2098ff9787dc3ff4af05bf82961_d2softplus_VLQY
#> #define dsoftplus _rx4e6ed916a2098ff9787dc3ff4af05bf82962_dsoftplus_M84J
#> #define softplus _rx4e6ed916a2098ff9787dc3ff4af05bf82963_softplus_Cnuy
#> #define d4GELU _rx4e6ed916a2098ff9787dc3ff4af05bf82964_d4GELU_CXPJ
#> #define d3GELU _rx4e6ed916a2098ff9787dc3ff4af05bf82965_d3GELU_UjvF
#> #define d2GELU _rx4e6ed916a2098ff9787dc3ff4af05bf82966_d2GELU_k7KJ
#> #define dGELU _rx4e6ed916a2098ff9787dc3ff4af05bf82967_dGELU_vDxq
#> #define GELU _rx4e6ed916a2098ff9787dc3ff4af05bf82968_GELU_bVSK
#> #define dReLU _rx4e6ed916a2098ff9787dc3ff4af05bf82969_dReLU_Vfr7
#> #define ReLU _rx4e6ed916a2098ff9787dc3ff4af05bf82970_ReLU_TcaU
#> #define riweibull _rx4e6ed916a2098ff9787dc3ff4af05bf82971_riweibull_EgpY
#> #define riunif _rx4e6ed916a2098ff9787dc3ff4af05bf82972_riunif_kNIq
#> #define rit_ _rx4e6ed916a2098ff9787dc3ff4af05bf82973_rit__nEnm
#> #define ripois _rx4e6ed916a2098ff9787dc3ff4af05bf82974_ripois_hSbo
#> #define ribeta _rx4e6ed916a2098ff9787dc3ff4af05bf82975_ribeta_MZYB
#> #define rigamma _rx4e6ed916a2098ff9787dc3ff4af05bf82976_rigamma_jkiC
#> #define rigeom _rx4e6ed916a2098ff9787dc3ff4af05bf82977_rigeom_oUqv
#> #define rif _rx4e6ed916a2098ff9787dc3ff4af05bf82978_rif_niI7
#> #define riexp _rx4e6ed916a2098ff9787dc3ff4af05bf82979_riexp_5Ui0
#> #define richisq _rx4e6ed916a2098ff9787dc3ff4af05bf82980_richisq_GVTs
#> #define ricauchy _rx4e6ed916a2098ff9787dc3ff4af05bf82981_ricauchy_34TN
#> #define rinbinomMu _rx4e6ed916a2098ff9787dc3ff4af05bf82982_rinbinomMu_HeFp
#> #define rinbinom _rx4e6ed916a2098ff9787dc3ff4af05bf82983_rinbinom_ue1Z
#> #define ribinom _rx4e6ed916a2098ff9787dc3ff4af05bf82984_ribinom_aEnJ
#> #define rinorm _rx4e6ed916a2098ff9787dc3ff4af05bf82985_rinorm_R6UH
#> #define rxweibull _rx4e6ed916a2098ff9787dc3ff4af05bf82986_rxweibull_ZKbS
#> #define rxunif _rx4e6ed916a2098ff9787dc3ff4af05bf82987_rxunif_smNY
#> #define rxt_ _rx4e6ed916a2098ff9787dc3ff4af05bf82988_rxt__lS2w
#> #define rxpois _rx4e6ed916a2098ff9787dc3ff4af05bf82989_rxpois_zYqw
#> #define rxbeta _rx4e6ed916a2098ff9787dc3ff4af05bf82990_rxbeta_2z8b
#> #define rxgamma _rx4e6ed916a2098ff9787dc3ff4af05bf82991_rxgamma_GMdj
#> #define rxgeom _rx4e6ed916a2098ff9787dc3ff4af05bf82992_rxgeom_LB3L
#> #define rxf _rx4e6ed916a2098ff9787dc3ff4af05bf82993_rxf_sOUR
#> #define rxexp _rx4e6ed916a2098ff9787dc3ff4af05bf82994_rxexp_bPYU
#> #define rxchisq _rx4e6ed916a2098ff9787dc3ff4af05bf82995_rxchisq_PmW8
#> #define rxcauchy _rx4e6ed916a2098ff9787dc3ff4af05bf82996_rxcauchy_sOmG
#> #define rxnbinomMu _rx4e6ed916a2098ff9787dc3ff4af05bf82997_rxnbinomMu_ypO5
#> #define rxnbinom _rx4e6ed916a2098ff9787dc3ff4af05bf82998_rxnbinom_5OnO
#> #define rxbinom _rx4e6ed916a2098ff9787dc3ff4af05bf82999_rxbinom_E9TF
#> #define rxnorm _rx4e6ed916a2098ff9787dc3ff4af05bf83000_rxnorm_yOK7
#> #define linCmtB _rx4e6ed916a2098ff9787dc3ff4af05bf83001_linCmtB_OqEK
#> #define linCmtA _rx4e6ed916a2098ff9787dc3ff4af05bf83002_linCmtA_wH4T
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){  if (_solveData->svar[_svari] == 0) {x = _PP[0];}; }
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){  if (_solveData->ovar[_ovari] == 0) {x = _PP[0];}; }
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  rx_6ef15e01012fd405b667863983221d4c__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *rx_6ef15e01012fd405b667863983221d4c__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP rx_6ef15e01012fd405b667863983221d4c__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void rx_6ef15e01012fd405b667863983221d4c__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double x = NA_REAL;
#>   double fx = NA_REAL;
#>   double dRate = NA_REAL;
#> 
#>   (void)t;
#>   (void)x;
#>   (void)fx;
#>   (void)dRate;
#> 
#>   fx = _PL[0];
#>   dRate = _PL[1];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#>   x = _PP[0];
#> 
#> 
#>   fx =llikExp(x,t);
#>   dRate =llikExpDrate(x,t);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void rx_6ef15e01012fd405b667863983221d4c__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void rx_6ef15e01012fd405b667863983221d4c__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void rx_6ef15e01012fd405b667863983221d4c__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double x = NA_REAL;
#>   double fx = NA_REAL;
#>   double dRate = NA_REAL;
#> 
#>   (void)t;
#>   (void)x;
#>   (void)fx;
#>   (void)dRate;
#> 
#>   fx = _PL[0];
#>   dRate = _PL[1];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#>   x = _PP[0];
#> 
#> 
#>   fx =llikExp(x,t);
#>   dRate =llikExpDrate(x,t);
#> 
#>   _lhs[_LHS_0_]=fx;
#>   _lhs[_LHS_1_]=dRate;
#> }
#> // Functional based bioavailability
#> double rx_6ef15e01012fd405b667863983221d4c__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double rx_6ef15e01012fd405b667863983221d4c__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double rx_6ef15e01012fd405b667863983221d4c__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double rx_6ef15e01012fd405b667863983221d4c__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void rx_6ef15e01012fd405b667863983221d4c__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void rx_6ef15e01012fd405b667863983221d4c__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void rx_6ef15e01012fd405b667863983221d4c__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP rx_6ef15e01012fd405b667863983221d4c__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_6ef15e01012fd405b667863983221d4c__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAAAAbAAA+:g6JiCU_8GA)mZ6!(S$5A11<i}WDX7tN[5}EA+4WnYTziTYa@1uv,Dz|MM;\?b}D|Y,7u|i+R=h>1;iWc>Q:kWl$bf;Z<Cw(IAhAG8Kp_|hKK,2i<a_A7:L3w_C8S~/]m.5Z0ov6tx@8]<n.}4n}p~Mma3xk:,,[m3e.=nZ_n$Vw2<~+.sA&MY1.C}Sjp;vC(^^CSz(Pr=P;io#L{5wf\"d\"t@jEv`)3aGv<SV8R7R7h6`|=!|IU&/FfnVu>[Qp%%*xC\"E3Xu]g_TXgzqY5z\"h<}):8cNRP[D8=nbQvHKOx>08Z#k1w7h(^l3Oq3UYuvs9hO=F!{1TEZ+S7S0HkX<E%%F*t@$ESd8)w8=~Y#/bn^0Bt=vaW:2Ys^$(#gC:cGOu+R]V#W#|RVXa7WuwcDZbF220`%%*xA48*MMMjS;U_|E]er\?5T)&z$YYQcS,>q#gX^CMvA%%#`$]\"52p[FU^)4.sFyw~%%5d#Pq0e/^Y6Val/}mi5IF<$Em4\"v#5\?QIre@JF}voN:oab8Dp{PbR`Qy2uNOskt|8Qs74Q3B/@N8~rsrOR9@yz)Z}0i;Hh\?ex&3*GnVhiwJ}.o[Xu\",YZL_:LsQkUtK#,]QeRbIV`/j[vHn#kke;tN9,_>s5e[o7`N[PFxojJY>F1U2o39khz:J/.%%>{7)0!o|w`sVl8]i|VFb8cV^\?ivKi5brHFW{k_../n}QyYU+v8uq`Q#\"z[RK#I9ldBG:7wX2%%HaMz<F+Wd58I9dU(~((~agG*Ki9xM,MwlE*Mvz/B{/ZtdcnOm<cD!G@vXGCM3L7FwDuZjz1ZOi#_Kam:8L\?Bv(!_U,=O}sC}pBRTuaq69vF$/*$rW>g~PD,mXtTSTrJ][\"Y@M\"Y\?xc0oba+y^q))WWRhn/DiYQr*z_5SoiPhHi#CCeL}\"w8R(A{hVeN|8]et\"kimC;|jtaa4Xx1hn;jNVadf)(CV^Uj;}VF};inY#/zo^X0&Hqze2TqyC+)c0u_NqXk>(4h]kfS/v{8>;XZQPrQ\"b}s|iucLE}(|_5JJ:i(W=4[t,gPD[^+c:.6VilDyvcQsGR}a*S/K;7B");
#>     char buf[1103];
#> #define __doBufN__ 1103
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void rx_6ef15e01012fd405b667863983221d4c__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   rx_6ef15e01012fd405b667863983221d4c__dydt(neq, *t, A, DADT);
#> }
#> extern int rx_6ef15e01012fd405b667863983221d4c__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   rx_6ef15e01012fd405b667863983221d4c__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void rx_6ef15e01012fd405b667863983221d4c__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   rx_6ef15e01012fd405b667863983221d4c__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void rx_6ef15e01012fd405b667863983221d4c__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_rx_6ef15e01012fd405b667863983221d4c_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("rx_6ef15e01012fd405b667863983221d4c_","rx_6ef15e01012fd405b667863983221d4c__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("rx_6ef15e01012fd405b667863983221d4c_","rx_6ef15e01012fd405b667863983221d4c__assignFuns", (DL_FUNC) rx_6ef15e01012fd405b667863983221d4c__assignFuns);
#>   R_RegisterCCallable("rx_6ef15e01012fd405b667863983221d4c_","rx_6ef15e01012fd405b667863983221d4c__inis",(DL_FUNC) rx_6ef15e01012fd405b667863983221d4c__inis);
#>   R_RegisterCCallable("rx_6ef15e01012fd405b667863983221d4c_","rx_6ef15e01012fd405b667863983221d4c__dydt",(DL_FUNC) rx_6ef15e01012fd405b667863983221d4c__dydt);
#>   R_RegisterCCallable("rx_6ef15e01012fd405b667863983221d4c_","rx_6ef15e01012fd405b667863983221d4c__calc_lhs",(DL_FUNC) rx_6ef15e01012fd405b667863983221d4c__calc_lhs);
#>   R_RegisterCCallable("rx_6ef15e01012fd405b667863983221d4c_","rx_6ef15e01012fd405b667863983221d4c__calc_jac",(DL_FUNC) rx_6ef15e01012fd405b667863983221d4c__calc_jac);
#>   R_RegisterCCallable("rx_6ef15e01012fd405b667863983221d4c_","rx_6ef15e01012fd405b667863983221d4c__dydt_lsoda", (DL_FUNC) rx_6ef15e01012fd405b667863983221d4c__dydt_lsoda);
#>   R_RegisterCCallable("rx_6ef15e01012fd405b667863983221d4c_","rx_6ef15e01012fd405b667863983221d4c__calc_jac_lsoda", (DL_FUNC) rx_6ef15e01012fd405b667863983221d4c__calc_jac_lsoda);
#>   R_RegisterCCallable("rx_6ef15e01012fd405b667863983221d4c_","rx_6ef15e01012fd405b667863983221d4c__ode_solver_solvedata", (DL_FUNC) rx_6ef15e01012fd405b667863983221d4c__ode_solver_solvedata);
#>   R_RegisterCCallable("rx_6ef15e01012fd405b667863983221d4c_","rx_6ef15e01012fd405b667863983221d4c__ode_solver_get_solvedata", (DL_FUNC) rx_6ef15e01012fd405b667863983221d4c__ode_solver_get_solvedata);
#>   R_RegisterCCallable("rx_6ef15e01012fd405b667863983221d4c_","rx_6ef15e01012fd405b667863983221d4c__F", (DL_FUNC) rx_6ef15e01012fd405b667863983221d4c__F);
#>   R_RegisterCCallable("rx_6ef15e01012fd405b667863983221d4c_","rx_6ef15e01012fd405b667863983221d4c__Lag", (DL_FUNC) rx_6ef15e01012fd405b667863983221d4c__Lag);
#>   R_RegisterCCallable("rx_6ef15e01012fd405b667863983221d4c_","rx_6ef15e01012fd405b667863983221d4c__Rate", (DL_FUNC) rx_6ef15e01012fd405b667863983221d4c__Rate);
#>   R_RegisterCCallable("rx_6ef15e01012fd405b667863983221d4c_","rx_6ef15e01012fd405b667863983221d4c__Dur", (DL_FUNC) rx_6ef15e01012fd405b667863983221d4c__Dur);
#>   R_RegisterCCallable("rx_6ef15e01012fd405b667863983221d4c_","rx_6ef15e01012fd405b667863983221d4c__mtime", (DL_FUNC) rx_6ef15e01012fd405b667863983221d4c__mtime);
#>   R_RegisterCCallable("rx_6ef15e01012fd405b667863983221d4c_","rx_6ef15e01012fd405b667863983221d4c__ME", (DL_FUNC) rx_6ef15e01012fd405b667863983221d4c__ME);
#>   R_RegisterCCallable("rx_6ef15e01012fd405b667863983221d4c_","rx_6ef15e01012fd405b667863983221d4c__IndF", (DL_FUNC) rx_6ef15e01012fd405b667863983221d4c__IndF);
#>   R_RegisterCCallable("rx_6ef15e01012fd405b667863983221d4c_","rx_6ef15e01012fd405b667863983221d4c__dydt_liblsoda", (DL_FUNC) rx_6ef15e01012fd405b667863983221d4c__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_rx_6ef15e01012fd405b667863983221d4c_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_rx_6ef15e01012fd405b667863983221d4c_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"rx_6ef15e01012fd405b667863983221d4c__model_vars", (DL_FUNC) &rx_6ef15e01012fd405b667863983221d4c__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_rx_6ef15e01012fd405b667863983221d4c_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_6ef15e01012fd405b667863983221d4c__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("rx_6ef15e01012fd405b667863983221d4c__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’
#> ── Solved rxode2 object ──
#> ── Parameters (value$params): ──
#> # A tibble: 1 × 0
#> ── Initial Conditions (value$inits): ──
#> named numeric(0)
#> ── First part of data (object): ──
#> # A tibble: 3 × 4
#>    time    fx  dRate     x
#>   <dbl> <dbl>  <dbl> <dbl>
#> 1     1 -1     0         1
#> 2     2 -1.31 -0.5       1
#> 3     3 -1.90 -0.667     1
# }