log likelihood and derivatives for F distribution

log likelihood and derivatives for F distribution

Usage

llikF(x, df1, df2, full = FALSE)

Arguments

x

variable that is distributed by f distribution

df1, df2

degrees of freedom. Inf is allowed.

full

Add the data frame showing x, mean, sd as well as the fx and derivatives

Value

data frame with fx for the log pdf value of with dDf1 and dDf2 that has the derivatives with respect to the df1/df2 parameters at the observation time-point

Details

In an rxode2() model, you can use llikF() but you have to use the x and rate arguments. You can also get the derivative of df1 and df2 with llikFDdf1() and llikFDdf2().

Author

Matthew L. Fidler

Examples

# \donttest{
x <- seq(0.001, 5, length.out = 100)

llikF(x^2, 1, 5)
#>               fx         dDf1          dDf2
#> 1    5.939135090 -5.769327755  0.0098138672
#> 2    1.996061339 -1.829698617  0.0098667834
#> 3    1.308027230 -1.151720596  0.0100203751
#> 4    0.898151814 -0.758467992  0.0102713542
#> 5    0.601410048 -0.484842238  0.0106143384
#> 6    0.365567397 -0.278507461  0.0110419510
#> 7    0.167275143 -0.115987118  0.0115449567
#> 8   -0.005928407  0.015333168  0.0121124287
#> 9   -0.161467803  0.123053577  0.0127319420
#> 10  -0.304110944  0.212138450  0.0133897879
#> 11  -0.437089645  0.286034532  0.0140712037
#> 12  -0.562684604  0.347253504  0.0147606108
#> 13  -0.682555520  0.397699568  0.0154418570
#> 14  -0.797939098  0.438864917  0.0160984557
#> 15  -0.909773848  0.471951996  0.0167138175
#> 16  -1.018782007  0.497952966  0.0172714700
#> 17  -1.125525217  0.517703019  0.0177552614
#> 18  -1.230443500  0.531917117  0.0181495454
#> 19  -1.333883243  0.541215913  0.0184393449
#> 20  -1.436117763  0.546144408  0.0186104926
#> 21  -1.537362702  0.547185650  0.0186497493
#> 22  -1.637787764  0.544770962  0.0185448973
#> 23  -1.737525805  0.539287723  0.0182848126
#> 24  -1.836679954  0.531085385  0.0178595139
#> 25  -1.935329266  0.520480216  0.0172601925
#> 26  -2.033533263  0.507759100  0.0164792220
#> 27  -2.131335592  0.493182655  0.0155101530
#> 28  -2.228767013  0.476987816  0.0143476919
#> 29  -2.325847849  0.459390036  0.0129876676
#> 30  -2.422589998  0.440585182  0.0114269877
#> 31  -2.518998603  0.420751193  0.0096635860
#> 32  -2.615073425  0.400049565  0.0076963639
#> 33  -2.710810000  0.378626673  0.0055251263
#> 34  -2.806200581  0.356614982  0.0031505147
#> 35  -2.901234940  0.334134152  0.0005739382
#> 36  -2.995901017  0.311292055 -0.0022024960
#> 37  -3.090185467  0.288185720 -0.0051760516
#> 38  -3.184074111  0.264902202 -0.0083434311
#> 39  -3.277552304  0.241519400 -0.0117008417
#> 40  -3.370605236  0.218106810 -0.0152440579
#> 41  -3.463218177  0.194726229 -0.0189684823
#> 42  -3.555376677  0.171432411 -0.0228692014
#> 43  -3.647066718  0.148273673 -0.0269410395
#> 44  -3.738274839  0.125292465 -0.0311786073
#> 45  -3.828988227  0.102525891 -0.0355763478
#> 46  -3.919194789  0.080006197 -0.0401285781
#> 47  -4.008883199  0.057761220 -0.0448295271
#> 48  -4.098042931  0.035814802 -0.0496733702
#> 49  -4.186664276  0.014187176 -0.0546542601
#> 50  -4.274738350 -0.007104684 -0.0597663547
#> 51  -4.362257091 -0.028046739 -0.0650038410
#> 52  -4.449213248 -0.048627585 -0.0703609577
#> 53  -4.535600366 -0.068838184 -0.0758320130
#> 54  -4.621412763 -0.088671614 -0.0814114015
#> 55  -4.706645505 -0.108122840 -0.0870936184
#> 56  -4.791294380 -0.127188508 -0.0928732707
#> 57  -4.875355868 -0.145866755 -0.0987450876
#> 58  -4.958827105 -0.164157039 -0.1047039284
#> 59  -5.041705859 -0.182059980 -0.1107447891
#> 60  -5.123990491 -0.199577223 -0.1168628074
#> 61  -5.205679923 -0.216711307 -0.1230532668
#> 62  -5.286773611 -0.233465550 -0.1293115988
#> 63  -5.367271506 -0.249843943 -0.1356333846
#> 64  -5.447174028 -0.265851061 -0.1420143563
#> 65  -5.526482033 -0.281491973 -0.1484503963
#> 66  -5.605196787 -0.296772169 -0.1549375368
#> 67  -5.683319932 -0.311697491 -0.1614719584
#> 68  -5.760853464 -0.326274073 -0.1680499885
#> 69  -5.837799706 -0.340508283 -0.1746680988
#> 70  -5.914161282 -0.354406680 -0.1813229026
#> 71  -5.989941093 -0.367975967 -0.1880111521
#> 72  -6.065142296 -0.381222952 -0.1947297352
#> 73  -6.139768282 -0.394154517 -0.2014756717
#> 74  -6.213822655 -0.406777586 -0.2082461105
#> 75  -6.287309215 -0.419099097 -0.2150383254
#> 76  -6.360231938 -0.431125983 -0.2218497115
#> 77  -6.432594959 -0.442865149 -0.2286777813
#> 78  -6.504402560 -0.454323453 -0.2355201612
#> 79  -6.575659147 -0.465507693 -0.2423745875
#> 80  -6.646369247 -0.476424593 -0.2492389023
#> 81  -6.716537483 -0.487080791 -0.2561110501
#> 82  -6.786168572 -0.497482830 -0.2629890742
#> 83  -6.855267306 -0.507637151 -0.2698711122
#> 84  -6.923838547 -0.517550083 -0.2767553934
#> 85  -6.991887212 -0.527227840 -0.2836402345
#> 86  -7.059418267 -0.536676516 -0.2905240366
#> 87  -7.126436719 -0.545902081 -0.2974052813
#> 88  -7.192947604 -0.554910380 -0.3042825282
#> 89  -7.258955983 -0.563707126 -0.3111544109
#> 90  -7.324466935 -0.572297907 -0.3180196342
#> 91  -7.389485548 -0.580688177 -0.3248769714
#> 92  -7.454016916 -0.588883261 -0.3317252608
#> 93  -7.518066132 -0.596888352 -0.3385634034
#> 94  -7.581638284 -0.604708517 -0.3453903600
#> 95  -7.644738447 -0.612348689 -0.3522051484
#> 96  -7.707371685 -0.619813679 -0.3590068413
#> 97  -7.769543041 -0.627108167 -0.3657945633
#> 98  -7.831257537 -0.634236713 -0.3725674892
#> 99  -7.892520171 -0.641203751 -0.3793248413
#> 100 -7.953335909 -0.648013598 -0.3860658874

model <- function(){
  model({
    fx <- llikF(time, df1, df2)
    dMean <- llikFDdf1(time, df1, df2)
    dSd <- llikFDdf2(time, df1, df2)
  })
}

et <- et(x)
et$df1 <- 1
et$df2 <- 5

rxSolve(model, et)
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#>  
#>  
#> C model:
#> // Define 3 LHS values
#> #define _LHS_0_ 0 // fx
#> #define _LHS_1_ 1 // dMean
#> #define _LHS_2_ 2 // dSd
#> #define _getRxSolve_ _rxb6b79cd890fa51a1c3b7172341ee26793003__getRxSolve__HF4t
#> #define _evalUdf _rxb6b79cd890fa51a1c3b7172341ee26793004__evalUdf_E5bq
#> #define _solveData _rxb6b79cd890fa51a1c3b7172341ee26793005__solveData_rWj2
#> #define _assign_ptr _rxb6b79cd890fa51a1c3b7172341ee26793006__assign_ptr_jIRW
#> #define _rxRmModelLib _rxb6b79cd890fa51a1c3b7172341ee26793007__rxRmModelLib_0crr
#> #define _rxGetModelLib _rxb6b79cd890fa51a1c3b7172341ee26793008__rxGetModelLib_gcGm
#> #define _old_c _rxb6b79cd890fa51a1c3b7172341ee26793009__old_c_EsJu
#> #define _ptrid _rxb6b79cd890fa51a1c3b7172341ee26793010__ptrid_9xah
#> #define _rxIsCurrentC _rxb6b79cd890fa51a1c3b7172341ee26793011__rxIsCurrentC_gBOZ
#> #define _sumPS _rxb6b79cd890fa51a1c3b7172341ee26793012__sumPS_l3zY
#> #define _prodPS _rxb6b79cd890fa51a1c3b7172341ee26793013__prodPS_vTa9
#> #define _prodType _rxb6b79cd890fa51a1c3b7172341ee26793014__prodType_7EhW
#> #define _sumType _rxb6b79cd890fa51a1c3b7172341ee26793015__sumType_3MG9
#> #define _update_par_ptr _rxb6b79cd890fa51a1c3b7172341ee26793016__update_par_ptr_V3pR
#> #define _getParCov _rxb6b79cd890fa51a1c3b7172341ee26793017__getParCov_1aNk
#> #define _rxode2_rxAssignPtr _rxb6b79cd890fa51a1c3b7172341ee26793018__rxode2_rxAssignPtr_G5Jw
#> #define _rxQr _rxb6b79cd890fa51a1c3b7172341ee26793019__rxQr_D9TK
#> #define _compareFactorVal _rxb6b79cd890fa51a1c3b7172341ee26793020__compareFactorVal_40NP
#> #define _sum _rxb6b79cd890fa51a1c3b7172341ee26793021__sum_JiAL
#> #define _udf _rxb6b79cd890fa51a1c3b7172341ee26793022__udf_ZBrh
#> #define _sign _rxb6b79cd890fa51a1c3b7172341ee26793023__sign_efgh
#> #define _prod _rxb6b79cd890fa51a1c3b7172341ee26793024__prod_Drm8
#> #define _max _rxb6b79cd890fa51a1c3b7172341ee26793025__max_5igv
#> #define _min _rxb6b79cd890fa51a1c3b7172341ee26793026__min_lJYg
#> #define _transit4P _rxb6b79cd890fa51a1c3b7172341ee26793027__transit4P_GqQx
#> #define _transit3P _rxb6b79cd890fa51a1c3b7172341ee26793028__transit3P_LoYl
#> #define _assignFuns0 _rxb6b79cd890fa51a1c3b7172341ee26793029__assignFuns0_FgJd
#> #define _assignFuns _rxb6b79cd890fa51a1c3b7172341ee26793030__assignFuns_KaFg
#> #define _rxord _rxb6b79cd890fa51a1c3b7172341ee26793031__rxord_mDR4
#> #define __assignFuns2 _rxb6b79cd890fa51a1c3b7172341ee26793032___assignFuns2_jxuo
#> #define _llikCauchyDscale _rxb6b79cd890fa51a1c3b7172341ee26793033__llikCauchyDscale_P8gm
#> #define _llikCauchyDlocation _rxb6b79cd890fa51a1c3b7172341ee26793034__llikCauchyDlocation_oUom
#> #define _llikCauchy _rxb6b79cd890fa51a1c3b7172341ee26793035__llikCauchy_wLtu
#> #define _llikGammaDrate _rxb6b79cd890fa51a1c3b7172341ee26793036__llikGammaDrate_DgOz
#> #define _llikGammaDshape _rxb6b79cd890fa51a1c3b7172341ee26793037__llikGammaDshape_tPZw
#> #define _llikGamma _rxb6b79cd890fa51a1c3b7172341ee26793038__llikGamma_Dqx2
#> #define _llikWeibullDscale _rxb6b79cd890fa51a1c3b7172341ee26793039__llikWeibullDscale_t0Ff
#> #define _llikWeibullDshape _rxb6b79cd890fa51a1c3b7172341ee26793040__llikWeibullDshape_CRup
#> #define _llikWeibull _rxb6b79cd890fa51a1c3b7172341ee26793041__llikWeibull_BfcL
#> #define _llikUnifDbeta _rxb6b79cd890fa51a1c3b7172341ee26793042__llikUnifDbeta_LlJk
#> #define _llikUnifDalpha _rxb6b79cd890fa51a1c3b7172341ee26793043__llikUnifDalpha_0HOu
#> #define _llikUnif _rxb6b79cd890fa51a1c3b7172341ee26793044__llikUnif_quMQ
#> #define _llikGeomDp _rxb6b79cd890fa51a1c3b7172341ee26793045__llikGeomDp_IP4s
#> #define _llikGeom _rxb6b79cd890fa51a1c3b7172341ee26793046__llikGeom_BKwl
#> #define _llikFDdf2 _rxb6b79cd890fa51a1c3b7172341ee26793047__llikFDdf2_ION1
#> #define _llikFDdf1 _rxb6b79cd890fa51a1c3b7172341ee26793048__llikFDdf1_tTX5
#> #define _llikF _rxb6b79cd890fa51a1c3b7172341ee26793049__llikF_DulH
#> #define _llikExpDrate _rxb6b79cd890fa51a1c3b7172341ee26793050__llikExpDrate_LxeA
#> #define _llikExp _rxb6b79cd890fa51a1c3b7172341ee26793051__llikExp_Ut0p
#> #define _llikChisqDdf _rxb6b79cd890fa51a1c3b7172341ee26793052__llikChisqDdf_o6Aq
#> #define _llikChisq _rxb6b79cd890fa51a1c3b7172341ee26793053__llikChisq_SQkn
#> #define _llikTDsd _rxb6b79cd890fa51a1c3b7172341ee26793054__llikTDsd_H8fP
#> #define _llikTDmean _rxb6b79cd890fa51a1c3b7172341ee26793055__llikTDmean_CAnN
#> #define _llikTDdf _rxb6b79cd890fa51a1c3b7172341ee26793056__llikTDdf_kUms
#> #define _llikT _rxb6b79cd890fa51a1c3b7172341ee26793057__llikT_wiOW
#> #define _llikBetaDshape2 _rxb6b79cd890fa51a1c3b7172341ee26793058__llikBetaDshape2_G16H
#> #define _llikBetaDshape1 _rxb6b79cd890fa51a1c3b7172341ee26793059__llikBetaDshape1_PUgn
#> #define _llikBeta _rxb6b79cd890fa51a1c3b7172341ee26793060__llikBeta_Cjuw
#> #define _llikNbinomMuDmu _rxb6b79cd890fa51a1c3b7172341ee26793061__llikNbinomMuDmu_A8iw
#> #define _llikNbinomMu _rxb6b79cd890fa51a1c3b7172341ee26793062__llikNbinomMu_bn2G
#> #define _llikNbinomDprob _rxb6b79cd890fa51a1c3b7172341ee26793063__llikNbinomDprob_TQjS
#> #define _llikNbinom _rxb6b79cd890fa51a1c3b7172341ee26793064__llikNbinom_rer5
#> #define _llikBinomDprob _rxb6b79cd890fa51a1c3b7172341ee26793065__llikBinomDprob_zexL
#> #define _llikBinom _rxb6b79cd890fa51a1c3b7172341ee26793066__llikBinom_gAwM
#> #define _llikPoisDlambda _rxb6b79cd890fa51a1c3b7172341ee26793067__llikPoisDlambda_E6nq
#> #define _llikPois _rxb6b79cd890fa51a1c3b7172341ee26793068__llikPois_R3DD
#> #define _llikNormDsd _rxb6b79cd890fa51a1c3b7172341ee26793069__llikNormDsd_9jN5
#> #define _llikNormDmean _rxb6b79cd890fa51a1c3b7172341ee26793070__llikNormDmean_O4r9
#> #define _llikNorm _rxb6b79cd890fa51a1c3b7172341ee26793071__llikNorm_D97m
#> #define simeps _rxb6b79cd890fa51a1c3b7172341ee26793072_simeps_e5yb
#> #define simeta _rxb6b79cd890fa51a1c3b7172341ee26793073_simeta_V8Eu
#> #define expit _rxb6b79cd890fa51a1c3b7172341ee26793074_expit_Z1xT
#> #define logit _rxb6b79cd890fa51a1c3b7172341ee26793075_logit_QYOQ
#> #define gammapDer _rxb6b79cd890fa51a1c3b7172341ee26793076_gammapDer_e6aq
#> #define lowergamma _rxb6b79cd890fa51a1c3b7172341ee26793077_lowergamma_t3Lb
#> #define uppergamma _rxb6b79cd890fa51a1c3b7172341ee26793078_uppergamma_sbea
#> #define gammaqInva _rxb6b79cd890fa51a1c3b7172341ee26793079_gammaqInva_w5m4
#> #define gammaqInv _rxb6b79cd890fa51a1c3b7172341ee26793080_gammaqInv_brD9
#> #define gammapInva _rxb6b79cd890fa51a1c3b7172341ee26793081_gammapInva_EZAJ
#> #define gammapInv _rxb6b79cd890fa51a1c3b7172341ee26793082_gammapInv_Ne7I
#> #define gammaq _rxb6b79cd890fa51a1c3b7172341ee26793083_gammaq_xUFA
#> #define gammap _rxb6b79cd890fa51a1c3b7172341ee26793084_gammap_5R2w
#> #define phi _rxb6b79cd890fa51a1c3b7172341ee26793085_phi_ZwwB
#> #define d2ELUa _rxb6b79cd890fa51a1c3b7172341ee26793086_d2ELUa_gRvb
#> #define dELUa _rxb6b79cd890fa51a1c3b7172341ee26793087_dELUa_lnk3
#> #define d2aELU _rxb6b79cd890fa51a1c3b7172341ee26793088_d2aELU_57cf
#> #define d2ELU _rxb6b79cd890fa51a1c3b7172341ee26793089_d2ELU_UwMa
#> #define dELU _rxb6b79cd890fa51a1c3b7172341ee26793090_dELU_rEA9
#> #define ELU _rxb6b79cd890fa51a1c3b7172341ee26793091_ELU_z1SE
#> #define dPReLUa1 _rxb6b79cd890fa51a1c3b7172341ee26793092_dPReLUa1_1KzA
#> #define dPReLUa _rxb6b79cd890fa51a1c3b7172341ee26793093_dPReLUa_f2H4
#> #define dPReLU _rxb6b79cd890fa51a1c3b7172341ee26793094_dPReLU_Q2oF
#> #define PReLU _rxb6b79cd890fa51a1c3b7172341ee26793095_PReLU_0joC
#> #define dSwish _rxb6b79cd890fa51a1c3b7172341ee26793096_dSwish_Gj0a
#> #define Swish _rxb6b79cd890fa51a1c3b7172341ee26793097_Swish_8wAb
#> #define dlReLU _rxb6b79cd890fa51a1c3b7172341ee26793098_dlReLU_QELO
#> #define lReLU _rxb6b79cd890fa51a1c3b7172341ee26793099_lReLU_lObq
#> #define dSELU _rxb6b79cd890fa51a1c3b7172341ee26793100_dSELU_Nath
#> #define SELU _rxb6b79cd890fa51a1c3b7172341ee26793101_SELU_ltiW
#> #define d4softplus _rxb6b79cd890fa51a1c3b7172341ee26793102_d4softplus_R2IG
#> #define d3softplus _rxb6b79cd890fa51a1c3b7172341ee26793103_d3softplus_VLQY
#> #define d2softplus _rxb6b79cd890fa51a1c3b7172341ee26793104_d2softplus_M84J
#> #define dsoftplus _rxb6b79cd890fa51a1c3b7172341ee26793105_dsoftplus_Cnuy
#> #define softplus _rxb6b79cd890fa51a1c3b7172341ee26793106_softplus_CXPJ
#> #define d4GELU _rxb6b79cd890fa51a1c3b7172341ee26793107_d4GELU_UjvF
#> #define d3GELU _rxb6b79cd890fa51a1c3b7172341ee26793108_d3GELU_k7KJ
#> #define d2GELU _rxb6b79cd890fa51a1c3b7172341ee26793109_d2GELU_vDxq
#> #define dGELU _rxb6b79cd890fa51a1c3b7172341ee26793110_dGELU_bVSK
#> #define GELU _rxb6b79cd890fa51a1c3b7172341ee26793111_GELU_Vfr7
#> #define dReLU _rxb6b79cd890fa51a1c3b7172341ee26793112_dReLU_TcaU
#> #define ReLU _rxb6b79cd890fa51a1c3b7172341ee26793113_ReLU_EgpY
#> #define riweibull _rxb6b79cd890fa51a1c3b7172341ee26793114_riweibull_kNIq
#> #define riunif _rxb6b79cd890fa51a1c3b7172341ee26793115_riunif_nEnm
#> #define rit_ _rxb6b79cd890fa51a1c3b7172341ee26793116_rit__hSbo
#> #define ripois _rxb6b79cd890fa51a1c3b7172341ee26793117_ripois_MZYB
#> #define ribeta _rxb6b79cd890fa51a1c3b7172341ee26793118_ribeta_jkiC
#> #define rigamma _rxb6b79cd890fa51a1c3b7172341ee26793119_rigamma_oUqv
#> #define rigeom _rxb6b79cd890fa51a1c3b7172341ee26793120_rigeom_niI7
#> #define rif _rxb6b79cd890fa51a1c3b7172341ee26793121_rif_5Ui0
#> #define riexp _rxb6b79cd890fa51a1c3b7172341ee26793122_riexp_GVTs
#> #define richisq _rxb6b79cd890fa51a1c3b7172341ee26793123_richisq_34TN
#> #define ricauchy _rxb6b79cd890fa51a1c3b7172341ee26793124_ricauchy_HeFp
#> #define rinbinomMu _rxb6b79cd890fa51a1c3b7172341ee26793125_rinbinomMu_ue1Z
#> #define rinbinom _rxb6b79cd890fa51a1c3b7172341ee26793126_rinbinom_aEnJ
#> #define ribinom _rxb6b79cd890fa51a1c3b7172341ee26793127_ribinom_R6UH
#> #define rinorm _rxb6b79cd890fa51a1c3b7172341ee26793128_rinorm_ZKbS
#> #define rxweibull _rxb6b79cd890fa51a1c3b7172341ee26793129_rxweibull_smNY
#> #define rxunif _rxb6b79cd890fa51a1c3b7172341ee26793130_rxunif_lS2w
#> #define rxt_ _rxb6b79cd890fa51a1c3b7172341ee26793131_rxt__zYqw
#> #define rxpois _rxb6b79cd890fa51a1c3b7172341ee26793132_rxpois_2z8b
#> #define rxbeta _rxb6b79cd890fa51a1c3b7172341ee26793133_rxbeta_GMdj
#> #define rxgamma _rxb6b79cd890fa51a1c3b7172341ee26793134_rxgamma_LB3L
#> #define rxgeom _rxb6b79cd890fa51a1c3b7172341ee26793135_rxgeom_sOUR
#> #define rxf _rxb6b79cd890fa51a1c3b7172341ee26793136_rxf_bPYU
#> #define rxexp _rxb6b79cd890fa51a1c3b7172341ee26793137_rxexp_PmW8
#> #define rxchisq _rxb6b79cd890fa51a1c3b7172341ee26793138_rxchisq_sOmG
#> #define rxcauchy _rxb6b79cd890fa51a1c3b7172341ee26793139_rxcauchy_ypO5
#> #define rxnbinomMu _rxb6b79cd890fa51a1c3b7172341ee26793140_rxnbinomMu_5OnO
#> #define rxnbinom _rxb6b79cd890fa51a1c3b7172341ee26793141_rxnbinom_E9TF
#> #define rxbinom _rxb6b79cd890fa51a1c3b7172341ee26793142_rxbinom_yOK7
#> #define rxnorm _rxb6b79cd890fa51a1c3b7172341ee26793143_rxnorm_OqEK
#> #define linCmtB _rxb6b79cd890fa51a1c3b7172341ee26793144_linCmtB_wH4T
#> #define linCmtA _rxb6b79cd890fa51a1c3b7172341ee26793145_linCmtA_tSJj
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){  if (_solveData->svar[_svari] == 0) {df1 = _PP[0];};   if (_solveData->svar[_svari] == 1) {df2 = _PP[1];}; }
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){  if (_solveData->ovar[_ovari] == 0) {df1 = _PP[0];};   if (_solveData->ovar[_ovari] == 1) {df2 = _PP[1];}; }
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  rx_925ff828bd7406d6d9c2ce4a75c951ba__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *rx_925ff828bd7406d6d9c2ce4a75c951ba__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP rx_925ff828bd7406d6d9c2ce4a75c951ba__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void rx_925ff828bd7406d6d9c2ce4a75c951ba__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double df1 = NA_REAL;
#>   double df2 = NA_REAL;
#>   double fx = NA_REAL;
#>   double dMean = NA_REAL;
#>   double dSd = NA_REAL;
#> 
#>   (void)t;
#>   (void)df1;
#>   (void)df2;
#>   (void)fx;
#>   (void)dMean;
#>   (void)dSd;
#> 
#>   fx = _PL[0];
#>   dMean = _PL[1];
#>   dSd = _PL[2];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#>   df1 = _PP[0];
#>   df2 = _PP[1];
#> 
#> 
#>   fx =llikF(t,df1,df2);
#>   dMean =llikFDdf1(t,df1,df2);
#>   dSd =llikFDdf2(t,df1,df2);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void rx_925ff828bd7406d6d9c2ce4a75c951ba__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void rx_925ff828bd7406d6d9c2ce4a75c951ba__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void rx_925ff828bd7406d6d9c2ce4a75c951ba__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double df1 = NA_REAL;
#>   double df2 = NA_REAL;
#>   double fx = NA_REAL;
#>   double dMean = NA_REAL;
#>   double dSd = NA_REAL;
#> 
#>   (void)t;
#>   (void)df1;
#>   (void)df2;
#>   (void)fx;
#>   (void)dMean;
#>   (void)dSd;
#> 
#>   fx = _PL[0];
#>   dMean = _PL[1];
#>   dSd = _PL[2];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#>   df1 = _PP[0];
#>   df2 = _PP[1];
#> 
#> 
#>   fx =llikF(t,df1,df2);
#>   dMean =llikFDdf1(t,df1,df2);
#>   dSd =llikFDdf2(t,df1,df2);
#> 
#>   _lhs[_LHS_0_]=fx;
#>   _lhs[_LHS_1_]=dMean;
#>   _lhs[_LHS_2_]=dSd;
#> }
#> // Functional based bioavailability
#> double rx_925ff828bd7406d6d9c2ce4a75c951ba__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double rx_925ff828bd7406d6d9c2ce4a75c951ba__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double rx_925ff828bd7406d6d9c2ce4a75c951ba__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double rx_925ff828bd7406d6d9c2ce4a75c951ba__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void rx_925ff828bd7406d6d9c2ce4a75c951ba__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void rx_925ff828bd7406d6d9c2ce4a75c951ba__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void rx_925ff828bd7406d6d9c2ce4a75c951ba__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP rx_925ff828bd7406d6d9c2ce4a75c951ba__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_925ff828bd7406d6d9c2ce4a75c951ba__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAA+>2AAAv7#aT)#pvzNA+]@a=WdwYB~1u%%C8Ae68jR&AE/ieng$<(Z(xKa,uPnd0nx6H:6%%gm4hAw+ZuRsI/@#|>#$y#MHM3*a,^DH.>RA(A5jF1$G1SW!LOm,3i5~I27QmG$=9(\?SBid;E2TWcBj33s{LjF$``b$B|=l6Eb3<#>TW)_4v]4C|mM@$4xMDVYUyG\"pQ$^\"]`t+pDdfjW3pvS5beHl^V{w7=i$bXy*hicizj>1!.5a@j1ZZ[~N4~X&AO&n^aAO~/if\?(su(zKam6{/P*b2RJ02a#Hf5i`0ZMx@atf>ZsWz\"{,8TTs]YX3t:@ZQvZ,>Xo\?N@Q{X<C|:9YOwiG_(1p`\?o:86VrPFARWUw12a|Li,!iE:W|JBZt<1,X#zq8ZnkBO/eM.GkV.cDuZR8\?o2[IYLUE#yO7jFWyb5lbo(afjK#GF=u&0\"\"#|icSpSuW2/RGaxRJG=N_{3<;;[Kd#/F6m}|W$||*:hzF)7\?]D+x#u\"#2ce\"TI@uZxbI9fR(S60_~5,`Tlfd!5UBo6DlCV2tM\?Y#bVP\?Y=p%%]&;e3)h]f@htOm9;X(xZH>o1BCN`ciZs]cEHY>M`c8d.;P>(skf=G7s9xE*m>kuBq>2esE\?Y+5m~*C[=p}0j=L^t[Z#4#WKkwd:<u,c7D<llftkvb4\?xgw/:|$hWf!|K@`~9Q8ku\?4V/m)\"A$[a:tvUOl/6G<Znc{s3cPWyi)wSPL={hyU&#`ld.iZlSfP~g5rzDu^!rG;SYaNaei.c:j&jIB)ikQlReC(j~vlE*Mvz/B{/a\"1dgY/%%/A}4;Py\?^vgGhX$MSweGFMC=6Czw|&qqpEGeJ~I{S^S4OoR@ATXLcQ>e#lh@G!Pbe|arHR7B\"rm{3;1uJ4YF$EYx:%%6H6d\"}j,ck+mZU[!_qr9iw@}Uk{ebRQ$8B|WJkE%%_XA]qi_{,|~0,x<c3Z|c8bktOafGx\"IJ(~@wn43WIwvzh&_q;)TF}V|4BL|jw\?q9Ct!!LQ;%%NElL41\?J.yccO8b&[n>a@|$D]OHfDa+!gy=K}~,j#FgAY)3ffJMt[w7:Ky0:y)!zYSR{YE_z9a%%X8SFH@il8Z^vXcF");
#>     char buf[1121];
#> #define __doBufN__ 1121
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void rx_925ff828bd7406d6d9c2ce4a75c951ba__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   rx_925ff828bd7406d6d9c2ce4a75c951ba__dydt(neq, *t, A, DADT);
#> }
#> extern int rx_925ff828bd7406d6d9c2ce4a75c951ba__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   rx_925ff828bd7406d6d9c2ce4a75c951ba__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void rx_925ff828bd7406d6d9c2ce4a75c951ba__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   rx_925ff828bd7406d6d9c2ce4a75c951ba__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void rx_925ff828bd7406d6d9c2ce4a75c951ba__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_rx_925ff828bd7406d6d9c2ce4a75c951ba_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("rx_925ff828bd7406d6d9c2ce4a75c951ba_","rx_925ff828bd7406d6d9c2ce4a75c951ba__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("rx_925ff828bd7406d6d9c2ce4a75c951ba_","rx_925ff828bd7406d6d9c2ce4a75c951ba__assignFuns", (DL_FUNC) rx_925ff828bd7406d6d9c2ce4a75c951ba__assignFuns);
#>   R_RegisterCCallable("rx_925ff828bd7406d6d9c2ce4a75c951ba_","rx_925ff828bd7406d6d9c2ce4a75c951ba__inis",(DL_FUNC) rx_925ff828bd7406d6d9c2ce4a75c951ba__inis);
#>   R_RegisterCCallable("rx_925ff828bd7406d6d9c2ce4a75c951ba_","rx_925ff828bd7406d6d9c2ce4a75c951ba__dydt",(DL_FUNC) rx_925ff828bd7406d6d9c2ce4a75c951ba__dydt);
#>   R_RegisterCCallable("rx_925ff828bd7406d6d9c2ce4a75c951ba_","rx_925ff828bd7406d6d9c2ce4a75c951ba__calc_lhs",(DL_FUNC) rx_925ff828bd7406d6d9c2ce4a75c951ba__calc_lhs);
#>   R_RegisterCCallable("rx_925ff828bd7406d6d9c2ce4a75c951ba_","rx_925ff828bd7406d6d9c2ce4a75c951ba__calc_jac",(DL_FUNC) rx_925ff828bd7406d6d9c2ce4a75c951ba__calc_jac);
#>   R_RegisterCCallable("rx_925ff828bd7406d6d9c2ce4a75c951ba_","rx_925ff828bd7406d6d9c2ce4a75c951ba__dydt_lsoda", (DL_FUNC) rx_925ff828bd7406d6d9c2ce4a75c951ba__dydt_lsoda);
#>   R_RegisterCCallable("rx_925ff828bd7406d6d9c2ce4a75c951ba_","rx_925ff828bd7406d6d9c2ce4a75c951ba__calc_jac_lsoda", (DL_FUNC) rx_925ff828bd7406d6d9c2ce4a75c951ba__calc_jac_lsoda);
#>   R_RegisterCCallable("rx_925ff828bd7406d6d9c2ce4a75c951ba_","rx_925ff828bd7406d6d9c2ce4a75c951ba__ode_solver_solvedata", (DL_FUNC) rx_925ff828bd7406d6d9c2ce4a75c951ba__ode_solver_solvedata);
#>   R_RegisterCCallable("rx_925ff828bd7406d6d9c2ce4a75c951ba_","rx_925ff828bd7406d6d9c2ce4a75c951ba__ode_solver_get_solvedata", (DL_FUNC) rx_925ff828bd7406d6d9c2ce4a75c951ba__ode_solver_get_solvedata);
#>   R_RegisterCCallable("rx_925ff828bd7406d6d9c2ce4a75c951ba_","rx_925ff828bd7406d6d9c2ce4a75c951ba__F", (DL_FUNC) rx_925ff828bd7406d6d9c2ce4a75c951ba__F);
#>   R_RegisterCCallable("rx_925ff828bd7406d6d9c2ce4a75c951ba_","rx_925ff828bd7406d6d9c2ce4a75c951ba__Lag", (DL_FUNC) rx_925ff828bd7406d6d9c2ce4a75c951ba__Lag);
#>   R_RegisterCCallable("rx_925ff828bd7406d6d9c2ce4a75c951ba_","rx_925ff828bd7406d6d9c2ce4a75c951ba__Rate", (DL_FUNC) rx_925ff828bd7406d6d9c2ce4a75c951ba__Rate);
#>   R_RegisterCCallable("rx_925ff828bd7406d6d9c2ce4a75c951ba_","rx_925ff828bd7406d6d9c2ce4a75c951ba__Dur", (DL_FUNC) rx_925ff828bd7406d6d9c2ce4a75c951ba__Dur);
#>   R_RegisterCCallable("rx_925ff828bd7406d6d9c2ce4a75c951ba_","rx_925ff828bd7406d6d9c2ce4a75c951ba__mtime", (DL_FUNC) rx_925ff828bd7406d6d9c2ce4a75c951ba__mtime);
#>   R_RegisterCCallable("rx_925ff828bd7406d6d9c2ce4a75c951ba_","rx_925ff828bd7406d6d9c2ce4a75c951ba__ME", (DL_FUNC) rx_925ff828bd7406d6d9c2ce4a75c951ba__ME);
#>   R_RegisterCCallable("rx_925ff828bd7406d6d9c2ce4a75c951ba_","rx_925ff828bd7406d6d9c2ce4a75c951ba__IndF", (DL_FUNC) rx_925ff828bd7406d6d9c2ce4a75c951ba__IndF);
#>   R_RegisterCCallable("rx_925ff828bd7406d6d9c2ce4a75c951ba_","rx_925ff828bd7406d6d9c2ce4a75c951ba__dydt_liblsoda", (DL_FUNC) rx_925ff828bd7406d6d9c2ce4a75c951ba__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_rx_925ff828bd7406d6d9c2ce4a75c951ba_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_rx_925ff828bd7406d6d9c2ce4a75c951ba_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"rx_925ff828bd7406d6d9c2ce4a75c951ba__model_vars", (DL_FUNC) &rx_925ff828bd7406d6d9c2ce4a75c951ba__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_rx_925ff828bd7406d6d9c2ce4a75c951ba_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_925ff828bd7406d6d9c2ce4a75c951ba__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("rx_925ff828bd7406d6d9c2ce4a75c951ba__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’
#> ── Solved rxode2 object ──
#> ── Parameters (value$params): ──
#> # A tibble: 1 × 0
#> ── Initial Conditions (value$inits): ──
#> named numeric(0)
#> ── First part of data (object): ──
#> # A tibble: 100 × 6
#>     time     fx   dMean     dSd   df1   df2
#>    <dbl>  <dbl>   <dbl>   <dbl> <dbl> <dbl>
#> 1 0.001   2.48  -2.32   0.00983     1     5
#> 2 0.0515  0.484 -0.380  0.0108      1     5
#> 3 0.102   0.112 -0.0731 0.0117      1     5
#> 4 0.152  -0.118  0.0943 0.0125      1     5
#> 5 0.203  -0.291  0.204  0.0133      1     5
#> 6 0.253  -0.431  0.283  0.0140      1     5
#> # ℹ 94 more rows
# }