Log likelihood of T and it's derivatives (from stan)

Log likelihood of T and it's derivatives (from stan)

Usage

llikT(x, df, mean = 0, sd = 1, full = FALSE)

Arguments

x

Observation

df

degrees of freedom (\(> 0\), maybe non-integer). df = Inf is allowed.

mean

Mean for the likelihood

sd

Standard deviation for the likelihood

full

Add the data frame showing x, mean, sd as well as the fx and derivatives

Value

data frame with fx for the log pdf value of with dDf dMean and dSd that has the derivatives with respect to the parameters at the observation time-point

Details

In an rxode2() model, you can use llikT() but you have to use all arguments. You can also get the derivative of df, mean and sd with llikTDdf(), llikTDmean() and llikTDsd().

Author

Matthew L. Fidler

Examples

# \donttest{
x <- seq(-3, 3, length.out = 21)

llikT(x, 7, 0, 1)
#>            fx          dDf      dMean        dSd
#> 1  -4.2612484 -0.086858773 -1.5000000  3.5000000
#> 2  -3.8091335 -0.060260433 -1.5115465  3.0811756
#> 3  -3.3561547 -0.037201674 -1.5047022  2.6112853
#> 4  -2.9088542 -0.018379169 -1.4723926  2.0920245
#> 5  -2.4761000 -0.004340221 -1.4062500  1.5312500
#> 6  -2.0693878  0.004691380 -1.2972973  0.9459459
#> 7  -1.7028228  0.009010785 -1.1374408  0.3649289
#> 8  -1.3925134  0.009569214 -0.9218950 -0.1702945
#> 9  -1.1551333  0.007927361 -0.6521739 -0.6086957
#> 10 -1.0056349  0.005918024 -0.3385049 -0.8984485
#> 11 -0.9545342  0.005051942  0.0000000 -1.0000000
#> 12 -1.0056349  0.005918024  0.3385049 -0.8984485
#> 13 -1.1551333  0.007927361  0.6521739 -0.6086957
#> 14 -1.3925134  0.009569214  0.9218950 -0.1702945
#> 15 -1.7028228  0.009010785  1.1374408  0.3649289
#> 16 -2.0693878  0.004691380  1.2972973  0.9459459
#> 17 -2.4761000 -0.004340221  1.4062500  1.5312500
#> 18 -2.9088542 -0.018379169  1.4723926  2.0920245
#> 19 -3.3561547 -0.037201674  1.5047022  2.6112853
#> 20 -3.8091335 -0.060260433  1.5115465  3.0811756
#> 21 -4.2612484 -0.086858773  1.5000000  3.5000000

llikT(x, 15, 0, 1, full=TRUE)
#>            fx           dDf      dMean        dSd
#> 1  -4.6956220 -0.0338931511 -2.0000000  5.0000000
#> 2  -4.1042965 -0.0225073150 -1.9380888  4.2328398
#> 3  -3.5354158 -0.0134033865 -1.8497110  3.4393064
#> 4  -2.9974985 -0.0065857823 -1.7310665  2.6352396
#> 5  -2.5001272 -0.0019378861 -1.5789474  1.8421053
#> 6  -2.0536885  0.0007929097 -1.3913043  1.0869565
#> 7  -1.6689304  0.0019903977 -1.1678832  0.4014599
#> 8  -1.3563325  0.0021369166 -0.9108159 -0.1802657
#> 9  -1.1253251  0.0017504002 -0.6250000 -0.6250000
#> 10 -0.9834495  0.0012985422 -0.3180915 -0.9045726
#> 11 -0.9355929  0.0011086635  0.0000000 -1.0000000
#> 12 -0.9834495  0.0012985422  0.3180915 -0.9045726
#> 13 -1.1253251  0.0017504002  0.6250000 -0.6250000
#> 14 -1.3563325  0.0021369166  0.9108159 -0.1802657
#> 15 -1.6689304  0.0019903977  1.1678832  0.4014599
#> 16 -2.0536885  0.0007929097  1.3913043  1.0869565
#> 17 -2.5001272 -0.0019378861  1.5789474  1.8421053
#> 18 -2.9974985 -0.0065857823  1.7310665  2.6352396
#> 19 -3.5354158 -0.0134033865  1.8497110  3.4393064
#> 20 -4.1042965 -0.0225073150  1.9380888  4.2328398
#> 21 -4.6956220 -0.0338931511  2.0000000  5.0000000

et <- et(-3, 3, length.out=10)
et$nu <- 7
et$mean <- 0
et$sd <- 1

model <- function() {
  model({
    fx <- llikT(time, nu, mean, sd)
    dDf <- llikTDdf(time, nu, mean, sd)
    dMean <- llikTDmean(time, nu, mean, sd)
    dSd   <- llikTDsd(time, nu, mean, sd)
  })
}

rxSolve(model, et)
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#>  
#>  
#> C model:
#> // Define 4 LHS values
#> #define _LHS_0_ 0 // fx
#> #define _LHS_1_ 1 // dDf
#> #define _LHS_2_ 2 // dMean
#> #define _LHS_3_ 3 // dSd
#> #define _getRxSolve_ _rxba708c89ec9ae5c319e4f4c6846419594004__getRxSolve__9xah
#> #define _evalUdf _rxba708c89ec9ae5c319e4f4c6846419594005__evalUdf_gBOZ
#> #define _solveData _rxba708c89ec9ae5c319e4f4c6846419594006__solveData_l3zY
#> #define _assign_ptr _rxba708c89ec9ae5c319e4f4c6846419594007__assign_ptr_vTa9
#> #define _rxRmModelLib _rxba708c89ec9ae5c319e4f4c6846419594008__rxRmModelLib_7EhW
#> #define _rxGetModelLib _rxba708c89ec9ae5c319e4f4c6846419594009__rxGetModelLib_3MG9
#> #define _old_c _rxba708c89ec9ae5c319e4f4c6846419594010__old_c_V3pR
#> #define _ptrid _rxba708c89ec9ae5c319e4f4c6846419594011__ptrid_1aNk
#> #define _rxIsCurrentC _rxba708c89ec9ae5c319e4f4c6846419594012__rxIsCurrentC_G5Jw
#> #define _sumPS _rxba708c89ec9ae5c319e4f4c6846419594013__sumPS_D9TK
#> #define _prodPS _rxba708c89ec9ae5c319e4f4c6846419594014__prodPS_40NP
#> #define _prodType _rxba708c89ec9ae5c319e4f4c6846419594015__prodType_JiAL
#> #define _sumType _rxba708c89ec9ae5c319e4f4c6846419594016__sumType_ZBrh
#> #define _update_par_ptr _rxba708c89ec9ae5c319e4f4c6846419594017__update_par_ptr_efgh
#> #define _getParCov _rxba708c89ec9ae5c319e4f4c6846419594018__getParCov_Drm8
#> #define _rxode2_rxAssignPtr _rxba708c89ec9ae5c319e4f4c6846419594019__rxode2_rxAssignPtr_5igv
#> #define _rxQr _rxba708c89ec9ae5c319e4f4c6846419594020__rxQr_lJYg
#> #define _compareFactorVal _rxba708c89ec9ae5c319e4f4c6846419594021__compareFactorVal_GqQx
#> #define _sum _rxba708c89ec9ae5c319e4f4c6846419594022__sum_LoYl
#> #define _udf _rxba708c89ec9ae5c319e4f4c6846419594023__udf_FgJd
#> #define _sign _rxba708c89ec9ae5c319e4f4c6846419594024__sign_KaFg
#> #define _prod _rxba708c89ec9ae5c319e4f4c6846419594025__prod_mDR4
#> #define _max _rxba708c89ec9ae5c319e4f4c6846419594026__max_jxuo
#> #define _min _rxba708c89ec9ae5c319e4f4c6846419594027__min_P8gm
#> #define _transit4P _rxba708c89ec9ae5c319e4f4c6846419594028__transit4P_oUom
#> #define _transit3P _rxba708c89ec9ae5c319e4f4c6846419594029__transit3P_wLtu
#> #define _assignFuns0 _rxba708c89ec9ae5c319e4f4c6846419594030__assignFuns0_DgOz
#> #define _assignFuns _rxba708c89ec9ae5c319e4f4c6846419594031__assignFuns_tPZw
#> #define _rxord _rxba708c89ec9ae5c319e4f4c6846419594032__rxord_Dqx2
#> #define __assignFuns2 _rxba708c89ec9ae5c319e4f4c6846419594033___assignFuns2_t0Ff
#> #define _llikCauchyDscale _rxba708c89ec9ae5c319e4f4c6846419594034__llikCauchyDscale_CRup
#> #define _llikCauchyDlocation _rxba708c89ec9ae5c319e4f4c6846419594035__llikCauchyDlocation_BfcL
#> #define _llikCauchy _rxba708c89ec9ae5c319e4f4c6846419594036__llikCauchy_LlJk
#> #define _llikGammaDrate _rxba708c89ec9ae5c319e4f4c6846419594037__llikGammaDrate_0HOu
#> #define _llikGammaDshape _rxba708c89ec9ae5c319e4f4c6846419594038__llikGammaDshape_quMQ
#> #define _llikGamma _rxba708c89ec9ae5c319e4f4c6846419594039__llikGamma_IP4s
#> #define _llikWeibullDscale _rxba708c89ec9ae5c319e4f4c6846419594040__llikWeibullDscale_BKwl
#> #define _llikWeibullDshape _rxba708c89ec9ae5c319e4f4c6846419594041__llikWeibullDshape_ION1
#> #define _llikWeibull _rxba708c89ec9ae5c319e4f4c6846419594042__llikWeibull_tTX5
#> #define _llikUnifDbeta _rxba708c89ec9ae5c319e4f4c6846419594043__llikUnifDbeta_DulH
#> #define _llikUnifDalpha _rxba708c89ec9ae5c319e4f4c6846419594044__llikUnifDalpha_LxeA
#> #define _llikUnif _rxba708c89ec9ae5c319e4f4c6846419594045__llikUnif_Ut0p
#> #define _llikGeomDp _rxba708c89ec9ae5c319e4f4c6846419594046__llikGeomDp_o6Aq
#> #define _llikGeom _rxba708c89ec9ae5c319e4f4c6846419594047__llikGeom_SQkn
#> #define _llikFDdf2 _rxba708c89ec9ae5c319e4f4c6846419594048__llikFDdf2_H8fP
#> #define _llikFDdf1 _rxba708c89ec9ae5c319e4f4c6846419594049__llikFDdf1_CAnN
#> #define _llikF _rxba708c89ec9ae5c319e4f4c6846419594050__llikF_kUms
#> #define _llikExpDrate _rxba708c89ec9ae5c319e4f4c6846419594051__llikExpDrate_wiOW
#> #define _llikExp _rxba708c89ec9ae5c319e4f4c6846419594052__llikExp_G16H
#> #define _llikChisqDdf _rxba708c89ec9ae5c319e4f4c6846419594053__llikChisqDdf_PUgn
#> #define _llikChisq _rxba708c89ec9ae5c319e4f4c6846419594054__llikChisq_Cjuw
#> #define _llikTDsd _rxba708c89ec9ae5c319e4f4c6846419594055__llikTDsd_A8iw
#> #define _llikTDmean _rxba708c89ec9ae5c319e4f4c6846419594056__llikTDmean_bn2G
#> #define _llikTDdf _rxba708c89ec9ae5c319e4f4c6846419594057__llikTDdf_TQjS
#> #define _llikT _rxba708c89ec9ae5c319e4f4c6846419594058__llikT_rer5
#> #define _llikBetaDshape2 _rxba708c89ec9ae5c319e4f4c6846419594059__llikBetaDshape2_zexL
#> #define _llikBetaDshape1 _rxba708c89ec9ae5c319e4f4c6846419594060__llikBetaDshape1_gAwM
#> #define _llikBeta _rxba708c89ec9ae5c319e4f4c6846419594061__llikBeta_E6nq
#> #define _llikNbinomMuDmu _rxba708c89ec9ae5c319e4f4c6846419594062__llikNbinomMuDmu_R3DD
#> #define _llikNbinomMu _rxba708c89ec9ae5c319e4f4c6846419594063__llikNbinomMu_9jN5
#> #define _llikNbinomDprob _rxba708c89ec9ae5c319e4f4c6846419594064__llikNbinomDprob_O4r9
#> #define _llikNbinom _rxba708c89ec9ae5c319e4f4c6846419594065__llikNbinom_D97m
#> #define _llikBinomDprob _rxba708c89ec9ae5c319e4f4c6846419594066__llikBinomDprob_e5yb
#> #define _llikBinom _rxba708c89ec9ae5c319e4f4c6846419594067__llikBinom_V8Eu
#> #define _llikPoisDlambda _rxba708c89ec9ae5c319e4f4c6846419594068__llikPoisDlambda_Z1xT
#> #define _llikPois _rxba708c89ec9ae5c319e4f4c6846419594069__llikPois_QYOQ
#> #define _llikNormDsd _rxba708c89ec9ae5c319e4f4c6846419594070__llikNormDsd_e6aq
#> #define _llikNormDmean _rxba708c89ec9ae5c319e4f4c6846419594071__llikNormDmean_t3Lb
#> #define _llikNorm _rxba708c89ec9ae5c319e4f4c6846419594072__llikNorm_sbea
#> #define simeps _rxba708c89ec9ae5c319e4f4c6846419594073_simeps_w5m4
#> #define simeta _rxba708c89ec9ae5c319e4f4c6846419594074_simeta_brD9
#> #define expit _rxba708c89ec9ae5c319e4f4c6846419594075_expit_EZAJ
#> #define logit _rxba708c89ec9ae5c319e4f4c6846419594076_logit_Ne7I
#> #define gammapDer _rxba708c89ec9ae5c319e4f4c6846419594077_gammapDer_xUFA
#> #define lowergamma _rxba708c89ec9ae5c319e4f4c6846419594078_lowergamma_5R2w
#> #define uppergamma _rxba708c89ec9ae5c319e4f4c6846419594079_uppergamma_ZwwB
#> #define gammaqInva _rxba708c89ec9ae5c319e4f4c6846419594080_gammaqInva_gRvb
#> #define gammaqInv _rxba708c89ec9ae5c319e4f4c6846419594081_gammaqInv_lnk3
#> #define gammapInva _rxba708c89ec9ae5c319e4f4c6846419594082_gammapInva_57cf
#> #define gammapInv _rxba708c89ec9ae5c319e4f4c6846419594083_gammapInv_UwMa
#> #define gammaq _rxba708c89ec9ae5c319e4f4c6846419594084_gammaq_rEA9
#> #define gammap _rxba708c89ec9ae5c319e4f4c6846419594085_gammap_z1SE
#> #define phi _rxba708c89ec9ae5c319e4f4c6846419594086_phi_1KzA
#> #define d2ELUa _rxba708c89ec9ae5c319e4f4c6846419594087_d2ELUa_f2H4
#> #define dELUa _rxba708c89ec9ae5c319e4f4c6846419594088_dELUa_Q2oF
#> #define d2aELU _rxba708c89ec9ae5c319e4f4c6846419594089_d2aELU_0joC
#> #define d2ELU _rxba708c89ec9ae5c319e4f4c6846419594090_d2ELU_Gj0a
#> #define dELU _rxba708c89ec9ae5c319e4f4c6846419594091_dELU_8wAb
#> #define ELU _rxba708c89ec9ae5c319e4f4c6846419594092_ELU_QELO
#> #define dPReLUa1 _rxba708c89ec9ae5c319e4f4c6846419594093_dPReLUa1_lObq
#> #define dPReLUa _rxba708c89ec9ae5c319e4f4c6846419594094_dPReLUa_Nath
#> #define dPReLU _rxba708c89ec9ae5c319e4f4c6846419594095_dPReLU_ltiW
#> #define PReLU _rxba708c89ec9ae5c319e4f4c6846419594096_PReLU_R2IG
#> #define dSwish _rxba708c89ec9ae5c319e4f4c6846419594097_dSwish_VLQY
#> #define Swish _rxba708c89ec9ae5c319e4f4c6846419594098_Swish_M84J
#> #define dlReLU _rxba708c89ec9ae5c319e4f4c6846419594099_dlReLU_Cnuy
#> #define lReLU _rxba708c89ec9ae5c319e4f4c6846419594100_lReLU_CXPJ
#> #define dSELU _rxba708c89ec9ae5c319e4f4c6846419594101_dSELU_UjvF
#> #define SELU _rxba708c89ec9ae5c319e4f4c6846419594102_SELU_k7KJ
#> #define d4softplus _rxba708c89ec9ae5c319e4f4c6846419594103_d4softplus_vDxq
#> #define d3softplus _rxba708c89ec9ae5c319e4f4c6846419594104_d3softplus_bVSK
#> #define d2softplus _rxba708c89ec9ae5c319e4f4c6846419594105_d2softplus_Vfr7
#> #define dsoftplus _rxba708c89ec9ae5c319e4f4c6846419594106_dsoftplus_TcaU
#> #define softplus _rxba708c89ec9ae5c319e4f4c6846419594107_softplus_EgpY
#> #define d4GELU _rxba708c89ec9ae5c319e4f4c6846419594108_d4GELU_kNIq
#> #define d3GELU _rxba708c89ec9ae5c319e4f4c6846419594109_d3GELU_nEnm
#> #define d2GELU _rxba708c89ec9ae5c319e4f4c6846419594110_d2GELU_hSbo
#> #define dGELU _rxba708c89ec9ae5c319e4f4c6846419594111_dGELU_MZYB
#> #define GELU _rxba708c89ec9ae5c319e4f4c6846419594112_GELU_jkiC
#> #define dReLU _rxba708c89ec9ae5c319e4f4c6846419594113_dReLU_oUqv
#> #define ReLU _rxba708c89ec9ae5c319e4f4c6846419594114_ReLU_niI7
#> #define riweibull _rxba708c89ec9ae5c319e4f4c6846419594115_riweibull_5Ui0
#> #define riunif _rxba708c89ec9ae5c319e4f4c6846419594116_riunif_GVTs
#> #define rit_ _rxba708c89ec9ae5c319e4f4c6846419594117_rit__34TN
#> #define ripois _rxba708c89ec9ae5c319e4f4c6846419594118_ripois_HeFp
#> #define ribeta _rxba708c89ec9ae5c319e4f4c6846419594119_ribeta_ue1Z
#> #define rigamma _rxba708c89ec9ae5c319e4f4c6846419594120_rigamma_aEnJ
#> #define rigeom _rxba708c89ec9ae5c319e4f4c6846419594121_rigeom_R6UH
#> #define rif _rxba708c89ec9ae5c319e4f4c6846419594122_rif_ZKbS
#> #define riexp _rxba708c89ec9ae5c319e4f4c6846419594123_riexp_smNY
#> #define richisq _rxba708c89ec9ae5c319e4f4c6846419594124_richisq_lS2w
#> #define ricauchy _rxba708c89ec9ae5c319e4f4c6846419594125_ricauchy_zYqw
#> #define rinbinomMu _rxba708c89ec9ae5c319e4f4c6846419594126_rinbinomMu_2z8b
#> #define rinbinom _rxba708c89ec9ae5c319e4f4c6846419594127_rinbinom_GMdj
#> #define ribinom _rxba708c89ec9ae5c319e4f4c6846419594128_ribinom_LB3L
#> #define rinorm _rxba708c89ec9ae5c319e4f4c6846419594129_rinorm_sOUR
#> #define rxweibull _rxba708c89ec9ae5c319e4f4c6846419594130_rxweibull_bPYU
#> #define rxunif _rxba708c89ec9ae5c319e4f4c6846419594131_rxunif_PmW8
#> #define rxt_ _rxba708c89ec9ae5c319e4f4c6846419594132_rxt__sOmG
#> #define rxpois _rxba708c89ec9ae5c319e4f4c6846419594133_rxpois_ypO5
#> #define rxbeta _rxba708c89ec9ae5c319e4f4c6846419594134_rxbeta_5OnO
#> #define rxgamma _rxba708c89ec9ae5c319e4f4c6846419594135_rxgamma_E9TF
#> #define rxgeom _rxba708c89ec9ae5c319e4f4c6846419594136_rxgeom_yOK7
#> #define rxf _rxba708c89ec9ae5c319e4f4c6846419594137_rxf_OqEK
#> #define rxexp _rxba708c89ec9ae5c319e4f4c6846419594138_rxexp_wH4T
#> #define rxchisq _rxba708c89ec9ae5c319e4f4c6846419594139_rxchisq_tSJj
#> #define rxcauchy _rxba708c89ec9ae5c319e4f4c6846419594140_rxcauchy_0v8j
#> #define rxnbinomMu _rxba708c89ec9ae5c319e4f4c6846419594141_rxnbinomMu_xwnn
#> #define rxnbinom _rxba708c89ec9ae5c319e4f4c6846419594142_rxnbinom_6QBD
#> #define rxbinom _rxba708c89ec9ae5c319e4f4c6846419594143_rxbinom_Uzzs
#> #define rxnorm _rxba708c89ec9ae5c319e4f4c6846419594144_rxnorm_oz5n
#> #define linCmtB _rxba708c89ec9ae5c319e4f4c6846419594145_linCmtB_36W2
#> #define linCmtA _rxba708c89ec9ae5c319e4f4c6846419594146_linCmtA_2yJ1
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){  if (_solveData->svar[_svari] == 0) {nu = _PP[0];};   if (_solveData->svar[_svari] == 1) {mean = _PP[1];};   if (_solveData->svar[_svari] == 2) {sd = _PP[2];}; }
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){  if (_solveData->ovar[_ovari] == 0) {nu = _PP[0];};   if (_solveData->ovar[_ovari] == 1) {mean = _PP[1];};   if (_solveData->ovar[_ovari] == 2) {sd = _PP[2];}; }
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  rx_ba654182f679c11135a5f8cda0697cf0__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *rx_ba654182f679c11135a5f8cda0697cf0__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP rx_ba654182f679c11135a5f8cda0697cf0__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void rx_ba654182f679c11135a5f8cda0697cf0__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double nu = NA_REAL;
#>   double mean = NA_REAL;
#>   double sd = NA_REAL;
#>   double fx = NA_REAL;
#>   double dDf = NA_REAL;
#>   double dMean = NA_REAL;
#>   double dSd = NA_REAL;
#> 
#>   (void)t;
#>   (void)nu;
#>   (void)mean;
#>   (void)sd;
#>   (void)fx;
#>   (void)dDf;
#>   (void)dMean;
#>   (void)dSd;
#> 
#>   fx = _PL[0];
#>   dDf = _PL[1];
#>   dMean = _PL[2];
#>   dSd = _PL[3];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#>   nu = _PP[0];
#>   mean = _PP[1];
#>   sd = _PP[2];
#> 
#> 
#>   fx =llikT(t,nu,mean,sd);
#>   dDf =llikTDdf(t,nu,mean,sd);
#>   dMean =llikTDmean(t,nu,mean,sd);
#>   dSd =llikTDsd(t,nu,mean,sd);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void rx_ba654182f679c11135a5f8cda0697cf0__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void rx_ba654182f679c11135a5f8cda0697cf0__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void rx_ba654182f679c11135a5f8cda0697cf0__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double nu = NA_REAL;
#>   double mean = NA_REAL;
#>   double sd = NA_REAL;
#>   double fx = NA_REAL;
#>   double dDf = NA_REAL;
#>   double dMean = NA_REAL;
#>   double dSd = NA_REAL;
#> 
#>   (void)t;
#>   (void)nu;
#>   (void)mean;
#>   (void)sd;
#>   (void)fx;
#>   (void)dDf;
#>   (void)dMean;
#>   (void)dSd;
#> 
#>   fx = _PL[0];
#>   dDf = _PL[1];
#>   dMean = _PL[2];
#>   dSd = _PL[3];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#>   nu = _PP[0];
#>   mean = _PP[1];
#>   sd = _PP[2];
#> 
#> 
#>   fx =llikT(t,nu,mean,sd);
#>   dDf =llikTDdf(t,nu,mean,sd);
#>   dMean =llikTDmean(t,nu,mean,sd);
#>   dSd =llikTDsd(t,nu,mean,sd);
#> 
#>   _lhs[_LHS_0_]=fx;
#>   _lhs[_LHS_1_]=dDf;
#>   _lhs[_LHS_2_]=dMean;
#>   _lhs[_LHS_3_]=dSd;
#> }
#> // Functional based bioavailability
#> double rx_ba654182f679c11135a5f8cda0697cf0__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double rx_ba654182f679c11135a5f8cda0697cf0__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double rx_ba654182f679c11135a5f8cda0697cf0__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double rx_ba654182f679c11135a5f8cda0697cf0__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void rx_ba654182f679c11135a5f8cda0697cf0__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void rx_ba654182f679c11135a5f8cda0697cf0__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void rx_ba654182f679c11135a5f8cda0697cf0__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP rx_ba654182f679c11135a5f8cda0697cf0__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_ba654182f679c11135a5f8cda0697cf0__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAA*h4AAAv7#aT)VrZ`NAa{e.=WdwYB~1u%%C8Ae687/8E~=5IhiJn$<i@KwXHo+H1Iz{rVsUGMpYY\"X@0xiWWj}D/\?~gh@!Nc|C\"JnGIY^XE\"JtWcI@_).K0pU*V~PSch92>^D<sK+y1aNlm~|C)UU~>Q<UZ/EVVI+{31cShkVWrE`)Ts6F0r7U`)@kuaGO#psA6{t_eL;HM$X0~zPtHzvG\?Rv#glZR@,p0@Of25R`H^OOR0#g4v$B^uGm~idT~D%%K%%idujQJB%%C=\"fU#cj^(sB_w/kBpyc5xJbCzPQ,dj1mS,M++cG_0JqD_.0enHd}ZVIDGJ@n[_+OCNZH2bAsYyw=1qxlXDz_cvzc^3^<@z&c^rI0n/~A}Oody>i`h>ZHGA0lnZFByv\?(iRs45LteOS5]6<v\?]$X[<sK}2HOHbVDo0\?CDZHe%%6&>i9B1Cav/d*8\"\"#}Pv=4CAAs@l0hfYwFdp3`Y5fhw+UdazpL;bW<;W_Mmi6Uxx&^%%Y<.e;h!6_HxBdc)MNU0;r*j^^f7~bQA=3jW*x]a+YPeKhOhy$U*U69$Ui=t9\?gep\"sUryt\"5[<3xpJYP}CgDpb0N11}:OM~Lb2NbrGLhWpA~8jib0}6ngMaodE9^b|dDjP8$]92%%1Cw=1FDZ\"Pg.yy@^I=7a(VKg$p.,8\?cqcds,3Q#HcLmp~ri(|@7&Vnxy*>%%z9p$s3<edBDSs1ligm8=UkfY$K#C/yxXiS0<&Q)y|TZ`h#H{6xGevne=0r6qN*zeh7@Cw5$p|0SeJ+MX52WQ0)cd[hz4:*5Z/PzJwsI5k)y5xw(P78Ir]7BBl#5Iw6L~XqitB|XUwDz^j|TJXbw>hOlJBX/[3VW0_~yGiG!*Y}Cl3^)lIXT;S}UIL(#PJCUzx,+lyUt*o_(X/[z>]RPviwM%%N@12tXo3)jFwiaD4RP(+v;HJM$7C!wIdoOM1h;b@F:aLk~#=Rf4VJ9)Y{zet8ptTqjWN%%rifN^vrDS&@i>hwlN*X34E`ug>ql4dp[<N|4@d<.=l:Yu&6H^m_[\";vU`x{X@IY^NhVr*BZ_)DWk[=J0:j#W2oF*arMyvD!Hs>VHm^Y<I9(16iGhjj06h53R[xr7JU^c`gcJRgRME\"J/");
#>     char buf[1153];
#> #define __doBufN__ 1153
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void rx_ba654182f679c11135a5f8cda0697cf0__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   rx_ba654182f679c11135a5f8cda0697cf0__dydt(neq, *t, A, DADT);
#> }
#> extern int rx_ba654182f679c11135a5f8cda0697cf0__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   rx_ba654182f679c11135a5f8cda0697cf0__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void rx_ba654182f679c11135a5f8cda0697cf0__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   rx_ba654182f679c11135a5f8cda0697cf0__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void rx_ba654182f679c11135a5f8cda0697cf0__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_rx_ba654182f679c11135a5f8cda0697cf0_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("rx_ba654182f679c11135a5f8cda0697cf0_","rx_ba654182f679c11135a5f8cda0697cf0__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("rx_ba654182f679c11135a5f8cda0697cf0_","rx_ba654182f679c11135a5f8cda0697cf0__assignFuns", (DL_FUNC) rx_ba654182f679c11135a5f8cda0697cf0__assignFuns);
#>   R_RegisterCCallable("rx_ba654182f679c11135a5f8cda0697cf0_","rx_ba654182f679c11135a5f8cda0697cf0__inis",(DL_FUNC) rx_ba654182f679c11135a5f8cda0697cf0__inis);
#>   R_RegisterCCallable("rx_ba654182f679c11135a5f8cda0697cf0_","rx_ba654182f679c11135a5f8cda0697cf0__dydt",(DL_FUNC) rx_ba654182f679c11135a5f8cda0697cf0__dydt);
#>   R_RegisterCCallable("rx_ba654182f679c11135a5f8cda0697cf0_","rx_ba654182f679c11135a5f8cda0697cf0__calc_lhs",(DL_FUNC) rx_ba654182f679c11135a5f8cda0697cf0__calc_lhs);
#>   R_RegisterCCallable("rx_ba654182f679c11135a5f8cda0697cf0_","rx_ba654182f679c11135a5f8cda0697cf0__calc_jac",(DL_FUNC) rx_ba654182f679c11135a5f8cda0697cf0__calc_jac);
#>   R_RegisterCCallable("rx_ba654182f679c11135a5f8cda0697cf0_","rx_ba654182f679c11135a5f8cda0697cf0__dydt_lsoda", (DL_FUNC) rx_ba654182f679c11135a5f8cda0697cf0__dydt_lsoda);
#>   R_RegisterCCallable("rx_ba654182f679c11135a5f8cda0697cf0_","rx_ba654182f679c11135a5f8cda0697cf0__calc_jac_lsoda", (DL_FUNC) rx_ba654182f679c11135a5f8cda0697cf0__calc_jac_lsoda);
#>   R_RegisterCCallable("rx_ba654182f679c11135a5f8cda0697cf0_","rx_ba654182f679c11135a5f8cda0697cf0__ode_solver_solvedata", (DL_FUNC) rx_ba654182f679c11135a5f8cda0697cf0__ode_solver_solvedata);
#>   R_RegisterCCallable("rx_ba654182f679c11135a5f8cda0697cf0_","rx_ba654182f679c11135a5f8cda0697cf0__ode_solver_get_solvedata", (DL_FUNC) rx_ba654182f679c11135a5f8cda0697cf0__ode_solver_get_solvedata);
#>   R_RegisterCCallable("rx_ba654182f679c11135a5f8cda0697cf0_","rx_ba654182f679c11135a5f8cda0697cf0__F", (DL_FUNC) rx_ba654182f679c11135a5f8cda0697cf0__F);
#>   R_RegisterCCallable("rx_ba654182f679c11135a5f8cda0697cf0_","rx_ba654182f679c11135a5f8cda0697cf0__Lag", (DL_FUNC) rx_ba654182f679c11135a5f8cda0697cf0__Lag);
#>   R_RegisterCCallable("rx_ba654182f679c11135a5f8cda0697cf0_","rx_ba654182f679c11135a5f8cda0697cf0__Rate", (DL_FUNC) rx_ba654182f679c11135a5f8cda0697cf0__Rate);
#>   R_RegisterCCallable("rx_ba654182f679c11135a5f8cda0697cf0_","rx_ba654182f679c11135a5f8cda0697cf0__Dur", (DL_FUNC) rx_ba654182f679c11135a5f8cda0697cf0__Dur);
#>   R_RegisterCCallable("rx_ba654182f679c11135a5f8cda0697cf0_","rx_ba654182f679c11135a5f8cda0697cf0__mtime", (DL_FUNC) rx_ba654182f679c11135a5f8cda0697cf0__mtime);
#>   R_RegisterCCallable("rx_ba654182f679c11135a5f8cda0697cf0_","rx_ba654182f679c11135a5f8cda0697cf0__ME", (DL_FUNC) rx_ba654182f679c11135a5f8cda0697cf0__ME);
#>   R_RegisterCCallable("rx_ba654182f679c11135a5f8cda0697cf0_","rx_ba654182f679c11135a5f8cda0697cf0__IndF", (DL_FUNC) rx_ba654182f679c11135a5f8cda0697cf0__IndF);
#>   R_RegisterCCallable("rx_ba654182f679c11135a5f8cda0697cf0_","rx_ba654182f679c11135a5f8cda0697cf0__dydt_liblsoda", (DL_FUNC) rx_ba654182f679c11135a5f8cda0697cf0__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_rx_ba654182f679c11135a5f8cda0697cf0_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_rx_ba654182f679c11135a5f8cda0697cf0_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"rx_ba654182f679c11135a5f8cda0697cf0__model_vars", (DL_FUNC) &rx_ba654182f679c11135a5f8cda0697cf0__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_rx_ba654182f679c11135a5f8cda0697cf0_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_ba654182f679c11135a5f8cda0697cf0__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("rx_ba654182f679c11135a5f8cda0697cf0__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’
#> ── Solved rxode2 object ──
#> ── Parameters (value$params): ──
#> # A tibble: 1 × 0
#> ── Initial Conditions (value$inits): ──
#> named numeric(0)
#> ── First part of data (object): ──
#> # A tibble: 10 × 8
#>     time    fx       dDf  dMean    dSd    nu  mean    sd
#>    <dbl> <dbl>     <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>
#> 1 -3     -4.26 -0.0869   -1.5    3.5       7     0     1
#> 2 -2.33  -3.26 -0.0326   -1.5    2.5       7     0     1
#> 3 -1.67  -2.29  0.000289 -1.36   1.27      7     0     1
#> 4 -1     -1.49  0.00971  -1      0         7     0     1
#> 5 -0.333 -1.02  0.00611  -0.375 -0.875     7     0     1
#> 6  0.333 -1.02  0.00611   0.375 -0.875     7     0     1
#> # ℹ 4 more rows
# }