log likelihood and derivatives for Unif distribution

log likelihood and derivatives for Unif distribution

Usage

llikUnif(x, alpha, beta, full = FALSE)

Arguments

x

variable distributed by a uniform distribution

alpha

is the lower limit of the uniform distribution

beta

is the upper limit of the distribution

full

Add the data frame showing x, mean, sd as well as the fx and derivatives

Value

data frame with fx for the log pdf value of with dProb that has the derivatives with respect to the prob parameters at the observation time-point

Details

In an rxode2() model, you can use llikUnif() but you have to use the x and rate arguments. You can also get the derivative of alpha or beta with llikUnifDalpha() and llikUnifDbeta().

Author

Matthew L. Fidler

Examples

# \donttest{

llikUnif(1, -2, 2)
#>          fx dAlpha dBeta
#> 1 -1.386294   0.25 -0.25

et  <- et(seq(1,1, length.out=4))
et$alpha <- -2
et$beta <- 2
 
model <- function() {
  model({
    fx <- llikUnif(time, alpha, beta)
    dAlpha<- llikUnifDalpha(time, alpha, beta)
    dBeta <- llikUnifDbeta(time, alpha, beta)
  })
}

rxSolve(model, et)
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#>  
#>  
#> C model:
#> // Define 3 LHS values
#> #define _LHS_0_ 0 // fx
#> #define _LHS_1_ 1 // dAlpha
#> #define _LHS_2_ 2 // dBeta
#> #define _getRxSolve_ _rx9c54122483901b6b54a4545874ccacf14147__getRxSolve__gBOZ
#> #define _evalUdf _rx9c54122483901b6b54a4545874ccacf14148__evalUdf_l3zY
#> #define _solveData _rx9c54122483901b6b54a4545874ccacf14149__solveData_vTa9
#> #define _assign_ptr _rx9c54122483901b6b54a4545874ccacf14150__assign_ptr_7EhW
#> #define _rxRmModelLib _rx9c54122483901b6b54a4545874ccacf14151__rxRmModelLib_3MG9
#> #define _rxGetModelLib _rx9c54122483901b6b54a4545874ccacf14152__rxGetModelLib_V3pR
#> #define _old_c _rx9c54122483901b6b54a4545874ccacf14153__old_c_1aNk
#> #define _ptrid _rx9c54122483901b6b54a4545874ccacf14154__ptrid_G5Jw
#> #define _rxIsCurrentC _rx9c54122483901b6b54a4545874ccacf14155__rxIsCurrentC_D9TK
#> #define _sumPS _rx9c54122483901b6b54a4545874ccacf14156__sumPS_40NP
#> #define _prodPS _rx9c54122483901b6b54a4545874ccacf14157__prodPS_JiAL
#> #define _prodType _rx9c54122483901b6b54a4545874ccacf14158__prodType_ZBrh
#> #define _sumType _rx9c54122483901b6b54a4545874ccacf14159__sumType_efgh
#> #define _update_par_ptr _rx9c54122483901b6b54a4545874ccacf14160__update_par_ptr_Drm8
#> #define _getParCov _rx9c54122483901b6b54a4545874ccacf14161__getParCov_5igv
#> #define _rxode2_rxAssignPtr _rx9c54122483901b6b54a4545874ccacf14162__rxode2_rxAssignPtr_lJYg
#> #define _rxQr _rx9c54122483901b6b54a4545874ccacf14163__rxQr_GqQx
#> #define _compareFactorVal _rx9c54122483901b6b54a4545874ccacf14164__compareFactorVal_LoYl
#> #define _sum _rx9c54122483901b6b54a4545874ccacf14165__sum_FgJd
#> #define _udf _rx9c54122483901b6b54a4545874ccacf14166__udf_KaFg
#> #define _sign _rx9c54122483901b6b54a4545874ccacf14167__sign_mDR4
#> #define _prod _rx9c54122483901b6b54a4545874ccacf14168__prod_jxuo
#> #define _max _rx9c54122483901b6b54a4545874ccacf14169__max_P8gm
#> #define _min _rx9c54122483901b6b54a4545874ccacf14170__min_oUom
#> #define _transit4P _rx9c54122483901b6b54a4545874ccacf14171__transit4P_wLtu
#> #define _transit3P _rx9c54122483901b6b54a4545874ccacf14172__transit3P_DgOz
#> #define _assignFuns0 _rx9c54122483901b6b54a4545874ccacf14173__assignFuns0_tPZw
#> #define _assignFuns _rx9c54122483901b6b54a4545874ccacf14174__assignFuns_Dqx2
#> #define _rxord _rx9c54122483901b6b54a4545874ccacf14175__rxord_t0Ff
#> #define __assignFuns2 _rx9c54122483901b6b54a4545874ccacf14176___assignFuns2_CRup
#> #define _llikCauchyDscale _rx9c54122483901b6b54a4545874ccacf14177__llikCauchyDscale_BfcL
#> #define _llikCauchyDlocation _rx9c54122483901b6b54a4545874ccacf14178__llikCauchyDlocation_LlJk
#> #define _llikCauchy _rx9c54122483901b6b54a4545874ccacf14179__llikCauchy_0HOu
#> #define _llikGammaDrate _rx9c54122483901b6b54a4545874ccacf14180__llikGammaDrate_quMQ
#> #define _llikGammaDshape _rx9c54122483901b6b54a4545874ccacf14181__llikGammaDshape_IP4s
#> #define _llikGamma _rx9c54122483901b6b54a4545874ccacf14182__llikGamma_BKwl
#> #define _llikWeibullDscale _rx9c54122483901b6b54a4545874ccacf14183__llikWeibullDscale_ION1
#> #define _llikWeibullDshape _rx9c54122483901b6b54a4545874ccacf14184__llikWeibullDshape_tTX5
#> #define _llikWeibull _rx9c54122483901b6b54a4545874ccacf14185__llikWeibull_DulH
#> #define _llikUnifDbeta _rx9c54122483901b6b54a4545874ccacf14186__llikUnifDbeta_LxeA
#> #define _llikUnifDalpha _rx9c54122483901b6b54a4545874ccacf14187__llikUnifDalpha_Ut0p
#> #define _llikUnif _rx9c54122483901b6b54a4545874ccacf14188__llikUnif_o6Aq
#> #define _llikGeomDp _rx9c54122483901b6b54a4545874ccacf14189__llikGeomDp_SQkn
#> #define _llikGeom _rx9c54122483901b6b54a4545874ccacf14190__llikGeom_H8fP
#> #define _llikFDdf2 _rx9c54122483901b6b54a4545874ccacf14191__llikFDdf2_CAnN
#> #define _llikFDdf1 _rx9c54122483901b6b54a4545874ccacf14192__llikFDdf1_kUms
#> #define _llikF _rx9c54122483901b6b54a4545874ccacf14193__llikF_wiOW
#> #define _llikExpDrate _rx9c54122483901b6b54a4545874ccacf14194__llikExpDrate_G16H
#> #define _llikExp _rx9c54122483901b6b54a4545874ccacf14195__llikExp_PUgn
#> #define _llikChisqDdf _rx9c54122483901b6b54a4545874ccacf14196__llikChisqDdf_Cjuw
#> #define _llikChisq _rx9c54122483901b6b54a4545874ccacf14197__llikChisq_A8iw
#> #define _llikTDsd _rx9c54122483901b6b54a4545874ccacf14198__llikTDsd_bn2G
#> #define _llikTDmean _rx9c54122483901b6b54a4545874ccacf14199__llikTDmean_TQjS
#> #define _llikTDdf _rx9c54122483901b6b54a4545874ccacf14200__llikTDdf_rer5
#> #define _llikT _rx9c54122483901b6b54a4545874ccacf14201__llikT_zexL
#> #define _llikBetaDshape2 _rx9c54122483901b6b54a4545874ccacf14202__llikBetaDshape2_gAwM
#> #define _llikBetaDshape1 _rx9c54122483901b6b54a4545874ccacf14203__llikBetaDshape1_E6nq
#> #define _llikBeta _rx9c54122483901b6b54a4545874ccacf14204__llikBeta_R3DD
#> #define _llikNbinomMuDmu _rx9c54122483901b6b54a4545874ccacf14205__llikNbinomMuDmu_9jN5
#> #define _llikNbinomMu _rx9c54122483901b6b54a4545874ccacf14206__llikNbinomMu_O4r9
#> #define _llikNbinomDprob _rx9c54122483901b6b54a4545874ccacf14207__llikNbinomDprob_D97m
#> #define _llikNbinom _rx9c54122483901b6b54a4545874ccacf14208__llikNbinom_e5yb
#> #define _llikBinomDprob _rx9c54122483901b6b54a4545874ccacf14209__llikBinomDprob_V8Eu
#> #define _llikBinom _rx9c54122483901b6b54a4545874ccacf14210__llikBinom_Z1xT
#> #define _llikPoisDlambda _rx9c54122483901b6b54a4545874ccacf14211__llikPoisDlambda_QYOQ
#> #define _llikPois _rx9c54122483901b6b54a4545874ccacf14212__llikPois_e6aq
#> #define _llikNormDsd _rx9c54122483901b6b54a4545874ccacf14213__llikNormDsd_t3Lb
#> #define _llikNormDmean _rx9c54122483901b6b54a4545874ccacf14214__llikNormDmean_sbea
#> #define _llikNorm _rx9c54122483901b6b54a4545874ccacf14215__llikNorm_w5m4
#> #define simeps _rx9c54122483901b6b54a4545874ccacf14216_simeps_brD9
#> #define simeta _rx9c54122483901b6b54a4545874ccacf14217_simeta_EZAJ
#> #define expit _rx9c54122483901b6b54a4545874ccacf14218_expit_Ne7I
#> #define logit _rx9c54122483901b6b54a4545874ccacf14219_logit_xUFA
#> #define gammapDer _rx9c54122483901b6b54a4545874ccacf14220_gammapDer_5R2w
#> #define lowergamma _rx9c54122483901b6b54a4545874ccacf14221_lowergamma_ZwwB
#> #define uppergamma _rx9c54122483901b6b54a4545874ccacf14222_uppergamma_gRvb
#> #define gammaqInva _rx9c54122483901b6b54a4545874ccacf14223_gammaqInva_lnk3
#> #define gammaqInv _rx9c54122483901b6b54a4545874ccacf14224_gammaqInv_57cf
#> #define gammapInva _rx9c54122483901b6b54a4545874ccacf14225_gammapInva_UwMa
#> #define gammapInv _rx9c54122483901b6b54a4545874ccacf14226_gammapInv_rEA9
#> #define gammaq _rx9c54122483901b6b54a4545874ccacf14227_gammaq_z1SE
#> #define gammap _rx9c54122483901b6b54a4545874ccacf14228_gammap_1KzA
#> #define phi _rx9c54122483901b6b54a4545874ccacf14229_phi_f2H4
#> #define d2ELUa _rx9c54122483901b6b54a4545874ccacf14230_d2ELUa_Q2oF
#> #define dELUa _rx9c54122483901b6b54a4545874ccacf14231_dELUa_0joC
#> #define d2aELU _rx9c54122483901b6b54a4545874ccacf14232_d2aELU_Gj0a
#> #define d2ELU _rx9c54122483901b6b54a4545874ccacf14233_d2ELU_8wAb
#> #define dELU _rx9c54122483901b6b54a4545874ccacf14234_dELU_QELO
#> #define ELU _rx9c54122483901b6b54a4545874ccacf14235_ELU_lObq
#> #define dPReLUa1 _rx9c54122483901b6b54a4545874ccacf14236_dPReLUa1_Nath
#> #define dPReLUa _rx9c54122483901b6b54a4545874ccacf14237_dPReLUa_ltiW
#> #define dPReLU _rx9c54122483901b6b54a4545874ccacf14238_dPReLU_R2IG
#> #define PReLU _rx9c54122483901b6b54a4545874ccacf14239_PReLU_VLQY
#> #define dSwish _rx9c54122483901b6b54a4545874ccacf14240_dSwish_M84J
#> #define Swish _rx9c54122483901b6b54a4545874ccacf14241_Swish_Cnuy
#> #define dlReLU _rx9c54122483901b6b54a4545874ccacf14242_dlReLU_CXPJ
#> #define lReLU _rx9c54122483901b6b54a4545874ccacf14243_lReLU_UjvF
#> #define dSELU _rx9c54122483901b6b54a4545874ccacf14244_dSELU_k7KJ
#> #define SELU _rx9c54122483901b6b54a4545874ccacf14245_SELU_vDxq
#> #define d4softplus _rx9c54122483901b6b54a4545874ccacf14246_d4softplus_bVSK
#> #define d3softplus _rx9c54122483901b6b54a4545874ccacf14247_d3softplus_Vfr7
#> #define d2softplus _rx9c54122483901b6b54a4545874ccacf14248_d2softplus_TcaU
#> #define dsoftplus _rx9c54122483901b6b54a4545874ccacf14249_dsoftplus_EgpY
#> #define softplus _rx9c54122483901b6b54a4545874ccacf14250_softplus_kNIq
#> #define d4GELU _rx9c54122483901b6b54a4545874ccacf14251_d4GELU_nEnm
#> #define d3GELU _rx9c54122483901b6b54a4545874ccacf14252_d3GELU_hSbo
#> #define d2GELU _rx9c54122483901b6b54a4545874ccacf14253_d2GELU_MZYB
#> #define dGELU _rx9c54122483901b6b54a4545874ccacf14254_dGELU_jkiC
#> #define GELU _rx9c54122483901b6b54a4545874ccacf14255_GELU_oUqv
#> #define dReLU _rx9c54122483901b6b54a4545874ccacf14256_dReLU_niI7
#> #define ReLU _rx9c54122483901b6b54a4545874ccacf14257_ReLU_5Ui0
#> #define riweibull _rx9c54122483901b6b54a4545874ccacf14258_riweibull_GVTs
#> #define riunif _rx9c54122483901b6b54a4545874ccacf14259_riunif_34TN
#> #define rit_ _rx9c54122483901b6b54a4545874ccacf14260_rit__HeFp
#> #define ripois _rx9c54122483901b6b54a4545874ccacf14261_ripois_ue1Z
#> #define ribeta _rx9c54122483901b6b54a4545874ccacf14262_ribeta_aEnJ
#> #define rigamma _rx9c54122483901b6b54a4545874ccacf14263_rigamma_R6UH
#> #define rigeom _rx9c54122483901b6b54a4545874ccacf14264_rigeom_ZKbS
#> #define rif _rx9c54122483901b6b54a4545874ccacf14265_rif_smNY
#> #define riexp _rx9c54122483901b6b54a4545874ccacf14266_riexp_lS2w
#> #define richisq _rx9c54122483901b6b54a4545874ccacf14267_richisq_zYqw
#> #define ricauchy _rx9c54122483901b6b54a4545874ccacf14268_ricauchy_2z8b
#> #define rinbinomMu _rx9c54122483901b6b54a4545874ccacf14269_rinbinomMu_GMdj
#> #define rinbinom _rx9c54122483901b6b54a4545874ccacf14270_rinbinom_LB3L
#> #define ribinom _rx9c54122483901b6b54a4545874ccacf14271_ribinom_sOUR
#> #define rinorm _rx9c54122483901b6b54a4545874ccacf14272_rinorm_bPYU
#> #define rxweibull _rx9c54122483901b6b54a4545874ccacf14273_rxweibull_PmW8
#> #define rxunif _rx9c54122483901b6b54a4545874ccacf14274_rxunif_sOmG
#> #define rxt_ _rx9c54122483901b6b54a4545874ccacf14275_rxt__ypO5
#> #define rxpois _rx9c54122483901b6b54a4545874ccacf14276_rxpois_5OnO
#> #define rxbeta _rx9c54122483901b6b54a4545874ccacf14277_rxbeta_E9TF
#> #define rxgamma _rx9c54122483901b6b54a4545874ccacf14278_rxgamma_yOK7
#> #define rxgeom _rx9c54122483901b6b54a4545874ccacf14279_rxgeom_OqEK
#> #define rxf _rx9c54122483901b6b54a4545874ccacf14280_rxf_wH4T
#> #define rxexp _rx9c54122483901b6b54a4545874ccacf14281_rxexp_tSJj
#> #define rxchisq _rx9c54122483901b6b54a4545874ccacf14282_rxchisq_0v8j
#> #define rxcauchy _rx9c54122483901b6b54a4545874ccacf14283_rxcauchy_xwnn
#> #define rxnbinomMu _rx9c54122483901b6b54a4545874ccacf14284_rxnbinomMu_6QBD
#> #define rxnbinom _rx9c54122483901b6b54a4545874ccacf14285_rxnbinom_Uzzs
#> #define rxbinom _rx9c54122483901b6b54a4545874ccacf14286_rxbinom_oz5n
#> #define rxnorm _rx9c54122483901b6b54a4545874ccacf14287_rxnorm_36W2
#> #define linCmtB _rx9c54122483901b6b54a4545874ccacf14288_linCmtB_2yJ1
#> #define linCmtA _rx9c54122483901b6b54a4545874ccacf14289_linCmtA_Kelq
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){  if (_solveData->svar[_svari] == 0) {alpha = _PP[0];};   if (_solveData->svar[_svari] == 1) {_rxNotFun_beta = _PP[1];}; }
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){  if (_solveData->ovar[_ovari] == 0) {alpha = _PP[0];};   if (_solveData->ovar[_ovari] == 1) {_rxNotFun_beta = _PP[1];}; }
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  rx_bf085b0932f776bdf68efc2bb27f8afc__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *rx_bf085b0932f776bdf68efc2bb27f8afc__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP rx_bf085b0932f776bdf68efc2bb27f8afc__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void rx_bf085b0932f776bdf68efc2bb27f8afc__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double alpha = NA_REAL;
#>   double _rxNotFun_beta = NA_REAL;
#>   double fx = NA_REAL;
#>   double dAlpha = NA_REAL;
#>   double dBeta = NA_REAL;
#> 
#>   (void)t;
#>   (void)alpha;
#>   (void)_rxNotFun_beta;
#>   (void)fx;
#>   (void)dAlpha;
#>   (void)dBeta;
#> 
#>   fx = _PL[0];
#>   dAlpha = _PL[1];
#>   dBeta = _PL[2];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#>   alpha = _PP[0];
#>   _rxNotFun_beta = _PP[1];
#> 
#> 
#>   fx =llikUnif(t,alpha,_rxNotFun_beta);
#>   dAlpha =llikUnifDalpha(t,alpha,_rxNotFun_beta);
#>   dBeta =llikUnifDbeta(t,alpha,_rxNotFun_beta);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void rx_bf085b0932f776bdf68efc2bb27f8afc__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void rx_bf085b0932f776bdf68efc2bb27f8afc__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void rx_bf085b0932f776bdf68efc2bb27f8afc__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double alpha = NA_REAL;
#>   double _rxNotFun_beta = NA_REAL;
#>   double fx = NA_REAL;
#>   double dAlpha = NA_REAL;
#>   double dBeta = NA_REAL;
#> 
#>   (void)t;
#>   (void)alpha;
#>   (void)_rxNotFun_beta;
#>   (void)fx;
#>   (void)dAlpha;
#>   (void)dBeta;
#> 
#>   fx = _PL[0];
#>   dAlpha = _PL[1];
#>   dBeta = _PL[2];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#>   alpha = _PP[0];
#>   _rxNotFun_beta = _PP[1];
#> 
#> 
#>   fx =llikUnif(t,alpha,_rxNotFun_beta);
#>   dAlpha =llikUnifDalpha(t,alpha,_rxNotFun_beta);
#>   dBeta =llikUnifDbeta(t,alpha,_rxNotFun_beta);
#> 
#>   _lhs[_LHS_0_]=fx;
#>   _lhs[_LHS_1_]=dAlpha;
#>   _lhs[_LHS_2_]=dBeta;
#> }
#> // Functional based bioavailability
#> double rx_bf085b0932f776bdf68efc2bb27f8afc__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double rx_bf085b0932f776bdf68efc2bb27f8afc__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double rx_bf085b0932f776bdf68efc2bb27f8afc__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double rx_bf085b0932f776bdf68efc2bb27f8afc__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void rx_bf085b0932f776bdf68efc2bb27f8afc__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void rx_bf085b0932f776bdf68efc2bb27f8afc__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void rx_bf085b0932f776bdf68efc2bb27f8afc__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP rx_bf085b0932f776bdf68efc2bb27f8afc__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_bf085b0932f776bdf68efc2bb27f8afc__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAABt3AAAv7#aT)4q1%%NA]>9,h\";*Tu1)=J=qaErnbmA`g#&<sf,3]sgho$eeW=MzAn4a)|MmYE|dv];,mS)t(Blb7&UzdRSJVQVcCQvLoINcIAOVIY&E\"Y_%%_}0``.z47:Kcv^^4GUu#X8`m\"v8~~%%suE`w0:)ew;<B@&e7s#@bN.r]X8nk^dPE323Ux:GG7tiP.1x^j8iG($I^DIT/sP8T6E#km!Nla)j%%/^fBivSLT%%8V^E1\?G(tkh[W^W3bdmCT.&/z3.fsWJRcVk4,{SB)GsXdZeC=l5~0o,!z@jadJ8GYv0u}hP){*2epK%%}5NB;]Xgp0,>VrF+)kCdZo*<b|;CcScc<pW{t@P>KpRJ@a~/l<|8oI#Iq\?,E&\"pNd`e(jc<Y*Z.b#MRT+Gt+z>fuNCFm6cDz>MOKHNEh$%%kb\?s7C%%<n6ntM[A1@49Hnu)I]GaFVR]}Cm/FI<YnhySk+<>6k^3v^{%%iW>HdfLA@mc43A&[|o:mnc+yz`D{r7f9aWNsgWyOYsp=]ST62&|;rAXgP;LQRGz/P}r^6/Pe~!6H4bshH}r%%@g5qVolYD@F!*0cuCO5p/qFeONHQX$k75/eZ6AqLq|C,x3rnId=wW~WwFZ`4m\?v~pungq{)j[jq{#x>=0;`+{.QdUcQ5$H$,[}v{c`/UNI&r8dVrdD_\"WX8eYe~j]@6|grN%%%%1,Yc[Od{@X5;F>|:i_6GyvK\?ZOmeb~\"Ml/.8nhO,*kk3jUIlK)w)U}zI`I[0G<z2\?0o,.,_@6:Uccl#fZky+bj)anx%%ZLFyuo+^y7ap1xVfxFJ1X_QAMRwDz^j$+A+v11|\"P_j2(+DYs7Z%%|3w!\"z;[hynVs0S.uQDv1RsD%%2j\?p#W#d.c|=9;tC@,xFLz<zs,Nu`@uUVvbm6)I<{<s1v+R.=5(7fL|f{);Q[r*!3(DzA@iq@oC~!3\?S6SezC<;=My(Xm6q+&jerAO;WMAzWbD`+O;u~mSEYo@[O8.=Tp\?fz5S0Fl@jt/86th@Q9E]_2`>oZ\"nLX(tZ}Y&+}NDI+7d<8WEI5xJGw`#y{_n7]:a.A=q!baBx\?lE`\?PM^%%WG)g:WPifT:}eU)a\"dKK~:`KTJ*X[<c");
#>     char buf[1142];
#> #define __doBufN__ 1142
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void rx_bf085b0932f776bdf68efc2bb27f8afc__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   rx_bf085b0932f776bdf68efc2bb27f8afc__dydt(neq, *t, A, DADT);
#> }
#> extern int rx_bf085b0932f776bdf68efc2bb27f8afc__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   rx_bf085b0932f776bdf68efc2bb27f8afc__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void rx_bf085b0932f776bdf68efc2bb27f8afc__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   rx_bf085b0932f776bdf68efc2bb27f8afc__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void rx_bf085b0932f776bdf68efc2bb27f8afc__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_rx_bf085b0932f776bdf68efc2bb27f8afc_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("rx_bf085b0932f776bdf68efc2bb27f8afc_","rx_bf085b0932f776bdf68efc2bb27f8afc__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("rx_bf085b0932f776bdf68efc2bb27f8afc_","rx_bf085b0932f776bdf68efc2bb27f8afc__assignFuns", (DL_FUNC) rx_bf085b0932f776bdf68efc2bb27f8afc__assignFuns);
#>   R_RegisterCCallable("rx_bf085b0932f776bdf68efc2bb27f8afc_","rx_bf085b0932f776bdf68efc2bb27f8afc__inis",(DL_FUNC) rx_bf085b0932f776bdf68efc2bb27f8afc__inis);
#>   R_RegisterCCallable("rx_bf085b0932f776bdf68efc2bb27f8afc_","rx_bf085b0932f776bdf68efc2bb27f8afc__dydt",(DL_FUNC) rx_bf085b0932f776bdf68efc2bb27f8afc__dydt);
#>   R_RegisterCCallable("rx_bf085b0932f776bdf68efc2bb27f8afc_","rx_bf085b0932f776bdf68efc2bb27f8afc__calc_lhs",(DL_FUNC) rx_bf085b0932f776bdf68efc2bb27f8afc__calc_lhs);
#>   R_RegisterCCallable("rx_bf085b0932f776bdf68efc2bb27f8afc_","rx_bf085b0932f776bdf68efc2bb27f8afc__calc_jac",(DL_FUNC) rx_bf085b0932f776bdf68efc2bb27f8afc__calc_jac);
#>   R_RegisterCCallable("rx_bf085b0932f776bdf68efc2bb27f8afc_","rx_bf085b0932f776bdf68efc2bb27f8afc__dydt_lsoda", (DL_FUNC) rx_bf085b0932f776bdf68efc2bb27f8afc__dydt_lsoda);
#>   R_RegisterCCallable("rx_bf085b0932f776bdf68efc2bb27f8afc_","rx_bf085b0932f776bdf68efc2bb27f8afc__calc_jac_lsoda", (DL_FUNC) rx_bf085b0932f776bdf68efc2bb27f8afc__calc_jac_lsoda);
#>   R_RegisterCCallable("rx_bf085b0932f776bdf68efc2bb27f8afc_","rx_bf085b0932f776bdf68efc2bb27f8afc__ode_solver_solvedata", (DL_FUNC) rx_bf085b0932f776bdf68efc2bb27f8afc__ode_solver_solvedata);
#>   R_RegisterCCallable("rx_bf085b0932f776bdf68efc2bb27f8afc_","rx_bf085b0932f776bdf68efc2bb27f8afc__ode_solver_get_solvedata", (DL_FUNC) rx_bf085b0932f776bdf68efc2bb27f8afc__ode_solver_get_solvedata);
#>   R_RegisterCCallable("rx_bf085b0932f776bdf68efc2bb27f8afc_","rx_bf085b0932f776bdf68efc2bb27f8afc__F", (DL_FUNC) rx_bf085b0932f776bdf68efc2bb27f8afc__F);
#>   R_RegisterCCallable("rx_bf085b0932f776bdf68efc2bb27f8afc_","rx_bf085b0932f776bdf68efc2bb27f8afc__Lag", (DL_FUNC) rx_bf085b0932f776bdf68efc2bb27f8afc__Lag);
#>   R_RegisterCCallable("rx_bf085b0932f776bdf68efc2bb27f8afc_","rx_bf085b0932f776bdf68efc2bb27f8afc__Rate", (DL_FUNC) rx_bf085b0932f776bdf68efc2bb27f8afc__Rate);
#>   R_RegisterCCallable("rx_bf085b0932f776bdf68efc2bb27f8afc_","rx_bf085b0932f776bdf68efc2bb27f8afc__Dur", (DL_FUNC) rx_bf085b0932f776bdf68efc2bb27f8afc__Dur);
#>   R_RegisterCCallable("rx_bf085b0932f776bdf68efc2bb27f8afc_","rx_bf085b0932f776bdf68efc2bb27f8afc__mtime", (DL_FUNC) rx_bf085b0932f776bdf68efc2bb27f8afc__mtime);
#>   R_RegisterCCallable("rx_bf085b0932f776bdf68efc2bb27f8afc_","rx_bf085b0932f776bdf68efc2bb27f8afc__ME", (DL_FUNC) rx_bf085b0932f776bdf68efc2bb27f8afc__ME);
#>   R_RegisterCCallable("rx_bf085b0932f776bdf68efc2bb27f8afc_","rx_bf085b0932f776bdf68efc2bb27f8afc__IndF", (DL_FUNC) rx_bf085b0932f776bdf68efc2bb27f8afc__IndF);
#>   R_RegisterCCallable("rx_bf085b0932f776bdf68efc2bb27f8afc_","rx_bf085b0932f776bdf68efc2bb27f8afc__dydt_liblsoda", (DL_FUNC) rx_bf085b0932f776bdf68efc2bb27f8afc__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_rx_bf085b0932f776bdf68efc2bb27f8afc_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_rx_bf085b0932f776bdf68efc2bb27f8afc_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"rx_bf085b0932f776bdf68efc2bb27f8afc__model_vars", (DL_FUNC) &rx_bf085b0932f776bdf68efc2bb27f8afc__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_rx_bf085b0932f776bdf68efc2bb27f8afc_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_bf085b0932f776bdf68efc2bb27f8afc__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("rx_bf085b0932f776bdf68efc2bb27f8afc__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’
#> ── Solved rxode2 object ──
#> ── Parameters (value$params): ──
#> # A tibble: 1 × 0
#> ── Initial Conditions (value$inits): ──
#> named numeric(0)
#> ── First part of data (object): ──
#> # A tibble: 4 × 6
#>    time    fx dAlpha dBeta alpha  beta
#>   <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#> 1     1 -1.39   0.25 -0.25    -2     2
#> 2     1 -1.39   0.25 -0.25    -2     2
#> 3     1 -1.39   0.25 -0.25    -2     2
#> 4     1 -1.39   0.25 -0.25    -2     2
# }