log likelihood and derivatives for Weibull distribution

log likelihood and derivatives for Weibull distribution

Usage

llikWeibull(x, shape, scale, full = FALSE)

Arguments

x

variable distributed by a Weibull distribution

shape, scale

shape and scale parameters, the latter defaulting to 1.

full

Add the data frame showing x, mean, sd as well as the fx and derivatives

Value

data frame with fx for the log pdf value of with dProb that has the derivatives with respect to the prob parameters at the observation time-point

Details

In an rxode2() model, you can use llikWeibull() but you have to use the x and rate arguments. You can also get the derivative of shape or scale with llikWeibullDshape() and llikWeibullDscale().

Author

Matthew L. Fidler

Examples

# \donttest{
llikWeibull(1, 1, 10)
#>          fx    dShape dScale
#> 1 -2.402585 -1.072327  -0.09

# rxode2 can use this too:

et  <- et(seq(0.001, 1, length.out=10))
et$shape <- 1
et$scale <- 10
 
model <- function() {
  model({
    fx <- llikWeibull(time, shape, scale)
    dShape<- llikWeibullDshape(time, shape, scale)
    dScale <- llikWeibullDscale(time, shape, scale)
  })
}

rxSolve(model, et)
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#>  
#>  
#> C model:
#> // Define 3 LHS values
#> #define _LHS_0_ 0 // fx
#> #define _LHS_1_ 1 // dShape
#> #define _LHS_2_ 2 // dScale
#> #define _getRxSolve_ _rxa7ee786d0578b339c8a662567a5d56064290__getRxSolve__l3zY
#> #define _evalUdf _rxa7ee786d0578b339c8a662567a5d56064291__evalUdf_vTa9
#> #define _solveData _rxa7ee786d0578b339c8a662567a5d56064292__solveData_7EhW
#> #define _assign_ptr _rxa7ee786d0578b339c8a662567a5d56064293__assign_ptr_3MG9
#> #define _rxRmModelLib _rxa7ee786d0578b339c8a662567a5d56064294__rxRmModelLib_V3pR
#> #define _rxGetModelLib _rxa7ee786d0578b339c8a662567a5d56064295__rxGetModelLib_1aNk
#> #define _old_c _rxa7ee786d0578b339c8a662567a5d56064296__old_c_G5Jw
#> #define _ptrid _rxa7ee786d0578b339c8a662567a5d56064297__ptrid_D9TK
#> #define _rxIsCurrentC _rxa7ee786d0578b339c8a662567a5d56064298__rxIsCurrentC_40NP
#> #define _sumPS _rxa7ee786d0578b339c8a662567a5d56064299__sumPS_JiAL
#> #define _prodPS _rxa7ee786d0578b339c8a662567a5d56064300__prodPS_ZBrh
#> #define _prodType _rxa7ee786d0578b339c8a662567a5d56064301__prodType_efgh
#> #define _sumType _rxa7ee786d0578b339c8a662567a5d56064302__sumType_Drm8
#> #define _update_par_ptr _rxa7ee786d0578b339c8a662567a5d56064303__update_par_ptr_5igv
#> #define _getParCov _rxa7ee786d0578b339c8a662567a5d56064304__getParCov_lJYg
#> #define _rxode2_rxAssignPtr _rxa7ee786d0578b339c8a662567a5d56064305__rxode2_rxAssignPtr_GqQx
#> #define _rxQr _rxa7ee786d0578b339c8a662567a5d56064306__rxQr_LoYl
#> #define _compareFactorVal _rxa7ee786d0578b339c8a662567a5d56064307__compareFactorVal_FgJd
#> #define _sum _rxa7ee786d0578b339c8a662567a5d56064308__sum_KaFg
#> #define _udf _rxa7ee786d0578b339c8a662567a5d56064309__udf_mDR4
#> #define _sign _rxa7ee786d0578b339c8a662567a5d56064310__sign_jxuo
#> #define _prod _rxa7ee786d0578b339c8a662567a5d56064311__prod_P8gm
#> #define _max _rxa7ee786d0578b339c8a662567a5d56064312__max_oUom
#> #define _min _rxa7ee786d0578b339c8a662567a5d56064313__min_wLtu
#> #define _transit4P _rxa7ee786d0578b339c8a662567a5d56064314__transit4P_DgOz
#> #define _transit3P _rxa7ee786d0578b339c8a662567a5d56064315__transit3P_tPZw
#> #define _assignFuns0 _rxa7ee786d0578b339c8a662567a5d56064316__assignFuns0_Dqx2
#> #define _assignFuns _rxa7ee786d0578b339c8a662567a5d56064317__assignFuns_t0Ff
#> #define _rxord _rxa7ee786d0578b339c8a662567a5d56064318__rxord_CRup
#> #define __assignFuns2 _rxa7ee786d0578b339c8a662567a5d56064319___assignFuns2_BfcL
#> #define _llikCauchyDscale _rxa7ee786d0578b339c8a662567a5d56064320__llikCauchyDscale_LlJk
#> #define _llikCauchyDlocation _rxa7ee786d0578b339c8a662567a5d56064321__llikCauchyDlocation_0HOu
#> #define _llikCauchy _rxa7ee786d0578b339c8a662567a5d56064322__llikCauchy_quMQ
#> #define _llikGammaDrate _rxa7ee786d0578b339c8a662567a5d56064323__llikGammaDrate_IP4s
#> #define _llikGammaDshape _rxa7ee786d0578b339c8a662567a5d56064324__llikGammaDshape_BKwl
#> #define _llikGamma _rxa7ee786d0578b339c8a662567a5d56064325__llikGamma_ION1
#> #define _llikWeibullDscale _rxa7ee786d0578b339c8a662567a5d56064326__llikWeibullDscale_tTX5
#> #define _llikWeibullDshape _rxa7ee786d0578b339c8a662567a5d56064327__llikWeibullDshape_DulH
#> #define _llikWeibull _rxa7ee786d0578b339c8a662567a5d56064328__llikWeibull_LxeA
#> #define _llikUnifDbeta _rxa7ee786d0578b339c8a662567a5d56064329__llikUnifDbeta_Ut0p
#> #define _llikUnifDalpha _rxa7ee786d0578b339c8a662567a5d56064330__llikUnifDalpha_o6Aq
#> #define _llikUnif _rxa7ee786d0578b339c8a662567a5d56064331__llikUnif_SQkn
#> #define _llikGeomDp _rxa7ee786d0578b339c8a662567a5d56064332__llikGeomDp_H8fP
#> #define _llikGeom _rxa7ee786d0578b339c8a662567a5d56064333__llikGeom_CAnN
#> #define _llikFDdf2 _rxa7ee786d0578b339c8a662567a5d56064334__llikFDdf2_kUms
#> #define _llikFDdf1 _rxa7ee786d0578b339c8a662567a5d56064335__llikFDdf1_wiOW
#> #define _llikF _rxa7ee786d0578b339c8a662567a5d56064336__llikF_G16H
#> #define _llikExpDrate _rxa7ee786d0578b339c8a662567a5d56064337__llikExpDrate_PUgn
#> #define _llikExp _rxa7ee786d0578b339c8a662567a5d56064338__llikExp_Cjuw
#> #define _llikChisqDdf _rxa7ee786d0578b339c8a662567a5d56064339__llikChisqDdf_A8iw
#> #define _llikChisq _rxa7ee786d0578b339c8a662567a5d56064340__llikChisq_bn2G
#> #define _llikTDsd _rxa7ee786d0578b339c8a662567a5d56064341__llikTDsd_TQjS
#> #define _llikTDmean _rxa7ee786d0578b339c8a662567a5d56064342__llikTDmean_rer5
#> #define _llikTDdf _rxa7ee786d0578b339c8a662567a5d56064343__llikTDdf_zexL
#> #define _llikT _rxa7ee786d0578b339c8a662567a5d56064344__llikT_gAwM
#> #define _llikBetaDshape2 _rxa7ee786d0578b339c8a662567a5d56064345__llikBetaDshape2_E6nq
#> #define _llikBetaDshape1 _rxa7ee786d0578b339c8a662567a5d56064346__llikBetaDshape1_R3DD
#> #define _llikBeta _rxa7ee786d0578b339c8a662567a5d56064347__llikBeta_9jN5
#> #define _llikNbinomMuDmu _rxa7ee786d0578b339c8a662567a5d56064348__llikNbinomMuDmu_O4r9
#> #define _llikNbinomMu _rxa7ee786d0578b339c8a662567a5d56064349__llikNbinomMu_D97m
#> #define _llikNbinomDprob _rxa7ee786d0578b339c8a662567a5d56064350__llikNbinomDprob_e5yb
#> #define _llikNbinom _rxa7ee786d0578b339c8a662567a5d56064351__llikNbinom_V8Eu
#> #define _llikBinomDprob _rxa7ee786d0578b339c8a662567a5d56064352__llikBinomDprob_Z1xT
#> #define _llikBinom _rxa7ee786d0578b339c8a662567a5d56064353__llikBinom_QYOQ
#> #define _llikPoisDlambda _rxa7ee786d0578b339c8a662567a5d56064354__llikPoisDlambda_e6aq
#> #define _llikPois _rxa7ee786d0578b339c8a662567a5d56064355__llikPois_t3Lb
#> #define _llikNormDsd _rxa7ee786d0578b339c8a662567a5d56064356__llikNormDsd_sbea
#> #define _llikNormDmean _rxa7ee786d0578b339c8a662567a5d56064357__llikNormDmean_w5m4
#> #define _llikNorm _rxa7ee786d0578b339c8a662567a5d56064358__llikNorm_brD9
#> #define simeps _rxa7ee786d0578b339c8a662567a5d56064359_simeps_EZAJ
#> #define simeta _rxa7ee786d0578b339c8a662567a5d56064360_simeta_Ne7I
#> #define expit _rxa7ee786d0578b339c8a662567a5d56064361_expit_xUFA
#> #define logit _rxa7ee786d0578b339c8a662567a5d56064362_logit_5R2w
#> #define gammapDer _rxa7ee786d0578b339c8a662567a5d56064363_gammapDer_ZwwB
#> #define lowergamma _rxa7ee786d0578b339c8a662567a5d56064364_lowergamma_gRvb
#> #define uppergamma _rxa7ee786d0578b339c8a662567a5d56064365_uppergamma_lnk3
#> #define gammaqInva _rxa7ee786d0578b339c8a662567a5d56064366_gammaqInva_57cf
#> #define gammaqInv _rxa7ee786d0578b339c8a662567a5d56064367_gammaqInv_UwMa
#> #define gammapInva _rxa7ee786d0578b339c8a662567a5d56064368_gammapInva_rEA9
#> #define gammapInv _rxa7ee786d0578b339c8a662567a5d56064369_gammapInv_z1SE
#> #define gammaq _rxa7ee786d0578b339c8a662567a5d56064370_gammaq_1KzA
#> #define gammap _rxa7ee786d0578b339c8a662567a5d56064371_gammap_f2H4
#> #define phi _rxa7ee786d0578b339c8a662567a5d56064372_phi_Q2oF
#> #define d2ELUa _rxa7ee786d0578b339c8a662567a5d56064373_d2ELUa_0joC
#> #define dELUa _rxa7ee786d0578b339c8a662567a5d56064374_dELUa_Gj0a
#> #define d2aELU _rxa7ee786d0578b339c8a662567a5d56064375_d2aELU_8wAb
#> #define d2ELU _rxa7ee786d0578b339c8a662567a5d56064376_d2ELU_QELO
#> #define dELU _rxa7ee786d0578b339c8a662567a5d56064377_dELU_lObq
#> #define ELU _rxa7ee786d0578b339c8a662567a5d56064378_ELU_Nath
#> #define dPReLUa1 _rxa7ee786d0578b339c8a662567a5d56064379_dPReLUa1_ltiW
#> #define dPReLUa _rxa7ee786d0578b339c8a662567a5d56064380_dPReLUa_R2IG
#> #define dPReLU _rxa7ee786d0578b339c8a662567a5d56064381_dPReLU_VLQY
#> #define PReLU _rxa7ee786d0578b339c8a662567a5d56064382_PReLU_M84J
#> #define dSwish _rxa7ee786d0578b339c8a662567a5d56064383_dSwish_Cnuy
#> #define Swish _rxa7ee786d0578b339c8a662567a5d56064384_Swish_CXPJ
#> #define dlReLU _rxa7ee786d0578b339c8a662567a5d56064385_dlReLU_UjvF
#> #define lReLU _rxa7ee786d0578b339c8a662567a5d56064386_lReLU_k7KJ
#> #define dSELU _rxa7ee786d0578b339c8a662567a5d56064387_dSELU_vDxq
#> #define SELU _rxa7ee786d0578b339c8a662567a5d56064388_SELU_bVSK
#> #define d4softplus _rxa7ee786d0578b339c8a662567a5d56064389_d4softplus_Vfr7
#> #define d3softplus _rxa7ee786d0578b339c8a662567a5d56064390_d3softplus_TcaU
#> #define d2softplus _rxa7ee786d0578b339c8a662567a5d56064391_d2softplus_EgpY
#> #define dsoftplus _rxa7ee786d0578b339c8a662567a5d56064392_dsoftplus_kNIq
#> #define softplus _rxa7ee786d0578b339c8a662567a5d56064393_softplus_nEnm
#> #define d4GELU _rxa7ee786d0578b339c8a662567a5d56064394_d4GELU_hSbo
#> #define d3GELU _rxa7ee786d0578b339c8a662567a5d56064395_d3GELU_MZYB
#> #define d2GELU _rxa7ee786d0578b339c8a662567a5d56064396_d2GELU_jkiC
#> #define dGELU _rxa7ee786d0578b339c8a662567a5d56064397_dGELU_oUqv
#> #define GELU _rxa7ee786d0578b339c8a662567a5d56064398_GELU_niI7
#> #define dReLU _rxa7ee786d0578b339c8a662567a5d56064399_dReLU_5Ui0
#> #define ReLU _rxa7ee786d0578b339c8a662567a5d56064400_ReLU_GVTs
#> #define riweibull _rxa7ee786d0578b339c8a662567a5d56064401_riweibull_34TN
#> #define riunif _rxa7ee786d0578b339c8a662567a5d56064402_riunif_HeFp
#> #define rit_ _rxa7ee786d0578b339c8a662567a5d56064403_rit__ue1Z
#> #define ripois _rxa7ee786d0578b339c8a662567a5d56064404_ripois_aEnJ
#> #define ribeta _rxa7ee786d0578b339c8a662567a5d56064405_ribeta_R6UH
#> #define rigamma _rxa7ee786d0578b339c8a662567a5d56064406_rigamma_ZKbS
#> #define rigeom _rxa7ee786d0578b339c8a662567a5d56064407_rigeom_smNY
#> #define rif _rxa7ee786d0578b339c8a662567a5d56064408_rif_lS2w
#> #define riexp _rxa7ee786d0578b339c8a662567a5d56064409_riexp_zYqw
#> #define richisq _rxa7ee786d0578b339c8a662567a5d56064410_richisq_2z8b
#> #define ricauchy _rxa7ee786d0578b339c8a662567a5d56064411_ricauchy_GMdj
#> #define rinbinomMu _rxa7ee786d0578b339c8a662567a5d56064412_rinbinomMu_LB3L
#> #define rinbinom _rxa7ee786d0578b339c8a662567a5d56064413_rinbinom_sOUR
#> #define ribinom _rxa7ee786d0578b339c8a662567a5d56064414_ribinom_bPYU
#> #define rinorm _rxa7ee786d0578b339c8a662567a5d56064415_rinorm_PmW8
#> #define rxweibull _rxa7ee786d0578b339c8a662567a5d56064416_rxweibull_sOmG
#> #define rxunif _rxa7ee786d0578b339c8a662567a5d56064417_rxunif_ypO5
#> #define rxt_ _rxa7ee786d0578b339c8a662567a5d56064418_rxt__5OnO
#> #define rxpois _rxa7ee786d0578b339c8a662567a5d56064419_rxpois_E9TF
#> #define rxbeta _rxa7ee786d0578b339c8a662567a5d56064420_rxbeta_yOK7
#> #define rxgamma _rxa7ee786d0578b339c8a662567a5d56064421_rxgamma_OqEK
#> #define rxgeom _rxa7ee786d0578b339c8a662567a5d56064422_rxgeom_wH4T
#> #define rxf _rxa7ee786d0578b339c8a662567a5d56064423_rxf_tSJj
#> #define rxexp _rxa7ee786d0578b339c8a662567a5d56064424_rxexp_0v8j
#> #define rxchisq _rxa7ee786d0578b339c8a662567a5d56064425_rxchisq_xwnn
#> #define rxcauchy _rxa7ee786d0578b339c8a662567a5d56064426_rxcauchy_6QBD
#> #define rxnbinomMu _rxa7ee786d0578b339c8a662567a5d56064427_rxnbinomMu_Uzzs
#> #define rxnbinom _rxa7ee786d0578b339c8a662567a5d56064428_rxnbinom_oz5n
#> #define rxbinom _rxa7ee786d0578b339c8a662567a5d56064429_rxbinom_36W2
#> #define rxnorm _rxa7ee786d0578b339c8a662567a5d56064430_rxnorm_2yJ1
#> #define linCmtB _rxa7ee786d0578b339c8a662567a5d56064431_linCmtB_Kelq
#> #define linCmtA _rxa7ee786d0578b339c8a662567a5d56064432_linCmtA_u4n1
#> #include <rxode2_model_shared.h>
#> #define __MAX_PROD__ 0
#> #define _CMT CMT
#> #define _SYNC_simeps_ for (int _svari=_solveData->neps; _svari--;){  if (_solveData->svar[_svari] == 0) {shape = _PP[0];};   if (_solveData->svar[_svari] == 1) {scale = _PP[1];}; }
#> #define _SYNC_simeta_ for (int _ovari=_solveData->neta; _ovari--;){  if (_solveData->ovar[_ovari] == 0) {shape = _PP[0];};   if (_solveData->ovar[_ovari] == 1) {scale = _PP[1];}; }
#> _getRxSolve_t _getRxSolve_;
#> _simfun simeps;
#> _simfun simeta;
#> _udf_type _evalUdf=NULL;
#> rx_solve *_solveData=NULL;
#> rxode2_assign_ptr _assign_ptr=NULL;
#> _rxRmModelLibType _rxRmModelLib=NULL;
#> _rxGetModelLibType _rxGetModelLib=NULL;
#> rxode2_ode_solver_old_c _old_c=NULL;
#> rxode2_fn0i _ptrid=NULL;
#> _rxIsCurrentC_type _rxIsCurrentC=NULL;
#> _rxSumType _sumPS=NULL;
#> _rxProdType _prodPS=NULL;
#> rxode2_fn0i _prodType=NULL;
#> rxode2_fn0i _sumType=NULL;
#> _update_par_ptr_p _update_par_ptr=NULL;
#> _getParCov_p _getParCov=NULL;
#> linCmtA_p linCmtA;
#> linCmtB_p linCmtB;
#> _rx_asgn _rxode2_rxAssignPtr=NULL;
#> _rx_asgn _rxQr=NULL;
#> rxode2_fn phi;
#> rxode2_fn ReLU;
#> rxode2_fn dReLU;
#> rxode2_fn GELU;
#> rxode2_fn dGELU;
#> rxode2_fn d2GELU;
#> rxode2_fn d3GELU;
#> rxode2_fn d4GELU;
#> rxode2_fn softplus;
#> rxode2_fn dsoftplus;
#> rxode2_fn d2softplus;
#> rxode2_fn d3softplus;
#> rxode2_fn d4softplus;
#> rxode2_fn SELU;
#> rxode2_fn dSELU;
#> rxode2_fn lReLU;
#> rxode2_fn dlReLU;
#> rxode2_fn Swish;
#> rxode2_fn dSwish;
#> rxode2_fn2 PReLU;
#> rxode2_fn2 dPReLU;
#> rxode2_fn2 dPReLUa;
#> rxode2_fn2 dPReLUa1;
#> rxode2_fn2 ELU;
#> rxode2_fn2 dELU;
#> rxode2_fn2 d2ELU;
#> rxode2_fn2 d2aELU;
#> rxode2_fn2 dELUa;
#> rxode2_fn2 d2ELUa;
#> rxode2_fn3 logit;
#> rxode2_fn3 expit;
#> rxode2_fn2 gammap;
#> rxode2_fn2 gammaq;
#> rxode2_fn2 lowergamma;
#> rxode2_fn2 uppergamma;
#> rxode2_fn2 gammapInv;
#> rxode2_fn2 gammapDer;
#> rxode2_fn2 gammapInva;
#> rxode2_fn2 gammaqInv;
#> rxode2_fn2 gammaqInva;
#> rxode2i_fn2 rxnorm;
#> rxode2i_rxbinom rxbinom;
#> rxode2i_rxbinom rxnbinom;
#> rxode2i_rxbinom rxnbinomMu;
#> rxode2i_fn2 rxcauchy;
#> rxode2i_fn rxchisq;
#> rxode2i_fn rxexp;
#> rxode2i_fn2 rxf;
#> rxode2i_ifn rxgeom;
#> rxode2i_fn2 rxgamma;
#> rxode2i_fn2 rxbeta;
#> rxode2i_ifn rxpois;
#> rxode2i_fn rxt_;
#> rxode2i_fn2 rxunif;
#> rxode2i_fn2 rxweibull;
#> rxode2i2_fn2 rinorm;
#> rxode2i2_ribinom ribinom;
#> rxode2i2_ribinom rinbinom;
#> rxode2i2_ribinom rinbinomMu;
#> rxode2i2_fn2 ricauchy;
#> rxode2i2_fn richisq;
#> rxode2i2_fn riexp;
#> rxode2i2_fn2 rif;
#> rxode2i2_ifn rigeom;
#> rxode2i2_fn2 rigamma;
#> rxode2i2_fn2 ribeta;
#> rxode2i2_ifn ripois;
#> rxode2i2_fn rit_;
#> rxode2i2_fn2 riunif;
#> rxode2i2_fn2 riweibull;
#> rxode2_llikNormFun _llikNorm;
#> rxode2_llikNormFun _llikNormDmean;
#> rxode2_llikNormFun _llikNormDsd;
#> rxode2_llikPoisFun _llikPois;
#> rxode2_llikPoisFun _llikPoisDlambda;
#> rxode2_llikBinomFun _llikBinom;
#> rxode2_llikBinomFun _llikBinomDprob;
#> rxode2_llikBinomFun _llikNbinom;
#> rxode2_llikBinomFun _llikNbinomDprob;
#> rxode2_llikBinomFun _llikNbinomMu;
#> rxode2_llikBinomFun _llikNbinomMuDmu;
#> rxode2_llikBetaFun _llikBeta;
#> rxode2_llikBetaFun _llikBetaDshape1;
#> rxode2_llikBetaFun _llikBetaDshape2;
#> rxode2_llikTFun _llikT;
#> rxode2_llikTFun _llikTDdf;
#> rxode2_llikTFun _llikTDmean;
#> rxode2_llikTFun _llikTDsd;
#> rxode2_llikChisqFun _llikChisq;
#> rxode2_llikChisqFun _llikChisqDdf;
#> rxode2_llikExpFun _llikExp;
#> rxode2_llikExpFun _llikExpDrate;
#> rxode2_llikFFun _llikF;
#> rxode2_llikFFun _llikFDdf1;
#> rxode2_llikFFun _llikFDdf2;
#> rxode2_llikGeomFun _llikGeom;
#> rxode2_llikGeomFun _llikGeomDp;
#> rxode2_llikUnifFun _llikUnif;
#> rxode2_llikUnifFun _llikUnifDalpha;
#> rxode2_llikUnifFun _llikUnifDbeta;
#> rxode2_llikWeibullFun _llikWeibull;
#> rxode2_llikWeibullFun _llikWeibullDshape;
#> rxode2_llikWeibullFun _llikWeibullDscale;
#> rxode2_llikGammaFun _llikGamma;
#> rxode2_llikGammaFun _llikGammaDshape;
#> rxode2_llikGammaFun _llikGammaDrate;
#> rxode2_llikCauchyFun _llikCauchy;
#> rxode2_llikCauchyFun _llikCauchyDlocation;
#> rxode2_llikCauchyFun _llikCauchyDscale;
#> rxode2_compareFactorVal_fn _compareFactorVal;
#> #include "extraC.h"
#> double _prod(double *input, double *p, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _prodPS(input, p, n, type);
#> }
#> double _udf(const char *funName, double *input, int n, ...) {
#>   if (n == -42) Rf_error("%s", "this has a ui user function that cannot be called directly");
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   return _evalUdf(funName, n, input);
#> }
#> double _sum(double *input, double *pld, int m, int type, int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   for (unsigned int i = 0; i < n; i++){
#>     input[i] = va_arg(valist, double);
#>   }
#>   va_end(valist);
#>   double ret = _sumPS(input, n, pld, m, type);
#>   if (type == 2 && m < 0){
#>     for (int i = -m; i--;){
#>       pld[i] = 0.0;
#>     }
#>   }
#>   return ret;
#> }
#> double _sign(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double s = 1;
#>   for (unsigned int i = 0; i < n; i++) {
#>     s = sign(va_arg(valist, double))*s;
#>     if (s == 0){
#>       break;
#>     }
#>   }
#>   va_end(valist);
#>   return s;
#> }
#> double _rxord(int _cSub, unsigned int n,  ...) {
#>   rx_solving_options_ind* ind = &(_solveData->subjects[_cSub]);
#>   if (!ind->inLhs) {
#>     return 1.0;
#>   }
#>   va_list valist;
#>   va_start(valist, n);
#>   double ret = 1.0;
#>   double p = 0.0;
#>   double u = rxunif(ind, 0.0, 1.0);
#>   int found = 0;
#>   for (unsigned int i = 0; i < n; i++) {
#>     p += va_arg(valist, double);
#>     if (!found) {
#>       if (u < p) {
#>         ret = (double)(i+1);
#>         found = 1;
#>       }
#>     }
#>   }
#>   if (!found) ret =(double)(n+1);
#>   va_end(valist);
#>   return ret;
#> }
#> double _max(unsigned int n, ...) {
#>   va_list valist;
#>   va_start(valist, n);
#>   double mx = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mx = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++) {
#>       tmp = va_arg(valist, double);
#>       if (tmp>mx) mx=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mx;
#> }
#> double _min(unsigned int n, ...){
#>   va_list valist;
#>   va_start(valist, n);
#>   double mn = NA_REAL;
#>   double tmp = 0;
#>   if (n >= 1){
#>     mn = va_arg(valist, double);
#>     for (unsigned int i = 1; i < n; i++){
#>       tmp = va_arg(valist, double);
#>       if (tmp<mn) mn=tmp;
#>     }
#>     va_end(valist);
#>   }
#>   return mn;
#> }
#> double _transit4P(int cmt, double t, unsigned int id, double n, double mtt, double bio){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   double dose = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(dose)) dose = 0.0;
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = (t-tlast);
#>   return exp(_safe_log(bio*dose)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> double _transit3P(int cmt, double t, unsigned int id, double n, double mtt){
#>   double nd = (double) n;
#>   double ktr = (nd+1)/mtt;
#>   double lktr = _safe_log(nd+1)-_safe_log(mtt);
#>   double tlast = _solveData->subjects[id].tlastS[cmt];
#>   if (ISNA(tlast)) tlast = 0.0;
#>   double tad = t-tlast;
#>   double podo = _solveData->subjects[id].curDoseS[cmt];
#>   if (ISNA(podo)) podo = 0.0;
#>   return exp(_safe_log(podo)+lktr+n*(lktr+_safe_log(tad))-ktr*(tad)-lgamma1p(nd));
#> }
#> void _assignFuns0(void) {
#>   _evalUdf = (_udf_type) R_GetCCallable("rxode2", "_rxode2_evalUdf");
#>   _getRxSolve_ = (_getRxSolve_t) R_GetCCallable("rxode2","getRxSolve_");
#>   _assign_ptr=(rxode2_assign_ptr) R_GetCCallable("rxode2","rxode2_assign_fn_pointers");
#>   _rxRmModelLib=(_rxRmModelLibType) R_GetCCallable("rxode2","rxRmModelLib");
#>   _rxGetModelLib=(_rxGetModelLibType) R_GetCCallable("rxode2","rxGetModelLib");
#>   _rxode2_rxAssignPtr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxAssignPtr");
#>   _rxQr=(_rx_asgn)R_GetCCallable("rxode2","_rxode2_rxQr");
#>   _rxIsCurrentC = (_rxIsCurrentC_type)R_GetCCallable("rxode2","rxIsCurrentC");
#>   _sumPS  = (_rxSumType) R_GetCCallable("PreciseSums","PreciseSums_sum_r");
#>   _prodPS = (_rxProdType) R_GetCCallable("PreciseSums","PreciseSums_prod_r");
#>   _prodType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_prod_get");
#>   _sumType=(rxode2_fn0i)R_GetCCallable("PreciseSums", "PreciseSums_sum_get");
#>   _ptrid=(rxode2_fn0i)R_GetCCallable("rxode2", "rxode2_current_fn_pointer_id");
#>   _compareFactorVal=(rxode2_compareFactorVal_fn) R_GetCCallable("rxode2", "compareFactorVal");
#>   _update_par_ptr = (_update_par_ptr_p) R_GetCCallable("rxode2","_update_par_ptr");
#>   _getParCov = (_getParCov_p) R_GetCCallable("rxode2","_getParCov");
#>   // dynamic start
#>   _llikCauchyDscale = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDscale");
#>   _llikCauchyDlocation = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchyDlocation");
#>   _llikCauchy = (rxode2_llikCauchyFun) R_GetCCallable("rxode2ll", "rxLlikCauchy");
#>   _llikGammaDrate = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDrate");
#>   _llikGammaDshape = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGammaDshape");
#>   _llikGamma = (rxode2_llikGammaFun) R_GetCCallable("rxode2ll", "rxLlikGamma");
#>   _llikWeibullDscale = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDscale");
#>   _llikWeibullDshape = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibullDshape");
#>   _llikWeibull = (rxode2_llikWeibullFun) R_GetCCallable("rxode2ll", "rxLlikWeibull");
#>   _llikUnifDbeta = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDbeta");
#>   _llikUnifDalpha = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnifDalpha");
#>   _llikUnif = (rxode2_llikUnifFun) R_GetCCallable("rxode2ll", "rxLlikUnif");
#>   _llikGeomDp = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeomDp");
#>   _llikGeom = (rxode2_llikGeomFun) R_GetCCallable("rxode2ll", "rxLlikGeom");
#>   _llikFDdf2 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf2");
#>   _llikFDdf1 = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikFDdf1");
#>   _llikF = (rxode2_llikFFun) R_GetCCallable("rxode2ll", "rxLlikF");
#>   _llikExpDrate = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExpDrate");
#>   _llikExp = (rxode2_llikExpFun) R_GetCCallable("rxode2ll", "rxLlikExp");
#>   _llikChisqDdf = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisqDdf");
#>   _llikChisq = (rxode2_llikChisqFun) R_GetCCallable("rxode2ll", "rxLlikChisq");
#>   _llikTDsd = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDsd");
#>   _llikTDmean = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDmean");
#>   _llikTDdf = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikTDdf");
#>   _llikT = (rxode2_llikTFun) R_GetCCallable("rxode2ll", "rxLlikT");
#>   _llikBetaDshape2 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape2");
#>   _llikBetaDshape1 = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBetaDshape1");
#>   _llikBeta = (rxode2_llikBetaFun) R_GetCCallable("rxode2ll", "rxLlikBeta");
#>   _llikNbinomMuDmu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMuDmu");
#>   _llikNbinomMu = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomMu");
#>   _llikNbinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinomDprob");
#>   _llikNbinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikNbinom");
#>   _llikBinomDprob = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinomDprob");
#>   _llikBinom = (rxode2_llikBinomFun) R_GetCCallable("rxode2ll", "rxLlikBinom");
#>   _llikPoisDlambda = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPoisDlambda");
#>   _llikPois = (rxode2_llikPoisFun) R_GetCCallable("rxode2ll", "rxLlikPois");
#>   _llikNormDsd = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDsd");
#>   _llikNormDmean = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNormDmean");
#>   _llikNorm = (rxode2_llikNormFun) R_GetCCallable("rxode2ll", "rxLlikNorm");
#>   simeps = (_simfun) R_GetCCallable("rxode2", "simeps");
#>   simeta = (_simfun) R_GetCCallable("rxode2", "simeta");
#>   expit = (rxode2_fn3) R_GetCCallable("rxode2", "expit");
#>   logit = (rxode2_fn3) R_GetCCallable("rxode2", "logit");
#>   gammapDer = (rxode2_fn2) R_GetCCallable("rxode2", "gammapDer");
#>   lowergamma = (rxode2_fn2) R_GetCCallable("rxode2", "lowergamma");
#>   uppergamma = (rxode2_fn2) R_GetCCallable("rxode2", "uppergamma");
#>   gammaqInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInva");
#>   gammaqInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammaqInv");
#>   gammapInva = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInva");
#>   gammapInv = (rxode2_fn2) R_GetCCallable("rxode2", "gammapInv");
#>   gammaq = (rxode2_fn2) R_GetCCallable("rxode2", "gammaq");
#>   gammap = (rxode2_fn2) R_GetCCallable("rxode2", "gammap");
#>   phi = (rxode2_fn) R_GetCCallable("rxode2", "phi");
#>   d2ELUa = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELUa");
#>   dELUa = (rxode2_fn2) R_GetCCallable("rxode2", "dELUa");
#>   d2aELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2aELU");
#>   d2ELU = (rxode2_fn2) R_GetCCallable("rxode2", "d2ELU");
#>   dELU = (rxode2_fn2) R_GetCCallable("rxode2", "dELU");
#>   ELU = (rxode2_fn2) R_GetCCallable("rxode2", "ELU");
#>   dPReLUa1 = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa1");
#>   dPReLUa = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLUa");
#>   dPReLU = (rxode2_fn2) R_GetCCallable("rxode2", "dPReLU");
#>   PReLU = (rxode2_fn2) R_GetCCallable("rxode2", "PReLU");
#>   dSwish = (rxode2_fn) R_GetCCallable("rxode2", "dSwish");
#>   Swish = (rxode2_fn) R_GetCCallable("rxode2", "Swish");
#>   dlReLU = (rxode2_fn) R_GetCCallable("rxode2", "dlReLU");
#>   lReLU = (rxode2_fn) R_GetCCallable("rxode2", "lReLU");
#>   dSELU = (rxode2_fn) R_GetCCallable("rxode2", "dSELU");
#>   SELU = (rxode2_fn) R_GetCCallable("rxode2", "SELU");
#>   d4softplus = (rxode2_fn) R_GetCCallable("rxode2", "d4softplus");
#>   d3softplus = (rxode2_fn) R_GetCCallable("rxode2", "d3softplus");
#>   d2softplus = (rxode2_fn) R_GetCCallable("rxode2", "d2softplus");
#>   dsoftplus = (rxode2_fn) R_GetCCallable("rxode2", "dsoftplus");
#>   softplus = (rxode2_fn) R_GetCCallable("rxode2", "softplus");
#>   d4GELU = (rxode2_fn) R_GetCCallable("rxode2", "d4GELU");
#>   d3GELU = (rxode2_fn) R_GetCCallable("rxode2", "d3GELU");
#>   d2GELU = (rxode2_fn) R_GetCCallable("rxode2", "d2GELU");
#>   dGELU = (rxode2_fn) R_GetCCallable("rxode2", "dGELU");
#>   GELU = (rxode2_fn) R_GetCCallable("rxode2", "GELU");
#>   dReLU = (rxode2_fn) R_GetCCallable("rxode2", "dReLU");
#>   ReLU = (rxode2_fn) R_GetCCallable("rxode2", "ReLU");
#>   riweibull = (rxode2i2_fn2) R_GetCCallable("rxode2", "riweibull");
#>   riunif = (rxode2i2_fn2) R_GetCCallable("rxode2", "riunif");
#>   rit_ = (rxode2i2_fn) R_GetCCallable("rxode2", "rit_");
#>   ripois = (rxode2i2_ifn) R_GetCCallable("rxode2", "ripois");
#>   ribeta = (rxode2i2_fn2) R_GetCCallable("rxode2", "ribeta");
#>   rigamma = (rxode2i2_fn2) R_GetCCallable("rxode2", "rigamma");
#>   rigeom = (rxode2i2_ifn) R_GetCCallable("rxode2", "rigeom");
#>   rif = (rxode2i2_fn2) R_GetCCallable("rxode2", "rif");
#>   riexp = (rxode2i2_fn) R_GetCCallable("rxode2", "riexp");
#>   richisq = (rxode2i2_fn) R_GetCCallable("rxode2", "richisq");
#>   ricauchy = (rxode2i2_fn2) R_GetCCallable("rxode2", "ricauchy");
#>   rinbinomMu = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinomMu");
#>   rinbinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "rinbinom");
#>   ribinom = (rxode2i2_ribinom) R_GetCCallable("rxode2", "ribinom");
#>   rinorm = (rxode2i2_fn2) R_GetCCallable("rxode2", "rinorm");
#>   rxweibull = (rxode2i_fn2) R_GetCCallable("rxode2", "rxweibull");
#>   rxunif = (rxode2i_fn2) R_GetCCallable("rxode2", "rxunif");
#>   rxt_ = (rxode2i_fn) R_GetCCallable("rxode2", "rxt_");
#>   rxpois = (rxode2i_ifn) R_GetCCallable("rxode2", "rxpois");
#>   rxbeta = (rxode2i_fn2) R_GetCCallable("rxode2", "rxbeta");
#>   rxgamma = (rxode2i_fn2) R_GetCCallable("rxode2", "rxgamma");
#>   rxgeom = (rxode2i_ifn) R_GetCCallable("rxode2", "rxgeom");
#>   rxf = (rxode2i_fn2) R_GetCCallable("rxode2", "rxf");
#>   rxexp = (rxode2i_fn) R_GetCCallable("rxode2", "rxexp");
#>   rxchisq = (rxode2i_fn) R_GetCCallable("rxode2", "rxchisq");
#>   rxcauchy = (rxode2i_fn2) R_GetCCallable("rxode2", "rxcauchy");
#>   rxnbinomMu = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinomMu");
#>   rxnbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxnbinom");
#>   rxbinom = (rxode2i_rxbinom) R_GetCCallable("rxode2", "rxbinom");
#>   rxnorm = (rxode2i_fn2) R_GetCCallable("rxode2", "rxnorm");
#>   linCmtB = (linCmtB_p) R_GetCCallable("rxode2", "linCmtB");
#>   linCmtA = (linCmtA_p) R_GetCCallable("rxode2", "linCmtA");
#>   // dynamic stop
#>   _solveData = _getRxSolve_();
#> }
#> void _assignFuns(void) {
#>   if (_assign_ptr == NULL){
#>     _assignFuns0();
#>   }
#> }
#> void __assignFuns2(rx_solve rx,
#>                    rx_solving_options op,
#>                    t_F f,
#>                    t_LAG lag,
#>                    t_RATE rate,
#>                    t_DUR dur,
#>                    t_calc_mtime mtime,
#>                    t_ME me,
#>                    t_IndF indf,
#>                    t_getTime gettime,
#>                    t_locateTimeIndex timeindex,
#>                    t_handle_evidL handleEvid,
#>                    t_getDur getdur) {
#>   // assign start
#>     static rxode2_assignFuns2_t rxode2_assignFuns2 = NULL;
#>     if (rxode2_assignFuns2 == NULL) rxode2_assignFuns2 = (rxode2_assignFuns2_t)(R_GetCCallable("rxode2", "_rxode2_assignFuns2"));
#>     rxode2_assignFuns2(rx, op, f, lag, rate, dur, mtime, me, indf, gettime, timeindex, handleEvid, getdur);
#>   // assign stop
#> }
#> extern void  rx_47ebddf599b2ee87c79d0a088472449a__ode_solver_solvedata (rx_solve *solve){
#>   _solveData = solve;
#> }
#> extern rx_solve *rx_47ebddf599b2ee87c79d0a088472449a__ode_solver_get_solvedata(void){
#>   return _solveData;
#> }
#> SEXP rx_47ebddf599b2ee87c79d0a088472449a__model_vars(void);
#> 
#> 
#> // prj-specific differential eqns
#> void rx_47ebddf599b2ee87c79d0a088472449a__dydt(int *_neq, double __t, double *__zzStateVar__, double *__DDtStateVar__)
#> {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   int _cSub = _neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=1;
#>   double shape = NA_REAL;
#>   double scale = NA_REAL;
#>   double fx = NA_REAL;
#>   double dShape = NA_REAL;
#>   double dScale = NA_REAL;
#> 
#>   (void)t;
#>   (void)shape;
#>   (void)scale;
#>   (void)fx;
#>   (void)dShape;
#>   (void)dScale;
#> 
#>   fx = _PL[0];
#>   dShape = _PL[1];
#>   dScale = _PL[2];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#>   shape = _PP[0];
#>   scale = _PP[1];
#> 
#> 
#>   fx =llikWeibull(t,shape,scale);
#>   dShape =llikWeibullDshape(t,shape,scale);
#>   dScale =llikWeibullDscale(t,shape,scale);
#>   (&_solveData->subjects[_cSub])->dadt_counter[0]++;
#> }
#> 
#> // Jacobian derived vars
#> void rx_47ebddf599b2ee87c79d0a088472449a__calc_jac(int *_neq, double __t, double *__zzStateVar__, double *__PDStateVar__, unsigned int __NROWPD__) {
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>     int _cSub=_neq[1];
#>   double t = __t + _solveData->subjects[_neq[1]].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=2;
#>   (&_solveData->subjects[_cSub])->jac_counter[0]++;
#> }
#> // Functional based initial conditions.
#> void rx_47ebddf599b2ee87c79d0a088472449a__inis(int _cSub, double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>  _ind->_rxFlag=3;
#> }
#> // prj-specific derived vars
#> void rx_47ebddf599b2ee87c79d0a088472449a__calc_lhs(int _cSub, double __t, double *__zzStateVar__, double *_lhs) {
#>     int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=11;
#>   double shape = NA_REAL;
#>   double scale = NA_REAL;
#>   double fx = NA_REAL;
#>   double dShape = NA_REAL;
#>   double dScale = NA_REAL;
#> 
#>   (void)t;
#>   (void)shape;
#>   (void)scale;
#>   (void)fx;
#>   (void)dShape;
#>   (void)dScale;
#> 
#>   fx = _PL[0];
#>   dShape = _PL[1];
#>   dScale = _PL[2];
#> 
#>   _update_par_ptr(__t, _cSub, _solveData, _idx);
#>   shape = _PP[0];
#>   scale = _PP[1];
#> 
#> 
#>   fx =llikWeibull(t,shape,scale);
#>   dShape =llikWeibullDshape(t,shape,scale);
#>   dScale =llikWeibullDscale(t,shape,scale);
#> 
#>   _lhs[_LHS_0_]=fx;
#>   _lhs[_LHS_1_]=dShape;
#>   _lhs[_LHS_2_]=dScale;
#> }
#> // Functional based bioavailability
#> double rx_47ebddf599b2ee87c79d0a088472449a__F(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return _amt;
#>   }
#> // Functional based absorption lag
#> double rx_47ebddf599b2ee87c79d0a088472449a__Lag(int _cSub,  int _cmt, double __t, double *__zzStateVar__){
#>  return __t;
#> }
#> // Modeled zero-order rate
#> double rx_47ebddf599b2ee87c79d0a088472449a__Rate(int _cSub,  int _cmt, double _amt, double __t, double *__zzStateVar__){
#>  return 0.0;
#> }
#> // Modeled zero-order duration
#> double rx_47ebddf599b2ee87c79d0a088472449a__Dur(int _cSub,  int _cmt, double _amt, double __t){
#>  return 0.0;
#> }
#> // Model Times
#> void rx_47ebddf599b2ee87c79d0a088472449a__mtime(int _cSub, double *_mtime){
#> }
#> // Matrix Exponential (0)
#> void rx_47ebddf599b2ee87c79d0a088472449a__ME(int _cSub, double _t, double __t, double *_mat, const double *__zzStateVar__){
#>   int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=9;
#> }
#> // Inductive linearization Matf
#> void rx_47ebddf599b2ee87c79d0a088472449a__IndF(int _cSub, double _t, double __t, double *_matf){
#>  int _itwhile = 0;
#>   (void)_itwhile;
#>   double t = __t + _solveData->subjects[_cSub].curShift;
#>   (void)t;
#>   rx_solving_options_ind *_ind = &(_solveData->subjects[_cSub]);
#>   _ind->_rxFlag=10;
#> }
#> extern SEXP rx_47ebddf599b2ee87c79d0a088472449a__model_vars(void){
#>   int pro=0;
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_47ebddf599b2ee87c79d0a088472449a__model_vars"));pro++;
#>   if (!_rxIsCurrentC(_mv)){
#>     SEXP hash    = PROTECT(Rf_allocVector(STRSXP, 1));pro++;
#> #define __doBuf__  snprintf(buf, __doBufN__, "un]\"BAAA@QRtHACAAAAAAA6y3AAAv7#aT)Fr\"&NA8@)a=WdwYB~1u%%C8Ae0!_iwDPd#i.)C5<zk$+5*F#kijHRGI`mXOVB|CQwgU4@$%%N`|tI&F>x)D.>4Ko1WIAkArBq6_P=nOZW6tj<U&lHWJ_a{k)&|lU46=,C:19P_nL&Qj8.`:PN>_RGMlWw[_c^&S_kGYP^)wr=2LYP_t!nn#VrMiq~G_yw.mq5@.V>K#nYaNI*DuIJmPS5Jbx;Qof`D`mHyp)EfdY}Cc#ad1S\?yDYZ|yT~+3!;UhM:V=VejMk,!;!(W{wr<(j;Rn#7h+D2OmjMRuHrMn`mI9bF;YoY]t`8NG[Dq:%%qN;hH>]br\?Vw&ZWdsu8\"&\"rL:OOl,Z,iys>ymS#LRg(>Kb2MzXVLUXuAJ\"F5q@UC)LiMr7KIBw^7qc\?LEj\"JIDgApHFM0608.9:uF3}tTT7B(Mq@&\?6]`ar0;y2DHpFBdxjp/7/F$4[O$tRUElJWaw/{f64{&v;q|4v{ULwd7>{*iOC!>}8NkLU]_)@zxU0X}Sld9&E&5%%mT@P7OQl~DXZE/Io.N4@4%%[GV^$DU#nht+WHPci1$I[$$Jw#4Lb*eO]otcBJn5PcaOD*bfs:qF:e]`u#,Kuesudp8<&gz4V|5U95Bn2&&#O{tXWeh6VH*{*/J4gfSnh<+}!6>}W6CYj[)$pv4]G0L9eT\"pcamkbqaZ}Z*Mz7dag/|f\?5My=0<]t@>r>NP_T!dV`=;RIwuVC`n@)SyQmQx#ax,impEvt<!c1/E)i!iIBRGj_YT77S@8MSZj7RiNoz%%Z6A844!d<rGd5ky$GOGC5eS7FgGFMex:\"\?n$)!\?y(%%y7CU8~sr~V+,oa9F%%vEBNg2gfxm{6m>ta@~|73.<kvvC,\?#\?Vp|k,$O>ZZeAuenr*^DbQ%%swx~I39buLw8w^J2/RBpj\?W3~ioLB2aOc4!@:m6OH~U3[aZxL5UbfgW)TO(hz#:(1(u*JUN)!6F}>{grz{/[eb\?Q3li:UrcwF5ujG_lh33D1;B&@vW|~5\?TJT~YV9bq(4\"C].FPPTLv`4.0P*9Js`~__ntP6S=qyf]xRiUfUz[AWQfn)K!/DRj\?2qQ3=5tHAJ&#$Q|(2O_c+erB");
#>     char buf[1136];
#> #define __doBufN__ 1136
#>     __doBuf__
#> #undef __doBuf__
#> #undef __doBufN__
#>     SET_STRING_ELT(hash, 0, Rf_mkChar(buf));
#>     SEXP lst      = PROTECT(_rxQr(hash));pro++;
#>     _assign_ptr(lst);
#>     UNPROTECT(pro);
#>     return lst;
#>   } else {
#>     UNPROTECT(pro);
#>     return _mv;
#>   }
#> }
#> extern void rx_47ebddf599b2ee87c79d0a088472449a__dydt_lsoda(int *neq, double *t, double *A, double *DADT)
#> {
#>   rx_47ebddf599b2ee87c79d0a088472449a__dydt(neq, *t, A, DADT);
#> }
#> extern int rx_47ebddf599b2ee87c79d0a088472449a__dydt_liblsoda(double __t, double *y, double *ydot, void *data)
#> {
#>   int *neq = (int*)(data);
#>   rx_47ebddf599b2ee87c79d0a088472449a__dydt(neq, __t, y, ydot);
#>   return(0);
#> }
#> extern void rx_47ebddf599b2ee87c79d0a088472449a__calc_jac_lsoda(int *neq, double *t, double *A,int *ml, int *mu, double *JAC, int *nrowpd){
#>   // Update all covariate parameters
#>   rx_47ebddf599b2ee87c79d0a088472449a__calc_jac(neq, *t, A, JAC, *nrowpd);
#> }
#> 
#> //Create function to call from R's main thread that assigns the required functions. Sometimes they don't get assigned.
#> extern void rx_47ebddf599b2ee87c79d0a088472449a__assignFuns(void){
#>   _assignFuns();
#> }
#> 
#> //Initialize the dll to match rxode2's calls
#> void R_init0_rx_47ebddf599b2ee87c79d0a088472449a_(void){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_RegisterCCallable("rx_47ebddf599b2ee87c79d0a088472449a_","rx_47ebddf599b2ee87c79d0a088472449a__assignFuns2", (DL_FUNC) __assignFuns2);
#>   R_RegisterCCallable("rx_47ebddf599b2ee87c79d0a088472449a_","rx_47ebddf599b2ee87c79d0a088472449a__assignFuns", (DL_FUNC) rx_47ebddf599b2ee87c79d0a088472449a__assignFuns);
#>   R_RegisterCCallable("rx_47ebddf599b2ee87c79d0a088472449a_","rx_47ebddf599b2ee87c79d0a088472449a__inis",(DL_FUNC) rx_47ebddf599b2ee87c79d0a088472449a__inis);
#>   R_RegisterCCallable("rx_47ebddf599b2ee87c79d0a088472449a_","rx_47ebddf599b2ee87c79d0a088472449a__dydt",(DL_FUNC) rx_47ebddf599b2ee87c79d0a088472449a__dydt);
#>   R_RegisterCCallable("rx_47ebddf599b2ee87c79d0a088472449a_","rx_47ebddf599b2ee87c79d0a088472449a__calc_lhs",(DL_FUNC) rx_47ebddf599b2ee87c79d0a088472449a__calc_lhs);
#>   R_RegisterCCallable("rx_47ebddf599b2ee87c79d0a088472449a_","rx_47ebddf599b2ee87c79d0a088472449a__calc_jac",(DL_FUNC) rx_47ebddf599b2ee87c79d0a088472449a__calc_jac);
#>   R_RegisterCCallable("rx_47ebddf599b2ee87c79d0a088472449a_","rx_47ebddf599b2ee87c79d0a088472449a__dydt_lsoda", (DL_FUNC) rx_47ebddf599b2ee87c79d0a088472449a__dydt_lsoda);
#>   R_RegisterCCallable("rx_47ebddf599b2ee87c79d0a088472449a_","rx_47ebddf599b2ee87c79d0a088472449a__calc_jac_lsoda", (DL_FUNC) rx_47ebddf599b2ee87c79d0a088472449a__calc_jac_lsoda);
#>   R_RegisterCCallable("rx_47ebddf599b2ee87c79d0a088472449a_","rx_47ebddf599b2ee87c79d0a088472449a__ode_solver_solvedata", (DL_FUNC) rx_47ebddf599b2ee87c79d0a088472449a__ode_solver_solvedata);
#>   R_RegisterCCallable("rx_47ebddf599b2ee87c79d0a088472449a_","rx_47ebddf599b2ee87c79d0a088472449a__ode_solver_get_solvedata", (DL_FUNC) rx_47ebddf599b2ee87c79d0a088472449a__ode_solver_get_solvedata);
#>   R_RegisterCCallable("rx_47ebddf599b2ee87c79d0a088472449a_","rx_47ebddf599b2ee87c79d0a088472449a__F", (DL_FUNC) rx_47ebddf599b2ee87c79d0a088472449a__F);
#>   R_RegisterCCallable("rx_47ebddf599b2ee87c79d0a088472449a_","rx_47ebddf599b2ee87c79d0a088472449a__Lag", (DL_FUNC) rx_47ebddf599b2ee87c79d0a088472449a__Lag);
#>   R_RegisterCCallable("rx_47ebddf599b2ee87c79d0a088472449a_","rx_47ebddf599b2ee87c79d0a088472449a__Rate", (DL_FUNC) rx_47ebddf599b2ee87c79d0a088472449a__Rate);
#>   R_RegisterCCallable("rx_47ebddf599b2ee87c79d0a088472449a_","rx_47ebddf599b2ee87c79d0a088472449a__Dur", (DL_FUNC) rx_47ebddf599b2ee87c79d0a088472449a__Dur);
#>   R_RegisterCCallable("rx_47ebddf599b2ee87c79d0a088472449a_","rx_47ebddf599b2ee87c79d0a088472449a__mtime", (DL_FUNC) rx_47ebddf599b2ee87c79d0a088472449a__mtime);
#>   R_RegisterCCallable("rx_47ebddf599b2ee87c79d0a088472449a_","rx_47ebddf599b2ee87c79d0a088472449a__ME", (DL_FUNC) rx_47ebddf599b2ee87c79d0a088472449a__ME);
#>   R_RegisterCCallable("rx_47ebddf599b2ee87c79d0a088472449a_","rx_47ebddf599b2ee87c79d0a088472449a__IndF", (DL_FUNC) rx_47ebddf599b2ee87c79d0a088472449a__IndF);
#>   R_RegisterCCallable("rx_47ebddf599b2ee87c79d0a088472449a_","rx_47ebddf599b2ee87c79d0a088472449a__dydt_liblsoda", (DL_FUNC) rx_47ebddf599b2ee87c79d0a088472449a__dydt_liblsoda);
#> }
#> //Initialize the dll to match rxode2's calls
#> void R_init_rx_47ebddf599b2ee87c79d0a088472449a_(DllInfo *info){
#>   // Get C callables on load; Otherwise it isn't thread safe
#>   R_init0_rx_47ebddf599b2ee87c79d0a088472449a_();
#>   static const R_CallMethodDef callMethods[]  = {
#>     {"rx_47ebddf599b2ee87c79d0a088472449a__model_vars", (DL_FUNC) &rx_47ebddf599b2ee87c79d0a088472449a__model_vars, 0},
#>     {NULL, NULL, 0}
#>   };
#> 
#>   R_registerRoutines(info, NULL, callMethods, NULL, NULL);
#>   R_useDynamicSymbols(info,FALSE);
#>   _assignFuns0();
#> 
#> }
#> 
#> void R_unload_rx_47ebddf599b2ee87c79d0a088472449a_ (DllInfo *info){
#>   // Free resources required for single subject solve.
#>   SEXP _mv = PROTECT(_rxGetModelLib("rx_47ebddf599b2ee87c79d0a088472449a__model_vars"));
#>   if (!Rf_isNull(_mv)){
#>     _rxRmModelLib("rx_47ebddf599b2ee87c79d0a088472449a__model_vars");
#>   }
#>   UNPROTECT(1);
#> }
#> 
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’
#> ── Solved rxode2 object ──
#> ── Parameters (value$params): ──
#> # A tibble: 1 × 0
#> ── Initial Conditions (value$inits): ──
#> named numeric(0)
#> ── First part of data (object): ──
#> # A tibble: 10 × 6
#>    time    fx dShape  dScale shape scale
#>   <dbl> <dbl>  <dbl>   <dbl> <dbl> <dbl>
#> 1 0.001 -2.30  -8.21 -0.1000     1    10
#> 2 0.112 -2.31  -3.44 -0.0989     1    10
#> 3 0.223 -2.32  -2.72 -0.0978     1    10
#> 4 0.334 -2.34  -2.29 -0.0967     1    10
#> 5 0.445 -2.35  -1.97 -0.0956     1    10
#> 6 0.556 -2.36  -1.73 -0.0944     1    10
#> # ℹ 4 more rows
# }