Simulate a from a Poisson process

Simulate a from a Poisson process

rxPp(
  n,
  lambda,
  gamma = 1,
  prob = NULL,
  t0 = 0,
  tmax = Inf,
  randomOrder = FALSE
)

Arguments

n

Number of time points to simulate in the Poisson process

lambda

Rate of Poisson process

gamma

Asymmetry rate of Poisson process. When gamma=1.0, this simulates a homogenous Poisson process. When gamma<1.0, the Poisson process has more events early, when gamma > 1.0, the Poisson process has more events late in the process.

When gamma is non-zero, the tmax should not be infinite but indicate the end of the Poisson process to be simulated. In most pharamcometric cases, this will be the end of the study. Internally this uses a rate of:

l(t) = lambdagamma(t/tmax)^(gamma-1)

prob

When specified, this is a probability function with one argument, time, that gives the probability that a Poisson time t is accepted as a rejection time.

t0

the starting time of the Poisson process

tmax

the maximum time of the Poisson process

randomOrder

when TRUE randomize the order of the Poisson events. By default (FALSE) it returns the Poisson process is in order of how the events occurred.

Value

This returns a vector of the Poisson process times; If the dropout is >= tmax, then all the rest of the times are = tmax to indicate the dropout is equal to or after tmax.

Author

Matthew Fidler

Examples

## Sample homogenous Poisson process of rate 1/10
rxPp(10, 1 / 10)
#>  [1]  22.52244  32.52278  35.67854  40.13887  42.27229  43.42258  59.96902
#>  [8]  83.65568 106.89140 114.00779

## Sample inhomogenous Poisson rate of 1/10

rxPp(10, 1 / 10, gamma = 2, tmax = 100)
#>  [1]   5.776723  14.314296  23.810958  60.695604  86.741636  87.245363
#>  [7]  93.626805 100.000000 100.000000 100.000000

## Typically the Poisson process times are in a sequential order,
## using randomOrder gives the Poisson process in random order

rxPp(10, 1 / 10, gamma = 2, tmax = 10, randomOrder = TRUE)
#>  [1] 10 10 10 10 10 10 10 10 10 10

## This uses an arbitrary function to sample a non-homogenous Poisson process

rxPp(10, 1 / 10, prob = function(x) {
  1 / x
})
#>  [1]   1.857834  55.443178 109.272974 124.328784 255.434906 271.852250
#>  [7] 386.936304 511.998214 544.051434 565.722050