Convert a Raw Vector or R object to C Code

This function converts a raw vector or R object to C code that is used in the rxode2 model

Usage

rxRawToC(raw, type = c("xz", "qs2", "qdata", "base", "bzip2"))

Arguments

raw

vector or R object to convert

type

serialization type; one of "qs2", "qdata", "base", "xz" or "bzip2".

Value

raw vector for conversion to C in the codegen procedure

Author

Matthew L. Fidler

Examples

rxRawToC(mtcars)
#>    [1] 42 5a 68 39 31 41 59 26 53 59 1c df ec 6b 00 04 b9 ff ff ff ff fe 71 d4
#>   [25] 77 ff ff ff f6 ff dd 3f ff ff fe 42 00 a0 34 4c 05 18 40 59 06 60 81 61
#>   [49] 0e 4b 73 50 04 78 34 1b 6b 46 da 55 49 90 c9 11 a4 34 48 f2 9e a3 d4 f5
#>   [73] 3c 05 3c a7 93 24 3d 43 d4 f4 80 d0 34 3d 4c 9a 06 9e a0 d3 c2 35 1a 1e
#>   [97] a1 b2 9a 00 00 00 7a 80 d1 ea 0d 04 20 09 a3 44 69 a2 64 da 8d 1a 7a 20
#>  [121] 34 01 93 40 34 34 00 06 80 68 d1 a6 d4 01 a0 1a 06 99 00 1a 00 91 22 24
#>  [145] d2 03 d4 00 32 1a 07 a3 50 64 03 40 00 34 00 00 00 03 d4 1a 00 00 00 01
#>  [169] a1 a6 88 30 98 23 26 02 60 13 04 c1 01 80 00 4c 00 4c 00 8d 30 86 23 01
#>  [193] 1a 30 8c 26 8d 00 c9 93 20 83 09 82 32 60 26 01 30 4c 10 18 00 04 c0 04
#>  [217] c0 08 d3 08 62 30 11 a3 08 c2 68 d0 0c 99 32 14 76 cf 0f a7 e7 25 7b 1f
#>  [241] 99 4a 67 c5 93 9c e5 4c 88 6a 1d 33 9b b8 90 c7 29 d2 2c 93 43 68 1b 1b
#>  [265] a3 17 2a cd b6 d8 26 4d 81 0d 93 21 b2 1d bb 2d a6 c2 2d e7 04 c7 04 0d
#>  [289] 44 21 b8 b0 5d 68 6f 34 0e f3 4f b5 bf be e4 9c 4b 39 20 58 ea cd e5 a1
#>  [313] 63 1a 2d d4 e6 90 08 b8 0c d4 20 81 8d 0c 4d 21 a6 95 89 c2 93 17 92 cd
#>  [337] 4d 8d 7a 16 21 31 54 b1 55 88 80 e6 b0 48 46 bd a4 5f 73 21 09 24 ac 30
#>  [361] 55 6c f5 2c e3 cb 67 25 dd b7 6b 81 b3 8f bd ba d1 20 76 54 6f f2 27 c8
#>  [385] 74 b5 39 c2 fa cd 8e 2b e2 5b 44 7e 7b 79 67 b6 15 8c 90 6a d2 48 68 49
#>  [409] 79 3e 02 21 17 36 77 f2 88 6d 8e 12 6c 50 c1 b0 86 84 d9 0d b4 4a 80 28
#>  [433] 15 2a 8a 10 95 2d 20 60 c0 4a 06 9a 71 10 36 a2 08 48 a6 b6 be 40 a8 19
#>  [457] 36 00 a4 c3 a6 d1 2b 5b 49 23 80 35 30 b2 e1 82 1a 68 69 24 34 c1 b6 1d
#>  [481] 6f 9a 44 92 52 13 4d 50 e0 18 3a ae a4 10 a2 24 49 5e c2 04 40 36 d0 c6
#>  [505] ea d5 89 0b d6 8d b0 40 54 48 79 1d bf 29 32 d0 87 51 10 01 8d 16 8b d2
#>  [529] 4b 82 74 c9 40 4d 5a 8c f9 4c 04 5f 4a ec 30 b7 17 ef 08 2b ba b9 88 9e
#>  [553] 38 e3 90 77 41 70 6d b5 6a 43 d4 64 03 9c 5e d8 53 af 76 80 49 e4 84 67
#>  [577] 79 96 29 76 9b 35 48 15 f9 ba 4b b7 34 17 4e 0a bd d8 2b 5a d0 96 20 03
#>  [601] 3c c0 0f 69 7e 37 06 30 4b 96 24 7c a2 32 13 1a 24 b1 1a 93 0a 32 d0 42
#>  [625] 90 c9 8f bc 03 98 ac 41 b9 2d d8 93 25 44 06 86 ba e6 49 75 b5 09 e8 f9
#>  [649] 2a a4 d4 25 42 c2 98 96 d4 a0 2d d6 de fc af f8 76 13 29 3d 06 01 70 b6
#>  [673] eb a1 f6 47 4f 2a af bd c6 a0 8a e0 c0 40 cf a5 1c fc 3c 6f 4b c8 6f e5
#>  [697] 1f c9 c0 e3 66 24 c6 6e cb 74 79 a8 c1 4a 61 12 8a 86 41 86 02 19 21 90
#>  [721] 08 64 16 67 fd 12 40 81 c4 39 b6 50 42 60 88 53 3b 4c 54 54 34 33 43 7d
#>  [745] 5a cb 93 a7 56 27 ee 9a ab 46 74 8d a2 35 a4 72 0b d3 75 3f ca 3a ad 58
#>  [769] ad 54 b4 9c 2a 65 87 90 4c 6c 90 e1 8d a6 ca 18 28 8a 24 85 26 4c 60 29
#>  [793] 01 53 ee 9b 39 f5 fc d7 a8 37 79 00 62 d5 65 15 b5 f4 2f d3 4e ee fd a1
#>  [817] 56 80 e3 a0 75 01 82 d6 23 21 1c 66 78 71 e1 0a 19 7e 7f 0c c7 45 0c e7
#>  [841] 14 24 12 51 79 4a ec 86 26 6d 07 06 00 03 c1 d6 a4 88 3d 9a d7 d6 15 7e
#>  [865] ce a9 c0 f1 06 15 52 0b 52 03 54 1c 30 f2 8a c8 0d 20 28 47 12 dd bb 59
#>  [889] d9 c1 b5 05 69 a5 b7 2c d6 5b 57 38 bb 86 3e 74 39 95 a1 9e dd da 98 8b
#>  [913] 70 6a d5 77 49 82 1b 48 1f 54 50 53 c6 89 68 68 e8 cd 9b 06 00 f7 60 d9
#>  [937] 9e 34 8c 86 10 f0 a5 9f 56 c8 8a 5a 8d 39 1b 6e a6 d9 c8 ae 72 22 0f 8a
#>  [961] 92 16 49 00 47 f6 47 e6 5d a5 5a ef 8f 65 4e 03 5b cd 06 d8 fc c1 8d 3a
#>  [985] 3c 2c 18 b4 28 aa 34 a4 49 11 84 a7 57 59 94 77 2e 31 e3 d8 22 d7 c7 48
#> [1009] 2c 62 84 09 5d ba 3b cb 0a 81 f4 21 9d 61 bf 38 1e bf b8 6a 9d 11 12 42
#> [1033] 86 c1 10 ad 55 e6 5a 2a 0c e5 59 a7 12 fb 5f 18 8c 0d 86 4c 8a 24 c8 7b
#> [1057] 98 23 d3 87 0b 68 32 d7 60 64 f0 6e ac 96 8a 0a 31 e4 1d 27 53 02 6b d5
#> [1081] 71 91 5d b8 d5 6e 98 14 c5 e6 13 54 4f 28 a6 c1 5d e8 f1 a1 ad 22 36 fa
#> [1105] 15 ae cb c5 88 1c a0 e2 64 86 a6 0d 8d 7d b9 70 d1 76 03 84 90 1d 8d 53
#> [1129] 03 9f 5a cc d4 34 50 80 63 90 c5 30 4e 19 c9 8f 8b 65 3a e2 32 52 96 57
#> [1153] 83 f8 4c c8 03 31 2b e0 6c 93 fa 07 36 b9 48 ab 44 20 ea 79 c9 07 b2 f7
#> [1177] 3d ff 0b 4f d7 f5 90 c3 b7 42 1e 63 09 f1 44 20 84 0c 7c 04 44 d7 c0 58
#> [1201] 39 b4 29 3f fa 12 e7 1a e5 8b 8d ea 20 95 de df 17 c6 ef 6d 1b c0 c1 16
#> [1225] 19 ad 5b 48 c0 a9 ef 00 57 c5 d3 2f ff bf b5 6b 80 57 f8 19 c5 9d 1b 18
#> [1249] c5 ac e1 51 fd 3f da 13 31 24 17 1b 63 07 6f 19 3e 92 38 03 c2 61 52 b9
#> [1273] 8c b9 b6 11 84 32 47 72 66 5b fa c0 80 2b 73 86 10 6f 57 d9 2c 92 c1 19
#> [1297] 24 a7 fc 5d c9 14 e1 42 40 73 7f b1 ac