Convert a Raw Vector or R object to C Code

This function converts a raw vector or R object to C code that is used in the rxode2 model

Usage

rxRawToC(raw, type = c("qs2", "qdata", "base"))

Arguments

raw

raw vector or R object to convert

type

serialization type; one of "qs2", "qdata", or "base".

Value

character string of C code

Author

Matthew L. Fidler

Examples

message(rxRawToC(mtcars))
#>     SEXP rw    = PROTECT(Rf_allocVector(RAWSXP, 1309));pro++;
#>     unsigned char r[]={
#>                 0x42, 0x5A, 0x68, 0x39, 0x31, 0x41, 0x59, 0x26, 0x53, 0x59, 
#>                 0x81, 0x7F, 0x5F, 0x7E, 0x00, 0x04, 0xB9, 0xFF, 0xFF, 0xFF, 
#>                 0xFF, 0xFE, 0x71, 0xD4, 0x77, 0xFF, 0xFF, 0xFF, 0xF6, 0xFF, 
#>                 0xDD, 0x3F, 0xFF, 0xFF, 0xFE, 0x42, 0x00, 0xA0, 0x34, 0x4C, 
#>                 0x05, 0x18, 0x40, 0x59, 0x06, 0x60, 0x81, 0x61, 0x0E, 0x4B, 
#>                 0x73, 0x50, 0x04, 0x76, 0x1A, 0x0B, 0x6B, 0x46, 0xDA, 0x55, 
#>                 0x49, 0x90, 0xC8, 0x91, 0xA2, 0x09, 0xA8, 0xF2, 0x04, 0xF5, 
#>                 0x3C, 0x13, 0x53, 0x26, 0xD3, 0x28, 0x34, 0xD3, 0x4C, 0x86, 
#>                 0x80, 0x0D, 0x3D, 0x46, 0x9A, 0x0C, 0x9E, 0xA1, 0x93, 0xC9, 
#>                 0xA3, 0x51, 0x91, 0xEA, 0x3D, 0x1A, 0x9A, 0x00, 0x00, 0x00, 
#>                 0x7A, 0x80, 0xD1, 0xA0, 0xD3, 0x44, 0x22, 0x61, 0x04, 0xC8, 
#>                 0x6A, 0x69, 0xA6, 0x4D, 0xA8, 0xD1, 0xA6, 0x9A, 0x03, 0x40, 
#>                 0x0D, 0x34, 0x03, 0x43, 0x40, 0x00, 0x68, 0x00, 0x03, 0x4D, 
#>                 0x01, 0xEA, 0x00, 0x06, 0x99, 0x00, 0x1A, 0x00, 0x91, 0x24, 
#>                 0x44, 0xD1, 0x34, 0x6D, 0x20, 0x03, 0x21, 0xA1, 0xA7, 0xA1, 
#>                 0x1A, 0x64, 0x00, 0xD0, 0x01, 0x93, 0x40, 0x34, 0x00, 0x32, 
#>                 0x00, 0x3D, 0x41, 0xA0, 0x0D, 0x00, 0x0D, 0x00, 0x1A, 0x32, 
#>                 0x68, 0x83, 0x09, 0x82, 0x32, 0x60, 0x26, 0x01, 0x30, 0x4C, 
#>                 0x10, 0x18, 0x00, 0x04, 0xC0, 0x04, 0xC0, 0x08, 0xD3, 0x08, 
#>                 0x62, 0x30, 0x11, 0xA3, 0x08, 0xC2, 0x68, 0xD0, 0x0C, 0x99, 
#>                 0x32, 0x08, 0x30, 0x98, 0x23, 0x26, 0x02, 0x60, 0x13, 0x04, 
#>                 0xC1, 0x01, 0x80, 0x00, 0x4C, 0x00, 0x4C, 0x00, 0x8D, 0x30, 
#>                 0x86, 0x23, 0x01, 0x1A, 0x30, 0x8C, 0x26, 0x8D, 0x00, 0xC9, 
#>                 0x93, 0x20, 0xFC, 0x81, 0x67, 0xD5, 0xF1, 0x64, 0xD7, 0xAF, 
#>                 0xDD, 0x99, 0xC3, 0x28, 0x87, 0x39, 0xCD, 0x06, 0x0A, 0x51, 
#>                 0x4C, 0x1C, 0xE3, 0x61, 0x81, 0x05, 0x9D, 0x00, 0x95, 0xCA, 
#>                 0x02, 0xD0, 0x36, 0x37, 0x3D, 0x84, 0xA9, 0xB6, 0xDB, 0x60, 
#>                 0x99, 0x33, 0x02, 0x1B, 0x26, 0x21, 0xB2, 0x1D, 0x9B, 0x2C, 
#>                 0xA6, 0x61, 0x16, 0x73, 0x41, 0x30, 0xE0, 0x81, 0xA8, 0x84, 
#>                 0x37, 0x17, 0x05, 0xBD, 0xCD, 0xDE, 0xA5, 0xDD, 0xEB, 0xF7, 
#>                 0xB8, 0x77, 0xFC, 0xE3, 0x93, 0x57, 0x38, 0x0B, 0x31, 0x4D, 
#>                 0xBD, 0x04, 0x2B, 0x16, 0x8B, 0x35, 0x34, 0xC9, 0x00, 0x8B, 
#>                 0x40, 0xD2, 0x42, 0x08, 0x18, 0xD0, 0xC4, 0xD2, 0x1A, 0x69, 
#>                 0x5C, 0x4D, 0x0A, 0x4C, 0x58, 0x95, 0x68, 0xEE, 0x6A, 0x4E, 
#>                 0xAB, 0xD3, 0x15, 0x1B, 0x8A, 0x4C, 0x44, 0x06, 0x13, 0x04, 
#>                 0x84, 0x6D, 0xDA, 0x45, 0xFE, 0x0C, 0x21, 0x24, 0x95, 0x73, 
#>                 0x05, 0x4A, 0xAE, 0x7D, 0x8E, 0xBC, 0x4B, 0x15, 0xDB, 0xDB, 
#>                 0x6E, 0x10, 0x47, 0x32, 0xB1, 0xE5, 0xC9, 0x50, 0xDE, 0xE4, 
#>                 0x37, 0xF2, 0xED, 0xF4, 0x9B, 0x12, 0xE7, 0x30, 0x50, 0xB1, 
#>                 0x3F, 0xA0, 0xF8, 0x89, 0x74, 0xAF, 0xDB, 0xC8, 0x89, 0x6F, 
#>                 0x1C, 0x63, 0x21, 0x41, 0x78, 0x92, 0x43, 0x42, 0x4B, 0x13, 
#>                 0x14, 0x44, 0x22, 0xDA, 0xAF, 0x16, 0x51, 0x0D, 0xB1, 0xC2, 
#>                 0x4D, 0x8A, 0x18, 0x36, 0x10, 0xD0, 0x9B, 0x21, 0xB6, 0x89, 
#>                 0x4E, 0x04, 0xE2, 0xA0, 0xA7, 0xDB, 0xD0, 0x42, 0x54, 0x5A, 
#>                 0x40, 0xC1, 0x80, 0x94, 0x86, 0x9A, 0x71, 0x10, 0x36, 0xA2, 
#>                 0x08, 0x48, 0xA3, 0x63, 0x00, 0xA6, 0x19, 0x26, 0x00, 0xA7, 
#>                 0x61, 0xD6, 0x68, 0x95, 0x6D, 0x64, 0x91, 0x80, 0x35, 0x30, 
#>                 0x55, 0x70, 0xC1, 0x0D, 0x34, 0x34, 0x92, 0x1A, 0x60, 0xDB, 
#>                 0x0E, 0xCF, 0xD1, 0x22, 0x49, 0x29, 0x09, 0xA6, 0xA7, 0x70, 
#>                 0x0C, 0x1D, 0x27, 0x46, 0x08, 0x51, 0x12, 0x24, 0xB6, 0x50, 
#>                 0x81, 0x10, 0x0D, 0xB4, 0x31, 0xBA, 0x54, 0xA2, 0x42, 0xF6, 
#>                 0xA2, 0xF4, 0x20, 0x2A, 0x48, 0x7E, 0x4B, 0xBE, 0x52, 0x6F, 
#>                 0x10, 0x88, 0x48, 0x88, 0x00, 0xDB, 0xC7, 0x9F, 0x00, 0x92, 
#>                 0xF0, 0xBE, 0x9B, 0x30, 0x0B, 0xD6, 0x29, 0xBF, 0x98, 0xD0, 
#>                 0x45, 0x95, 0xB0, 0xE2, 0x1A, 0xE7, 0x02, 0x21, 0x42, 0xC4, 
#>                 0x2C, 0x11, 0x2C, 0x46, 0x69, 0xA7, 0x19, 0x01, 0x37, 0xB9, 
#>                 0xAD, 0x3B, 0x20, 0xF4, 0x94, 0x02, 0x10, 0x4F, 0x6C, 0x8B, 
#>                 0xD4, 0xBC, 0x50, 0x27, 0xF1, 0xC2, 0x2D, 0x70, 0x6E, 0x28, 
#>                 0x3A, 0xCD, 0xDA, 0xA0, 0x05, 0x4C, 0x2D, 0x65, 0xD3, 0x99, 
#>                 0x05, 0xBB, 0x82, 0x97, 0x82, 0x0A, 0x75, 0xB3, 0xA5, 0x5E, 
#>                 0x01, 0xA7, 0x60, 0x07, 0xBD, 0x3F, 0x1B, 0x42, 0x30, 0x93, 
#>                 0x8E, 0x27, 0x8D, 0x25, 0x28, 0x29, 0x95, 0x5A, 0x32, 0x96, 
#>                 0x47, 0x24, 0xC9, 0x98, 0xA9, 0x39, 0x47, 0x47, 0x10, 0x10, 
#>                 0x73, 0x39, 0x46, 0xD5, 0x36, 0xE2, 0x5C, 0xB5, 0x20, 0x38, 
#>                 0x2E, 0xAB, 0x71, 0x8C, 0x27, 0x54, 0x5F, 0xA4, 0x14, 0xC4, 
#>                 0x4B, 0xD4, 0x25, 0x45, 0x35, 0x6A, 0x55, 0x55, 0x02, 0xAD, 
#>                 0x1C, 0x6F, 0x95, 0x9F, 0x0E, 0xCE, 0x6D, 0x08, 0x00, 0xDB, 
#>                 0x26, 0xFB, 0x66, 0xD9, 0x9F, 0x65, 0x85, 0x5A, 0x2C, 0xDE, 
#>                 0xF8, 0x16, 0x08, 0x97, 0x24, 0x05, 0x0E, 0xFA, 0x54, 0xCB, 
#>                 0xC3, 0xC6, 0xF4, 0x44, 0xA6, 0x3E, 0xD9, 0xF9, 0x40, 0x19, 
#>                 0x9D, 0x89, 0x4C, 0x98, 0x59, 0x84, 0xD9, 0x2A, 0xE1, 0x69, 
#>                 0x4C, 0x4A, 0xC2, 0xD4, 0xA0, 0x71, 0x01, 0x4A, 0x29, 0x40, 
#>                 0x14, 0xA0, 0x6A, 0x71, 0xF9, 0xE7, 0xA0, 0xA2, 0x08, 0x39, 
#>                 0x38, 0xEA, 0x24, 0x21, 0x10, 0x56, 0x36, 0xD8, 0xAA, 0xCA, 
#>                 0x54, 0x95, 0x31, 0xB3, 0xDD, 0x74, 0x6B, 0x57, 0x96, 0x3B, 
#>                 0xB7, 0xB3, 0x54, 0xAA, 0xB3, 0x3C, 0x4C, 0xD6, 0x86, 0x71, 
#>                 0x7C, 0x61, 0x5B, 0xF2, 0xB0, 0xAF, 0x4A, 0x29, 0xD2, 0x2B, 
#>                 0x26, 0x85, 0x42, 0x59, 0x49, 0x04, 0xC3, 0x64, 0x87, 0x0C, 
#>                 0x6D, 0x36, 0x4E, 0xC1, 0x44, 0x4F, 0x24, 0x29, 0x32, 0x61, 
#>                 0x80, 0xA8, 0x0A, 0xC1, 0xCC, 0x6F, 0x4A, 0x97, 0x9B, 0x52, 
#>                 0x73, 0xBA, 0x88, 0x03, 0x58, 0xB5, 0x11, 0x77, 0x3E, 0x86, 
#>                 0x7A, 0x70, 0x6B, 0xF0, 0x5C, 0x18, 0x93, 0x87, 0x29, 0x04, 
#>                 0x24, 0x06, 0x2B, 0x91, 0x1B, 0x18, 0xCC, 0x4E, 0x1C, 0xD8, 
#>                 0x42, 0x95, 0x9F, 0x37, 0x82, 0x99, 0xA8, 0xA4, 0xD5, 0x14, 
#>                 0xA7, 0x13, 0xD4, 0x89, 0x92, 0xF1, 0x06, 0x26, 0x4D, 0x28, 
#>                 0x07, 0x00, 0x04, 0x42, 0x16, 0xE8, 0xAA, 0x3D, 0x97, 0x30, 
#>                 0x69, 0xAD, 0x9A, 0x9B, 0x07, 0x83, 0xE8, 0x35, 0xD6, 0x80, 
#>                 0x62, 0xA0, 0x38, 0x80, 0xEB, 0xDC, 0x73, 0x0B, 0x08, 0x0E, 
#>                 0x20, 0x28, 0x47, 0x1E, 0xAD, 0x5C, 0xEC, 0xDC, 0xD0, 0xEA, 
#>                 0x85, 0xD5, 0x4A, 0xAF, 0x54, 0xA2, 0xC7, 0x2B, 0x72, 0x77, 
#>                 0x6D, 0x83, 0x08, 0x3D, 0xCC, 0xD0, 0xCF, 0x76, 0xBE, 0x1C, 
#>                 0x46, 0x3C, 0x36, 0x2B, 0x5F, 0x4D, 0xA2, 0xE1, 0xC4, 0x10, 
#>                 0x70, 0xD4, 0x2B, 0xEE, 0x12, 0xE0, 0xD1, 0xD1, 0xBD, 0xA7, 
#>                 0xB2, 0xA0, 0x22, 0xEA, 0x97, 0x65, 0x8C, 0xF3, 0xA4, 0x90, 
#>                 0x44, 0x2D, 0x32, 0xEB, 0xE3, 0x89, 0x68, 0xDA, 0xBC, 0xCE, 
#>                 0xFC, 0xDB, 0x17, 0xD5, 0x44, 0xBA, 0xEB, 0x4A, 0x23, 0x16, 
#>                 0xB3, 0xA7, 0x65, 0x00, 0x43, 0xFB, 0x23, 0xF3, 0x79, 0x9C, 
#>                 0xB9, 0xDD, 0xF1, 0xEC, 0xC3, 0xE1, 0xB9, 0xBF, 0xD0, 0x75, 
#>                 0xB0, 0xCC, 0x59, 0x17, 0xD2, 0x63, 0x5A, 0xC0, 0x9D, 0x3D, 
#>                 0x64, 0x4C, 0x44, 0x95, 0x20, 0x4C, 0x84, 0xCA, 0x2D, 0x9F, 
#>                 0xBC, 0x79, 0xAF, 0xC0, 0x8A, 0x2D, 0xFC, 0x56, 0x81, 0xA6, 
#>                 0x14, 0xA8, 0x2D, 0x87, 0xE7, 0x6F, 0x6F, 0xD9, 0x44, 0x6A, 
#>                 0x83, 0x2A, 0xFD, 0xFF, 0x3A, 0x11, 0x60, 0x5C, 0xBD, 0x8D, 
#>                 0x41, 0x2A, 0xC8, 0xA5, 0xC2, 0x51, 0x71, 0x5E, 0x27, 0x5B, 
#>                 0x2C, 0x5B, 0xE5, 0xA9, 0x83, 0x77, 0x65, 0xCF, 0x8D, 0xDB, 
#>                 0x43, 0xA1, 0xB3, 0x63, 0x89, 0x72, 0xA0, 0x64, 0x89, 0x38, 
#>                 0x37, 0x11, 0xB7, 0x03, 0x21, 0x94, 0xC7, 0x23, 0x89, 0x5D, 
#>                 0x84, 0xB8, 0x50, 0x4F, 0x93, 0x28, 0xE9, 0xC2, 0x5C, 0x1A, 
#>                 0x77, 0xCD, 0x02, 0x89, 0x79, 0x02, 0xCD, 0x76, 0xE0, 0xC4, 
#>                 0x5E, 0xE1, 0x3D, 0x85, 0x5A, 0x22, 0xB3, 0x85, 0x88, 0xA6, 
#>                 0xC6, 0x92, 0xE2, 0x75, 0x7E, 0x8E, 0x46, 0xB0, 0xEB, 0xD6, 
#>                 0x94, 0x1A, 0x20, 0xCA, 0xE9, 0xCB, 0x10, 0x5C, 0xF8, 0xDF, 
#>                 0x66, 0x4A, 0x90, 0xC1, 0x80, 0x9E, 0x81, 0xBF, 0xD2, 0xB8, 
#>                 0x42, 0x36, 0xB8, 0x76, 0x89, 0xE2, 0x95, 0x01, 0xD3, 0x41, 
#>                 0x8A, 0x42, 0x40, 0x65, 0x26, 0x2E, 0x05, 0x44, 0xD2, 0x12, 
#>                 0x12, 0x85, 0xE1, 0x65, 0xAF, 0xE1, 0x33, 0x6D, 0x86, 0x4A, 
#>                 0x52, 0xB4, 0xE9, 0x27, 0xF4, 0x0E, 0x5D, 0x22, 0x91, 0xD8, 
#>                 0x88, 0x41, 0xD6, 0xF3, 0x92, 0x0F, 0x6A, 0x06, 0x7C, 0x3E, 
#>                 0x2E, 0xAF, 0xAF, 0xEB, 0x29, 0xA7, 0x09, 0x08, 0x7F, 0x6C, 
#>                 0x2F, 0xE3, 0x88, 0xC1, 0x08, 0x18, 0xB6, 0x88, 0xBB, 0xD9, 
#>                 0x5B, 0x30, 0x1C, 0xB9, 0xD4, 0x21, 0xFD, 0x09, 0x7B, 0x1E, 
#>                 0xB5, 0x3A, 0xCE, 0xEB, 0x20, 0x95, 0xD5, 0xF5, 0x87, 0x2F, 
#>                 0xC3, 0xBC, 0x6F, 0x1B, 0x20, 0x55, 0xCD, 0x5E, 0x34, 0x8C, 
#>                 0x6A, 0x3F, 0x10, 0x05, 0x7E, 0x5B, 0xB3, 0x0F, 0x1F, 0xF8, 
#>                 0xAB, 0x71, 0x8A, 0x98, 0xA6, 0x99, 0x5A, 0x8D, 0x8C, 0x62, 
#>                 0xDA, 0x71, 0xE7, 0xC7, 0x3F, 0xD7, 0x26, 0x8A, 0x48, 0x2D, 
#>                 0x2F, 0x4C, 0x8B, 0xE8, 0xCE, 0xF5, 0x51, 0x80, 0x3C, 0x96, 
#>                 0x14, 0x6A, 0x16, 0x36, 0xD6, 0x42, 0x32, 0x46, 0x48, 0xB1, 
#>                 0x37, 0x22, 0x1D, 0x10, 0x40, 0x14, 0x72, 0x86, 0xB8, 0x3A, 
#>                 0x94, 0xB6, 0xAC, 0x24, 0xB4, 0x48, 0x49, 0x69, 0xFF, 0x17, 
#>                 0x72, 0x45, 0x38, 0x50, 0x90, 0x81, 0x7F, 0x5F, 0x7E
#>                 };
#>     memcpy(RAW(rw), r, sizeof(r));