Convert a Raw Vector or R object to C Code

This function converts a raw vector or R object to C code that is used in the rxode2 model

Usage

rxRawToC(raw, type = c("qdata", "qs2", "base"))

Arguments

raw

raw vector or R object to convert

type

serialization type; one of "qs2", "qdata", or "base".

Value

character string of C code

Author

Matthew L. Fidler

Examples

message(rxRawToC(mtcars))
#>     SEXP rw    = PROTECT(Rf_allocVector(RAWSXP, 1199));pro++;
#>     unsigned char r[]={
#>                 0x0B, 0x0E, 0x0A, 0xCD, 0x01, 0x01, 0x02, 0x01, 0x00, 0x00, 
#>                 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x54, 0x9E, 0x92, 0xAB, 
#>                 0x06, 0x46, 0x2E, 0xCC, 0x93, 0x04, 0x00, 0x00, 0x28, 0xB5, 
#>                 0x2F, 0xFD, 0x60, 0x07, 0x0C, 0x4D, 0x24, 0x00, 0x26, 0xB0, 
#>                 0xB2, 0x4C, 0xE0, 0xB2, 0x38, 0x07, 0xD0, 0x22, 0x51, 0xD9, 
#>                 0x8B, 0x16, 0x06, 0x50, 0x43, 0x15, 0x01, 0x49, 0x15, 0x79, 
#>                 0x4C, 0x15, 0x53, 0x10, 0x54, 0x10, 0x12, 0x0E, 0x39, 0x3C, 
#>                 0x63, 0xFB, 0xBE, 0xC1, 0x9A, 0xC9, 0x89, 0x5E, 0xD9, 0xC9, 
#>                 0xC0, 0x5A, 0x14, 0x36, 0xE3, 0x5A, 0x75, 0x55, 0x11, 0xDC, 
#>                 0x69, 0x02, 0x20, 0xDE, 0xA7, 0x01, 0x00, 0x25, 0xFE, 0x77, 
#>                 0x37, 0xB5, 0x08, 0x0F, 0x6A, 0x39, 0x94, 0x02, 0x86, 0x40, 
#>                 0xB6, 0xDF, 0x7E, 0x6E, 0xBA, 0xFE, 0x62, 0x07, 0x96, 0x00, 
#>                 0x8E, 0x00, 0xA8, 0x00, 0x7A, 0xCF, 0x52, 0xF4, 0x65, 0x72, 
#>                 0xEE, 0xBD, 0xD3, 0x73, 0xD6, 0x2F, 0x77, 0xF6, 0x5A, 0xA3, 
#>                 0xE7, 0x9B, 0x02, 0xCC, 0xD2, 0xCB, 0x85, 0x46, 0x5C, 0xA9, 
#>                 0x48, 0xA6, 0xC8, 0x62, 0x89, 0xE5, 0xE7, 0x12, 0x55, 0xE4, 
#>                 0x9D, 0xF1, 0x1C, 0x03, 0x8A, 0x2D, 0xE5, 0xE5, 0xB9, 0xAF, 
#>                 0x37, 0x17, 0xA9, 0xA7, 0x8F, 0xEA, 0x8F, 0x91, 0xF4, 0xDB, 
#>                 0x96, 0x7D, 0xE7, 0x5B, 0xA5, 0x29, 0xC9, 0x38, 0x94, 0xBB, 
#>                 0xAE, 0xEA, 0xF6, 0x9C, 0xF2, 0x0D, 0x13, 0x8A, 0xA3, 0x82, 
#>                 0x9B, 0xDE, 0xD6, 0xFE, 0x2E, 0x12, 0xED, 0x6B, 0xEE, 0x49, 
#>                 0x65, 0x33, 0x73, 0xFF, 0xEC, 0xF9, 0x18, 0x50, 0xC6, 0x7F, 
#>                 0x36, 0xE2, 0x9F, 0x69, 0x23, 0xDB, 0x8B, 0x56, 0x9C, 0x68, 
#>                 0x96, 0x57, 0x9D, 0x51, 0xDF, 0x11, 0xF5, 0xE7, 0xE7, 0x48, 
#>                 0xAD, 0x89, 0xD4, 0x15, 0x3D, 0x47, 0x9A, 0x67, 0x16, 0x42, 
#>                 0x53, 0x6D, 0x21, 0x57, 0xA4, 0xC4, 0xBA, 0xBE, 0x9E, 0x89, 
#>                 0xDC, 0x92, 0x95, 0x9D, 0xCD, 0x41, 0x14, 0x6B, 0xB6, 0x7D, 
#>                 0x2A, 0x82, 0xA0, 0x46, 0xCF, 0xB7, 0x84, 0x8B, 0xF6, 0xE7, 
#>                 0x79, 0xFD, 0x98, 0xD3, 0x4E, 0x51, 0xD7, 0x45, 0xF9, 0x73, 
#>                 0x9E, 0x5F, 0x4C, 0xAF, 0xFB, 0x3C, 0xF7, 0xF4, 0x61, 0xF7, 
#>                 0xE4, 0xFB, 0x25, 0x7C, 0x75, 0xF3, 0x9B, 0xCB, 0x5D, 0x6F, 
#>                 0x7D, 0xBF, 0x4C, 0x64, 0x4A, 0x35, 0x2E, 0xB8, 0xB9, 0x36, 
#>                 0x5B, 0xF9, 0x86, 0x9D, 0x16, 0xA4, 0x46, 0x1B, 0xF0, 0xE4, 
#>                 0xA8, 0x2C, 0x73, 0x14, 0x66, 0x59, 0x24, 0x86, 0xBB, 0xF9, 
#>                 0xDB, 0xDE, 0x24, 0xA5, 0xA8, 0xCE, 0xB8, 0xAF, 0x70, 0xCD, 
#>                 0x98, 0x08, 0x25, 0x9E, 0x69, 0x69, 0xD6, 0xFA, 0x6B, 0xE6, 
#>                 0xCD, 0xB4, 0x73, 0xBB, 0xEC, 0xA7, 0xA7, 0xD7, 0x7A, 0xF3, 
#>                 0x6D, 0xDE, 0xEF, 0x9B, 0x67, 0x5F, 0x08, 0x67, 0x28, 0x0E, 
#>                 0x35, 0xEF, 0x9F, 0xE9, 0x86, 0x29, 0xB1, 0xC8, 0xF1, 0xF6, 
#>                 0x47, 0x5F, 0x3E, 0xFD, 0x28, 0xCF, 0xAA, 0x8D, 0xC2, 0xE5, 
#>                 0x59, 0xAF, 0x3C, 0xBC, 0x66, 0x96, 0xE8, 0xEF, 0x69, 0x4E, 
#>                 0xD1, 0x9E, 0x6E, 0xF3, 0xA6, 0x01, 0x18, 0xFC, 0x4C, 0xC8, 
#>                 0x20, 0x3D, 0x50, 0x4F, 0xE0, 0x30, 0x28, 0x40, 0x99, 0x01, 
#>                 0xF7, 0x21, 0x35, 0x77, 0x13, 0xDB, 0xC9, 0x76, 0x36, 0x30, 
#>                 0x5C, 0xC1, 0x83, 0xE5, 0xF4, 0x4E, 0x83, 0x10, 0x4A, 0x22, 
#>                 0x27, 0x87, 0xB3, 0xAB, 0xD9, 0x59, 0xF0, 0x52, 0x12, 0x18, 
#>                 0x70, 0x16, 0xAA, 0xC1, 0x39, 0x1C, 0xCC, 0x61, 0xC8, 0xEE, 
#>                 0x6D, 0xA8, 0xC6, 0x05, 0x8E, 0x09, 0xCD, 0x8E, 0xC7, 0xBB, 
#>                 0x09, 0x79, 0xAB, 0x10, 0xCE, 0x88, 0x8C, 0x8B, 0x19, 0x1C, 
#>                 0x34, 0x29, 0x3A, 0xE4, 0x3C, 0xEC, 0x3E, 0xE8, 0x38, 0x38, 
#>                 0x75, 0xBC, 0x2C, 0x86, 0xCB, 0xE9, 0x8B, 0xCB, 0xA9, 0x37, 
#>                 0xC3, 0x79, 0x51, 0xD7, 0x5E, 0x7B, 0x51, 0xF6, 0xA2, 0x2E, 
#>                 0xCA, 0xA2, 0x56, 0xCA, 0x1F, 0x16, 0x06, 0x5E, 0x83, 0xD7, 
#>                 0x01, 0x96, 0x1C, 0xD6, 0x01, 0x01, 0xA7, 0x0A, 0x04, 0xAD, 
#>                 0xBC, 0xC4, 0xB1, 0x2C, 0x59, 0x09, 0x74, 0x03, 0x74, 0x3A, 
#>                 0x9D, 0x4E, 0x04, 0x26, 0x7F, 0x4A, 0xBC, 0xCA, 0xC1, 0x2D, 
#>                 0x41, 0x2F, 0xF0, 0xCA, 0xC3, 0x2B, 0x2F, 0xDB, 0x9E, 0xE7, 
#>                 0xAC, 0x7C, 0xFF, 0x93, 0xE4, 0x42, 0xAA, 0xC6, 0xBB, 0x71, 
#>                 0x7A, 0xC7, 0x48, 0x01, 0xD8, 0x43, 0x59, 0xFB, 0x59, 0x0F, 
#>                 0x7B, 0xD6, 0x9E, 0x9C, 0xD8, 0xEF, 0xB3, 0xF6, 0xFB, 0xAC, 
#>                 0xFD, 0x3E, 0xDB, 0x1A, 0x22, 0x53, 0x65, 0x4A, 0x14, 0x7F, 
#>                 0x46, 0x53, 0x8C, 0x19, 0x58, 0x25, 0x2A, 0x6D, 0xF8, 0x70, 
#>                 0xD8, 0x80, 0x03, 0x8F, 0x31, 0xC6, 0x1F, 0xE4, 0x38, 0xE6, 
#>                 0x90, 0xAF, 0x16, 0x1B, 0x2A, 0x16, 0x15, 0xAB, 0x24, 0x47, 
#>                 0x8C, 0x22, 0xE2, 0x6D, 0x64, 0x04, 0xEC, 0xB1, 0x6A, 0x79, 
#>                 0xC8, 0x11, 0xA2, 0x86, 0x01, 0xA4, 0x10, 0x68, 0x05, 0x09, 
#>                 0x77, 0xB4, 0xC0, 0x61, 0xA9, 0xA7, 0x03, 0x36, 0x85, 0x00, 
#>                 0xB5, 0x80, 0xDC, 0xA8, 0xB4, 0x7C, 0x00, 0x40, 0x79, 0x94, 
#>                 0x15, 0xA9, 0x30, 0x36, 0x1A, 0x80, 0x51, 0xC2, 0x07, 0x16, 
#>                 0x06, 0x34, 0x95, 0x4A, 0x22, 0x92, 0x91, 0xC9, 0xA4, 0xC2, 
#>                 0x23, 0x34, 0x80, 0x4B, 0x4E, 0x27, 0x4B, 0x63, 0x02, 0x36, 
#>                 0x01, 0xC2, 0x56, 0xE6, 0x86, 0x4F, 0x0A, 0xD0, 0x0C, 0x06, 
#>                 0x4D, 0xC8, 0xC0, 0x9F, 0x47, 0xA2, 0xEA, 0x7C, 0x20, 0x02, 
#>                 0xDC, 0xA8, 0xEC, 0x60, 0x01, 0xE2, 0x06, 0x9F, 0x7C, 0x52, 
#>                 0x80, 0x64, 0x6E, 0xF8, 0x60, 0x50, 0x05, 0x1E, 0x39, 0xE9, 
#>                 0xA1, 0x31, 0x41, 0xC5, 0x49, 0x81, 0x0C, 0xA8, 0x11, 0x49, 
#>                 0x72, 0x44, 0x49, 0x92, 0x92, 0x92, 0x52, 0xA1, 0x03, 0x81, 
#>                 0x05, 0xC4, 0x31, 0x07, 0x69, 0x75, 0x12, 0xE0, 0xB0, 0xF8, 
#>                 0x8C, 0x11, 0x12, 0x89, 0x86, 0x4A, 0x22, 0x09, 0x32, 0xCA, 
#>                 0x4E, 0x41, 0xB5, 0xB2, 0x19, 0x03, 0xD9, 0xD3, 0x59, 0xDC, 
#>                 0x5C, 0x63, 0x5E, 0xA5, 0x7D, 0x1A, 0x85, 0x69, 0x61, 0x2F, 
#>                 0x50, 0xC0, 0x9B, 0x90, 0xCD, 0xC8, 0x04, 0xE6, 0x3C, 0x3B, 
#>                 0x21, 0xA4, 0xDD, 0x18, 0x8E, 0x8C, 0x71, 0xE6, 0x5D, 0xE1, 
#>                 0xD5, 0x9A, 0x6D, 0x6D, 0x6B, 0x31, 0x16, 0x56, 0xB7, 0xB3, 
#>                 0xF3, 0x76, 0x74, 0x8E, 0xF3, 0xDA, 0xCB, 0x53, 0xAD, 0x96, 
#>                 0x61, 0xEC, 0x18, 0x2B, 0x50, 0xA1, 0xE2, 0x43, 0x0C, 0x06, 
#>                 0x2A, 0x38, 0x24, 0xF2, 0x2C, 0xD8, 0x51, 0x19, 0xBB, 0xD4, 
#>                 0x33, 0xC2, 0x80, 0xD7, 0x5A, 0x64, 0xC3, 0x70, 0x34, 0x8B, 
#>                 0x21, 0xE5, 0xBB, 0x7A, 0xB1, 0x42, 0xD2, 0x9B, 0xDD, 0xF9, 
#>                 0x29, 0x75, 0x9E, 0x4F, 0x7D, 0xA7, 0x54, 0xD1, 0xED, 0x75, 
#>                 0xEC, 0x1F, 0xB2, 0x66, 0xCB, 0x38, 0xC8, 0xC9, 0xE4, 0x2C, 
#>                 0x86, 0x86, 0x97, 0x93, 0xB2, 0x05, 0x6E, 0x6C, 0x40, 0xDA, 
#>                 0xF1, 0xFF, 0xDA, 0xD7, 0x72, 0x65, 0xC1, 0xD0, 0x62, 0x91, 
#>                 0x05, 0x1E, 0x49, 0x59, 0x4E, 0xEE, 0x8E, 0x36, 0x63, 0x82, 
#>                 0x37, 0x0C, 0x0D, 0x3C, 0x5E, 0x32, 0x22, 0x59, 0x8E, 0x41, 
#>                 0x3D, 0x6E, 0x66, 0x14, 0x93, 0x35, 0x88, 0xD3, 0xC0, 0x28, 
#>                 0x0B, 0x12, 0x0B, 0xCD, 0xE3, 0x1A, 0xDA, 0xA3, 0xD3, 0x9A, 
#>                 0x99, 0xEC, 0x70, 0x05, 0x80, 0x18, 0xB5, 0xC1, 0xAB, 0xB1, 
#>                 0x91, 0xC9, 0x01, 0xE2, 0x78, 0xEE, 0x21, 0xD7, 0x7F, 0x61, 
#>                 0x98, 0xD9, 0x26, 0x06, 0x76, 0xB3, 0x80, 0xAC, 0x03, 0x11, 
#>                 0xA0, 0x25, 0xE9, 0x9C, 0x95, 0x1B, 0xEC, 0xAC, 0x30, 0xF7, 
#>                 0xD7, 0x04, 0x61, 0x70, 0x46, 0x87, 0x3F, 0x98, 0xD4, 0x6C, 
#>                 0xEA, 0x88, 0xB2, 0xE5, 0x85, 0xF3, 0x5A, 0x02, 0x5A, 0x6B, 
#>                 0x45, 0xCC, 0xC8, 0x52, 0x25, 0x74, 0x2D, 0x19, 0xF4, 0x70, 
#>                 0x09, 0x33, 0xD1, 0xD5, 0xC2, 0xBF, 0x42, 0x0E, 0x6D, 0x23, 
#>                 0xF7, 0x85, 0x8D, 0x44, 0x35, 0xDB, 0xAB, 0xDE, 0x52, 0xD1, 
#>                 0xEF, 0xA6, 0x84, 0xA4, 0x60, 0x4A, 0xA7, 0xF9, 0xE6, 0xDC, 
#>                 0xA0, 0x5F, 0x56, 0xDA, 0xB1, 0x81, 0xC4, 0x40, 0x23, 0x3C, 
#>                 0x33, 0x6C, 0xB8, 0x71, 0x14, 0x9B, 0x51, 0xD8, 0xDC, 0x60, 
#>                 0x40, 0x27, 0x20, 0xBA, 0x4E, 0x74, 0x53, 0x4C, 0x1A, 0x83, 
#>                 0xA7, 0x9D, 0xB9, 0x9D, 0xB6, 0x96, 0xB2, 0x29, 0x10, 0x25, 
#>                 0x34, 0xD4, 0x48, 0x28, 0xC2, 0x48, 0x42, 0x12, 0xA3, 0xB0, 
#>                 0x09, 0x9E, 0x21, 0x97, 0x16, 0xF5, 0x65, 0x7C, 0x52, 0x6A, 
#>                 0x9F, 0x2C, 0x1C, 0x2D, 0x51, 0xA3, 0xDD, 0xE6, 0x7D, 0x4D, 
#>                 0xC5, 0xE0, 0x59, 0x73, 0x96, 0x96, 0x0E, 0xA1, 0x0C, 0x9C, 
#>                 0x01, 0x6A, 0x51, 0x99, 0xD2, 0x83, 0x40, 0x3B, 0x3B, 0x33, 
#>                 0xD5, 0xC4, 0x78, 0x21, 0x13, 0xA4, 0x54, 0xB9, 0x75, 0x2B, 
#>                 0xCE, 0x04, 0xF7, 0xFC, 0xFC, 0x9F, 0x80, 0x2B, 0xE5, 0xC7, 
#>                 0xD6, 0x09, 0xF3, 0xC5, 0x8C, 0x36, 0xFA, 0xFD, 0x8B, 0x8C, 
#>                 0xC2, 0x7E, 0x46, 0xB5, 0xE6, 0x64, 0x0C, 0xED, 0x07
#>                 };
#>     memcpy(RAW(rw), r, sizeof(r));