Convert a Raw Vector or R object to C Code

This function converts a raw vector or R object to C code that is used in the rxode2 model

Usage

rxRawToC(raw, type = c("xz", "qs2", "qdata", "base", "bzip2"))

Arguments

raw

vector or R object to convert

type

serialization type; one of "qs2", "qdata", "base", "xz" or "bzip2".

Value

raw vector for conversion to C in the codegen procedure

Author

Matthew L. Fidler

Examples

rxRawToC(mtcars)
#>    [1] 42 5a 68 39 31 41 59 26 53 59 81 7f 5f 7e 00 04 b9 ff ff ff ff fe 71 d4
#>   [25] 77 ff ff ff f6 ff dd 3f ff ff fe 42 00 a0 34 4c 05 18 40 59 06 60 81 61
#>   [49] 0e 4b 73 50 04 76 1a 0b 6b 46 da 55 49 90 c8 91 a2 09 a8 f2 04 f5 3c 13
#>   [73] 53 26 d3 28 34 d3 4c 86 80 0d 3d 46 9a 0c 9e a1 93 c9 a3 51 91 ea 3d 1a
#>   [97] 9a 00 00 00 7a 80 d1 a0 d3 44 22 61 04 c8 6a 69 a6 4d a8 d1 a6 9a 03 40
#>  [121] 0d 34 03 43 40 00 68 00 03 4d 01 ea 00 06 99 00 1a 00 91 24 44 d1 34 6d
#>  [145] 20 03 21 a1 a7 a1 1a 64 00 d0 01 93 40 34 00 32 00 3d 41 a0 0d 00 0d 00
#>  [169] 1a 32 68 83 09 82 32 60 26 01 30 4c 10 18 00 04 c0 04 c0 08 d3 08 62 30
#>  [193] 11 a3 08 c2 68 d0 0c 99 32 08 30 98 23 26 02 60 13 04 c1 01 80 00 4c 00
#>  [217] 4c 00 8d 30 86 23 01 1a 30 8c 26 8d 00 c9 93 20 fc 81 67 d5 f1 64 d7 af
#>  [241] dd 99 c3 28 87 39 cd 06 0a 51 4c 1c e3 61 81 05 9d 00 95 ca 02 d0 36 37
#>  [265] 3d 84 a9 b6 db 60 99 33 02 1b 26 21 b2 1d 9b 2c a6 61 16 73 41 30 e0 81
#>  [289] a8 84 37 17 05 bd cd de a5 dd eb f7 b8 77 fc e3 93 57 38 0b 31 4d bd 04
#>  [313] 2b 16 8b 35 34 c9 00 8b 40 d2 42 08 18 d0 c4 d2 1a 69 5c 4d 0a 4c 58 95
#>  [337] 68 ee 6a 4e ab d3 15 1b 8a 4c 44 06 13 04 84 6d da 45 fe 0c 21 24 95 73
#>  [361] 05 4a ae 7d 8e bc 4b 15 db db 6e 10 47 32 b1 e5 c9 50 de e4 37 f2 ed f4
#>  [385] 9b 12 e7 30 50 b1 3f a0 f8 89 74 af db c8 89 6f 1c 63 21 41 78 92 43 42
#>  [409] 4b 13 14 44 22 da af 16 51 0d b1 c2 4d 8a 18 36 10 d0 9b 21 b6 89 4e 04
#>  [433] e2 a0 a7 db d0 42 54 5a 40 c1 80 94 86 9a 71 10 36 a2 08 48 a3 63 00 a6
#>  [457] 19 26 00 a7 61 d6 68 95 6d 64 91 80 35 30 55 70 c1 0d 34 34 92 1a 60 db
#>  [481] 0e cf d1 22 49 29 09 a6 a7 70 0c 1d 27 46 08 51 12 24 b6 50 81 10 0d b4
#>  [505] 31 ba 54 a2 42 f6 a2 f4 20 2a 48 7e 4b be 52 6f 10 88 48 88 00 db c7 9f
#>  [529] 00 92 f0 be 9b 30 0b d6 29 bf 98 d0 45 95 b0 e2 1a e7 02 21 42 c4 2c 11
#>  [553] 2c 46 69 a7 19 01 37 b9 ad 3b 20 f4 94 02 10 4f 6c 8b d4 bc 50 27 f1 c2
#>  [577] 2d 70 6e 28 3a cd da a0 05 4c 2d 65 d3 99 05 bb 82 97 82 0a 75 b3 a5 5e
#>  [601] 01 a7 60 07 bd 3f 1b 42 30 93 8e 27 8d 25 28 29 95 5a 32 96 47 24 c9 98
#>  [625] a9 39 47 47 10 10 73 39 46 d5 36 e2 5c b5 20 38 2e ab 71 8c 27 54 5f a4
#>  [649] 14 c4 4b d4 25 45 35 6a 55 55 02 ad 1c 6f 95 9f 0e ce 6d 08 00 db 26 fb
#>  [673] 66 d9 9f 65 85 5a 2c de f8 16 08 97 24 05 0e fa 54 cb c3 c6 f4 44 a6 3e
#>  [697] d9 f9 40 19 9d 89 4c 98 59 84 d9 2a e1 69 4c 4a c2 d4 a0 71 01 4a 29 40
#>  [721] 14 a0 6a 71 f9 e7 a0 a2 08 39 38 ea 24 21 10 56 36 d8 aa ca 54 95 31 b3
#>  [745] dd 74 6b 57 96 3b b7 b3 54 aa b3 3c 4c d6 86 71 7c 61 5b f2 b0 af 4a 29
#>  [769] d2 2b 26 85 42 59 49 04 c3 64 87 0c 6d 36 4e c1 44 4f 24 29 32 61 80 a8
#>  [793] 0a c1 cc 6f 4a 97 9b 52 73 ba 88 03 58 b5 11 77 3e 86 7a 70 6b f0 5c 18
#>  [817] 93 87 29 04 24 06 2b 91 1b 18 cc 4e 1c d8 42 95 9f 37 82 99 a8 a4 d5 14
#>  [841] a7 13 d4 89 92 f1 06 26 4d 28 07 00 04 42 16 e8 aa 3d 97 30 69 ad 9a 9b
#>  [865] 07 83 e8 35 d6 80 62 a0 38 80 eb dc 73 0b 08 0e 20 28 47 1e ad 5c ec dc
#>  [889] d0 ea 85 d5 4a af 54 a2 c7 2b 72 77 6d 83 08 3d cc d0 cf 76 be 1c 46 3c
#>  [913] 36 2b 5f 4d a2 e1 c4 10 70 d4 2b ee 12 e0 d1 d1 bd a7 b2 a0 22 ea 97 65
#>  [937] 8c f3 a4 90 44 2d 32 eb e3 89 68 da bc ce fc db 17 d5 44 ba eb 4a 23 16
#>  [961] b3 a7 65 00 43 fb 23 f3 79 9c b9 dd f1 ec c3 e1 b9 bf d0 75 b0 cc 59 17
#>  [985] d2 63 5a c0 9d 3d 64 4c 44 95 20 4c 84 ca 2d 9f bc 79 af c0 8a 2d fc 56
#> [1009] 81 a6 14 a8 2d 87 e7 6f 6f d9 44 6a 83 2a fd ff 3a 11 60 5c bd 8d 41 2a
#> [1033] c8 a5 c2 51 71 5e 27 5b 2c 5b e5 a9 83 77 65 cf 8d db 43 a1 b3 63 89 72
#> [1057] a0 64 89 38 37 11 b7 03 21 94 c7 23 89 5d 84 b8 50 4f 93 28 e9 c2 5c 1a
#> [1081] 77 cd 02 89 79 02 cd 76 e0 c4 5e e1 3d 85 5a 22 b3 85 88 a6 c6 92 e2 75
#> [1105] 7e 8e 46 b0 eb d6 94 1a 20 ca e9 cb 10 5c f8 df 66 4a 90 c1 80 9e 81 bf
#> [1129] d2 b8 42 36 b8 76 89 e2 95 01 d3 41 8a 42 40 65 26 2e 05 44 d2 12 12 85
#> [1153] e1 65 af e1 33 6d 86 4a 52 b4 e9 27 f4 0e 5d 22 91 d8 88 41 d6 f3 92 0f
#> [1177] 6a 06 7c 3e 2e af af eb 29 a7 09 08 7f 6c 2f e3 88 c1 08 18 b6 88 bb d9
#> [1201] 5b 30 1c b9 d4 21 fd 09 7b 1e b5 3a ce eb 20 95 d5 f5 87 2f c3 bc 6f 1b
#> [1225] 20 55 cd 5e 34 8c 6a 3f 10 05 7e 5b b3 0f 1f f8 ab 71 8a 98 a6 99 5a 8d
#> [1249] 8c 62 da 71 e7 c7 3f d7 26 8a 48 2d 2f 4c 8b e8 ce f5 51 80 3c 96 14 6a
#> [1273] 16 36 d6 42 32 46 48 b1 37 22 1d 10 40 14 72 86 b8 3a 94 b6 ac 24 b4 48
#> [1297] 49 69 ff 17 72 45 38 50 90 81 7f 5f 7e