Convert a Raw Vector or R object to C Code

This function converts a raw vector or R object to C code that is used in the rxode2 model

Usage

rxRawToC(raw, type = c("xz", "qs2", "qdata", "base", "bzip2"))

Arguments

raw

raw vector or R object to convert

type

serialization type; one of "qs2", "qdata", "base", "xz" or "bzip2".

Value

character string of C code

Author

Matthew L. Fidler

Examples

message(rxRawToC(mtcars))
#>     SEXP rw    = PROTECT(Rf_allocVector(RAWSXP, 1148));pro++;
#>     unsigned char r[]={
#>                 0xFD, 0x37, 0x7A, 0x58, 0x5A, 0x00, 0x00, 0x01, 0x69, 0x22, 
#>                 0xDE, 0x36, 0x02, 0x00, 0x21, 0x01, 0x1C, 0x00, 0x00, 0x00, 
#>                 0x10, 0xCF, 0x58, 0xCC, 0xE0, 0x0E, 0xDE, 0x04, 0x3F, 0x5D, 
#>                 0x00, 0x2C, 0x02, 0x7C, 0x19, 0xF9, 0xAF, 0xF4, 0x1D, 0x47, 
#>                 0x1B, 0x6D, 0xDA, 0x57, 0x86, 0xCA, 0x28, 0x98, 0xA5, 0x72, 
#>                 0x2D, 0xB5, 0x8A, 0x1B, 0xCC, 0xF1, 0x0B, 0x92, 0x01, 0x0D, 
#>                 0x46, 0xF0, 0x3A, 0x43, 0x43, 0xC7, 0x73, 0xAF, 0x7D, 0x11, 
#>                 0x82, 0x8A, 0x46, 0xCF, 0x56, 0x2C, 0x3C, 0x67, 0xEA, 0x82, 
#>                 0xEC, 0xFA, 0x8A, 0x28, 0xA1, 0x79, 0x6A, 0xA4, 0xFB, 0x97, 
#>                 0x47, 0x77, 0xD3, 0x3B, 0x9C, 0x72, 0x76, 0x94, 0x62, 0xEE, 
#>                 0x61, 0x91, 0x8E, 0x35, 0x54, 0xB9, 0x08, 0x3A, 0xC9, 0xCF, 
#>                 0xE9, 0x6B, 0xDD, 0x3F, 0x9E, 0x00, 0x1E, 0xD1, 0x06, 0xB7, 
#>                 0xCD, 0x3F, 0x39, 0x23, 0x51, 0x77, 0x9E, 0x57, 0x40, 0x84, 
#>                 0x5F, 0x7D, 0x60, 0x72, 0x59, 0xEC, 0x58, 0x0A, 0xD2, 0xBE, 
#>                 0x90, 0x3E, 0x9A, 0x25, 0xCB, 0xF6, 0x4B, 0x50, 0x8F, 0xAF, 
#>                 0x5A, 0x49, 0x9C, 0xA4, 0x98, 0x34, 0x7C, 0x00, 0x74, 0x8C, 
#>                 0x1A, 0xBF, 0x73, 0x30, 0x98, 0x3B, 0xA1, 0x44, 0xF6, 0xD7, 
#>                 0x86, 0x06, 0xBA, 0xF1, 0x42, 0x10, 0xB7, 0xA4, 0x0C, 0xBA, 
#>                 0x7A, 0xFE, 0xC6, 0x0A, 0x18, 0xA0, 0xED, 0x4D, 0x56, 0x0A, 
#>                 0xAF, 0x1D, 0xAC, 0x9B, 0x49, 0x19, 0x87, 0x36, 0x0C, 0xA7, 
#>                 0x6C, 0x0A, 0xF3, 0xA1, 0x62, 0xDA, 0x66, 0xCD, 0xC3, 0xD7, 
#>                 0x0A, 0x6F, 0x66, 0xBE, 0xA1, 0xDC, 0x52, 0x3D, 0xC2, 0xEE, 
#>                 0x09, 0x38, 0x59, 0xAA, 0xEA, 0xE8, 0x27, 0x3A, 0x5A, 0xFD, 
#>                 0x5B, 0x12, 0xA3, 0x3B, 0x86, 0x76, 0x8E, 0x78, 0x31, 0x4E, 
#>                 0xA5, 0x13, 0x7D, 0xEA, 0x62, 0x56, 0x2F, 0xC9, 0x25, 0xAB, 
#>                 0x06, 0x78, 0x0A, 0x4E, 0x21, 0xFA, 0xFC, 0xAA, 0x96, 0xC6, 
#>                 0x2A, 0x0D, 0x69, 0x33, 0x29, 0x0E, 0x7D, 0xAF, 0xF9, 0x30, 
#>                 0xC4, 0xDC, 0x49, 0x6E, 0xA0, 0x3C, 0x28, 0xBA, 0x34, 0xE2, 
#>                 0xCE, 0x3B, 0x8C, 0xA8, 0x21, 0x8C, 0x9B, 0xE6, 0x5F, 0x74, 
#>                 0x7F, 0x53, 0xB6, 0x19, 0x07, 0x75, 0x76, 0x79, 0xF3, 0x89, 
#>                 0x00, 0xCE, 0xDB, 0x58, 0xF4, 0x3C, 0xED, 0x4F, 0x20, 0xCD, 
#>                 0x70, 0xF9, 0x42, 0x97, 0x2C, 0xE8, 0x8E, 0xE0, 0x91, 0xC7, 
#>                 0x9D, 0x26, 0x81, 0x89, 0x0E, 0x8F, 0x8C, 0x46, 0x49, 0x26, 
#>                 0x9A, 0xCD, 0x5C, 0xDA, 0x1D, 0x6A, 0x0E, 0x52, 0x65, 0x76, 
#>                 0xD3, 0xA2, 0xEB, 0xCA, 0xD2, 0xA9, 0x02, 0x25, 0xB2, 0x30, 
#>                 0x28, 0x99, 0x9F, 0x36, 0x64, 0x36, 0xC5, 0x7B, 0x7B, 0xD8, 
#>                 0x6F, 0x2A, 0x0A, 0xED, 0x28, 0xAE, 0x16, 0x56, 0x87, 0xB0, 
#>                 0x9F, 0x25, 0xD8, 0x4B, 0xCD, 0x26, 0x0C, 0x19, 0x88, 0x3F, 
#>                 0x63, 0x27, 0xF1, 0x19, 0xBA, 0x0C, 0x15, 0x3A, 0x79, 0x64, 
#>                 0xA0, 0x3F, 0x09, 0xB1, 0x64, 0x9A, 0x27, 0x10, 0x09, 0x49, 
#>                 0x0C, 0x7B, 0x47, 0xE4, 0x47, 0xC6, 0xC2, 0xB4, 0x28, 0xBC, 
#>                 0xFE, 0x68, 0xC2, 0x1D, 0x33, 0x56, 0xBD, 0xF9, 0xAF, 0xB9, 
#>                 0xC2, 0x97, 0x8B, 0x5F, 0xEA, 0x05, 0x1E, 0x89, 0xD2, 0xE6, 
#>                 0xA0, 0x24, 0xD2, 0x1D, 0x27, 0xD8, 0x3D, 0x32, 0x09, 0xD8, 
#>                 0xFD, 0xCB, 0xB4, 0x22, 0xA2, 0x0C, 0x92, 0x7D, 0x75, 0xBB, 
#>                 0x7E, 0x5A, 0xE1, 0xCA, 0x68, 0x94, 0x1A, 0xD8, 0xE7, 0xC9, 
#>                 0x80, 0x3C, 0x63, 0x87, 0x79, 0x4B, 0x5E, 0x6D, 0xCB, 0xFB, 
#>                 0x7B, 0x33, 0xA4, 0x6C, 0x01, 0xCA, 0x6B, 0x9C, 0x48, 0x69, 
#>                 0xA6, 0x33, 0x82, 0x6E, 0x82, 0x38, 0x1E, 0xC1, 0xFB, 0x6D, 
#>                 0x65, 0x30, 0x67, 0xE0, 0x9F, 0xDF, 0x49, 0x59, 0xD0, 0xE2, 
#>                 0xEC, 0xFA, 0x8D, 0x56, 0x86, 0x22, 0xEF, 0xEE, 0xFF, 0xFF, 
#>                 0xF8, 0x06, 0x6B, 0x9B, 0x77, 0xFC, 0x8D, 0xC9, 0x7B, 0xF9, 
#>                 0xC9, 0xCF, 0x81, 0x47, 0x56, 0xD7, 0xD6, 0x1A, 0x5D, 0x17, 
#>                 0x70, 0x5E, 0x80, 0x07, 0xA3, 0x50, 0xDC, 0xA6, 0xF1, 0xAA, 
#>                 0xC6, 0x7D, 0xCF, 0xEC, 0xD7, 0xF6, 0xF9, 0x90, 0x98, 0x09, 
#>                 0x3D, 0x67, 0xD7, 0xF2, 0x90, 0x26, 0xD0, 0xF8, 0xFF, 0x0B, 
#>                 0x13, 0xF9, 0x8D, 0x27, 0x1E, 0x3A, 0xA4, 0x52, 0x57, 0x44, 
#>                 0x99, 0xBE, 0x25, 0x8F, 0x26, 0x3B, 0x25, 0x5C, 0xDF, 0x3E, 
#>                 0x13, 0x3A, 0x58, 0x51, 0x90, 0x0A, 0x81, 0x82, 0xF4, 0x0D, 
#>                 0x54, 0x16, 0xF0, 0xFE, 0x57, 0x7D, 0x8A, 0x13, 0xF4, 0xFC, 
#>                 0x7D, 0xA5, 0x57, 0xD8, 0x35, 0xBA, 0x13, 0xFE, 0x57, 0x5C, 
#>                 0xD1, 0xFE, 0x3F, 0x80, 0x14, 0xAC, 0x9B, 0x63, 0xCA, 0xB8, 
#>                 0x51, 0xD9, 0x31, 0x2F, 0x7A, 0x9E, 0x5F, 0x33, 0xD2, 0x68, 
#>                 0x77, 0xEC, 0x61, 0xA6, 0x73, 0xF0, 0x21, 0x6B, 0x98, 0x67, 
#>                 0xBA, 0xCF, 0x7C, 0x2B, 0x7B, 0x20, 0x34, 0x5A, 0xFD, 0xB8, 
#>                 0xCD, 0xF3, 0x03, 0x3C, 0xEA, 0x3E, 0x51, 0xAD, 0x23, 0x15, 
#>                 0x3E, 0x21, 0xB7, 0x79, 0x55, 0x51, 0xFB, 0x25, 0xD6, 0xA8, 
#>                 0x50, 0x6B, 0xAD, 0xB8, 0x10, 0x40, 0x21, 0x29, 0x82, 0x57, 
#>                 0xFF, 0x46, 0x3D, 0xCC, 0x22, 0x9A, 0x93, 0x3A, 0xB4, 0x1D, 
#>                 0xFD, 0xF7, 0xAC, 0xF7, 0x8D, 0xD5, 0xDA, 0x8C, 0x2E, 0x4F, 
#>                 0x73, 0x2C, 0x02, 0xE3, 0x1A, 0x3B, 0xF2, 0xA9, 0xC1, 0x39, 
#>                 0xB3, 0xA4, 0x1D, 0x67, 0xC1, 0xBC, 0x9C, 0xDB, 0x58, 0x8F, 
#>                 0x16, 0x92, 0xE0, 0x6F, 0x03, 0x91, 0xD7, 0xB3, 0x4F, 0x27, 
#>                 0x3B, 0x2F, 0xEF, 0xF8, 0xE3, 0x14, 0x9E, 0xAC, 0xAE, 0x95, 
#>                 0x3F, 0x8F, 0xE8, 0xDF, 0x36, 0x65, 0x63, 0xC6, 0xE8, 0xA4, 
#>                 0x68, 0xEB, 0xD0, 0x37, 0x0E, 0xA8, 0x77, 0x04, 0xC8, 0x8D, 
#>                 0xD8, 0x9E, 0xF0, 0x2E, 0xDF, 0x3B, 0x26, 0x95, 0x26, 0x02, 
#>                 0x6D, 0xFD, 0x4B, 0x6D, 0x70, 0x8C, 0x9E, 0xF8, 0xB0, 0x57, 
#>                 0x29, 0xFA, 0x0F, 0xC9, 0xC0, 0xDD, 0xB8, 0x86, 0x11, 0x0F, 
#>                 0xFA, 0x3B, 0xF0, 0xB7, 0x17, 0x31, 0x11, 0xE5, 0x39, 0x9A, 
#>                 0x12, 0xF5, 0x5C, 0xA9, 0x1E, 0x36, 0x14, 0x17, 0x02, 0x52, 
#>                 0x51, 0x39, 0x23, 0xDF, 0x0E, 0xC3, 0x8C, 0x83, 0x78, 0xD4, 
#>                 0x04, 0xF2, 0x96, 0xF6, 0xD9, 0xCF, 0x47, 0x7A, 0x6C, 0x8F, 
#>                 0xD1, 0xF1, 0x9C, 0x30, 0x34, 0xBE, 0xE8, 0xF1, 0xAD, 0x01, 
#>                 0xA1, 0x65, 0x19, 0x8B, 0x1C, 0x27, 0x97, 0x40, 0x4F, 0x6D, 
#>                 0x6C, 0x10, 0xA6, 0xD6, 0xF6, 0xBF, 0xF3, 0xBA, 0x3E, 0x26, 
#>                 0x5D, 0x47, 0x80, 0x0C, 0xA1, 0x51, 0x76, 0xD2, 0xC3, 0x7A, 
#>                 0x2A, 0x66, 0x06, 0x40, 0x8C, 0xAF, 0x01, 0x88, 0xEE, 0xF5, 
#>                 0xDC, 0xB0, 0xA4, 0x4B, 0xF6, 0x6D, 0xA2, 0x8F, 0xE4, 0x06, 
#>                 0x13, 0xC9, 0x4B, 0x8C, 0x55, 0x0B, 0xE0, 0xAD, 0x84, 0x16, 
#>                 0xF4, 0xDC, 0x3D, 0xA3, 0x1D, 0x99, 0x13, 0xB4, 0x51, 0xC2, 
#>                 0xB8, 0x2B, 0x25, 0x24, 0x55, 0x83, 0xC1, 0xB7, 0x60, 0x8B, 
#>                 0x79, 0xA0, 0x64, 0x6C, 0xC0, 0x65, 0xEF, 0x50, 0x53, 0xD2, 
#>                 0xD0, 0x4C, 0x65, 0x0D, 0x29, 0x1B, 0xD2, 0xEF, 0xD8, 0xB6, 
#>                 0x84, 0xAA, 0x6F, 0xAE, 0xB7, 0x3A, 0xFB, 0x83, 0xA2, 0x77, 
#>                 0xAC, 0xFF, 0x41, 0x31, 0xA2, 0x5F, 0xBB, 0x42, 0x14, 0x63, 
#>                 0x93, 0xF0, 0x45, 0x50, 0xF3, 0xC1, 0x3C, 0xDE, 0xCD, 0x70, 
#>                 0xF3, 0x9B, 0x73, 0x01, 0x09, 0x8C, 0x1F, 0x67, 0xF2, 0x88, 
#>                 0x88, 0xE4, 0x7E, 0xAD, 0xB7, 0x32, 0x9D, 0x32, 0xB5, 0xB1, 
#>                 0x66, 0xA6, 0x47, 0xEE, 0x3C, 0x1C, 0x78, 0x4D, 0x2E, 0x6F, 
#>                 0xFB, 0xD4, 0xDA, 0x20, 0x2D, 0x30, 0x65, 0x3C, 0xBD, 0xE8, 
#>                 0xA7, 0x24, 0xFD, 0x13, 0x91, 0xCC, 0x4A, 0xD3, 0xDF, 0x24, 
#>                 0x7E, 0x51, 0x3D, 0xDC, 0x54, 0x3A, 0xF9, 0xBA, 0xDB, 0xD4, 
#>                 0x1B, 0x7D, 0x3F, 0x58, 0x1C, 0x4C, 0x8D, 0x94, 0xE6, 0xAA, 
#>                 0x76, 0x67, 0xC8, 0xA9, 0x90, 0x8B, 0xE0, 0xED, 0x9B, 0x32, 
#>                 0x8C, 0xE3, 0xC0, 0x77, 0x6C, 0x68, 0x0B, 0x9D, 0xA0, 0x2C, 
#>                 0x63, 0x35, 0xFC, 0xB2, 0x56, 0x16, 0x3E, 0xAD, 0x62, 0x03, 
#>                 0xF8, 0xAB, 0x95, 0x6A, 0x52, 0x5E, 0x4F, 0x6F, 0x44, 0xF1, 
#>                 0x11, 0xB8, 0x1A, 0x1F, 0x56, 0xE1, 0x40, 0x49, 0x01, 0xB1, 
#>                 0x2E, 0x5A, 0x79, 0xBA, 0x8B, 0xC2, 0x4D, 0x1A, 0xE3, 0x74, 
#>                 0x2E, 0xD5, 0xAA, 0x2E, 0xC0, 0x13, 0x94, 0xFB, 0x6A, 0x82, 
#>                 0xB2, 0xB1, 0x75, 0x03, 0xAA, 0xDE, 0xC0, 0x00, 0x00, 0x00, 
#>                 0x95, 0x5D, 0x00, 0x11, 0x00, 0x01, 0xD7, 0x08, 0xDF, 0x1D, 
#>                 0x00, 0x00, 0x67, 0x2A, 0xD7, 0x96, 0x3E, 0x30, 0x0D, 0x8B, 
#>                 0x02, 0x00, 0x00, 0x00, 0x00, 0x01, 0x59, 0x5A
#>                 };
#>     memcpy(RAW(rw), r, sizeof(r));