Simulate from a (truncated) multivariate normal

This is simulated with the fast, thread-safe threefry simulator and can use multiple cores to generate the random deviates.

Usage

rxRmvn(
  n,
  mu = NULL,
  sigma,
  lower = -Inf,
  upper = Inf,
  ncores = 1,
  isChol = FALSE,
  keepNames = TRUE,
  a = 0.4,
  tol = 2.05,
  nlTol = 1e-10,
  nlMaxiter = 100L
)

Arguments

n

Number of random row vectors to be simulated OR the matrix to use for simulation (faster).

mu

mean vector

sigma

Covariance matrix for multivariate normal or a list of covariance matrices. If a list of covariance matrix, each matrix will simulate n matrices and combine them to a full matrix

lower

is a vector of the lower bound for the truncated multivariate norm

upper

is a vector of the upper bound for the truncated multivariate norm

ncores

Number of cores used in the simulation

isChol

A boolean indicating if sigma is a cholesky decomposition of the covariance matrix.

keepNames

Keep the names from either the mean or covariance matrix.

a

threshold for switching between methods; They can be tuned for maximum speed; There are three cases that are considered:

case 1: a < l < u

case 2: l < u < -a

case 3: otherwise

where l=lower and u = upper

tol

When case 3 is used from the above possibilities, the tol value controls the acceptance rejection and inverse-transformation;

When abs(u-l)>tol, uses accept-reject from randn

nlTol

Tolerance for newton line-search

nlMaxiter

Maximum iterations for newton line-search

Value

If n==integer (default) the output is an (n x d) matrix where the i-th row is the i-th simulated vector.

If is.matrix(n) then the random vector are store in n, which is provided by the user, and the function returns NULL invisibly.

References

John K. Salmon, Mark A. Moraes, Ron O. Dror, and David E. Shaw (2011). Parallel Random Numbers: As Easy as 1, 2, 3. D. E. Shaw Research, New York, NY 10036, USA.

The thread safe multivariate normal was inspired from the mvnfast package by Matteo Fasiolo https://CRAN.R-project.org/package=mvnfast

The concept of the truncated multivariate normal was taken from Zdravko Botev Botev (2017) doi:10.1111/rssb.12162 and Botev and L'Ecuyer (2015) doi:10.1109/WSC.2015.7408180 and converted to thread safe simulation;

Author

Matthew Fidler, Zdravko Botev and some from Matteo Fasiolo

Examples

## From mvnfast
## Unlike mvnfast, uses threefry simulation

d <- 5
mu <- 1:d

# Creating covariance matrix
tmp <- matrix(rnorm(d^2), d, d)
mcov <- tcrossprod(tmp, tmp)

set.seed(414)
rxRmvn(4, 1:d, mcov)
#>             [,1]      [,2]      [,3]      [,4]     [,5]
#> [1,]  0.44551640  1.135299 4.1930872 5.3342689 3.294011
#> [2,] -0.07398444 -1.636221 0.4627023 0.5493825 4.363167
#> [3,]  4.03533473 -1.139955 1.6960875 3.4654586 6.553250
#> [4,] -0.14596643  3.120244 1.4508551 1.2927681 5.818908

set.seed(414)
rxRmvn(4, 1:d, mcov)
#>             [,1]      [,2]      [,3]      [,4]     [,5]
#> [1,]  0.44551640  1.135299 4.1930872 5.3342689 3.294011
#> [2,] -0.07398444 -1.636221 0.4627023 0.5493825 4.363167
#> [3,]  4.03533473 -1.139955 1.6960875 3.4654586 6.553250
#> [4,] -0.14596643  3.120244 1.4508551 1.2927681 5.818908

set.seed(414)
rxRmvn(4, 1:d, mcov, ncores = 2) # r.v. generated on the second core are different
#>             [,1]      [,2]       [,3]      [,4]     [,5]
#> [1,]  0.44551640 5.4982942  3.3790994 2.1528232 4.125642
#> [2,]  0.08902736 2.6807564  3.8555336 3.9941609 3.851401
#> [3,] -0.07398444 0.2691971 -0.5120344 0.1186391 6.680531
#> [4,]  1.45423479 2.0331871  2.2334239 2.1622739 4.851160

###### Here we create the matrix that will hold the simulated
#  random variables upfront.
A <- matrix(NA, 4, d)
class(A) <- "numeric" # This is important. We need the elements of A to be of class "numeric".

set.seed(414)
rxRmvn(A, 1:d, mcov, ncores = 2) # This returns NULL ...
A # ... but the result is here
#>             [,1]      [,2]       [,3]      [,4]     [,5]
#> [1,]  0.44551640 5.4982942  3.3790994 2.1528232 4.125642
#> [2,]  0.08902736 2.6807564  3.8555336 3.9941609 3.851401
#> [3,] -0.07398444 0.2691971 -0.5120344 0.1186391 6.680531
#> [4,]  1.45423479 2.0331871  2.2334239 2.1622739 4.851160

## You can also simulate from a truncated normal:

rxRmvn(10, 1:d, mcov, lower = 1:d - 1, upper = 1:d + 1)
#>            [,1]     [,2]     [,3]     [,4]     [,5]
#>  [1,] 1.3653338 1.913500 2.884704 4.638532 4.970611
#>  [2,] 0.2552094 1.829579 2.804913 4.462161 5.188182
#>  [3,] 1.1799136 2.349490 3.155009 4.984464 4.397339
#>  [4,] 1.1431108 2.364839 2.743327 4.907312 4.580010
#>  [5,] 0.8902283 2.002904 3.183637 3.298237 5.390240
#>  [6,] 1.2708036 2.143880 3.167478 4.401943 4.566324
#>  [7,] 0.8913711 1.983748 2.772283 4.462135 5.133781
#>  [8,] 1.7430533 1.608873 2.990598 3.182227 5.089337
#>  [9,] 1.6952039 1.876700 2.842330 4.519336 4.667154
#> [10,] 0.1794205 2.316141 2.990005 4.659106 5.336159

# You can also simulate from different matrices (if they match
# dimensions) by using a list of matrices.

matL <- lapply(1:4, function(...) {
  tmp <- matrix(rnorm(d^2), d, d)
  tcrossprod(tmp, tmp)
})

rxRmvn(4, setNames(1:d, paste0("a", 1:d)), matL)
#>               a1         a2         a3       a4       a5
#>  [1,]  2.1027689 -3.1277802 -1.7058664 3.877449 6.160238
#>  [2,] -0.1528749  0.3372425  1.4079033 3.204818 6.771887
#>  [3,] -1.6178920  2.4185331  4.2854089 4.204923 6.115151
#>  [4,]  5.2850177  6.0162640  1.1522397 6.009218 8.423744
#>  [5,] -0.8389005  1.8108360  3.9390837 2.635257 3.078044
#>  [6,]  2.4044119  0.2584573  2.4396218 4.687467 7.747550
#>  [7,]  7.2635180  5.5533803 -0.2744354 9.627871 4.788654
#>  [8,] -0.3358839 -1.7501935  4.1309667 2.961912 6.318101
#>  [9,]  3.4028706  1.1584222  4.5627803 3.408984 3.627014
#> [10,]  3.4927361 -1.5055971  1.0890217 2.867575 4.036955
#> [11,]  2.2089477 -1.0108105 -0.6234824 1.745858 4.178149
#> [12,]  1.9045759 -3.5353019  3.6738287 3.085417 2.755276
#> [13,] -1.0910176  1.6896066 -0.1344867 5.810620 3.303876
#> [14,]  0.7032309  2.5654064  2.7642640 3.706315 5.057585
#> [15,] -3.4067706  2.3051387  4.4260234 6.692458 8.170364
#> [16,] -3.7727314  4.2737997 -0.8379328 6.990844 1.255809