Simulate geometric variable from threefry generator

Care should be taken with this method not to encounter the birthday problem, described https://www.johndcook.com/blog/2016/01/29/random-number-generator-seed-mistakes/. Since the sitmo threefry, this currently generates one random deviate from the uniform distribution to seed the engine threefry and then run the code.

rxgeom(prob, n = 1L, ncores = 1L)

Arguments

prob

probability of success in each trial. 0 < prob <= 1.

n

number of observations. If length(n) > 1, the length is taken to be the number required.

ncores

Number of cores for the simulation

rxnorm simulates using the threefry sitmo generator; rxnormV uses the vandercorput generator

Value

geometric random deviates

Details

Therefore, a simple call to the random number generated followed by a second call to random number generated may have identical seeds. As the number of random number generator calls are increased the probability that the birthday problem will increase.

The key to avoid this problem is to either run all simulations in the rxode2 environment once (therefore one seed or series of seeds for the whole simulation), pre-generate all random variables used for the simulation, or seed the rxode2 engine with rxSetSeed()

Also care should be made that the computer you will be running on can run the same number of cores as you are running so they can reproduce your results.

Examples

# \donttest{

## Use threefry engine

rxgeom(0.5, n = 10) # with rxgeom you have to explicitly state n
#>  [1] 1 2 1 0 2 0 1 1 0 0
rxgeom(0.25, n = 10, ncores = 2) # You can parallelize the simulation using openMP
#>  [1] 9 0 4 1 7 1 0 4 0 5

rxgeom(0.75)
#> [1] 1

## This example uses `rxgeom` directly in the model

rx <- rxode2({
  a <- rxgeom(0.24)
})
#>  

et <- et(1, id = 1:2)

s <- rxSolve(rx, et)
# }